% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_control.R
\docType{class}
\name{AccessControl}
\alias{AccessControl}
\title{AccessControl}
\format{
An \code{R6Class} generator object
}
\description{
AccessControl Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{canShare}}{character Whether the recipient can share the role further.}

\item{\code{role}}{character The role of the user on this entity.}

\item{\code{userId}}{character The identifier of the user that has access to this entity.}

\item{\code{username}}{character The username of the user that has access to the entity.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AccessControl-new}{\code{AccessControl$new()}}
\item \href{#method-AccessControl-validate}{\code{AccessControl$validate()}}
\item \href{#method-AccessControl-toJSON}{\code{AccessControl$toJSON()}}
\item \href{#method-AccessControl-fromJSON}{\code{AccessControl$fromJSON()}}
\item \href{#method-AccessControl-clone}{\code{AccessControl$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessControl-new"></a>}}
\if{latex}{\out{\hypertarget{method-AccessControl-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessControl$new(
  canShare = NULL,
  role = NULL,
  userId = NULL,
  username = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{canShare}}{Whether the recipient can share the role further.}

\item{\code{role}}{The role of the user on this entity.}

\item{\code{userId}}{The identifier of the user that has access to this entity.}

\item{\code{username}}{The username of the user that has access to the entity.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessControl-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AccessControl-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessControl$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessControl-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AccessControl-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessControl$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessControl-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AccessControl-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessControl$fromJSON(AccessControlJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AccessControlJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessControl-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AccessControl-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessControl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
