% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_control_with_grant.R
\docType{class}
\name{AccessControlWithGrant}
\alias{AccessControlWithGrant}
\title{AccessControlWithGrant}
\format{
An \code{R6Class} generator object
}
\description{
AccessControlWithGrant Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{canShare}}{character Whether the recipient can share the role further.}

\item{\code{id}}{character The identifier of the recipient.}

\item{\code{name}}{character The name of the recipient.}

\item{\code{role}}{character The role of the recipient on this entity.}

\item{\code{shareRecipientType}}{character The type of the recipient.}

\item{\code{userFullName}}{character \link{optional} Full name of the recipient user.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AccessControlWithGrant-new}{\code{AccessControlWithGrant$new()}}
\item \href{#method-AccessControlWithGrant-validate}{\code{AccessControlWithGrant$validate()}}
\item \href{#method-AccessControlWithGrant-toJSON}{\code{AccessControlWithGrant$toJSON()}}
\item \href{#method-AccessControlWithGrant-fromJSON}{\code{AccessControlWithGrant$fromJSON()}}
\item \href{#method-AccessControlWithGrant-clone}{\code{AccessControlWithGrant$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessControlWithGrant-new"></a>}}
\if{latex}{\out{\hypertarget{method-AccessControlWithGrant-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessControlWithGrant$new(
  canShare = NULL,
  id = NULL,
  name = NULL,
  role = NULL,
  shareRecipientType = NULL,
  userFullName = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{canShare}}{Whether the recipient can share the role further.}

\item{\code{id}}{The identifier of the recipient.}

\item{\code{name}}{The name of the recipient.}

\item{\code{role}}{The role of the recipient on this entity.}

\item{\code{shareRecipientType}}{The type of the recipient.}

\item{\code{userFullName}}{Full name of the recipient user.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessControlWithGrant-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AccessControlWithGrant-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessControlWithGrant$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessControlWithGrant-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AccessControlWithGrant-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessControlWithGrant$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessControlWithGrant-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AccessControlWithGrant-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessControlWithGrant$fromJSON(
  AccessControlWithGrantJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AccessControlWithGrantJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccessControlWithGrant-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AccessControlWithGrant-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccessControlWithGrant$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
