% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_over_time_bucket.R
\docType{class}
\name{AccuracyOverTimeBucket}
\alias{AccuracyOverTimeBucket}
\title{AccuracyOverTimeBucket}
\format{
An \code{R6Class} generator object
}
\description{
AccuracyOverTimeBucket Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{period}}{\link{TimeRange}}

\item{\code{sampleSize}}{integer Number of predictions used to calculate the metric.}

\item{\code{value}}{numeric Value of the metric, null if no value.}

\item{\code{valuePerClass}}{object \link{optional} A dict keyed by class names with metric calculated for specific classes as values, if targetClasses is set.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AccuracyOverTimeBucket$new()}}
\item \href{#method-validate}{\code{AccuracyOverTimeBucket$validate()}}
\item \href{#method-toJSON}{\code{AccuracyOverTimeBucket$toJSON()}}
\item \href{#method-fromJSON}{\code{AccuracyOverTimeBucket$fromJSON()}}
\item \href{#method-clone}{\code{AccuracyOverTimeBucket$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccuracyOverTimeBucket$new(
  period = NULL,
  sampleSize = NULL,
  value = NULL,
  valuePerClass = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sampleSize}}{Number of predictions used to calculate the metric.}

\item{\code{value}}{Value of the metric, null if no value.}

\item{\code{valuePerClass}}{A dict keyed by class names with metric calculated for specific classes as values, if targetClasses is set.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccuracyOverTimeBucket$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccuracyOverTimeBucket$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccuracyOverTimeBucket$fromJSON(
  AccuracyOverTimeBucketJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AccuracyOverTimeBucketJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccuracyOverTimeBucket$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
