% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_over_time_plots_statistics.R
\docType{class}
\name{AccuracyOverTimePlotsStatistics}
\alias{AccuracyOverTimePlotsStatistics}
\title{AccuracyOverTimePlotsStatistics}
\format{
An \code{R6Class} generator object
}
\description{
AccuracyOverTimePlotsStatistics Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{durbinWatson}}{numeric The Durbin-Watson statistic for the chart data. Value is between 0 and 4. Durbin-Watson statistic is a test statistic used to detect the presence of autocorrelation at lag 1 in the residuals (prediction errors) from a regression analysis. More info https://wikipedia.org/wiki/Durbin\%E2\%80\%93Watson_statistic}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AccuracyOverTimePlotsStatistics-new}{\code{AccuracyOverTimePlotsStatistics$new()}}
\item \href{#method-AccuracyOverTimePlotsStatistics-validate}{\code{AccuracyOverTimePlotsStatistics$validate()}}
\item \href{#method-AccuracyOverTimePlotsStatistics-toJSON}{\code{AccuracyOverTimePlotsStatistics$toJSON()}}
\item \href{#method-AccuracyOverTimePlotsStatistics-fromJSON}{\code{AccuracyOverTimePlotsStatistics$fromJSON()}}
\item \href{#method-AccuracyOverTimePlotsStatistics-clone}{\code{AccuracyOverTimePlotsStatistics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccuracyOverTimePlotsStatistics-new"></a>}}
\if{latex}{\out{\hypertarget{method-AccuracyOverTimePlotsStatistics-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccuracyOverTimePlotsStatistics$new(
  durbinWatson = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{durbinWatson}}{The Durbin-Watson statistic for the chart data. Value is between 0 and 4. Durbin-Watson statistic is a test statistic used to detect the presence of autocorrelation at lag 1 in the residuals (prediction errors) from a regression analysis. More info https://wikipedia.org/wiki/Durbin\%E2\%80\%93Watson_statistic}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccuracyOverTimePlotsStatistics-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AccuracyOverTimePlotsStatistics-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccuracyOverTimePlotsStatistics$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccuracyOverTimePlotsStatistics-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AccuracyOverTimePlotsStatistics-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccuracyOverTimePlotsStatistics$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccuracyOverTimePlotsStatistics-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AccuracyOverTimePlotsStatistics-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccuracyOverTimePlotsStatistics$fromJSON(
  AccuracyOverTimePlotsStatisticsJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AccuracyOverTimePlotsStatisticsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AccuracyOverTimePlotsStatistics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AccuracyOverTimePlotsStatistics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AccuracyOverTimePlotsStatistics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
