% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activation_maps_compute_response.R
\docType{class}
\name{ActivationMapsComputeResponse}
\alias{ActivationMapsComputeResponse}
\title{ActivationMapsComputeResponse}
\format{
An \code{R6Class} generator object
}
\description{
ActivationMapsComputeResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character The job ID.}

\item{\code{isBlocked}}{character True if the job is waiting for its dependencies to be resolved first.}

\item{\code{jobType}}{character the type of the job}

\item{\code{message}}{character Error message in case of failure.}

\item{\code{modelId}}{character the model ID of the target model}

\item{\code{projectId}}{character The project the job belongs to.}

\item{\code{status}}{character The job status.}

\item{\code{url}}{character A URL that can be used to request details about the job.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ActivationMapsComputeResponse-new}{\code{ActivationMapsComputeResponse$new()}}
\item \href{#method-ActivationMapsComputeResponse-validate}{\code{ActivationMapsComputeResponse$validate()}}
\item \href{#method-ActivationMapsComputeResponse-toJSON}{\code{ActivationMapsComputeResponse$toJSON()}}
\item \href{#method-ActivationMapsComputeResponse-fromJSON}{\code{ActivationMapsComputeResponse$fromJSON()}}
\item \href{#method-ActivationMapsComputeResponse-clone}{\code{ActivationMapsComputeResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivationMapsComputeResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ActivationMapsComputeResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivationMapsComputeResponse$new(
  id = NULL,
  isBlocked = NULL,
  jobType = NULL,
  message = NULL,
  modelId = NULL,
  projectId = NULL,
  status = NULL,
  url = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The job ID.}

\item{\code{isBlocked}}{True if the job is waiting for its dependencies to be resolved first.}

\item{\code{jobType}}{the type of the job}

\item{\code{message}}{Error message in case of failure.}

\item{\code{modelId}}{the model ID of the target model}

\item{\code{projectId}}{The project the job belongs to.}

\item{\code{status}}{The job status.}

\item{\code{url}}{A URL that can be used to request details about the job.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivationMapsComputeResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ActivationMapsComputeResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivationMapsComputeResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivationMapsComputeResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ActivationMapsComputeResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivationMapsComputeResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivationMapsComputeResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ActivationMapsComputeResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivationMapsComputeResponse$fromJSON(
  ActivationMapsComputeResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ActivationMapsComputeResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivationMapsComputeResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ActivationMapsComputeResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivationMapsComputeResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
