% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actual.R
\docType{class}
\name{Actual}
\alias{Actual}
\title{Actual}
\format{
An \code{R6Class} generator object
}
\description{
Actual Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actualValue}}{\link{AnyOfintegernumberstring} The actual value of a prediction. Will be numeric for regression models and a string label for classification models.}

\item{\code{associationId}}{character A unique identifier used with a predicted row.}

\item{\code{timestamp}}{character \link{optional} The datetime when actual values were obtained, formatted according to RFC3339.}

\item{\code{wasActedOn}}{character \link{optional} Indicates if the prediction was acted on in a way that could have affected the actual outcome. For example, if a hospital patient is predicted to be readmitted in 30 days, extra procedures or new medication might be given to mitigate this problem, influencing the actual outcome of the prediction.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Actual-new}{\code{Actual$new()}}
\item \href{#method-Actual-validate}{\code{Actual$validate()}}
\item \href{#method-Actual-toJSON}{\code{Actual$toJSON()}}
\item \href{#method-Actual-fromJSON}{\code{Actual$fromJSON()}}
\item \href{#method-Actual-clone}{\code{Actual$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Actual-new"></a>}}
\if{latex}{\out{\hypertarget{method-Actual-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Actual$new(
  actualValue = NULL,
  associationId = NULL,
  timestamp = NULL,
  wasActedOn = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{actualValue}}{The actual value of a prediction. Will be numeric for regression models and a string label for classification models.}

\item{\code{associationId}}{A unique identifier used with a predicted row.}

\item{\code{timestamp}}{The datetime when actual values were obtained, formatted according to RFC3339.}

\item{\code{wasActedOn}}{Indicates if the prediction was acted on in a way that could have affected the actual outcome. For example, if a hospital patient is predicted to be readmitted in 30 days, extra procedures or new medication might be given to mitigate this problem, influencing the actual outcome of the prediction.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Actual-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Actual-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Actual$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Actual-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Actual-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Actual$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Actual-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Actual-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Actual$fromJSON(ActualJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ActualJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Actual-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Actual-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Actual$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
