% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{AddEureqaSolution}
\alias{AddEureqaSolution}
\title{AddEureqaSolution}
\usage{
AddEureqaSolution(project, eureqaSolutionId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a new model from an existing eureqa solution.
}
\details{
Create a new model from an existing eureqa solution.

This method invokes \verb{POST /projects/\{projectId\}/eureqaModels/} in the DataRobot Public API.
}
\seealso{
The method ProjectsEureqaModelsCreate in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
