% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_users_to_group_response.R
\docType{class}
\name{AddUsersToGroupResponse}
\alias{AddUsersToGroupResponse}
\title{AddUsersToGroupResponse}
\format{
An \code{R6Class} generator object
}
\description{
AddUsersToGroupResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer The number of items returned on this page.}

\item{\code{data}}{list( \link{UserInGroup} ) List of users in the group.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AddUsersToGroupResponse-new}{\code{AddUsersToGroupResponse$new()}}
\item \href{#method-AddUsersToGroupResponse-validate}{\code{AddUsersToGroupResponse$validate()}}
\item \href{#method-AddUsersToGroupResponse-toJSON}{\code{AddUsersToGroupResponse$toJSON()}}
\item \href{#method-AddUsersToGroupResponse-fromJSON}{\code{AddUsersToGroupResponse$fromJSON()}}
\item \href{#method-AddUsersToGroupResponse-clone}{\code{AddUsersToGroupResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddUsersToGroupResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-AddUsersToGroupResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddUsersToGroupResponse$new(
  count = NULL,
  data = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{The number of items returned on this page.}

\item{\code{data}}{List of users in the group.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddUsersToGroupResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AddUsersToGroupResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddUsersToGroupResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddUsersToGroupResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AddUsersToGroupResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddUsersToGroupResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddUsersToGroupResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AddUsersToGroupResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddUsersToGroupResponse$fromJSON(
  AddUsersToGroupResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AddUsersToGroupResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddUsersToGroupResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AddUsersToGroupResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddUsersToGroupResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
