% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_sample_size.R
\docType{class}
\name{AdminSampleSize}
\alias{AdminSampleSize}
\title{AdminSampleSize}
\format{
An \code{R6Class} generator object
}
\description{
AdminSampleSize Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{character Sample size can be specified only as a number of rows for now.}

\item{\code{value}}{integer Number of rows to ingest during dataset registration.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AdminSampleSize-new}{\code{AdminSampleSize$new()}}
\item \href{#method-AdminSampleSize-validate}{\code{AdminSampleSize$validate()}}
\item \href{#method-AdminSampleSize-toJSON}{\code{AdminSampleSize$toJSON()}}
\item \href{#method-AdminSampleSize-fromJSON}{\code{AdminSampleSize$fromJSON()}}
\item \href{#method-AdminSampleSize-clone}{\code{AdminSampleSize$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdminSampleSize-new"></a>}}
\if{latex}{\out{\hypertarget{method-AdminSampleSize-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdminSampleSize$new(type = NULL, value = NULL, validateParams = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Sample size can be specified only as a number of rows for now.}

\item{\code{value}}{Number of rows to ingest during dataset registration.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdminSampleSize-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AdminSampleSize-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdminSampleSize$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdminSampleSize-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AdminSampleSize-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdminSampleSize$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdminSampleSize-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AdminSampleSize-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdminSampleSize$fromJSON(AdminSampleSizeJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AdminSampleSizeJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdminSampleSize-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AdminSampleSize-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdminSampleSize$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
