% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_tuning_arguments_retrieve_response.R
\docType{class}
\name{AdvancedTuningArgumentsRetrieveResponse}
\alias{AdvancedTuningArgumentsRetrieveResponse}
\title{AdvancedTuningArgumentsRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
AdvancedTuningArgumentsRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tuningDescription}}{character Human-readable description of the tuned model, if specified by the user. `null` if unspecified.}

\item{\code{tuningParameters}}{list( \link{TuningParameters} ) An array of objects containing information about tuning parameters that are supported by the specified model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AdvancedTuningArgumentsRetrieveResponse-new}{\code{AdvancedTuningArgumentsRetrieveResponse$new()}}
\item \href{#method-AdvancedTuningArgumentsRetrieveResponse-validate}{\code{AdvancedTuningArgumentsRetrieveResponse$validate()}}
\item \href{#method-AdvancedTuningArgumentsRetrieveResponse-toJSON}{\code{AdvancedTuningArgumentsRetrieveResponse$toJSON()}}
\item \href{#method-AdvancedTuningArgumentsRetrieveResponse-fromJSON}{\code{AdvancedTuningArgumentsRetrieveResponse$fromJSON()}}
\item \href{#method-AdvancedTuningArgumentsRetrieveResponse-clone}{\code{AdvancedTuningArgumentsRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdvancedTuningArgumentsRetrieveResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-AdvancedTuningArgumentsRetrieveResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedTuningArgumentsRetrieveResponse$new(
  tuningDescription = NULL,
  tuningParameters = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tuningDescription}}{Human-readable description of the tuned model, if specified by the user. `null` if unspecified.}

\item{\code{tuningParameters}}{An array of objects containing information about tuning parameters that are supported by the specified model.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdvancedTuningArgumentsRetrieveResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AdvancedTuningArgumentsRetrieveResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedTuningArgumentsRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdvancedTuningArgumentsRetrieveResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AdvancedTuningArgumentsRetrieveResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedTuningArgumentsRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdvancedTuningArgumentsRetrieveResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AdvancedTuningArgumentsRetrieveResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedTuningArgumentsRetrieveResponse$fromJSON(
  AdvancedTuningArgumentsRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AdvancedTuningArgumentsRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdvancedTuningArgumentsRetrieveResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AdvancedTuningArgumentsRetrieveResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedTuningArgumentsRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
