% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim.R
\docType{class}
\name{Aim}
\alias{Aim}
\title{Aim}
\format{
An \code{R6Class} generator object
}
\description{
Aim Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{accuracyOptimizedMb}}{character \link{optional} Include additional, longer-running models that will be run by the autopilot and available to run manually.}

\item{\code{aggregationType}}{character \link{optional} For multiseries projects only. The aggregation type to apply when creating cross-series features.}

\item{\code{allowPartialHistoryTimeSeriesPredictions}}{character \link{optional} Specifies whether the time series predictions can use partial historical data.}

\item{\code{allowedPairwiseInteractionGroups}}{list( \link{array[character]} ) \link{optional} For GAM models - specify groups of columns for which pairwise interactions will be allowed. E.g. if set to [\link{'A', 'B', 'C'}, \link{'C', 'D'} ] then GAM models will allow interactions between columns AxB, BxC, AxC, CxD. All others (AxD, BxD) will not be considered. If not specified - all possible interactions will be considered by model.}

\item{\code{autopilotClusterList}}{list( integer ) \link{optional} A list of integers where each value will be used as the number of clusters in Autopilot model(s) for unsupervised clustering projects. Cannot be specified unless `unsupervisedMode` is true and `unsupervisedType` is set to `clustering`.}

\item{\code{autopilotDataSamplingMethod}}{character \link{optional} Defines how autopilot will select subsample from training dataset in OTV/TS projects. Defaults to 'latest' for 'rowCount' dataSelectionMethod and to 'random' for 'duration'.}

\item{\code{autopilotDataSelectionMethod}}{character The Data Selection method to be used by autopilot when creating models for datetime-partitioned datasets.}

\item{\code{autopilotWithFeatureDiscovery}}{character \link{optional} If true, autopilot will run on a feature list that includes features found via search for interactions.}

\item{\code{backtests}}{list( \link{Backtest} ) \link{optional} An array specifying the format of the backtests.}

\item{\code{biasMitigationFeatureName}}{character \link{optional} The name of the protected feature used to mitigate bias on models.}

\item{\code{biasMitigationTechnique}}{character \link{optional} Method applied to perform bias mitigation.}

\item{\code{blendBestModels}}{character \link{optional} Blend best models during Autopilot run. This option is not supported in SHAP-only mode or for multilabel projects.}

\item{\code{blueprintThreshold}}{integer \link{optional} The runtime (in hours) which if exceeded will exclude a model from autopilot runs.}

\item{\code{calendarId}}{character \link{optional} The ID of the calendar to be used in this project.}

\item{\code{classMappingAggregationSettings}}{\link{ClassMappingAggregationSettings} \link{optional}}

\item{\code{considerBlendersInRecommendation}}{character \link{optional} Include blenders when selecting a model to prepare for deployment in an Autopilot Run. This option is not supported in SHAP-only mode or for multilabel projects.}

\item{\code{credentials}}{list( \link{OneOfPasswordCredentialsCredentialId} ) \link{optional} List of credentials for the secondary datasets used in feature discovery project.}

\item{\code{crossSeriesGroupByColumns}}{list( character ) \link{optional} For multiseries projects with cross-series features enabled only. List of columns (currently of length 1). Setting that indicates how to further split series into related groups. For example, if every series is sales of an individual product, the series group-by could be the product category with values like \"men's clothing\", \"sports equipment\", etc.}

\item{\code{cvHoldoutLevel}}{\link{AnyOfstringintegernumber} \link{optional} The value of the partition column indicating a row is part of the holdout set. This level is optional - if not specified or if provided as ``null``, then no holdout will be used in the project. The rest of the levels indicate which cross validation fold each row should fall into.}

\item{\code{cvMethod}}{character \link{optional} The partitioning method to be applied to the training data.}

\item{\code{dateRemoval}}{character \link{optional} If true, enable creating additional feature lists without dates (does not apply to time-aware projects).}

\item{\code{datetimePartitionColumn}}{character \link{optional} The date column that will be used as a datetime partition column.}

\item{\code{defaultToAPriori}}{character \link{optional} Renamed to `defaultToKnownInAdvance`.}

\item{\code{defaultToDoNotDerive}}{character \link{optional} For time series projects only. Sets whether all features default to being treated as do-not-derive features, excluding them from feature derivation. Individual features can be set to a value different than the default by using the `featureSettings` parameter.}

\item{\code{defaultToKnownInAdvance}}{character \link{optional} For time series projects only. Sets whether all features default to being treated as known in advance features, which are features that are known into the future. Features marked as known in advance must be specified into the future when making predictions. The default is false, all features are not known in advance. Individual features can be set to a value different than the default using the `featureSettings` parameter. See the :ref:`Time Series Overview <time_series_overview>` for more context.}

\item{\code{differencingMethod}}{character \link{optional} For time series projects only. Used to specify which differencing method to apply if the data is stationary. For classification problems `simple` and `seasonal` are not allowed. Parameter `periodicities` must be specified if `seasonal` is chosen. Defaults to `auto`.}

\item{\code{disableHoldout}}{character \link{optional} Whether to suppress allocating a holdout fold. If `disableHoldout` is set to true, `holdoutStartDate` and `holdoutDuration` must not be set.}

\item{\code{eventsCount}}{character \link{optional} The name of a column specifying events count. The data in this column must be pure numeric and non negative without missing values}

\item{\code{exponentiallyWeightedMovingAlpha}}{numeric \link{optional} Discount factor (alpha) used for exponentially weighted moving features}

\item{\code{exposure}}{character \link{optional} The name of a column specifying row exposure.The data in this column must be pure numeric (e.g. not currency, date, length, etc.) and without missing values}

\item{\code{externalPredictions}}{list( character ) \link{optional} List of external prediction columns from the dataset.}

\item{\code{externalTimeSeriesBaselineDatasetId}}{character \link{optional} Catalog version id for external prediction data that can be used as a baseline to calculate new metrics.}

\item{\code{externalTimeSeriesBaselineDatasetName}}{character \link{optional} The name of the time series baseline dataset for the project.}

\item{\code{fairnessMetricsSet}}{character \link{optional} Metric to use for calculating fairness. Can be one of ``proportionalParity``, ``equalParity``, ``predictionBalance``, ``trueFavorableAndUnfavorableRateParity`` or ``FavorableAndUnfavorablePredictiveValueParity``. Used and required only if \emph{Bias & Fairness in AutoML} feature is enabled.}

\item{\code{fairnessThreshold}}{numeric \link{optional} Threshold value of the fairness metric. Can be in a range of ``\link{0:1}``. If the actual metric value is below the threshold, the user will be notified}

\item{\code{featureDerivationWindowEnd}}{integer \link{optional} For time series projects only. How many timeUnits of the datetimePartitionColumn into the past relative to the forecast point the feature derivation window should end.}

\item{\code{featureDerivationWindowStart}}{integer \link{optional} For time series projects only. How many timeUnits of the datetimePartitionColumn into the past relative to the forecast point the feature derivation window should begin.}

\item{\code{featureDiscoverySupervisedFeatureReduction}}{character \link{optional} Run supervised feature reduction for feature discovery projects.}

\item{\code{featureEngineeringPredictionPoint}}{character \link{optional} The date column to be used as prediction point for time-based feature engineering.}

\item{\code{featureSettings}}{list( \link{FeatureSetting} ) \link{optional} An array specifying per feature settings. Features can be left unspecified.}

\item{\code{featurelistId}}{character \link{optional} The ID of a featurelist to use for autopilot.}

\item{\code{forecastWindowEnd}}{integer \link{optional} For time series projects only. How many timeUnits of the datetimePartitionColumn into the future relative to the forecast point the forecast window should end.}

\item{\code{forecastWindowStart}}{integer \link{optional} For time series projects only. How many timeUnits of the datetimePartitionColumn into the future relative to the forecast point the forecast window should start.}

\item{\code{gapDuration}}{character \link{optional} The duration of the gap between holdout training and holdout scoring data. For time series projects, defaults to the duration of the gap between the end of the feature derivation window and the beginning of the forecast window. For OTV projects, defaults to a zero duration (P0Y0M0D).}

\item{\code{holdoutDuration}}{character \link{optional} The duration of holdout scoring data. When specifying `holdoutDuration`, `holdoutStartDate` must also be specified. This attribute cannot be specified when `disableHoldout` is true.}

\item{\code{holdoutEndDate}}{character \link{optional} The end date of holdout scoring data. When specifying `holdoutEndDate`, `holdoutStartDate` must also be specified. This attribute cannot be specified when `disableHoldout` is true.}

\item{\code{holdoutLevel}}{\link{AnyOfstringintegernumber} \link{optional} The value of the partition column indicating a row is part of the holdout set. This level is optional - if not specified or if provided as ``null``, then no holdout will be used in the project. However, the column must have exactly 2 values in order for this option to be valid}

\item{\code{holdoutPct}}{numeric \link{optional} The percentage of the dataset to assign to the holdout set}

\item{\code{holdoutStartDate}}{character \link{optional} The start date of holdout scoring data. When specifying `holdoutStartDate`, one of `holdoutEndDate` or `holdoutDuration` must also be specified. This attribute cannot be specified when `disableHoldout` is true.}

\item{\code{includeBiasMitigationFeatureAsPredictorVariable}}{character \link{optional} Specifies whether the mitigation feature will be used as a predictor variable (i.e., treated like other categorical features in the input to train the modeler), in addition to being used for bias mitigation. If false, the mitigation feature will be used only for bias mitigation, and not for training the modeler task.}

\item{\code{isHoldoutModified}}{character \link{optional} A boolean value indicating whether holdout settings (start/end dates) have been modified by user.}

\item{\code{majorityDownsamplingRate}}{numeric \link{optional} The percentage between 0 and 100 of the majority rows that should be kept. Must be specified only if using smart downsampling. If not specified, a default will be selected based on the dataset distribution. The chosen rate may not cause the majority class to become smaller than the minority class.}

\item{\code{metric}}{character \link{optional} The metric to use to select the best models. See `/api/v2/projects/(projectId)/features/metrics/` for the metrics that may be valid for a potential target.  Note that weighted metrics must be used with a weights column.}

\item{\code{minSecondaryValidationModelCount}}{integer \link{optional} Compute 'All backtest' scores (datetime models) or cross validation scores for the specified number of highest ranking models on the Leaderboard, if over the Autopilot default.}

\item{\code{mode}}{character \link{optional} The autopilot mode to use.  Either 'quick', 'auto', 'manual' or 'comprehensive'}

\item{\code{modelSplits}}{integer \link{optional} Sets the cap on the number of jobs per model used when building models to control number of jobs in the queue. Higher number of modelSplits will allow for less downsampling leading to the use of more post-processed data.}

\item{\code{monotonicDecreasingFeaturelistId}}{character \link{optional} The ID of the featurelist that defines the set of features with a monotonically decreasing relationship to the target.  If null, no such constraints are enforced. When specified, this will set a default for the project that can be overriden at model submission time if desired.}

\item{\code{monotonicIncreasingFeaturelistId}}{character \link{optional} The ID of the featurelist that defines the set of features with a monotonically increasing relationship to the target. If null, no such constraints are enforced. When specified, this will set a default for the project that can be overriden at model submission time if desired.}

\item{\code{multiseriesIdColumns}}{list( character ) \link{optional} May be used only with time series projects. An array of the column names identifying  the series to which each row of the dataset belongs. Currently only one multiseries ID column is supported. See the :ref:`multiseries <multiseries>` section of the time series documentation for more context.}

\item{\code{numberOfBacktests}}{integer \link{optional} The number of backtests to use. If omitted, defaults to a positive value selected by the server based on the validation and gap durations.}

\item{\code{offset}}{list( character ) \link{optional} An array of strings with names of a columns specifying row offsets.The data in this column must be pure numeric (e.g. not currency, date, length, etc.) and without missing values}

\item{\code{onlyIncludeMonotonicBlueprints}}{character When true, only blueprints that support enforcing montonic constraints will be available in the project or selected for autopilot.}

\item{\code{partitionKeyCols}}{list( character ) \link{optional} An array containing a single string - the name of the group partition column}

\item{\code{periodicities}}{list( \link{Periodicity} ) \link{optional} A list of periodicities for time series projects only. For classification problems periodicities are not allowed. If this is provided, parameter 'differencing_method' will default to 'seasonal' if not provided or 'auto'.}

\item{\code{positiveClass}}{\link{AnyOfstringintegernumber} \link{optional} A value from the target column to use for the positive class. May only be specified for projects doing binary classification.If not specified, a positive class is selected automatically.}

\item{\code{preferableTargetValue}}{\link{AnyOfstringintegernumber} \link{optional} A target value that should be treated as a positive outcome for the prediction. For example if we want to check gender discrimination for giving a loan and our target named ``is_bad``, then the positive outcome for the prediction would be ``No``, which means that the loan is good and that's what we treat as a preferable result for the loaner. Used and required only if \emph{Bias & Fairness in AutoML} feature is enabled.}

\item{\code{prepareModelForDeployment}}{character \link{optional} Prepare model for deployment during Autopilot run. The preparation includes creating reduced feature list models, retraining best model on higher sample size, computing insights and assigning 'RECOMMENDED FOR DEPLOYMENT' label.}

\item{\code{primaryLocationColumn}}{character \link{optional} Primary geospatial location column.}

\item{\code{protectedFeatures}}{list( character ) \link{optional} A list of project feature to mark as protected for Bias metric calculation and Fairness correction. Used and required only if \emph{Bias & Fairness in AutoML} feature is enabled.}

\item{\code{quantileLevel}}{numeric \link{optional} The quantile level between 0.01 and 0.99 for specifying the Quantile metric.}

\item{\code{quickrun}}{character \link{optional} (Deprecated): 'quick' should be used in the `mode` parameter instead of using this parameter. If set to `true`, autopilot mode will be set to 'quick'.Cannot be set to `true` when `mode` is set to 'comprehensive' or 'manual'.}

\item{\code{rateTopPctThreshold}}{numeric \link{optional} The percentage threshold between 0.1 and 50 for specifying the Rate@Top\% metric.}

\item{\code{relationshipsConfigurationId}}{character \link{optional} Relationships configuration id to be used for Feature Discovery projects.}

\item{\code{reps}}{integer \link{optional} The number of cross validation folds to use.}

\item{\code{responseCap}}{numeric \link{optional} Used to cap the maximum response of a model}

\item{\code{runLeakageRemovedFeatureList}}{character \link{optional} Run Autopilot on Leakage Removed feature list (if exists).}

\item{\code{sampleStepPct}}{numeric \link{optional} A float between 0 and 100 indicating the desired percentage of data to sample when training models in comprehensive Autopilot. Note: this only supported for comprehensive Autopilot and the specified value may be lowered in order to be compatible with the project's dataset and partition settings.}

\item{\code{scoringCodeOnly}}{character \link{optional} Keep only models that can be converted to scorable java code during Autopilot run.}

\item{\code{seed}}{integer \link{optional} A seed to use for randomization.}

\item{\code{segmentationTaskId}}{character \link{optional} Specifies the SegmentationTask that will be used for dividing the project up into multiple segmented projects.}

\item{\code{shapOnlyMode}}{character \link{optional} Keep only models that support SHAP values during Autopilot run. Use SHAP-based insights wherever possible.}

\item{\code{smartDownsampled}}{character \link{optional} Whether to use smart downsampling to throw away excess rows of the majority class. Only applicable to classification and zero-boosted regression projects.}

\item{\code{stopWords}}{list( character ) \link{optional} A list of stop words to be used for text blueprints. Note: ``stop_words=True`` must be set in the blueprint preprocessing parameters for this list of stop words to actually be used during preprocessing.}

\item{\code{target}}{character \link{optional} The name of the target feature.}

\item{\code{targetType}}{character \link{optional} Used to specify the targetType to use for a project when it is ambiguous, i.e. a numeric target with a few unique values that could be used for either regression or multiclass.}

\item{\code{trainingLevel}}{\link{AnyOfstringintegernumber} \link{optional} The value of the partition column indicating a row is part of the training set.}

\item{\code{treatAsExponential}}{character \link{optional} For time series projects only. Used to specify whether to treat data as exponential trend and apply transformations like log-transform. For classification problems `always` is not allowed.}

\item{\code{unsupervisedMode}}{character \link{optional} If True, unsupervised project (without target) will be created. ``target`` cannot be specified if ``unsupervisedMode`` is True.}

\item{\code{unsupervisedType}}{character \link{optional} The type of unsupervised project. Only valid when `unsupervisedMode` is true. If `unsupervisedMode`, defaults to `anomaly`.}

\item{\code{useCrossSeriesFeatures}}{character \link{optional} Indicating if user wants to use cross-series features.}

\item{\code{useProjectSettings}}{character \link{optional} Specifies whether datetime-partitioned project should use project settings (i.e. backtests configuration has been modified by the user).}

\item{\code{useSupervisedFeatureReduction}}{character \link{optional} When true, during feature generation DataRobot runs a supervised algorithm feature lists using only qualifying features. Setting false can severely impact autopilot duration, especially for datasets with many features.}

\item{\code{useTimeSeries}}{character \link{optional} A boolean value indicating whether a time series project should be created instead of a regular project which uses datetime partitioning.}

\item{\code{userPartitionCol}}{character \link{optional} The name of the column containing the partition assignments.}

\item{\code{validationDuration}}{character \link{optional} The default validation duration for all backtests. If the primary date/time feature in a time series project is irregular, you cannot set a default validation length. Instead, set each duration individually. For an OTV project setting the validation duration will always use regular partitioning. Omitting it will use irregular partitioning if the date/time feature is irregular.}

\item{\code{validationLevel}}{\link{AnyOfstringintegernumber} \link{optional} The value of the partition column indicating a row is part of the validation set.}

\item{\code{validationPct}}{numeric \link{optional} The percentage of the dataset to assign to the validation set}

\item{\code{validationType}}{character \link{optional} The validation method to be used.  CV for cross validation or TVH for train-validation-holdout split.}

\item{\code{weights}}{character \link{optional} The name of a column specifying row weights. The data in this column must be pure numeric (e.g. not currency, date, length, etc.) and without missing values}

\item{\code{windowsBasisUnit}}{character \link{optional} For time series projects only. Indicates which unit is basis for feature derivation window and forecast window. Valid options are detected time unit or `ROW`. If omitted, the default value is detected time unit.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Aim-new}{\code{Aim$new()}}
\item \href{#method-Aim-validate}{\code{Aim$validate()}}
\item \href{#method-Aim-toJSON}{\code{Aim$toJSON()}}
\item \href{#method-Aim-fromJSON}{\code{Aim$fromJSON()}}
\item \href{#method-Aim-clone}{\code{Aim$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Aim-new"></a>}}
\if{latex}{\out{\hypertarget{method-Aim-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Aim$new(
  autopilotDataSelectionMethod = NULL,
  onlyIncludeMonotonicBlueprints = NULL,
  accuracyOptimizedMb = NULL,
  aggregationType = NULL,
  allowPartialHistoryTimeSeriesPredictions = NULL,
  allowedPairwiseInteractionGroups = NULL,
  autopilotClusterList = NULL,
  autopilotDataSamplingMethod = NULL,
  autopilotWithFeatureDiscovery = NULL,
  backtests = NULL,
  biasMitigationFeatureName = NULL,
  biasMitigationTechnique = NULL,
  blendBestModels = NULL,
  blueprintThreshold = NULL,
  calendarId = NULL,
  classMappingAggregationSettings = NULL,
  considerBlendersInRecommendation = NULL,
  credentials = NULL,
  crossSeriesGroupByColumns = NULL,
  cvHoldoutLevel = NULL,
  cvMethod = NULL,
  dateRemoval = NULL,
  datetimePartitionColumn = NULL,
  defaultToAPriori = NULL,
  defaultToDoNotDerive = NULL,
  defaultToKnownInAdvance = NULL,
  differencingMethod = NULL,
  disableHoldout = FALSE,
  eventsCount = NULL,
  exponentiallyWeightedMovingAlpha = NULL,
  exposure = NULL,
  externalPredictions = NULL,
  externalTimeSeriesBaselineDatasetId = NULL,
  externalTimeSeriesBaselineDatasetName = NULL,
  fairnessMetricsSet = NULL,
  fairnessThreshold = NULL,
  featureDerivationWindowEnd = NULL,
  featureDerivationWindowStart = NULL,
  featureDiscoverySupervisedFeatureReduction = NULL,
  featureEngineeringPredictionPoint = NULL,
  featureSettings = NULL,
  featurelistId = NULL,
  forecastWindowEnd = NULL,
  forecastWindowStart = NULL,
  gapDuration = NULL,
  holdoutDuration = NULL,
  holdoutEndDate = NULL,
  holdoutLevel = NULL,
  holdoutPct = NULL,
  holdoutStartDate = NULL,
  includeBiasMitigationFeatureAsPredictorVariable = NULL,
  isHoldoutModified = NULL,
  majorityDownsamplingRate = NULL,
  metric = NULL,
  minSecondaryValidationModelCount = NULL,
  mode = NULL,
  modelSplits = 5,
  monotonicDecreasingFeaturelistId = NULL,
  monotonicIncreasingFeaturelistId = NULL,
  multiseriesIdColumns = NULL,
  numberOfBacktests = NULL,
  offset = NULL,
  partitionKeyCols = NULL,
  periodicities = NULL,
  positiveClass = NULL,
  preferableTargetValue = NULL,
  prepareModelForDeployment = NULL,
  primaryLocationColumn = NULL,
  protectedFeatures = NULL,
  quantileLevel = NULL,
  quickrun = NULL,
  rateTopPctThreshold = NULL,
  relationshipsConfigurationId = NULL,
  reps = NULL,
  responseCap = NULL,
  runLeakageRemovedFeatureList = NULL,
  sampleStepPct = NULL,
  scoringCodeOnly = NULL,
  seed = NULL,
  segmentationTaskId = NULL,
  shapOnlyMode = NULL,
  smartDownsampled = NULL,
  stopWords = NULL,
  target = NULL,
  targetType = NULL,
  trainingLevel = NULL,
  treatAsExponential = "auto",
  unsupervisedMode = FALSE,
  unsupervisedType = NULL,
  useCrossSeriesFeatures = NULL,
  useProjectSettings = NULL,
  useSupervisedFeatureReduction = TRUE,
  useTimeSeries = FALSE,
  userPartitionCol = NULL,
  validationDuration = NULL,
  validationLevel = NULL,
  validationPct = NULL,
  validationType = NULL,
  weights = NULL,
  windowsBasisUnit = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{autopilotDataSelectionMethod}}{The Data Selection method to be used by autopilot when creating models for datetime-partitioned datasets.}

\item{\code{onlyIncludeMonotonicBlueprints}}{When true, only blueprints that support enforcing montonic constraints will be available in the project or selected for autopilot.}

\item{\code{accuracyOptimizedMb}}{Include additional, longer-running models that will be run by the autopilot and available to run manually.}

\item{\code{aggregationType}}{For multiseries projects only. The aggregation type to apply when creating cross-series features.}

\item{\code{allowPartialHistoryTimeSeriesPredictions}}{Specifies whether the time series predictions can use partial historical data.}

\item{\code{allowedPairwiseInteractionGroups}}{For GAM models - specify groups of columns for which pairwise interactions will be allowed. E.g. if set to [\link{'A', 'B', 'C'}, \link{'C', 'D'} ] then GAM models will allow interactions between columns AxB, BxC, AxC, CxD. All others (AxD, BxD) will not be considered. If not specified - all possible interactions will be considered by model.}

\item{\code{autopilotClusterList}}{A list of integers where each value will be used as the number of clusters in Autopilot model(s) for unsupervised clustering projects. Cannot be specified unless `unsupervisedMode` is true and `unsupervisedType` is set to `clustering`.}

\item{\code{autopilotDataSamplingMethod}}{Defines how autopilot will select subsample from training dataset in OTV/TS projects. Defaults to 'latest' for 'rowCount' dataSelectionMethod and to 'random' for 'duration'.}

\item{\code{autopilotWithFeatureDiscovery}}{If true, autopilot will run on a feature list that includes features found via search for interactions.}

\item{\code{backtests}}{An array specifying the format of the backtests.}

\item{\code{biasMitigationFeatureName}}{The name of the protected feature used to mitigate bias on models.}

\item{\code{biasMitigationTechnique}}{Method applied to perform bias mitigation.}

\item{\code{blendBestModels}}{Blend best models during Autopilot run. This option is not supported in SHAP-only mode or for multilabel projects.}

\item{\code{blueprintThreshold}}{The runtime (in hours) which if exceeded will exclude a model from autopilot runs.}

\item{\code{calendarId}}{The ID of the calendar to be used in this project.}

\item{\code{considerBlendersInRecommendation}}{Include blenders when selecting a model to prepare for deployment in an Autopilot Run. This option is not supported in SHAP-only mode or for multilabel projects.}

\item{\code{credentials}}{List of credentials for the secondary datasets used in feature discovery project.}

\item{\code{crossSeriesGroupByColumns}}{For multiseries projects with cross-series features enabled only. List of columns (currently of length 1). Setting that indicates how to further split series into related groups. For example, if every series is sales of an individual product, the series group-by could be the product category with values like \"men's clothing\", \"sports equipment\", etc.}

\item{\code{cvHoldoutLevel}}{The value of the partition column indicating a row is part of the holdout set. This level is optional - if not specified or if provided as ``null``, then no holdout will be used in the project. The rest of the levels indicate which cross validation fold each row should fall into.}

\item{\code{cvMethod}}{The partitioning method to be applied to the training data.}

\item{\code{dateRemoval}}{If true, enable creating additional feature lists without dates (does not apply to time-aware projects).}

\item{\code{datetimePartitionColumn}}{The date column that will be used as a datetime partition column.}

\item{\code{defaultToAPriori}}{Renamed to `defaultToKnownInAdvance`.}

\item{\code{defaultToDoNotDerive}}{For time series projects only. Sets whether all features default to being treated as do-not-derive features, excluding them from feature derivation. Individual features can be set to a value different than the default by using the `featureSettings` parameter.}

\item{\code{defaultToKnownInAdvance}}{For time series projects only. Sets whether all features default to being treated as known in advance features, which are features that are known into the future. Features marked as known in advance must be specified into the future when making predictions. The default is false, all features are not known in advance. Individual features can be set to a value different than the default using the `featureSettings` parameter. See the :ref:`Time Series Overview <time_series_overview>` for more context.}

\item{\code{differencingMethod}}{For time series projects only. Used to specify which differencing method to apply if the data is stationary. For classification problems `simple` and `seasonal` are not allowed. Parameter `periodicities` must be specified if `seasonal` is chosen. Defaults to `auto`.}

\item{\code{disableHoldout}}{Whether to suppress allocating a holdout fold. If `disableHoldout` is set to true, `holdoutStartDate` and `holdoutDuration` must not be set.}

\item{\code{eventsCount}}{The name of a column specifying events count. The data in this column must be pure numeric and non negative without missing values}

\item{\code{exponentiallyWeightedMovingAlpha}}{Discount factor (alpha) used for exponentially weighted moving features}

\item{\code{exposure}}{The name of a column specifying row exposure.The data in this column must be pure numeric (e.g. not currency, date, length, etc.) and without missing values}

\item{\code{externalPredictions}}{List of external prediction columns from the dataset.}

\item{\code{externalTimeSeriesBaselineDatasetId}}{Catalog version id for external prediction data that can be used as a baseline to calculate new metrics.}

\item{\code{externalTimeSeriesBaselineDatasetName}}{The name of the time series baseline dataset for the project.}

\item{\code{fairnessMetricsSet}}{Metric to use for calculating fairness. Can be one of ``proportionalParity``, ``equalParity``, ``predictionBalance``, ``trueFavorableAndUnfavorableRateParity`` or ``FavorableAndUnfavorablePredictiveValueParity``. Used and required only if \emph{Bias & Fairness in AutoML} feature is enabled.}

\item{\code{fairnessThreshold}}{Threshold value of the fairness metric. Can be in a range of ``\link{0:1}``. If the actual metric value is below the threshold, the user will be notified}

\item{\code{featureDerivationWindowEnd}}{For time series projects only. How many timeUnits of the datetimePartitionColumn into the past relative to the forecast point the feature derivation window should end.}

\item{\code{featureDerivationWindowStart}}{For time series projects only. How many timeUnits of the datetimePartitionColumn into the past relative to the forecast point the feature derivation window should begin.}

\item{\code{featureDiscoverySupervisedFeatureReduction}}{Run supervised feature reduction for feature discovery projects.}

\item{\code{featureEngineeringPredictionPoint}}{The date column to be used as prediction point for time-based feature engineering.}

\item{\code{featureSettings}}{An array specifying per feature settings. Features can be left unspecified.}

\item{\code{featurelistId}}{The ID of a featurelist to use for autopilot.}

\item{\code{forecastWindowEnd}}{For time series projects only. How many timeUnits of the datetimePartitionColumn into the future relative to the forecast point the forecast window should end.}

\item{\code{forecastWindowStart}}{For time series projects only. How many timeUnits of the datetimePartitionColumn into the future relative to the forecast point the forecast window should start.}

\item{\code{gapDuration}}{The duration of the gap between holdout training and holdout scoring data. For time series projects, defaults to the duration of the gap between the end of the feature derivation window and the beginning of the forecast window. For OTV projects, defaults to a zero duration (P0Y0M0D).}

\item{\code{holdoutDuration}}{The duration of holdout scoring data. When specifying `holdoutDuration`, `holdoutStartDate` must also be specified. This attribute cannot be specified when `disableHoldout` is true.}

\item{\code{holdoutEndDate}}{The end date of holdout scoring data. When specifying `holdoutEndDate`, `holdoutStartDate` must also be specified. This attribute cannot be specified when `disableHoldout` is true.}

\item{\code{holdoutLevel}}{The value of the partition column indicating a row is part of the holdout set. This level is optional - if not specified or if provided as ``null``, then no holdout will be used in the project. However, the column must have exactly 2 values in order for this option to be valid}

\item{\code{holdoutPct}}{The percentage of the dataset to assign to the holdout set}

\item{\code{holdoutStartDate}}{The start date of holdout scoring data. When specifying `holdoutStartDate`, one of `holdoutEndDate` or `holdoutDuration` must also be specified. This attribute cannot be specified when `disableHoldout` is true.}

\item{\code{includeBiasMitigationFeatureAsPredictorVariable}}{Specifies whether the mitigation feature will be used as a predictor variable (i.e., treated like other categorical features in the input to train the modeler), in addition to being used for bias mitigation. If false, the mitigation feature will be used only for bias mitigation, and not for training the modeler task.}

\item{\code{isHoldoutModified}}{A boolean value indicating whether holdout settings (start/end dates) have been modified by user.}

\item{\code{majorityDownsamplingRate}}{The percentage between 0 and 100 of the majority rows that should be kept. Must be specified only if using smart downsampling. If not specified, a default will be selected based on the dataset distribution. The chosen rate may not cause the majority class to become smaller than the minority class.}

\item{\code{metric}}{The metric to use to select the best models. See `/api/v2/projects/(projectId)/features/metrics/` for the metrics that may be valid for a potential target.  Note that weighted metrics must be used with a weights column.}

\item{\code{minSecondaryValidationModelCount}}{Compute 'All backtest' scores (datetime models) or cross validation scores for the specified number of highest ranking models on the Leaderboard, if over the Autopilot default.}

\item{\code{mode}}{The autopilot mode to use.  Either 'quick', 'auto', 'manual' or 'comprehensive'}

\item{\code{modelSplits}}{Sets the cap on the number of jobs per model used when building models to control number of jobs in the queue. Higher number of modelSplits will allow for less downsampling leading to the use of more post-processed data.}

\item{\code{monotonicDecreasingFeaturelistId}}{The ID of the featurelist that defines the set of features with a monotonically decreasing relationship to the target.  If null, no such constraints are enforced. When specified, this will set a default for the project that can be overriden at model submission time if desired.}

\item{\code{monotonicIncreasingFeaturelistId}}{The ID of the featurelist that defines the set of features with a monotonically increasing relationship to the target. If null, no such constraints are enforced. When specified, this will set a default for the project that can be overriden at model submission time if desired.}

\item{\code{multiseriesIdColumns}}{May be used only with time series projects. An array of the column names identifying  the series to which each row of the dataset belongs. Currently only one multiseries ID column is supported. See the :ref:`multiseries <multiseries>` section of the time series documentation for more context.}

\item{\code{numberOfBacktests}}{The number of backtests to use. If omitted, defaults to a positive value selected by the server based on the validation and gap durations.}

\item{\code{offset}}{An array of strings with names of a columns specifying row offsets.The data in this column must be pure numeric (e.g. not currency, date, length, etc.) and without missing values}

\item{\code{partitionKeyCols}}{An array containing a single string - the name of the group partition column}

\item{\code{periodicities}}{A list of periodicities for time series projects only. For classification problems periodicities are not allowed. If this is provided, parameter 'differencing_method' will default to 'seasonal' if not provided or 'auto'.}

\item{\code{positiveClass}}{A value from the target column to use for the positive class. May only be specified for projects doing binary classification.If not specified, a positive class is selected automatically.}

\item{\code{preferableTargetValue}}{A target value that should be treated as a positive outcome for the prediction. For example if we want to check gender discrimination for giving a loan and our target named ``is_bad``, then the positive outcome for the prediction would be ``No``, which means that the loan is good and that's what we treat as a preferable result for the loaner. Used and required only if \emph{Bias & Fairness in AutoML} feature is enabled.}

\item{\code{prepareModelForDeployment}}{Prepare model for deployment during Autopilot run. The preparation includes creating reduced feature list models, retraining best model on higher sample size, computing insights and assigning 'RECOMMENDED FOR DEPLOYMENT' label.}

\item{\code{primaryLocationColumn}}{Primary geospatial location column.}

\item{\code{protectedFeatures}}{A list of project feature to mark as protected for Bias metric calculation and Fairness correction. Used and required only if \emph{Bias & Fairness in AutoML} feature is enabled.}

\item{\code{quantileLevel}}{The quantile level between 0.01 and 0.99 for specifying the Quantile metric.}

\item{\code{quickrun}}{(Deprecated): 'quick' should be used in the `mode` parameter instead of using this parameter. If set to `true`, autopilot mode will be set to 'quick'.Cannot be set to `true` when `mode` is set to 'comprehensive' or 'manual'.}

\item{\code{rateTopPctThreshold}}{The percentage threshold between 0.1 and 50 for specifying the Rate@Top\% metric.}

\item{\code{relationshipsConfigurationId}}{Relationships configuration id to be used for Feature Discovery projects.}

\item{\code{reps}}{The number of cross validation folds to use.}

\item{\code{responseCap}}{Used to cap the maximum response of a model}

\item{\code{runLeakageRemovedFeatureList}}{Run Autopilot on Leakage Removed feature list (if exists).}

\item{\code{sampleStepPct}}{A float between 0 and 100 indicating the desired percentage of data to sample when training models in comprehensive Autopilot. Note: this only supported for comprehensive Autopilot and the specified value may be lowered in order to be compatible with the project's dataset and partition settings.}

\item{\code{scoringCodeOnly}}{Keep only models that can be converted to scorable java code during Autopilot run.}

\item{\code{seed}}{A seed to use for randomization.}

\item{\code{segmentationTaskId}}{Specifies the SegmentationTask that will be used for dividing the project up into multiple segmented projects.}

\item{\code{shapOnlyMode}}{Keep only models that support SHAP values during Autopilot run. Use SHAP-based insights wherever possible.}

\item{\code{smartDownsampled}}{Whether to use smart downsampling to throw away excess rows of the majority class. Only applicable to classification and zero-boosted regression projects.}

\item{\code{stopWords}}{A list of stop words to be used for text blueprints. Note: ``stop_words=True`` must be set in the blueprint preprocessing parameters for this list of stop words to actually be used during preprocessing.}

\item{\code{target}}{The name of the target feature.}

\item{\code{targetType}}{Used to specify the targetType to use for a project when it is ambiguous, i.e. a numeric target with a few unique values that could be used for either regression or multiclass.}

\item{\code{trainingLevel}}{The value of the partition column indicating a row is part of the training set.}

\item{\code{treatAsExponential}}{For time series projects only. Used to specify whether to treat data as exponential trend and apply transformations like log-transform. For classification problems `always` is not allowed.}

\item{\code{unsupervisedMode}}{If True, unsupervised project (without target) will be created. ``target`` cannot be specified if ``unsupervisedMode`` is True.}

\item{\code{unsupervisedType}}{The type of unsupervised project. Only valid when `unsupervisedMode` is true. If `unsupervisedMode`, defaults to `anomaly`.}

\item{\code{useCrossSeriesFeatures}}{Indicating if user wants to use cross-series features.}

\item{\code{useProjectSettings}}{Specifies whether datetime-partitioned project should use project settings (i.e. backtests configuration has been modified by the user).}

\item{\code{useSupervisedFeatureReduction}}{When true, during feature generation DataRobot runs a supervised algorithm feature lists using only qualifying features. Setting false can severely impact autopilot duration, especially for datasets with many features.}

\item{\code{useTimeSeries}}{A boolean value indicating whether a time series project should be created instead of a regular project which uses datetime partitioning.}

\item{\code{userPartitionCol}}{The name of the column containing the partition assignments.}

\item{\code{validationDuration}}{The default validation duration for all backtests. If the primary date/time feature in a time series project is irregular, you cannot set a default validation length. Instead, set each duration individually. For an OTV project setting the validation duration will always use regular partitioning. Omitting it will use irregular partitioning if the date/time feature is irregular.}

\item{\code{validationLevel}}{The value of the partition column indicating a row is part of the validation set.}

\item{\code{validationPct}}{The percentage of the dataset to assign to the validation set}

\item{\code{validationType}}{The validation method to be used.  CV for cross validation or TVH for train-validation-holdout split.}

\item{\code{weights}}{The name of a column specifying row weights. The data in this column must be pure numeric (e.g. not currency, date, length, etc.) and without missing values}

\item{\code{windowsBasisUnit}}{For time series projects only. Indicates which unit is basis for feature derivation window and forecast window. Valid options are detected time unit or `ROW`. If omitted, the default value is detected time unit.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Aim-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Aim-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Aim$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Aim-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Aim-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Aim$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Aim-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Aim-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Aim$fromJSON(AimJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AimJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Aim-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Aim-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Aim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
