% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_assessment_explanations_response.R
\docType{class}
\name{AnomalyAssessmentExplanationsResponse}
\alias{AnomalyAssessmentExplanationsResponse}
\title{AnomalyAssessmentExplanationsResponse}
\format{
An \code{R6Class} generator object
}
\description{
AnomalyAssessmentExplanationsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{backtest}}{\link{OneOfintegerstring} The backtest of the record.}

\item{\code{count}}{integer The count of points.}

\item{\code{data}}{list( \link{DataPointResponse} ) Each is a `DataPoint` corresponding to a row in the specified range.}

\item{\code{endDate}}{character ISO-formatted last timestamp in the response. For example: ``2019-08-30T00:00:00.000000Z``.}

\item{\code{modelId}}{character The model ID of the record.}

\item{\code{projectId}}{character The project ID of the record.}

\item{\code{recordId}}{character The ID of the anomaly assessment record.}

\item{\code{seriesId}}{character The series id of the record. Applicable in multiseries projects}

\item{\code{shapBaseValue}}{numeric shap base value}

\item{\code{source}}{character The source of the record}

\item{\code{startDate}}{character ISO-formatted first timestamp in the response. For example: ``2019-08-01T00:00:00.000000Z``.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AnomalyAssessmentExplanationsResponse$new()}}
\item \href{#method-validate}{\code{AnomalyAssessmentExplanationsResponse$validate()}}
\item \href{#method-toJSON}{\code{AnomalyAssessmentExplanationsResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{AnomalyAssessmentExplanationsResponse$fromJSON()}}
\item \href{#method-clone}{\code{AnomalyAssessmentExplanationsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyAssessmentExplanationsResponse$new(
  backtest = NULL,
  count = NULL,
  data = NULL,
  endDate = NULL,
  modelId = NULL,
  projectId = NULL,
  recordId = NULL,
  seriesId = NULL,
  shapBaseValue = NULL,
  source = NULL,
  startDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{backtest}}{The backtest of the record.}

\item{\code{count}}{The count of points.}

\item{\code{data}}{Each is a `DataPoint` corresponding to a row in the specified range.}

\item{\code{endDate}}{ISO-formatted last timestamp in the response. For example: ``2019-08-30T00:00:00.000000Z``.}

\item{\code{modelId}}{The model ID of the record.}

\item{\code{projectId}}{The project ID of the record.}

\item{\code{recordId}}{The ID of the anomaly assessment record.}

\item{\code{seriesId}}{The series id of the record. Applicable in multiseries projects}

\item{\code{shapBaseValue}}{shap base value}

\item{\code{source}}{The source of the record}

\item{\code{startDate}}{ISO-formatted first timestamp in the response. For example: ``2019-08-01T00:00:00.000000Z``.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyAssessmentExplanationsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyAssessmentExplanationsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyAssessmentExplanationsResponse$fromJSON(
  AnomalyAssessmentExplanationsResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AnomalyAssessmentExplanationsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyAssessmentExplanationsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
