% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_assessment_initialize.R
\docType{class}
\name{AnomalyAssessmentInitialize}
\alias{AnomalyAssessmentInitialize}
\title{AnomalyAssessmentInitialize}
\format{
An \code{R6Class} generator object
}
\description{
AnomalyAssessmentInitialize Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{backtest}}{\link{OneOfintegerstring} The backtest to compute insight for.}

\item{\code{seriesId}}{character \link{optional} Required for multiseries projects. The series id to compute insight for.}

\item{\code{source}}{character The source to compute insight for.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnomalyAssessmentInitialize-new}{\code{AnomalyAssessmentInitialize$new()}}
\item \href{#method-AnomalyAssessmentInitialize-validate}{\code{AnomalyAssessmentInitialize$validate()}}
\item \href{#method-AnomalyAssessmentInitialize-toJSON}{\code{AnomalyAssessmentInitialize$toJSON()}}
\item \href{#method-AnomalyAssessmentInitialize-fromJSON}{\code{AnomalyAssessmentInitialize$fromJSON()}}
\item \href{#method-AnomalyAssessmentInitialize-clone}{\code{AnomalyAssessmentInitialize$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyAssessmentInitialize-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyAssessmentInitialize-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyAssessmentInitialize$new(
  backtest = NULL,
  source = NULL,
  seriesId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{backtest}}{The backtest to compute insight for.}

\item{\code{source}}{The source to compute insight for.}

\item{\code{seriesId}}{Required for multiseries projects. The series id to compute insight for.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyAssessmentInitialize-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyAssessmentInitialize-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyAssessmentInitialize$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyAssessmentInitialize-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyAssessmentInitialize-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyAssessmentInitialize$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyAssessmentInitialize-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyAssessmentInitialize-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyAssessmentInitialize$fromJSON(
  AnomalyAssessmentInitializeJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AnomalyAssessmentInitializeJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyAssessmentInitialize-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyAssessmentInitialize-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyAssessmentInitialize$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
