% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_assessment_record_response.R
\docType{class}
\name{AnomalyAssessmentRecordResponse}
\alias{AnomalyAssessmentRecordResponse}
\title{AnomalyAssessmentRecordResponse}
\format{
An \code{R6Class} generator object
}
\description{
AnomalyAssessmentRecordResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{backtest}}{\link{OneOfintegerstring} The backtest of the record.}

\item{\code{deleteLocation}}{character URL to delete anomaly assessment record.}

\item{\code{endDate}}{character ISO-formatted last timestamp in the subset. For example: ``2019-08-30T00:00:00.000000Z``.}

\item{\code{latestExplanationsLocation}}{character URL to retrieve the latest predictions with the shap explanations.}

\item{\code{modelId}}{character The model ID of the record.}

\item{\code{predictionThreshold}}{numeric The threshold, all rows with anomaly scores greater or equal to it have Shapley explanations computed.}

\item{\code{previewLocation}}{character URL to retrieve predictions preview for the record.}

\item{\code{projectId}}{character The project ID of the record.}

\item{\code{recordId}}{character The ID of the anomaly assessment record.}

\item{\code{seriesId}}{character The series id of the record. Applicable in multiseries projects}

\item{\code{source}}{character The source of the record}

\item{\code{startDate}}{character ISO-formatted first timestamp in the subset. For example: ``2019-08-01T00:00:00.000000Z``.}

\item{\code{status}}{character The status of the anomaly assessment record.}

\item{\code{statusDetails}}{character The status details.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AnomalyAssessmentRecordResponse$new()}}
\item \href{#method-validate}{\code{AnomalyAssessmentRecordResponse$validate()}}
\item \href{#method-toJSON}{\code{AnomalyAssessmentRecordResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{AnomalyAssessmentRecordResponse$fromJSON()}}
\item \href{#method-clone}{\code{AnomalyAssessmentRecordResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyAssessmentRecordResponse$new(
  backtest = NULL,
  deleteLocation = NULL,
  endDate = NULL,
  latestExplanationsLocation = NULL,
  modelId = NULL,
  predictionThreshold = NULL,
  previewLocation = NULL,
  projectId = NULL,
  recordId = NULL,
  seriesId = NULL,
  source = NULL,
  startDate = NULL,
  status = NULL,
  statusDetails = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{backtest}}{The backtest of the record.}

\item{\code{deleteLocation}}{URL to delete anomaly assessment record.}

\item{\code{endDate}}{ISO-formatted last timestamp in the subset. For example: ``2019-08-30T00:00:00.000000Z``.}

\item{\code{latestExplanationsLocation}}{URL to retrieve the latest predictions with the shap explanations.}

\item{\code{modelId}}{The model ID of the record.}

\item{\code{predictionThreshold}}{The threshold, all rows with anomaly scores greater or equal to it have Shapley explanations computed.}

\item{\code{previewLocation}}{URL to retrieve predictions preview for the record.}

\item{\code{projectId}}{The project ID of the record.}

\item{\code{recordId}}{The ID of the anomaly assessment record.}

\item{\code{seriesId}}{The series id of the record. Applicable in multiseries projects}

\item{\code{source}}{The source of the record}

\item{\code{startDate}}{ISO-formatted first timestamp in the subset. For example: ``2019-08-01T00:00:00.000000Z``.}

\item{\code{status}}{The status of the anomaly assessment record.}

\item{\code{statusDetails}}{The status details.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyAssessmentRecordResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyAssessmentRecordResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyAssessmentRecordResponse$fromJSON(
  AnomalyAssessmentRecordResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AnomalyAssessmentRecordResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyAssessmentRecordResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
