% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_over_time_plots_bins.R
\docType{class}
\name{AnomalyOverTimePlotsBins}
\alias{AnomalyOverTimePlotsBins}
\title{AnomalyOverTimePlotsBins}
\format{
An \code{R6Class} generator object
}
\description{
AnomalyOverTimePlotsBins Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{endDate}}{character The datetime of the end of the bin (exclusive).}

\item{\code{frequency}}{integer Indicates number of values averaged in bin in case of a resolution change.}

\item{\code{predicted}}{numeric Average prediction of the model in the bin. `null` if there are no entries in the bin.}

\item{\code{startDate}}{character The datetime of the start of the bin (inclusive).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnomalyOverTimePlotsBins-new}{\code{AnomalyOverTimePlotsBins$new()}}
\item \href{#method-AnomalyOverTimePlotsBins-validate}{\code{AnomalyOverTimePlotsBins$validate()}}
\item \href{#method-AnomalyOverTimePlotsBins-toJSON}{\code{AnomalyOverTimePlotsBins$toJSON()}}
\item \href{#method-AnomalyOverTimePlotsBins-fromJSON}{\code{AnomalyOverTimePlotsBins$fromJSON()}}
\item \href{#method-AnomalyOverTimePlotsBins-clone}{\code{AnomalyOverTimePlotsBins$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyOverTimePlotsBins-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyOverTimePlotsBins-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyOverTimePlotsBins$new(
  endDate = NULL,
  frequency = NULL,
  predicted = NULL,
  startDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endDate}}{The datetime of the end of the bin (exclusive).}

\item{\code{frequency}}{Indicates number of values averaged in bin in case of a resolution change.}

\item{\code{predicted}}{Average prediction of the model in the bin. `null` if there are no entries in the bin.}

\item{\code{startDate}}{The datetime of the start of the bin (inclusive).}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyOverTimePlotsBins-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyOverTimePlotsBins-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyOverTimePlotsBins$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyOverTimePlotsBins-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyOverTimePlotsBins-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyOverTimePlotsBins$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyOverTimePlotsBins-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyOverTimePlotsBins-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyOverTimePlotsBins$fromJSON(
  AnomalyOverTimePlotsBinsJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AnomalyOverTimePlotsBinsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyOverTimePlotsBins-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyOverTimePlotsBins-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyOverTimePlotsBins$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
