% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_over_time_plots_data_response.R
\docType{class}
\name{AnomalyOverTimePlotsDataResponse}
\alias{AnomalyOverTimePlotsDataResponse}
\title{AnomalyOverTimePlotsDataResponse}
\format{
An \code{R6Class} generator object
}
\description{
AnomalyOverTimePlotsDataResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bins}}{list( \link{AnomalyOverTimePlotsBins} ) An array of bins for the retrieved plots.}

\item{\code{calendarEvents}}{list( \link{CalendarEvent} ) An array of calendar events for a retrieved plot.}

\item{\code{endDate}}{character The datetime of the end of the chart data (exclusive).}

\item{\code{resolution}}{character The resolution that is used for binning.}

\item{\code{startDate}}{character The datetime of the start of the chart data (inclusive).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnomalyOverTimePlotsDataResponse-new}{\code{AnomalyOverTimePlotsDataResponse$new()}}
\item \href{#method-AnomalyOverTimePlotsDataResponse-validate}{\code{AnomalyOverTimePlotsDataResponse$validate()}}
\item \href{#method-AnomalyOverTimePlotsDataResponse-toJSON}{\code{AnomalyOverTimePlotsDataResponse$toJSON()}}
\item \href{#method-AnomalyOverTimePlotsDataResponse-fromJSON}{\code{AnomalyOverTimePlotsDataResponse$fromJSON()}}
\item \href{#method-AnomalyOverTimePlotsDataResponse-clone}{\code{AnomalyOverTimePlotsDataResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyOverTimePlotsDataResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyOverTimePlotsDataResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyOverTimePlotsDataResponse$new(
  bins = NULL,
  calendarEvents = NULL,
  endDate = NULL,
  resolution = NULL,
  startDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bins}}{An array of bins for the retrieved plots.}

\item{\code{calendarEvents}}{An array of calendar events for a retrieved plot.}

\item{\code{endDate}}{The datetime of the end of the chart data (exclusive).}

\item{\code{resolution}}{The resolution that is used for binning.}

\item{\code{startDate}}{The datetime of the start of the chart data (inclusive).}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyOverTimePlotsDataResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyOverTimePlotsDataResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyOverTimePlotsDataResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyOverTimePlotsDataResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyOverTimePlotsDataResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyOverTimePlotsDataResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyOverTimePlotsDataResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyOverTimePlotsDataResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyOverTimePlotsDataResponse$fromJSON(
  AnomalyOverTimePlotsDataResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AnomalyOverTimePlotsDataResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyOverTimePlotsDataResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyOverTimePlotsDataResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyOverTimePlotsDataResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
