% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_over_time_plots_metadata_response.R
\docType{class}
\name{AnomalyOverTimePlotsMetadataResponse}
\alias{AnomalyOverTimePlotsMetadataResponse}
\title{AnomalyOverTimePlotsMetadataResponse}
\format{
An \code{R6Class} generator object
}
\description{
AnomalyOverTimePlotsMetadataResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{backtestMetadata}}{list( \link{DatetimeTrendPlotsBacktestMetadata} ) An array of metadata information for each backtest. The array index of metadata object is the backtest index.}

\item{\code{backtestStatuses}}{list( \link{AccuracyOverTimePlotsStatus} ) An array of status information for each backtest. The array index of status object is the backtest index.}

\item{\code{estimatedSeriesLimit}}{integer \link{optional} Estimated number of series that can be calculated in one request for 1 FD.}

\item{\code{holdoutMetadata}}{\link{DatetimeTrendPlotsBacktestMetadata}}

\item{\code{holdoutStatuses}}{\link{AccuracyOverTimePlotsStatus}}

\item{\code{resolutions}}{list( character ) An array of available time resolutions for which plots can be retrieved.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnomalyOverTimePlotsMetadataResponse-new}{\code{AnomalyOverTimePlotsMetadataResponse$new()}}
\item \href{#method-AnomalyOverTimePlotsMetadataResponse-validate}{\code{AnomalyOverTimePlotsMetadataResponse$validate()}}
\item \href{#method-AnomalyOverTimePlotsMetadataResponse-toJSON}{\code{AnomalyOverTimePlotsMetadataResponse$toJSON()}}
\item \href{#method-AnomalyOverTimePlotsMetadataResponse-fromJSON}{\code{AnomalyOverTimePlotsMetadataResponse$fromJSON()}}
\item \href{#method-AnomalyOverTimePlotsMetadataResponse-clone}{\code{AnomalyOverTimePlotsMetadataResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyOverTimePlotsMetadataResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyOverTimePlotsMetadataResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyOverTimePlotsMetadataResponse$new(
  backtestMetadata = NULL,
  backtestStatuses = NULL,
  holdoutMetadata = NULL,
  holdoutStatuses = NULL,
  resolutions = NULL,
  estimatedSeriesLimit = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{backtestMetadata}}{An array of metadata information for each backtest. The array index of metadata object is the backtest index.}

\item{\code{backtestStatuses}}{An array of status information for each backtest. The array index of status object is the backtest index.}

\item{\code{resolutions}}{An array of available time resolutions for which plots can be retrieved.}

\item{\code{estimatedSeriesLimit}}{Estimated number of series that can be calculated in one request for 1 FD.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyOverTimePlotsMetadataResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyOverTimePlotsMetadataResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyOverTimePlotsMetadataResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyOverTimePlotsMetadataResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyOverTimePlotsMetadataResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyOverTimePlotsMetadataResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyOverTimePlotsMetadataResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyOverTimePlotsMetadataResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyOverTimePlotsMetadataResponse$fromJSON(
  AnomalyOverTimePlotsMetadataResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AnomalyOverTimePlotsMetadataResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyOverTimePlotsMetadataResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyOverTimePlotsMetadataResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyOverTimePlotsMetadataResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
