% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_over_time_plots_preview_bins.R
\docType{class}
\name{AnomalyOverTimePlotsPreviewBins}
\alias{AnomalyOverTimePlotsPreviewBins}
\title{AnomalyOverTimePlotsPreviewBins}
\format{
An \code{R6Class} generator object
}
\description{
AnomalyOverTimePlotsPreviewBins Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{endDate}}{character The datetime of the end of the bin (exclusive).}

\item{\code{startDate}}{character The datetime of the start of the bin (inclusive).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnomalyOverTimePlotsPreviewBins-new}{\code{AnomalyOverTimePlotsPreviewBins$new()}}
\item \href{#method-AnomalyOverTimePlotsPreviewBins-validate}{\code{AnomalyOverTimePlotsPreviewBins$validate()}}
\item \href{#method-AnomalyOverTimePlotsPreviewBins-toJSON}{\code{AnomalyOverTimePlotsPreviewBins$toJSON()}}
\item \href{#method-AnomalyOverTimePlotsPreviewBins-fromJSON}{\code{AnomalyOverTimePlotsPreviewBins$fromJSON()}}
\item \href{#method-AnomalyOverTimePlotsPreviewBins-clone}{\code{AnomalyOverTimePlotsPreviewBins$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyOverTimePlotsPreviewBins-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyOverTimePlotsPreviewBins-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyOverTimePlotsPreviewBins$new(
  endDate = NULL,
  startDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endDate}}{The datetime of the end of the bin (exclusive).}

\item{\code{startDate}}{The datetime of the start of the bin (inclusive).}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyOverTimePlotsPreviewBins-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyOverTimePlotsPreviewBins-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyOverTimePlotsPreviewBins$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyOverTimePlotsPreviewBins-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyOverTimePlotsPreviewBins-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyOverTimePlotsPreviewBins$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyOverTimePlotsPreviewBins-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyOverTimePlotsPreviewBins-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyOverTimePlotsPreviewBins$fromJSON(
  AnomalyOverTimePlotsPreviewBinsJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AnomalyOverTimePlotsPreviewBinsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyOverTimePlotsPreviewBins-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyOverTimePlotsPreviewBins-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyOverTimePlotsPreviewBins$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
