% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aot_chart_bins.R
\docType{class}
\name{AOTChartBins}
\alias{AOTChartBins}
\title{AOTChartBins}
\format{
An \code{R6Class} generator object
}
\description{
AOTChartBins Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual}}{numeric The average actual value of the target in the bin. ``null`` if there are no entries in the bin or if this is an anomaly detection project.}

\item{\code{endDate}}{character The datetime of the end of the bin (exclusive).}

\item{\code{frequency}}{numeric As indicated by the frequencyType in the Metadata, used to determine what the averages mentioned above are taken over. ``null`` if there are no entries in the bin.}

\item{\code{predicted}}{numeric The average prediction of the model in the bin. ``null`` if there are no entries in the bin.}

\item{\code{startDate}}{character The datetime of the start of the bin (inclusive).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AOTChartBins-new}{\code{AOTChartBins$new()}}
\item \href{#method-AOTChartBins-validate}{\code{AOTChartBins$validate()}}
\item \href{#method-AOTChartBins-toJSON}{\code{AOTChartBins$toJSON()}}
\item \href{#method-AOTChartBins-fromJSON}{\code{AOTChartBins$fromJSON()}}
\item \href{#method-AOTChartBins-clone}{\code{AOTChartBins$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AOTChartBins-new"></a>}}
\if{latex}{\out{\hypertarget{method-AOTChartBins-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOTChartBins$new(
  actual = NULL,
  endDate = NULL,
  frequency = NULL,
  predicted = NULL,
  startDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{actual}}{The average actual value of the target in the bin. ``null`` if there are no entries in the bin or if this is an anomaly detection project.}

\item{\code{endDate}}{The datetime of the end of the bin (exclusive).}

\item{\code{frequency}}{As indicated by the frequencyType in the Metadata, used to determine what the averages mentioned above are taken over. ``null`` if there are no entries in the bin.}

\item{\code{predicted}}{The average prediction of the model in the bin. ``null`` if there are no entries in the bin.}

\item{\code{startDate}}{The datetime of the start of the bin (inclusive).}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AOTChartBins-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AOTChartBins-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOTChartBins$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AOTChartBins-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AOTChartBins-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOTChartBins$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AOTChartBins-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AOTChartBins-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOTChartBins$fromJSON(AOTChartBinsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AOTChartBinsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AOTChartBins-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AOTChartBins-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOTChartBins$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
