% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aot_chart_metadata_response.R
\docType{class}
\name{AOTChartMetadataResponse}
\alias{AOTChartMetadataResponse}
\title{AOTChartMetadataResponse}
\format{
An \code{R6Class} generator object
}
\description{
AOTChartMetadataResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datasetId}}{character The dataset id which was used to compute the AOT chart.}

\item{\code{datasetMetadata}}{\link{AOTChartMetadataDatasetMetadata}}

\item{\code{frequencyType}}{character How to interpret the frequency attribute of each datetimeTrendBin. One of \link{'rowCount', 'weightedRowCount', 'exposure', 'weightedExposure'}.}

\item{\code{metricName}}{character The metric used to score each bin and the calculate the metric attribute of each datetimeTrendBin.}

\item{\code{modelId}}{character The model id which was used to compute the AOT chart.}

\item{\code{projectId}}{character The project id which was used to compute the AOT chart.}

\item{\code{resolutions}}{list( character ) Suggested time resolutions where a resolution is one of \link{'milliseconds', 'seconds', 'minutes', 'hours', 'days', 'weeks', 'months', 'years'}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AOTChartMetadataResponse-new}{\code{AOTChartMetadataResponse$new()}}
\item \href{#method-AOTChartMetadataResponse-validate}{\code{AOTChartMetadataResponse$validate()}}
\item \href{#method-AOTChartMetadataResponse-toJSON}{\code{AOTChartMetadataResponse$toJSON()}}
\item \href{#method-AOTChartMetadataResponse-fromJSON}{\code{AOTChartMetadataResponse$fromJSON()}}
\item \href{#method-AOTChartMetadataResponse-clone}{\code{AOTChartMetadataResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AOTChartMetadataResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-AOTChartMetadataResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOTChartMetadataResponse$new(
  datasetId = NULL,
  datasetMetadata = NULL,
  frequencyType = NULL,
  metricName = NULL,
  modelId = NULL,
  projectId = NULL,
  resolutions = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasetId}}{The dataset id which was used to compute the AOT chart.}

\item{\code{frequencyType}}{How to interpret the frequency attribute of each datetimeTrendBin. One of \link{'rowCount', 'weightedRowCount', 'exposure', 'weightedExposure'}.}

\item{\code{metricName}}{The metric used to score each bin and the calculate the metric attribute of each datetimeTrendBin.}

\item{\code{modelId}}{The model id which was used to compute the AOT chart.}

\item{\code{projectId}}{The project id which was used to compute the AOT chart.}

\item{\code{resolutions}}{Suggested time resolutions where a resolution is one of \link{'milliseconds', 'seconds', 'minutes', 'hours', 'days', 'weeks', 'months', 'years'}.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AOTChartMetadataResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AOTChartMetadataResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOTChartMetadataResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AOTChartMetadataResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AOTChartMetadataResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOTChartMetadataResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AOTChartMetadataResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AOTChartMetadataResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOTChartMetadataResponse$fromJSON(
  AOTChartMetadataResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AOTChartMetadataResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AOTChartMetadataResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AOTChartMetadataResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOTChartMetadataResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
