% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aot_chart_preview_response.R
\docType{class}
\name{AOTChartPreviewResponse}
\alias{AOTChartPreviewResponse}
\title{AOTChartPreviewResponse}
\format{
An \code{R6Class} generator object
}
\description{
AOTChartPreviewResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bins}}{list( \link{AOTChartBins} ) The datetime chart data for that source.}

\item{\code{datasetId}}{character The dataset id which was used to compute the AOT chart.}

\item{\code{modelId}}{character The model id which was used to compute the AOT chart.}

\item{\code{projectId}}{character The project id which was used to compute the AOT chart.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AOTChartPreviewResponse-new}{\code{AOTChartPreviewResponse$new()}}
\item \href{#method-AOTChartPreviewResponse-validate}{\code{AOTChartPreviewResponse$validate()}}
\item \href{#method-AOTChartPreviewResponse-toJSON}{\code{AOTChartPreviewResponse$toJSON()}}
\item \href{#method-AOTChartPreviewResponse-fromJSON}{\code{AOTChartPreviewResponse$fromJSON()}}
\item \href{#method-AOTChartPreviewResponse-clone}{\code{AOTChartPreviewResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AOTChartPreviewResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-AOTChartPreviewResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOTChartPreviewResponse$new(
  bins = NULL,
  datasetId = NULL,
  modelId = NULL,
  projectId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bins}}{The datetime chart data for that source.}

\item{\code{datasetId}}{The dataset id which was used to compute the AOT chart.}

\item{\code{modelId}}{The model id which was used to compute the AOT chart.}

\item{\code{projectId}}{The project id which was used to compute the AOT chart.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AOTChartPreviewResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AOTChartPreviewResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOTChartPreviewResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AOTChartPreviewResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AOTChartPreviewResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOTChartPreviewResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AOTChartPreviewResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AOTChartPreviewResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOTChartPreviewResponse$fromJSON(
  AOTChartPreviewResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AOTChartPreviewResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AOTChartPreviewResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AOTChartPreviewResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOTChartPreviewResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
