% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aot_chart_retrieve_response.R
\docType{class}
\name{AOTChartRetrieveResponse}
\alias{AOTChartRetrieveResponse}
\title{AOTChartRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
AOTChartRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bins}}{list( \link{AOTChartBins} ) The datetime chart data for that source.}

\item{\code{datasetId}}{character The dataset id to which the chart data belongs.}

\item{\code{endDate}}{character \link{optional} The requested `endDate`, or, if not specified, the end date for this dataset (exclusive). Example: '2010-05-13T00:00:00.000000Z'.}

\item{\code{modelId}}{character The model id to which the chart data belongs.}

\item{\code{projectId}}{character The project id to which the chart data belongs.}

\item{\code{resolution}}{character The resolution used for binning where a resolution is one of \link{'milliseconds', 'seconds', 'minutes', 'hours', 'days', 'weeks', 'months', 'years'}.}

\item{\code{startDate}}{character \link{optional} The requested `startDate`, or, if not specified, the start date for this dataset. Example: '2010-05-13T00:00:00.000000Z'.}

\item{\code{statistics}}{\link{AOTChartStatistics}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AOTChartRetrieveResponse$new()}}
\item \href{#method-validate}{\code{AOTChartRetrieveResponse$validate()}}
\item \href{#method-toJSON}{\code{AOTChartRetrieveResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{AOTChartRetrieveResponse$fromJSON()}}
\item \href{#method-clone}{\code{AOTChartRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOTChartRetrieveResponse$new(
  bins = NULL,
  datasetId = NULL,
  modelId = NULL,
  projectId = NULL,
  resolution = NULL,
  statistics = NULL,
  endDate = NULL,
  startDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bins}}{The datetime chart data for that source.}

\item{\code{datasetId}}{The dataset id to which the chart data belongs.}

\item{\code{modelId}}{The model id to which the chart data belongs.}

\item{\code{projectId}}{The project id to which the chart data belongs.}

\item{\code{resolution}}{The resolution used for binning where a resolution is one of \link{'milliseconds', 'seconds', 'minutes', 'hours', 'days', 'weeks', 'months', 'years'}.}

\item{\code{endDate}}{The requested `endDate`, or, if not specified, the end date for this dataset (exclusive). Example: '2010-05-13T00:00:00.000000Z'.}

\item{\code{startDate}}{The requested `startDate`, or, if not specified, the start date for this dataset. Example: '2010-05-13T00:00:00.000000Z'.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOTChartRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOTChartRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOTChartRetrieveResponse$fromJSON(
  AOTChartRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AOTChartRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOTChartRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
