% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application.R
\docType{class}
\name{Application}
\alias{Application}
\title{Application}
\format{
An \code{R6Class} generator object
}
\description{
Application Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{applicationTemplateType}}{character Application template type, purpose}

\item{\code{applicationTypeId}}{character The ID of the type of the application}

\item{\code{cloudProvider}}{character The host of this application}

\item{\code{createdAt}}{character The timestamp when the application was created}

\item{\code{createdBy}}{character The username of who created the application}

\item{\code{creatorFirstName}}{character Application creator first name}

\item{\code{creatorLastName}}{character Application creator last name}

\item{\code{creatorUserhash}}{character Application creator userhash}

\item{\code{datasets}}{list( character ) The list of datasets IDs associated with the application}

\item{\code{deactivationStatusId}}{character The ID of the status object to track the asynchronous app deactivation process status. Will be null if the app was never deactivated.}

\item{\code{deploymentIds}}{list( character ) A list of deployment IDs for this app}

\item{\code{deploymentName}}{character Name of the deployment}

\item{\code{deploymentState}}{character The state of the application deployment}

\item{\code{deploymentStatusId}}{character  The ID of the status object to track the asynchronous deployment process status}

\item{\code{deployments}}{list( \link{ApplicationDeployment} ) A list of deployment details}

\item{\code{description}}{character A description of the application}

\item{\code{id}}{character The ID of the created application}

\item{\code{modelDeploymentId}}{character The ID of the associated model deployment}

\item{\code{name}}{character The name of the application}

\item{\code{permissions}}{list( character ) The list of permitted actions, which the authenticated user can perform on this application.}

\item{\code{poolUsed}}{character Whether the pool where used for last app deployment}

\item{\code{recentEvents}}{list( character ) The list of deployment events of the application.}

\item{\code{updatedAt}}{character The timestamp when the application was updated}

\item{\code{url}}{character The URL of the active application}

\item{\code{userId}}{character The ID of the user which created the application}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Application$new()}}
\item \href{#method-validate}{\code{Application$validate()}}
\item \href{#method-toJSON}{\code{Application$toJSON()}}
\item \href{#method-fromJSON}{\code{Application$fromJSON()}}
\item \href{#method-clone}{\code{Application$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Application$new(
  applicationTemplateType = NULL,
  applicationTypeId = NULL,
  cloudProvider = NULL,
  createdAt = NULL,
  createdBy = NULL,
  creatorFirstName = NULL,
  creatorLastName = NULL,
  creatorUserhash = NULL,
  datasets = NULL,
  deactivationStatusId = NULL,
  deploymentIds = NULL,
  deploymentName = NULL,
  deploymentState = NULL,
  deploymentStatusId = NULL,
  deployments = NULL,
  description = NULL,
  id = NULL,
  modelDeploymentId = NULL,
  name = NULL,
  permissions = NULL,
  poolUsed = NULL,
  recentEvents = NULL,
  updatedAt = NULL,
  url = NULL,
  userId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{applicationTemplateType}}{Application template type, purpose}

\item{\code{applicationTypeId}}{The ID of the type of the application}

\item{\code{cloudProvider}}{The host of this application}

\item{\code{createdAt}}{The timestamp when the application was created}

\item{\code{createdBy}}{The username of who created the application}

\item{\code{creatorFirstName}}{Application creator first name}

\item{\code{creatorLastName}}{Application creator last name}

\item{\code{creatorUserhash}}{Application creator userhash}

\item{\code{datasets}}{The list of datasets IDs associated with the application}

\item{\code{deactivationStatusId}}{The ID of the status object to track the asynchronous app deactivation process status. Will be null if the app was never deactivated.}

\item{\code{deploymentIds}}{A list of deployment IDs for this app}

\item{\code{deploymentName}}{Name of the deployment}

\item{\code{deploymentState}}{The state of the application deployment}

\item{\code{deploymentStatusId}}{The ID of the status object to track the asynchronous deployment process status}

\item{\code{deployments}}{A list of deployment details}

\item{\code{description}}{A description of the application}

\item{\code{id}}{The ID of the created application}

\item{\code{modelDeploymentId}}{The ID of the associated model deployment}

\item{\code{name}}{The name of the application}

\item{\code{permissions}}{The list of permitted actions, which the authenticated user can perform on this application.}

\item{\code{poolUsed}}{Whether the pool where used for last app deployment}

\item{\code{recentEvents}}{The list of deployment events of the application.}

\item{\code{updatedAt}}{The timestamp when the application was updated}

\item{\code{url}}{The URL of the active application}

\item{\code{userId}}{The ID of the user which created the application}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Application$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Application$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Application$fromJSON(ApplicationJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ApplicationJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Application$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
