% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_create.R
\docType{class}
\name{ApplicationCreate}
\alias{ApplicationCreate}
\title{ApplicationCreate}
\format{
An \code{R6Class} generator object
}
\description{
ApplicationCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{applicationTypeId}}{character \link{optional} The ID of the of application to be created.}

\item{\code{authenticationType}}{character \link{optional} Authentication type}

\item{\code{cloudProvider}}{character \link{optional} The optional cloud provider}

\item{\code{description}}{character \link{optional} The description of the application}

\item{\code{modelDeploymentId}}{character \link{optional} The ID of the model deployment. The deployed application will use this deployment to make predictions.}

\item{\code{name}}{character \link{optional} The name of the app}

\item{\code{purpose}}{character \link{optional} An optional field to describe the purpose of the application.}

\item{\code{sources}}{list( \link{ApplicationCreateSources} ) \link{optional} The sources for this application}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApplicationCreate-new}{\code{ApplicationCreate$new()}}
\item \href{#method-ApplicationCreate-validate}{\code{ApplicationCreate$validate()}}
\item \href{#method-ApplicationCreate-toJSON}{\code{ApplicationCreate$toJSON()}}
\item \href{#method-ApplicationCreate-fromJSON}{\code{ApplicationCreate$fromJSON()}}
\item \href{#method-ApplicationCreate-clone}{\code{ApplicationCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationCreate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationCreate$new(
  applicationTypeId = NULL,
  authenticationType = "invitedUsersOnly",
  cloudProvider = "drcloud",
  description = NULL,
  modelDeploymentId = NULL,
  name = NULL,
  purpose = NULL,
  sources = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{applicationTypeId}}{The ID of the of application to be created.}

\item{\code{authenticationType}}{Authentication type}

\item{\code{cloudProvider}}{The optional cloud provider}

\item{\code{description}}{The description of the application}

\item{\code{modelDeploymentId}}{The ID of the model deployment. The deployed application will use this deployment to make predictions.}

\item{\code{name}}{The name of the app}

\item{\code{purpose}}{An optional field to describe the purpose of the application.}

\item{\code{sources}}{The sources for this application}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationCreate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationCreate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationCreate$fromJSON(ApplicationCreateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ApplicationCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
