% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_retrieve.R
\docType{class}
\name{ApplicationRetrieve}
\alias{ApplicationRetrieve}
\title{ApplicationRetrieve}
\format{
An \code{R6Class} generator object
}
\description{
ApplicationRetrieve Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{applicationTemplateType}}{character Application template type, purpose}

\item{\code{applicationTypeId}}{character The ID of the type of the application}

\item{\code{cloudProvider}}{character The host of this application}

\item{\code{createdAt}}{character The timestamp when the application was created}

\item{\code{createdBy}}{character The username of who created the application}

\item{\code{creatorFirstName}}{character Application creator first name}

\item{\code{creatorLastName}}{character Application creator last name}

\item{\code{creatorUserhash}}{character Application creator userhash}

\item{\code{datasets}}{list( character ) The list of datasets IDs associated with the application}

\item{\code{deactivationStatusId}}{character The ID of the status object to track the asynchronous app deactivation process status. Will be null if the app was never deactivated.}

\item{\code{deploymentIds}}{list( character ) A list of deployment IDs for this app}

\item{\code{deploymentName}}{character Name of the deployment}

\item{\code{deploymentStatusId}}{character  The ID of the status object to track the asynchronous deployment process status}

\item{\code{deployments}}{list( \link{ApplicationDeployment} ) A list of deployment details}

\item{\code{description}}{character A description of the application}

\item{\code{hasCustomLogo}}{character Whether the app has a custom logo}

\item{\code{id}}{character The ID of the created application}

\item{\code{modelDeploymentId}}{character \link{optional} The ID of the model deployment. The deployed application will use this deployment to make predictions.}

\item{\code{name}}{character The name of the application}

\item{\code{permissions}}{list( character ) The list of permitted actions, which the authenticated user can perform on this application.}

\item{\code{poolUsed}}{character Whether the pool where used for last app deployment}

\item{\code{updatedAt}}{character The timestamp when the application was updated}

\item{\code{userId}}{character The ID of the user which created the application}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApplicationRetrieve-new}{\code{ApplicationRetrieve$new()}}
\item \href{#method-ApplicationRetrieve-validate}{\code{ApplicationRetrieve$validate()}}
\item \href{#method-ApplicationRetrieve-toJSON}{\code{ApplicationRetrieve$toJSON()}}
\item \href{#method-ApplicationRetrieve-fromJSON}{\code{ApplicationRetrieve$fromJSON()}}
\item \href{#method-ApplicationRetrieve-clone}{\code{ApplicationRetrieve$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationRetrieve-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationRetrieve-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationRetrieve$new(
  applicationTemplateType = NULL,
  applicationTypeId = NULL,
  cloudProvider = NULL,
  createdAt = NULL,
  createdBy = NULL,
  creatorFirstName = NULL,
  creatorLastName = NULL,
  creatorUserhash = NULL,
  datasets = NULL,
  deactivationStatusId = NULL,
  deploymentIds = NULL,
  deploymentName = NULL,
  deploymentStatusId = NULL,
  deployments = NULL,
  description = NULL,
  hasCustomLogo = NULL,
  id = NULL,
  name = NULL,
  permissions = NULL,
  poolUsed = NULL,
  updatedAt = NULL,
  userId = NULL,
  modelDeploymentId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{applicationTemplateType}}{Application template type, purpose}

\item{\code{applicationTypeId}}{The ID of the type of the application}

\item{\code{cloudProvider}}{The host of this application}

\item{\code{createdAt}}{The timestamp when the application was created}

\item{\code{createdBy}}{The username of who created the application}

\item{\code{creatorFirstName}}{Application creator first name}

\item{\code{creatorLastName}}{Application creator last name}

\item{\code{creatorUserhash}}{Application creator userhash}

\item{\code{datasets}}{The list of datasets IDs associated with the application}

\item{\code{deactivationStatusId}}{The ID of the status object to track the asynchronous app deactivation process status. Will be null if the app was never deactivated.}

\item{\code{deploymentIds}}{A list of deployment IDs for this app}

\item{\code{deploymentName}}{Name of the deployment}

\item{\code{deploymentStatusId}}{The ID of the status object to track the asynchronous deployment process status}

\item{\code{deployments}}{A list of deployment details}

\item{\code{description}}{A description of the application}

\item{\code{hasCustomLogo}}{Whether the app has a custom logo}

\item{\code{id}}{The ID of the created application}

\item{\code{name}}{The name of the application}

\item{\code{permissions}}{The list of permitted actions, which the authenticated user can perform on this application.}

\item{\code{poolUsed}}{Whether the pool where used for last app deployment}

\item{\code{updatedAt}}{The timestamp when the application was updated}

\item{\code{userId}}{The ID of the user which created the application}

\item{\code{modelDeploymentId}}{The ID of the model deployment. The deployed application will use this deployment to make predictions.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationRetrieve-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationRetrieve-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationRetrieve$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationRetrieve-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationRetrieve-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationRetrieve$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationRetrieve-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationRetrieve-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationRetrieve$fromJSON(ApplicationRetrieveJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ApplicationRetrieveJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationRetrieve-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationRetrieve-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationRetrieve$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
