% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_shared_roles_list.R
\docType{class}
\name{ApplicationSharedRolesList}
\alias{ApplicationSharedRolesList}
\title{ApplicationSharedRolesList}
\format{
An \code{R6Class} generator object
}
\description{
ApplicationSharedRolesList Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer Number of items returned on this page}

\item{\code{data}}{list( \link{ApplicationSharedRolesEntry} ) Details about the Shared Role entries}

\item{\code{next_}}{character URL pointing to the next page (if null, there is no next page)}

\item{\code{previous}}{character URL pointing to the previous page (if null, there is no previous page)}

\item{\code{totalCount}}{integer Number of items matching to the query condition}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApplicationSharedRolesList-new}{\code{ApplicationSharedRolesList$new()}}
\item \href{#method-ApplicationSharedRolesList-validate}{\code{ApplicationSharedRolesList$validate()}}
\item \href{#method-ApplicationSharedRolesList-toJSON}{\code{ApplicationSharedRolesList$toJSON()}}
\item \href{#method-ApplicationSharedRolesList-fromJSON}{\code{ApplicationSharedRolesList$fromJSON()}}
\item \href{#method-ApplicationSharedRolesList-clone}{\code{ApplicationSharedRolesList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationSharedRolesList-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationSharedRolesList-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationSharedRolesList$new(
  count = NULL,
  data = NULL,
  next_ = NULL,
  previous = NULL,
  totalCount = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{Number of items returned on this page}

\item{\code{data}}{Details about the Shared Role entries}

\item{\code{next_}}{URL pointing to the next page (if null, there is no next page)}

\item{\code{previous}}{URL pointing to the previous page (if null, there is no previous page)}

\item{\code{totalCount}}{Number of items matching to the query condition}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationSharedRolesList-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationSharedRolesList-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationSharedRolesList$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationSharedRolesList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationSharedRolesList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationSharedRolesList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationSharedRolesList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationSharedRolesList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationSharedRolesList$fromJSON(
  ApplicationSharedRolesListJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ApplicationSharedRolesListJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationSharedRolesList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationSharedRolesList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationSharedRolesList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
