% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_sharing_update_or_remove.R
\docType{class}
\name{ApplicationSharingUpdateOrRemove}
\alias{ApplicationSharingUpdateOrRemove}
\title{ApplicationSharingUpdateOrRemove}
\format{
An \code{R6Class} generator object
}
\description{
ApplicationSharingUpdateOrRemove Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{note}}{character \link{optional} A note to go with the project share}

\item{\code{operation}}{character Name of the action being taken. The only operation is 'updateRoles'.}

\item{\code{roles}}{list( \link{OneOfGrantAccessControlWithUsernameGrantAccessControlWithId} ) Array of GrantAccessControl objects., up to maximum 100 objects.}

\item{\code{sendNotification}}{character \link{optional} Send a notification?}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApplicationSharingUpdateOrRemove-new}{\code{ApplicationSharingUpdateOrRemove$new()}}
\item \href{#method-ApplicationSharingUpdateOrRemove-validate}{\code{ApplicationSharingUpdateOrRemove$validate()}}
\item \href{#method-ApplicationSharingUpdateOrRemove-toJSON}{\code{ApplicationSharingUpdateOrRemove$toJSON()}}
\item \href{#method-ApplicationSharingUpdateOrRemove-fromJSON}{\code{ApplicationSharingUpdateOrRemove$fromJSON()}}
\item \href{#method-ApplicationSharingUpdateOrRemove-clone}{\code{ApplicationSharingUpdateOrRemove$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationSharingUpdateOrRemove-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationSharingUpdateOrRemove-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationSharingUpdateOrRemove$new(
  operation = NULL,
  roles = NULL,
  note = "",
  sendNotification = FALSE,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{operation}}{Name of the action being taken. The only operation is 'updateRoles'.}

\item{\code{roles}}{Array of GrantAccessControl objects., up to maximum 100 objects.}

\item{\code{note}}{A note to go with the project share}

\item{\code{sendNotification}}{Send a notification?}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationSharingUpdateOrRemove-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationSharingUpdateOrRemove-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationSharingUpdateOrRemove$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationSharingUpdateOrRemove-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationSharingUpdateOrRemove-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationSharingUpdateOrRemove$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationSharingUpdateOrRemove-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationSharingUpdateOrRemove-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationSharingUpdateOrRemove$fromJSON(
  ApplicationSharingUpdateOrRemoveJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ApplicationSharingUpdateOrRemoveJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationSharingUpdateOrRemove-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationSharingUpdateOrRemove-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationSharingUpdateOrRemove$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
