% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_type_configuration_deployment.R
\docType{class}
\name{ApplicationTypeConfigurationDeployment}
\alias{ApplicationTypeConfigurationDeployment}
\title{ApplicationTypeConfigurationDeployment}
\format{
An \code{R6Class} generator object
}
\description{
ApplicationTypeConfigurationDeployment Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{description}}{character \link{optional} A description of the deployment configuration}

\item{\code{internalName}}{character The internal name of this deployment}

\item{\code{name}}{character The name of this deployment}

\item{\code{targetTypes}}{list( character ) \link{optional} The type of the deployment}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ApplicationTypeConfigurationDeployment$new()}}
\item \href{#method-validate}{\code{ApplicationTypeConfigurationDeployment$validate()}}
\item \href{#method-toJSON}{\code{ApplicationTypeConfigurationDeployment$toJSON()}}
\item \href{#method-fromJSON}{\code{ApplicationTypeConfigurationDeployment$fromJSON()}}
\item \href{#method-clone}{\code{ApplicationTypeConfigurationDeployment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationTypeConfigurationDeployment$new(
  internalName = NULL,
  name = NULL,
  description = NULL,
  targetTypes = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{internalName}}{The internal name of this deployment}

\item{\code{name}}{The name of this deployment}

\item{\code{description}}{A description of the deployment configuration}

\item{\code{targetTypes}}{The type of the deployment}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationTypeConfigurationDeployment$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationTypeConfigurationDeployment$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationTypeConfigurationDeployment$fromJSON(
  ApplicationTypeConfigurationDeploymentJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ApplicationTypeConfigurationDeploymentJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationTypeConfigurationDeployment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
