% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_type_list_response.R
\docType{class}
\name{ApplicationTypeListResponse}
\alias{ApplicationTypeListResponse}
\title{ApplicationTypeListResponse}
\format{
An \code{R6Class} generator object
}
\description{
ApplicationTypeListResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer Number of items}

\item{\code{data}}{list( \link{ApplicationTypeRetrieveResponse} ) each has the same schema as when application types retrieved individually from :http:get:`/api/v2/applicationTypes/(applicationTypeId)/`.}

\item{\code{next_}}{character URL pointing to the next page (if null, there is no next page)}

\item{\code{previous}}{character URL pointing to the previous page (if null, there is no previous page)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ApplicationTypeListResponse$new()}}
\item \href{#method-validate}{\code{ApplicationTypeListResponse$validate()}}
\item \href{#method-toJSON}{\code{ApplicationTypeListResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{ApplicationTypeListResponse$fromJSON()}}
\item \href{#method-clone}{\code{ApplicationTypeListResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationTypeListResponse$new(
  count = NULL,
  data = NULL,
  next_ = NULL,
  previous = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{Number of items}

\item{\code{data}}{each has the same schema as when application types retrieved individually from :http:get:`/api/v2/applicationTypes/(applicationTypeId)/`.}

\item{\code{next_}}{URL pointing to the next page (if null, there is no next page)}

\item{\code{previous}}{URL pointing to the previous page (if null, there is no previous page)}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationTypeListResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationTypeListResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationTypeListResponse$fromJSON(
  ApplicationTypeListResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ApplicationTypeListResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationTypeListResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
