% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_type_retrieve_response.R
\docType{class}
\name{ApplicationTypeRetrieveResponse}
\alias{ApplicationTypeRetrieveResponse}
\title{ApplicationTypeRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
ApplicationTypeRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{configuration}}{\link{ApplicationTypeConfiguration}}

\item{\code{demoAvailable}}{character Indicates if app type has the ability to launch a demo version}

\item{\code{demoDatasetLink}}{character Link to demo dataset that usable with app}

\item{\code{description}}{character the description of the application}

\item{\code{id}}{character The ID of the application type}

\item{\code{isLaunchable}}{character Indicates if app of this type could be deployed by current user}

\item{\code{logoImage}}{character The logo of the application encoded as base64 string PNG image}

\item{\code{maturity}}{character Indicates app maturity level}

\item{\code{name}}{character The name of the application type}

\item{\code{version}}{character The version}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ApplicationTypeRetrieveResponse$new()}}
\item \href{#method-validate}{\code{ApplicationTypeRetrieveResponse$validate()}}
\item \href{#method-toJSON}{\code{ApplicationTypeRetrieveResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{ApplicationTypeRetrieveResponse$fromJSON()}}
\item \href{#method-clone}{\code{ApplicationTypeRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationTypeRetrieveResponse$new(
  configuration = NULL,
  demoAvailable = NULL,
  demoDatasetLink = NULL,
  description = NULL,
  id = NULL,
  isLaunchable = NULL,
  logoImage = NULL,
  maturity = NULL,
  name = NULL,
  version = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{demoAvailable}}{Indicates if app type has the ability to launch a demo version}

\item{\code{demoDatasetLink}}{Link to demo dataset that usable with app}

\item{\code{description}}{the description of the application}

\item{\code{id}}{The ID of the application type}

\item{\code{isLaunchable}}{Indicates if app of this type could be deployed by current user}

\item{\code{logoImage}}{The logo of the application encoded as base64 string PNG image}

\item{\code{maturity}}{Indicates app maturity level}

\item{\code{name}}{The name of the application type}

\item{\code{version}}{The version}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationTypeRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationTypeRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationTypeRetrieveResponse$fromJSON(
  ApplicationTypeRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ApplicationTypeRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationTypeRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
