% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approval_policy_automatic_action.R
\docType{class}
\name{ApprovalPolicyAutomaticAction}
\alias{ApprovalPolicyAutomaticAction}
\title{ApprovalPolicyAutomaticAction}
\format{
An \code{R6Class} generator object
}
\description{
ApprovalPolicyAutomaticAction Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{action}}{character Action of the workflow automation.}

\item{\code{period}}{character Period (ISO 8601) after which an action is executed on a Change Request if it is not resolved or cancelled.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApprovalPolicyAutomaticAction-new}{\code{ApprovalPolicyAutomaticAction$new()}}
\item \href{#method-ApprovalPolicyAutomaticAction-validate}{\code{ApprovalPolicyAutomaticAction$validate()}}
\item \href{#method-ApprovalPolicyAutomaticAction-toJSON}{\code{ApprovalPolicyAutomaticAction$toJSON()}}
\item \href{#method-ApprovalPolicyAutomaticAction-fromJSON}{\code{ApprovalPolicyAutomaticAction$fromJSON()}}
\item \href{#method-ApprovalPolicyAutomaticAction-clone}{\code{ApprovalPolicyAutomaticAction$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyAutomaticAction-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyAutomaticAction-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyAutomaticAction$new(
  action = NULL,
  period = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action}}{Action of the workflow automation.}

\item{\code{period}}{Period (ISO 8601) after which an action is executed on a Change Request if it is not resolved or cancelled.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyAutomaticAction-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyAutomaticAction-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyAutomaticAction$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyAutomaticAction-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyAutomaticAction-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyAutomaticAction$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyAutomaticAction-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyAutomaticAction-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyAutomaticAction$fromJSON(
  ApprovalPolicyAutomaticActionJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ApprovalPolicyAutomaticActionJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyAutomaticAction-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyAutomaticAction-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyAutomaticAction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
