% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approval_policy_intended_action_condition.R
\docType{class}
\name{ApprovalPolicyIntendedActionCondition}
\alias{ApprovalPolicyIntendedActionCondition}
\title{ApprovalPolicyIntendedActionCondition}
\format{
An \code{R6Class} generator object
}
\description{
ApprovalPolicyIntendedActionCondition Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{condition}}{character Condition for the field content to trigger on.}

\item{\code{fieldName}}{character Name of the attribute of the entity to filter entities by. An example value is ``importance`` attribute for the deployment entity.}

\item{\code{values}}{list( character ) Array of field values to apply condition to trigger on. If ``null`` for ``equals`` condition, then any value for the field is accepted.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ApprovalPolicyIntendedActionCondition$new()}}
\item \href{#method-validate}{\code{ApprovalPolicyIntendedActionCondition$validate()}}
\item \href{#method-toJSON}{\code{ApprovalPolicyIntendedActionCondition$toJSON()}}
\item \href{#method-fromJSON}{\code{ApprovalPolicyIntendedActionCondition$fromJSON()}}
\item \href{#method-clone}{\code{ApprovalPolicyIntendedActionCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyIntendedActionCondition$new(
  condition = NULL,
  fieldName = NULL,
  values = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{condition}}{Condition for the field content to trigger on.}

\item{\code{fieldName}}{Name of the attribute of the entity to filter entities by. An example value is ``importance`` attribute for the deployment entity.}

\item{\code{values}}{Array of field values to apply condition to trigger on. If ``null`` for ``equals`` condition, then any value for the field is accepted.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyIntendedActionCondition$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyIntendedActionCondition$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyIntendedActionCondition$fromJSON(
  ApprovalPolicyIntendedActionConditionJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ApprovalPolicyIntendedActionConditionJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyIntendedActionCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
