% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approval_policy_match_response.R
\docType{class}
\name{ApprovalPolicyMatchResponse}
\alias{ApprovalPolicyMatchResponse}
\title{ApprovalPolicyMatchResponse}
\format{
An \code{R6Class} generator object
}
\description{
ApprovalPolicyMatchResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{action}}{character Searched policy action.}

\item{\code{entityType}}{character Searched typed of the entity.}

\item{\code{fieldName}}{character \link{optional} Name of the entity field to filter policies by.}

\item{\code{fieldValue}}{character \link{optional} Value of the entity field to filter policies by.}

\item{\code{policyId}}{character ID of the matching approval policy. ``null`` if no matching policies found.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApprovalPolicyMatchResponse-new}{\code{ApprovalPolicyMatchResponse$new()}}
\item \href{#method-ApprovalPolicyMatchResponse-validate}{\code{ApprovalPolicyMatchResponse$validate()}}
\item \href{#method-ApprovalPolicyMatchResponse-toJSON}{\code{ApprovalPolicyMatchResponse$toJSON()}}
\item \href{#method-ApprovalPolicyMatchResponse-fromJSON}{\code{ApprovalPolicyMatchResponse$fromJSON()}}
\item \href{#method-ApprovalPolicyMatchResponse-clone}{\code{ApprovalPolicyMatchResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyMatchResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyMatchResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyMatchResponse$new(
  action = NULL,
  entityType = NULL,
  policyId = NULL,
  fieldName = NULL,
  fieldValue = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action}}{Searched policy action.}

\item{\code{entityType}}{Searched typed of the entity.}

\item{\code{policyId}}{ID of the matching approval policy. ``null`` if no matching policies found.}

\item{\code{fieldName}}{Name of the entity field to filter policies by.}

\item{\code{fieldValue}}{Value of the entity field to filter policies by.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyMatchResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyMatchResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyMatchResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyMatchResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyMatchResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyMatchResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyMatchResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyMatchResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyMatchResponse$fromJSON(
  ApprovalPolicyMatchResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ApprovalPolicyMatchResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyMatchResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyMatchResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyMatchResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
