% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approval_policy_response.R
\docType{class}
\name{ApprovalPolicyResponse}
\alias{ApprovalPolicyResponse}
\title{ApprovalPolicyResponse}
\format{
An \code{R6Class} generator object
}
\description{
ApprovalPolicyResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{active}}{character Whether this policy is active.}

\item{\code{automaticAction}}{\link{ApprovalPolicyAutomaticAction} \link{optional}}

\item{\code{id}}{character ID of the Approval Policy.}

\item{\code{name}}{character Name of the Approval Policy.}

\item{\code{openRequests}}{integer Number of open Change Requests associated with the policy.}

\item{\code{review}}{\link{ApprovalPolicyReview} \link{optional}}

\item{\code{trigger}}{\link{ApprovalPolicyTrigger}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApprovalPolicyResponse-new}{\code{ApprovalPolicyResponse$new()}}
\item \href{#method-ApprovalPolicyResponse-validate}{\code{ApprovalPolicyResponse$validate()}}
\item \href{#method-ApprovalPolicyResponse-toJSON}{\code{ApprovalPolicyResponse$toJSON()}}
\item \href{#method-ApprovalPolicyResponse-fromJSON}{\code{ApprovalPolicyResponse$fromJSON()}}
\item \href{#method-ApprovalPolicyResponse-clone}{\code{ApprovalPolicyResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyResponse$new(
  active = NULL,
  id = NULL,
  name = NULL,
  openRequests = NULL,
  trigger = NULL,
  automaticAction = NULL,
  review = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{active}}{Whether this policy is active.}

\item{\code{id}}{ID of the Approval Policy.}

\item{\code{name}}{Name of the Approval Policy.}

\item{\code{openRequests}}{Number of open Change Requests associated with the policy.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyResponse$fromJSON(
  ApprovalPolicyResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ApprovalPolicyResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
