% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approval_policy_review.R
\docType{class}
\name{ApprovalPolicyReview}
\alias{ApprovalPolicyReview}
\title{ApprovalPolicyReview}
\format{
An \code{R6Class} generator object
}
\description{
ApprovalPolicyReview Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{groups}}{list( \link{PolicyUserGroup} ) \link{optional} A list of user groups that will be added as required reviewers on Change Requests for the entities that match the policy.}

\item{\code{reminderPeriod}}{character \link{optional} Duration period in ISO 8601 format that indicates when to send a reminder for reviewing a Change Request after its creation or last reminder if it hasn't been approved yet. If ``null``, no review reminders are sent to the reviewers.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApprovalPolicyReview-new}{\code{ApprovalPolicyReview$new()}}
\item \href{#method-ApprovalPolicyReview-validate}{\code{ApprovalPolicyReview$validate()}}
\item \href{#method-ApprovalPolicyReview-toJSON}{\code{ApprovalPolicyReview$toJSON()}}
\item \href{#method-ApprovalPolicyReview-fromJSON}{\code{ApprovalPolicyReview$fromJSON()}}
\item \href{#method-ApprovalPolicyReview-clone}{\code{ApprovalPolicyReview$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyReview-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyReview-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyReview$new(
  groups = NULL,
  reminderPeriod = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groups}}{A list of user groups that will be added as required reviewers on Change Requests for the entities that match the policy.}

\item{\code{reminderPeriod}}{Duration period in ISO 8601 format that indicates when to send a reminder for reviewing a Change Request after its creation or last reminder if it hasn't been approved yet. If ``null``, no review reminders are sent to the reviewers.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyReview-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyReview-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyReview$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyReview-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyReview-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyReview$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyReview-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyReview-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyReview$fromJSON(ApprovalPolicyReviewJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ApprovalPolicyReviewJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyReview-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyReview-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyReview$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
