% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approval_policy_trigger.R
\docType{class}
\name{ApprovalPolicyTrigger}
\alias{ApprovalPolicyTrigger}
\title{ApprovalPolicyTrigger}
\format{
An \code{R6Class} generator object
}
\description{
ApprovalPolicyTrigger Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entityType}}{character Type of entity to trigger on.}

\item{\code{filterGroups}}{list( \link{PolicyUserGroup} ) \link{optional} A list of user groups to apply Approval Policy for. If User 'A' and User 'B' are both members of the same organisation, and User 'A' is a member of one of the groups listed in this field,but User 'B' is not, then an approval workflow will be triggerred for User 'A' on an action to the entity that matches policy condition, but not for User 'B'. If ``null``, approvals workflow will be triggered for all users.}

\item{\code{intendedAction}}{\link{ApprovalPolicyIntendedAction}}

\item{\code{labels}}{\link{ApprovalPolicyTriggerLabel} \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApprovalPolicyTrigger-new}{\code{ApprovalPolicyTrigger$new()}}
\item \href{#method-ApprovalPolicyTrigger-validate}{\code{ApprovalPolicyTrigger$validate()}}
\item \href{#method-ApprovalPolicyTrigger-toJSON}{\code{ApprovalPolicyTrigger$toJSON()}}
\item \href{#method-ApprovalPolicyTrigger-fromJSON}{\code{ApprovalPolicyTrigger$fromJSON()}}
\item \href{#method-ApprovalPolicyTrigger-clone}{\code{ApprovalPolicyTrigger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyTrigger-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyTrigger-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyTrigger$new(
  entityType = NULL,
  intendedAction = NULL,
  filterGroups = NULL,
  labels = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entityType}}{Type of entity to trigger on.}

\item{\code{filterGroups}}{A list of user groups to apply Approval Policy for. If User 'A' and User 'B' are both members of the same organisation, and User 'A' is a member of one of the groups listed in this field,but User 'B' is not, then an approval workflow will be triggerred for User 'A' on an action to the entity that matches policy condition, but not for User 'B'. If ``null``, approvals workflow will be triggered for all users.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyTrigger-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyTrigger-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyTrigger$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyTrigger-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyTrigger-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyTrigger$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyTrigger-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyTrigger-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyTrigger$fromJSON(
  ApprovalPolicyTriggerJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ApprovalPolicyTriggerJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApprovalPolicyTrigger-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApprovalPolicyTrigger-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApprovalPolicyTrigger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
