% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_id.R
\docType{class}
\name{AssociationID}
\alias{AssociationID}
\title{AssociationID}
\format{
An \code{R6Class} generator object
}
\description{
AssociationID Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{columnNames}}{list( character ) Indicates the columns that will be used to associate actuals with predictions.}

\item{\code{requiredInPredictionRequests}}{character Indicates whether the association ID is required in prediction requests.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AssociationID-new}{\code{AssociationID$new()}}
\item \href{#method-AssociationID-validate}{\code{AssociationID$validate()}}
\item \href{#method-AssociationID-toJSON}{\code{AssociationID$toJSON()}}
\item \href{#method-AssociationID-fromJSON}{\code{AssociationID$fromJSON()}}
\item \href{#method-AssociationID-clone}{\code{AssociationID$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssociationID-new"></a>}}
\if{latex}{\out{\hypertarget{method-AssociationID-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssociationID$new(
  columnNames = NULL,
  requiredInPredictionRequests = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columnNames}}{Indicates the columns that will be used to associate actuals with predictions.}

\item{\code{requiredInPredictionRequests}}{Indicates whether the association ID is required in prediction requests.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssociationID-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AssociationID-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssociationID$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssociationID-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AssociationID-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssociationID$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssociationID-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AssociationID-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssociationID$fromJSON(AssociationIDJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AssociationIDJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssociationID-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AssociationID-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssociationID$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
