% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_id_update.R
\docType{class}
\name{AssociationIDUpdate}
\alias{AssociationIDUpdate}
\title{AssociationIDUpdate}
\format{
An \code{R6Class} generator object
}
\description{
AssociationIDUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{columnNames}}{list( character ) \link{optional} List of column names used to represent an association id. Currently, exactly one column name may be provided.}

\item{\code{requiredInPredictionRequests}}{character \link{optional} Whether or not association id values are required in each prediction request for this deployment. Note that you may not change an association ID's column names after they have been set and predictions including those columns have been made}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AssociationIDUpdate-new}{\code{AssociationIDUpdate$new()}}
\item \href{#method-AssociationIDUpdate-validate}{\code{AssociationIDUpdate$validate()}}
\item \href{#method-AssociationIDUpdate-toJSON}{\code{AssociationIDUpdate$toJSON()}}
\item \href{#method-AssociationIDUpdate-fromJSON}{\code{AssociationIDUpdate$fromJSON()}}
\item \href{#method-AssociationIDUpdate-clone}{\code{AssociationIDUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssociationIDUpdate-new"></a>}}
\if{latex}{\out{\hypertarget{method-AssociationIDUpdate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssociationIDUpdate$new(
  columnNames = NULL,
  requiredInPredictionRequests = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columnNames}}{List of column names used to represent an association id. Currently, exactly one column name may be provided.}

\item{\code{requiredInPredictionRequests}}{Whether or not association id values are required in each prediction request for this deployment. Note that you may not change an association ID's column names after they have been set and predictions including those columns have been made}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssociationIDUpdate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AssociationIDUpdate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssociationIDUpdate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssociationIDUpdate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AssociationIDUpdate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssociationIDUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssociationIDUpdate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AssociationIDUpdate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssociationIDUpdate$fromJSON(AssociationIDUpdateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AssociationIDUpdateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssociationIDUpdate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AssociationIDUpdate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssociationIDUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
