% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_log_context_object_response.R
\docType{class}
\name{AuditLogContextObjectResponse}
\alias{AuditLogContextObjectResponse}
\title{AuditLogContextObjectResponse}
\format{
An \code{R6Class} generator object
}
\description{
AuditLogContextObjectResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{orgId}}{character \link{optional} The id of the organization.}

\item{\code{orgName}}{character \link{optional} The name of the organization.}

\item{\code{projectName}}{character \link{optional} The name of the project.}

\item{\code{projectType}}{character \link{optional} The type of project.}

\item{\code{userProjectRole}}{character \link{optional} The role of the user associated with this project.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AuditLogContextObjectResponse-new}{\code{AuditLogContextObjectResponse$new()}}
\item \href{#method-AuditLogContextObjectResponse-validate}{\code{AuditLogContextObjectResponse$validate()}}
\item \href{#method-AuditLogContextObjectResponse-toJSON}{\code{AuditLogContextObjectResponse$toJSON()}}
\item \href{#method-AuditLogContextObjectResponse-fromJSON}{\code{AuditLogContextObjectResponse$fromJSON()}}
\item \href{#method-AuditLogContextObjectResponse-clone}{\code{AuditLogContextObjectResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLogContextObjectResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLogContextObjectResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLogContextObjectResponse$new(
  orgId = NULL,
  orgName = NULL,
  projectName = NULL,
  projectType = NULL,
  userProjectRole = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{orgId}}{The id of the organization.}

\item{\code{orgName}}{The name of the organization.}

\item{\code{projectName}}{The name of the project.}

\item{\code{projectType}}{The type of project.}

\item{\code{userProjectRole}}{The role of the user associated with this project.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLogContextObjectResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLogContextObjectResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLogContextObjectResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLogContextObjectResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLogContextObjectResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLogContextObjectResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLogContextObjectResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLogContextObjectResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLogContextObjectResponse$fromJSON(
  AuditLogContextObjectResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AuditLogContextObjectResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLogContextObjectResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLogContextObjectResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLogContextObjectResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
