% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_logs_retrieve_one_response.R
\docType{class}
\name{AuditLogsRetrieveOneResponse}
\alias{AuditLogsRetrieveOneResponse}
\title{AuditLogsRetrieveOneResponse}
\format{
An \code{R6Class} generator object
}
\description{
AuditLogsRetrieveOneResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{context}}{\link{AuditLogContextObjectResponse}}

\item{\code{event}}{character The record event label, e.g. 'Start Autopilot'.}

\item{\code{id}}{character The record ID.}

\item{\code{ip}}{character \link{optional} The IP address of the server where the record event happened. Will be empty if ``includeIdentifyingFields`` is True.}

\item{\code{projectId}}{character The record event project id.}

\item{\code{timestamp}}{character The record timestamp.}

\item{\code{userId}}{character The user id of the record events user.}

\item{\code{username}}{character \link{optional} The username of the record events user. Will be empty if ``includeIdentifyingFields`` is True.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AuditLogsRetrieveOneResponse$new()}}
\item \href{#method-validate}{\code{AuditLogsRetrieveOneResponse$validate()}}
\item \href{#method-toJSON}{\code{AuditLogsRetrieveOneResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{AuditLogsRetrieveOneResponse$fromJSON()}}
\item \href{#method-clone}{\code{AuditLogsRetrieveOneResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLogsRetrieveOneResponse$new(
  context = NULL,
  event = NULL,
  id = NULL,
  projectId = NULL,
  timestamp = NULL,
  userId = NULL,
  ip = NULL,
  username = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event}}{The record event label, e.g. 'Start Autopilot'.}

\item{\code{id}}{The record ID.}

\item{\code{projectId}}{The record event project id.}

\item{\code{timestamp}}{The record timestamp.}

\item{\code{userId}}{The user id of the record events user.}

\item{\code{ip}}{The IP address of the server where the record event happened. Will be empty if ``includeIdentifyingFields`` is True.}

\item{\code{username}}{The username of the record events user. Will be empty if ``includeIdentifyingFields`` is True.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLogsRetrieveOneResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLogsRetrieveOneResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLogsRetrieveOneResponse$fromJSON(
  AuditLogsRetrieveOneResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AuditLogsRetrieveOneResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLogsRetrieveOneResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
