% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_logs_retrieve_response.R
\docType{class}
\name{AuditLogsRetrieveResponse}
\alias{AuditLogsRetrieveResponse}
\title{AuditLogsRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
AuditLogsRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer The number of returned items.}

\item{\code{data}}{list( \link{AuditLogsRetrieveOneResponse} ) A json array of audit log records structured in the same form as the response :http:get:`/api/v2/eventLogs/(recordId)/`}

\item{\code{next_}}{character A URL to the next page (if null, no next page exists).}

\item{\code{previous}}{character A URL to the previous page (if null, no previous page exists).}

\item{\code{totalCount}}{integer The number of items matching the query condition.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AuditLogsRetrieveResponse-new}{\code{AuditLogsRetrieveResponse$new()}}
\item \href{#method-AuditLogsRetrieveResponse-validate}{\code{AuditLogsRetrieveResponse$validate()}}
\item \href{#method-AuditLogsRetrieveResponse-toJSON}{\code{AuditLogsRetrieveResponse$toJSON()}}
\item \href{#method-AuditLogsRetrieveResponse-fromJSON}{\code{AuditLogsRetrieveResponse$fromJSON()}}
\item \href{#method-AuditLogsRetrieveResponse-clone}{\code{AuditLogsRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLogsRetrieveResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLogsRetrieveResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLogsRetrieveResponse$new(
  count = NULL,
  data = NULL,
  next_ = NULL,
  previous = NULL,
  totalCount = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{The number of returned items.}

\item{\code{data}}{A json array of audit log records structured in the same form as the response :http:get:`/api/v2/eventLogs/(recordId)/`}

\item{\code{next_}}{A URL to the next page (if null, no next page exists).}

\item{\code{previous}}{A URL to the previous page (if null, no previous page exists).}

\item{\code{totalCount}}{The number of items matching the query condition.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLogsRetrieveResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLogsRetrieveResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLogsRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLogsRetrieveResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLogsRetrieveResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLogsRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLogsRetrieveResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLogsRetrieveResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLogsRetrieveResponse$fromJSON(
  AuditLogsRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AuditLogsRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLogsRetrieveResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLogsRetrieveResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLogsRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
