% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_doc_create.R
\docType{class}
\name{AutomatedDocCreate}
\alias{AutomatedDocCreate}
\title{AutomatedDocCreate}
\format{
An \code{R6Class} generator object
}
\description{
AutomatedDocCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{documentType}}{character Type of the automated document you want to generate.}

\item{\code{documentTypeSpecificParameters}}{\link{OneOfAutomatedDocCreateDeploymentReport} \link{optional} Set parameters unique for a specific document type. Currently, only these document types can have document-specific parameters: \link{'DEPLOYMENT_REPORT'}}

\item{\code{entityId}}{character ID of the entity to generate the document for. It can be a model ID, a project ID.}

\item{\code{locale}}{character \link{optional} Localization of the document. Defaults to EN_US.}

\item{\code{outputFormat}}{character Format to generate the document in.}

\item{\code{templateId}}{character \link{optional} Template ID to use for the document outline. Defaults to standard Datarobot template.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AutomatedDocCreate-new}{\code{AutomatedDocCreate$new()}}
\item \href{#method-AutomatedDocCreate-validate}{\code{AutomatedDocCreate$validate()}}
\item \href{#method-AutomatedDocCreate-toJSON}{\code{AutomatedDocCreate$toJSON()}}
\item \href{#method-AutomatedDocCreate-fromJSON}{\code{AutomatedDocCreate$fromJSON()}}
\item \href{#method-AutomatedDocCreate-clone}{\code{AutomatedDocCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutomatedDocCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-AutomatedDocCreate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedDocCreate$new(
  documentType = NULL,
  entityId = NULL,
  outputFormat = NULL,
  documentTypeSpecificParameters = NULL,
  locale = NULL,
  templateId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{documentType}}{Type of the automated document you want to generate.}

\item{\code{entityId}}{ID of the entity to generate the document for. It can be a model ID, a project ID.}

\item{\code{outputFormat}}{Format to generate the document in.}

\item{\code{documentTypeSpecificParameters}}{Set parameters unique for a specific document type. Currently, only these document types can have document-specific parameters: \link{'DEPLOYMENT_REPORT'}}

\item{\code{locale}}{Localization of the document. Defaults to EN_US.}

\item{\code{templateId}}{Template ID to use for the document outline. Defaults to standard Datarobot template.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutomatedDocCreate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AutomatedDocCreate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedDocCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutomatedDocCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AutomatedDocCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedDocCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutomatedDocCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AutomatedDocCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedDocCreate$fromJSON(AutomatedDocCreateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AutomatedDocCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutomatedDocCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AutomatedDocCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedDocCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
