% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_doc_create_deployment_report.R
\docType{class}
\name{AutomatedDocCreateDeploymentReport}
\alias{AutomatedDocCreateDeploymentReport}
\title{AutomatedDocCreateDeploymentReport}
\format{
An \code{R6Class} generator object
}
\description{
AutomatedDocCreateDeploymentReport Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bucketSize}}{character \link{optional} You can regulate the size of the buckets in charts. One bucket reflects some duration period and you can set the exact duration with this parameter. Bucket size gets defaulted to either a month, a week, or a day, based on the time range of the report. We use `ISO 8601 <https://www.digi.com/resources/documentation/digidocs/90001437-13/reference/r_iso_8601_duration_format.htm>`_ duration format to specify values.}

\item{\code{end}}{character \link{optional} RFC3339 datetime. End of time period to retrieve the features drift for. If not set, defaults to the next top of the hour from now.Note: this field only accepts top of the hour RFC3339 datetime strings For example: ``2019-08-01T00:00:00Z``.}

\item{\code{modelId}}{character \link{optional} Provide a model ID to generate a report for a previously deployed model.}

\item{\code{start}}{character \link{optional} RFC3339 datetime. Start of time period to retrieve the features drift for. If not set, defaults to 1 week from the end of the period.Note: this field only accepts top of the hour RFC3339 datetime strings For example: ``2019-08-01T00:00:00Z``.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AutomatedDocCreateDeploymentReport$new()}}
\item \href{#method-validate}{\code{AutomatedDocCreateDeploymentReport$validate()}}
\item \href{#method-toJSON}{\code{AutomatedDocCreateDeploymentReport$toJSON()}}
\item \href{#method-fromJSON}{\code{AutomatedDocCreateDeploymentReport$fromJSON()}}
\item \href{#method-clone}{\code{AutomatedDocCreateDeploymentReport$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedDocCreateDeploymentReport$new(
  bucketSize = NULL,
  end = NULL,
  modelId = NULL,
  start = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bucketSize}}{You can regulate the size of the buckets in charts. One bucket reflects some duration period and you can set the exact duration with this parameter. Bucket size gets defaulted to either a month, a week, or a day, based on the time range of the report. We use `ISO 8601 <https://www.digi.com/resources/documentation/digidocs/90001437-13/reference/r_iso_8601_duration_format.htm>`_ duration format to specify values.}

\item{\code{end}}{RFC3339 datetime. End of time period to retrieve the features drift for. If not set, defaults to the next top of the hour from now.Note: this field only accepts top of the hour RFC3339 datetime strings For example: ``2019-08-01T00:00:00Z``.}

\item{\code{modelId}}{Provide a model ID to generate a report for a previously deployed model.}

\item{\code{start}}{RFC3339 datetime. Start of time period to retrieve the features drift for. If not set, defaults to 1 week from the end of the period.Note: this field only accepts top of the hour RFC3339 datetime strings For example: ``2019-08-01T00:00:00Z``.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedDocCreateDeploymentReport$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedDocCreateDeploymentReport$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedDocCreateDeploymentReport$fromJSON(
  AutomatedDocCreateDeploymentReportJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AutomatedDocCreateDeploymentReportJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedDocCreateDeploymentReport$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
