% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_doc_options.R
\docType{class}
\name{AutomatedDocOptions}
\alias{AutomatedDocOptions}
\title{AutomatedDocOptions}
\format{
An \code{R6Class} generator object
}
\description{
AutomatedDocOptions Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{documentType}}{character Type of document available for generation.}

\item{\code{locale}}{character Locale available for the document generation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AutomatedDocOptions-new}{\code{AutomatedDocOptions$new()}}
\item \href{#method-AutomatedDocOptions-validate}{\code{AutomatedDocOptions$validate()}}
\item \href{#method-AutomatedDocOptions-toJSON}{\code{AutomatedDocOptions$toJSON()}}
\item \href{#method-AutomatedDocOptions-fromJSON}{\code{AutomatedDocOptions$fromJSON()}}
\item \href{#method-AutomatedDocOptions-clone}{\code{AutomatedDocOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutomatedDocOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-AutomatedDocOptions-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedDocOptions$new(
  documentType = NULL,
  locale = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{documentType}}{Type of document available for generation.}

\item{\code{locale}}{Locale available for the document generation.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutomatedDocOptions-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AutomatedDocOptions-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedDocOptions$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutomatedDocOptions-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AutomatedDocOptions-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedDocOptions$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutomatedDocOptions-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AutomatedDocOptions-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedDocOptions$fromJSON(AutomatedDocOptionsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AutomatedDocOptionsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutomatedDocOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AutomatedDocOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedDocOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
