% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autopilot.R
\docType{class}
\name{Autopilot}
\alias{Autopilot}
\title{Autopilot}
\format{
An \code{R6Class} generator object
}
\description{
Autopilot Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{command}}{character If `start`, will unpause the autopilot and run queued jobs if workers are available.  If `stop`, will pause the autopilot so no new jobs will be started.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Autopilot-new}{\code{Autopilot$new()}}
\item \href{#method-Autopilot-validate}{\code{Autopilot$validate()}}
\item \href{#method-Autopilot-toJSON}{\code{Autopilot$toJSON()}}
\item \href{#method-Autopilot-fromJSON}{\code{Autopilot$fromJSON()}}
\item \href{#method-Autopilot-clone}{\code{Autopilot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Autopilot-new"></a>}}
\if{latex}{\out{\hypertarget{method-Autopilot-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Autopilot$new(command = NULL, validateParams = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{command}}{If `start`, will unpause the autopilot and run queued jobs if workers are available.  If `stop`, will pause the autopilot so no new jobs will be started.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Autopilot-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Autopilot-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Autopilot$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Autopilot-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Autopilot-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Autopilot$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Autopilot-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Autopilot-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Autopilot$fromJSON(AutopilotJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AutopilotJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Autopilot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Autopilot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Autopilot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
