% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autopilot_options.R
\docType{class}
\name{AutopilotOptions}
\alias{AutopilotOptions}
\title{AutopilotOptions}
\format{
An \code{R6Class} generator object
}
\description{
AutopilotOptions Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{blendBestModels}}{character \link{optional} Blend best models during Autopilot run. This option is not supported in SHAP-only mode.}

\item{\code{mode}}{character \link{optional} The autopilot mode.}

\item{\code{runLeakageRemovedFeatureList}}{character \link{optional} Run Autopilot on Leakage Removed feature list (if exists).}

\item{\code{scoringCodeOnly}}{character \link{optional} Keep only models that can be converted to scorable java code during Autopilot run.}

\item{\code{shapOnlyMode}}{character \link{optional} Include only models with SHAP value support.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AutopilotOptions-new}{\code{AutopilotOptions$new()}}
\item \href{#method-AutopilotOptions-validate}{\code{AutopilotOptions$validate()}}
\item \href{#method-AutopilotOptions-toJSON}{\code{AutopilotOptions$toJSON()}}
\item \href{#method-AutopilotOptions-fromJSON}{\code{AutopilotOptions$fromJSON()}}
\item \href{#method-AutopilotOptions-clone}{\code{AutopilotOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutopilotOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-AutopilotOptions-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutopilotOptions$new(
  blendBestModels = NULL,
  mode = NULL,
  runLeakageRemovedFeatureList = NULL,
  scoringCodeOnly = NULL,
  shapOnlyMode = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blendBestModels}}{Blend best models during Autopilot run. This option is not supported in SHAP-only mode.}

\item{\code{mode}}{The autopilot mode.}

\item{\code{runLeakageRemovedFeatureList}}{Run Autopilot on Leakage Removed feature list (if exists).}

\item{\code{scoringCodeOnly}}{Keep only models that can be converted to scorable java code during Autopilot run.}

\item{\code{shapOnlyMode}}{Include only models with SHAP value support.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutopilotOptions-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AutopilotOptions-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutopilotOptions$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutopilotOptions-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AutopilotOptions-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutopilotOptions$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutopilotOptions-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AutopilotOptions-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutopilotOptions$fromJSON(AutopilotOptionsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AutopilotOptionsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutopilotOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AutopilotOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutopilotOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
