% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autopilot_start.R
\docType{class}
\name{AutopilotStart}
\alias{AutopilotStart}
\title{AutopilotStart}
\format{
An \code{R6Class} generator object
}
\description{
AutopilotStart Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{autopilotClusterList}}{list( integer ) \link{optional} Optional. A list of integers where each value will be used as the number of clusters in Autopilot model(s) for unsupervised clustering projects. Cannot be specified unless unsupervisedMode is true and unsupervisedType is set to 'clustering'.}

\item{\code{blendBestModels}}{character \link{optional} Blend best models during Autopilot run. This option is not supported in SHAP-only mode or for multilabel projects.}

\item{\code{considerBlendersInRecommendation}}{character \link{optional} Include blenders when selecting a model to prepare for deployment in an Autopilot Run. This option is not supported in SHAP-only mode or for multilabel projects.}

\item{\code{featurelistId}}{character The ID of a featurelist that should be used for autopilot.}

\item{\code{mode}}{character \link{optional} The autopilot mode.}

\item{\code{prepareModelForDeployment}}{character \link{optional} Prepare model for deployment during Autopilot run. The preparation includes creating reduced feature list models, retraining best model on higher sample size, computing insights and assigning \"RECOMMENDED FOR DEPLOYMENT\" label.}

\item{\code{runLeakageRemovedFeatureList}}{character \link{optional} Run Autopilot on Leakage Removed feature list (if exists).}

\item{\code{scoringCodeOnly}}{character \link{optional} Keep only models that can be converted to scorable java code during Autopilot run.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AutopilotStart-new}{\code{AutopilotStart$new()}}
\item \href{#method-AutopilotStart-validate}{\code{AutopilotStart$validate()}}
\item \href{#method-AutopilotStart-toJSON}{\code{AutopilotStart$toJSON()}}
\item \href{#method-AutopilotStart-fromJSON}{\code{AutopilotStart$fromJSON()}}
\item \href{#method-AutopilotStart-clone}{\code{AutopilotStart$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutopilotStart-new"></a>}}
\if{latex}{\out{\hypertarget{method-AutopilotStart-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutopilotStart$new(
  featurelistId = NULL,
  autopilotClusterList = NULL,
  blendBestModels = NULL,
  considerBlendersInRecommendation = NULL,
  mode = "auto",
  prepareModelForDeployment = NULL,
  runLeakageRemovedFeatureList = NULL,
  scoringCodeOnly = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{featurelistId}}{The ID of a featurelist that should be used for autopilot.}

\item{\code{autopilotClusterList}}{Optional. A list of integers where each value will be used as the number of clusters in Autopilot model(s) for unsupervised clustering projects. Cannot be specified unless unsupervisedMode is true and unsupervisedType is set to 'clustering'.}

\item{\code{blendBestModels}}{Blend best models during Autopilot run. This option is not supported in SHAP-only mode or for multilabel projects.}

\item{\code{considerBlendersInRecommendation}}{Include blenders when selecting a model to prepare for deployment in an Autopilot Run. This option is not supported in SHAP-only mode or for multilabel projects.}

\item{\code{mode}}{The autopilot mode.}

\item{\code{prepareModelForDeployment}}{Prepare model for deployment during Autopilot run. The preparation includes creating reduced feature list models, retraining best model on higher sample size, computing insights and assigning \"RECOMMENDED FOR DEPLOYMENT\" label.}

\item{\code{runLeakageRemovedFeatureList}}{Run Autopilot on Leakage Removed feature list (if exists).}

\item{\code{scoringCodeOnly}}{Keep only models that can be converted to scorable java code during Autopilot run.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutopilotStart-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AutopilotStart-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutopilotStart$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutopilotStart-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AutopilotStart-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutopilotStart$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutopilotStart-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AutopilotStart-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutopilotStart$fromJSON(AutopilotStartJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AutopilotStartJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutopilotStart-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AutopilotStart-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutopilotStart$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
