% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_model_metrics_field.R
\docType{class}
\name{AverageModelMetricsField}
\alias{AverageModelMetricsField}
\title{AverageModelMetricsField}
\format{
An \code{R6Class} generator object
}
\description{
AverageModelMetricsField Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metrics}}{list( \link{AverageModelMetricsValues} ) Average model metrics for the given thresholds.}

\item{\code{source}}{character Chart source.}

\item{\code{thresholds}}{list( numeric ) Threshold values for which model metrics are available.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AverageModelMetricsField-new}{\code{AverageModelMetricsField$new()}}
\item \href{#method-AverageModelMetricsField-validate}{\code{AverageModelMetricsField$validate()}}
\item \href{#method-AverageModelMetricsField-toJSON}{\code{AverageModelMetricsField$toJSON()}}
\item \href{#method-AverageModelMetricsField-fromJSON}{\code{AverageModelMetricsField$fromJSON()}}
\item \href{#method-AverageModelMetricsField-clone}{\code{AverageModelMetricsField$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AverageModelMetricsField-new"></a>}}
\if{latex}{\out{\hypertarget{method-AverageModelMetricsField-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AverageModelMetricsField$new(
  metrics = NULL,
  source = NULL,
  thresholds = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metrics}}{Average model metrics for the given thresholds.}

\item{\code{source}}{Chart source.}

\item{\code{thresholds}}{Threshold values for which model metrics are available.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AverageModelMetricsField-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AverageModelMetricsField-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AverageModelMetricsField$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AverageModelMetricsField-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AverageModelMetricsField-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AverageModelMetricsField$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AverageModelMetricsField-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AverageModelMetricsField-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AverageModelMetricsField$fromJSON(
  AverageModelMetricsFieldJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AverageModelMetricsFieldJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AverageModelMetricsField-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AverageModelMetricsField-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AverageModelMetricsField$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
