% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_output_adaptor.R
\docType{class}
\name{AzureOutputAdaptor}
\alias{AzureOutputAdaptor}
\title{AzureOutputAdaptor}
\format{
An \code{R6Class} generator object
}
\description{
AzureOutputAdaptor Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{credentialId}}{\link{OneOfstringstring} \link{optional} Either the populated value of the field or \link{redacted} due to permission settings}

\item{\code{format}}{character \link{optional} Type of output file format}

\item{\code{partitionColumns}}{list( character ) \link{optional} For Parquet directory-scoring only. The column names of the intake data of which to partition the dataset. Columns are partitioned in the order they are given. At least one value is required if scoring to a directory (meaning the output url ends with a slash (\"/\").}

\item{\code{type}}{character Type name for this output type}

\item{\code{url}}{character URL for the file or directory}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AzureOutputAdaptor-new}{\code{AzureOutputAdaptor$new()}}
\item \href{#method-AzureOutputAdaptor-validate}{\code{AzureOutputAdaptor$validate()}}
\item \href{#method-AzureOutputAdaptor-toJSON}{\code{AzureOutputAdaptor$toJSON()}}
\item \href{#method-AzureOutputAdaptor-fromJSON}{\code{AzureOutputAdaptor$fromJSON()}}
\item \href{#method-AzureOutputAdaptor-clone}{\code{AzureOutputAdaptor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AzureOutputAdaptor-new"></a>}}
\if{latex}{\out{\hypertarget{method-AzureOutputAdaptor-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AzureOutputAdaptor$new(
  type = NULL,
  url = NULL,
  credentialId = NULL,
  format = "csv",
  partitionColumns = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Type name for this output type}

\item{\code{url}}{URL for the file or directory}

\item{\code{credentialId}}{Either the populated value of the field or \link{redacted} due to permission settings}

\item{\code{format}}{Type of output file format}

\item{\code{partitionColumns}}{For Parquet directory-scoring only. The column names of the intake data of which to partition the dataset. Columns are partitioned in the order they are given. At least one value is required if scoring to a directory (meaning the output url ends with a slash (\"/\").}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AzureOutputAdaptor-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AzureOutputAdaptor-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AzureOutputAdaptor$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AzureOutputAdaptor-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AzureOutputAdaptor-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AzureOutputAdaptor$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AzureOutputAdaptor-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AzureOutputAdaptor-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AzureOutputAdaptor$fromJSON(AzureOutputAdaptorJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AzureOutputAdaptorJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AzureOutputAdaptor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AzureOutputAdaptor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AzureOutputAdaptor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
