% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_new_method_for_open_api.R
\docType{class}
\name{BacktestNewMethodForOpenApi}
\alias{BacktestNewMethodForOpenApi}
\title{BacktestNewMethodForOpenApi}
\format{
An \code{R6Class} generator object
}
\description{
BacktestNewMethodForOpenApi Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{index}}{integer The index from zero of the backtest.}

\item{\code{primaryTrainingEndDate}}{character A datetime string representing the end date of the primary training data for this backtest.}

\item{\code{primaryTrainingStartDate}}{character A datetime string representing the start date of the primary training data for this backtest.}

\item{\code{validationEndDate}}{character A datetime string representing the end date of the validation data for this backtest.}

\item{\code{validationStartDate}}{character A datetime string representing the start date of the validation data for this backtest.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BacktestNewMethodForOpenApi-new}{\code{BacktestNewMethodForOpenApi$new()}}
\item \href{#method-BacktestNewMethodForOpenApi-validate}{\code{BacktestNewMethodForOpenApi$validate()}}
\item \href{#method-BacktestNewMethodForOpenApi-toJSON}{\code{BacktestNewMethodForOpenApi$toJSON()}}
\item \href{#method-BacktestNewMethodForOpenApi-fromJSON}{\code{BacktestNewMethodForOpenApi$fromJSON()}}
\item \href{#method-BacktestNewMethodForOpenApi-clone}{\code{BacktestNewMethodForOpenApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BacktestNewMethodForOpenApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-BacktestNewMethodForOpenApi-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestNewMethodForOpenApi$new(
  index = NULL,
  primaryTrainingEndDate = NULL,
  primaryTrainingStartDate = NULL,
  validationEndDate = NULL,
  validationStartDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{The index from zero of the backtest.}

\item{\code{primaryTrainingEndDate}}{A datetime string representing the end date of the primary training data for this backtest.}

\item{\code{primaryTrainingStartDate}}{A datetime string representing the start date of the primary training data for this backtest.}

\item{\code{validationEndDate}}{A datetime string representing the end date of the validation data for this backtest.}

\item{\code{validationStartDate}}{A datetime string representing the start date of the validation data for this backtest.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BacktestNewMethodForOpenApi-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BacktestNewMethodForOpenApi-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestNewMethodForOpenApi$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BacktestNewMethodForOpenApi-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BacktestNewMethodForOpenApi-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestNewMethodForOpenApi$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BacktestNewMethodForOpenApi-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BacktestNewMethodForOpenApi-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestNewMethodForOpenApi$fromJSON(
  BacktestNewMethodForOpenApiJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BacktestNewMethodForOpenApiJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BacktestNewMethodForOpenApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BacktestNewMethodForOpenApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestNewMethodForOpenApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
