% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_response.R
\docType{class}
\name{BacktestResponse}
\alias{BacktestResponse}
\title{BacktestResponse}
\format{
An \code{R6Class} generator object
}
\description{
BacktestResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{availableTrainingDuration}}{character The duration of the available training data for this backtest.}

\item{\code{availableTrainingEndDate}}{character The end date of the available training data for this backtest.}

\item{\code{availableTrainingStartDate}}{character The start date of the available training data for this backtest.}

\item{\code{gapDuration}}{character The duration of the gap between the training and the validation scoring data for this backtest.}

\item{\code{gapEndDate}}{character The end date of the gap between the training and validation scoring data for this backtest.}

\item{\code{gapStartDate}}{character The start date of the gap between the training and validation scoring data for this backtest.}

\item{\code{index}}{integer The index from zero of this backtest.}

\item{\code{primaryTrainingDuration}}{character The duration of the primary training data for this backtest.}

\item{\code{primaryTrainingEndDate}}{character The end date of the primary training data for this backtest.}

\item{\code{primaryTrainingStartDate}}{character The start date of the primary training data for this backtest.}

\item{\code{validationDuration}}{character The duration of the validation scoring data for this backtest.}

\item{\code{validationEndDate}}{character The end date of the validation scoring data for this backtest.}

\item{\code{validationStartDate}}{character The start date of the validation scoring data for this backtest.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BacktestResponse-new}{\code{BacktestResponse$new()}}
\item \href{#method-BacktestResponse-validate}{\code{BacktestResponse$validate()}}
\item \href{#method-BacktestResponse-toJSON}{\code{BacktestResponse$toJSON()}}
\item \href{#method-BacktestResponse-fromJSON}{\code{BacktestResponse$fromJSON()}}
\item \href{#method-BacktestResponse-clone}{\code{BacktestResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BacktestResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-BacktestResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestResponse$new(
  availableTrainingDuration = NULL,
  availableTrainingEndDate = NULL,
  availableTrainingStartDate = NULL,
  gapDuration = NULL,
  gapEndDate = NULL,
  gapStartDate = NULL,
  index = NULL,
  primaryTrainingDuration = NULL,
  primaryTrainingEndDate = NULL,
  primaryTrainingStartDate = NULL,
  validationDuration = NULL,
  validationEndDate = NULL,
  validationStartDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{availableTrainingDuration}}{The duration of the available training data for this backtest.}

\item{\code{availableTrainingEndDate}}{The end date of the available training data for this backtest.}

\item{\code{availableTrainingStartDate}}{The start date of the available training data for this backtest.}

\item{\code{gapDuration}}{The duration of the gap between the training and the validation scoring data for this backtest.}

\item{\code{gapEndDate}}{The end date of the gap between the training and validation scoring data for this backtest.}

\item{\code{gapStartDate}}{The start date of the gap between the training and validation scoring data for this backtest.}

\item{\code{index}}{The index from zero of this backtest.}

\item{\code{primaryTrainingDuration}}{The duration of the primary training data for this backtest.}

\item{\code{primaryTrainingEndDate}}{The end date of the primary training data for this backtest.}

\item{\code{primaryTrainingStartDate}}{The start date of the primary training data for this backtest.}

\item{\code{validationDuration}}{The duration of the validation scoring data for this backtest.}

\item{\code{validationEndDate}}{The end date of the validation scoring data for this backtest.}

\item{\code{validationStartDate}}{The start date of the validation scoring data for this backtest.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BacktestResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BacktestResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BacktestResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BacktestResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BacktestResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BacktestResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestResponse$fromJSON(BacktestResponseJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BacktestResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BacktestResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BacktestResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
