% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtests.R
\docType{class}
\name{Backtests}
\alias{Backtests}
\title{Backtests}
\format{
An \code{R6Class} generator object
}
\description{
Backtests Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{validationEndDate}}{character The end date of the validation scoring data for this backtest.}

\item{\code{validationStartDate}}{character The start date of the validation scoring data for this backtest.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Backtests-new}{\code{Backtests$new()}}
\item \href{#method-Backtests-validate}{\code{Backtests$validate()}}
\item \href{#method-Backtests-toJSON}{\code{Backtests$toJSON()}}
\item \href{#method-Backtests-fromJSON}{\code{Backtests$fromJSON()}}
\item \href{#method-Backtests-clone}{\code{Backtests$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Backtests-new"></a>}}
\if{latex}{\out{\hypertarget{method-Backtests-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Backtests$new(
  validationEndDate = NULL,
  validationStartDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{validationEndDate}}{The end date of the validation scoring data for this backtest.}

\item{\code{validationStartDate}}{The start date of the validation scoring data for this backtest.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Backtests-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Backtests-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Backtests$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Backtests-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Backtests-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Backtests$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Backtests-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Backtests-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Backtests$fromJSON(BacktestsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BacktestsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Backtests-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Backtests-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Backtests$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
