% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_stability_plot_data.R
\docType{class}
\name{BacktestStabilityPlotData}
\alias{BacktestStabilityPlotData}
\title{BacktestStabilityPlotData}
\format{
An \code{R6Class} generator object
}
\description{
BacktestStabilityPlotData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{backtestIndex}}{integer An integer representing the index of the backtest, starting from 0. For holdout, this field will be null.}

\item{\code{partition}}{character Identifier of the partition. Can either identify a specific backtest (\"backtest0\", \"backtest1\", ...) or the holdout set (\"holdout\").}

\item{\code{score}}{numeric Score for this partition. Can be null if the score is unavailable for this partition (e.g. holdout is locked or backtesting has not been run yet).}

\item{\code{scoringEndDate}}{character End date of the subset used for scoring.}

\item{\code{scoringStartDate}}{character Start date of the subset used for scoring.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BacktestStabilityPlotData$new()}}
\item \href{#method-validate}{\code{BacktestStabilityPlotData$validate()}}
\item \href{#method-toJSON}{\code{BacktestStabilityPlotData$toJSON()}}
\item \href{#method-fromJSON}{\code{BacktestStabilityPlotData$fromJSON()}}
\item \href{#method-clone}{\code{BacktestStabilityPlotData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestStabilityPlotData$new(
  backtestIndex = NULL,
  partition = NULL,
  score = NULL,
  scoringEndDate = NULL,
  scoringStartDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{backtestIndex}}{An integer representing the index of the backtest, starting from 0. For holdout, this field will be null.}

\item{\code{partition}}{Identifier of the partition. Can either identify a specific backtest (\"backtest0\", \"backtest1\", ...) or the holdout set (\"holdout\").}

\item{\code{score}}{Score for this partition. Can be null if the score is unavailable for this partition (e.g. holdout is locked or backtesting has not been run yet).}

\item{\code{scoringEndDate}}{End date of the subset used for scoring.}

\item{\code{scoringStartDate}}{Start date of the subset used for scoring.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestStabilityPlotData$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestStabilityPlotData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestStabilityPlotData$fromJSON(
  BacktestStabilityPlotDataJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BacktestStabilityPlotDataJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestStabilityPlotData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
