% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_stability_plot_response.R
\docType{class}
\name{BacktestStabilityPlotResponse}
\alias{BacktestStabilityPlotResponse}
\title{BacktestStabilityPlotResponse}
\format{
An \code{R6Class} generator object
}
\description{
BacktestStabilityPlotResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{backtestPlotData}}{list( \link{BacktestStabilityPlotData} ) An array of objects containing the details of the scores for each partition defined for the project.}

\item{\code{endDate}}{character End date of the project dataset.}

\item{\code{metricName}}{character Name of the metric used to compute the scores.}

\item{\code{startDate}}{character Start date of the project dataset.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BacktestStabilityPlotResponse$new()}}
\item \href{#method-validate}{\code{BacktestStabilityPlotResponse$validate()}}
\item \href{#method-toJSON}{\code{BacktestStabilityPlotResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{BacktestStabilityPlotResponse$fromJSON()}}
\item \href{#method-clone}{\code{BacktestStabilityPlotResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestStabilityPlotResponse$new(
  backtestPlotData = NULL,
  endDate = NULL,
  metricName = NULL,
  startDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{backtestPlotData}}{An array of objects containing the details of the scores for each partition defined for the project.}

\item{\code{endDate}}{End date of the project dataset.}

\item{\code{metricName}}{Name of the metric used to compute the scores.}

\item{\code{startDate}}{Start date of the project dataset.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestStabilityPlotResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestStabilityPlotResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestStabilityPlotResponse$fromJSON(
  BacktestStabilityPlotResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BacktestStabilityPlotResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestStabilityPlotResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
