% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_status_response.R
\docType{class}
\name{BacktestStatusResponse}
\alias{BacktestStatusResponse}
\title{BacktestStatusResponse}
\format{
An \code{R6Class} generator object
}
\description{
BacktestStatusResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{index}}{integer the index of the fold}

\item{\code{score}}{numeric the score of the model for this backtesting fold, if computed}

\item{\code{status}}{character the status of the current backtest model job}

\item{\code{trainingDuration}}{character the duration of the data used to train the model for this backtesting fold}

\item{\code{trainingEndDate}}{character the end date of the training for this backtesting fold}

\item{\code{trainingRowCount}}{integer the number of rows used to train the model for this backtesting fold}

\item{\code{trainingStartDate}}{character the start date of the training for this backtesting fold}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BacktestStatusResponse-new}{\code{BacktestStatusResponse$new()}}
\item \href{#method-BacktestStatusResponse-validate}{\code{BacktestStatusResponse$validate()}}
\item \href{#method-BacktestStatusResponse-toJSON}{\code{BacktestStatusResponse$toJSON()}}
\item \href{#method-BacktestStatusResponse-fromJSON}{\code{BacktestStatusResponse$fromJSON()}}
\item \href{#method-BacktestStatusResponse-clone}{\code{BacktestStatusResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BacktestStatusResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-BacktestStatusResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestStatusResponse$new(
  index = NULL,
  score = NULL,
  status = NULL,
  trainingDuration = NULL,
  trainingEndDate = NULL,
  trainingRowCount = NULL,
  trainingStartDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{the index of the fold}

\item{\code{score}}{the score of the model for this backtesting fold, if computed}

\item{\code{status}}{the status of the current backtest model job}

\item{\code{trainingDuration}}{the duration of the data used to train the model for this backtesting fold}

\item{\code{trainingEndDate}}{the end date of the training for this backtesting fold}

\item{\code{trainingRowCount}}{the number of rows used to train the model for this backtesting fold}

\item{\code{trainingStartDate}}{the start date of the training for this backtesting fold}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BacktestStatusResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BacktestStatusResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestStatusResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BacktestStatusResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BacktestStatusResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestStatusResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BacktestStatusResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BacktestStatusResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestStatusResponse$fromJSON(
  BacktestStatusResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BacktestStatusResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BacktestStatusResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BacktestStatusResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BacktestStatusResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
