% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_blender_builder_create.R
\docType{class}
\name{BaseBlenderBuilderCreate}
\alias{BaseBlenderBuilderCreate}
\title{BaseBlenderBuilderCreate}
\format{
An \code{R6Class} generator object
}
\description{
BaseBlenderBuilderCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{blendMethods}}{list( character ) The blender method names, either GLM, ENET, AVERAGE, LGBM, TF, RF, PLS, MAE, MAEL1, MAX, or MIN.}

\item{\code{correlationWeight}}{numeric \link{optional} The weighting between the correlation score and accuracy score}

\item{\code{heuristic}}{character The blender builder heuristic name, either VALIDATION or CORRELATION.}

\item{\code{maxModels}}{integer The number of models to blend}

\item{\code{models}}{list( character ) \link{optional} The leaderboard IDs for the models to be considered for blending.}

\item{\code{topN}}{integer \link{optional} Use the Top N models in the leaderboard instead of a list of models.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BaseBlenderBuilderCreate$new()}}
\item \href{#method-validate}{\code{BaseBlenderBuilderCreate$validate()}}
\item \href{#method-toJSON}{\code{BaseBlenderBuilderCreate$toJSON()}}
\item \href{#method-fromJSON}{\code{BaseBlenderBuilderCreate$fromJSON()}}
\item \href{#method-clone}{\code{BaseBlenderBuilderCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseBlenderBuilderCreate$new(
  blendMethods = NULL,
  heuristic = NULL,
  maxModels = NULL,
  correlationWeight = NULL,
  models = NULL,
  topN = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blendMethods}}{The blender method names, either GLM, ENET, AVERAGE, LGBM, TF, RF, PLS, MAE, MAEL1, MAX, or MIN.}

\item{\code{heuristic}}{The blender builder heuristic name, either VALIDATION or CORRELATION.}

\item{\code{maxModels}}{The number of models to blend}

\item{\code{correlationWeight}}{The weighting between the correlation score and accuracy score}

\item{\code{models}}{The leaderboard IDs for the models to be considered for blending.}

\item{\code{topN}}{Use the Top N models in the leaderboard instead of a list of models.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseBlenderBuilderCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseBlenderBuilderCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseBlenderBuilderCreate$fromJSON(
  BaseBlenderBuilderCreateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BaseBlenderBuilderCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseBlenderBuilderCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
