% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_dependency_build_metadata_response.R
\docType{class}
\name{BaseDependencyBuildMetadataResponse}
\alias{BaseDependencyBuildMetadataResponse}
\title{BaseDependencyBuildMetadataResponse}
\format{
An \code{R6Class} generator object
}
\description{
BaseDependencyBuildMetadataResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{buildEnd}}{character The ISO-8601 encoded time when this build completed.}

\item{\code{buildLogLocation}}{character The URL to download the build logs from this build.}

\item{\code{buildStart}}{character The ISO-8601 encoded time when this build started.}

\item{\code{buildStatus}}{character The current status of the dependency build.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaseDependencyBuildMetadataResponse-new}{\code{BaseDependencyBuildMetadataResponse$new()}}
\item \href{#method-BaseDependencyBuildMetadataResponse-validate}{\code{BaseDependencyBuildMetadataResponse$validate()}}
\item \href{#method-BaseDependencyBuildMetadataResponse-toJSON}{\code{BaseDependencyBuildMetadataResponse$toJSON()}}
\item \href{#method-BaseDependencyBuildMetadataResponse-fromJSON}{\code{BaseDependencyBuildMetadataResponse$fromJSON()}}
\item \href{#method-BaseDependencyBuildMetadataResponse-clone}{\code{BaseDependencyBuildMetadataResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseDependencyBuildMetadataResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-BaseDependencyBuildMetadataResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseDependencyBuildMetadataResponse$new(
  buildEnd = NULL,
  buildLogLocation = NULL,
  buildStart = NULL,
  buildStatus = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{buildEnd}}{The ISO-8601 encoded time when this build completed.}

\item{\code{buildLogLocation}}{The URL to download the build logs from this build.}

\item{\code{buildStart}}{The ISO-8601 encoded time when this build started.}

\item{\code{buildStatus}}{The current status of the dependency build.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseDependencyBuildMetadataResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BaseDependencyBuildMetadataResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseDependencyBuildMetadataResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseDependencyBuildMetadataResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BaseDependencyBuildMetadataResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseDependencyBuildMetadataResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseDependencyBuildMetadataResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BaseDependencyBuildMetadataResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseDependencyBuildMetadataResponse$fromJSON(
  BaseDependencyBuildMetadataResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BaseDependencyBuildMetadataResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseDependencyBuildMetadataResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaseDependencyBuildMetadataResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseDependencyBuildMetadataResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
