% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_dataset_with_size_response.R
\docType{class}
\name{BasicDatasetWithSizeResponse}
\alias{BasicDatasetWithSizeResponse}
\title{BasicDatasetWithSizeResponse}
\format{
An \code{R6Class} generator object
}
\description{
BasicDatasetWithSizeResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{categories}}{list( character ) An array of strings describing the intended use of the dataset.}

\item{\code{createdBy}}{character Username of the user who created the dataset.}

\item{\code{creationDate}}{character The date when the dataset was created.}

\item{\code{dataPersisted}}{character If true, user is allowed to view extended data profile (which includes data statistics like min/max/median/mean, histogram, etc.) and download data. If false, download is not allowed and only the data schema (feature names and types) will be available.}

\item{\code{datasetId}}{character The ID of this dataset.}

\item{\code{datasetSize}}{integer The size of the dataset as a CSV in bytes.}

\item{\code{isDataEngineEligible}}{character Whether this dataset can be a data source of a data engine query.}

\item{\code{isLatestVersion}}{character Whether this dataset version is the latest version of this dataset.}

\item{\code{isSnapshot}}{character Whether the dataset is an immutable snapshot of data which has previously been retrieved and saved to DataRobot.}

\item{\code{name}}{character The name of this dataset in the catalog.}

\item{\code{processingState}}{character Current ingestion process state of dataset.}

\item{\code{rowCount}}{integer The number of rows in the dataset.}

\item{\code{sampleSize}}{\link{SampleSize} \link{optional}}

\item{\code{timeSeriesProperties}}{\link{TimeSeriesProperties}}

\item{\code{versionId}}{character The object ID of the catalog_version the dataset belongs to.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BasicDatasetWithSizeResponse-new}{\code{BasicDatasetWithSizeResponse$new()}}
\item \href{#method-BasicDatasetWithSizeResponse-validate}{\code{BasicDatasetWithSizeResponse$validate()}}
\item \href{#method-BasicDatasetWithSizeResponse-toJSON}{\code{BasicDatasetWithSizeResponse$toJSON()}}
\item \href{#method-BasicDatasetWithSizeResponse-fromJSON}{\code{BasicDatasetWithSizeResponse$fromJSON()}}
\item \href{#method-BasicDatasetWithSizeResponse-clone}{\code{BasicDatasetWithSizeResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicDatasetWithSizeResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-BasicDatasetWithSizeResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicDatasetWithSizeResponse$new(
  categories = NULL,
  createdBy = NULL,
  creationDate = NULL,
  dataPersisted = NULL,
  datasetId = NULL,
  datasetSize = NULL,
  isDataEngineEligible = NULL,
  isLatestVersion = NULL,
  isSnapshot = NULL,
  name = NULL,
  processingState = NULL,
  rowCount = NULL,
  timeSeriesProperties = NULL,
  versionId = NULL,
  sampleSize = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{categories}}{An array of strings describing the intended use of the dataset.}

\item{\code{createdBy}}{Username of the user who created the dataset.}

\item{\code{creationDate}}{The date when the dataset was created.}

\item{\code{dataPersisted}}{If true, user is allowed to view extended data profile (which includes data statistics like min/max/median/mean, histogram, etc.) and download data. If false, download is not allowed and only the data schema (feature names and types) will be available.}

\item{\code{datasetId}}{The ID of this dataset.}

\item{\code{datasetSize}}{The size of the dataset as a CSV in bytes.}

\item{\code{isDataEngineEligible}}{Whether this dataset can be a data source of a data engine query.}

\item{\code{isLatestVersion}}{Whether this dataset version is the latest version of this dataset.}

\item{\code{isSnapshot}}{Whether the dataset is an immutable snapshot of data which has previously been retrieved and saved to DataRobot.}

\item{\code{name}}{The name of this dataset in the catalog.}

\item{\code{processingState}}{Current ingestion process state of dataset.}

\item{\code{rowCount}}{The number of rows in the dataset.}

\item{\code{versionId}}{The object ID of the catalog_version the dataset belongs to.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicDatasetWithSizeResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BasicDatasetWithSizeResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicDatasetWithSizeResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicDatasetWithSizeResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BasicDatasetWithSizeResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicDatasetWithSizeResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicDatasetWithSizeResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BasicDatasetWithSizeResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicDatasetWithSizeResponse$fromJSON(
  BasicDatasetWithSizeResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BasicDatasetWithSizeResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicDatasetWithSizeResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BasicDatasetWithSizeResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicDatasetWithSizeResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
