% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{BatchFeaturesTypeTransform}
\alias{BatchFeaturesTypeTransform}
\title{BatchFeaturesTypeTransform}
\usage{
BatchFeaturesTypeTransform(
  project,
  parentNames,
  variableType,
  prefix = NULL,
  suffix = NULL,
  maxWait = 600,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create multiple new features by changing the type of existing features.
}
\details{
Create multiple new features by changing the type of existing features.

This method invokes \verb{POST /projects/\{projectId\}/batchTypeTransformFeatures/} in the DataRobot Public API.
}
\seealso{
The method BatchTypeTransformFeaturesCreate in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
