% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_feature_transform.R
\docType{class}
\name{BatchFeatureTransform}
\alias{BatchFeatureTransform}
\title{BatchFeatureTransform}
\format{
An \code{R6Class} generator object
}
\description{
BatchFeatureTransform Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{parentNames}}{list( character ) List of feature names that will be transformed into a new variable type.}

\item{\code{prefix}}{character \link{optional} The string that will preface all feature names. Optional if suffix is present. (One or both are required.)}

\item{\code{suffix}}{character \link{optional} The string that will be appended at the end to all feature names. Optional if prefix is present. (One or both are required.)}

\item{\code{variableType}}{character The type of the new feature. Must be one of `text`, `categorical` (Deprecated in version v2.21), `numeric`, or `categoricalInt`.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BatchFeatureTransform-new}{\code{BatchFeatureTransform$new()}}
\item \href{#method-BatchFeatureTransform-validate}{\code{BatchFeatureTransform$validate()}}
\item \href{#method-BatchFeatureTransform-toJSON}{\code{BatchFeatureTransform$toJSON()}}
\item \href{#method-BatchFeatureTransform-fromJSON}{\code{BatchFeatureTransform$fromJSON()}}
\item \href{#method-BatchFeatureTransform-clone}{\code{BatchFeatureTransform$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchFeatureTransform-new"></a>}}
\if{latex}{\out{\hypertarget{method-BatchFeatureTransform-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchFeatureTransform$new(
  parentNames = NULL,
  variableType = NULL,
  prefix = NULL,
  suffix = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parentNames}}{List of feature names that will be transformed into a new variable type.}

\item{\code{variableType}}{The type of the new feature. Must be one of `text`, `categorical` (Deprecated in version v2.21), `numeric`, or `categoricalInt`.}

\item{\code{prefix}}{The string that will preface all feature names. Optional if suffix is present. (One or both are required.)}

\item{\code{suffix}}{The string that will be appended at the end to all feature names. Optional if prefix is present. (One or both are required.)}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchFeatureTransform-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BatchFeatureTransform-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchFeatureTransform$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchFeatureTransform-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BatchFeatureTransform-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchFeatureTransform$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchFeatureTransform-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BatchFeatureTransform-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchFeatureTransform$fromJSON(
  BatchFeatureTransformJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BatchFeatureTransformJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchFeatureTransform-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BatchFeatureTransform-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchFeatureTransform$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
