% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_feature_transform_retrieve_response.R
\docType{class}
\name{BatchFeatureTransformRetrieveResponse}
\alias{BatchFeatureTransformRetrieveResponse}
\title{BatchFeatureTransformRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
BatchFeatureTransformRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{failures}}{object An object keyed by original feature names, the values are strings indicating why the transformation failed.}

\item{\code{newFeatureNames}}{list( character ) List of new feature names.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BatchFeatureTransformRetrieveResponse-new}{\code{BatchFeatureTransformRetrieveResponse$new()}}
\item \href{#method-BatchFeatureTransformRetrieveResponse-validate}{\code{BatchFeatureTransformRetrieveResponse$validate()}}
\item \href{#method-BatchFeatureTransformRetrieveResponse-toJSON}{\code{BatchFeatureTransformRetrieveResponse$toJSON()}}
\item \href{#method-BatchFeatureTransformRetrieveResponse-fromJSON}{\code{BatchFeatureTransformRetrieveResponse$fromJSON()}}
\item \href{#method-BatchFeatureTransformRetrieveResponse-clone}{\code{BatchFeatureTransformRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchFeatureTransformRetrieveResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-BatchFeatureTransformRetrieveResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchFeatureTransformRetrieveResponse$new(
  failures = NULL,
  newFeatureNames = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{failures}}{An object keyed by original feature names, the values are strings indicating why the transformation failed.}

\item{\code{newFeatureNames}}{List of new feature names.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchFeatureTransformRetrieveResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BatchFeatureTransformRetrieveResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchFeatureTransformRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchFeatureTransformRetrieveResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BatchFeatureTransformRetrieveResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchFeatureTransformRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchFeatureTransformRetrieveResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BatchFeatureTransformRetrieveResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchFeatureTransformRetrieveResponse$fromJSON(
  BatchFeatureTransformRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BatchFeatureTransformRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchFeatureTransformRetrieveResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BatchFeatureTransformRetrieveResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchFeatureTransformRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
