% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_prediction_created_by.R
\docType{class}
\name{BatchPredictionCreatedBy}
\alias{BatchPredictionCreatedBy}
\title{BatchPredictionCreatedBy}
\format{
An \code{R6Class} generator object
}
\description{
BatchPredictionCreatedBy Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fullName}}{character The full name of the user who created this job (if defined by the user)}

\item{\code{userId}}{character The User ID of the user who created this job}

\item{\code{username}}{character The username (e-mail address) of the user who created this job}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BatchPredictionCreatedBy-new}{\code{BatchPredictionCreatedBy$new()}}
\item \href{#method-BatchPredictionCreatedBy-validate}{\code{BatchPredictionCreatedBy$validate()}}
\item \href{#method-BatchPredictionCreatedBy-toJSON}{\code{BatchPredictionCreatedBy$toJSON()}}
\item \href{#method-BatchPredictionCreatedBy-fromJSON}{\code{BatchPredictionCreatedBy$fromJSON()}}
\item \href{#method-BatchPredictionCreatedBy-clone}{\code{BatchPredictionCreatedBy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionCreatedBy-new"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionCreatedBy-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionCreatedBy$new(
  fullName = NULL,
  userId = NULL,
  username = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fullName}}{The full name of the user who created this job (if defined by the user)}

\item{\code{userId}}{The User ID of the user who created this job}

\item{\code{username}}{The username (e-mail address) of the user who created this job}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionCreatedBy-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionCreatedBy-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionCreatedBy$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionCreatedBy-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionCreatedBy-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionCreatedBy$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionCreatedBy-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionCreatedBy-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionCreatedBy$fromJSON(
  BatchPredictionCreatedByJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BatchPredictionCreatedByJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionCreatedBy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionCreatedBy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionCreatedBy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
