% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_prediction_job_create.R
\docType{class}
\name{BatchPredictionJobCreate}
\alias{BatchPredictionJobCreate}
\title{BatchPredictionJobCreate}
\format{
An \code{R6Class} generator object
}
\description{
BatchPredictionJobCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{abortOnError}}{character Should this job abort if too many errors are encountered}

\item{\code{chunkSize}}{\link{OneOfstringinteger} \link{optional} Which strategy should be used to determine the chunk size. Can be either a named strategy or a fixed size in bytes.}

\item{\code{columnNamesRemapping}}{\link{OneOfobjectarray} \link{optional} Remap (rename or remove columns from) the output from this job}

\item{\code{csvSettings}}{\link{BatchPredictionJobCSVSettings}}

\item{\code{deploymentId}}{character \link{optional} ID of deployment which is used in job for processing predictions dataset}

\item{\code{disableRowLevelErrorHandling}}{character Skip row by row error handling}

\item{\code{explanationAlgorithm}}{character \link{optional} Which algorithm will be used to calculate prediction explanations}

\item{\code{includePredictionStatus}}{character Include prediction status column in the output}

\item{\code{includeProbabilities}}{character Include probabilities for all classes}

\item{\code{includeProbabilitiesClasses}}{list( character ) Include only probabilities for these specific class names.}

\item{\code{intakeSettings}}{\link{OneOfAzureIntakeBigQueryIntakeDataStageIntakeCatalogDSSFileSystemIntakeGCPIntakeHTTPIntakeJDBCIntakeLocalFileIntakeS3IntakeSnowflakeIntakeSynapseIntake} The intake option configured for this job}

\item{\code{maxExplanations}}{integer Number of explanations requested. Will be ordered by strength.}

\item{\code{modelId}}{character \link{optional} ID of leaderboard model which is used in job for processing predictions dataset}

\item{\code{modelPackageId}}{character \link{optional} ID of model package from registry is used in job for processing predictions dataset}

\item{\code{numConcurrent}}{integer \link{optional} Number of simultaneous requests to run against the prediction instance}

\item{\code{outputSettings}}{\link{OneOfAzureOutputBigQueryOutputFileSystemOutputGCPOutputHTTPOutputJDBCOutputLocalFileOutputS3OutputSnowflakeOutputSynapseOutputTableau} The output option configured for this job}

\item{\code{passthroughColumns}}{list( character ) \link{optional} Pass through columns from the original dataset}

\item{\code{passthroughColumnsSet}}{character \link{optional} Pass through all columns from the original dataset}

\item{\code{pinnedModelId}}{character \link{optional} Specify a model ID used for scoring}

\item{\code{predictionInstance}}{\link{BatchPredictionJobPredictionInstance} \link{optional}}

\item{\code{predictionWarningEnabled}}{character \link{optional} Enable prediction warnings.}

\item{\code{skipDriftTracking}}{character Skip drift tracking for this job.}

\item{\code{thresholdHigh}}{numeric \link{optional} Compute explanations for predictions above this threshold}

\item{\code{thresholdLow}}{numeric \link{optional} Compute explanations for predictions below this threshold}

\item{\code{timeseriesSettings}}{\link{OneOfBatchPredictionJobTimeSeriesSettingsForecastBatchPredictionJobTimeSeriesSettingsHistorical} \link{optional} Time Series settings included of this job is a Time Series job.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BatchPredictionJobCreate$new()}}
\item \href{#method-validate}{\code{BatchPredictionJobCreate$validate()}}
\item \href{#method-toJSON}{\code{BatchPredictionJobCreate$toJSON()}}
\item \href{#method-fromJSON}{\code{BatchPredictionJobCreate$fromJSON()}}
\item \href{#method-clone}{\code{BatchPredictionJobCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobCreate$new(
  abortOnError = NULL,
  csvSettings = NULL,
  disableRowLevelErrorHandling = NULL,
  includePredictionStatus = NULL,
  includeProbabilities = NULL,
  includeProbabilitiesClasses = NULL,
  intakeSettings = NULL,
  maxExplanations = NULL,
  outputSettings = NULL,
  skipDriftTracking = NULL,
  chunkSize = NULL,
  columnNamesRemapping = NULL,
  deploymentId = NULL,
  explanationAlgorithm = NULL,
  modelId = NULL,
  modelPackageId = NULL,
  numConcurrent = NULL,
  passthroughColumns = NULL,
  passthroughColumnsSet = NULL,
  pinnedModelId = NULL,
  predictionInstance = NULL,
  predictionWarningEnabled = NULL,
  thresholdHigh = NULL,
  thresholdLow = NULL,
  timeseriesSettings = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{abortOnError}}{Should this job abort if too many errors are encountered}

\item{\code{disableRowLevelErrorHandling}}{Skip row by row error handling}

\item{\code{includePredictionStatus}}{Include prediction status column in the output}

\item{\code{includeProbabilities}}{Include probabilities for all classes}

\item{\code{includeProbabilitiesClasses}}{Include only probabilities for these specific class names.}

\item{\code{intakeSettings}}{The intake option configured for this job}

\item{\code{maxExplanations}}{Number of explanations requested. Will be ordered by strength.}

\item{\code{outputSettings}}{The output option configured for this job}

\item{\code{skipDriftTracking}}{Skip drift tracking for this job.}

\item{\code{chunkSize}}{Which strategy should be used to determine the chunk size. Can be either a named strategy or a fixed size in bytes.}

\item{\code{columnNamesRemapping}}{Remap (rename or remove columns from) the output from this job}

\item{\code{deploymentId}}{ID of deployment which is used in job for processing predictions dataset}

\item{\code{explanationAlgorithm}}{Which algorithm will be used to calculate prediction explanations}

\item{\code{modelId}}{ID of leaderboard model which is used in job for processing predictions dataset}

\item{\code{modelPackageId}}{ID of model package from registry is used in job for processing predictions dataset}

\item{\code{numConcurrent}}{Number of simultaneous requests to run against the prediction instance}

\item{\code{passthroughColumns}}{Pass through columns from the original dataset}

\item{\code{passthroughColumnsSet}}{Pass through all columns from the original dataset}

\item{\code{pinnedModelId}}{Specify a model ID used for scoring}

\item{\code{predictionWarningEnabled}}{Enable prediction warnings.}

\item{\code{thresholdHigh}}{Compute explanations for predictions above this threshold}

\item{\code{thresholdLow}}{Compute explanations for predictions below this threshold}

\item{\code{timeseriesSettings}}{Time Series settings included of this job is a Time Series job.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobCreate$fromJSON(
  BatchPredictionJobCreateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BatchPredictionJobCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
