% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_prediction_job_csv_settings.R
\docType{class}
\name{BatchPredictionJobCSVSettings}
\alias{BatchPredictionJobCSVSettings}
\title{BatchPredictionJobCSVSettings}
\format{
An \code{R6Class} generator object
}
\description{
BatchPredictionJobCSVSettings Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{delimiter}}{\link{OneOfstringstring} CSV fields are delimited by this character. Use the string \"tab\" to denote TSV (TAB separated values).}

\item{\code{encoding}}{character The encoding to be used for intake and output. For example (but not limited to): \"shift_jis\", \"latin_1\" or \"mskanji\".}

\item{\code{quotechar}}{character Fields containing the delimiter or newlines must be quoted using this character.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BatchPredictionJobCSVSettings-new}{\code{BatchPredictionJobCSVSettings$new()}}
\item \href{#method-BatchPredictionJobCSVSettings-validate}{\code{BatchPredictionJobCSVSettings$validate()}}
\item \href{#method-BatchPredictionJobCSVSettings-toJSON}{\code{BatchPredictionJobCSVSettings$toJSON()}}
\item \href{#method-BatchPredictionJobCSVSettings-fromJSON}{\code{BatchPredictionJobCSVSettings$fromJSON()}}
\item \href{#method-BatchPredictionJobCSVSettings-clone}{\code{BatchPredictionJobCSVSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionJobCSVSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionJobCSVSettings-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobCSVSettings$new(
  delimiter = NULL,
  encoding = NULL,
  quotechar = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delimiter}}{CSV fields are delimited by this character. Use the string \"tab\" to denote TSV (TAB separated values).}

\item{\code{encoding}}{The encoding to be used for intake and output. For example (but not limited to): \"shift_jis\", \"latin_1\" or \"mskanji\".}

\item{\code{quotechar}}{Fields containing the delimiter or newlines must be quoted using this character.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionJobCSVSettings-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionJobCSVSettings-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobCSVSettings$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionJobCSVSettings-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionJobCSVSettings-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobCSVSettings$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionJobCSVSettings-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionJobCSVSettings-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobCSVSettings$fromJSON(
  BatchPredictionJobCSVSettingsJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BatchPredictionJobCSVSettingsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionJobCSVSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionJobCSVSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobCSVSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
