% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_prediction_job_definitions_response.R
\docType{class}
\name{BatchPredictionJobDefinitionsResponse}
\alias{BatchPredictionJobDefinitionsResponse}
\title{BatchPredictionJobDefinitionsResponse}
\format{
An \code{R6Class} generator object
}
\description{
BatchPredictionJobDefinitionsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{batchPredictionJob}}{\link{BatchPredictionJobDefinitionJobSpecResponse}}

\item{\code{created}}{character When was this job created}

\item{\code{createdBy}}{\link{BatchPredictionCreatedBy}}

\item{\code{enabled}}{character If this job definition is enabled as a scheduled job.}

\item{\code{id}}{character The ID of the Batch Prediction job definition}

\item{\code{lastFailedRunTime}}{character \link{optional} Last time this job had a failed run}

\item{\code{lastScheduledRunTime}}{character \link{optional} Last time this job was scheduled to run (though not guaranteed it actually ran at that time)}

\item{\code{lastStartedJobStatus}}{character The status of the latest job launched to the queue (if any).}

\item{\code{lastStartedJobTime}}{character The last time (if any) a job was launched.}

\item{\code{lastSuccessfulRunTime}}{character \link{optional} Last time this job had a successful run}

\item{\code{name}}{character A human-readable name for the definition, must be unique across organisations}

\item{\code{nextScheduledRunTime}}{character \link{optional} Next time this job is scheduled to run}

\item{\code{schedule}}{\link{Schedule} \link{optional}}

\item{\code{updated}}{character When was this job last updated}

\item{\code{updatedBy}}{\link{BatchPredictionCreatedBy}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BatchPredictionJobDefinitionsResponse$new()}}
\item \href{#method-validate}{\code{BatchPredictionJobDefinitionsResponse$validate()}}
\item \href{#method-toJSON}{\code{BatchPredictionJobDefinitionsResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{BatchPredictionJobDefinitionsResponse$fromJSON()}}
\item \href{#method-clone}{\code{BatchPredictionJobDefinitionsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobDefinitionsResponse$new(
  batchPredictionJob = NULL,
  created = NULL,
  createdBy = NULL,
  enabled = NULL,
  id = NULL,
  lastStartedJobStatus = NULL,
  lastStartedJobTime = NULL,
  name = NULL,
  updated = NULL,
  updatedBy = NULL,
  lastFailedRunTime = NULL,
  lastScheduledRunTime = NULL,
  lastSuccessfulRunTime = NULL,
  nextScheduledRunTime = NULL,
  schedule = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{created}}{When was this job created}

\item{\code{enabled}}{If this job definition is enabled as a scheduled job.}

\item{\code{id}}{The ID of the Batch Prediction job definition}

\item{\code{lastStartedJobStatus}}{The status of the latest job launched to the queue (if any).}

\item{\code{lastStartedJobTime}}{The last time (if any) a job was launched.}

\item{\code{name}}{A human-readable name for the definition, must be unique across organisations}

\item{\code{updated}}{When was this job last updated}

\item{\code{lastFailedRunTime}}{Last time this job had a failed run}

\item{\code{lastScheduledRunTime}}{Last time this job was scheduled to run (though not guaranteed it actually ran at that time)}

\item{\code{lastSuccessfulRunTime}}{Last time this job had a successful run}

\item{\code{nextScheduledRunTime}}{Next time this job is scheduled to run}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobDefinitionsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobDefinitionsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobDefinitionsResponse$fromJSON(
  BatchPredictionJobDefinitionsResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BatchPredictionJobDefinitionsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobDefinitionsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
