% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_prediction_job_definitions_update.R
\docType{class}
\name{BatchPredictionJobDefinitionsUpdate}
\alias{BatchPredictionJobDefinitionsUpdate}
\title{BatchPredictionJobDefinitionsUpdate}
\format{
An \code{R6Class} generator object
}
\description{
BatchPredictionJobDefinitionsUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{abortOnError}}{character \link{optional} Should this job abort if too many errors are encountered}

\item{\code{chunkSize}}{\link{OneOfstringinteger} \link{optional} Which strategy should be used to determine the chunk size. Can be either a named strategy or a fixed size in bytes.}

\item{\code{columnNamesRemapping}}{\link{OneOfobjectarray} \link{optional} Remap (rename or remove columns from) the output from this job}

\item{\code{csvSettings}}{\link{BatchPredictionJobCSVSettings} \link{optional}}

\item{\code{deploymentId}}{character \link{optional} ID of deployment which is used in job for processing predictions dataset}

\item{\code{disableRowLevelErrorHandling}}{character \link{optional} Skip row by row error handling}

\item{\code{enabled}}{character \link{optional} If this job definition is enabled as a scheduled job. Optional if no schedule is supplied.}

\item{\code{explanationAlgorithm}}{character \link{optional} Which algorithm will be used to calculate prediction explanations}

\item{\code{explanationClassNames}}{list( character ) \link{optional} List of class names that will be explained for each row for multiclass. Mutually exclusive with explanationNumTopClasses. If neither specified - we assume explanationNumTopClasses=1}

\item{\code{explanationNumTopClasses}}{integer \link{optional} Number of top predicted classes for each row that will be explained for multiclass. Mutually exclusive with explanationClassNames. If neither specified - we assume explanationNumTopClasses=1}

\item{\code{includePredictionStatus}}{character \link{optional} Include prediction status column in the output}

\item{\code{includeProbabilities}}{character \link{optional} Include probabilities for all classes}

\item{\code{includeProbabilitiesClasses}}{list( character ) \link{optional} Include only probabilities for these specific class names.}

\item{\code{intakeSettings}}{\link{OneOfAzureIntakeBigQueryIntakeDataStageIntakeCatalogDSSFileSystemIntakeGCPIntakeHTTPIntakeJDBCIntakeLocalFileIntakeS3IntakeSnowflakeIntakeSynapseIntake} \link{optional} The intake option configured for this job}

\item{\code{maxExplanations}}{integer \link{optional} Number of explanations requested. Will be ordered by strength.}

\item{\code{modelId}}{character \link{optional} ID of leaderboard model which is used in job for processing predictions dataset}

\item{\code{modelPackageId}}{character \link{optional} ID of model package from registry is used in job for processing predictions dataset}

\item{\code{name}}{character \link{optional} A human-readable name for the definition, must be unique across organisations, if left out the backend will generate one for you.}

\item{\code{numConcurrent}}{integer \link{optional} Number of simultaneous requests to run against the prediction instance}

\item{\code{outputSettings}}{\link{OneOfAzureOutputBigQueryOutputFileSystemOutputGCPOutputHTTPOutputJDBCOutputLocalFileOutputS3OutputSnowflakeOutputSynapseOutputTableau} \link{optional} The output option configured for this job}

\item{\code{passthroughColumns}}{list( character ) \link{optional} Pass through columns from the original dataset}

\item{\code{passthroughColumnsSet}}{character \link{optional} Pass through all columns from the original dataset}

\item{\code{pinnedModelId}}{character \link{optional} Specify a model ID used for scoring}

\item{\code{predictionInstance}}{\link{BatchPredictionJobPredictionInstance} \link{optional}}

\item{\code{predictionWarningEnabled}}{character \link{optional} Enable prediction warnings.}

\item{\code{schedule}}{\link{Schedule} \link{optional}}

\item{\code{skipDriftTracking}}{character \link{optional} Skip drift tracking for this job.}

\item{\code{thresholdHigh}}{numeric \link{optional} Compute explanations for predictions above this threshold}

\item{\code{thresholdLow}}{numeric \link{optional} Compute explanations for predictions below this threshold}

\item{\code{timeseriesSettings}}{\link{OneOfBatchPredictionJobTimeSeriesSettingsForecastBatchPredictionJobTimeSeriesSettingsForecastWithPolicyBatchPredictionJobTimeSeriesSettingsHistorical} \link{optional} Time Series settings included of this job is a Time Series job.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BatchPredictionJobDefinitionsUpdate-new}{\code{BatchPredictionJobDefinitionsUpdate$new()}}
\item \href{#method-BatchPredictionJobDefinitionsUpdate-validate}{\code{BatchPredictionJobDefinitionsUpdate$validate()}}
\item \href{#method-BatchPredictionJobDefinitionsUpdate-toJSON}{\code{BatchPredictionJobDefinitionsUpdate$toJSON()}}
\item \href{#method-BatchPredictionJobDefinitionsUpdate-fromJSON}{\code{BatchPredictionJobDefinitionsUpdate$fromJSON()}}
\item \href{#method-BatchPredictionJobDefinitionsUpdate-clone}{\code{BatchPredictionJobDefinitionsUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionJobDefinitionsUpdate-new"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionJobDefinitionsUpdate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobDefinitionsUpdate$new(
  abortOnError = TRUE,
  chunkSize = NULL,
  columnNamesRemapping = NULL,
  csvSettings = NULL,
  deploymentId = NULL,
  disableRowLevelErrorHandling = FALSE,
  enabled = NULL,
  explanationAlgorithm = NULL,
  explanationClassNames = NULL,
  explanationNumTopClasses = NULL,
  includePredictionStatus = FALSE,
  includeProbabilities = TRUE,
  includeProbabilitiesClasses = c(),
  intakeSettings = NULL,
  maxExplanations = 0,
  modelId = NULL,
  modelPackageId = NULL,
  name = NULL,
  numConcurrent = NULL,
  outputSettings = NULL,
  passthroughColumns = NULL,
  passthroughColumnsSet = NULL,
  pinnedModelId = NULL,
  predictionInstance = NULL,
  predictionWarningEnabled = NULL,
  schedule = NULL,
  skipDriftTracking = FALSE,
  thresholdHigh = NULL,
  thresholdLow = NULL,
  timeseriesSettings = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{abortOnError}}{Should this job abort if too many errors are encountered}

\item{\code{chunkSize}}{Which strategy should be used to determine the chunk size. Can be either a named strategy or a fixed size in bytes.}

\item{\code{columnNamesRemapping}}{Remap (rename or remove columns from) the output from this job}

\item{\code{deploymentId}}{ID of deployment which is used in job for processing predictions dataset}

\item{\code{disableRowLevelErrorHandling}}{Skip row by row error handling}

\item{\code{enabled}}{If this job definition is enabled as a scheduled job. Optional if no schedule is supplied.}

\item{\code{explanationAlgorithm}}{Which algorithm will be used to calculate prediction explanations}

\item{\code{explanationClassNames}}{List of class names that will be explained for each row for multiclass. Mutually exclusive with explanationNumTopClasses. If neither specified - we assume explanationNumTopClasses=1}

\item{\code{explanationNumTopClasses}}{Number of top predicted classes for each row that will be explained for multiclass. Mutually exclusive with explanationClassNames. If neither specified - we assume explanationNumTopClasses=1}

\item{\code{includePredictionStatus}}{Include prediction status column in the output}

\item{\code{includeProbabilities}}{Include probabilities for all classes}

\item{\code{includeProbabilitiesClasses}}{Include only probabilities for these specific class names.}

\item{\code{intakeSettings}}{The intake option configured for this job}

\item{\code{maxExplanations}}{Number of explanations requested. Will be ordered by strength.}

\item{\code{modelId}}{ID of leaderboard model which is used in job for processing predictions dataset}

\item{\code{modelPackageId}}{ID of model package from registry is used in job for processing predictions dataset}

\item{\code{name}}{A human-readable name for the definition, must be unique across organisations, if left out the backend will generate one for you.}

\item{\code{numConcurrent}}{Number of simultaneous requests to run against the prediction instance}

\item{\code{outputSettings}}{The output option configured for this job}

\item{\code{passthroughColumns}}{Pass through columns from the original dataset}

\item{\code{passthroughColumnsSet}}{Pass through all columns from the original dataset}

\item{\code{pinnedModelId}}{Specify a model ID used for scoring}

\item{\code{predictionWarningEnabled}}{Enable prediction warnings.}

\item{\code{skipDriftTracking}}{Skip drift tracking for this job.}

\item{\code{thresholdHigh}}{Compute explanations for predictions above this threshold}

\item{\code{thresholdLow}}{Compute explanations for predictions below this threshold}

\item{\code{timeseriesSettings}}{Time Series settings included of this job is a Time Series job.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionJobDefinitionsUpdate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionJobDefinitionsUpdate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobDefinitionsUpdate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionJobDefinitionsUpdate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionJobDefinitionsUpdate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobDefinitionsUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionJobDefinitionsUpdate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionJobDefinitionsUpdate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobDefinitionsUpdate$fromJSON(
  BatchPredictionJobDefinitionsUpdateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BatchPredictionJobDefinitionsUpdateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionJobDefinitionsUpdate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionJobDefinitionsUpdate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobDefinitionsUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
