% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_prediction_job_links.R
\docType{class}
\name{BatchPredictionJobLinks}
\alias{BatchPredictionJobLinks}
\title{BatchPredictionJobLinks}
\format{
An \code{R6Class} generator object
}
\description{
BatchPredictionJobLinks Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{csvUpload}}{character \link{optional} The URL used to upload the dataset for this job. Only available for localFile intake.}

\item{\code{download}}{character \link{optional} The URL used to download the results from this job. Only available for localFile outputs. Will be null if the download is not yet available.}

\item{\code{self_}}{character The URL used access this job.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BatchPredictionJobLinks$new()}}
\item \href{#method-validate}{\code{BatchPredictionJobLinks$validate()}}
\item \href{#method-toJSON}{\code{BatchPredictionJobLinks$toJSON()}}
\item \href{#method-fromJSON}{\code{BatchPredictionJobLinks$fromJSON()}}
\item \href{#method-clone}{\code{BatchPredictionJobLinks$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobLinks$new(
  self_ = NULL,
  csvUpload = NULL,
  download = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{self_}}{The URL used access this job.}

\item{\code{csvUpload}}{The URL used to upload the dataset for this job. Only available for localFile intake.}

\item{\code{download}}{The URL used to download the results from this job. Only available for localFile outputs. Will be null if the download is not yet available.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobLinks$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobLinks$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobLinks$fromJSON(
  BatchPredictionJobLinksJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BatchPredictionJobLinksJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobLinks$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
