% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_prediction_job_response.R
\docType{class}
\name{BatchPredictionJobResponse}
\alias{BatchPredictionJobResponse}
\title{BatchPredictionJobResponse}
\format{
An \code{R6Class} generator object
}
\description{
BatchPredictionJobResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{batchPredictionJobDefinition}}{\link{BatchPredictionJobDefinitionResponse} \link{optional}}

\item{\code{created}}{character When was this job created}

\item{\code{createdBy}}{\link{BatchPredictionCreatedBy}}

\item{\code{elapsedTimeSec}}{integer Number of seconds the job has been processing for}

\item{\code{failedRows}}{integer Number of rows that have failed scoring}

\item{\code{hidden}}{character \link{optional} When was this job was hidden last, blank if visible}

\item{\code{id}}{character The ID of the Batch Prediction job}

\item{\code{intakeDatasetDisplayName}}{character \link{optional} If applicable (e.g. for AI catalog), will contain the dataset name used for the intake dataset.}

\item{\code{jobIntakeSize}}{integer Number of bytes in the intake dataset for this job}

\item{\code{jobOutputSize}}{integer Number of bytes in the output dataset for this job}

\item{\code{jobSpec}}{\link{BatchPredictionJobSpecResponse}}

\item{\code{links}}{\link{BatchPredictionJobLinks}}

\item{\code{logs}}{list( character ) The job log.}

\item{\code{percentageCompleted}}{numeric Indicates job progress which is based on number of already processed rows in dataset}

\item{\code{queuePosition}}{integer \link{optional} To ensure a dedicated prediction instance is not overloaded, only one job will be run against it at a time. This is the number of jobs that are awaiting processing before this job start running. May not be available in all environments.}

\item{\code{queued}}{character The job has been put on the queue for execution.}

\item{\code{resultsDeleted}}{character \link{optional} Indicates if the job was subject to garbage collection and had its artifacts deleted (output files, if any, and scoring data on local storage)}

\item{\code{scoredRows}}{integer Number of rows that have been successfully scored}

\item{\code{skippedRows}}{integer Number of rows that have been skipped during scoring. May contain non-zero value only in time-series predictions case if provided dataset contains more than required historical rows.}

\item{\code{source}}{character \link{optional} Source from which batch job was started}

\item{\code{status}}{character The current job status}

\item{\code{statusDetails}}{character Explanation for current status}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BatchPredictionJobResponse$new()}}
\item \href{#method-validate}{\code{BatchPredictionJobResponse$validate()}}
\item \href{#method-toJSON}{\code{BatchPredictionJobResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{BatchPredictionJobResponse$fromJSON()}}
\item \href{#method-clone}{\code{BatchPredictionJobResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobResponse$new(
  created = NULL,
  createdBy = NULL,
  elapsedTimeSec = NULL,
  failedRows = NULL,
  id = NULL,
  jobIntakeSize = NULL,
  jobOutputSize = NULL,
  jobSpec = NULL,
  links = NULL,
  logs = NULL,
  percentageCompleted = NULL,
  queued = NULL,
  scoredRows = NULL,
  skippedRows = NULL,
  status = NULL,
  statusDetails = NULL,
  batchPredictionJobDefinition = NULL,
  hidden = NULL,
  intakeDatasetDisplayName = NULL,
  queuePosition = NULL,
  resultsDeleted = NULL,
  source = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{created}}{When was this job created}

\item{\code{elapsedTimeSec}}{Number of seconds the job has been processing for}

\item{\code{failedRows}}{Number of rows that have failed scoring}

\item{\code{id}}{The ID of the Batch Prediction job}

\item{\code{jobIntakeSize}}{Number of bytes in the intake dataset for this job}

\item{\code{jobOutputSize}}{Number of bytes in the output dataset for this job}

\item{\code{logs}}{The job log.}

\item{\code{percentageCompleted}}{Indicates job progress which is based on number of already processed rows in dataset}

\item{\code{queued}}{The job has been put on the queue for execution.}

\item{\code{scoredRows}}{Number of rows that have been successfully scored}

\item{\code{skippedRows}}{Number of rows that have been skipped during scoring. May contain non-zero value only in time-series predictions case if provided dataset contains more than required historical rows.}

\item{\code{status}}{The current job status}

\item{\code{statusDetails}}{Explanation for current status}

\item{\code{hidden}}{When was this job was hidden last, blank if visible}

\item{\code{intakeDatasetDisplayName}}{If applicable (e.g. for AI catalog), will contain the dataset name used for the intake dataset.}

\item{\code{queuePosition}}{To ensure a dedicated prediction instance is not overloaded, only one job will be run against it at a time. This is the number of jobs that are awaiting processing before this job start running. May not be available in all environments.}

\item{\code{resultsDeleted}}{Indicates if the job was subject to garbage collection and had its artifacts deleted (output files, if any, and scoring data on local storage)}

\item{\code{source}}{Source from which batch job was started}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobResponse$fromJSON(
  BatchPredictionJobResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BatchPredictionJobResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
