% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/batch_prediction_job_time_series_settings_forecast.R
\docType{class}
\name{BatchPredictionJobTimeSeriesSettingsForecast}
\alias{BatchPredictionJobTimeSeriesSettingsForecast}
\title{BatchPredictionJobTimeSeriesSettingsForecast}
\format{
An \code{R6Class} generator object
}
\description{
BatchPredictionJobTimeSeriesSettingsForecast Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{forecastPoint}}{character \link{optional} Used for forecast predictions in order to override the inferred forecast point from the dataset.}

\item{\code{relaxKnownInAdvanceFeaturesCheck}}{character \link{optional} If activated, missing values in the known in advance features are allowed in the forecast window at prediction time. If omitted or false, missing values are not allowed.}

\item{\code{type}}{character Forecast mode makes predictions using forecastPoint or rows in the dataset without target.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BatchPredictionJobTimeSeriesSettingsForecast-new}{\code{BatchPredictionJobTimeSeriesSettingsForecast$new()}}
\item \href{#method-BatchPredictionJobTimeSeriesSettingsForecast-validate}{\code{BatchPredictionJobTimeSeriesSettingsForecast$validate()}}
\item \href{#method-BatchPredictionJobTimeSeriesSettingsForecast-toJSON}{\code{BatchPredictionJobTimeSeriesSettingsForecast$toJSON()}}
\item \href{#method-BatchPredictionJobTimeSeriesSettingsForecast-fromJSON}{\code{BatchPredictionJobTimeSeriesSettingsForecast$fromJSON()}}
\item \href{#method-BatchPredictionJobTimeSeriesSettingsForecast-clone}{\code{BatchPredictionJobTimeSeriesSettingsForecast$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionJobTimeSeriesSettingsForecast-new"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionJobTimeSeriesSettingsForecast-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobTimeSeriesSettingsForecast$new(
  type = NULL,
  forecastPoint = NULL,
  relaxKnownInAdvanceFeaturesCheck = FALSE,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Forecast mode makes predictions using forecastPoint or rows in the dataset without target.}

\item{\code{forecastPoint}}{Used for forecast predictions in order to override the inferred forecast point from the dataset.}

\item{\code{relaxKnownInAdvanceFeaturesCheck}}{If activated, missing values in the known in advance features are allowed in the forecast window at prediction time. If omitted or false, missing values are not allowed.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionJobTimeSeriesSettingsForecast-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionJobTimeSeriesSettingsForecast-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobTimeSeriesSettingsForecast$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionJobTimeSeriesSettingsForecast-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionJobTimeSeriesSettingsForecast-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobTimeSeriesSettingsForecast$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionJobTimeSeriesSettingsForecast-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionJobTimeSeriesSettingsForecast-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobTimeSeriesSettingsForecast$fromJSON(
  BatchPredictionJobTimeSeriesSettingsForecastJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BatchPredictionJobTimeSeriesSettingsForecastJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionJobTimeSeriesSettingsForecast-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionJobTimeSeriesSettingsForecast-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobTimeSeriesSettingsForecast$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
