% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_prediction_job_update.R
\docType{class}
\name{BatchPredictionJobUpdate}
\alias{BatchPredictionJobUpdate}
\title{BatchPredictionJobUpdate}
\format{
An \code{R6Class} generator object
}
\description{
BatchPredictionJobUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aborted}}{character \link{optional} Time when job abortion happened}

\item{\code{completed}}{character \link{optional} Time when job completed scoring}

\item{\code{failedRows}}{integer \link{optional} Number of rows that have failed scoring}

\item{\code{hidden}}{character \link{optional} Hides or unhides the job from the job list}

\item{\code{jobIntakeSize}}{integer \link{optional} Number of bytes in the intake dataset for this job}

\item{\code{jobOutputSize}}{integer \link{optional} Number of bytes in the output dataset for this job}

\item{\code{logs}}{list( character ) \link{optional} The job log.}

\item{\code{scoredRows}}{integer \link{optional} Number of rows that have been successfully scored}

\item{\code{skippedRows}}{integer \link{optional} Number of rows that have been skipped during scoring. May contain non-zero value only in time-series predictions case if provided dataset contains more than required historical rows.}

\item{\code{started}}{character \link{optional} Time when job scoring begin}

\item{\code{status}}{character \link{optional} The current job status}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BatchPredictionJobUpdate-new}{\code{BatchPredictionJobUpdate$new()}}
\item \href{#method-BatchPredictionJobUpdate-validate}{\code{BatchPredictionJobUpdate$validate()}}
\item \href{#method-BatchPredictionJobUpdate-toJSON}{\code{BatchPredictionJobUpdate$toJSON()}}
\item \href{#method-BatchPredictionJobUpdate-fromJSON}{\code{BatchPredictionJobUpdate$fromJSON()}}
\item \href{#method-BatchPredictionJobUpdate-clone}{\code{BatchPredictionJobUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionJobUpdate-new"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionJobUpdate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobUpdate$new(
  aborted = NULL,
  completed = NULL,
  failedRows = NULL,
  hidden = NULL,
  jobIntakeSize = NULL,
  jobOutputSize = NULL,
  logs = NULL,
  scoredRows = NULL,
  skippedRows = NULL,
  started = NULL,
  status = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aborted}}{Time when job abortion happened}

\item{\code{completed}}{Time when job completed scoring}

\item{\code{failedRows}}{Number of rows that have failed scoring}

\item{\code{hidden}}{Hides or unhides the job from the job list}

\item{\code{jobIntakeSize}}{Number of bytes in the intake dataset for this job}

\item{\code{jobOutputSize}}{Number of bytes in the output dataset for this job}

\item{\code{logs}}{The job log.}

\item{\code{scoredRows}}{Number of rows that have been successfully scored}

\item{\code{skippedRows}}{Number of rows that have been skipped during scoring. May contain non-zero value only in time-series predictions case if provided dataset contains more than required historical rows.}

\item{\code{started}}{Time when job scoring begin}

\item{\code{status}}{The current job status}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionJobUpdate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionJobUpdate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobUpdate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionJobUpdate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionJobUpdate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionJobUpdate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionJobUpdate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobUpdate$fromJSON(
  BatchPredictionJobUpdateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BatchPredictionJobUpdateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchPredictionJobUpdate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BatchPredictionJobUpdate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchPredictionJobUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
