% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_mitigated_models_data_response.R
\docType{class}
\name{BiasMitigatedModelsDataResponse}
\alias{BiasMitigatedModelsDataResponse}
\title{BiasMitigatedModelsDataResponse}
\format{
An \code{R6Class} generator object
}
\description{
BiasMitigatedModelsDataResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{biasMitigationTechnique}}{character Method applied to perform bias mitigation.}

\item{\code{includeBiasMitigationFeatureAsPredictorVariable}}{character Specifies whether the mitigation feature will be used as a predictor variable (i.e., treated like other categorical features in the input to train the modeler), in addition to being used for bias mitigation. If false, the mitigation feature will be used only for bias mitigation, and not for training the modeler task.}

\item{\code{modelId}}{character Mitigated model ID}

\item{\code{parentModelId}}{character Parent model ID}

\item{\code{protectedFeature}}{character Protected feature that will be used in a bias mitigation task to mitigate bias}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BiasMitigatedModelsDataResponse-new}{\code{BiasMitigatedModelsDataResponse$new()}}
\item \href{#method-BiasMitigatedModelsDataResponse-validate}{\code{BiasMitigatedModelsDataResponse$validate()}}
\item \href{#method-BiasMitigatedModelsDataResponse-toJSON}{\code{BiasMitigatedModelsDataResponse$toJSON()}}
\item \href{#method-BiasMitigatedModelsDataResponse-fromJSON}{\code{BiasMitigatedModelsDataResponse$fromJSON()}}
\item \href{#method-BiasMitigatedModelsDataResponse-clone}{\code{BiasMitigatedModelsDataResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiasMitigatedModelsDataResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-BiasMitigatedModelsDataResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasMitigatedModelsDataResponse$new(
  biasMitigationTechnique = NULL,
  includeBiasMitigationFeatureAsPredictorVariable = NULL,
  modelId = NULL,
  parentModelId = NULL,
  protectedFeature = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{biasMitigationTechnique}}{Method applied to perform bias mitigation.}

\item{\code{includeBiasMitigationFeatureAsPredictorVariable}}{Specifies whether the mitigation feature will be used as a predictor variable (i.e., treated like other categorical features in the input to train the modeler), in addition to being used for bias mitigation. If false, the mitigation feature will be used only for bias mitigation, and not for training the modeler task.}

\item{\code{modelId}}{Mitigated model ID}

\item{\code{parentModelId}}{Parent model ID}

\item{\code{protectedFeature}}{Protected feature that will be used in a bias mitigation task to mitigate bias}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiasMitigatedModelsDataResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BiasMitigatedModelsDataResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasMitigatedModelsDataResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiasMitigatedModelsDataResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BiasMitigatedModelsDataResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasMitigatedModelsDataResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiasMitigatedModelsDataResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BiasMitigatedModelsDataResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasMitigatedModelsDataResponse$fromJSON(
  BiasMitigatedModelsDataResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BiasMitigatedModelsDataResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiasMitigatedModelsDataResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BiasMitigatedModelsDataResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasMitigatedModelsDataResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
