% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_mitigation_model_create.R
\docType{class}
\name{BiasMitigationModelCreate}
\alias{BiasMitigationModelCreate}
\title{BiasMitigationModelCreate}
\format{
An \code{R6Class} generator object
}
\description{
BiasMitigationModelCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{biasMitigationFeature}}{character The name of the protected feature used to mitigate bias on models.}

\item{\code{biasMitigationParentLid}}{character The ID of the model to modify with a bias-mitigation task.}

\item{\code{biasMitigationTechnique}}{character Method applied to perform bias mitigation.}

\item{\code{includeBiasMitigationFeatureAsPredictorVariable}}{character Specifies whether the mitigation feature will be used as a predictor variable (i.e., treated like other categorical features in the input to train the modeler), in addition to being used for bias mitigation. If false, the mitigation feature will be used only for bias mitigation, and not for training the modeler task.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BiasMitigationModelCreate-new}{\code{BiasMitigationModelCreate$new()}}
\item \href{#method-BiasMitigationModelCreate-validate}{\code{BiasMitigationModelCreate$validate()}}
\item \href{#method-BiasMitigationModelCreate-toJSON}{\code{BiasMitigationModelCreate$toJSON()}}
\item \href{#method-BiasMitigationModelCreate-fromJSON}{\code{BiasMitigationModelCreate$fromJSON()}}
\item \href{#method-BiasMitigationModelCreate-clone}{\code{BiasMitigationModelCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiasMitigationModelCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-BiasMitigationModelCreate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasMitigationModelCreate$new(
  biasMitigationFeature = NULL,
  biasMitigationParentLid = NULL,
  biasMitigationTechnique = NULL,
  includeBiasMitigationFeatureAsPredictorVariable = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{biasMitigationFeature}}{The name of the protected feature used to mitigate bias on models.}

\item{\code{biasMitigationParentLid}}{The ID of the model to modify with a bias-mitigation task.}

\item{\code{biasMitigationTechnique}}{Method applied to perform bias mitigation.}

\item{\code{includeBiasMitigationFeatureAsPredictorVariable}}{Specifies whether the mitigation feature will be used as a predictor variable (i.e., treated like other categorical features in the input to train the modeler), in addition to being used for bias mitigation. If false, the mitigation feature will be used only for bias mitigation, and not for training the modeler task.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiasMitigationModelCreate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BiasMitigationModelCreate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasMitigationModelCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiasMitigationModelCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BiasMitigationModelCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasMitigationModelCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiasMitigationModelCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BiasMitigationModelCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasMitigationModelCreate$fromJSON(
  BiasMitigationModelCreateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BiasMitigationModelCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BiasMitigationModelCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BiasMitigationModelCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasMitigationModelCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
