% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_vs_accuracy_insight.R
\docType{class}
\name{BiasVsAccuracyInsight}
\alias{BiasVsAccuracyInsight}
\title{BiasVsAccuracyInsight}
\format{
An \code{R6Class} generator object
}
\description{
BiasVsAccuracyInsight Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{accuracyMetric}}{character \link{optional} The metric to return model accuracy scores. Defaults to the optimization metric configured in project options.}

\item{\code{fairnessMetric}}{\link{OneOfstringarray} \link{optional} The fairness metric used to calculate the fairness scores.}

\item{\code{fairnessThreshold}}{numeric Value of the fairness threshold, defined in project options.}

\item{\code{models}}{list( \link{BiasVsAccuracyModels} ) An array of models of the insight.}

\item{\code{protectedFeature}}{\link{OneOfstringarray} \link{optional} Name of the protected feature.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BiasVsAccuracyInsight$new()}}
\item \href{#method-validate}{\code{BiasVsAccuracyInsight$validate()}}
\item \href{#method-toJSON}{\code{BiasVsAccuracyInsight$toJSON()}}
\item \href{#method-fromJSON}{\code{BiasVsAccuracyInsight$fromJSON()}}
\item \href{#method-clone}{\code{BiasVsAccuracyInsight$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasVsAccuracyInsight$new(
  fairnessThreshold = NULL,
  models = NULL,
  accuracyMetric = NULL,
  fairnessMetric = NULL,
  protectedFeature = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fairnessThreshold}}{Value of the fairness threshold, defined in project options.}

\item{\code{models}}{An array of models of the insight.}

\item{\code{accuracyMetric}}{The metric to return model accuracy scores. Defaults to the optimization metric configured in project options.}

\item{\code{fairnessMetric}}{The fairness metric used to calculate the fairness scores.}

\item{\code{protectedFeature}}{Name of the protected feature.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasVsAccuracyInsight$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasVsAccuracyInsight$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasVsAccuracyInsight$fromJSON(
  BiasVsAccuracyInsightJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BiasVsAccuracyInsightJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiasVsAccuracyInsight$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
