% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_query_intake.R
\docType{class}
\name{BigQueryIntake}
\alias{BigQueryIntake}
\title{BigQueryIntake}
\format{
An \code{R6Class} generator object
}
\description{
BigQueryIntake Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bucket}}{character The name of gcs bucket for data export}

\item{\code{credentialId}}{character \link{optional} The ID of the GCP credentials}

\item{\code{dataset}}{character The name of the specified big query dataset to read input data from}

\item{\code{table}}{character The name of the specified big query table to read input data from}

\item{\code{type}}{character Type name for this intake type}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BigQueryIntake-new}{\code{BigQueryIntake$new()}}
\item \href{#method-BigQueryIntake-validate}{\code{BigQueryIntake$validate()}}
\item \href{#method-BigQueryIntake-toJSON}{\code{BigQueryIntake$toJSON()}}
\item \href{#method-BigQueryIntake-fromJSON}{\code{BigQueryIntake$fromJSON()}}
\item \href{#method-BigQueryIntake-clone}{\code{BigQueryIntake$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BigQueryIntake-new"></a>}}
\if{latex}{\out{\hypertarget{method-BigQueryIntake-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigQueryIntake$new(
  bucket = NULL,
  dataset = NULL,
  table = NULL,
  type = NULL,
  credentialId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bucket}}{The name of gcs bucket for data export}

\item{\code{dataset}}{The name of the specified big query dataset to read input data from}

\item{\code{table}}{The name of the specified big query table to read input data from}

\item{\code{type}}{Type name for this intake type}

\item{\code{credentialId}}{The ID of the GCP credentials}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BigQueryIntake-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BigQueryIntake-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigQueryIntake$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BigQueryIntake-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BigQueryIntake-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigQueryIntake$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BigQueryIntake-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BigQueryIntake-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigQueryIntake$fromJSON(BigQueryIntakeJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BigQueryIntakeJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BigQueryIntake-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BigQueryIntake-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigQueryIntake$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
