% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_query_output.R
\docType{class}
\name{BigQueryOutput}
\alias{BigQueryOutput}
\title{BigQueryOutput}
\format{
An \code{R6Class} generator object
}
\description{
BigQueryOutput Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bucket}}{character The name of gcs bucket for data loading}

\item{\code{credentialId}}{character \link{optional} The ID of the GCP credentials}

\item{\code{dataset}}{character The name of the specified big query dataset to write data back}

\item{\code{table}}{character The name of the specified big query table to write data back}

\item{\code{type}}{character Type name for this output type}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BigQueryOutput-new}{\code{BigQueryOutput$new()}}
\item \href{#method-BigQueryOutput-validate}{\code{BigQueryOutput$validate()}}
\item \href{#method-BigQueryOutput-toJSON}{\code{BigQueryOutput$toJSON()}}
\item \href{#method-BigQueryOutput-fromJSON}{\code{BigQueryOutput$fromJSON()}}
\item \href{#method-BigQueryOutput-clone}{\code{BigQueryOutput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BigQueryOutput-new"></a>}}
\if{latex}{\out{\hypertarget{method-BigQueryOutput-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigQueryOutput$new(
  bucket = NULL,
  dataset = NULL,
  table = NULL,
  type = NULL,
  credentialId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bucket}}{The name of gcs bucket for data loading}

\item{\code{dataset}}{The name of the specified big query dataset to write data back}

\item{\code{table}}{The name of the specified big query table to write data back}

\item{\code{type}}{Type name for this output type}

\item{\code{credentialId}}{The ID of the GCP credentials}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BigQueryOutput-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BigQueryOutput-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigQueryOutput$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BigQueryOutput-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BigQueryOutput-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigQueryOutput$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BigQueryOutput-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BigQueryOutput-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigQueryOutput$fromJSON(BigQueryOutputJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BigQueryOutputJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BigQueryOutput-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BigQueryOutput-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigQueryOutput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
