% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_query_output_adaptor.R
\docType{class}
\name{BigQueryOutputAdaptor}
\alias{BigQueryOutputAdaptor}
\title{BigQueryOutputAdaptor}
\format{
An \code{R6Class} generator object
}
\description{
BigQueryOutputAdaptor Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bucket}}{character The name of gcs bucket for data loading}

\item{\code{credentialId}}{\link{OneOfstringstring} \link{optional} Either the populated value of the field or \link{redacted} due to permission settings}

\item{\code{dataset}}{character The name of the specified big query dataset to write data back}

\item{\code{table}}{character The name of the specified big query table to write data back}

\item{\code{type}}{character Type name for this output type}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BigQueryOutputAdaptor$new()}}
\item \href{#method-validate}{\code{BigQueryOutputAdaptor$validate()}}
\item \href{#method-toJSON}{\code{BigQueryOutputAdaptor$toJSON()}}
\item \href{#method-fromJSON}{\code{BigQueryOutputAdaptor$fromJSON()}}
\item \href{#method-clone}{\code{BigQueryOutputAdaptor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigQueryOutputAdaptor$new(
  bucket = NULL,
  dataset = NULL,
  table = NULL,
  type = NULL,
  credentialId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bucket}}{The name of gcs bucket for data loading}

\item{\code{dataset}}{The name of the specified big query dataset to write data back}

\item{\code{table}}{The name of the specified big query table to write data back}

\item{\code{type}}{Type name for this output type}

\item{\code{credentialId}}{Either the populated value of the field or \link{redacted} due to permission settings}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigQueryOutputAdaptor$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigQueryOutputAdaptor$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigQueryOutputAdaptor$fromJSON(
  BigQueryOutputAdaptorJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BigQueryOutputAdaptorJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BigQueryOutputAdaptor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
