% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_response.R
\docType{class}
\name{BinResponse}
\alias{BinResponse}
\title{BinResponse}
\format{
An \code{R6Class} generator object
}
\description{
BinResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{avgPredicted}}{numeric Average prediction of the model in the bin. Null if there are no entries in the bin.}

\item{\code{endDate}}{character ISO-formatted datetime of the end of the bin (exclusive).}

\item{\code{frequency}}{integer Number of the rows in the bin.}

\item{\code{maxPredicted}}{numeric Maximum prediction of the model in the bin. Null if there are no entries in the bin.}

\item{\code{startDate}}{character ISO-formatted datetime of the start of the bin (inclusive).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BinResponse-new}{\code{BinResponse$new()}}
\item \href{#method-BinResponse-validate}{\code{BinResponse$validate()}}
\item \href{#method-BinResponse-toJSON}{\code{BinResponse$toJSON()}}
\item \href{#method-BinResponse-fromJSON}{\code{BinResponse$fromJSON()}}
\item \href{#method-BinResponse-clone}{\code{BinResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-BinResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinResponse$new(
  avgPredicted = NULL,
  endDate = NULL,
  frequency = NULL,
  maxPredicted = NULL,
  startDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{avgPredicted}}{Average prediction of the model in the bin. Null if there are no entries in the bin.}

\item{\code{endDate}}{ISO-formatted datetime of the end of the bin (exclusive).}

\item{\code{frequency}}{Number of the rows in the bin.}

\item{\code{maxPredicted}}{Maximum prediction of the model in the bin. Null if there are no entries in the bin.}

\item{\code{startDate}}{ISO-formatted datetime of the start of the bin (inclusive).}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BinResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BinResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BinResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinResponse$fromJSON(BinResponseJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BinResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BinResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
