% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blender_create.R
\docType{class}
\name{BlenderCreate}
\alias{BlenderCreate}
\title{BlenderCreate}
\format{
An \code{R6Class} generator object
}
\description{
BlenderCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{blenderMethod}}{character The blender method, one of \"PLS\", \"GLM\", \"AVG\", \"ENET\", \"MED\", \"MAE\", \"MAEL1\", \"TF\", \"RF\", \"LGBM\", \"FORECAST_DISTANCE_ENET\" (new in v2.18), \"FORECAST_DISTANCE_AVG\" (new in v2.18), \"MIN\", \"MAX\".}

\item{\code{modelIds}}{list( character ) The list of models to use in blender.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BlenderCreate-new}{\code{BlenderCreate$new()}}
\item \href{#method-BlenderCreate-validate}{\code{BlenderCreate$validate()}}
\item \href{#method-BlenderCreate-toJSON}{\code{BlenderCreate$toJSON()}}
\item \href{#method-BlenderCreate-fromJSON}{\code{BlenderCreate$fromJSON()}}
\item \href{#method-BlenderCreate-clone}{\code{BlenderCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlenderCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-BlenderCreate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlenderCreate$new(
  blenderMethod = NULL,
  modelIds = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blenderMethod}}{The blender method, one of \"PLS\", \"GLM\", \"AVG\", \"ENET\", \"MED\", \"MAE\", \"MAEL1\", \"TF\", \"RF\", \"LGBM\", \"FORECAST_DISTANCE_ENET\" (new in v2.18), \"FORECAST_DISTANCE_AVG\" (new in v2.18), \"MIN\", \"MAX\".}

\item{\code{modelIds}}{The list of models to use in blender.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlenderCreate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BlenderCreate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlenderCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlenderCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BlenderCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlenderCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlenderCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BlenderCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlenderCreate$fromJSON(BlenderCreateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BlenderCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlenderCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BlenderCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlenderCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
