% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blender_retrieve_response.R
\docType{class}
\name{BlenderRetrieveResponse}
\alias{BlenderRetrieveResponse}
\title{BlenderRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
BlenderRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{blenderMethod}}{character Method used to blend results of underlying models.}

\item{\code{blueprintId}}{character The blueprint used to construct the model - note this is not an ObjectId.}

\item{\code{featurelistId}}{character The ID of the featurelist used by the model. If null, the model was trained on multiple feature lists.}

\item{\code{featurelistName}}{character The name of the featurelist used by the model. If null, the model was trained on multiple feature lists.}

\item{\code{id}}{character The ID of the model.}

\item{\code{isFrozen}}{character Indicates whether the model is frozen, i.e. uses tuning parameters from a parent model}

\item{\code{isStarred}}{character Whether the model has been starred.}

\item{\code{linkFunction}}{character The link function the final modeler uses in the blueprint. If no link function exists, returns null.}

\item{\code{metrics}}{object The performance of the model according to various metrics.}

\item{\code{modelCategory}}{character Indicates what kind of model it is - will be `blend` for blenders.}

\item{\code{modelFamily}}{character The family model belongs to, e.g. SVM, GMB, etc.}

\item{\code{modelIds}}{list( character ) List of models used in blender.}

\item{\code{modelNumber}}{integer The model number from the leaderboard.}

\item{\code{modelType}}{character Identifies the model, e.g. `ENET Blender`.}

\item{\code{monotonicDecreasingFeaturelistId}}{character The ID of the featurelist that defines the set of features with a monotonically decreasing relationship to the target. If null, no such constraints are enforced.}

\item{\code{monotonicIncreasingFeaturelistId}}{character The ID of the featurelist that defines the set of features with a monotonically increasing relationship to the target. If null, no such constraints are enforced.}

\item{\code{parentModelId}}{character If this model is frozen, this is the ID of the parent model. Otherwise Null.}

\item{\code{predictionThreshold}}{numeric Threshold used for binary classification in predictions.}

\item{\code{predictionThresholdReadOnly}}{character Indicates whether modification of a predictions threshold is forbidden. Since v2.22 threshold modification is allowed.}

\item{\code{processes}}{list( character ) A list of processes used by the model.}

\item{\code{projectId}}{character The ID of the project to which the model belongs.}

\item{\code{samplePct}}{numeric The percentage of the dataset used in training the model.}

\item{\code{supportsComposableMl}}{character indicates whether this model is supported in Composable ML.}

\item{\code{supportsMonotonicConstraints}}{character Whether this model supports enforcing montonic constraints.}

\item{\code{trainingDuration}}{character The duration spanned by the dates in the partition column for the data used to train the model.}

\item{\code{trainingEndDate}}{character The end date of the dates in the partition column for the data used to train the model.}

\item{\code{trainingRowCount}}{integer The number of rows used to train the model.}

\item{\code{trainingStartDate}}{character The start date of the dates in the partition column for the data used to train the model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BlenderRetrieveResponse$new()}}
\item \href{#method-validate}{\code{BlenderRetrieveResponse$validate()}}
\item \href{#method-toJSON}{\code{BlenderRetrieveResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{BlenderRetrieveResponse$fromJSON()}}
\item \href{#method-clone}{\code{BlenderRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlenderRetrieveResponse$new(
  blenderMethod = NULL,
  blueprintId = NULL,
  featurelistId = NULL,
  featurelistName = NULL,
  id = NULL,
  isFrozen = NULL,
  isStarred = NULL,
  linkFunction = NULL,
  metrics = NULL,
  modelCategory = NULL,
  modelFamily = NULL,
  modelIds = NULL,
  modelNumber = NULL,
  modelType = NULL,
  monotonicDecreasingFeaturelistId = NULL,
  monotonicIncreasingFeaturelistId = NULL,
  parentModelId = NULL,
  predictionThreshold = NULL,
  predictionThresholdReadOnly = NULL,
  processes = NULL,
  projectId = NULL,
  samplePct = NULL,
  supportsComposableMl = NULL,
  supportsMonotonicConstraints = NULL,
  trainingDuration = NULL,
  trainingEndDate = NULL,
  trainingRowCount = NULL,
  trainingStartDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blenderMethod}}{Method used to blend results of underlying models.}

\item{\code{blueprintId}}{The blueprint used to construct the model - note this is not an ObjectId.}

\item{\code{featurelistId}}{The ID of the featurelist used by the model. If null, the model was trained on multiple feature lists.}

\item{\code{featurelistName}}{The name of the featurelist used by the model. If null, the model was trained on multiple feature lists.}

\item{\code{id}}{The ID of the model.}

\item{\code{isFrozen}}{Indicates whether the model is frozen, i.e. uses tuning parameters from a parent model}

\item{\code{isStarred}}{Whether the model has been starred.}

\item{\code{linkFunction}}{The link function the final modeler uses in the blueprint. If no link function exists, returns null.}

\item{\code{metrics}}{The performance of the model according to various metrics.}

\item{\code{modelCategory}}{Indicates what kind of model it is - will be `blend` for blenders.}

\item{\code{modelFamily}}{The family model belongs to, e.g. SVM, GMB, etc.}

\item{\code{modelIds}}{List of models used in blender.}

\item{\code{modelNumber}}{The model number from the leaderboard.}

\item{\code{modelType}}{Identifies the model, e.g. `ENET Blender`.}

\item{\code{monotonicDecreasingFeaturelistId}}{The ID of the featurelist that defines the set of features with a monotonically decreasing relationship to the target. If null, no such constraints are enforced.}

\item{\code{monotonicIncreasingFeaturelistId}}{The ID of the featurelist that defines the set of features with a monotonically increasing relationship to the target. If null, no such constraints are enforced.}

\item{\code{parentModelId}}{If this model is frozen, this is the ID of the parent model. Otherwise Null.}

\item{\code{predictionThreshold}}{Threshold used for binary classification in predictions.}

\item{\code{predictionThresholdReadOnly}}{Indicates whether modification of a predictions threshold is forbidden. Since v2.22 threshold modification is allowed.}

\item{\code{processes}}{A list of processes used by the model.}

\item{\code{projectId}}{The ID of the project to which the model belongs.}

\item{\code{samplePct}}{The percentage of the dataset used in training the model.}

\item{\code{supportsComposableMl}}{indicates whether this model is supported in Composable ML.}

\item{\code{supportsMonotonicConstraints}}{Whether this model supports enforcing montonic constraints.}

\item{\code{trainingDuration}}{The duration spanned by the dates in the partition column for the data used to train the model.}

\item{\code{trainingEndDate}}{The end date of the dates in the partition column for the data used to train the model.}

\item{\code{trainingRowCount}}{The number of rows used to train the model.}

\item{\code{trainingStartDate}}{The start date of the dates in the partition column for the data used to train the model.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlenderRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlenderRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlenderRetrieveResponse$fromJSON(
  BlenderRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BlenderRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlenderRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
