% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint_doc_parameters.R
\docType{class}
\name{BlueprintDocParameters}
\alias{BlueprintDocParameters}
\title{BlueprintDocParameters}
\format{
An \code{R6Class} generator object
}
\description{
BlueprintDocParameters Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{description}}{character A description of what the parameter does.}

\item{\code{name}}{character The name of the parameter.}

\item{\code{type}}{character The type (and default value) of the parameter.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BlueprintDocParameters-new}{\code{BlueprintDocParameters$new()}}
\item \href{#method-BlueprintDocParameters-validate}{\code{BlueprintDocParameters$validate()}}
\item \href{#method-BlueprintDocParameters-toJSON}{\code{BlueprintDocParameters$toJSON()}}
\item \href{#method-BlueprintDocParameters-fromJSON}{\code{BlueprintDocParameters$fromJSON()}}
\item \href{#method-BlueprintDocParameters-clone}{\code{BlueprintDocParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlueprintDocParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-BlueprintDocParameters-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlueprintDocParameters$new(
  description = NULL,
  name = NULL,
  type = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{A description of what the parameter does.}

\item{\code{name}}{The name of the parameter.}

\item{\code{type}}{The type (and default value) of the parameter.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlueprintDocParameters-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BlueprintDocParameters-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlueprintDocParameters$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlueprintDocParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BlueprintDocParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlueprintDocParameters$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlueprintDocParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BlueprintDocParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlueprintDocParameters$fromJSON(
  BlueprintDocParametersJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BlueprintDocParametersJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlueprintDocParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BlueprintDocParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlueprintDocParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
