% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint_doc_references.R
\docType{class}
\name{BlueprintDocReferences}
\alias{BlueprintDocReferences}
\title{BlueprintDocReferences}
\format{
An \code{R6Class} generator object
}
\description{
BlueprintDocReferences Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character The name of the reference.}

\item{\code{url}}{character The URL at which the reference can be found.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BlueprintDocReferences-new}{\code{BlueprintDocReferences$new()}}
\item \href{#method-BlueprintDocReferences-validate}{\code{BlueprintDocReferences$validate()}}
\item \href{#method-BlueprintDocReferences-toJSON}{\code{BlueprintDocReferences$toJSON()}}
\item \href{#method-BlueprintDocReferences-fromJSON}{\code{BlueprintDocReferences$fromJSON()}}
\item \href{#method-BlueprintDocReferences-clone}{\code{BlueprintDocReferences$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlueprintDocReferences-new"></a>}}
\if{latex}{\out{\hypertarget{method-BlueprintDocReferences-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlueprintDocReferences$new(
  name = NULL,
  url = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the reference.}

\item{\code{url}}{The URL at which the reference can be found.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlueprintDocReferences-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BlueprintDocReferences-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlueprintDocReferences$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlueprintDocReferences-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BlueprintDocReferences-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlueprintDocReferences$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlueprintDocReferences-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BlueprintDocReferences-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlueprintDocReferences$fromJSON(
  BlueprintDocReferencesJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BlueprintDocReferencesJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlueprintDocReferences-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BlueprintDocReferences-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlueprintDocReferences$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
