% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_catalog_append_tags_payload.R
\docType{class}
\name{BulkCatalogAppendTagsPayload}
\alias{BulkCatalogAppendTagsPayload}
\title{BulkCatalogAppendTagsPayload}
\format{
An \code{R6Class} generator object
}
\description{
BulkCatalogAppendTagsPayload Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{action}}{character The action to execute on the datasets. Has to be 'tag' for this payload.}

\item{\code{tags}}{list( character ) The tags to append to the datasets. Tags will not be duplicated.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BulkCatalogAppendTagsPayload-new}{\code{BulkCatalogAppendTagsPayload$new()}}
\item \href{#method-BulkCatalogAppendTagsPayload-validate}{\code{BulkCatalogAppendTagsPayload$validate()}}
\item \href{#method-BulkCatalogAppendTagsPayload-toJSON}{\code{BulkCatalogAppendTagsPayload$toJSON()}}
\item \href{#method-BulkCatalogAppendTagsPayload-fromJSON}{\code{BulkCatalogAppendTagsPayload$fromJSON()}}
\item \href{#method-BulkCatalogAppendTagsPayload-clone}{\code{BulkCatalogAppendTagsPayload$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BulkCatalogAppendTagsPayload-new"></a>}}
\if{latex}{\out{\hypertarget{method-BulkCatalogAppendTagsPayload-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BulkCatalogAppendTagsPayload$new(
  action = NULL,
  tags = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action}}{The action to execute on the datasets. Has to be 'tag' for this payload.}

\item{\code{tags}}{The tags to append to the datasets. Tags will not be duplicated.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BulkCatalogAppendTagsPayload-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BulkCatalogAppendTagsPayload-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BulkCatalogAppendTagsPayload$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BulkCatalogAppendTagsPayload-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BulkCatalogAppendTagsPayload-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BulkCatalogAppendTagsPayload$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BulkCatalogAppendTagsPayload-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BulkCatalogAppendTagsPayload-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BulkCatalogAppendTagsPayload$fromJSON(
  BulkCatalogAppendTagsPayloadJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BulkCatalogAppendTagsPayloadJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BulkCatalogAppendTagsPayload-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BulkCatalogAppendTagsPayload-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BulkCatalogAppendTagsPayload$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
