% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_event.R
\docType{class}
\name{CalendarEvent}
\alias{CalendarEvent}
\title{CalendarEvent}
\format{
An \code{R6Class} generator object
}
\description{
CalendarEvent Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{date}}{character The date of the calendar event.}

\item{\code{name}}{character Name of the calendar event.}

\item{\code{seriesId}}{character The series ID for the event. If this event does not specify a series ID, then this will be `null`, indicating that the event applies to all series.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CalendarEvent-new}{\code{CalendarEvent$new()}}
\item \href{#method-CalendarEvent-validate}{\code{CalendarEvent$validate()}}
\item \href{#method-CalendarEvent-toJSON}{\code{CalendarEvent$toJSON()}}
\item \href{#method-CalendarEvent-fromJSON}{\code{CalendarEvent$fromJSON()}}
\item \href{#method-CalendarEvent-clone}{\code{CalendarEvent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarEvent-new"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarEvent-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarEvent$new(
  date = NULL,
  name = NULL,
  seriesId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date}}{The date of the calendar event.}

\item{\code{name}}{Name of the calendar event.}

\item{\code{seriesId}}{The series ID for the event. If this event does not specify a series ID, then this will be `null`, indicating that the event applies to all series.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarEvent-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarEvent-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarEvent$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarEvent-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarEvent-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarEvent$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarEvent-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarEvent-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarEvent$fromJSON(CalendarEventJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CalendarEventJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarEvent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarEvent-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarEvent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
