% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_from_dataset.R
\docType{class}
\name{CalendarFromDataset}
\alias{CalendarFromDataset}
\title{CalendarFromDataset}
\format{
An \code{R6Class} generator object
}
\description{
CalendarFromDataset Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datasetId}}{character The ID of the dataset from which to create the calendar.}

\item{\code{datasetVersionId}}{character \link{optional} The ID of the dataset version from which to create the calendar.}

\item{\code{deleteOnError}}{character \link{optional} Whether delete calendar file from Catalog if it's not valid.}

\item{\code{multiseriesIdColumns}}{list( character ) \link{optional} Optional multiseries id columns for calendar.}

\item{\code{name}}{character \link{optional} Optional name for catalog.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CalendarFromDataset-new}{\code{CalendarFromDataset$new()}}
\item \href{#method-CalendarFromDataset-validate}{\code{CalendarFromDataset$validate()}}
\item \href{#method-CalendarFromDataset-toJSON}{\code{CalendarFromDataset$toJSON()}}
\item \href{#method-CalendarFromDataset-fromJSON}{\code{CalendarFromDataset$fromJSON()}}
\item \href{#method-CalendarFromDataset-clone}{\code{CalendarFromDataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarFromDataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarFromDataset-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarFromDataset$new(
  datasetId = NULL,
  datasetVersionId = NULL,
  deleteOnError = NULL,
  multiseriesIdColumns = NULL,
  name = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasetId}}{The ID of the dataset from which to create the calendar.}

\item{\code{datasetVersionId}}{The ID of the dataset version from which to create the calendar.}

\item{\code{deleteOnError}}{Whether delete calendar file from Catalog if it's not valid.}

\item{\code{multiseriesIdColumns}}{Optional multiseries id columns for calendar.}

\item{\code{name}}{Optional name for catalog.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarFromDataset-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarFromDataset-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarFromDataset$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarFromDataset-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarFromDataset-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarFromDataset$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarFromDataset-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarFromDataset-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarFromDataset$fromJSON(CalendarFromDatasetJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CalendarFromDatasetJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarFromDataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarFromDataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarFromDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
