% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_record.R
\docType{class}
\name{CalendarRecord}
\alias{CalendarRecord}
\title{CalendarRecord}
\format{
An \code{R6Class} generator object
}
\description{
CalendarRecord Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{created}}{character An ISO-8601 string with the time that this calendar was created.}

\item{\code{datetimeFormat}}{character The datetime format detected for the uploaded calendar file.}

\item{\code{earliestEvent}}{character An ISO-8601 date string of the earliest event seen in this calendar.}

\item{\code{id}}{character The ID of this calendar.}

\item{\code{latestEvent}}{character An ISO-8601 date string of the latest event seen in this calendar.}

\item{\code{multiseriesIdColumns}}{list( character ) An array of multiseries ID column names in this calendar file. Currently only one multiseries ID column is supported. Will be `null` if this calendar is single-series.}

\item{\code{name}}{character The name of this calendar. This will be the same as `source` if no name was specified when the calendar was created.}

\item{\code{numEventTypes}}{integer The number of distinct eventTypes in this calendar.}

\item{\code{numEvents}}{integer The number of dates that are marked as having an event in this calendar.}

\item{\code{projectId}}{list( character ) The project IDs of projects currently using this calendar.}

\item{\code{role}}{character The role the requesting user has on this calendar.}

\item{\code{source}}{character The name of the source file that was used to create this calendar.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CalendarRecord-new}{\code{CalendarRecord$new()}}
\item \href{#method-CalendarRecord-validate}{\code{CalendarRecord$validate()}}
\item \href{#method-CalendarRecord-toJSON}{\code{CalendarRecord$toJSON()}}
\item \href{#method-CalendarRecord-fromJSON}{\code{CalendarRecord$fromJSON()}}
\item \href{#method-CalendarRecord-clone}{\code{CalendarRecord$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarRecord-new"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarRecord-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarRecord$new(
  created = NULL,
  datetimeFormat = NULL,
  earliestEvent = NULL,
  id = NULL,
  latestEvent = NULL,
  multiseriesIdColumns = NULL,
  name = NULL,
  numEventTypes = NULL,
  numEvents = NULL,
  projectId = NULL,
  role = NULL,
  source = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{created}}{An ISO-8601 string with the time that this calendar was created.}

\item{\code{datetimeFormat}}{The datetime format detected for the uploaded calendar file.}

\item{\code{earliestEvent}}{An ISO-8601 date string of the earliest event seen in this calendar.}

\item{\code{id}}{The ID of this calendar.}

\item{\code{latestEvent}}{An ISO-8601 date string of the latest event seen in this calendar.}

\item{\code{multiseriesIdColumns}}{An array of multiseries ID column names in this calendar file. Currently only one multiseries ID column is supported. Will be `null` if this calendar is single-series.}

\item{\code{name}}{The name of this calendar. This will be the same as `source` if no name was specified when the calendar was created.}

\item{\code{numEventTypes}}{The number of distinct eventTypes in this calendar.}

\item{\code{numEvents}}{The number of dates that are marked as having an event in this calendar.}

\item{\code{projectId}}{The project IDs of projects currently using this calendar.}

\item{\code{role}}{The role the requesting user has on this calendar.}

\item{\code{source}}{The name of the source file that was used to create this calendar.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarRecord-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarRecord-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarRecord$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarRecord-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarRecord-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarRecord$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarRecord-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarRecord-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarRecord$fromJSON(CalendarRecordJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CalendarRecordJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarRecord-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarRecord-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarRecord$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
