% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_username_role.R
\docType{class}
\name{CalendarUsernameRole}
\alias{CalendarUsernameRole}
\title{CalendarUsernameRole}
\format{
An \code{R6Class} generator object
}
\description{
CalendarUsernameRole Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{role}}{character The new role to assign to the specified user.}

\item{\code{username}}{character The username of the user to modify access for.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CalendarUsernameRole-new}{\code{CalendarUsernameRole$new()}}
\item \href{#method-CalendarUsernameRole-validate}{\code{CalendarUsernameRole$validate()}}
\item \href{#method-CalendarUsernameRole-toJSON}{\code{CalendarUsernameRole$toJSON()}}
\item \href{#method-CalendarUsernameRole-fromJSON}{\code{CalendarUsernameRole$fromJSON()}}
\item \href{#method-CalendarUsernameRole-clone}{\code{CalendarUsernameRole$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarUsernameRole-new"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarUsernameRole-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarUsernameRole$new(
  role = NULL,
  username = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{The new role to assign to the specified user.}

\item{\code{username}}{The username of the user to modify access for.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarUsernameRole-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarUsernameRole-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarUsernameRole$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarUsernameRole-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarUsernameRole-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarUsernameRole$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarUsernameRole-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarUsernameRole-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarUsernameRole$fromJSON(CalendarUsernameRoleJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CalendarUsernameRoleJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarUsernameRole-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarUsernameRole-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarUsernameRole$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
