% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\docType{class}
\name{Catalog}
\alias{Catalog}
\title{Catalog}
\format{
An \code{R6Class} generator object
}
\description{
Catalog Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datasetId}}{character The ID of the AI catalog dataset}

\item{\code{datasetVersionId}}{character \link{optional} The ID of the AI catalog dataset version}

\item{\code{type}}{character Type name for this intake type}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Catalog-new}{\code{Catalog$new()}}
\item \href{#method-Catalog-validate}{\code{Catalog$validate()}}
\item \href{#method-Catalog-toJSON}{\code{Catalog$toJSON()}}
\item \href{#method-Catalog-fromJSON}{\code{Catalog$fromJSON()}}
\item \href{#method-Catalog-clone}{\code{Catalog$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Catalog-new"></a>}}
\if{latex}{\out{\hypertarget{method-Catalog-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Catalog$new(
  datasetId = NULL,
  type = NULL,
  datasetVersionId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasetId}}{The ID of the AI catalog dataset}

\item{\code{type}}{Type name for this intake type}

\item{\code{datasetVersionId}}{The ID of the AI catalog dataset version}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Catalog-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Catalog-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Catalog$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Catalog-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Catalog-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Catalog$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Catalog-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Catalog-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Catalog$fromJSON(CatalogJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CatalogJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Catalog-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Catalog-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Catalog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
