% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_details_retrieve_response.R
\docType{class}
\name{CatalogDetailsRetrieveResponse}
\alias{CatalogDetailsRetrieveResponse}
\title{CatalogDetailsRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
CatalogDetailsRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{createdAt}}{character The ISO 8601-formatted date and time indicating when this item was created in the catalog.}

\item{\code{createdBy}}{character The full name or username of the user who added this item to the catalog.}

\item{\code{description}}{character Catalog item description.}

\item{\code{id}}{character Catalog item ID.}

\item{\code{message}}{character Details of exception(s) raised during ingestion process, if any.}

\item{\code{modifiedAt}}{character The ISO 8601-formatted date and time indicating changes to the Info field(s) of this catalog item.}

\item{\code{modifiedBy}}{character The full name or username of the user who last modified the Info field(s) of this catalog item.}

\item{\code{name}}{character Catalog item name.}

\item{\code{status}}{character For datasets, the current ingestion process state of this catalog item.}

\item{\code{tags}}{list( character ) List of catalog item tags in the lower case with no spaces.}

\item{\code{type}}{character Catalog item type.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CatalogDetailsRetrieveResponse$new()}}
\item \href{#method-validate}{\code{CatalogDetailsRetrieveResponse$validate()}}
\item \href{#method-toJSON}{\code{CatalogDetailsRetrieveResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{CatalogDetailsRetrieveResponse$fromJSON()}}
\item \href{#method-clone}{\code{CatalogDetailsRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogDetailsRetrieveResponse$new(
  createdAt = NULL,
  createdBy = NULL,
  description = NULL,
  id = NULL,
  message = NULL,
  modifiedAt = NULL,
  modifiedBy = NULL,
  name = NULL,
  status = NULL,
  tags = NULL,
  type = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{createdAt}}{The ISO 8601-formatted date and time indicating when this item was created in the catalog.}

\item{\code{createdBy}}{The full name or username of the user who added this item to the catalog.}

\item{\code{description}}{Catalog item description.}

\item{\code{id}}{Catalog item ID.}

\item{\code{message}}{Details of exception(s) raised during ingestion process, if any.}

\item{\code{modifiedAt}}{The ISO 8601-formatted date and time indicating changes to the Info field(s) of this catalog item.}

\item{\code{modifiedBy}}{The full name or username of the user who last modified the Info field(s) of this catalog item.}

\item{\code{name}}{Catalog item name.}

\item{\code{status}}{For datasets, the current ingestion process state of this catalog item.}

\item{\code{tags}}{List of catalog item tags in the lower case with no spaces.}

\item{\code{type}}{Catalog item type.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogDetailsRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogDetailsRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogDetailsRetrieveResponse$fromJSON(
  CatalogDetailsRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CatalogDetailsRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogDetailsRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
