% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_list_search_response.R
\docType{class}
\name{CatalogListSearchResponse}
\alias{CatalogListSearchResponse}
\title{CatalogListSearchResponse}
\format{
An \code{R6Class} generator object
}
\description{
CatalogListSearchResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cacheHit}}{character Indicates if the catalog item is returned from the cache.}

\item{\code{count}}{integer Number of catalog items returned on this page.}

\item{\code{data}}{list( \link{CatalogExtendedDetailsResponse} ) Detailed information for every found catalog item.}

\item{\code{next_}}{character Location of the next page.}

\item{\code{previous}}{character Location of the previous page.}

\item{\code{totalCount}}{integer \link{optional} Total number of catalog items.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CatalogListSearchResponse$new()}}
\item \href{#method-validate}{\code{CatalogListSearchResponse$validate()}}
\item \href{#method-toJSON}{\code{CatalogListSearchResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{CatalogListSearchResponse$fromJSON()}}
\item \href{#method-clone}{\code{CatalogListSearchResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogListSearchResponse$new(
  cacheHit = NULL,
  count = NULL,
  data = NULL,
  next_ = NULL,
  previous = NULL,
  totalCount = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cacheHit}}{Indicates if the catalog item is returned from the cache.}

\item{\code{count}}{Number of catalog items returned on this page.}

\item{\code{data}}{Detailed information for every found catalog item.}

\item{\code{next_}}{Location of the next page.}

\item{\code{previous}}{Location of the previous page.}

\item{\code{totalCount}}{Total number of catalog items.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogListSearchResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogListSearchResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogListSearchResponse$fromJSON(
  CatalogListSearchResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CatalogListSearchResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatalogListSearchResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
