% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/challenger_score.R
\docType{class}
\name{ChallengerScore}
\alias{ChallengerScore}
\title{ChallengerScore}
\format{
An \code{R6Class} generator object
}
\description{
ChallengerScore Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{timestamp}}{character \link{optional} The date and time in ISO8601 format, challenger models will be scored on data starting from deployment creation until timestamp. If not specified UTC current time is used.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChallengerScore-new}{\code{ChallengerScore$new()}}
\item \href{#method-ChallengerScore-validate}{\code{ChallengerScore$validate()}}
\item \href{#method-ChallengerScore-toJSON}{\code{ChallengerScore$toJSON()}}
\item \href{#method-ChallengerScore-fromJSON}{\code{ChallengerScore$fromJSON()}}
\item \href{#method-ChallengerScore-clone}{\code{ChallengerScore$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChallengerScore-new"></a>}}
\if{latex}{\out{\hypertarget{method-ChallengerScore-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChallengerScore$new(timestamp = NULL, validateParams = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timestamp}}{The date and time in ISO8601 format, challenger models will be scored on data starting from deployment creation until timestamp. If not specified UTC current time is used.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChallengerScore-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ChallengerScore-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChallengerScore$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChallengerScore-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ChallengerScore-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChallengerScore$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChallengerScore-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ChallengerScore-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChallengerScore$fromJSON(ChallengerScoreJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ChallengerScoreJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChallengerScore-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChallengerScore-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChallengerScore$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
