% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_request_create_one_of5.R
\docType{class}
\name{ChangeRequestCreateOneOf5}
\alias{ChangeRequestCreateOneOf5}
\title{ChangeRequestCreateOneOf5}
\format{
An \code{R6Class} generator object
}
\description{
ChangeRequestCreateOneOf5 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{action}}{character Actions the user can take on the entity. Each entity type has a specific set of actions.}

\item{\code{autoApply}}{character Whether to automatically apply the change when the request is approved. If `true`, the requested changes will be applied on approval.}

\item{\code{change}}{\link{DeploymentReplaceModelChange}}

\item{\code{comment}}{character \link{optional} Free form text to comment on the requested changes.}

\item{\code{entityId}}{character ID of the Product Entity the request is intended to change.}

\item{\code{entityType}}{character Type of the Product Entity that is requested to be changed.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ChangeRequestCreateOneOf5$new()}}
\item \href{#method-validate}{\code{ChangeRequestCreateOneOf5$validate()}}
\item \href{#method-toJSON}{\code{ChangeRequestCreateOneOf5$toJSON()}}
\item \href{#method-fromJSON}{\code{ChangeRequestCreateOneOf5$fromJSON()}}
\item \href{#method-clone}{\code{ChangeRequestCreateOneOf5$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangeRequestCreateOneOf5$new(
  action = NULL,
  autoApply = NULL,
  change = NULL,
  entityId = NULL,
  entityType = NULL,
  comment = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action}}{Actions the user can take on the entity. Each entity type has a specific set of actions.}

\item{\code{autoApply}}{Whether to automatically apply the change when the request is approved. If `true`, the requested changes will be applied on approval.}

\item{\code{entityId}}{ID of the Product Entity the request is intended to change.}

\item{\code{entityType}}{Type of the Product Entity that is requested to be changed.}

\item{\code{comment}}{Free form text to comment on the requested changes.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangeRequestCreateOneOf5$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangeRequestCreateOneOf5$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangeRequestCreateOneOf5$fromJSON(
  ChangeRequestCreateOneOf5Json,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ChangeRequestCreateOneOf5Json}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangeRequestCreateOneOf5$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
