% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_request_diff.R
\docType{class}
\name{ChangeRequestDiff}
\alias{ChangeRequestDiff}
\title{ChangeRequestDiff}
\format{
An \code{R6Class} generator object
}
\description{
ChangeRequestDiff Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{changesFrom}}{list( character ) List of human readable messages describing the state of the entity before changes are applied.}

\item{\code{changesTo}}{list( character ) List of human readable messages describing the state of the entity after changes are applied.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChangeRequestDiff-new}{\code{ChangeRequestDiff$new()}}
\item \href{#method-ChangeRequestDiff-validate}{\code{ChangeRequestDiff$validate()}}
\item \href{#method-ChangeRequestDiff-toJSON}{\code{ChangeRequestDiff$toJSON()}}
\item \href{#method-ChangeRequestDiff-fromJSON}{\code{ChangeRequestDiff$fromJSON()}}
\item \href{#method-ChangeRequestDiff-clone}{\code{ChangeRequestDiff$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChangeRequestDiff-new"></a>}}
\if{latex}{\out{\hypertarget{method-ChangeRequestDiff-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangeRequestDiff$new(
  changesFrom = NULL,
  changesTo = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{changesFrom}}{List of human readable messages describing the state of the entity before changes are applied.}

\item{\code{changesTo}}{List of human readable messages describing the state of the entity after changes are applied.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChangeRequestDiff-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ChangeRequestDiff-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangeRequestDiff$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChangeRequestDiff-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ChangeRequestDiff-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangeRequestDiff$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChangeRequestDiff-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ChangeRequestDiff-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangeRequestDiff$fromJSON(ChangeRequestDiffJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ChangeRequestDiffJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChangeRequestDiff-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChangeRequestDiff-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangeRequestDiff$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
