% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_request_info_response.R
\docType{class}
\name{ChangeRequestInfoResponse}
\alias{ChangeRequestInfoResponse}
\title{ChangeRequestInfoResponse}
\format{
An \code{R6Class} generator object
}
\description{
ChangeRequestInfoResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{changeRequestId}}{character ID of the Change Request.}

\item{\code{createDate}}{character Change Request creation date.}

\item{\code{entityId}}{character ID of the modified entity.}

\item{\code{entityName}}{character Name of the modified entity.}

\item{\code{requester}}{character Username of the account that initiated a Change Request.}

\item{\code{state}}{character Status of the Change Request.}

\item{\code{updateDate}}{character Last date when Change Request was modified.}

\item{\code{updatedBy}}{character Username of the account that last updated the Change Request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChangeRequestInfoResponse-new}{\code{ChangeRequestInfoResponse$new()}}
\item \href{#method-ChangeRequestInfoResponse-validate}{\code{ChangeRequestInfoResponse$validate()}}
\item \href{#method-ChangeRequestInfoResponse-toJSON}{\code{ChangeRequestInfoResponse$toJSON()}}
\item \href{#method-ChangeRequestInfoResponse-fromJSON}{\code{ChangeRequestInfoResponse$fromJSON()}}
\item \href{#method-ChangeRequestInfoResponse-clone}{\code{ChangeRequestInfoResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChangeRequestInfoResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ChangeRequestInfoResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangeRequestInfoResponse$new(
  changeRequestId = NULL,
  createDate = NULL,
  entityId = NULL,
  entityName = NULL,
  requester = NULL,
  state = NULL,
  updateDate = NULL,
  updatedBy = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{changeRequestId}}{ID of the Change Request.}

\item{\code{createDate}}{Change Request creation date.}

\item{\code{entityId}}{ID of the modified entity.}

\item{\code{entityName}}{Name of the modified entity.}

\item{\code{requester}}{Username of the account that initiated a Change Request.}

\item{\code{state}}{Status of the Change Request.}

\item{\code{updateDate}}{Last date when Change Request was modified.}

\item{\code{updatedBy}}{Username of the account that last updated the Change Request.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChangeRequestInfoResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ChangeRequestInfoResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangeRequestInfoResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChangeRequestInfoResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ChangeRequestInfoResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangeRequestInfoResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChangeRequestInfoResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ChangeRequestInfoResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangeRequestInfoResponse$fromJSON(
  ChangeRequestInfoResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ChangeRequestInfoResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChangeRequestInfoResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChangeRequestInfoResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangeRequestInfoResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
