% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_request_response.R
\docType{class}
\name{ChangeRequestResponse}
\alias{ChangeRequestResponse}
\title{ChangeRequestResponse}
\format{
An \code{R6Class} generator object
}
\description{
ChangeRequestResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{action}}{character Actions the user can take on the entity. Each entity type has a specific set of actions.}

\item{\code{autoApply}}{character Whether to automatically apply the change when the request is approved. If `true`, the requested changes will be applied on approval.}

\item{\code{change}}{\link{OneOfDeploymentApproveChangeDeploymentChangeStatusChangeDeploymentChangeImportanceChangeDeploymentCleanupStatsChangeDeploymentReplaceModelChangeDeploymentReplaceModelPackageChangeDeploymentUpdateSecondaryDatasetConfigChange} Change that the user wants to apply to the entity. Needs to be provided if the action, like `approve` action for a deployment, requires additional parameters . `null` if the action does not require any additional parameters to be applied.}

\item{\code{changeVersionId}}{character ID of the current version of change within this Change Request. It's possible to modify the changes that have been requested. At the same time, we need to make sure that review is associated with the correct changes, that's why we implement versioning on the changes and associate user reviews with the specific Change Request versions.}

\item{\code{comment}}{character \link{optional} Free form text to comment on the requested changes.}

\item{\code{createdAt}}{character Timestamp when the request was created.}

\item{\code{diff}}{\link{ChangeRequestDiff}}

\item{\code{entityId}}{character ID of the Product Entity the request is intended to change.}

\item{\code{entityType}}{character Type of the Product Entity that is requested to be changed.}

\item{\code{id}}{character ID of the Change Request.}

\item{\code{numApprovalsRequired}}{integer Number of approving reviews required for the Change Request to be considered approved.}

\item{\code{processedAt}}{character Timestamp when the request was processed.}

\item{\code{status}}{character Change Request Status.}

\item{\code{statusChangedAt}}{character Timestamp when the current request status was set. `null` if status is set by the system.}

\item{\code{statusChangedBy}}{character ID of the user who set the current request status. `null` if the status was set by the system.}

\item{\code{updatedAt}}{character Timestamp when the request was last modified.}

\item{\code{userId}}{character ID of the user, who created the Change Request.}

\item{\code{userName}}{character Email of the user, who created the Change Request}

\item{\code{userOperations}}{\link{UserOperations}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ChangeRequestResponse$new()}}
\item \href{#method-validate}{\code{ChangeRequestResponse$validate()}}
\item \href{#method-toJSON}{\code{ChangeRequestResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{ChangeRequestResponse$fromJSON()}}
\item \href{#method-clone}{\code{ChangeRequestResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangeRequestResponse$new(
  action = NULL,
  autoApply = NULL,
  change = NULL,
  changeVersionId = NULL,
  createdAt = NULL,
  diff = NULL,
  entityId = NULL,
  entityType = NULL,
  id = NULL,
  numApprovalsRequired = NULL,
  processedAt = NULL,
  status = NULL,
  statusChangedAt = NULL,
  statusChangedBy = NULL,
  updatedAt = NULL,
  userId = NULL,
  userName = NULL,
  userOperations = NULL,
  comment = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action}}{Actions the user can take on the entity. Each entity type has a specific set of actions.}

\item{\code{autoApply}}{Whether to automatically apply the change when the request is approved. If `true`, the requested changes will be applied on approval.}

\item{\code{change}}{Change that the user wants to apply to the entity. Needs to be provided if the action, like `approve` action for a deployment, requires additional parameters . `null` if the action does not require any additional parameters to be applied.}

\item{\code{changeVersionId}}{ID of the current version of change within this Change Request. It's possible to modify the changes that have been requested. At the same time, we need to make sure that review is associated with the correct changes, that's why we implement versioning on the changes and associate user reviews with the specific Change Request versions.}

\item{\code{createdAt}}{Timestamp when the request was created.}

\item{\code{entityId}}{ID of the Product Entity the request is intended to change.}

\item{\code{entityType}}{Type of the Product Entity that is requested to be changed.}

\item{\code{id}}{ID of the Change Request.}

\item{\code{numApprovalsRequired}}{Number of approving reviews required for the Change Request to be considered approved.}

\item{\code{processedAt}}{Timestamp when the request was processed.}

\item{\code{status}}{Change Request Status.}

\item{\code{statusChangedAt}}{Timestamp when the current request status was set. `null` if status is set by the system.}

\item{\code{statusChangedBy}}{ID of the user who set the current request status. `null` if the status was set by the system.}

\item{\code{updatedAt}}{Timestamp when the request was last modified.}

\item{\code{userId}}{ID of the user, who created the Change Request.}

\item{\code{userName}}{Email of the user, who created the Change Request}

\item{\code{comment}}{Free form text to comment on the requested changes.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangeRequestResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangeRequestResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangeRequestResponse$fromJSON(
  ChangeRequestResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ChangeRequestResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChangeRequestResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
