% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mapping_aggregation_settings.R
\docType{class}
\name{ClassMappingAggregationSettings}
\alias{ClassMappingAggregationSettings}
\title{ClassMappingAggregationSettings}
\format{
An \code{R6Class} generator object
}
\description{
ClassMappingAggregationSettings Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aggregationClassName}}{character \link{optional} The name of the class that will be assigned to all rows with aggregated classes. Should not match any excluded_from_aggregation or we will have 2 classes with the same name and no way to distinguish between them. This option is only available formulticlass projects. By default 'DR_RARE_TARGET_VALUES' is used.}

\item{\code{excludedFromAggregation}}{list( character ) \link{optional} List of target values that should be guaranteed to kept as is, regardless of other settings.}

\item{\code{maxUnaggregatedClassValues}}{integer \link{optional} The maximum number of unique labels before aggregation kicks in. Should be at least len(excludedFromAggregation) + 1 for multiclass and at least len(excludedFromAggregation) for multilabel.}

\item{\code{minClassSupport}}{integer \link{optional} Minimum number of instances necessary for each target value in the dataset. All values with fewer instances than this value will be aggregated}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClassMappingAggregationSettings-new}{\code{ClassMappingAggregationSettings$new()}}
\item \href{#method-ClassMappingAggregationSettings-validate}{\code{ClassMappingAggregationSettings$validate()}}
\item \href{#method-ClassMappingAggregationSettings-toJSON}{\code{ClassMappingAggregationSettings$toJSON()}}
\item \href{#method-ClassMappingAggregationSettings-fromJSON}{\code{ClassMappingAggregationSettings$fromJSON()}}
\item \href{#method-ClassMappingAggregationSettings-clone}{\code{ClassMappingAggregationSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassMappingAggregationSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClassMappingAggregationSettings-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassMappingAggregationSettings$new(
  aggregationClassName = NULL,
  excludedFromAggregation = c(),
  maxUnaggregatedClassValues = 1000,
  minClassSupport = 1,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aggregationClassName}}{The name of the class that will be assigned to all rows with aggregated classes. Should not match any excluded_from_aggregation or we will have 2 classes with the same name and no way to distinguish between them. This option is only available formulticlass projects. By default 'DR_RARE_TARGET_VALUES' is used.}

\item{\code{excludedFromAggregation}}{List of target values that should be guaranteed to kept as is, regardless of other settings.}

\item{\code{maxUnaggregatedClassValues}}{The maximum number of unique labels before aggregation kicks in. Should be at least len(excludedFromAggregation) + 1 for multiclass and at least len(excludedFromAggregation) for multilabel.}

\item{\code{minClassSupport}}{Minimum number of instances necessary for each target value in the dataset. All values with fewer instances than this value will be aggregated}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassMappingAggregationSettings-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ClassMappingAggregationSettings-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassMappingAggregationSettings$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassMappingAggregationSettings-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ClassMappingAggregationSettings-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassMappingAggregationSettings$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassMappingAggregationSettings-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ClassMappingAggregationSettings-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassMappingAggregationSettings$fromJSON(
  ClassMappingAggregationSettingsJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ClassMappingAggregationSettingsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassMappingAggregationSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClassMappingAggregationSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassMappingAggregationSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
