% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{CloneProject}
\alias{CloneProject}
\title{CloneProject}
\usage{
CloneProject(project, newProjectName = NULL, maxWait = 600, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Clone a project
}
\details{
Create a clone of an existing project. > > The resultant project will begin the initial exploratory data analysis > and will be ready to set the target of the new project shortly.

This method invokes \verb{POST /projectClones/} in the DataRobot Public API.
}
\seealso{
The method ClonesCreate in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
