% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_license_retrieve_response_upload_info.R
\docType{class}
\name{ClusterLicenseRetrieveResponseUploadInfo}
\alias{ClusterLicenseRetrieveResponseUploadInfo}
\title{ClusterLicenseRetrieveResponseUploadInfo}
\format{
An \code{R6Class} generator object
}
\description{
ClusterLicenseRetrieveResponseUploadInfo Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uploadTimestamp}}{character the time when the current license was uploaded in UTC ISO format, ex. '2020-12-21T23:59:59.000000Z'}

\item{\code{uploaderUserId}}{character Id of the user who uploaded the current license}

\item{\code{uploaderUsername}}{character The username of the user who uploaded the current license}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClusterLicenseRetrieveResponseUploadInfo-new}{\code{ClusterLicenseRetrieveResponseUploadInfo$new()}}
\item \href{#method-ClusterLicenseRetrieveResponseUploadInfo-validate}{\code{ClusterLicenseRetrieveResponseUploadInfo$validate()}}
\item \href{#method-ClusterLicenseRetrieveResponseUploadInfo-toJSON}{\code{ClusterLicenseRetrieveResponseUploadInfo$toJSON()}}
\item \href{#method-ClusterLicenseRetrieveResponseUploadInfo-fromJSON}{\code{ClusterLicenseRetrieveResponseUploadInfo$fromJSON()}}
\item \href{#method-ClusterLicenseRetrieveResponseUploadInfo-clone}{\code{ClusterLicenseRetrieveResponseUploadInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusterLicenseRetrieveResponseUploadInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClusterLicenseRetrieveResponseUploadInfo-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusterLicenseRetrieveResponseUploadInfo$new(
  uploadTimestamp = NULL,
  uploaderUserId = NULL,
  uploaderUsername = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uploadTimestamp}}{the time when the current license was uploaded in UTC ISO format, ex. '2020-12-21T23:59:59.000000Z'}

\item{\code{uploaderUserId}}{Id of the user who uploaded the current license}

\item{\code{uploaderUsername}}{The username of the user who uploaded the current license}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusterLicenseRetrieveResponseUploadInfo-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ClusterLicenseRetrieveResponseUploadInfo-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusterLicenseRetrieveResponseUploadInfo$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusterLicenseRetrieveResponseUploadInfo-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ClusterLicenseRetrieveResponseUploadInfo-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusterLicenseRetrieveResponseUploadInfo$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusterLicenseRetrieveResponseUploadInfo-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ClusterLicenseRetrieveResponseUploadInfo-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusterLicenseRetrieveResponseUploadInfo$fromJSON(
  ClusterLicenseRetrieveResponseUploadInfoJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ClusterLicenseRetrieveResponseUploadInfoJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusterLicenseRetrieveResponseUploadInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClusterLicenseRetrieveResponseUploadInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusterLicenseRetrieveResponseUploadInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
