% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_model_segments_response.R
\docType{class}
\name{CombinedModelSegmentsResponse}
\alias{CombinedModelSegmentsResponse}
\title{CombinedModelSegmentsResponse}
\format{
An \code{R6Class} generator object
}
\description{
CombinedModelSegmentsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{autopilotDone}}{character Is autopilot done for the project.}

\item{\code{holdoutUnlocked}}{character Is holdout unlocked for the project.}

\item{\code{isFrozen}}{character Indicates whether the segment champion model is frozen, i.e. uses tuning parameters from a parent model}

\item{\code{modelAssignedBy}}{character Who assigned model as segment champion. Default is ``DataRobot``.}

\item{\code{modelAwardTime}}{character Time when model was awarded as segment champion.}

\item{\code{modelCount}}{integer Count of trained models in project.}

\item{\code{modelIcon}}{list( integer ) The number for the icon representing the given champion model.}

\item{\code{modelId}}{character ID of segment champion model.}

\item{\code{modelMetrics}}{object The performance of the model according to various metrics, where each metric has validation, crossValidation, holdout, and training scores reported, or ``null`` if they have not been computed.}

\item{\code{modelType}}{character The description of the model type of the given champion model.}

\item{\code{projectId}}{character The ID of the project.}

\item{\code{projectPaused}}{character \link{optional} Is project paused right now.}

\item{\code{projectStage}}{character The current stage of the project, where modeling indicates that the target has been successfully set and modeling and predictions may proceed.}

\item{\code{projectStageDescription}}{character A description of the current stage of the project.}

\item{\code{projectStatusError}}{character \link{optional} Project status error message.}

\item{\code{rowCount}}{integer Count of rows in project's dataset.}

\item{\code{rowPercentage}}{numeric Percentage of rows in segment project's dataset comparing to original dataset.}

\item{\code{segment}}{character Segment name.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CombinedModelSegmentsResponse$new()}}
\item \href{#method-validate}{\code{CombinedModelSegmentsResponse$validate()}}
\item \href{#method-toJSON}{\code{CombinedModelSegmentsResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{CombinedModelSegmentsResponse$fromJSON()}}
\item \href{#method-clone}{\code{CombinedModelSegmentsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedModelSegmentsResponse$new(
  autopilotDone = NULL,
  holdoutUnlocked = NULL,
  isFrozen = NULL,
  modelAssignedBy = NULL,
  modelAwardTime = NULL,
  modelCount = NULL,
  modelIcon = NULL,
  modelId = NULL,
  modelMetrics = NULL,
  modelType = NULL,
  projectId = NULL,
  projectStage = NULL,
  projectStageDescription = NULL,
  rowCount = NULL,
  rowPercentage = NULL,
  segment = NULL,
  projectPaused = NULL,
  projectStatusError = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{autopilotDone}}{Is autopilot done for the project.}

\item{\code{holdoutUnlocked}}{Is holdout unlocked for the project.}

\item{\code{isFrozen}}{Indicates whether the segment champion model is frozen, i.e. uses tuning parameters from a parent model}

\item{\code{modelAssignedBy}}{Who assigned model as segment champion. Default is ``DataRobot``.}

\item{\code{modelAwardTime}}{Time when model was awarded as segment champion.}

\item{\code{modelCount}}{Count of trained models in project.}

\item{\code{modelIcon}}{The number for the icon representing the given champion model.}

\item{\code{modelId}}{ID of segment champion model.}

\item{\code{modelMetrics}}{The performance of the model according to various metrics, where each metric has validation, crossValidation, holdout, and training scores reported, or ``null`` if they have not been computed.}

\item{\code{modelType}}{The description of the model type of the given champion model.}

\item{\code{projectId}}{The ID of the project.}

\item{\code{projectStage}}{The current stage of the project, where modeling indicates that the target has been successfully set and modeling and predictions may proceed.}

\item{\code{projectStageDescription}}{A description of the current stage of the project.}

\item{\code{rowCount}}{Count of rows in project's dataset.}

\item{\code{rowPercentage}}{Percentage of rows in segment project's dataset comparing to original dataset.}

\item{\code{segment}}{Segment name.}

\item{\code{projectPaused}}{Is project paused right now.}

\item{\code{projectStatusError}}{Project status error message.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedModelSegmentsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedModelSegmentsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedModelSegmentsResponse$fromJSON(
  CombinedModelSegmentsResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CombinedModelSegmentsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedModelSegmentsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
