% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_retrieve.R
\docType{class}
\name{CommentRetrieve}
\alias{CommentRetrieve}
\title{CommentRetrieve}
\format{
An \code{R6Class} generator object
}
\description{
CommentRetrieve Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{content}}{character Content of the comment}

\item{\code{createdAt}}{character Timestamp when the comment was created}

\item{\code{createdBy}}{\link{CommentUser}}

\item{\code{entityId}}{character ID of the entity the comment posted to}

\item{\code{entityType}}{character Type of the entity to post the comment to, currently only useCase is supported}

\item{\code{id}}{character ID of the comment}

\item{\code{mentions}}{list( \link{CommentUser} ) A list of users objects (see below) mentioned in the content}

\item{\code{updatedAt}}{character Timestamp when the comment was updated}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CommentRetrieve$new()}}
\item \href{#method-validate}{\code{CommentRetrieve$validate()}}
\item \href{#method-toJSON}{\code{CommentRetrieve$toJSON()}}
\item \href{#method-fromJSON}{\code{CommentRetrieve$fromJSON()}}
\item \href{#method-clone}{\code{CommentRetrieve$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentRetrieve$new(
  content = NULL,
  createdAt = NULL,
  createdBy = NULL,
  entityId = NULL,
  entityType = NULL,
  id = NULL,
  mentions = NULL,
  updatedAt = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{Content of the comment}

\item{\code{createdAt}}{Timestamp when the comment was created}

\item{\code{entityId}}{ID of the entity the comment posted to}

\item{\code{entityType}}{Type of the entity to post the comment to, currently only useCase is supported}

\item{\code{id}}{ID of the comment}

\item{\code{mentions}}{A list of users objects (see below) mentioned in the content}

\item{\code{updatedAt}}{Timestamp when the comment was updated}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentRetrieve$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentRetrieve$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentRetrieve$fromJSON(CommentRetrieveJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CommentRetrieveJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentRetrieve$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
