% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_update.R
\docType{class}
\name{CommentUpdate}
\alias{CommentUpdate}
\title{CommentUpdate}
\format{
An \code{R6Class} generator object
}
\description{
CommentUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{content}}{character Updated content of the comment, 10000 symbols max}

\item{\code{mentions}}{list( character ) \link{optional} A list of users IDs mentioned in the content}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CommentUpdate-new}{\code{CommentUpdate$new()}}
\item \href{#method-CommentUpdate-validate}{\code{CommentUpdate$validate()}}
\item \href{#method-CommentUpdate-toJSON}{\code{CommentUpdate$toJSON()}}
\item \href{#method-CommentUpdate-fromJSON}{\code{CommentUpdate$fromJSON()}}
\item \href{#method-CommentUpdate-clone}{\code{CommentUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommentUpdate-new"></a>}}
\if{latex}{\out{\hypertarget{method-CommentUpdate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentUpdate$new(content = NULL, mentions = NULL, validateParams = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{Updated content of the comment, 10000 symbols max}

\item{\code{mentions}}{A list of users IDs mentioned in the content}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommentUpdate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CommentUpdate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentUpdate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommentUpdate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CommentUpdate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommentUpdate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CommentUpdate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentUpdate$fromJSON(CommentUpdateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CommentUpdateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommentUpdate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CommentUpdate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
