% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_doc_template_create.R
\docType{class}
\name{ComplianceDocTemplateCreate}
\alias{ComplianceDocTemplateCreate}
\title{ComplianceDocTemplateCreate}
\format{
An \code{R6Class} generator object
}
\description{
ComplianceDocTemplateCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{labels}}{list( character ) \link{optional} Names of the labels to assign to the template.}

\item{\code{name}}{character Name of the new template. Must be unique among templates created by the user.}

\item{\code{projectType}}{character \link{optional} Type of project template.}

\item{\code{sections}}{list( \link{OneOfSectionDataRobotSectionUserSectionCustomSectionTableOfContents} ) List of section objects representing the structure of the document. Each section can have sub-sections that have the same schema as the parent section, i.e., the structure is recursive. The number of nested sections allowed is 5. The total number of sections allowed is 500.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ComplianceDocTemplateCreate-new}{\code{ComplianceDocTemplateCreate$new()}}
\item \href{#method-ComplianceDocTemplateCreate-validate}{\code{ComplianceDocTemplateCreate$validate()}}
\item \href{#method-ComplianceDocTemplateCreate-toJSON}{\code{ComplianceDocTemplateCreate$toJSON()}}
\item \href{#method-ComplianceDocTemplateCreate-fromJSON}{\code{ComplianceDocTemplateCreate$fromJSON()}}
\item \href{#method-ComplianceDocTemplateCreate-clone}{\code{ComplianceDocTemplateCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceDocTemplateCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceDocTemplateCreate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceDocTemplateCreate$new(
  name = NULL,
  sections = NULL,
  labels = NULL,
  projectType = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the new template. Must be unique among templates created by the user.}

\item{\code{sections}}{List of section objects representing the structure of the document. Each section can have sub-sections that have the same schema as the parent section, i.e., the structure is recursive. The number of nested sections allowed is 5. The total number of sections allowed is 500.}

\item{\code{labels}}{Names of the labels to assign to the template.}

\item{\code{projectType}}{Type of project template.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceDocTemplateCreate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceDocTemplateCreate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceDocTemplateCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceDocTemplateCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceDocTemplateCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceDocTemplateCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceDocTemplateCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceDocTemplateCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceDocTemplateCreate$fromJSON(
  ComplianceDocTemplateCreateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ComplianceDocTemplateCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceDocTemplateCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceDocTemplateCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceDocTemplateCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
