% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_doc_template_create_response.R
\docType{class}
\name{ComplianceDocTemplateCreateResponse}
\alias{ComplianceDocTemplateCreateResponse}
\title{ComplianceDocTemplateCreateResponse}
\format{
An \code{R6Class} generator object
}
\description{
ComplianceDocTemplateCreateResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dateModified}}{character Timestamp of the created template}

\item{\code{id}}{character Created template Id}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ComplianceDocTemplateCreateResponse-new}{\code{ComplianceDocTemplateCreateResponse$new()}}
\item \href{#method-ComplianceDocTemplateCreateResponse-validate}{\code{ComplianceDocTemplateCreateResponse$validate()}}
\item \href{#method-ComplianceDocTemplateCreateResponse-toJSON}{\code{ComplianceDocTemplateCreateResponse$toJSON()}}
\item \href{#method-ComplianceDocTemplateCreateResponse-fromJSON}{\code{ComplianceDocTemplateCreateResponse$fromJSON()}}
\item \href{#method-ComplianceDocTemplateCreateResponse-clone}{\code{ComplianceDocTemplateCreateResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceDocTemplateCreateResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceDocTemplateCreateResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceDocTemplateCreateResponse$new(
  dateModified = NULL,
  id = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dateModified}}{Timestamp of the created template}

\item{\code{id}}{Created template Id}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceDocTemplateCreateResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceDocTemplateCreateResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceDocTemplateCreateResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceDocTemplateCreateResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceDocTemplateCreateResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceDocTemplateCreateResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceDocTemplateCreateResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceDocTemplateCreateResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceDocTemplateCreateResponse$fromJSON(
  ComplianceDocTemplateCreateResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ComplianceDocTemplateCreateResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceDocTemplateCreateResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceDocTemplateCreateResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceDocTemplateCreateResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
