% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_doc_template_list_response.R
\docType{class}
\name{ComplianceDocTemplateListResponse}
\alias{ComplianceDocTemplateListResponse}
\title{ComplianceDocTemplateListResponse}
\format{
An \code{R6Class} generator object
}
\description{
ComplianceDocTemplateListResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer Number of items returned on this page.}

\item{\code{data}}{list( \link{TemplateResponse} ) List of templates.}

\item{\code{next_}}{character URL pointing to the previous page. If null, there is no next page.}

\item{\code{previous}}{character URL pointing to the previous page. If null, there is no previous page.}

\item{\code{totalCount}}{integer Total number of templates.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ComplianceDocTemplateListResponse-new}{\code{ComplianceDocTemplateListResponse$new()}}
\item \href{#method-ComplianceDocTemplateListResponse-validate}{\code{ComplianceDocTemplateListResponse$validate()}}
\item \href{#method-ComplianceDocTemplateListResponse-toJSON}{\code{ComplianceDocTemplateListResponse$toJSON()}}
\item \href{#method-ComplianceDocTemplateListResponse-fromJSON}{\code{ComplianceDocTemplateListResponse$fromJSON()}}
\item \href{#method-ComplianceDocTemplateListResponse-clone}{\code{ComplianceDocTemplateListResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceDocTemplateListResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceDocTemplateListResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceDocTemplateListResponse$new(
  count = NULL,
  data = NULL,
  next_ = NULL,
  previous = NULL,
  totalCount = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{Number of items returned on this page.}

\item{\code{data}}{List of templates.}

\item{\code{next_}}{URL pointing to the previous page. If null, there is no next page.}

\item{\code{previous}}{URL pointing to the previous page. If null, there is no previous page.}

\item{\code{totalCount}}{Total number of templates.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceDocTemplateListResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceDocTemplateListResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceDocTemplateListResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceDocTemplateListResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceDocTemplateListResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceDocTemplateListResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceDocTemplateListResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceDocTemplateListResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceDocTemplateListResponse$fromJSON(
  ComplianceDocTemplateListResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ComplianceDocTemplateListResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceDocTemplateListResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceDocTemplateListResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceDocTemplateListResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
