% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_chart_retrieve_for_datasets.R
\docType{class}
\name{ConfusionChartRetrieveForDatasets}
\alias{ConfusionChartRetrieveForDatasets}
\title{ConfusionChartRetrieveForDatasets}
\format{
An \code{R6Class} generator object
}
\description{
ConfusionChartRetrieveForDatasets Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{columns}}{list( integer ) \link{colStart, colEnd} column dimension of confusion matrix in response}

\item{\code{data}}{\link{ConfusionChartData}}

\item{\code{datasetId}}{character The datasetId to retrieve a Confusion chart from.}

\item{\code{numberOfClasses}}{integer count of classes in full confusion matrix.}

\item{\code{rows}}{list( integer ) \link{rowStart, rowEnd} row dimension of confusion matrix in response}

\item{\code{totalMatrixSum}}{integer sum of all values in full confusion matrix}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConfusionChartRetrieveForDatasets-new}{\code{ConfusionChartRetrieveForDatasets$new()}}
\item \href{#method-ConfusionChartRetrieveForDatasets-validate}{\code{ConfusionChartRetrieveForDatasets$validate()}}
\item \href{#method-ConfusionChartRetrieveForDatasets-toJSON}{\code{ConfusionChartRetrieveForDatasets$toJSON()}}
\item \href{#method-ConfusionChartRetrieveForDatasets-fromJSON}{\code{ConfusionChartRetrieveForDatasets$fromJSON()}}
\item \href{#method-ConfusionChartRetrieveForDatasets-clone}{\code{ConfusionChartRetrieveForDatasets$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfusionChartRetrieveForDatasets-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConfusionChartRetrieveForDatasets-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfusionChartRetrieveForDatasets$new(
  columns = NULL,
  data = NULL,
  datasetId = NULL,
  numberOfClasses = NULL,
  rows = NULL,
  totalMatrixSum = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{\link{colStart, colEnd} column dimension of confusion matrix in response}

\item{\code{datasetId}}{The datasetId to retrieve a Confusion chart from.}

\item{\code{numberOfClasses}}{count of classes in full confusion matrix.}

\item{\code{rows}}{\link{rowStart, rowEnd} row dimension of confusion matrix in response}

\item{\code{totalMatrixSum}}{sum of all values in full confusion matrix}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfusionChartRetrieveForDatasets-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ConfusionChartRetrieveForDatasets-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfusionChartRetrieveForDatasets$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfusionChartRetrieveForDatasets-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConfusionChartRetrieveForDatasets-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfusionChartRetrieveForDatasets$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfusionChartRetrieveForDatasets-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConfusionChartRetrieveForDatasets-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfusionChartRetrieveForDatasets$fromJSON(
  ConfusionChartRetrieveForDatasetsJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ConfusionChartRetrieveForDatasetsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfusionChartRetrieveForDatasets-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConfusionChartRetrieveForDatasets-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfusionChartRetrieveForDatasets$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
