% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/confusion_chart_retrieve_metadata_for_datasets.R
\docType{class}
\name{ConfusionChartRetrieveMetadataForDatasets}
\alias{ConfusionChartRetrieveMetadataForDatasets}
\title{ConfusionChartRetrieveMetadataForDatasets}
\format{
An \code{R6Class} generator object
}
\description{
ConfusionChartRetrieveMetadataForDatasets Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{classNames}}{list( character ) list of all class names in the full confusion matrix, sorted by the `orderBy` parameter.}

\item{\code{datasetId}}{character The dataset to retrieve a Confusion chart from.}

\item{\code{modelId}}{character The model to retrieve a Confusion chart from.}

\item{\code{projectId}}{character The project to retrieve a Confusion chart from.}

\item{\code{relevantClassesPositions}}{list( \link{array[integer]} ) Matrix to highlight important cell blocks in the confusion chart. Intended to represent a thumbnail view, where the relevantClassesPositions array has a 1 in thumbnail cells that are of interest, and 0 otherwise. The dimensions of the implied thumbnail will not match those of the confusion matrix, e.g. a twenty-class confusion matrix may have a 2x2 thumbnail.}

\item{\code{totalMatrixSum}}{integer Sum of all values in the full confusion matrix (equal to the number of points considered).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConfusionChartRetrieveMetadataForDatasets$new()}}
\item \href{#method-validate}{\code{ConfusionChartRetrieveMetadataForDatasets$validate()}}
\item \href{#method-toJSON}{\code{ConfusionChartRetrieveMetadataForDatasets$toJSON()}}
\item \href{#method-fromJSON}{\code{ConfusionChartRetrieveMetadataForDatasets$fromJSON()}}
\item \href{#method-clone}{\code{ConfusionChartRetrieveMetadataForDatasets$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfusionChartRetrieveMetadataForDatasets$new(
  classNames = NULL,
  datasetId = NULL,
  modelId = NULL,
  projectId = NULL,
  relevantClassesPositions = NULL,
  totalMatrixSum = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{classNames}}{list of all class names in the full confusion matrix, sorted by the `orderBy` parameter.}

\item{\code{datasetId}}{The dataset to retrieve a Confusion chart from.}

\item{\code{modelId}}{The model to retrieve a Confusion chart from.}

\item{\code{projectId}}{The project to retrieve a Confusion chart from.}

\item{\code{relevantClassesPositions}}{Matrix to highlight important cell blocks in the confusion chart. Intended to represent a thumbnail view, where the relevantClassesPositions array has a 1 in thumbnail cells that are of interest, and 0 otherwise. The dimensions of the implied thumbnail will not match those of the confusion matrix, e.g. a twenty-class confusion matrix may have a 2x2 thumbnail.}

\item{\code{totalMatrixSum}}{Sum of all values in the full confusion matrix (equal to the number of points considered).}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfusionChartRetrieveMetadataForDatasets$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfusionChartRetrieveMetadataForDatasets$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfusionChartRetrieveMetadataForDatasets$fromJSON(
  ConfusionChartRetrieveMetadataForDatasetsJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ConfusionChartRetrieveMetadataForDatasetsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfusionChartRetrieveMetadataForDatasets$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
