% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_response.R
\docType{class}
\name{ConversionResponse}
\alias{ConversionResponse}
\title{ConversionResponse}
\format{
An \code{R6Class} generator object
}
\description{
ConversionResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{conversionInProgress}}{character \link{optional} Whether a custom model conversion is in progress or not.}

\item{\code{conversionSucceeded}}{character \link{optional} Indication for a successful custom model conversion.}

\item{\code{created}}{character ISO-8601 timestamp of when the custom model conversion created.}

\item{\code{customModelVersionId}}{character ID of the custom model version.}

\item{\code{generatedMetadata}}{\link{GeneratedMetadata} \link{optional}}

\item{\code{id}}{character ID of the custom model version.}

\item{\code{logMessage}}{character \link{optional} The output log message from the custom model conversion process.}

\item{\code{mainProgramItemId}}{character \link{optional} The main program file item ID.}

\item{\code{shouldStop}}{character \link{optional} Whether the user requested to stop the given conversion.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConversionResponse-new}{\code{ConversionResponse$new()}}
\item \href{#method-ConversionResponse-validate}{\code{ConversionResponse$validate()}}
\item \href{#method-ConversionResponse-toJSON}{\code{ConversionResponse$toJSON()}}
\item \href{#method-ConversionResponse-fromJSON}{\code{ConversionResponse$fromJSON()}}
\item \href{#method-ConversionResponse-clone}{\code{ConversionResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConversionResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConversionResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConversionResponse$new(
  created = NULL,
  customModelVersionId = NULL,
  id = NULL,
  conversionInProgress = NULL,
  conversionSucceeded = NULL,
  generatedMetadata = NULL,
  logMessage = NULL,
  mainProgramItemId = NULL,
  shouldStop = FALSE,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{created}}{ISO-8601 timestamp of when the custom model conversion created.}

\item{\code{customModelVersionId}}{ID of the custom model version.}

\item{\code{id}}{ID of the custom model version.}

\item{\code{conversionInProgress}}{Whether a custom model conversion is in progress or not.}

\item{\code{conversionSucceeded}}{Indication for a successful custom model conversion.}

\item{\code{logMessage}}{The output log message from the custom model conversion process.}

\item{\code{mainProgramItemId}}{The main program file item ID.}

\item{\code{shouldStop}}{Whether the user requested to stop the given conversion.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConversionResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ConversionResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConversionResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConversionResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConversionResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConversionResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConversionResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConversionResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConversionResponse$fromJSON(ConversionResponseJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ConversionResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConversionResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConversionResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConversionResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
