% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications_apiWrapper.R
\name{CreateApplicationsDeployments}
\alias{CreateApplicationsDeployments}
\title{CreateApplicationsDeployments}
\usage{
CreateApplicationsDeployments(modelDeploymentId, applicationId, linkName, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Links a deployment to an application
}
\details{
If application creates deployment during its lifetime, we want to have an API to link deployment with application.

This method invokes \verb{POST /applications/\{applicationId\}/deployments/} in the DataRobot Public API.
}
\seealso{
The method ApplicationsDeploymentsCreate in \link{ApplicationsApi}, which invokes the same underlying API endpoint.
}
