% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{CreateBatchPredictions}
\alias{CreateBatchPredictions}
\title{CreateBatchPredictions}
\usage{
CreateBatchPredictions(
  columnNamesRemapping,
  timeseriesSettings,
  modelId,
  passthroughColumns,
  outputSettings,
  explanationNumTopClasses,
  predictionWarningEnabled,
  intakeSettings,
  deploymentId,
  predictionInstance,
  csvSettings,
  chunkSize,
  explanationClassNames,
  modelPackageId,
  numConcurrent,
  passthroughColumnsSet,
  explanationAlgorithm,
  thresholdHigh,
  thresholdLow,
  pinnedModelId,
  includeProbabilitiesClasses = c(),
  includePredictionStatus = FALSE,
  skipDriftTracking = FALSE,
  disableRowLevelErrorHandling = FALSE,
  abortOnError = TRUE,
  maxExplanations = 0,
  includeProbabilities = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Creates a new Batch Prediction job
}
\details{
Submit the configuration for the job and it will be submitted to the queue

This method invokes \verb{POST /batchPredictions/} in the DataRobot Public API.
}
\seealso{
The method BatchPredictionsCreate in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
