% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{CreateBatchPredictionsCsvUploadFinalizeMultipart}
\alias{CreateBatchPredictionsCsvUploadFinalizeMultipart}
\title{CreateBatchPredictionsCsvUploadFinalizeMultipart}
\usage{
CreateBatchPredictionsCsvUploadFinalizeMultipart(
  predictionJobId,
  partNumber = 0,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Finalize a multipart upload
}
\details{
Finalize a multipart upload, indicating that no further chunks will be sent

This method invokes \verb{POST /batchPredictions/\{predictionJobId\}/csvUpload/finalizeMultipart/} in the DataRobot Public API.
}
\seealso{
The method BatchPredictionsCsvUploadFinalizeMultipartCreate in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
