% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{CreateCalendar}
\alias{CreateCalendar}
\title{CreateCalendar}
\usage{
CreateCalendar(
  dataSource,
  name = NULL,
  multiSeriesIdColumn = NULL,
  maxWait = 600,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a calendar from a file.
}
\details{
Create a calendar from a file in a csv or xlsx format. The calendar file specifies the dates or events in a dataset such that DataRobot automatically derives and creates special features based on the calendar events (e.g., time until the next event, labeling the most recent event).

This method invokes \verb{POST /calendars/fileUpload/} in the DataRobot Public API.
}
\seealso{
The method CalendarsFileUploadCreate in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
