% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{CreateChangeRequestsReviews}
\alias{CreateChangeRequestsReviews}
\title{CreateChangeRequestsReviews}
\usage{
CreateChangeRequestsReviews(
  changeRequestId,
  comment,
  changeVersionId,
  status,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create review.
}
\details{
Review the Change Request.

This method invokes \verb{POST /changeRequests/\{changeRequestId\}/reviews/} in the DataRobot Public API.
}
\seealso{
The method ChangeRequestsReviewsCreate in \link{GovernanceApi}, which invokes the same underlying API endpoint.
}
