% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure_apiWrapper.R
\name{CreateClusterLicenseValidation}
\alias{CreateClusterLicenseValidation}
\title{CreateClusterLicenseValidation}
\usage{
CreateClusterLicenseValidation(licenseKey, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Check if a license is valid
}
\details{
Check if a cluster license is valid.

This method invokes \verb{POST /clusterLicenseValidation/} in the DataRobot Public API.
}
\seealso{
The method ClusterLicenseValidationCreate in \link{InfrastructureApi}, which invokes the same underlying API endpoint.
}
