% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_apiWrapper.R
\name{CreateCredentials}
\alias{CreateCredentials}
\title{CreateCredentials}
\usage{
CreateCredentials(credentialsBody, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Store a new set of credentials which can be used for data sources creation.
}
\details{
Store a new set of credentials.

This method invokes \verb{POST /credentials/} in the DataRobot Public API.
}
\seealso{
The method CredentialsCreate in \link{CredentialsApi}, which invokes the same underlying API endpoint.
}
