% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_credentials_response.R
\docType{class}
\name{CreateCredentialsResponse}
\alias{CreateCredentialsResponse}
\title{CreateCredentialsResponse}
\format{
An \code{R6Class} generator object
}
\description{
CreateCredentialsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{creationDate}}{character ISO-8601 formatted date/time when these credentials were created.}

\item{\code{credentialId}}{character ID of these credentials.}

\item{\code{credentialType}}{character \link{optional} Type of credentials.}

\item{\code{description}}{character \link{optional} Description of these credentials.}

\item{\code{name}}{character Name of these credentials.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CreateCredentialsResponse-new}{\code{CreateCredentialsResponse$new()}}
\item \href{#method-CreateCredentialsResponse-validate}{\code{CreateCredentialsResponse$validate()}}
\item \href{#method-CreateCredentialsResponse-toJSON}{\code{CreateCredentialsResponse$toJSON()}}
\item \href{#method-CreateCredentialsResponse-fromJSON}{\code{CreateCredentialsResponse$fromJSON()}}
\item \href{#method-CreateCredentialsResponse-clone}{\code{CreateCredentialsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateCredentialsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-CreateCredentialsResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateCredentialsResponse$new(
  creationDate = NULL,
  credentialId = NULL,
  name = NULL,
  credentialType = "basic",
  description = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{creationDate}}{ISO-8601 formatted date/time when these credentials were created.}

\item{\code{credentialId}}{ID of these credentials.}

\item{\code{name}}{Name of these credentials.}

\item{\code{credentialType}}{Type of credentials.}

\item{\code{description}}{Description of these credentials.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateCredentialsResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CreateCredentialsResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateCredentialsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateCredentialsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CreateCredentialsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateCredentialsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateCredentialsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CreateCredentialsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateCredentialsResponse$fromJSON(
  CreateCredentialsResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CreateCredentialsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateCredentialsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CreateCredentialsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateCredentialsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
