% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{CreateCustomModelsVersionsConversions}
\alias{CreateCustomModelsVersionsConversions}
\title{CreateCustomModelsVersionsConversions}
\usage{
CreateCustomModelsVersionsConversions(
  customModelId,
  mainProgramItemId,
  customModelVersionId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Generates JAR file from particular files.
}
\details{
Converts files in the given custom model version to a JAR file.

This method invokes \verb{POST /customModels/\{customModelId\}/versions/\{customModelVersionId\}/conversions/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsVersionsConversionsCreate in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
