% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{CreateCustomModelsVersionsPredictionExplanationsInitialization}
\alias{CreateCustomModelsVersionsPredictionExplanationsInitialization}
\title{CreateCustomModelsVersionsPredictionExplanationsInitialization}
\usage{
CreateCustomModelsVersionsPredictionExplanationsInitialization(
  customModelId,
  customModelVersionId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a new prediction explanations initialization for custom model version.
}
\details{
Create a new prediction explanations initialization for custom model version. This is a necessary prerequisite for generating prediction explanations.

This method invokes \verb{POST /customModels/\{customModelId\}/versions/\{customModelVersionId\}/predictionExplanationsInitialization/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsVersionsPredictionExplanationsInitializationCreate in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
