% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{CreateCustomModelTests}
\alias{CreateCustomModelTests}
\title{CreateCustomModelTests}
\usage{
CreateCustomModelTests(
  networkEgressPolicy,
  requiresHa,
  environmentId,
  customModelId,
  configuration,
  replicas,
  maximumMemory,
  datasetId,
  desiredMemory,
  environmentVersionId,
  datasetVersionId,
  customModelVersionId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create custom model test.
}
\details{
Test a custom inference model. This will start a job to check that the custom model can make predictions against the supplied dataset without breaking.

This method invokes \verb{POST /customModelTests/} in the DataRobot Public API.
}
\seealso{
The method CustomModelTestsCreate in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
