% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{CreateCustomTasks}
\alias{CreateCustomTasks}
\title{CreateCustomTasks}
\usage{
CreateCustomTasks(
  name,
  maximumMemory,
  description,
  targetType,
  language,
  calibratePredictions = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a custom task
}
\details{
Creates a new custom task and returns the newly created metadata record for it.  A custom task may either be an estimator or a transform. Estimators must support a single target type (e.g. binaryClassification, regression). Regression and anomaly detection models are expected to produce predictions that are arbitrary floating-point or integer numbers. A classification model is expected to return predictions with probability scores for each class. For example, a binary classification model might return:  \verb{Python \{     positiveClassLabel: probability,     negativeClassLabel: 1.0 - probability \}}  Transforms are expected to return a dataframe or sparse matrix with the same number of rows as the input feature matrix. At this time, only numeric outputs are supported for custom transforms.

This method invokes \verb{POST /customTasks/} in the DataRobot Public API.
}
\seealso{
The method CustomTasksCreate in \link{CustomTasksApi}, which invokes the same underlying API endpoint.
}
