% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{CreateCustomTasksVersionsFromRepository}
\alias{CreateCustomTasksVersionsFromRepository}
\title{CreateCustomTasksVersionsFromRepository}
\usage{
CreateCustomTasksVersionsFromRepository(
  requiredMetadata,
  ref,
  baseEnvironmentId,
  customTaskId,
  repositoryId,
  requiredMetadataValues,
  sourcePath,
  isMajorUpdate = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create custom task version from remote repository.
}
\details{
Create a new custom task version with only files added from the specified remote repository.

This method invokes \verb{POST /customTasks/\{customTaskId\}/versions/fromRepository/} in the DataRobot Public API.
}
\seealso{
The method CustomTasksVersionsFromRepositoryCreate in \link{CustomTasksApi}, which invokes the same underlying API endpoint.
}
