% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{CreateCustomTrainingBlueprints}
\alias{CreateCustomTrainingBlueprints}
\title{CreateCustomTrainingBlueprints}
\usage{
CreateCustomTrainingBlueprints(customModelVersionId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a blueprint from a single custom training estimator.
}
\details{
This route creates a blueprint from a custom training estimator with an environment so that it can be trained via blueprint ID.

This method invokes \verb{POST /customTrainingBlueprints/} in the DataRobot Public API.
}
\seealso{
The method CustomTrainingBlueprintsCreate in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
