% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R,
%   R/create_data_engine_query_generator.R
\docType{class}
\name{CreateDataEngineQueryGenerator}
\alias{CreateDataEngineQueryGenerator}
\title{CreateDataEngineQueryGenerator}
\format{
An \code{R6Class} generator object
}
\description{
Create a data engine query generator

CreateDataEngineQueryGenerator Class
}
\details{
Create a data engine query generator.

This method invokes \verb{POST /dataEngineQueryGenerators/} in the DataRobot Public API.
}
\seealso{
The method DataEngineQueryGeneratorsCreate in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CreateDataEngineQueryGenerator-new}{\code{CreateDataEngineQueryGenerator$new()}}
\item \href{#method-CreateDataEngineQueryGenerator-validate}{\code{CreateDataEngineQueryGenerator$validate()}}
\item \href{#method-CreateDataEngineQueryGenerator-toJSON}{\code{CreateDataEngineQueryGenerator$toJSON()}}
\item \href{#method-CreateDataEngineQueryGenerator-fromJSON}{\code{CreateDataEngineQueryGenerator$fromJSON()}}
\item \href{#method-CreateDataEngineQueryGenerator-clone}{\code{CreateDataEngineQueryGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateDataEngineQueryGenerator-new"></a>}}
\if{latex}{\out{\hypertarget{method-CreateDataEngineQueryGenerator-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateDataEngineQueryGenerator$new(
  datasets = NULL,
  generatorSettings = NULL,
  generatorType = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateDataEngineQueryGenerator-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CreateDataEngineQueryGenerator-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateDataEngineQueryGenerator$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateDataEngineQueryGenerator-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CreateDataEngineQueryGenerator-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateDataEngineQueryGenerator$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateDataEngineQueryGenerator-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CreateDataEngineQueryGenerator-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateDataEngineQueryGenerator$fromJSON(
  CreateDataEngineQueryGeneratorJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateDataEngineQueryGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CreateDataEngineQueryGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateDataEngineQueryGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datasets}}{list( \link{DataEngineDataset} ) Source datasets in the Data Engine workspace.}

\item{\code{generatorSettings}}{\link{GeneratorSettings}}

\item{\code{generatorType}}{character Type of data engine query generator}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CreateDataEngineQueryGenerator-new}{\code{CreateDataEngineQueryGenerator$new()}}
\item \href{#method-CreateDataEngineQueryGenerator-validate}{\code{CreateDataEngineQueryGenerator$validate()}}
\item \href{#method-CreateDataEngineQueryGenerator-toJSON}{\code{CreateDataEngineQueryGenerator$toJSON()}}
\item \href{#method-CreateDataEngineQueryGenerator-fromJSON}{\code{CreateDataEngineQueryGenerator$fromJSON()}}
\item \href{#method-CreateDataEngineQueryGenerator-clone}{\code{CreateDataEngineQueryGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateDataEngineQueryGenerator-new"></a>}}
\if{latex}{\out{\hypertarget{method-CreateDataEngineQueryGenerator-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateDataEngineQueryGenerator$new(
  datasets = NULL,
  generatorSettings = NULL,
  generatorType = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasets}}{Source datasets in the Data Engine workspace.}

\item{\code{generatorType}}{Type of data engine query generator}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateDataEngineQueryGenerator-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CreateDataEngineQueryGenerator-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateDataEngineQueryGenerator$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateDataEngineQueryGenerator-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CreateDataEngineQueryGenerator-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateDataEngineQueryGenerator$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateDataEngineQueryGenerator-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CreateDataEngineQueryGenerator-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateDataEngineQueryGenerator$fromJSON(
  CreateDataEngineQueryGeneratorJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CreateDataEngineQueryGeneratorJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateDataEngineQueryGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CreateDataEngineQueryGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateDataEngineQueryGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
