% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{CreateDataEngineWorkspaceStates}
\alias{CreateDataEngineWorkspaceStates}
\title{CreateDataEngineWorkspaceStates}
\usage{
CreateDataEngineWorkspaceStates(query, language, datasets, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create Data Engine workspace state
}
\details{
Create Data Engine workspace state in database.

This method invokes \verb{POST /dataEngineWorkspaceStates/} in the DataRobot Public API.
}
\seealso{
The method DataEngineWorkspaceStatesCreate in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
