% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{CreateDatasetFromDataEngineQueryGenerator}
\alias{CreateDatasetFromDataEngineQueryGenerator}
\title{CreateDatasetFromDataEngineQueryGenerator}
\usage{
CreateDatasetFromDataEngineQueryGenerator(
  generatorId,
  datasetId = NULL,
  datasetVersionId = NULL,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create Data Engine workspace state from a query generator
}
\details{
Create Data Engine workspace state in database from a query generator.

This method invokes \verb{POST /dataEngineWorkspaceStates/fromDataEngineQueryGenerator/} in the DataRobot Public API.
}
\seealso{
The method DataEngineWorkspaceStatesFromDataEngineQueryGeneratorCreate in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
