% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{CreateDatasetsFeaturelists}
\alias{CreateDatasetsFeaturelists}
\title{CreateDatasetsFeaturelists}
\usage{
CreateDatasetsFeaturelists(features, name, description, datasetId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create dataset featurelist
}
\details{
Create featurelist for specified dataset.

This method invokes \verb{POST /datasets/\{datasetId\}/featurelists/} in the DataRobot Public API.
}
\seealso{
The method DatasetsFeaturelistsCreate in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
