% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{CreateDatasetsFeatureTransforms}
\alias{CreateDatasetsFeatureTransforms}
\title{CreateDatasetsFeatureTransforms}
\usage{
CreateDatasetsFeatureTransforms(
  variableType,
  parentName,
  name,
  datasetId,
  dateExtraction,
  replacement,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create dataset feature transform
}
\details{
Create a new feature by changing the type of an existing one.

This method invokes \verb{POST /datasets/\{datasetId\}/featureTransforms/} in the DataRobot Public API.
}
\seealso{
The method DatasetsFeatureTransformsCreate in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
