% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{CreateDatasetsFromDataSource}
\alias{CreateDatasetsFromDataSource}
\title{CreateDatasetsFromDataSource}
\usage{
CreateDatasetsFromDataSource(
  password,
  dataSourceId,
  persistDataAfterIngestion,
  credentialId,
  credentialData,
  categories,
  sampleSize,
  user,
  doSnapshot = TRUE,
  useKerberos = FALSE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create dataset from data source
}
\details{
Create a Dataset Item from a data source.

This method invokes \verb{POST /datasets/fromDataSource/} in the DataRobot Public API.
}
\seealso{
The method DatasetsFromDataSourceCreate in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
