% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{CreateDatasetsFromHDFS}
\alias{CreateDatasetsFromHDFS}
\title{CreateDatasetsFromHDFS}
\usage{
CreateDatasetsFromHDFS(
  password,
  persistDataAfterIngestion,
  namenodeWebhdfsPort,
  categories,
  user,
  url,
  doSnapshot = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create dataset from HDFS URL
}
\details{
Create a Dataset Item from an HDFS URL.

This method invokes \verb{POST /datasets/fromHDFS/} in the DataRobot Public API.
}
\seealso{
The method DatasetsFromHDFSCreate in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
