% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{CreateDatasetsRelationships}
\alias{CreateDatasetsRelationships}
\title{CreateDatasetsRelationships}
\usage{
CreateDatasetsRelationships(
  linkedFeatures,
  linkedDatasetId,
  datasetId,
  sourceFeatures,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create dataset relationship.
}
\details{
Create a dataset relationship.

This method invokes \verb{POST /datasets/\{datasetId\}/relationships/} in the DataRobot Public API.
}
\seealso{
The method DatasetsRelationshipsCreate in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
