% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{CreateDatasetsVersionsFromDataEngineWorkspaceState}
\alias{CreateDatasetsVersionsFromDataEngineWorkspaceState}
\title{CreateDatasetsVersionsFromDataEngineWorkspaceState}
\usage{
CreateDatasetsVersionsFromDataEngineWorkspaceState(
  credentials,
  datasetName,
  datasetId,
  workspaceStateId,
  doSnapshot,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create dataset version from Data Engine workspace
}
\details{
Create a new dataset version for a specified dataset from a Data Engine workspace state. The new dataset version should have the same schema as the specified dataset.

This method invokes \verb{POST /datasets/\{datasetId\}/versions/fromDataEngineWorkspaceState/} in the DataRobot Public API.
}
\seealso{
The method DatasetsVersionsFromDataEngineWorkspaceStateCreate in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
