% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{CreateDatasetsVersionsFromFile}
\alias{CreateDatasetsVersionsFromFile}
\title{CreateDatasetsVersionsFromFile}
\usage{
CreateDatasetsVersionsFromFile(file, datasetId, categories, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create dataset version from file
}
\details{
Create a new version for the specified dataset from a file.

This method invokes \verb{POST /datasets/\{datasetId\}/versions/fromFile/} in the DataRobot Public API.
}
\seealso{
The method DatasetsVersionsFromFileCreate in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
