% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{CreateDataSource}
\alias{CreateDataSource}
\title{CreateDataSource}
\usage{
CreateDataSource(
  type,
  canonicalName,
  dataStoreId,
  query = NULL,
  table = NULL,
  schema = NULL,
  partitionColumn = NULL,
  fetchSize = NULL,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a data source.
}
\details{
Create a fully configured source of data which could be used for datasets and projects creation.A data source specifies, via SQL query or selected table and schema data, which data to extract from the data connection (the location of data within a given endpoint) to use for modeling or predictions. A data source has one data connection and one connector but can have many datasets. To test the SQL query before creating the data source,use :\url{http:post}:/api/v2/externalDataStores/(dataStoreId)/verifySQL/.

This method invokes \verb{POST /externalDataSources/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataSourcesCreate in \link{DataConnectivityApi}, which invokes the same underlying API endpoint.
}
