% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{CreateDataStore}
\alias{CreateDataStore}
\title{CreateDataStore}
\usage{
CreateDataStore(type, canonicalName, driverId, jdbcUrl, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a data store.
}
\details{
Create a data store which includes a name and a driver ID or a connector ID. The driver would be configured by a JDBC URL or by jdbc fields; The connector would be configured by connection fields.

This method invokes \verb{POST /externalDataStores/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataStoresCreate in \link{DataConnectivityApi}, which invokes the same underlying API endpoint.
}
