% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{CreateDeployment}
\alias{CreateDeployment}
\title{CreateDeployment}
\usage{
CreateDeployment(
  model,
  label = "",
  description = "",
  defaultPredictionServerId = NULL,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create deployment from DataRobot model
}
\details{
Create a deployment from a DataRobot model.

This method invokes \verb{POST /deployments/fromLearningModel/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsFromLearningModelCreate in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
