% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{CreateDeploymentsActualsFromDataset}
\alias{CreateDeploymentsActualsFromDataset}
\title{CreateDeploymentsActualsFromDataset}
\usage{
CreateDeploymentsActualsFromDataset(
  password,
  wasActedOnColumn,
  associationIdColumn,
  actualValueColumn,
  deploymentId,
  datasetId,
  user,
  datasetVersionId,
  timestampColumn,
  keepActualsWithoutPredictions = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Submit actuals values from AI Catalog
}
\details{
Submit actuals values for processing using catalog item. Submission of actuals is limited to 10,000,000 actuals per hour. For time series deployments, total actuals = number of actuals \\* number of forecast distances. For example, submitting 10 actuals for a deployment with 50 forecast distances = 500 total actuals. For multiclass deployments, a similar calculation is made where total actuals = number of actuals \\* number of classes. For example, submitting 10 actuals for a deployment with 20 classes = 200 actuals.

This method invokes \verb{POST /deployments/\{deploymentId\}/actuals/fromDataset/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsActualsFromDatasetCreate in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
