% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{CreateDeploymentsChallengerPredictions}
\alias{CreateDeploymentsChallengerPredictions}
\title{CreateDeploymentsChallengerPredictions}
\usage{
CreateDeploymentsChallengerPredictions(deploymentId, timestamp, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Score challenger models
}
\details{
Score main model prediction requests against challenger model requests.

This method invokes \verb{POST /deployments/\{deploymentId\}/challengerPredictions/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsChallengerPredictionsCreate in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
