% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{CreateDeploymentsChallengers}
\alias{CreateDeploymentsChallengers}
\title{CreateDeploymentsChallengers}
\usage{
CreateDeploymentsChallengers(
  deploymentId,
  name,
  predictionEnvironmentId,
  modelPackageId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create challenger model
}
\details{
Create new challenger model.

This method invokes \verb{POST /deployments/\{deploymentId\}/challengers/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsChallengersCreate in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
