% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{CreateDeploymentsPredictions}
\alias{CreateDeploymentsPredictions}
\title{CreateDeploymentsPredictions}
\usage{
CreateDeploymentsPredictions(
  predictionsEndDate,
  maxExplanations,
  predictionWarningEnabled,
  maxCodes,
  relaxKnownInAdvanceFeaturesCheck,
  predictionsStartDate,
  passthroughColumnsSet,
  passthroughColumns,
  deploymentId,
  forecastPoint,
  thresholdHigh,
  thresholdLow,
  excludeAdjustedPredictions = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Compute predictions synchronously
}
\details{
The route can be used for low-latency (hence synchronous) predictions. Scoring dataset, however, must be up to 50 MB and is limited by comma-delimited CSV and record oriented JSON. All other formats are not supported.

This method invokes \verb{POST /deployments/\{deploymentId\}/predictions/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsPredictionsCreate in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
