% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{CreateDeploymentsRetrainingPoliciesRuns}
\alias{CreateDeploymentsRetrainingPoliciesRuns}
\title{CreateDeploymentsRetrainingPoliciesRuns}
\usage{
CreateDeploymentsRetrainingPoliciesRuns(retrainingPolicyId, deploymentId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Endpoint for initiating a deployment retraining policy run.
}
\details{
Initiate a deployment retraining policy run.

This method invokes \verb{POST /deployments/\{deploymentId\}/retrainingPolicies/\{retrainingPolicyId\}/runs/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsRetrainingPoliciesRunsCreate in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
