% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_driver_request.R
\docType{class}
\name{CreateDriverRequest}
\alias{CreateDriverRequest}
\title{CreateDriverRequest}
\format{
An \code{R6Class} generator object
}
\description{
CreateDriverRequest Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{baseNames}}{list( character ) Original file name(s) of the uploaded JAR file(s). If there are multiple JAR files required for the driver, each of the original file names should be present in the list in the same order as found in 'localJarUrls'}

\item{\code{canonicalName}}{character \link{optional} User-friendly driver name.}

\item{\code{className}}{character \link{optional} Driver class name. For example 'com.amazon.redshift.jdbc.Driver')}

\item{\code{configurationId}}{character \link{optional} Driver configuration ID if it was provided during driver upload.}

\item{\code{localJarUrls}}{list( character ) File path(s) for the driver files.This path is returned in the response as 'local_url' by the driverUpload route.If there are multiple JAR files required for the driver, each uploaded JAR must be present in this list. If specified, values will replace any previous settings.}

\item{\code{version}}{character \link{optional} Driver version, which is used to construct the canonical name if driver configuration ID was provided during driver upload.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CreateDriverRequest$new()}}
\item \href{#method-validate}{\code{CreateDriverRequest$validate()}}
\item \href{#method-toJSON}{\code{CreateDriverRequest$toJSON()}}
\item \href{#method-fromJSON}{\code{CreateDriverRequest$fromJSON()}}
\item \href{#method-clone}{\code{CreateDriverRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateDriverRequest$new(
  baseNames = NULL,
  localJarUrls = NULL,
  canonicalName = NULL,
  className = NULL,
  configurationId = NULL,
  version = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{baseNames}}{Original file name(s) of the uploaded JAR file(s). If there are multiple JAR files required for the driver, each of the original file names should be present in the list in the same order as found in 'localJarUrls'}

\item{\code{localJarUrls}}{File path(s) for the driver files.This path is returned in the response as 'local_url' by the driverUpload route.If there are multiple JAR files required for the driver, each uploaded JAR must be present in this list. If specified, values will replace any previous settings.}

\item{\code{canonicalName}}{User-friendly driver name.}

\item{\code{className}}{Driver class name. For example 'com.amazon.redshift.jdbc.Driver')}

\item{\code{configurationId}}{Driver configuration ID if it was provided during driver upload.}

\item{\code{version}}{Driver version, which is used to construct the canonical name if driver configuration ID was provided during driver upload.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateDriverRequest$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateDriverRequest$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateDriverRequest$fromJSON(CreateDriverRequestJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CreateDriverRequestJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateDriverRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
