% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{CreateExternalDataDriverFile}
\alias{CreateExternalDataDriverFile}
\title{CreateExternalDataDriverFile}
\usage{
CreateExternalDataDriverFile(file, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Upload JDBC driver from file.
}
\details{
Upload JDBC driver from file. Only Java archive (.jar) files are supported.

This method invokes \verb{POST /externalDataDriverFile/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataDriverFileCreate in \link{DataConnectivityApi}, which invokes the same underlying API endpoint.
}
