% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{CreateExternalDataStoresVerifySQL}
\alias{CreateExternalDataStoresVerifySQL}
\title{CreateExternalDataStoresVerifySQL}
\usage{
CreateExternalDataStoresVerifySQL(
  password,
  maxRows,
  dataStoreId,
  query,
  credentialId,
  user,
  useKerberos = FALSE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Verifies a SQL query for a data store.
}
\details{
Executes the SQL query on the data store, returning a small number of rows (max 999). Use this for quick query execution validation and exploring results, not for capturing an entire result set.

This method invokes \verb{POST /externalDataStores/\{dataStoreId\}/verifySQL/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataStoresVerifySQLCreate in \link{DataConnectivityApi}, which invokes the same underlying API endpoint.
}
