% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_featurelist.R, R/projects_apiWrapper.R
\docType{class}
\name{CreateFeaturelist}
\alias{CreateFeaturelist}
\title{CreateFeaturelist}
\format{
An \code{R6Class} generator object
}
\usage{
CreateFeaturelist(project, listName, featureNames, ...)

CreateFeaturelist(project, listName, featureNames, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}

\item{features}{List of features for new featurelist.}

\item{name}{New featurelist name.}

\item{skipDatetimePartitionColumn}{Whether featurelist should contain datetime partition column.}

\item{CreateFeaturelistJson}{A JSON encoded string representation of a class instance.}

\item{validateParams}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\description{
CreateFeaturelist Class

A function used to initialize an instance of this class.

A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.

A helper function that serializes this object into a JSON encoded string.

A helper function that deserializes a JSON string into an instance of this class.

Create a new featurelist.
}
\details{
Create a new featurelist from list of feature names.

This method invokes \verb{POST /projects/\{projectId\}/featurelists/} in the DataRobot Public API.
}
\section{Fields}{

\describe{
\item{\code{features}}{list( character ) List of features for new featurelist.}

\item{\code{name}}{character New featurelist name.}

\item{\code{skipDatetimePartitionColumn}}{character \link{optional} Whether featurelist should contain datetime partition column.}
}}

\seealso{
The method FeaturelistsCreate in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
