% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{CreateFromLatestCustomTaskVersion}
\alias{CreateFromLatestCustomTaskVersion}
\title{CreateFromLatestCustomTaskVersion}
\usage{
CreateFromLatestCustomTaskVersion(
  requiredMetadata,
  file,
  baseEnvironmentId,
  filePath,
  customTaskId,
  maximumMemory,
  filesToDelete,
  requiredMetadataValues,
  isMajorUpdate = "true",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update custom task version files.
}
\details{
Create a new custom task version with files added, replaced or deleted. Files from the previous version of a custom task will be used as a basis.

This method invokes \verb{PATCH /customTasks/\{customTaskId\}/versions/} in the DataRobot Public API.
}
\seealso{
The method CustomTaskVersionCreateFromLatest in \link{CustomTasksApi}, which invokes the same underlying API endpoint.
}
