% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{CreateGroupsUsers}
\alias{CreateGroupsUsers}
\title{CreateGroupsUsers}
\usage{
CreateGroupsUsers(groupId, users, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Add users to group
}
\details{
Add users to the group.

This method invokes \verb{POST /groups/\{groupId\}/users/} in the DataRobot Public API.
}
\seealso{
The method GroupsUsersCreate in \link{UserManagementApi}, which invokes the same underlying API endpoint.
}
