% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{CreateHdfsProjects}
\alias{CreateHdfsProjects}
\title{CreateHdfsProjects}
\usage{
CreateHdfsProjects(password, port, projectName, user, url, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a project from an HDFS file source.
}
\details{
Create a project from an HDFS file via WebHDFS API. Represent the file using URL, optionally, port, and optionally, user/password credentials. For example, {\"url\": \"hdfs://\\<ip\\>/path/to/file.csv\", \"port\": \"50070\"}.

This method invokes \verb{POST /hdfsProjects/} in the DataRobot Public API.
}
\seealso{
The method HdfsProjectsCreate in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
