% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R
\name{CreateImageAugmentationListsSamples}
\alias{CreateImageAugmentationListsSamples}
\title{CreateImageAugmentationListsSamples}
\usage{
CreateImageAugmentationListsSamples(augmentationId, numberOfRows, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Requests the creation of sample augmentations based on the augmentation list passed in.
}
\details{
This endpoint will schedule a job to augment the specified images from a project's dataset and return a link to monitor the status of the job, as well as a link to retrieve the resulting augmented images.

This method invokes \verb{POST /imageAugmentationLists/\{augmentationId\}/samples/} in the DataRobot Public API.
}
\seealso{
The method ImageAugmentationListsSamplesCreate in \link{ImagesApi}, which invokes the same underlying API endpoint.
}
