% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{CreateModelPackagesFromLearningModel}
\alias{CreateModelPackagesFromLearningModel}
\title{CreateModelPackagesFromLearningModel}
\usage{
CreateModelPackagesFromLearningModel(predictionThreshold, modelId, name, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create model package from DataRobot model.
}
\details{
Create model package from DataRobot model.

This method invokes \verb{POST /modelPackages/fromLearningModel/} in the DataRobot Public API.
}
\seealso{
The method ModelPackagesFromLearningModelCreate in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
