% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{CreateNotificationEmailChannelVerification}
\alias{CreateNotificationEmailChannelVerification}
\title{CreateNotificationEmailChannelVerification}
\usage{
CreateNotificationEmailChannelVerification(
  emailAddress,
  name,
  channelType,
  orgId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Sending a 6 digit verification code to user's email
}
\details{
Sending a 6 digit verification code to user's email

This method invokes \verb{POST /notificationEmailChannelVerification/} in the DataRobot Public API.
}
\seealso{
The method NotificationEmailChannelVerificationCreate in \link{NotificationsApi}, which invokes the same underlying API endpoint.
}
