% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{CreateNotificationPolicies}
\alias{CreateNotificationPolicies}
\title{CreateNotificationPolicies}
\usage{
CreateNotificationPolicies(
  name,
  active,
  eventType,
  eventGroup,
  orgId,
  channelId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create notification policy
}
\details{
Create a new notification policy.

This method invokes \verb{POST /notificationPolicies/} in the DataRobot Public API.
}
\seealso{
The method NotificationPoliciesCreate in \link{NotificationsApi}, which invokes the same underlying API endpoint.
}
