% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{CreateOneProjectsBiasMitigationFeatureInfo}
\alias{CreateOneProjectsBiasMitigationFeatureInfo}
\title{CreateOneProjectsBiasMitigationFeatureInfo}
\usage{
CreateOneProjectsBiasMitigationFeatureInfo(featureName, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Submit a job to create bias mitigation data quality information for a given projectId and featureName.
}
\details{
Submit a job to create bias mitigation data quality information for a given projectId and featureName.

This method invokes \verb{POST /projects/\{projectId\}/biasMitigationFeatureInfo/\{featureName\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsBiasMitigationFeatureInfoCreateOne in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
