% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prediction_from_dataset.R
\docType{class}
\name{CreatePredictionFromDataset}
\alias{CreatePredictionFromDataset}
\title{CreatePredictionFromDataset}
\format{
An \code{R6Class} generator object
}
\description{
CreatePredictionFromDataset Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actualValueColumn}}{character \link{optional} For time series projects only. Actual value column name, valid for the prediction files if the project is unsupervised and the dataset is considered as bulk predictions dataset. This value is optional.}

\item{\code{datasetId}}{character The dataset to compute predictions for - must have previously been uploaded.}

\item{\code{explanationAlgorithm}}{character \link{optional} If set to `shap`, the response will include prediction explanations based on the SHAP explainer (SHapley Additive exPlanations). Defaults to null (no prediction explanations).}

\item{\code{forecastPoint}}{character \link{optional} For time series projects only. The time in the dataset relative to which predictions are generated. This value is optional. If not specified the default value is the value in the row with the latest specified timestamp. Specifying this value for a project that is not a time series project will result in an error.}

\item{\code{includeFdwCounts}}{character \link{optional} For time series projects with partial history only. Indicates if feature derivation window counts `featureDerivationWindowCounts` will be part of the response.}

\item{\code{includePredictionIntervals}}{character \link{optional} Specifies whether prediction intervals should be calculated for this request. Defaults to True if `predictionIntervalsSize` is specified, otherwise defaults to False.}

\item{\code{maxExplanations}}{integer \link{optional} Specifies the maximum number of explanation values that should be returned for each row, ordered by absolute value, greatest to least. In the case of 'shap': If not set, explanations are returned for all features. If the number of features is greater than the 'maxExplanations', the sum of remaining values will also be returned as 'shapRemainingTotal'. Defaults to null for datasets narrower than 100 columns, defaults to 100 for datasets wider than 100 columns. Cannot be set if 'explanationAlgorithm' is omitted.}

\item{\code{modelId}}{character The model to make predictions on.}

\item{\code{predictionIntervalsSize}}{integer \link{optional} Represents the percentile to use for the size of the prediction intervals. Defaults to 80 if `includePredictionIntervals` is True.}

\item{\code{predictionThreshold}}{numeric \link{optional} Threshold used for binary classification in predictions. Accepts values from 0.0 to 1.0. If not specified, model default prediction threshold will be used.}

\item{\code{predictionsEndDate}}{character \link{optional} The end date for bulk predictions, exclusive. Used for time series projects only. Note that this parameter is used for generating historical predictions using the training data, not for future predictions. If not specified, the dataset is not considered as a bulk predictions dataset. This parameter should be provided in conjunction with a ``predictionsStartDate``, and cannot be provided with the ``forecastPoint`` parameter.}

\item{\code{predictionsStartDate}}{character \link{optional} The start date for bulk predictions. Used for time series projects only. Note that this parameter is used for generating historical predictions using the training data, not for future predictions. If not specified, the dataset is not considered as a bulk predictions dataset. This parameter should be provided in conjunction with a ``predictionsEndDate``, and cannot be provided with the ``forecastPoint`` parameter.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CreatePredictionFromDataset$new()}}
\item \href{#method-validate}{\code{CreatePredictionFromDataset$validate()}}
\item \href{#method-toJSON}{\code{CreatePredictionFromDataset$toJSON()}}
\item \href{#method-fromJSON}{\code{CreatePredictionFromDataset$fromJSON()}}
\item \href{#method-clone}{\code{CreatePredictionFromDataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreatePredictionFromDataset$new(
  datasetId = NULL,
  modelId = NULL,
  actualValueColumn = NULL,
  explanationAlgorithm = NULL,
  forecastPoint = NULL,
  includeFdwCounts = FALSE,
  includePredictionIntervals = NULL,
  maxExplanations = NULL,
  predictionIntervalsSize = NULL,
  predictionThreshold = NULL,
  predictionsEndDate = NULL,
  predictionsStartDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasetId}}{The dataset to compute predictions for - must have previously been uploaded.}

\item{\code{modelId}}{The model to make predictions on.}

\item{\code{actualValueColumn}}{For time series projects only. Actual value column name, valid for the prediction files if the project is unsupervised and the dataset is considered as bulk predictions dataset. This value is optional.}

\item{\code{explanationAlgorithm}}{If set to `shap`, the response will include prediction explanations based on the SHAP explainer (SHapley Additive exPlanations). Defaults to null (no prediction explanations).}

\item{\code{forecastPoint}}{For time series projects only. The time in the dataset relative to which predictions are generated. This value is optional. If not specified the default value is the value in the row with the latest specified timestamp. Specifying this value for a project that is not a time series project will result in an error.}

\item{\code{includeFdwCounts}}{For time series projects with partial history only. Indicates if feature derivation window counts `featureDerivationWindowCounts` will be part of the response.}

\item{\code{includePredictionIntervals}}{Specifies whether prediction intervals should be calculated for this request. Defaults to True if `predictionIntervalsSize` is specified, otherwise defaults to False.}

\item{\code{maxExplanations}}{Specifies the maximum number of explanation values that should be returned for each row, ordered by absolute value, greatest to least. In the case of 'shap': If not set, explanations are returned for all features. If the number of features is greater than the 'maxExplanations', the sum of remaining values will also be returned as 'shapRemainingTotal'. Defaults to null for datasets narrower than 100 columns, defaults to 100 for datasets wider than 100 columns. Cannot be set if 'explanationAlgorithm' is omitted.}

\item{\code{predictionIntervalsSize}}{Represents the percentile to use for the size of the prediction intervals. Defaults to 80 if `includePredictionIntervals` is True.}

\item{\code{predictionThreshold}}{Threshold used for binary classification in predictions. Accepts values from 0.0 to 1.0. If not specified, model default prediction threshold will be used.}

\item{\code{predictionsEndDate}}{The end date for bulk predictions, exclusive. Used for time series projects only. Note that this parameter is used for generating historical predictions using the training data, not for future predictions. If not specified, the dataset is not considered as a bulk predictions dataset. This parameter should be provided in conjunction with a ``predictionsStartDate``, and cannot be provided with the ``forecastPoint`` parameter.}

\item{\code{predictionsStartDate}}{The start date for bulk predictions. Used for time series projects only. Note that this parameter is used for generating historical predictions using the training data, not for future predictions. If not specified, the dataset is not considered as a bulk predictions dataset. This parameter should be provided in conjunction with a ``predictionsEndDate``, and cannot be provided with the ``forecastPoint`` parameter.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreatePredictionFromDataset$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreatePredictionFromDataset$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreatePredictionFromDataset$fromJSON(
  CreatePredictionFromDatasetJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CreatePredictionFromDatasetJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreatePredictionFromDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
