% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{CreatePrimeCode}
\alias{CreatePrimeCode}
\title{CreatePrimeCode}
\usage{
CreatePrimeCode(project, primeModelId, language, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a Prime File
}
\details{
Request creation and validation of source code from a Prime model

This method invokes \verb{POST /projects/\{projectId\}/primeFiles/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPrimeFilesCreate in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
