% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{CreateProjectsDatetimeModelsFeatureFit}
\alias{CreateProjectsDatetimeModelsFeatureFit}
\title{CreateProjectsDatetimeModelsFeatureFit}
\usage{
CreateProjectsDatetimeModelsFeatureFit(backtestIndex, modelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Add a request to the queue to calculate Feature Fit for a backtest.
}
\details{
NOTE: Starting from API v2.29 this endpoint will be deprecated. Add a request to the queue to calculate Feature Fit for a backtest. If the job has been previously submitted, the request fails, returning the jobId of the previously submitted job. Use this jobId to check status of the previously submitted job.

This method invokes \verb{POST /projects/\{projectId\}/datetimeModels/\{modelId\}/featureFit/} in the DataRobot Public API.
}
\seealso{
The method ProjectsDatetimeModelsFeatureFitCreate in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
