% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{CreateProjectsDatetimeModelsFromModel}
\alias{CreateProjectsDatetimeModelsFromModel}
\title{CreateProjectsDatetimeModelsFromModel}
\usage{
CreateProjectsDatetimeModelsFromModel(
  timeWindowSamplePct,
  monotonicIncreasingFeaturelistId,
  monotonicDecreasingFeaturelistId,
  trainingDuration,
  trainingEndDate,
  trainingStartDate,
  modelId,
  trainingRowCount,
  samplingMethod,
  featurelistId,
  projectId,
  useProjectSettings,
  nClusters,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrain an existing datetime model with specified parameters.
}
\details{
Retrain an existing datetime model using a new training period for the model training set (with optional time window sampling) or different feature list.  All durations and datetimes should be specified in accordance with the \verb{timestamp and duration formatting rules<time_format>}.  Note that only one of trainingDuration or trainingRowCount or trainingStartDate and trainingEndDate should be specified. If trainingStartDate and trainingEndDate are specified, the source model must be frozen.

This method invokes \verb{POST /projects/\{projectId\}/datetimeModels/fromModel/} in the DataRobot Public API.
}
\seealso{
The method ProjectsDatetimeModelsFromModelCreate in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
