% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_partitioning_apiWrapper.R
\name{CreateProjectsDatetimePartitioning}
\alias{CreateProjectsDatetimePartitioning}
\title{CreateProjectsDatetimePartitioning}
\usage{
CreateProjectsDatetimePartitioning(
  aggregationType,
  holdoutEndDate,
  featureDerivationWindowEnd,
  autopilotDataSelectionMethod,
  periodicities,
  datetimePartitionColumn,
  featureSettings,
  treatAsExponential,
  differencingMethod,
  multiseriesIdColumns,
  forecastWindowStart,
  unsupervisedType,
  defaultToKnownInAdvance,
  isHoldoutModified,
  backtests,
  gapDuration,
  defaultToAPriori,
  defaultToDoNotDerive,
  autopilotClusterList,
  forecastWindowEnd,
  calendarId,
  crossSeriesGroupByColumns,
  windowsBasisUnit,
  holdoutDuration,
  holdoutStartDate,
  validationDuration,
  numberOfBacktests,
  projectId,
  featureDerivationWindowStart,
  allowPartialHistoryTimeSeriesPredictions = FALSE,
  modelSplits = 5,
  useCrossSeriesFeatures = FALSE,
  disableHoldout = FALSE,
  unsupervisedMode = FALSE,
  useSupervisedFeatureReduction = TRUE,
  useTimeSeries = FALSE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Preview the fully specified datetime partitioning generated by the requested configuration.
}
\details{
Preview the fully specified datetime partitioning generated by the requested configuration.  Populates the full datetime partitioning that would be used if the same arguments were passed to :\url{http:patch}:/api/v2/projects/(projectId)/aim/ based on the requested configuration, generating defaults for all non-specified values, so that potential configurations can be tested prior to setting the target and applying a configuration.  useTimeSeries controls whether a time series project should be created or a normal project that uses datetime partitioning. See \verb{Time-Series Projects<time_series_overview>} for more detail on the differences between time series projects and datetime partitioned projects. Time-series projects are only available to some users and use the additional settings of featureDerivationWindowStart and featureDerivationWindowEnd to establish feature derivation window and forecastWindowStart and forecastWindowEnd to establish a forecast window. The overview referenced above provides more information about using feature derivation and forecast windows.  When specifying a feature derivation window of a forecast window, the number of units it spans (end - start) must be an integer multiple of the timeStep of the datetimePartitionColumn.  All durations and datetimes should be specified in accordance with the \verb{timestamp and duration formatting rules<time_format>}.

This method invokes \verb{POST /projects/\{projectId\}/datetimePartitioning/} in the DataRobot Public API.
}
\seealso{
The method ProjectsDatetimePartitioningCreate in \link{DatetimePartitioningApi}, which invokes the same underlying API endpoint.
}
