% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{CreateProjectsModelsClusterInsights}
\alias{CreateProjectsModelsClusterInsights}
\title{CreateProjectsModelsClusterInsights}
\usage{
CreateProjectsModelsClusterInsights(modelId, body, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Compute Cluster Insights.
}
\details{
Compute Cluster Insights for a clustering project model.The number of features computed for cluster insights are capped at 100, starting with the features used to train the model sorted by feature impact (high to low), and then the remaining features in the dataset alphabetically.

This method invokes \verb{POST /projects/\{projectId\}/models/\{modelId\}/clusterInsights/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsClusterInsightsCreate in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
