% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{CreateProjectsModelsFairnessInsights}
\alias{CreateProjectsModelsFairnessInsights}
\title{CreateProjectsModelsFairnessInsights}
\usage{
CreateProjectsModelsFairnessInsights(
  fairnessMetricsSet,
  modelId,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Start insight calculations.
}
\details{
Submits a job to start Per Class Bias insight calculations for the model.

This method invokes \verb{POST /projects/\{projectId\}/models/\{modelId\}/fairnessInsights/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsFairnessInsightsCreate in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
