% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{CreateProjectsModelsFeatureFit}
\alias{CreateProjectsModelsFeatureFit}
\title{CreateProjectsModelsFeatureFit}
\usage{
CreateProjectsModelsFeatureFit(modelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Add a request to the queue to calculate Feature Fit.
}
\details{
NOTE: Starting from API v2.29 this endpoint will be deprecated. Add a request to the queue to calculate Feature Fit. If the job has been previously submitted, the request fails, returning the jobId of the previously submitted job. Use this jobId to check status of the previously submitted job.

This method invokes \verb{POST /projects/\{projectId\}/models/\{modelId\}/featureFit/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsFeatureFitCreate in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
