% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R, R/insights_apiWrapper.R
\name{CreateProjectsModelsImageActivationMaps}
\alias{CreateProjectsModelsImageActivationMaps}
\title{CreateProjectsModelsImageActivationMaps}
\usage{
CreateProjectsModelsImageActivationMaps(modelId, projectId, ...)

CreateProjectsModelsImageActivationMaps(modelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Request the computation of image activation maps for the specified model.

Request the computation of image activation maps for the specified model.
}
\details{
Request the computation of image activation maps for the specified model.

This method invokes \verb{POST /projects/\{projectId\}/models/\{modelId\}/imageActivationMaps/} in the DataRobot Public API.

Request the computation of image activation maps for the specified model.

This method invokes \verb{POST /projects/\{projectId\}/models/\{modelId\}/imageActivationMaps/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsImageActivationMapsCreate in \link{ImagesApi}, which invokes the same underlying API endpoint.

The method ProjectsModelsImageActivationMapsCreate in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
