% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R, R/insights_apiWrapper.R
\name{CreateProjectsModelsImageEmbeddings}
\alias{CreateProjectsModelsImageEmbeddings}
\title{CreateProjectsModelsImageEmbeddings}
\usage{
CreateProjectsModelsImageEmbeddings(modelId, projectId, ...)

CreateProjectsModelsImageEmbeddings(modelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Request the computation of image embeddings for the specified model.

Request the computation of image embeddings for the specified model.
}
\details{
Request the computation of image embeddings for the specified model.

This method invokes \verb{POST /projects/\{projectId\}/models/\{modelId\}/imageEmbeddings/} in the DataRobot Public API.

Request the computation of image embeddings for the specified model.

This method invokes \verb{POST /projects/\{projectId\}/models/\{modelId\}/imageEmbeddings/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsImageEmbeddingsCreate in \link{ImagesApi}, which invokes the same underlying API endpoint.

The method ProjectsModelsImageEmbeddingsCreate in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
