% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{CreateProjectsModelsPredictionIntervals}
\alias{CreateProjectsModelsPredictionIntervals}
\title{CreateProjectsModelsPredictionIntervals}
\usage{
CreateProjectsModelsPredictionIntervals(percentiles, modelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Calculate prediction intervals for the specified percentiles for this model.
}
\details{
Submit a job to calculate prediction intervals for the specified percentiles for this model. Note that the project this model belongs to must be a time series project.

This method invokes \verb{POST /projects/\{projectId\}/models/\{modelId\}/predictionIntervals/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsPredictionIntervalsCreate in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
