% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{CreateProjectsSecondaryDatasetsConfigurations}
\alias{CreateProjectsSecondaryDatasetsConfigurations}
\title{CreateProjectsSecondaryDatasetsConfigurations}
\usage{
CreateProjectsSecondaryDatasetsConfigurations(
  modelId,
  name,
  featurelistId,
  config,
  projectId,
  secondaryDatasets,
  save = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create secondary dataset configurations for a project.
}
\details{
Create secondary dataset configurations for a project.

This method invokes \verb{POST /projects/\{projectId\}/secondaryDatasetsConfigurations/} in the DataRobot Public API.
}
\seealso{
The method SecondaryDatasetsConfigurationsCreate in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
