% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{CreateProjectsShapMatrices}
\alias{CreateProjectsShapMatrices}
\title{CreateProjectsShapMatrices}
\usage{
CreateProjectsShapMatrices(modelId, datasetId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Calculate a matrix with SHAP based prediction explanations scores.
}
\details{
Submit a request to calculate a matrix with SHAP based prediction explanations scores.

This method invokes \verb{POST /projects/\{projectId\}/shapMatrices/} in the DataRobot Public API.
}
\seealso{
The method ProjectsShapMatricesCreate in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
