% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{CreateRatingTable}
\alias{CreateRatingTable}
\title{CreateRatingTable}
\usage{
CreateRatingTable(
  project,
  parentModelId,
  dataSource,
  ratingTableName = "UploadedRatingTable",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Upload Modified Rating Table File
}
\details{
Create a new rating table from a rating table file This will create a new rating table, regardless of whether the validation succeeds. The rating table object will have a validationError which will be left blank in the case of successful validation.

This method invokes \verb{POST /projects/\{projectId\}/ratingTables/} in the DataRobot Public API.
}
\seealso{
The method ProjectsRatingTablesCreate in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
