% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{CreateRelationshipsConfigurations}
\alias{CreateRelationshipsConfigurations}
\title{CreateRelationshipsConfigurations}
\usage{
CreateRelationshipsConfigurations(
  relationships,
  featureDiscoverySettings,
  datasetDefinitions,
  featureDiscoveryMode,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a relationships configuration
}
\details{
Create a relationships configuration

This method invokes \verb{POST /relationshipsConfigurations/} in the DataRobot Public API.
}
\seealso{
The method RelationshipsConfigurationsCreate in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
