% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_training_prediction.R
\docType{class}
\name{CreateTrainingPrediction}
\alias{CreateTrainingPrediction}
\title{CreateTrainingPrediction}
\format{
An \code{R6Class} generator object
}
\description{
CreateTrainingPrediction Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dataSubset}}{character Subset of data predicted on: The value \"all\" returns predictions for all rows in the dataset including data used for training, validation, holdout and any rows discarded. This is not available for large datasets or projects created with Date/Time partitioning. The value \"validationAndHoldout\" returns predictions for the rows used to calculate the validation score and the holdout score. Not available for large projects or Date/Time projects for models trained into the validation set. The value \"holdout\" returns predictions for the rows used to calculate the holdout score. Not available for projects created without a holdout or for models trained into holdout for large datasets or created with Date/Time partitioning. The value \"allBacktests\" returns predictions for the rows used to calculate the backtesting scores for Date/Time projects. The value \"validation\" returns predictions for the rows used to calculate the validation score.}

\item{\code{explanationAlgorithm}}{character \link{optional} If set to \"shap\", the response will include prediction explanations based on the SHAP explainer (SHapley Additive exPlanations). Defaults to null (no prediction explanations)}

\item{\code{maxExplanations}}{integer \link{optional} Specifies the maximum number of explanation values that should be returned for each row, ordered by absolute value, greatest to least. In the case of \"shap\": If not set, explanations are returned for all features. If the number of features is greater than the \"maxExplanations\", the sum of remaining values will also be returned as \"shapRemainingTotal\". Defaults to null for datasets narrower than 100 columns, defaults to 100 for datasets wider than 100 columns. Cannot be set if \"explanationAlgorithm\" is omitted.}

\item{\code{modelId}}{character The model to make predictions on}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CreateTrainingPrediction$new()}}
\item \href{#method-validate}{\code{CreateTrainingPrediction$validate()}}
\item \href{#method-toJSON}{\code{CreateTrainingPrediction$toJSON()}}
\item \href{#method-fromJSON}{\code{CreateTrainingPrediction$fromJSON()}}
\item \href{#method-clone}{\code{CreateTrainingPrediction$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateTrainingPrediction$new(
  dataSubset = NULL,
  modelId = NULL,
  explanationAlgorithm = NULL,
  maxExplanations = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataSubset}}{Subset of data predicted on: The value \"all\" returns predictions for all rows in the dataset including data used for training, validation, holdout and any rows discarded. This is not available for large datasets or projects created with Date/Time partitioning. The value \"validationAndHoldout\" returns predictions for the rows used to calculate the validation score and the holdout score. Not available for large projects or Date/Time projects for models trained into the validation set. The value \"holdout\" returns predictions for the rows used to calculate the holdout score. Not available for projects created without a holdout or for models trained into holdout for large datasets or created with Date/Time partitioning. The value \"allBacktests\" returns predictions for the rows used to calculate the backtesting scores for Date/Time projects. The value \"validation\" returns predictions for the rows used to calculate the validation score.}

\item{\code{modelId}}{The model to make predictions on}

\item{\code{explanationAlgorithm}}{If set to \"shap\", the response will include prediction explanations based on the SHAP explainer (SHapley Additive exPlanations). Defaults to null (no prediction explanations)}

\item{\code{maxExplanations}}{Specifies the maximum number of explanation values that should be returned for each row, ordered by absolute value, greatest to least. In the case of \"shap\": If not set, explanations are returned for all features. If the number of features is greater than the \"maxExplanations\", the sum of remaining values will also be returned as \"shapRemainingTotal\". Defaults to null for datasets narrower than 100 columns, defaults to 100 for datasets wider than 100 columns. Cannot be set if \"explanationAlgorithm\" is omitted.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateTrainingPrediction$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateTrainingPrediction$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateTrainingPrediction$fromJSON(
  CreateTrainingPredictionJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CreateTrainingPredictionJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateTrainingPrediction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
