% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_apiWrapper.R
\name{CreateUsageDataExports}
\alias{CreateUsageDataExports}
\title{CreateUsageDataExports}
\usage{
CreateUsageDataExports(
  include,
  noCache,
  start,
  end,
  userId,
  projectId,
  includeReport,
  includeIdentifyingFields = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a customer usage data artifact request. Requires \"CAN_ACCESS_USER_ACTIVITY\" permission.
}
\details{
Create a customer usage data artifact request. Requires \"CAN_ACCESS_USER_ACTIVITY\" permission. Returns async task status URL as a \"Location\" header. Async task status should be polled in order to retrieve artifact id.

This method invokes \verb{POST /usageDataExports/} in the DataRobot Public API.
}
\seealso{
The method UsageDataExportsCreate in \link{AnalyticsApi}, which invokes the same underlying API endpoint.
}
