% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{CreateUseCases}
\alias{CreateUseCases}
\title{CreateUseCases}
\usage{
CreateUseCases(
  owner,
  notes,
  description,
  businessImpact,
  potentialValueTemplate,
  targetDates,
  content,
  feasibility,
  predictionTargets,
  stage,
  realizedValue,
  permissions,
  mentions,
  name,
  commentId,
  id,
  potentialValue,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a new use case.
}
\details{
Create a new use case.

This method invokes \verb{POST /useCases/} in the DataRobot Public API.
}
\seealso{
The method UseCasesCreate in \link{UseCaseApi}, which invokes the same underlying API endpoint.
}
