% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R,
%   R/blueprints_apiWrapper.R
\name{CreateUserBlueprintsBulkValidations}
\alias{CreateUserBlueprintsBulkValidations}
\title{CreateUserBlueprintsBulkValidations}
\usage{
CreateUserBlueprintsBulkValidations(userBlueprintIds, projectId, ...)

CreateUserBlueprintsBulkValidations(userBlueprintIds, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Validate many user blueprints.

Validate many user blueprints.
}
\details{
Validate many user blueprints, optionally using a specific project. Any non-existent or inaccessible user blueprints will be ignored.

This method invokes \verb{POST /userBlueprintsBulkValidations/} in the DataRobot Public API.

Validate many user blueprints, optionally using a specific project. Any non-existent or inaccessible user blueprints will be ignored.

This method invokes \verb{POST /userBlueprintsBulkValidations/} in the DataRobot Public API.
}
\seealso{
The method UserBlueprintsBulkValidationsCreate in \link{AiCatalogApi}, which invokes the same underlying API endpoint.

The method UserBlueprintsBulkValidationsCreate in \link{BlueprintsApi}, which invokes the same underlying API endpoint.
}
