% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R,
%   R/blueprints_apiWrapper.R
\name{CreateUserBlueprintsFromCustomTaskVersionId}
\alias{CreateUserBlueprintsFromCustomTaskVersionId}
\title{CreateUserBlueprintsFromCustomTaskVersionId}
\usage{
CreateUserBlueprintsFromCustomTaskVersionId(
  customTaskVersionId,
  description,
  decompressedBlueprint = FALSE,
  saveToCatalog = TRUE,
  ...
)

CreateUserBlueprintsFromCustomTaskVersionId(
  customTaskVersionId,
  description,
  decompressedBlueprint = FALSE,
  saveToCatalog = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a user blueprint from a single custom task.

Create a user blueprint from a single custom task.
}
\details{
Create a user blueprint from a single custom task.

This method invokes \verb{POST /userBlueprints/fromCustomTaskVersionId/} in the DataRobot Public API.

Create a user blueprint from a single custom task.

This method invokes \verb{POST /userBlueprints/fromCustomTaskVersionId/} in the DataRobot Public API.
}
\seealso{
The method UserBlueprintsFromCustomTaskVersionIdCreate in \link{AiCatalogApi}, which invokes the same underlying API endpoint.

The method UserBlueprintsFromCustomTaskVersionIdCreate in \link{BlueprintsApi}, which invokes the same underlying API endpoint.
}
