% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R,
%   R/blueprints_apiWrapper.R
\name{CreateUserBlueprintsProjectBlueprints}
\alias{CreateUserBlueprintsProjectBlueprints}
\title{CreateUserBlueprintsProjectBlueprints}
\usage{
CreateUserBlueprintsProjectBlueprints(
  userBlueprintIds,
  projectId,
  deleteAfter = FALSE,
  describeFailures = FALSE,
  ...
)

CreateUserBlueprintsProjectBlueprints(
  userBlueprintIds,
  projectId,
  deleteAfter = FALSE,
  describeFailures = FALSE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Add user blueprints to a project.

Add user blueprints to a project.
}
\details{
Add a list of user blueprints, by id, to a specified (by id) project's repository.

This method invokes \verb{POST /userBlueprintsProjectBlueprints/} in the DataRobot Public API.

Add a list of user blueprints, by id, to a specified (by id) project's repository.

This method invokes \verb{POST /userBlueprintsProjectBlueprints/} in the DataRobot Public API.
}
\seealso{
The method UserBlueprintsProjectBlueprintsCreate in \link{AiCatalogApi}, which invokes the same underlying API endpoint.

The method UserBlueprintsProjectBlueprintsCreate in \link{BlueprintsApi}, which invokes the same underlying API endpoint.
}
