% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R,
%   R/blueprints_apiWrapper.R
\name{CreateUserBlueprintsValidations}
\alias{CreateUserBlueprintsValidations}
\title{CreateUserBlueprintsValidations}
\usage{
CreateUserBlueprintsValidations(
  blueprint,
  projectId,
  isInplaceEditor = FALSE,
  ...
)

CreateUserBlueprintsValidations(
  blueprint,
  projectId,
  isInplaceEditor = FALSE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Validate a user blueprint.

Validate a user blueprint.
}
\details{
Validate a user blueprint

This method invokes \verb{POST /userBlueprintsValidations/} in the DataRobot Public API.

Validate a user blueprint

This method invokes \verb{POST /userBlueprintsValidations/} in the DataRobot Public API.
}
\seealso{
The method UserBlueprintsValidationsCreate in \link{AiCatalogApi}, which invokes the same underlying API endpoint.

The method UserBlueprintsValidationsCreate in \link{BlueprintsApi}, which invokes the same underlying API endpoint.
}
