% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{CreateUsers}
\alias{CreateUsers}
\title{CreateUsers}
\usage{
CreateUsers(
  lastName,
  country,
  maxUploadSize,
  externalId,
  maxIdleWorkers,
  language,
  maxGpuWorkers,
  organizationId,
  maxWorkers,
  firstName,
  password,
  maxUploadSizeCatalog,
  requireClickthroughAgreement,
  userType,
  defaultCatalogDatasetSampleSize,
  username,
  expirationDate,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a User and add them to an organisation.
}
\details{
Create a User and add them to an organisation.

This method invokes \verb{POST /users/} in the DataRobot Public API.
}
\seealso{
The method UsersCreate in \link{UserManagementApi}, which invokes the same underlying API endpoint.
}
