% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_associations_data.R
\docType{class}
\name{CredentialsAssociationsData}
\alias{CredentialsAssociationsData}
\title{CredentialsAssociationsData}
\format{
An \code{R6Class} generator object
}
\description{
CredentialsAssociationsData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{isDefault}}{character \link{optional} Whether this credential association with the given object is default for given session user.}

\item{\code{link}}{character Link to get more details about associated object.}

\item{\code{objectId}}{character Associated object ID.}

\item{\code{objectType}}{character Associated object type.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CredentialsAssociationsData-new}{\code{CredentialsAssociationsData$new()}}
\item \href{#method-CredentialsAssociationsData-validate}{\code{CredentialsAssociationsData$validate()}}
\item \href{#method-CredentialsAssociationsData-toJSON}{\code{CredentialsAssociationsData$toJSON()}}
\item \href{#method-CredentialsAssociationsData-fromJSON}{\code{CredentialsAssociationsData$fromJSON()}}
\item \href{#method-CredentialsAssociationsData-clone}{\code{CredentialsAssociationsData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsAssociationsData-new"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsAssociationsData-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsAssociationsData$new(
  link = NULL,
  objectId = NULL,
  objectType = NULL,
  isDefault = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{link}}{Link to get more details about associated object.}

\item{\code{objectId}}{Associated object ID.}

\item{\code{objectType}}{Associated object type.}

\item{\code{isDefault}}{Whether this credential association with the given object is default for given session user.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsAssociationsData-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsAssociationsData-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsAssociationsData$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsAssociationsData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsAssociationsData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsAssociationsData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsAssociationsData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsAssociationsData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsAssociationsData$fromJSON(
  CredentialsAssociationsDataJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CredentialsAssociationsDataJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsAssociationsData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsAssociationsData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsAssociationsData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
