% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_association_update.R
\docType{class}
\name{CredentialsAssociationUpdate}
\alias{CredentialsAssociationUpdate}
\title{CredentialsAssociationUpdate}
\format{
An \code{R6Class} generator object
}
\description{
CredentialsAssociationUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{credentialsToAdd}}{list( \link{CredentialsToAdd} ) \link{optional} Objects to associate with given credentials.}

\item{\code{credentialsToRemove}}{list( character ) \link{optional} Object IDs, each of which identifies an object to be disassociated from this credential. To see which objects are currently associated, see the response from :http:get:`/api/v2/credentials/(credentialId)/associations/`.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CredentialsAssociationUpdate-new}{\code{CredentialsAssociationUpdate$new()}}
\item \href{#method-CredentialsAssociationUpdate-validate}{\code{CredentialsAssociationUpdate$validate()}}
\item \href{#method-CredentialsAssociationUpdate-toJSON}{\code{CredentialsAssociationUpdate$toJSON()}}
\item \href{#method-CredentialsAssociationUpdate-fromJSON}{\code{CredentialsAssociationUpdate$fromJSON()}}
\item \href{#method-CredentialsAssociationUpdate-clone}{\code{CredentialsAssociationUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsAssociationUpdate-new"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsAssociationUpdate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsAssociationUpdate$new(
  credentialsToAdd = NULL,
  credentialsToRemove = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{credentialsToAdd}}{Objects to associate with given credentials.}

\item{\code{credentialsToRemove}}{Object IDs, each of which identifies an object to be disassociated from this credential. To see which objects are currently associated, see the response from :http:get:`/api/v2/credentials/(credentialId)/associations/`.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsAssociationUpdate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsAssociationUpdate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsAssociationUpdate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsAssociationUpdate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsAssociationUpdate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsAssociationUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsAssociationUpdate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsAssociationUpdate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsAssociationUpdate$fromJSON(
  CredentialsAssociationUpdateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CredentialsAssociationUpdateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsAssociationUpdate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsAssociationUpdate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsAssociationUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
