% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_body.R
\docType{class}
\name{CredentialsBody}
\alias{CredentialsBody}
\title{CredentialsBody}
\format{
An \code{R6Class} generator object
}
\description{
CredentialsBody Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{credentialType}}{character Credentials type.}

\item{\code{description}}{character \link{optional} Credentials description.}

\item{\code{name}}{character Credentials name.}

\item{\code{password}}{character Password to store for this credentials.}

\item{\code{snowflakeAccountName}}{character \link{optional} Snowflake account name.}

\item{\code{user}}{character Username to store for this credentials.}

\item{\code{refreshToken}}{character OAUTH token.}

\item{\code{token}}{character OAUTH token.}

\item{\code{awsAccessKeyId}}{character \link{optional} AWS access key ID.}

\item{\code{awsSecretAccessKey}}{character \link{optional} AWS secret access key.}

\item{\code{awsSessionToken}}{character \link{optional} AWS session token.}

\item{\code{gcpKey}}{\link{GCPKey}}

\item{\code{azureConnectionString}}{character Azure connection string.}

\item{\code{clientId}}{character ADLS Gen2 OAuth client ID.}

\item{\code{clientSecret}}{character ADLS Gen2 OAuth client secret.}

\item{\code{oauthIssuerType}}{character \link{optional} OAuth issuer type.}

\item{\code{oauthIssuerUrl}}{character \link{optional} OAuth issuer URL.}

\item{\code{oauthScopes}}{list( character ) ADLS Gen2 OAuth scopes.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CredentialsBody$new()}}
\item \href{#method-validate}{\code{CredentialsBody$validate()}}
\item \href{#method-toJSON}{\code{CredentialsBody$toJSON()}}
\item \href{#method-fromJSON}{\code{CredentialsBody$fromJSON()}}
\item \href{#method-clone}{\code{CredentialsBody$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsBody$new(
  credentialType = NULL,
  name = NULL,
  password = NULL,
  user = NULL,
  refreshToken = NULL,
  token = NULL,
  gcpKey = NULL,
  azureConnectionString = NULL,
  clientId = NULL,
  clientSecret = NULL,
  oauthScopes = NULL,
  description = NULL,
  snowflakeAccountName = NULL,
  awsAccessKeyId = NULL,
  awsSecretAccessKey = NULL,
  awsSessionToken = NULL,
  oauthIssuerType = NULL,
  oauthIssuerUrl = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{credentialType}}{Credentials type.}

\item{\code{name}}{Credentials name.}

\item{\code{password}}{Password to store for this credentials.}

\item{\code{user}}{Username to store for this credentials.}

\item{\code{refreshToken}}{OAUTH token.}

\item{\code{token}}{OAUTH token.}

\item{\code{azureConnectionString}}{Azure connection string.}

\item{\code{clientId}}{ADLS Gen2 OAuth client ID.}

\item{\code{clientSecret}}{ADLS Gen2 OAuth client secret.}

\item{\code{oauthScopes}}{ADLS Gen2 OAuth scopes.}

\item{\code{description}}{Credentials description.}

\item{\code{snowflakeAccountName}}{Snowflake account name.}

\item{\code{awsAccessKeyId}}{AWS access key ID.}

\item{\code{awsSecretAccessKey}}{AWS secret access key.}

\item{\code{awsSessionToken}}{AWS session token.}

\item{\code{oauthIssuerType}}{OAuth issuer type.}

\item{\code{oauthIssuerUrl}}{OAuth issuer URL.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsBody$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsBody$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsBody$fromJSON(CredentialsBodyJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CredentialsBodyJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsBody$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
