% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_body_one_of1.R
\docType{class}
\name{CredentialsBodyOneOf1}
\alias{CredentialsBodyOneOf1}
\title{CredentialsBodyOneOf1}
\format{
An \code{R6Class} generator object
}
\description{
CredentialsBodyOneOf1 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{credentialType}}{character Credentials type.}

\item{\code{description}}{character \link{optional} Credentials description.}

\item{\code{name}}{character Credentials name.}

\item{\code{refreshToken}}{character OAUTH token.}

\item{\code{token}}{character OAUTH token.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CredentialsBodyOneOf1$new()}}
\item \href{#method-validate}{\code{CredentialsBodyOneOf1$validate()}}
\item \href{#method-toJSON}{\code{CredentialsBodyOneOf1$toJSON()}}
\item \href{#method-fromJSON}{\code{CredentialsBodyOneOf1$fromJSON()}}
\item \href{#method-clone}{\code{CredentialsBodyOneOf1$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsBodyOneOf1$new(
  credentialType = NULL,
  name = NULL,
  refreshToken = NULL,
  token = NULL,
  description = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{credentialType}}{Credentials type.}

\item{\code{name}}{Credentials name.}

\item{\code{refreshToken}}{OAUTH token.}

\item{\code{token}}{OAUTH token.}

\item{\code{description}}{Credentials description.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsBodyOneOf1$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsBodyOneOf1$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsBodyOneOf1$fromJSON(
  CredentialsBodyOneOf1Json,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CredentialsBodyOneOf1Json}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsBodyOneOf1$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
