% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_body_one_of5.R
\docType{class}
\name{CredentialsBodyOneOf5}
\alias{CredentialsBodyOneOf5}
\title{CredentialsBodyOneOf5}
\format{
An \code{R6Class} generator object
}
\description{
CredentialsBodyOneOf5 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{clientId}}{character \link{optional} Snowflake OAUTH client ID.}

\item{\code{clientSecret}}{character \link{optional} Snowflake OAUTH client secret.}

\item{\code{credentialType}}{character Credentials type.}

\item{\code{description}}{character \link{optional} Credentials description.}

\item{\code{name}}{character Credentials name.}

\item{\code{oauthIssuerType}}{character \link{optional} OAuth issuer type.}

\item{\code{oauthIssuerUrl}}{character \link{optional} OAuth issuer URL.}

\item{\code{oauthScopes}}{list( character ) \link{optional} OAuth scopes.}

\item{\code{snowflakeAccountName}}{character \link{optional} Snowflake account name.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CredentialsBodyOneOf5$new()}}
\item \href{#method-validate}{\code{CredentialsBodyOneOf5$validate()}}
\item \href{#method-toJSON}{\code{CredentialsBodyOneOf5$toJSON()}}
\item \href{#method-fromJSON}{\code{CredentialsBodyOneOf5$fromJSON()}}
\item \href{#method-clone}{\code{CredentialsBodyOneOf5$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsBodyOneOf5$new(
  credentialType = NULL,
  name = NULL,
  clientId = NULL,
  clientSecret = NULL,
  description = NULL,
  oauthIssuerType = NULL,
  oauthIssuerUrl = NULL,
  oauthScopes = NULL,
  snowflakeAccountName = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{credentialType}}{Credentials type.}

\item{\code{name}}{Credentials name.}

\item{\code{clientId}}{Snowflake OAUTH client ID.}

\item{\code{clientSecret}}{Snowflake OAUTH client secret.}

\item{\code{description}}{Credentials description.}

\item{\code{oauthIssuerType}}{OAuth issuer type.}

\item{\code{oauthIssuerUrl}}{OAuth issuer URL.}

\item{\code{oauthScopes}}{OAuth scopes.}

\item{\code{snowflakeAccountName}}{Snowflake account name.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsBodyOneOf5$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsBodyOneOf5$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsBodyOneOf5$fromJSON(
  CredentialsBodyOneOf5Json,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CredentialsBodyOneOf5Json}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsBodyOneOf5$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
