% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_to_add.R
\docType{class}
\name{CredentialsToAdd}
\alias{CredentialsToAdd}
\title{CredentialsToAdd}
\format{
An \code{R6Class} generator object
}
\description{
CredentialsToAdd Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{objectId}}{character Object ID identifying the object to be associated with the credentials.}

\item{\code{objectType}}{character Type of object associated with the credentials.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CredentialsToAdd-new}{\code{CredentialsToAdd$new()}}
\item \href{#method-CredentialsToAdd-validate}{\code{CredentialsToAdd$validate()}}
\item \href{#method-CredentialsToAdd-toJSON}{\code{CredentialsToAdd$toJSON()}}
\item \href{#method-CredentialsToAdd-fromJSON}{\code{CredentialsToAdd$fromJSON()}}
\item \href{#method-CredentialsToAdd-clone}{\code{CredentialsToAdd$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsToAdd-new"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsToAdd-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsToAdd$new(
  objectId = NULL,
  objectType = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objectId}}{Object ID identifying the object to be associated with the credentials.}

\item{\code{objectType}}{Type of object associated with the credentials.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsToAdd-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsToAdd-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsToAdd$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsToAdd-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsToAdd-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsToAdd$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsToAdd-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsToAdd-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsToAdd$fromJSON(CredentialsToAddJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CredentialsToAddJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsToAdd-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsToAdd-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsToAdd$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
