% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_update.R
\docType{class}
\name{CredentialsUpdate}
\alias{CredentialsUpdate}
\title{CredentialsUpdate}
\format{
An \code{R6Class} generator object
}
\description{
CredentialsUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{awsAccessKeyId}}{character \link{optional} AWS access key ID (applicable for credentialType `s3` only).}

\item{\code{awsSecretAccessKey}}{character \link{optional} The AWS secret access key (applicable for credentialType `s3` only).}

\item{\code{awsSessionToken}}{character \link{optional} The AWS session token (applicable for credentialType `s3` only).}

\item{\code{azureConnectionString}}{character \link{optional} Azure connection string (applicable for credentialType `azure` only).}

\item{\code{clientId}}{character \link{optional} Snowflake OAUTH client ID (applicable for credentialType `snowflake_oauth_user_account` only).}

\item{\code{clientSecret}}{character \link{optional} Snowflake OAUTH client secret (applicable for credentialType `snowflake_oauth_user_account` only).}

\item{\code{description}}{character \link{optional} Description of credentials. If omitted, and name is not omitted, clears any previous description for that name.}

\item{\code{gcpKey}}{\link{GCPKey} \link{optional}}

\item{\code{name}}{character \link{optional} Name of credentials.}

\item{\code{oauthIssuerType}}{character \link{optional} Snowflake IDP issuer type (applicable for credentialType `snowflake_oauth_user_account` only).}

\item{\code{oauthIssuerUrl}}{character \link{optional} Snowflake External IDP issuer URL (applicable for Snowflake External OAUTH connections only).}

\item{\code{oauthScopes}}{list( character ) \link{optional} Snowflake External OAUTH scopes (applicable for Snowflake External OAUTH connections only).}

\item{\code{password}}{character \link{optional} Password to update for this set of credentials (applicable for credentialType `basic` only).}

\item{\code{personalAccessToken}}{character \link{optional} Tableau personal access token (PAT)(applicable for Tableau AUTH connections only).}

\item{\code{refreshToken}}{character \link{optional} OAUTH refresh token (applicable for credentialType `oauth` only).}

\item{\code{snowflakeAccountName}}{character \link{optional} Snowflake account name (applicable for `snowflake_oauth_user_account` only).}

\item{\code{token}}{character \link{optional} OAUTH token (applicable for credentialType `oauth` only).}

\item{\code{tokenName}}{character \link{optional} Tableau token name(applicable for Tableau AUTH connections only).}

\item{\code{user}}{character \link{optional} Username to update for this set of credentials (applicable for credentialType `basic` only).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CredentialsUpdate-new}{\code{CredentialsUpdate$new()}}
\item \href{#method-CredentialsUpdate-validate}{\code{CredentialsUpdate$validate()}}
\item \href{#method-CredentialsUpdate-toJSON}{\code{CredentialsUpdate$toJSON()}}
\item \href{#method-CredentialsUpdate-fromJSON}{\code{CredentialsUpdate$fromJSON()}}
\item \href{#method-CredentialsUpdate-clone}{\code{CredentialsUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsUpdate-new"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsUpdate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsUpdate$new(
  awsAccessKeyId = NULL,
  awsSecretAccessKey = NULL,
  awsSessionToken = NULL,
  azureConnectionString = NULL,
  clientId = NULL,
  clientSecret = NULL,
  description = NULL,
  gcpKey = NULL,
  name = NULL,
  oauthIssuerType = NULL,
  oauthIssuerUrl = NULL,
  oauthScopes = NULL,
  password = NULL,
  personalAccessToken = NULL,
  refreshToken = NULL,
  snowflakeAccountName = NULL,
  token = NULL,
  tokenName = NULL,
  user = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{awsAccessKeyId}}{AWS access key ID (applicable for credentialType `s3` only).}

\item{\code{awsSecretAccessKey}}{The AWS secret access key (applicable for credentialType `s3` only).}

\item{\code{awsSessionToken}}{The AWS session token (applicable for credentialType `s3` only).}

\item{\code{azureConnectionString}}{Azure connection string (applicable for credentialType `azure` only).}

\item{\code{clientId}}{Snowflake OAUTH client ID (applicable for credentialType `snowflake_oauth_user_account` only).}

\item{\code{clientSecret}}{Snowflake OAUTH client secret (applicable for credentialType `snowflake_oauth_user_account` only).}

\item{\code{description}}{Description of credentials. If omitted, and name is not omitted, clears any previous description for that name.}

\item{\code{name}}{Name of credentials.}

\item{\code{oauthIssuerType}}{Snowflake IDP issuer type (applicable for credentialType `snowflake_oauth_user_account` only).}

\item{\code{oauthIssuerUrl}}{Snowflake External IDP issuer URL (applicable for Snowflake External OAUTH connections only).}

\item{\code{oauthScopes}}{Snowflake External OAUTH scopes (applicable for Snowflake External OAUTH connections only).}

\item{\code{password}}{Password to update for this set of credentials (applicable for credentialType `basic` only).}

\item{\code{personalAccessToken}}{Tableau personal access token (PAT)(applicable for Tableau AUTH connections only).}

\item{\code{refreshToken}}{OAUTH refresh token (applicable for credentialType `oauth` only).}

\item{\code{snowflakeAccountName}}{Snowflake account name (applicable for `snowflake_oauth_user_account` only).}

\item{\code{token}}{OAUTH token (applicable for credentialType `oauth` only).}

\item{\code{tokenName}}{Tableau token name(applicable for Tableau AUTH connections only).}

\item{\code{user}}{Username to update for this set of credentials (applicable for credentialType `basic` only).}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsUpdate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsUpdate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsUpdate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsUpdate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsUpdate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsUpdate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsUpdate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsUpdate$fromJSON(CredentialsUpdateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CredentialsUpdateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsUpdate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsUpdate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
