% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_class_accuracy.R
\docType{class}
\name{CrossClassAccuracy}
\alias{CrossClassAccuracy}
\title{CrossClassAccuracy}
\format{
An \code{R6Class} generator object
}
\description{
CrossClassAccuracy Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{feature}}{character The name of the categorical feature.}

\item{\code{modelId}}{character ID of the model for the cross-class accuracy scores.}

\item{\code{perClassAccuracyScores}}{list( \link{PerClassAccuracy} ) An array of metric scores for each class of the feature.}

\item{\code{predictionThreshold}}{numeric Value of the prediction threshold for the model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CrossClassAccuracy-new}{\code{CrossClassAccuracy$new()}}
\item \href{#method-CrossClassAccuracy-validate}{\code{CrossClassAccuracy$validate()}}
\item \href{#method-CrossClassAccuracy-toJSON}{\code{CrossClassAccuracy$toJSON()}}
\item \href{#method-CrossClassAccuracy-fromJSON}{\code{CrossClassAccuracy$fromJSON()}}
\item \href{#method-CrossClassAccuracy-clone}{\code{CrossClassAccuracy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossClassAccuracy-new"></a>}}
\if{latex}{\out{\hypertarget{method-CrossClassAccuracy-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossClassAccuracy$new(
  feature = NULL,
  modelId = NULL,
  perClassAccuracyScores = NULL,
  predictionThreshold = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{The name of the categorical feature.}

\item{\code{modelId}}{ID of the model for the cross-class accuracy scores.}

\item{\code{perClassAccuracyScores}}{An array of metric scores for each class of the feature.}

\item{\code{predictionThreshold}}{Value of the prediction threshold for the model.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossClassAccuracy-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CrossClassAccuracy-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossClassAccuracy$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossClassAccuracy-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CrossClassAccuracy-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossClassAccuracy$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossClassAccuracy-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CrossClassAccuracy-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossClassAccuracy$fromJSON(CrossClassAccuracyJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CrossClassAccuracyJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossClassAccuracy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CrossClassAccuracy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossClassAccuracy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
