% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_series_group_by_columns_list_item.R
\docType{class}
\name{CrossSeriesGroupByColumnsListItem}
\alias{CrossSeriesGroupByColumnsListItem}
\title{CrossSeriesGroupByColumnsListItem}
\format{
An \code{R6Class} generator object
}
\description{
CrossSeriesGroupByColumnsListItem Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{eligibility}}{character Information about the column's eligibility. If the column is not eligible, this will include the reason why.}

\item{\code{isEligible}}{character Indicates whether this column can be used as a group-by column.}

\item{\code{name}}{character The name of the column.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CrossSeriesGroupByColumnsListItem-new}{\code{CrossSeriesGroupByColumnsListItem$new()}}
\item \href{#method-CrossSeriesGroupByColumnsListItem-validate}{\code{CrossSeriesGroupByColumnsListItem$validate()}}
\item \href{#method-CrossSeriesGroupByColumnsListItem-toJSON}{\code{CrossSeriesGroupByColumnsListItem$toJSON()}}
\item \href{#method-CrossSeriesGroupByColumnsListItem-fromJSON}{\code{CrossSeriesGroupByColumnsListItem$fromJSON()}}
\item \href{#method-CrossSeriesGroupByColumnsListItem-clone}{\code{CrossSeriesGroupByColumnsListItem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossSeriesGroupByColumnsListItem-new"></a>}}
\if{latex}{\out{\hypertarget{method-CrossSeriesGroupByColumnsListItem-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSeriesGroupByColumnsListItem$new(
  eligibility = NULL,
  isEligible = NULL,
  name = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{eligibility}}{Information about the column's eligibility. If the column is not eligible, this will include the reason why.}

\item{\code{isEligible}}{Indicates whether this column can be used as a group-by column.}

\item{\code{name}}{The name of the column.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossSeriesGroupByColumnsListItem-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CrossSeriesGroupByColumnsListItem-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSeriesGroupByColumnsListItem$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossSeriesGroupByColumnsListItem-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CrossSeriesGroupByColumnsListItem-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSeriesGroupByColumnsListItem$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossSeriesGroupByColumnsListItem-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CrossSeriesGroupByColumnsListItem-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSeriesGroupByColumnsListItem$fromJSON(
  CrossSeriesGroupByColumnsListItemJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CrossSeriesGroupByColumnsListItemJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossSeriesGroupByColumnsListItem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CrossSeriesGroupByColumnsListItem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSeriesGroupByColumnsListItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
