% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_series_group_by_column_validate_payload.R
\docType{class}
\name{CrossSeriesGroupByColumnValidatePayload}
\alias{CrossSeriesGroupByColumnValidatePayload}
\title{CrossSeriesGroupByColumnValidatePayload}
\format{
An \code{R6Class} generator object
}
\description{
CrossSeriesGroupByColumnValidatePayload Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{crossSeriesGroupByColumns}}{list( character ) \link{optional} If specified, these columns will be validated for usage as the group-by column for creating cross-series features. If not present, then all columns from the dataset will be validated and only the eligible ones returned. To be valid, a column should be categorical or numerical (but not float), not be the series ID or equivalent to the series ID, not split any series, and not consist of only one value.}

\item{\code{datetimePartitionColumn}}{character The name of the column that will be used as the datetime partitioning column.}

\item{\code{multiseriesIdColumn}}{character The name of the column that wil be used as the multiseries ID column for this project.}

\item{\code{userDefinedSegmentIdColumn}}{character \link{optional} The name of the column that wil be used as the user defined segment ID column for this project.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CrossSeriesGroupByColumnValidatePayload-new}{\code{CrossSeriesGroupByColumnValidatePayload$new()}}
\item \href{#method-CrossSeriesGroupByColumnValidatePayload-validate}{\code{CrossSeriesGroupByColumnValidatePayload$validate()}}
\item \href{#method-CrossSeriesGroupByColumnValidatePayload-toJSON}{\code{CrossSeriesGroupByColumnValidatePayload$toJSON()}}
\item \href{#method-CrossSeriesGroupByColumnValidatePayload-fromJSON}{\code{CrossSeriesGroupByColumnValidatePayload$fromJSON()}}
\item \href{#method-CrossSeriesGroupByColumnValidatePayload-clone}{\code{CrossSeriesGroupByColumnValidatePayload$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossSeriesGroupByColumnValidatePayload-new"></a>}}
\if{latex}{\out{\hypertarget{method-CrossSeriesGroupByColumnValidatePayload-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSeriesGroupByColumnValidatePayload$new(
  datetimePartitionColumn = NULL,
  multiseriesIdColumn = NULL,
  crossSeriesGroupByColumns = NULL,
  userDefinedSegmentIdColumn = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datetimePartitionColumn}}{The name of the column that will be used as the datetime partitioning column.}

\item{\code{multiseriesIdColumn}}{The name of the column that wil be used as the multiseries ID column for this project.}

\item{\code{crossSeriesGroupByColumns}}{If specified, these columns will be validated for usage as the group-by column for creating cross-series features. If not present, then all columns from the dataset will be validated and only the eligible ones returned. To be valid, a column should be categorical or numerical (but not float), not be the series ID or equivalent to the series ID, not split any series, and not consist of only one value.}

\item{\code{userDefinedSegmentIdColumn}}{The name of the column that wil be used as the user defined segment ID column for this project.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossSeriesGroupByColumnValidatePayload-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CrossSeriesGroupByColumnValidatePayload-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSeriesGroupByColumnValidatePayload$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossSeriesGroupByColumnValidatePayload-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CrossSeriesGroupByColumnValidatePayload-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSeriesGroupByColumnValidatePayload$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossSeriesGroupByColumnValidatePayload-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CrossSeriesGroupByColumnValidatePayload-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSeriesGroupByColumnValidatePayload$fromJSON(
  CrossSeriesGroupByColumnValidatePayloadJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CrossSeriesGroupByColumnValidatePayloadJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossSeriesGroupByColumnValidatePayload-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CrossSeriesGroupByColumnValidatePayload-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSeriesGroupByColumnValidatePayload$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
