% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cross_series_group_by_column_validate_response.R
\docType{class}
\name{CrossSeriesGroupByColumnValidateResponse}
\alias{CrossSeriesGroupByColumnValidateResponse}
\title{CrossSeriesGroupByColumnValidateResponse}
\format{
An \code{R6Class} generator object
}
\description{
CrossSeriesGroupByColumnValidateResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{message}}{character An extended message about the result. For example, if a job is submitted that is a duplicate of a job that has already been added to the queue, the message will mention that no new job was created.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CrossSeriesGroupByColumnValidateResponse-new}{\code{CrossSeriesGroupByColumnValidateResponse$new()}}
\item \href{#method-CrossSeriesGroupByColumnValidateResponse-validate}{\code{CrossSeriesGroupByColumnValidateResponse$validate()}}
\item \href{#method-CrossSeriesGroupByColumnValidateResponse-toJSON}{\code{CrossSeriesGroupByColumnValidateResponse$toJSON()}}
\item \href{#method-CrossSeriesGroupByColumnValidateResponse-fromJSON}{\code{CrossSeriesGroupByColumnValidateResponse$fromJSON()}}
\item \href{#method-CrossSeriesGroupByColumnValidateResponse-clone}{\code{CrossSeriesGroupByColumnValidateResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossSeriesGroupByColumnValidateResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-CrossSeriesGroupByColumnValidateResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSeriesGroupByColumnValidateResponse$new(
  message = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{An extended message about the result. For example, if a job is submitted that is a duplicate of a job that has already been added to the queue, the message will mention that no new job was created.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossSeriesGroupByColumnValidateResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CrossSeriesGroupByColumnValidateResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSeriesGroupByColumnValidateResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossSeriesGroupByColumnValidateResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CrossSeriesGroupByColumnValidateResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSeriesGroupByColumnValidateResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossSeriesGroupByColumnValidateResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CrossSeriesGroupByColumnValidateResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSeriesGroupByColumnValidateResponse$fromJSON(
  CrossSeriesGroupByColumnValidateResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CrossSeriesGroupByColumnValidateResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossSeriesGroupByColumnValidateResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CrossSeriesGroupByColumnValidateResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSeriesGroupByColumnValidateResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
