% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_boundaries.R
\docType{class}
\name{CustomBoundaries}
\alias{CustomBoundaries}
\title{CustomBoundaries}
\format{
An \code{R6Class} generator object
}
\description{
CustomBoundaries Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lower}}{numeric All predictions less than provided value are considered anomalous.}

\item{\code{upper}}{numeric All predictions greater than provided value are considered anomalous.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CustomBoundaries-new}{\code{CustomBoundaries$new()}}
\item \href{#method-CustomBoundaries-validate}{\code{CustomBoundaries$validate()}}
\item \href{#method-CustomBoundaries-toJSON}{\code{CustomBoundaries$toJSON()}}
\item \href{#method-CustomBoundaries-fromJSON}{\code{CustomBoundaries$fromJSON()}}
\item \href{#method-CustomBoundaries-clone}{\code{CustomBoundaries$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomBoundaries-new"></a>}}
\if{latex}{\out{\hypertarget{method-CustomBoundaries-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomBoundaries$new(lower = NULL, upper = NULL, validateParams = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lower}}{All predictions less than provided value are considered anomalous.}

\item{\code{upper}}{All predictions greater than provided value are considered anomalous.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomBoundaries-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CustomBoundaries-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomBoundaries$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomBoundaries-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomBoundaries-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomBoundaries$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomBoundaries-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomBoundaries-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomBoundaries$fromJSON(CustomBoundariesJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CustomBoundariesJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomBoundaries-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CustomBoundaries-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomBoundaries$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
