% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer_header.R
\docType{class}
\name{CustomerHeader}
\alias{CustomerHeader}
\title{CustomerHeader}
\format{
An \code{R6Class} generator object
}
\description{
CustomerHeader Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character The name of the header.}

\item{\code{value}}{character The value of the header.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CustomerHeader-new}{\code{CustomerHeader$new()}}
\item \href{#method-CustomerHeader-validate}{\code{CustomerHeader$validate()}}
\item \href{#method-CustomerHeader-toJSON}{\code{CustomerHeader$toJSON()}}
\item \href{#method-CustomerHeader-fromJSON}{\code{CustomerHeader$fromJSON()}}
\item \href{#method-CustomerHeader-clone}{\code{CustomerHeader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomerHeader-new"></a>}}
\if{latex}{\out{\hypertarget{method-CustomerHeader-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomerHeader$new(name = NULL, value = NULL, validateParams = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the header.}

\item{\code{value}}{The value of the header.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomerHeader-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CustomerHeader-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomerHeader$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomerHeader-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomerHeader-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomerHeader$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomerHeader-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomerHeader-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomerHeader$fromJSON(CustomerHeaderJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CustomerHeaderJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomerHeader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CustomerHeader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomerHeader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
