% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_model_create.R
\docType{class}
\name{CustomModelCreate}
\alias{CustomModelCreate}
\title{CustomModelCreate}
\format{
An \code{R6Class} generator object
}
\description{
CustomModelCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{calibratePredictions}}{character \link{optional} Whether model predictions should be calibrated by DataRobot.Only applies to anomaly detection training tasks; we recommend this if you have not already included calibration in your model code.Calibration improves the probability estimates of a model, and modifies the predictions of non-probabilistic models to be interpretable as probabilities. This will facilitate comparison to DataRobot models, and give access to ROC curve insights on external data.}

\item{\code{classLabels}}{list( character ) \link{optional} The class labels for multiclass classification. Required for multiclass inference models. If using one of the \link{DataRobot} base environments and your model produces an ndarray of unlabeled class probabilities, the order of the labels should match the order of the predicted output}

\item{\code{customModelType}}{character The type of custom model.}

\item{\code{description}}{character \link{optional} The user-friendly description of the model.}

\item{\code{desiredMemory}}{integer \link{optional} The amount of memory that is expected to be allocated by the custom model.}

\item{\code{language}}{character \link{optional} Programming language name in which model is written.}

\item{\code{maximumMemory}}{integer \link{optional} The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{name}}{character The user-friendly name for the model.}

\item{\code{negativeClassLabel}}{character \link{optional} The negative class label for custom models that support binary classification. If specified, `positiveClassLabel` must also be specified. Default value is \"0\".}

\item{\code{networkEgressPolicy}}{character \link{optional} Network egress policy.}

\item{\code{positiveClassLabel}}{character \link{optional} The positive class label for custom models that support binary classification. If specified, `negativeClassLabel` must also be specified. Default value is \"1\".}

\item{\code{predictionThreshold}}{numeric \link{optional} The prediction threshold which will be used for binary classification custom model.}

\item{\code{replicas}}{integer \link{optional} A fixed number of replicas that will be set for the given custom-model.}

\item{\code{requiresHa}}{character \link{optional} Require all custom model replicas to be deployed on different Kubernetes nodes for predictions fault tolerance.}

\item{\code{supportsBinaryClassification}}{character \link{optional} Whether the model supports binary classification.}

\item{\code{supportsRegression}}{character \link{optional} Whether the model supports regression.}

\item{\code{targetName}}{character \link{optional} The name of the target for labeling predictions. Required for model type 'inference'. Specifying this value for a model type 'training' will result in an error.}

\item{\code{targetType}}{character \link{optional} The target type of the custom model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CustomModelCreate$new()}}
\item \href{#method-validate}{\code{CustomModelCreate$validate()}}
\item \href{#method-toJSON}{\code{CustomModelCreate$toJSON()}}
\item \href{#method-fromJSON}{\code{CustomModelCreate$fromJSON()}}
\item \href{#method-clone}{\code{CustomModelCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelCreate$new(
  customModelType = NULL,
  name = NULL,
  calibratePredictions = TRUE,
  classLabels = NULL,
  description = NULL,
  desiredMemory = NULL,
  language = NULL,
  maximumMemory = NULL,
  negativeClassLabel = NULL,
  networkEgressPolicy = NULL,
  positiveClassLabel = NULL,
  predictionThreshold = 0.5,
  replicas = NULL,
  requiresHa = NULL,
  supportsBinaryClassification = NULL,
  supportsRegression = NULL,
  targetName = NULL,
  targetType = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{customModelType}}{The type of custom model.}

\item{\code{name}}{The user-friendly name for the model.}

\item{\code{calibratePredictions}}{Whether model predictions should be calibrated by DataRobot.Only applies to anomaly detection training tasks; we recommend this if you have not already included calibration in your model code.Calibration improves the probability estimates of a model, and modifies the predictions of non-probabilistic models to be interpretable as probabilities. This will facilitate comparison to DataRobot models, and give access to ROC curve insights on external data.}

\item{\code{classLabels}}{The class labels for multiclass classification. Required for multiclass inference models. If using one of the \link{DataRobot} base environments and your model produces an ndarray of unlabeled class probabilities, the order of the labels should match the order of the predicted output}

\item{\code{description}}{The user-friendly description of the model.}

\item{\code{desiredMemory}}{The amount of memory that is expected to be allocated by the custom model.}

\item{\code{language}}{Programming language name in which model is written.}

\item{\code{maximumMemory}}{The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{negativeClassLabel}}{The negative class label for custom models that support binary classification. If specified, `positiveClassLabel` must also be specified. Default value is \"0\".}

\item{\code{networkEgressPolicy}}{Network egress policy.}

\item{\code{positiveClassLabel}}{The positive class label for custom models that support binary classification. If specified, `negativeClassLabel` must also be specified. Default value is \"1\".}

\item{\code{predictionThreshold}}{The prediction threshold which will be used for binary classification custom model.}

\item{\code{replicas}}{A fixed number of replicas that will be set for the given custom-model.}

\item{\code{requiresHa}}{Require all custom model replicas to be deployed on different Kubernetes nodes for predictions fault tolerance.}

\item{\code{supportsBinaryClassification}}{Whether the model supports binary classification.}

\item{\code{supportsRegression}}{Whether the model supports regression.}

\item{\code{targetName}}{The name of the target for labeling predictions. Required for model type 'inference'. Specifying this value for a model type 'training' will result in an error.}

\item{\code{targetType}}{The target type of the custom model}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelCreate$fromJSON(CustomModelCreateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CustomModelCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
