% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_model_deployment_response.R
\docType{class}
\name{CustomModelDeploymentResponse}
\alias{CustomModelDeploymentResponse}
\title{CustomModelDeploymentResponse}
\format{
An \code{R6Class} generator object
}
\description{
CustomModelDeploymentResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{customModel}}{\link{CustomModelShortResponse}}

\item{\code{customModelImageId}}{character The id of the custom model image associated with this deployment.}

\item{\code{customModelVersion}}{\link{CustomModelVersionShortResponse}}

\item{\code{deployed}}{character ISO-8601 timestamp of when deployment was created.}

\item{\code{deployedBy}}{character The username of the user that deployed the custom model.}

\item{\code{executionEnvironment}}{\link{ExecutionEnvironmentShortResponse}}

\item{\code{executionEnvironmentVersion}}{\link{ExecutionEnvironmentVersionShortResponse}}

\item{\code{id}}{character The ID of the deployment.}

\item{\code{imageType}}{character \link{optional} The type of the image, either customModelImage if the testing attempt is using a customModelImage as its model or customModelVersion if the testing attempt is using a customModelVersion with dependency management.}

\item{\code{label}}{character User-friendly name of the model deployment.}

\item{\code{status}}{character Deployment status.}

\item{\code{testingStatus}}{character Latest testing status of the deployed custom model image.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CustomModelDeploymentResponse$new()}}
\item \href{#method-validate}{\code{CustomModelDeploymentResponse$validate()}}
\item \href{#method-toJSON}{\code{CustomModelDeploymentResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{CustomModelDeploymentResponse$fromJSON()}}
\item \href{#method-clone}{\code{CustomModelDeploymentResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelDeploymentResponse$new(
  customModel = NULL,
  customModelImageId = NULL,
  customModelVersion = NULL,
  deployed = NULL,
  deployedBy = NULL,
  executionEnvironment = NULL,
  executionEnvironmentVersion = NULL,
  id = NULL,
  label = NULL,
  status = NULL,
  testingStatus = NULL,
  imageType = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{customModelImageId}}{The id of the custom model image associated with this deployment.}

\item{\code{deployed}}{ISO-8601 timestamp of when deployment was created.}

\item{\code{deployedBy}}{The username of the user that deployed the custom model.}

\item{\code{id}}{The ID of the deployment.}

\item{\code{label}}{User-friendly name of the model deployment.}

\item{\code{status}}{Deployment status.}

\item{\code{testingStatus}}{Latest testing status of the deployed custom model image.}

\item{\code{imageType}}{The type of the image, either customModelImage if the testing attempt is using a customModelImage as its model or customModelVersion if the testing attempt is using a customModelVersion with dependency management.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelDeploymentResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelDeploymentResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelDeploymentResponse$fromJSON(
  CustomModelDeploymentResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CustomModelDeploymentResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelDeploymentResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
