% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_model_prediction_explanations.R
\docType{class}
\name{CustomModelPredictionExplanations}
\alias{CustomModelPredictionExplanations}
\title{CustomModelPredictionExplanations}
\format{
An \code{R6Class} generator object
}
\description{
CustomModelPredictionExplanations Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{customModelId}}{character ID of the custom model.}

\item{\code{customModelVersionId}}{character ID of the custom model version.}

\item{\code{environmentId}}{character \link{optional} The ID of environment to use. If not specified, the customModelVersion's dependency environment will be used.}

\item{\code{environmentVersionId}}{character \link{optional} The ID of environment version to use. Defaults to the latest successfully built version.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CustomModelPredictionExplanations-new}{\code{CustomModelPredictionExplanations$new()}}
\item \href{#method-CustomModelPredictionExplanations-validate}{\code{CustomModelPredictionExplanations$validate()}}
\item \href{#method-CustomModelPredictionExplanations-toJSON}{\code{CustomModelPredictionExplanations$toJSON()}}
\item \href{#method-CustomModelPredictionExplanations-fromJSON}{\code{CustomModelPredictionExplanations$fromJSON()}}
\item \href{#method-CustomModelPredictionExplanations-clone}{\code{CustomModelPredictionExplanations$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomModelPredictionExplanations-new"></a>}}
\if{latex}{\out{\hypertarget{method-CustomModelPredictionExplanations-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelPredictionExplanations$new(
  customModelId = NULL,
  customModelVersionId = NULL,
  environmentId = NULL,
  environmentVersionId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{customModelId}}{ID of the custom model.}

\item{\code{customModelVersionId}}{ID of the custom model version.}

\item{\code{environmentId}}{The ID of environment to use. If not specified, the customModelVersion's dependency environment will be used.}

\item{\code{environmentVersionId}}{The ID of environment version to use. Defaults to the latest successfully built version.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomModelPredictionExplanations-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CustomModelPredictionExplanations-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelPredictionExplanations$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomModelPredictionExplanations-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomModelPredictionExplanations-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelPredictionExplanations$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomModelPredictionExplanations-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomModelPredictionExplanations-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelPredictionExplanations$fromJSON(
  CustomModelPredictionExplanationsJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CustomModelPredictionExplanationsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomModelPredictionExplanations-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CustomModelPredictionExplanations-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelPredictionExplanations$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
