% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_model_response.R
\docType{class}
\name{CustomModelResponse}
\alias{CustomModelResponse}
\title{CustomModelResponse}
\format{
An \code{R6Class} generator object
}
\description{
CustomModelResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{calibratePredictions}}{character \link{optional} Determines whether ot not predictions should be calibrated by DataRobot.Only applies to anomaly detection.}

\item{\code{classLabels}}{list( character ) \link{optional} If the model is a multiclass classifier, these are the model's class labels}

\item{\code{created}}{character ISO-8601 timestamp of when the model was created.}

\item{\code{createdBy}}{character The username of the custom model creator.}

\item{\code{customModelType}}{character \link{optional} The type of custom model.}

\item{\code{deploymentsCount}}{integer The number of models deployed.}

\item{\code{description}}{character The description of the model.}

\item{\code{desiredMemory}}{integer \link{optional} The amount of memory that is expected to be allocated by the custom model.}

\item{\code{id}}{character The ID of the custom model.}

\item{\code{language}}{character The programming language used to write the model.}

\item{\code{latestVersion}}{\link{CustomModelVersionResponse}}

\item{\code{maximumMemory}}{integer \link{optional} The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{name}}{character The name of the model.}

\item{\code{negativeClassLabel}}{character \link{optional} If the model is a binary classifier, this is the negative class label.}

\item{\code{networkEgressPolicy}}{character \link{optional} Network egress policy.}

\item{\code{positiveClassLabel}}{character \link{optional} If the model is a binary classifier, this is the positive class label.}

\item{\code{predictionThreshold}}{numeric \link{optional} If the model is a binary classifier, this is the prediction threshold.}

\item{\code{replicas}}{integer \link{optional} A fixed number of replicas that will be set for the given custom-model.}

\item{\code{requiresHa}}{character \link{optional} Require all custom model replicas to be deployed on different Kubernetes nodes for predictions fault tolerance.}

\item{\code{supportsAnomalyDetection}}{character \link{optional} Whether the model supports anomaly detection.}

\item{\code{supportsBinaryClassification}}{character Whether the model supports binary classification.}

\item{\code{supportsRegression}}{character Whether the model supports regression.}

\item{\code{targetName}}{character \link{optional} The name of the target for labeling predictions.}

\item{\code{targetType}}{character The target type of custom model.}

\item{\code{trainingDataAssignmentInProgress}}{character \link{optional} Indicates if training data is currently being assigned to the custom model.}

\item{\code{trainingDataFileName}}{character \link{optional} The name of the file that was used as training data if it was assigned previously.}

\item{\code{trainingDataPartitionColumn}}{character \link{optional} The name of the column containing the partition assignments in training data if it was assigned previously and partitioning was provided.}

\item{\code{trainingDatasetId}}{character \link{optional} The ID of the dataset that was used as training data if it was assigned previously.}

\item{\code{trainingDatasetVersionId}}{character \link{optional} The ID of the dataset version that was used as training data if it was assigned previously.}

\item{\code{updated}}{character ISO-8601 timestamp of when model was last updated.}

\item{\code{userProvidedId}}{character \link{optional} A user-provided unique ID associated with the given custom inference model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CustomModelResponse-new}{\code{CustomModelResponse$new()}}
\item \href{#method-CustomModelResponse-validate}{\code{CustomModelResponse$validate()}}
\item \href{#method-CustomModelResponse-toJSON}{\code{CustomModelResponse$toJSON()}}
\item \href{#method-CustomModelResponse-fromJSON}{\code{CustomModelResponse$fromJSON()}}
\item \href{#method-CustomModelResponse-clone}{\code{CustomModelResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomModelResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-CustomModelResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelResponse$new(
  created = NULL,
  createdBy = NULL,
  deploymentsCount = NULL,
  description = NULL,
  id = NULL,
  language = NULL,
  latestVersion = NULL,
  name = NULL,
  supportsBinaryClassification = NULL,
  supportsRegression = NULL,
  targetType = NULL,
  updated = NULL,
  calibratePredictions = NULL,
  classLabels = NULL,
  customModelType = NULL,
  desiredMemory = NULL,
  maximumMemory = NULL,
  negativeClassLabel = NULL,
  networkEgressPolicy = NULL,
  positiveClassLabel = NULL,
  predictionThreshold = NULL,
  replicas = NULL,
  requiresHa = NULL,
  supportsAnomalyDetection = NULL,
  targetName = NULL,
  trainingDataAssignmentInProgress = NULL,
  trainingDataFileName = NULL,
  trainingDataPartitionColumn = NULL,
  trainingDatasetId = NULL,
  trainingDatasetVersionId = NULL,
  userProvidedId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{created}}{ISO-8601 timestamp of when the model was created.}

\item{\code{createdBy}}{The username of the custom model creator.}

\item{\code{deploymentsCount}}{The number of models deployed.}

\item{\code{description}}{The description of the model.}

\item{\code{id}}{The ID of the custom model.}

\item{\code{language}}{The programming language used to write the model.}

\item{\code{name}}{The name of the model.}

\item{\code{supportsBinaryClassification}}{Whether the model supports binary classification.}

\item{\code{supportsRegression}}{Whether the model supports regression.}

\item{\code{targetType}}{The target type of custom model.}

\item{\code{updated}}{ISO-8601 timestamp of when model was last updated.}

\item{\code{calibratePredictions}}{Determines whether ot not predictions should be calibrated by DataRobot.Only applies to anomaly detection.}

\item{\code{classLabels}}{If the model is a multiclass classifier, these are the model's class labels}

\item{\code{customModelType}}{The type of custom model.}

\item{\code{desiredMemory}}{The amount of memory that is expected to be allocated by the custom model.}

\item{\code{maximumMemory}}{The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{negativeClassLabel}}{If the model is a binary classifier, this is the negative class label.}

\item{\code{networkEgressPolicy}}{Network egress policy.}

\item{\code{positiveClassLabel}}{If the model is a binary classifier, this is the positive class label.}

\item{\code{predictionThreshold}}{If the model is a binary classifier, this is the prediction threshold.}

\item{\code{replicas}}{A fixed number of replicas that will be set for the given custom-model.}

\item{\code{requiresHa}}{Require all custom model replicas to be deployed on different Kubernetes nodes for predictions fault tolerance.}

\item{\code{supportsAnomalyDetection}}{Whether the model supports anomaly detection.}

\item{\code{targetName}}{The name of the target for labeling predictions.}

\item{\code{trainingDataAssignmentInProgress}}{Indicates if training data is currently being assigned to the custom model.}

\item{\code{trainingDataFileName}}{The name of the file that was used as training data if it was assigned previously.}

\item{\code{trainingDataPartitionColumn}}{The name of the column containing the partition assignments in training data if it was assigned previously and partitioning was provided.}

\item{\code{trainingDatasetId}}{The ID of the dataset that was used as training data if it was assigned previously.}

\item{\code{trainingDatasetVersionId}}{The ID of the dataset version that was used as training data if it was assigned previously.}

\item{\code{userProvidedId}}{A user-provided unique ID associated with the given custom inference model.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomModelResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CustomModelResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomModelResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomModelResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomModelResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomModelResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelResponse$fromJSON(CustomModelResponseJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CustomModelResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomModelResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CustomModelResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
