% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_model_tests.R
\docType{class}
\name{CustomModelTests}
\alias{CustomModelTests}
\title{CustomModelTests}
\format{
An \code{R6Class} generator object
}
\description{
CustomModelTests Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{configuration}}{\link{CustomModelTestsConfig} \link{optional}}

\item{\code{customModelId}}{character The ID of the custom model to test.}

\item{\code{customModelVersionId}}{character The ID of custom model version to use.}

\item{\code{datasetId}}{character \link{optional} The ID of the dataset to use for testing. Dataset ID is required for regular (non-unstructured) custom models.}

\item{\code{datasetVersionId}}{character \link{optional} The ID of the version of the dataset item to use as the testing dataset. Defaults to the latest version.}

\item{\code{desiredMemory}}{integer \link{optional} The amount of memory that is expected to be allocated by the custom model.}

\item{\code{environmentId}}{character \link{optional} The ID of environment to use. If not specified, the customModelVersion's dependency environment will be used.}

\item{\code{environmentVersionId}}{character \link{optional} The ID of environment version to use. Defaults to the latest successfully built version.}

\item{\code{maximumMemory}}{integer \link{optional} The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{networkEgressPolicy}}{character \link{optional} Network egress policy.}

\item{\code{replicas}}{integer \link{optional} A fixed number of replicas that will be set for the given custom-model.}

\item{\code{requiresHa}}{character \link{optional} Require all custom model replicas to be deployed on different Kubernetes nodes for predictions fault tolerance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CustomModelTests$new()}}
\item \href{#method-validate}{\code{CustomModelTests$validate()}}
\item \href{#method-toJSON}{\code{CustomModelTests$toJSON()}}
\item \href{#method-fromJSON}{\code{CustomModelTests$fromJSON()}}
\item \href{#method-clone}{\code{CustomModelTests$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelTests$new(
  customModelId = NULL,
  customModelVersionId = NULL,
  configuration = NULL,
  datasetId = NULL,
  datasetVersionId = NULL,
  desiredMemory = NULL,
  environmentId = NULL,
  environmentVersionId = NULL,
  maximumMemory = NULL,
  networkEgressPolicy = NULL,
  replicas = NULL,
  requiresHa = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{customModelId}}{The ID of the custom model to test.}

\item{\code{customModelVersionId}}{The ID of custom model version to use.}

\item{\code{datasetId}}{The ID of the dataset to use for testing. Dataset ID is required for regular (non-unstructured) custom models.}

\item{\code{datasetVersionId}}{The ID of the version of the dataset item to use as the testing dataset. Defaults to the latest version.}

\item{\code{desiredMemory}}{The amount of memory that is expected to be allocated by the custom model.}

\item{\code{environmentId}}{The ID of environment to use. If not specified, the customModelVersion's dependency environment will be used.}

\item{\code{environmentVersionId}}{The ID of environment version to use. Defaults to the latest successfully built version.}

\item{\code{maximumMemory}}{The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{networkEgressPolicy}}{Network egress policy.}

\item{\code{replicas}}{A fixed number of replicas that will be set for the given custom-model.}

\item{\code{requiresHa}}{Require all custom model replicas to be deployed on different Kubernetes nodes for predictions fault tolerance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelTests$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelTests$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelTests$fromJSON(CustomModelTestsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CustomModelTestsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelTests$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
