% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_model_tests_config.R
\docType{class}
\name{CustomModelTestsConfig}
\alias{CustomModelTestsConfig}
\title{CustomModelTestsConfig}
\format{
An \code{R6Class} generator object
}
\description{
CustomModelTestsConfig Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{errorCheck}}{character \link{optional} Ensures that the model can make predictions on the provided test dataset.}

\item{\code{longRunningService}}{character \link{optional} Ensures that the custom model image can build and launch. If it cannot, the test is marked as Failed and subsequent test are aborted.}

\item{\code{nullValueImputation}}{character \link{optional} Verifies that the model can impute null values. Required for Feature Impact.}

\item{\code{sideEffects}}{character \link{optional} Verifies that predictions made on the dataset match row-wise predictions for the same dataset. Fails if the predictions do not match.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CustomModelTestsConfig$new()}}
\item \href{#method-validate}{\code{CustomModelTestsConfig$validate()}}
\item \href{#method-toJSON}{\code{CustomModelTestsConfig$toJSON()}}
\item \href{#method-fromJSON}{\code{CustomModelTestsConfig$fromJSON()}}
\item \href{#method-clone}{\code{CustomModelTestsConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelTestsConfig$new(
  errorCheck = "fail",
  longRunningService = "fail",
  nullValueImputation = "warn",
  sideEffects = "warn",
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{errorCheck}}{Ensures that the model can make predictions on the provided test dataset.}

\item{\code{longRunningService}}{Ensures that the custom model image can build and launch. If it cannot, the test is marked as Failed and subsequent test are aborted.}

\item{\code{nullValueImputation}}{Verifies that the model can impute null values. Required for Feature Impact.}

\item{\code{sideEffects}}{Verifies that predictions made on the dataset match row-wise predictions for the same dataset. Fails if the predictions do not match.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelTestsConfig$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelTestsConfig$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelTestsConfig$fromJSON(
  CustomModelTestsConfigJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CustomModelTestsConfigJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelTestsConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
