% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_model_tests_response.R
\docType{class}
\name{CustomModelTestsResponse}
\alias{CustomModelTestsResponse}
\title{CustomModelTestsResponse}
\format{
An \code{R6Class} generator object
}
\description{
CustomModelTestsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{completedAt}}{character ISO-8601 timestamp of when the testing attempt was completed.}

\item{\code{created}}{character ISO-8601 timestamp of when the testing entry was created.}

\item{\code{createdBy}}{character The username of the user that started the custom model test.}

\item{\code{customModel}}{\link{CustomModelShortResponse}}

\item{\code{customModelImageId}}{character If testing was successful, ID of the custom inference model image that can be used for a deployment, otherwise null.}

\item{\code{customModelVersion}}{\link{CustomModelVersionShortResponse}}

\item{\code{datasetId}}{character ID of the dataset used for testing.}

\item{\code{datasetVersionId}}{character ID of the specific dataset version used for testing.}

\item{\code{desiredMemory}}{integer \link{optional} The amount of memory that is expected to be allocated by the custom model.}

\item{\code{executionEnvironment}}{\link{ExecutionEnvironmentShortResponse}}

\item{\code{executionEnvironmentVersion}}{\link{ExecutionEnvironmentVersionShortResponse}}

\item{\code{id}}{character ID of the testing history entry.}

\item{\code{imageType}}{character \link{optional} The type of the image, either customModelImage if the testing attempt is using a customModelImage as its model or customModelVersion if the testing attempt is using a customModelVersion with dependency management.}

\item{\code{maximumMemory}}{integer \link{optional} The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{networkEgressPolicy}}{character \link{optional} Network egress policy.}

\item{\code{overallStatus}}{character The overall status of the testing history entry.}

\item{\code{replicas}}{integer \link{optional} A fixed number of replicas that will be set for the given custom-model.}

\item{\code{requiresHa}}{character \link{optional} Require all custom model replicas to be deployed on different Kubernetes nodes for predictions fault tolerance.}

\item{\code{testingStatus}}{\link{CustomModelTesting}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CustomModelTestsResponse$new()}}
\item \href{#method-validate}{\code{CustomModelTestsResponse$validate()}}
\item \href{#method-toJSON}{\code{CustomModelTestsResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{CustomModelTestsResponse$fromJSON()}}
\item \href{#method-clone}{\code{CustomModelTestsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelTestsResponse$new(
  completedAt = NULL,
  created = NULL,
  createdBy = NULL,
  customModel = NULL,
  customModelImageId = NULL,
  customModelVersion = NULL,
  datasetId = NULL,
  datasetVersionId = NULL,
  executionEnvironment = NULL,
  executionEnvironmentVersion = NULL,
  id = NULL,
  overallStatus = NULL,
  testingStatus = NULL,
  desiredMemory = NULL,
  imageType = NULL,
  maximumMemory = NULL,
  networkEgressPolicy = NULL,
  replicas = NULL,
  requiresHa = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{completedAt}}{ISO-8601 timestamp of when the testing attempt was completed.}

\item{\code{created}}{ISO-8601 timestamp of when the testing entry was created.}

\item{\code{createdBy}}{The username of the user that started the custom model test.}

\item{\code{customModelImageId}}{If testing was successful, ID of the custom inference model image that can be used for a deployment, otherwise null.}

\item{\code{datasetId}}{ID of the dataset used for testing.}

\item{\code{datasetVersionId}}{ID of the specific dataset version used for testing.}

\item{\code{id}}{ID of the testing history entry.}

\item{\code{overallStatus}}{The overall status of the testing history entry.}

\item{\code{desiredMemory}}{The amount of memory that is expected to be allocated by the custom model.}

\item{\code{imageType}}{The type of the image, either customModelImage if the testing attempt is using a customModelImage as its model or customModelVersion if the testing attempt is using a customModelVersion with dependency management.}

\item{\code{maximumMemory}}{The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{networkEgressPolicy}}{Network egress policy.}

\item{\code{replicas}}{A fixed number of replicas that will be set for the given custom-model.}

\item{\code{requiresHa}}{Require all custom model replicas to be deployed on different Kubernetes nodes for predictions fault tolerance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelTestsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelTestsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelTestsResponse$fromJSON(
  CustomModelTestsResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CustomModelTestsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelTestsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
