% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_model_version_response.R
\docType{class}
\name{CustomModelVersionResponse}
\alias{CustomModelVersionResponse}
\title{CustomModelVersionResponse}
\format{
An \code{R6Class} generator object
}
\description{
CustomModelVersionResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{baseEnvironmentId}}{character \link{optional} The base environment to use with this model version.}

\item{\code{baseEnvironmentVersionId}}{character \link{optional} The base environment version to use with this model version.}

\item{\code{created}}{character ISO-8601 timestamp of when the model was created.}

\item{\code{customModelId}}{character the ID of the custom model.}

\item{\code{dependencies}}{list( \link{DependencyResponse} ) \link{optional} The parsed dependencies of the custom model version if the version has a valid requirements.txt file.}

\item{\code{description}}{character Description of a custom model version.}

\item{\code{desiredMemory}}{integer \link{optional} The amount of memory that is expected to be allocated by the custom model.}

\item{\code{id}}{character the ID of the custom model version created.}

\item{\code{isFrozen}}{character If the version is frozen and immutable (i.e. it is either deployed or has been edited, causing a newer version to be yielded).}

\item{\code{items}}{list( \link{DeprecatedCustomModelWorkspaceItemResponse} ) List of file items.}

\item{\code{label}}{character A semantic version number of the major and minor version.}

\item{\code{maximumMemory}}{integer \link{optional} The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{networkEgressPolicy}}{character \link{optional} Network egress policy.}

\item{\code{replicas}}{integer \link{optional} A fixed number of replicas that will be set for the given custom-model.}

\item{\code{requiredMetadata}}{object \link{optional} Additional parameters required by the execution environment. The required keys are defined by the fieldNames in the base environment's requiredMetadataKeys. Once set, they cannot be changed. If you to change them, make a new version.}

\item{\code{requiredMetadataValues}}{list( \link{RequiredMetadataValue} ) \link{optional} Additional parameters required by the execution environment. The required fieldNames are defined by the fieldNames in the base environment's requiredMetadataKeys.}

\item{\code{requiresHa}}{character \link{optional} Require all custom model replicas to be deployed on different Kubernetes nodes for predictions fault tolerance.}

\item{\code{versionMajor}}{integer The major version number, incremented on deployments or larger file changes.}

\item{\code{versionMinor}}{integer The minor version number, incremented on general file changes.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CustomModelVersionResponse-new}{\code{CustomModelVersionResponse$new()}}
\item \href{#method-CustomModelVersionResponse-validate}{\code{CustomModelVersionResponse$validate()}}
\item \href{#method-CustomModelVersionResponse-toJSON}{\code{CustomModelVersionResponse$toJSON()}}
\item \href{#method-CustomModelVersionResponse-fromJSON}{\code{CustomModelVersionResponse$fromJSON()}}
\item \href{#method-CustomModelVersionResponse-clone}{\code{CustomModelVersionResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomModelVersionResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-CustomModelVersionResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelVersionResponse$new(
  created = NULL,
  customModelId = NULL,
  description = NULL,
  id = NULL,
  isFrozen = NULL,
  items = NULL,
  label = NULL,
  versionMajor = NULL,
  versionMinor = NULL,
  baseEnvironmentId = NULL,
  baseEnvironmentVersionId = NULL,
  dependencies = NULL,
  desiredMemory = NULL,
  maximumMemory = NULL,
  networkEgressPolicy = NULL,
  replicas = NULL,
  requiredMetadata = NULL,
  requiredMetadataValues = NULL,
  requiresHa = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{created}}{ISO-8601 timestamp of when the model was created.}

\item{\code{customModelId}}{the ID of the custom model.}

\item{\code{description}}{Description of a custom model version.}

\item{\code{id}}{the ID of the custom model version created.}

\item{\code{isFrozen}}{If the version is frozen and immutable (i.e. it is either deployed or has been edited, causing a newer version to be yielded).}

\item{\code{items}}{List of file items.}

\item{\code{label}}{A semantic version number of the major and minor version.}

\item{\code{versionMajor}}{The major version number, incremented on deployments or larger file changes.}

\item{\code{versionMinor}}{The minor version number, incremented on general file changes.}

\item{\code{baseEnvironmentId}}{The base environment to use with this model version.}

\item{\code{baseEnvironmentVersionId}}{The base environment version to use with this model version.}

\item{\code{dependencies}}{The parsed dependencies of the custom model version if the version has a valid requirements.txt file.}

\item{\code{desiredMemory}}{The amount of memory that is expected to be allocated by the custom model.}

\item{\code{maximumMemory}}{The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{networkEgressPolicy}}{Network egress policy.}

\item{\code{replicas}}{A fixed number of replicas that will be set for the given custom-model.}

\item{\code{requiredMetadata}}{Additional parameters required by the execution environment. The required keys are defined by the fieldNames in the base environment's requiredMetadataKeys. Once set, they cannot be changed. If you to change them, make a new version.}

\item{\code{requiredMetadataValues}}{Additional parameters required by the execution environment. The required fieldNames are defined by the fieldNames in the base environment's requiredMetadataKeys.}

\item{\code{requiresHa}}{Require all custom model replicas to be deployed on different Kubernetes nodes for predictions fault tolerance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomModelVersionResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CustomModelVersionResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelVersionResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomModelVersionResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomModelVersionResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelVersionResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomModelVersionResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomModelVersionResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelVersionResponse$fromJSON(
  CustomModelVersionResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CustomModelVersionResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomModelVersionResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CustomModelVersionResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelVersionResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
