% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_task_access_control_response.R
\docType{class}
\name{CustomTaskAccessControlResponse}
\alias{CustomTaskAccessControlResponse}
\title{CustomTaskAccessControlResponse}
\format{
An \code{R6Class} generator object
}
\description{
CustomTaskAccessControlResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{canShare}}{character Whether this user can share this custom task}

\item{\code{role}}{character This users role.}

\item{\code{userId}}{character This user's userId.}

\item{\code{username}}{character The username for this user's entry.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CustomTaskAccessControlResponse-new}{\code{CustomTaskAccessControlResponse$new()}}
\item \href{#method-CustomTaskAccessControlResponse-validate}{\code{CustomTaskAccessControlResponse$validate()}}
\item \href{#method-CustomTaskAccessControlResponse-toJSON}{\code{CustomTaskAccessControlResponse$toJSON()}}
\item \href{#method-CustomTaskAccessControlResponse-fromJSON}{\code{CustomTaskAccessControlResponse$fromJSON()}}
\item \href{#method-CustomTaskAccessControlResponse-clone}{\code{CustomTaskAccessControlResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskAccessControlResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskAccessControlResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskAccessControlResponse$new(
  canShare = NULL,
  role = NULL,
  userId = NULL,
  username = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{canShare}}{Whether this user can share this custom task}

\item{\code{role}}{This users role.}

\item{\code{userId}}{This user's userId.}

\item{\code{username}}{The username for this user's entry.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskAccessControlResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskAccessControlResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskAccessControlResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskAccessControlResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskAccessControlResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskAccessControlResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskAccessControlResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskAccessControlResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskAccessControlResponse$fromJSON(
  CustomTaskAccessControlResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CustomTaskAccessControlResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskAccessControlResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskAccessControlResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskAccessControlResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
