% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_task_create.R
\docType{class}
\name{CustomTaskCreate}
\alias{CustomTaskCreate}
\title{CustomTaskCreate}
\format{
An \code{R6Class} generator object
}
\description{
CustomTaskCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{calibratePredictions}}{character \link{optional} Whether model predictions should be calibrated by DataRobot.Only applies to anomaly detection; we recommend this if you have not already included calibration in your model code.Calibration improves the probability estimates of a model, and modifies the predictions of non-probabilistic models to be interpretable as probabilities. This will facilitate comparison to DataRobot models, and give access to ROC curve insights on external data.}

\item{\code{description}}{character \link{optional} The user-friendly description of the task.}

\item{\code{language}}{character \link{optional} Programming language name in which task is written.}

\item{\code{maximumMemory}}{integer \link{optional} The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{name}}{character The user-friendly name for the task.}

\item{\code{targetType}}{character The target type of the custom task}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CustomTaskCreate$new()}}
\item \href{#method-validate}{\code{CustomTaskCreate$validate()}}
\item \href{#method-toJSON}{\code{CustomTaskCreate$toJSON()}}
\item \href{#method-fromJSON}{\code{CustomTaskCreate$fromJSON()}}
\item \href{#method-clone}{\code{CustomTaskCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskCreate$new(
  name = NULL,
  targetType = NULL,
  calibratePredictions = TRUE,
  description = NULL,
  language = NULL,
  maximumMemory = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The user-friendly name for the task.}

\item{\code{targetType}}{The target type of the custom task}

\item{\code{calibratePredictions}}{Whether model predictions should be calibrated by DataRobot.Only applies to anomaly detection; we recommend this if you have not already included calibration in your model code.Calibration improves the probability estimates of a model, and modifies the predictions of non-probabilistic models to be interpretable as probabilities. This will facilitate comparison to DataRobot models, and give access to ROC curve insights on external data.}

\item{\code{description}}{The user-friendly description of the task.}

\item{\code{language}}{Programming language name in which task is written.}

\item{\code{maximumMemory}}{The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskCreate$fromJSON(CustomTaskCreateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CustomTaskCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
