% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_task_response.R
\docType{class}
\name{CustomTaskResponse}
\alias{CustomTaskResponse}
\title{CustomTaskResponse}
\format{
An \code{R6Class} generator object
}
\description{
CustomTaskResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{calibratePredictions}}{character \link{optional} Determines whether or not predictions should be calibrated by DataRobot. Only applies to anomaly detection.}

\item{\code{created}}{character ISO-8601 timestamp of when the task was created.}

\item{\code{createdBy}}{character The username of the custom task creator.}

\item{\code{customModelType}}{character The type of custom task.}

\item{\code{description}}{character The description of the task.}

\item{\code{id}}{character The ID of the custom task.}

\item{\code{language}}{character The programming language used by the task.}

\item{\code{latestVersion}}{\link{CustomTaskVersionResponse}}

\item{\code{maximumMemory}}{integer \link{optional} The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{name}}{character The name of the task.}

\item{\code{targetType}}{character The target type of the custom task.}

\item{\code{updated}}{character ISO-8601 timestamp of when task was last updated.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CustomTaskResponse-new}{\code{CustomTaskResponse$new()}}
\item \href{#method-CustomTaskResponse-validate}{\code{CustomTaskResponse$validate()}}
\item \href{#method-CustomTaskResponse-toJSON}{\code{CustomTaskResponse$toJSON()}}
\item \href{#method-CustomTaskResponse-fromJSON}{\code{CustomTaskResponse$fromJSON()}}
\item \href{#method-CustomTaskResponse-clone}{\code{CustomTaskResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskResponse$new(
  created = NULL,
  createdBy = NULL,
  customModelType = NULL,
  description = NULL,
  id = NULL,
  language = NULL,
  latestVersion = NULL,
  name = NULL,
  targetType = NULL,
  updated = NULL,
  calibratePredictions = NULL,
  maximumMemory = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{created}}{ISO-8601 timestamp of when the task was created.}

\item{\code{createdBy}}{The username of the custom task creator.}

\item{\code{customModelType}}{The type of custom task.}

\item{\code{description}}{The description of the task.}

\item{\code{id}}{The ID of the custom task.}

\item{\code{language}}{The programming language used by the task.}

\item{\code{name}}{The name of the task.}

\item{\code{targetType}}{The target type of the custom task.}

\item{\code{updated}}{ISO-8601 timestamp of when task was last updated.}

\item{\code{calibratePredictions}}{Determines whether or not predictions should be calibrated by DataRobot. Only applies to anomaly detection.}

\item{\code{maximumMemory}}{The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskResponse$fromJSON(CustomTaskResponseJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CustomTaskResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
