% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_task_update.R
\docType{class}
\name{CustomTaskUpdate}
\alias{CustomTaskUpdate}
\title{CustomTaskUpdate}
\format{
An \code{R6Class} generator object
}
\description{
CustomTaskUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{description}}{character \link{optional} The user-friendly description of the task.}

\item{\code{language}}{character \link{optional} Programming language name in which task is written.}

\item{\code{maximumMemory}}{integer \link{optional} The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{name}}{character \link{optional} The user-friendly name for the task.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CustomTaskUpdate-new}{\code{CustomTaskUpdate$new()}}
\item \href{#method-CustomTaskUpdate-validate}{\code{CustomTaskUpdate$validate()}}
\item \href{#method-CustomTaskUpdate-toJSON}{\code{CustomTaskUpdate$toJSON()}}
\item \href{#method-CustomTaskUpdate-fromJSON}{\code{CustomTaskUpdate$fromJSON()}}
\item \href{#method-CustomTaskUpdate-clone}{\code{CustomTaskUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskUpdate-new"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskUpdate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskUpdate$new(
  description = NULL,
  language = NULL,
  maximumMemory = NULL,
  name = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{The user-friendly description of the task.}

\item{\code{language}}{Programming language name in which task is written.}

\item{\code{maximumMemory}}{The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{name}}{The user-friendly name for the task.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskUpdate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskUpdate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskUpdate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskUpdate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskUpdate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskUpdate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskUpdate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskUpdate$fromJSON(CustomTaskUpdateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CustomTaskUpdateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskUpdate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskUpdate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
