% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_task_version_create_from_repository.R
\docType{class}
\name{CustomTaskVersionCreateFromRepository}
\alias{CustomTaskVersionCreateFromRepository}
\title{CustomTaskVersionCreateFromRepository}
\format{
An \code{R6Class} generator object
}
\description{
CustomTaskVersionCreateFromRepository Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{baseEnvironmentId}}{character \link{optional} The base environment to use with this version.}

\item{\code{isMajorUpdate}}{character \link{optional} If set to true, new major version will created, otherwise minor version will be created.}

\item{\code{ref}}{character \link{optional} Remote reference (branch, commit, etc). Latest, if not specified.}

\item{\code{repositoryId}}{character The ID of remote repository used to pull sources. This ID can be found using the /api/v2/remoteRepositories/ endpoint.}

\item{\code{requiredMetadata}}{object \link{optional} Additional parameters required by the execution environment. The required keys are defined by the fieldNames in the base environment's requiredMetadataKeys. Once set, they cannot be changed. If you to change them, make a new version.}

\item{\code{requiredMetadataValues}}{list( \link{RequiredMetadataValue} ) \link{optional} Additional parameters required by the execution environment. The required fieldNames are defined by the fieldNames in the base environment's requiredMetadataKeys.}

\item{\code{sourcePath}}{\link{OneOfstringarray} \link{optional} A remote repository file path to be pulled into a custom model or custom task.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CustomTaskVersionCreateFromRepository-new}{\code{CustomTaskVersionCreateFromRepository$new()}}
\item \href{#method-CustomTaskVersionCreateFromRepository-validate}{\code{CustomTaskVersionCreateFromRepository$validate()}}
\item \href{#method-CustomTaskVersionCreateFromRepository-toJSON}{\code{CustomTaskVersionCreateFromRepository$toJSON()}}
\item \href{#method-CustomTaskVersionCreateFromRepository-fromJSON}{\code{CustomTaskVersionCreateFromRepository$fromJSON()}}
\item \href{#method-CustomTaskVersionCreateFromRepository-clone}{\code{CustomTaskVersionCreateFromRepository$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskVersionCreateFromRepository-new"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskVersionCreateFromRepository-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskVersionCreateFromRepository$new(
  repositoryId = NULL,
  baseEnvironmentId = NULL,
  isMajorUpdate = TRUE,
  ref = NULL,
  requiredMetadata = NULL,
  requiredMetadataValues = NULL,
  sourcePath = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{repositoryId}}{The ID of remote repository used to pull sources. This ID can be found using the /api/v2/remoteRepositories/ endpoint.}

\item{\code{baseEnvironmentId}}{The base environment to use with this version.}

\item{\code{isMajorUpdate}}{If set to true, new major version will created, otherwise minor version will be created.}

\item{\code{ref}}{Remote reference (branch, commit, etc). Latest, if not specified.}

\item{\code{requiredMetadata}}{Additional parameters required by the execution environment. The required keys are defined by the fieldNames in the base environment's requiredMetadataKeys. Once set, they cannot be changed. If you to change them, make a new version.}

\item{\code{requiredMetadataValues}}{Additional parameters required by the execution environment. The required fieldNames are defined by the fieldNames in the base environment's requiredMetadataKeys.}

\item{\code{sourcePath}}{A remote repository file path to be pulled into a custom model or custom task.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskVersionCreateFromRepository-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskVersionCreateFromRepository-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskVersionCreateFromRepository$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskVersionCreateFromRepository-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskVersionCreateFromRepository-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskVersionCreateFromRepository$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskVersionCreateFromRepository-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskVersionCreateFromRepository-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskVersionCreateFromRepository$fromJSON(
  CustomTaskVersionCreateFromRepositoryJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CustomTaskVersionCreateFromRepositoryJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskVersionCreateFromRepository-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskVersionCreateFromRepository-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskVersionCreateFromRepository$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
