% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_task_version_metadata_update.R
\docType{class}
\name{CustomTaskVersionMetadataUpdate}
\alias{CustomTaskVersionMetadataUpdate}
\title{CustomTaskVersionMetadataUpdate}
\format{
An \code{R6Class} generator object
}
\description{
CustomTaskVersionMetadataUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{description}}{character \link{optional} New description for the custom task or model.}

\item{\code{requiredMetadata}}{object \link{optional} Additional parameters required by the execution environment. The required keys are defined by the fieldNames in the base environment's requiredMetadataKeys.}

\item{\code{requiredMetadataValues}}{list( \link{RequiredMetadataValue} ) \link{optional} Additional parameters required by the execution environment. The required fieldNames are defined by the fieldNames in the base environment's requiredMetadataKeys.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CustomTaskVersionMetadataUpdate-new}{\code{CustomTaskVersionMetadataUpdate$new()}}
\item \href{#method-CustomTaskVersionMetadataUpdate-validate}{\code{CustomTaskVersionMetadataUpdate$validate()}}
\item \href{#method-CustomTaskVersionMetadataUpdate-toJSON}{\code{CustomTaskVersionMetadataUpdate$toJSON()}}
\item \href{#method-CustomTaskVersionMetadataUpdate-fromJSON}{\code{CustomTaskVersionMetadataUpdate$fromJSON()}}
\item \href{#method-CustomTaskVersionMetadataUpdate-clone}{\code{CustomTaskVersionMetadataUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskVersionMetadataUpdate-new"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskVersionMetadataUpdate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskVersionMetadataUpdate$new(
  description = NULL,
  requiredMetadata = NULL,
  requiredMetadataValues = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{New description for the custom task or model.}

\item{\code{requiredMetadata}}{Additional parameters required by the execution environment. The required keys are defined by the fieldNames in the base environment's requiredMetadataKeys.}

\item{\code{requiredMetadataValues}}{Additional parameters required by the execution environment. The required fieldNames are defined by the fieldNames in the base environment's requiredMetadataKeys.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskVersionMetadataUpdate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskVersionMetadataUpdate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskVersionMetadataUpdate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskVersionMetadataUpdate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskVersionMetadataUpdate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskVersionMetadataUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskVersionMetadataUpdate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskVersionMetadataUpdate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskVersionMetadataUpdate$fromJSON(
  CustomTaskVersionMetadataUpdateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CustomTaskVersionMetadataUpdateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTaskVersionMetadataUpdate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTaskVersionMetadataUpdate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskVersionMetadataUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
