% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_task_version_response.R
\docType{class}
\name{CustomTaskVersionResponse}
\alias{CustomTaskVersionResponse}
\title{CustomTaskVersionResponse}
\format{
An \code{R6Class} generator object
}
\description{
CustomTaskVersionResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{baseEnvironmentId}}{character \link{optional} The base environment to use with this task version.}

\item{\code{baseEnvironmentVersionId}}{character \link{optional} The base environment version to use with this task version.}

\item{\code{created}}{character ISO-8601 timestamp of when the task was created.}

\item{\code{customModelId}}{character an alias for customTaskId}

\item{\code{customTaskId}}{character the ID of the custom task.}

\item{\code{dependencies}}{list( \link{DependencyResponse} ) \link{optional} The parsed dependencies of the custom task version if the version has a valid requirements.txt file.}

\item{\code{description}}{character Description of a custom task version.}

\item{\code{id}}{character the ID of the custom model version created.}

\item{\code{isFrozen}}{character If the version is frozen and immutable (i.e. it is either deployed or has been edited, causing a newer version to be yielded).}

\item{\code{items}}{list( \link{WorkspaceItemResponse} ) List of file items.}

\item{\code{label}}{character A semantic version number of the major and minor version.}

\item{\code{maximumMemory}}{integer \link{optional} The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{requiredMetadata}}{object \link{optional} Additional parameters required by the execution environment. The required keys are defined by the fieldNames in the base environment's requiredMetadataKeys. Once set, they cannot be changed. If you want to change them, make a new version.}

\item{\code{requiredMetadataValues}}{list( \link{RequiredMetadataValue} ) \link{optional} Additional parameters required by the execution environment. The required fieldNames are defined by the fieldNames in the base environment's requiredMetadataKeys.}

\item{\code{versionMajor}}{integer The major version number, incremented on deployments or larger file changes.}

\item{\code{versionMinor}}{integer The minor version number, incremented on general file changes.}

\item{\code{warning}}{list( character ) \link{optional} Warnings about the custom task version}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CustomTaskVersionResponse$new()}}
\item \href{#method-validate}{\code{CustomTaskVersionResponse$validate()}}
\item \href{#method-toJSON}{\code{CustomTaskVersionResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{CustomTaskVersionResponse$fromJSON()}}
\item \href{#method-clone}{\code{CustomTaskVersionResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskVersionResponse$new(
  created = NULL,
  customModelId = NULL,
  customTaskId = NULL,
  description = NULL,
  id = NULL,
  isFrozen = NULL,
  items = NULL,
  label = NULL,
  versionMajor = NULL,
  versionMinor = NULL,
  baseEnvironmentId = NULL,
  baseEnvironmentVersionId = NULL,
  dependencies = NULL,
  maximumMemory = NULL,
  requiredMetadata = NULL,
  requiredMetadataValues = NULL,
  warning = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{created}}{ISO-8601 timestamp of when the task was created.}

\item{\code{customModelId}}{an alias for customTaskId}

\item{\code{customTaskId}}{the ID of the custom task.}

\item{\code{description}}{Description of a custom task version.}

\item{\code{id}}{the ID of the custom model version created.}

\item{\code{isFrozen}}{If the version is frozen and immutable (i.e. it is either deployed or has been edited, causing a newer version to be yielded).}

\item{\code{items}}{List of file items.}

\item{\code{label}}{A semantic version number of the major and minor version.}

\item{\code{versionMajor}}{The major version number, incremented on deployments or larger file changes.}

\item{\code{versionMinor}}{The minor version number, incremented on general file changes.}

\item{\code{baseEnvironmentId}}{The base environment to use with this task version.}

\item{\code{baseEnvironmentVersionId}}{The base environment version to use with this task version.}

\item{\code{dependencies}}{The parsed dependencies of the custom task version if the version has a valid requirements.txt file.}

\item{\code{maximumMemory}}{The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{requiredMetadata}}{Additional parameters required by the execution environment. The required keys are defined by the fieldNames in the base environment's requiredMetadataKeys. Once set, they cannot be changed. If you want to change them, make a new version.}

\item{\code{requiredMetadataValues}}{Additional parameters required by the execution environment. The required fieldNames are defined by the fieldNames in the base environment's requiredMetadataKeys.}

\item{\code{warning}}{Warnings about the custom task version}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskVersionResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskVersionResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskVersionResponse$fromJSON(
  CustomTaskVersionResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CustomTaskVersionResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTaskVersionResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
