% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_training_blueprint_response.R
\docType{class}
\name{CustomTrainingBlueprintResponse}
\alias{CustomTrainingBlueprintResponse}
\title{CustomTrainingBlueprintResponse}
\format{
An \code{R6Class} generator object
}
\description{
CustomTrainingBlueprintResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{createdAt}}{character ISO-8601 timestamp of when the blueprint was created.}

\item{\code{customModel}}{\link{CustomModelShortResponse}}

\item{\code{customModelVersion}}{\link{CustomModelVersionShortResponse}}

\item{\code{executionEnvironment}}{\link{ExecutionEnvironmentShortResponse}}

\item{\code{executionEnvironmentVersion}}{\link{ExecutionEnvironmentVersionShortResponse}}

\item{\code{targetType}}{character The target type of the training model.}

\item{\code{trainingHistory}}{list( \link{TrainingHistoryEntry} ) List of instances of this blueprint having been trained.}

\item{\code{userBlueprintId}}{character User Blueprint ID that can be used to train the model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CustomTrainingBlueprintResponse-new}{\code{CustomTrainingBlueprintResponse$new()}}
\item \href{#method-CustomTrainingBlueprintResponse-validate}{\code{CustomTrainingBlueprintResponse$validate()}}
\item \href{#method-CustomTrainingBlueprintResponse-toJSON}{\code{CustomTrainingBlueprintResponse$toJSON()}}
\item \href{#method-CustomTrainingBlueprintResponse-fromJSON}{\code{CustomTrainingBlueprintResponse$fromJSON()}}
\item \href{#method-CustomTrainingBlueprintResponse-clone}{\code{CustomTrainingBlueprintResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTrainingBlueprintResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTrainingBlueprintResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTrainingBlueprintResponse$new(
  createdAt = NULL,
  customModel = NULL,
  customModelVersion = NULL,
  executionEnvironment = NULL,
  executionEnvironmentVersion = NULL,
  targetType = NULL,
  trainingHistory = NULL,
  userBlueprintId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{createdAt}}{ISO-8601 timestamp of when the blueprint was created.}

\item{\code{targetType}}{The target type of the training model.}

\item{\code{trainingHistory}}{List of instances of this blueprint having been trained.}

\item{\code{userBlueprintId}}{User Blueprint ID that can be used to train the model.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTrainingBlueprintResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTrainingBlueprintResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTrainingBlueprintResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTrainingBlueprintResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTrainingBlueprintResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTrainingBlueprintResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTrainingBlueprintResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTrainingBlueprintResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTrainingBlueprintResponse$fromJSON(
  CustomTrainingBlueprintResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CustomTrainingBlueprintResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomTrainingBlueprintResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CustomTrainingBlueprintResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomTrainingBlueprintResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
