% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_disparity_create_payload.R
\docType{class}
\name{DataDisparityCreatePayload}
\alias{DataDisparityCreatePayload}
\title{DataDisparityCreatePayload}
\format{
An \code{R6Class} generator object
}
\description{
DataDisparityCreatePayload Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{comparedClassNames}}{list( character ) An array of classes to calculate data disparity for.}

\item{\code{feature}}{character Feature for which insight is computed.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataDisparityCreatePayload-new}{\code{DataDisparityCreatePayload$new()}}
\item \href{#method-DataDisparityCreatePayload-validate}{\code{DataDisparityCreatePayload$validate()}}
\item \href{#method-DataDisparityCreatePayload-toJSON}{\code{DataDisparityCreatePayload$toJSON()}}
\item \href{#method-DataDisparityCreatePayload-fromJSON}{\code{DataDisparityCreatePayload$fromJSON()}}
\item \href{#method-DataDisparityCreatePayload-clone}{\code{DataDisparityCreatePayload$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataDisparityCreatePayload-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataDisparityCreatePayload-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataDisparityCreatePayload$new(
  comparedClassNames = NULL,
  feature = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comparedClassNames}}{An array of classes to calculate data disparity for.}

\item{\code{feature}}{Feature for which insight is computed.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataDisparityCreatePayload-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DataDisparityCreatePayload-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataDisparityCreatePayload$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataDisparityCreatePayload-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataDisparityCreatePayload-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataDisparityCreatePayload$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataDisparityCreatePayload-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataDisparityCreatePayload-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataDisparityCreatePayload$fromJSON(
  DataDisparityCreatePayloadJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DataDisparityCreatePayloadJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataDisparityCreatePayload-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataDisparityCreatePayload-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataDisparityCreatePayload$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
