% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_disparity_insights.R
\docType{class}
\name{DataDisparityInsights}
\alias{DataDisparityInsights}
\title{DataDisparityInsights}
\format{
An \code{R6Class} generator object
}
\description{
DataDisparityInsights Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{features}}{list( \link{AnalyzedFeature} ) \link{optional} A mapping of the feature name to the corresponding values on the graph.}

\item{\code{metric}}{character \link{optional} Metric used to calculate the impact of a feature on data disparity.}

\item{\code{protectedFeature}}{character \link{optional} Feature for which insights were computed.}

\item{\code{values}}{list( \link{FeatureCounts} ) \link{optional} Class count details for each class being compared.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataDisparityInsights$new()}}
\item \href{#method-validate}{\code{DataDisparityInsights$validate()}}
\item \href{#method-toJSON}{\code{DataDisparityInsights$toJSON()}}
\item \href{#method-fromJSON}{\code{DataDisparityInsights$fromJSON()}}
\item \href{#method-clone}{\code{DataDisparityInsights$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataDisparityInsights$new(
  features = NULL,
  metric = NULL,
  protectedFeature = NULL,
  values = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{features}}{A mapping of the feature name to the corresponding values on the graph.}

\item{\code{metric}}{Metric used to calculate the impact of a feature on data disparity.}

\item{\code{protectedFeature}}{Feature for which insights were computed.}

\item{\code{values}}{Class count details for each class being compared.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataDisparityInsights$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataDisparityInsights$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataDisparityInsights$fromJSON(
  DataDisparityInsightsJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DataDisparityInsightsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataDisparityInsights$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
