% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_warnings_record.R
\docType{class}
\name{DataQualityWarningsRecord}
\alias{DataQualityWarningsRecord}
\title{DataQualityWarningsRecord}
\format{
An \code{R6Class} generator object
}
\description{
DataQualityWarningsRecord Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{hasKiaMissingValuesInForecastWindow}}{character \link{optional} If true, known-in-advance features in this dataset have missing values in the forecast window. Absence of the known-in-advance values can negatively impact prediction quality. Only applies for time series projects.}

\item{\code{insufficientRowsForEvaluatingModels}}{character \link{optional} If true, the dataset has a target column present indicating it can be used to evaluate model performance but too few rows to be trustworthy in so doing. If false, either it has no target column at all or it has sufficient rows for model evaluation. Only applies for regression, binary classification, multiclass classification projects and time series unsupervised projects.}

\item{\code{singleClassActualValueColumn}}{character \link{optional} If true, actual value column has only one class and such insights as ROC curve can not be calculated. Only applies for binary classification projects or unsupervised projects.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataQualityWarningsRecord-new}{\code{DataQualityWarningsRecord$new()}}
\item \href{#method-DataQualityWarningsRecord-validate}{\code{DataQualityWarningsRecord$validate()}}
\item \href{#method-DataQualityWarningsRecord-toJSON}{\code{DataQualityWarningsRecord$toJSON()}}
\item \href{#method-DataQualityWarningsRecord-fromJSON}{\code{DataQualityWarningsRecord$fromJSON()}}
\item \href{#method-DataQualityWarningsRecord-clone}{\code{DataQualityWarningsRecord$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataQualityWarningsRecord-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataQualityWarningsRecord-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataQualityWarningsRecord$new(
  hasKiaMissingValuesInForecastWindow = NULL,
  insufficientRowsForEvaluatingModels = NULL,
  singleClassActualValueColumn = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hasKiaMissingValuesInForecastWindow}}{If true, known-in-advance features in this dataset have missing values in the forecast window. Absence of the known-in-advance values can negatively impact prediction quality. Only applies for time series projects.}

\item{\code{insufficientRowsForEvaluatingModels}}{If true, the dataset has a target column present indicating it can be used to evaluate model performance but too few rows to be trustworthy in so doing. If false, either it has no target column at all or it has sufficient rows for model evaluation. Only applies for regression, binary classification, multiclass classification projects and time series unsupervised projects.}

\item{\code{singleClassActualValueColumn}}{If true, actual value column has only one class and such insights as ROC curve can not be calculated. Only applies for binary classification projects or unsupervised projects.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataQualityWarningsRecord-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DataQualityWarningsRecord-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataQualityWarningsRecord$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataQualityWarningsRecord-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataQualityWarningsRecord-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataQualityWarningsRecord$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataQualityWarningsRecord-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataQualityWarningsRecord-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataQualityWarningsRecord$fromJSON(
  DataQualityWarningsRecordJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DataQualityWarningsRecordJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataQualityWarningsRecord-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataQualityWarningsRecord-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataQualityWarningsRecord$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
