% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datarobot.apicore-package.R
\docType{package}
\name{datarobot.apicore}
\alias{datarobot.apicore}
\alias{datarobot.apicore-package}
\title{datarobot.apicore: R Package Client for Public API}
\description{
DataRobot's Public facing API
}
\section{Package options}{


The \code{datarobot.apicore} package uses the following \code{\link[=options]{options()}} to configure behavior:

\itemize{
\item \code{datarobot.apicore.returnS3}: logical. If \code{TRUE}, API responses will be returned to calling functions as S3 lists instead of R6 objects. See the helper methods \code{\link[=.ReturnResponse]{.ReturnResponse()}} and \code{\link[=.toS3]{.toS3()}} for more information. Default is \code{TRUE}, which we recommend for consistency if you have been using the datarobot client prior to 2.27.0.

}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datarobot/rsdk}
  \item Report bugs at \url{https://github.com/datarobot/rsdk/issues}
}

}
\author{
\strong{Maintainer}: AJ Alon \email{api-maintainer@datarobot.com}

}
\keyword{internal}
