% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_access_control_response.R
\docType{class}
\name{DatasetAccessControlResponse}
\alias{DatasetAccessControlResponse}
\title{DatasetAccessControlResponse}
\format{
An \code{R6Class} generator object
}
\description{
DatasetAccessControlResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{canShare}}{character True if this user can share with other users}

\item{\code{canUseData}}{character True if the user can view, download and process data (use to create projects, predictions, etc)}

\item{\code{role}}{character The role of the user on this data source.}

\item{\code{userFullName}}{character The full name of a user with access to this dataset. If the full name is not available, username is returned instead.}

\item{\code{username}}{character `username` of a user with access to this dataset.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DatasetAccessControlResponse$new()}}
\item \href{#method-validate}{\code{DatasetAccessControlResponse$validate()}}
\item \href{#method-toJSON}{\code{DatasetAccessControlResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{DatasetAccessControlResponse$fromJSON()}}
\item \href{#method-clone}{\code{DatasetAccessControlResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetAccessControlResponse$new(
  canShare = NULL,
  canUseData = NULL,
  role = NULL,
  userFullName = NULL,
  username = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{canShare}}{True if this user can share with other users}

\item{\code{canUseData}}{True if the user can view, download and process data (use to create projects, predictions, etc)}

\item{\code{role}}{The role of the user on this data source.}

\item{\code{userFullName}}{The full name of a user with access to this dataset. If the full name is not available, username is returned instead.}

\item{\code{username}}{`username` of a user with access to this dataset.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetAccessControlResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetAccessControlResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetAccessControlResponse$fromJSON(
  DatasetAccessControlResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatasetAccessControlResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetAccessControlResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
