% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_access_inner.R
\docType{class}
\name{DatasetAccessInner}
\alias{DatasetAccessInner}
\title{DatasetAccessInner}
\format{
An \code{R6Class} generator object
}
\description{
DatasetAccessInner Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{canShare}}{character \link{optional} whether the user should be able to share with other users. If true, the user will be able to grant any role up to and including their own to other users. If `role` is empty `canShare` is ignored.}

\item{\code{canUseData}}{character \link{optional} Whether the user should be able to view, download and process data (use to create projects, predictions, etc). For OWNER `canUseData` is always True. If `role` is empty `canUseData` is ignored.}

\item{\code{role}}{character the role to grant to the user, or \"\" (empty string) to remove the users access}

\item{\code{username}}{character username of the user to update the access role for.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatasetAccessInner-new}{\code{DatasetAccessInner$new()}}
\item \href{#method-DatasetAccessInner-validate}{\code{DatasetAccessInner$validate()}}
\item \href{#method-DatasetAccessInner-toJSON}{\code{DatasetAccessInner$toJSON()}}
\item \href{#method-DatasetAccessInner-fromJSON}{\code{DatasetAccessInner$fromJSON()}}
\item \href{#method-DatasetAccessInner-clone}{\code{DatasetAccessInner$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetAccessInner-new"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetAccessInner-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetAccessInner$new(
  role = NULL,
  username = NULL,
  canShare = NULL,
  canUseData = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{the role to grant to the user, or \"\" (empty string) to remove the users access}

\item{\code{username}}{username of the user to update the access role for.}

\item{\code{canShare}}{whether the user should be able to share with other users. If true, the user will be able to grant any role up to and including their own to other users. If `role` is empty `canShare` is ignored.}

\item{\code{canUseData}}{Whether the user should be able to view, download and process data (use to create projects, predictions, etc). For OWNER `canUseData` is always True. If `role` is empty `canUseData` is ignored.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetAccessInner-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetAccessInner-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetAccessInner$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetAccessInner-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetAccessInner-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetAccessInner$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetAccessInner-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetAccessInner-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetAccessInner$fromJSON(DatasetAccessInnerJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatasetAccessInnerJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetAccessInner-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetAccessInner-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetAccessInner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
