% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_access_set.R
\docType{class}
\name{DatasetAccessSet}
\alias{DatasetAccessSet}
\title{DatasetAccessSet}
\format{
An \code{R6Class} generator object
}
\description{
DatasetAccessSet Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{applyGrantToLinkedObjects}}{character \link{optional} If true for any users being granted access to the dataset, grant the user read access to any linked objects such as DataSources and DataStores that may be used by this dataset. Ignored if no such objects are relevant for dataset. Will not result in access being lowered for a user if the user already has higher access to linked objects than read access. However, if the target user does not have sharing permissions to the linked object, they will be given sharing access without lowering existing permissions. May result in an error if user making call does not have sufficient permissions to complete grant. Default value is false.}

\item{\code{data}}{list( \link{DatasetAccessInner} ) array of DatasetAccessControl objects.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatasetAccessSet-new}{\code{DatasetAccessSet$new()}}
\item \href{#method-DatasetAccessSet-validate}{\code{DatasetAccessSet$validate()}}
\item \href{#method-DatasetAccessSet-toJSON}{\code{DatasetAccessSet$toJSON()}}
\item \href{#method-DatasetAccessSet-fromJSON}{\code{DatasetAccessSet$fromJSON()}}
\item \href{#method-DatasetAccessSet-clone}{\code{DatasetAccessSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetAccessSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetAccessSet-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetAccessSet$new(
  data = NULL,
  applyGrantToLinkedObjects = FALSE,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{array of DatasetAccessControl objects.}

\item{\code{applyGrantToLinkedObjects}}{If true for any users being granted access to the dataset, grant the user read access to any linked objects such as DataSources and DataStores that may be used by this dataset. Ignored if no such objects are relevant for dataset. Will not result in access being lowered for a user if the user already has higher access to linked objects than read access. However, if the target user does not have sharing permissions to the linked object, they will be given sharing access without lowering existing permissions. May result in an error if user making call does not have sufficient permissions to complete grant. Default value is false.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetAccessSet-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetAccessSet-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetAccessSet$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetAccessSet-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetAccessSet-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetAccessSet$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetAccessSet-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetAccessSet-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetAccessSet$fromJSON(DatasetAccessSetJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatasetAccessSetJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetAccessSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetAccessSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetAccessSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
