% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_create_from_workspace_state.R
\docType{class}
\name{DatasetCreateFromWorkspaceState}
\alias{DatasetCreateFromWorkspaceState}
\title{DatasetCreateFromWorkspaceState}
\format{
An \code{R6Class} generator object
}
\description{
DatasetCreateFromWorkspaceState Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{credentials}}{character \link{optional} JDBC credentials.}

\item{\code{datasetName}}{character \link{optional} Custom name for the created dataset.}

\item{\code{doSnapshot}}{character \link{optional} If true, create a snapshot dataset; if false, create a remote dataset. Creating snapshots from non-file sources requires an additional permission, `Enable Create Snapshot Data Source`.}

\item{\code{workspaceStateId}}{character ID of the workspace state to use as the source of data.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DatasetCreateFromWorkspaceState$new()}}
\item \href{#method-validate}{\code{DatasetCreateFromWorkspaceState$validate()}}
\item \href{#method-toJSON}{\code{DatasetCreateFromWorkspaceState$toJSON()}}
\item \href{#method-fromJSON}{\code{DatasetCreateFromWorkspaceState$fromJSON()}}
\item \href{#method-clone}{\code{DatasetCreateFromWorkspaceState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetCreateFromWorkspaceState$new(
  workspaceStateId = NULL,
  credentials = NULL,
  datasetName = NULL,
  doSnapshot = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{workspaceStateId}}{ID of the workspace state to use as the source of data.}

\item{\code{credentials}}{JDBC credentials.}

\item{\code{datasetName}}{Custom name for the created dataset.}

\item{\code{doSnapshot}}{If true, create a snapshot dataset; if false, create a remote dataset. Creating snapshots from non-file sources requires an additional permission, `Enable Create Snapshot Data Source`.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetCreateFromWorkspaceState$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetCreateFromWorkspaceState$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetCreateFromWorkspaceState$fromJSON(
  DatasetCreateFromWorkspaceStateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatasetCreateFromWorkspaceStateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetCreateFromWorkspaceState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
