% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_definition.R
\docType{class}
\name{DatasetDefinition}
\alias{DatasetDefinition}
\title{DatasetDefinition}
\format{
An \code{R6Class} generator object
}
\description{
DatasetDefinition Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalogId}}{character ID of the catalog item.}

\item{\code{catalogVersionId}}{character ID of the catalog item version.}

\item{\code{featureListId}}{character \link{optional} ID of the feature list. This decides which columns in the dataset are used for feature generation.}

\item{\code{identifier}}{character Short name of the dataset (used directly as part of the generated feature names).}

\item{\code{primaryTemporalKey}}{character \link{optional} Name of the column indicating time of record creation.}

\item{\code{snapshotPolicy}}{character \link{optional} Policy for using dataset snapshots when creating a project or making predictions. Must be one of the following values: 'specified': Use specific snapshot specified by catalogVersionId. 'latest': Use latest snapshot from the same catalog item. 'dynamic': Get data from the source (only applicable for JDBC datasets).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DatasetDefinition$new()}}
\item \href{#method-validate}{\code{DatasetDefinition$validate()}}
\item \href{#method-toJSON}{\code{DatasetDefinition$toJSON()}}
\item \href{#method-fromJSON}{\code{DatasetDefinition$fromJSON()}}
\item \href{#method-clone}{\code{DatasetDefinition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetDefinition$new(
  catalogId = NULL,
  catalogVersionId = NULL,
  identifier = NULL,
  featureListId = NULL,
  primaryTemporalKey = NULL,
  snapshotPolicy = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalogId}}{ID of the catalog item.}

\item{\code{catalogVersionId}}{ID of the catalog item version.}

\item{\code{identifier}}{Short name of the dataset (used directly as part of the generated feature names).}

\item{\code{featureListId}}{ID of the feature list. This decides which columns in the dataset are used for feature generation.}

\item{\code{primaryTemporalKey}}{Name of the column indicating time of record creation.}

\item{\code{snapshotPolicy}}{Policy for using dataset snapshots when creating a project or making predictions. Must be one of the following values: 'specified': Use specific snapshot specified by catalogVersionId. 'latest': Use latest snapshot from the same catalog item. 'dynamic': Get data from the source (only applicable for JDBC datasets).}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetDefinition$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetDefinition$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetDefinition$fromJSON(DatasetDefinitionJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatasetDefinitionJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetDefinition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
