% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_definition_response.R
\docType{class}
\name{DatasetDefinitionResponse}
\alias{DatasetDefinitionResponse}
\title{DatasetDefinitionResponse}
\format{
An \code{R6Class} generator object
}
\description{
DatasetDefinitionResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalogId}}{character ID of the catalog item.}

\item{\code{catalogVersionId}}{character ID of the catalog item version.}

\item{\code{dataSource}}{\link{DataSource} \link{optional}}

\item{\code{dataSources}}{list( \link{DataSource} ) \link{optional} Data source details for a JDBC dataset}

\item{\code{featureListId}}{character ID of the feature list. This decides which columns in the dataset are used for feature generation.}

\item{\code{featureLists}}{list( character ) \link{optional} List of available feature list ids for the dataset}

\item{\code{identifier}}{character Short name of the dataset (used directly as part of the generated feature names).}

\item{\code{isDeleted}}{character \link{optional} Is this dataset deleted?}

\item{\code{originalIdentifier}}{character \link{optional} Original identifier of the dataset if it was updated to resolve name conflicts.}

\item{\code{primaryTemporalKey}}{character \link{optional} Name of the column indicating time of record creation.}

\item{\code{snapshotPolicy}}{character Policy for using dataset snapshots when creating a project or making predictions. Must be one of the following values: 'specified': Use specific snapshot specified by catalogVersionId. 'latest': Use latest snapshot from the same catalog item. 'dynamic': Get data from the source (only applicable for JDBC datasets).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DatasetDefinitionResponse$new()}}
\item \href{#method-validate}{\code{DatasetDefinitionResponse$validate()}}
\item \href{#method-toJSON}{\code{DatasetDefinitionResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{DatasetDefinitionResponse$fromJSON()}}
\item \href{#method-clone}{\code{DatasetDefinitionResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetDefinitionResponse$new(
  catalogId = NULL,
  catalogVersionId = NULL,
  featureListId = NULL,
  identifier = NULL,
  snapshotPolicy = NULL,
  dataSource = NULL,
  dataSources = NULL,
  featureLists = NULL,
  isDeleted = NULL,
  originalIdentifier = NULL,
  primaryTemporalKey = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalogId}}{ID of the catalog item.}

\item{\code{catalogVersionId}}{ID of the catalog item version.}

\item{\code{featureListId}}{ID of the feature list. This decides which columns in the dataset are used for feature generation.}

\item{\code{identifier}}{Short name of the dataset (used directly as part of the generated feature names).}

\item{\code{snapshotPolicy}}{Policy for using dataset snapshots when creating a project or making predictions. Must be one of the following values: 'specified': Use specific snapshot specified by catalogVersionId. 'latest': Use latest snapshot from the same catalog item. 'dynamic': Get data from the source (only applicable for JDBC datasets).}

\item{\code{dataSources}}{Data source details for a JDBC dataset}

\item{\code{featureLists}}{List of available feature list ids for the dataset}

\item{\code{isDeleted}}{Is this dataset deleted?}

\item{\code{originalIdentifier}}{Original identifier of the dataset if it was updated to resolve name conflicts.}

\item{\code{primaryTemporalKey}}{Name of the column indicating time of record creation.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetDefinitionResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetDefinitionResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetDefinitionResponse$fromJSON(
  DatasetDefinitionResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatasetDefinitionResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetDefinitionResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
