% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_describe_permissions_response.R
\docType{class}
\name{DatasetDescribePermissionsResponse}
\alias{DatasetDescribePermissionsResponse}
\title{DatasetDescribePermissionsResponse}
\format{
An \code{R6Class} generator object
}
\description{
DatasetDescribePermissionsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{canCreateFeaturelist}}{character True if the user can create a new featurelist for this dataset.}

\item{\code{canDeleteDataset}}{character True if the user can delete dataset.}

\item{\code{canDownloadDatasetData}}{character True if the user can download data.}

\item{\code{canGetCatalogItemInfo}}{character True if the user can view catalog info.}

\item{\code{canGetDatasetInfo}}{character True if the user can view dataset info.}

\item{\code{canGetDatasetPermissions}}{character True if the user can view dataset permissions.}

\item{\code{canGetFeatureInfo}}{character True if the user can retrieve feature info of dataset.}

\item{\code{canGetFeaturelists}}{character True if the user can view featurelist for this dataset.}

\item{\code{canPatchCatalogInfo}}{character True if the user can modify catalog info.}

\item{\code{canPatchDatasetAliases}}{character True if the user can modify dataset feature aliases.}

\item{\code{canPatchDatasetInfo}}{character True if the user can modify dataset info.}

\item{\code{canPatchDatasetPermissions}}{character True if the user can modify dataset permissions.}

\item{\code{canPatchFeaturelists}}{character True if the user can modify featurelists for this dataset.}

\item{\code{canPostDataset}}{character True if the user can create a new dataset.}

\item{\code{canReloadDataset}}{character True if the user can reload dataset.}

\item{\code{canShareDataset}}{character True if the user can share the dataset.}

\item{\code{canSnapshotDataset}}{character True if the user can save snapshot of dataset.}

\item{\code{canUndeleteDataset}}{character True if the user can undelete dataset.}

\item{\code{canUseDatasetData}}{character True if the user can use dataset data to create projects, train custom models or provide predictions.}

\item{\code{canUseFeaturelists}}{character True if the user can use featurelists for this dataset. (for project creation)}

\item{\code{datasetId}}{character The ID of the dataset.}

\item{\code{uid}}{character The ID of the user identified by username.}

\item{\code{username}}{character `username` of a user with access to this dataset.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DatasetDescribePermissionsResponse$new()}}
\item \href{#method-validate}{\code{DatasetDescribePermissionsResponse$validate()}}
\item \href{#method-toJSON}{\code{DatasetDescribePermissionsResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{DatasetDescribePermissionsResponse$fromJSON()}}
\item \href{#method-clone}{\code{DatasetDescribePermissionsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetDescribePermissionsResponse$new(
  canCreateFeaturelist = NULL,
  canDeleteDataset = NULL,
  canDownloadDatasetData = NULL,
  canGetCatalogItemInfo = NULL,
  canGetDatasetInfo = NULL,
  canGetDatasetPermissions = NULL,
  canGetFeatureInfo = NULL,
  canGetFeaturelists = NULL,
  canPatchCatalogInfo = NULL,
  canPatchDatasetAliases = NULL,
  canPatchDatasetInfo = NULL,
  canPatchDatasetPermissions = NULL,
  canPatchFeaturelists = NULL,
  canPostDataset = NULL,
  canReloadDataset = NULL,
  canShareDataset = NULL,
  canSnapshotDataset = NULL,
  canUndeleteDataset = NULL,
  canUseDatasetData = NULL,
  canUseFeaturelists = NULL,
  datasetId = NULL,
  uid = NULL,
  username = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{canCreateFeaturelist}}{True if the user can create a new featurelist for this dataset.}

\item{\code{canDeleteDataset}}{True if the user can delete dataset.}

\item{\code{canDownloadDatasetData}}{True if the user can download data.}

\item{\code{canGetCatalogItemInfo}}{True if the user can view catalog info.}

\item{\code{canGetDatasetInfo}}{True if the user can view dataset info.}

\item{\code{canGetDatasetPermissions}}{True if the user can view dataset permissions.}

\item{\code{canGetFeatureInfo}}{True if the user can retrieve feature info of dataset.}

\item{\code{canGetFeaturelists}}{True if the user can view featurelist for this dataset.}

\item{\code{canPatchCatalogInfo}}{True if the user can modify catalog info.}

\item{\code{canPatchDatasetAliases}}{True if the user can modify dataset feature aliases.}

\item{\code{canPatchDatasetInfo}}{True if the user can modify dataset info.}

\item{\code{canPatchDatasetPermissions}}{True if the user can modify dataset permissions.}

\item{\code{canPatchFeaturelists}}{True if the user can modify featurelists for this dataset.}

\item{\code{canPostDataset}}{True if the user can create a new dataset.}

\item{\code{canReloadDataset}}{True if the user can reload dataset.}

\item{\code{canShareDataset}}{True if the user can share the dataset.}

\item{\code{canSnapshotDataset}}{True if the user can save snapshot of dataset.}

\item{\code{canUndeleteDataset}}{True if the user can undelete dataset.}

\item{\code{canUseDatasetData}}{True if the user can use dataset data to create projects, train custom models or provide predictions.}

\item{\code{canUseFeaturelists}}{True if the user can use featurelists for this dataset. (for project creation)}

\item{\code{datasetId}}{The ID of the dataset.}

\item{\code{uid}}{The ID of the user identified by username.}

\item{\code{username}}{`username` of a user with access to this dataset.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetDescribePermissionsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetDescribePermissionsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetDescribePermissionsResponse$fromJSON(
  DatasetDescribePermissionsResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatasetDescribePermissionsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetDescribePermissionsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
