% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_featurelist_response.R
\docType{class}
\name{DatasetFeaturelistResponse}
\alias{DatasetFeaturelistResponse}
\title{DatasetFeaturelistResponse}
\format{
An \code{R6Class} generator object
}
\description{
DatasetFeaturelistResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{createdBy}}{character `username` of the user who created the dataset.}

\item{\code{creationDate}}{character the ISO 8601 formatted date and time when the dataset was created.}

\item{\code{datasetId}}{character The ID of the dataset.}

\item{\code{datasetVersionId}}{character The ID of the dataset version if the featurelist is associated with a specific dataset version, for example Informative Features, or null otherwise.}

\item{\code{features}}{list( character ) Features in the featurelist.}

\item{\code{id}}{character The ID of the featurelist}

\item{\code{name}}{character The name of the featurelist}

\item{\code{userCreated}}{character True if the featurelist was created by a user vs the system.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatasetFeaturelistResponse-new}{\code{DatasetFeaturelistResponse$new()}}
\item \href{#method-DatasetFeaturelistResponse-validate}{\code{DatasetFeaturelistResponse$validate()}}
\item \href{#method-DatasetFeaturelistResponse-toJSON}{\code{DatasetFeaturelistResponse$toJSON()}}
\item \href{#method-DatasetFeaturelistResponse-fromJSON}{\code{DatasetFeaturelistResponse$fromJSON()}}
\item \href{#method-DatasetFeaturelistResponse-clone}{\code{DatasetFeaturelistResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetFeaturelistResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetFeaturelistResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetFeaturelistResponse$new(
  createdBy = NULL,
  creationDate = NULL,
  datasetId = NULL,
  datasetVersionId = NULL,
  features = NULL,
  id = NULL,
  name = NULL,
  userCreated = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{createdBy}}{`username` of the user who created the dataset.}

\item{\code{creationDate}}{the ISO 8601 formatted date and time when the dataset was created.}

\item{\code{datasetId}}{The ID of the dataset.}

\item{\code{datasetVersionId}}{The ID of the dataset version if the featurelist is associated with a specific dataset version, for example Informative Features, or null otherwise.}

\item{\code{features}}{Features in the featurelist.}

\item{\code{id}}{The ID of the featurelist}

\item{\code{name}}{The name of the featurelist}

\item{\code{userCreated}}{True if the featurelist was created by a user vs the system.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetFeaturelistResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetFeaturelistResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetFeaturelistResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetFeaturelistResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetFeaturelistResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetFeaturelistResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetFeaturelistResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetFeaturelistResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetFeaturelistResponse$fromJSON(
  DatasetFeaturelistResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatasetFeaturelistResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetFeaturelistResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetFeaturelistResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetFeaturelistResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
