% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_feature_response.R
\docType{class}
\name{DatasetFeatureResponse}
\alias{DatasetFeatureResponse}
\title{DatasetFeatureResponse}
\format{
An \code{R6Class} generator object
}
\description{
DatasetFeatureResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datasetId}}{character The ID of the dataset the feature belongs to}

\item{\code{datasetVersionId}}{character The ID of the dataset version the feature belongs to.}

\item{\code{dateFormat}}{character The date format string for how this feature was interpreted (or null if not a date feature). If not null, it will be compatible with https://docs.python.org/2/library/time.html#time.strftime .}

\item{\code{featureType}}{character Feature type.}

\item{\code{id}}{integer The number of the column in the dataset.}

\item{\code{isZeroInflated}}{character \link{optional} whether feature has an excessive number of zeros}

\item{\code{keySummary}}{\link{OneOfFeatureKeySummaryResponseValidatorSummarizedCategoricalarray} \link{optional} Per key summaries for Summarized Categorical or Multicategorical columns}

\item{\code{lowInformation}}{character \link{optional} Whether feature has too few values to be informative.}

\item{\code{max}}{numeric \link{optional} Maximum value of the EDA sample of the feature.}

\item{\code{mean}}{numeric \link{optional} Arithmetic mean of the EDA sample of the feature.}

\item{\code{median}}{numeric \link{optional} Median of the EDA sample of the feature.}

\item{\code{min}}{numeric \link{optional} Minimum value of the EDA sample of the feature.}

\item{\code{naCount}}{integer \link{optional} Number of missing values.}

\item{\code{name}}{character Feature name}

\item{\code{stdDev}}{numeric \link{optional} Standard deviation of EDA sample of the feature.}

\item{\code{timeSeriesEligibilityReason}}{character \link{optional} why the feature is ineligible for time series projects, or 'suitable' if it is eligible.}

\item{\code{timeSeriesEligibilityReasonAggregation}}{character \link{optional} why the feature is ineligible for aggregation, or 'suitable' if it is eligible.}

\item{\code{timeSeriesEligible}}{character \link{optional} whether this feature can be used as a datetime partitioning feature for time series projects.  Only sufficiently regular date features can be selected as the datetime feature for time series projects.  Always false for non-date features. Date features that cannot be used in datetime partitioning for a time series project may be eligible for an OTV project, which has less stringent requirements.}

\item{\code{timeSeriesEligibleAggregation}}{character \link{optional} whether this feature can be used as a datetime feature for aggregationfor time series data prep.  Always false for non-date features.}

\item{\code{timeStep}}{integer \link{optional} The minimum time step that can be used to specify time series windows.  The units for this value are the ``timeUnit``.  When specifying windows for time series projects, all windows must have durations that are integer multiples of this number. Only present for date features that are eligible for time series projects and null otherwise.}

\item{\code{timeStepAggregation}}{integer \link{optional} The minimum time step that can be used to aggregate using this feature for time series data prep. The units for this value are the ``timeUnit``.  Only present for date features that are eligible for aggregation in time series data prep and null otherwise.}

\item{\code{timeUnit}}{character \link{optional} The unit for the interval between values of this feature, e.g. DAY, MONTH, HOUR.  When specifying windows for time series projects, the windows are expressed in terms of this unit.  Only present for date features eligible for time series projects, and null otherwise.}

\item{\code{timeUnitAggregation}}{character \link{optional} The unit for the interval between values of this feature, e.g. DAY, MONTH, HOUR.  Only present for date features eligible for aggregation, and null otherwise.}

\item{\code{uniqueCount}}{integer \link{optional} Number of unique values.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatasetFeatureResponse-new}{\code{DatasetFeatureResponse$new()}}
\item \href{#method-DatasetFeatureResponse-validate}{\code{DatasetFeatureResponse$validate()}}
\item \href{#method-DatasetFeatureResponse-toJSON}{\code{DatasetFeatureResponse$toJSON()}}
\item \href{#method-DatasetFeatureResponse-fromJSON}{\code{DatasetFeatureResponse$fromJSON()}}
\item \href{#method-DatasetFeatureResponse-clone}{\code{DatasetFeatureResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetFeatureResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetFeatureResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetFeatureResponse$new(
  datasetId = NULL,
  datasetVersionId = NULL,
  dateFormat = NULL,
  featureType = NULL,
  id = NULL,
  name = NULL,
  isZeroInflated = NULL,
  keySummary = NULL,
  lowInformation = NULL,
  max = NULL,
  mean = NULL,
  median = NULL,
  min = NULL,
  naCount = NULL,
  stdDev = NULL,
  timeSeriesEligibilityReason = NULL,
  timeSeriesEligibilityReasonAggregation = NULL,
  timeSeriesEligible = NULL,
  timeSeriesEligibleAggregation = NULL,
  timeStep = NULL,
  timeStepAggregation = NULL,
  timeUnit = NULL,
  timeUnitAggregation = NULL,
  uniqueCount = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasetId}}{The ID of the dataset the feature belongs to}

\item{\code{datasetVersionId}}{The ID of the dataset version the feature belongs to.}

\item{\code{dateFormat}}{The date format string for how this feature was interpreted (or null if not a date feature). If not null, it will be compatible with https://docs.python.org/2/library/time.html#time.strftime .}

\item{\code{featureType}}{Feature type.}

\item{\code{id}}{The number of the column in the dataset.}

\item{\code{name}}{Feature name}

\item{\code{isZeroInflated}}{whether feature has an excessive number of zeros}

\item{\code{keySummary}}{Per key summaries for Summarized Categorical or Multicategorical columns}

\item{\code{lowInformation}}{Whether feature has too few values to be informative.}

\item{\code{max}}{Maximum value of the EDA sample of the feature.}

\item{\code{mean}}{Arithmetic mean of the EDA sample of the feature.}

\item{\code{median}}{Median of the EDA sample of the feature.}

\item{\code{min}}{Minimum value of the EDA sample of the feature.}

\item{\code{naCount}}{Number of missing values.}

\item{\code{stdDev}}{Standard deviation of EDA sample of the feature.}

\item{\code{timeSeriesEligibilityReason}}{why the feature is ineligible for time series projects, or 'suitable' if it is eligible.}

\item{\code{timeSeriesEligibilityReasonAggregation}}{why the feature is ineligible for aggregation, or 'suitable' if it is eligible.}

\item{\code{timeSeriesEligible}}{whether this feature can be used as a datetime partitioning feature for time series projects.  Only sufficiently regular date features can be selected as the datetime feature for time series projects.  Always false for non-date features. Date features that cannot be used in datetime partitioning for a time series project may be eligible for an OTV project, which has less stringent requirements.}

\item{\code{timeSeriesEligibleAggregation}}{whether this feature can be used as a datetime feature for aggregationfor time series data prep.  Always false for non-date features.}

\item{\code{timeStep}}{The minimum time step that can be used to specify time series windows.  The units for this value are the ``timeUnit``.  When specifying windows for time series projects, all windows must have durations that are integer multiples of this number. Only present for date features that are eligible for time series projects and null otherwise.}

\item{\code{timeStepAggregation}}{The minimum time step that can be used to aggregate using this feature for time series data prep. The units for this value are the ``timeUnit``.  Only present for date features that are eligible for aggregation in time series data prep and null otherwise.}

\item{\code{timeUnit}}{The unit for the interval between values of this feature, e.g. DAY, MONTH, HOUR.  When specifying windows for time series projects, the windows are expressed in terms of this unit.  Only present for date features eligible for time series projects, and null otherwise.}

\item{\code{timeUnitAggregation}}{The unit for the interval between values of this feature, e.g. DAY, MONTH, HOUR.  Only present for date features eligible for aggregation, and null otherwise.}

\item{\code{uniqueCount}}{Number of unique values.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetFeatureResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetFeatureResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetFeatureResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetFeatureResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetFeatureResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetFeatureResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetFeatureResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetFeatureResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetFeatureResponse$fromJSON(
  DatasetFeatureResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatasetFeatureResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetFeatureResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetFeatureResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetFeatureResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
