% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_from_file.R
\docType{class}
\name{DatasetFromFile}
\alias{DatasetFromFile}
\title{DatasetFromFile}
\format{
An \code{R6Class} generator object
}
\description{
DatasetFromFile Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{categories}}{\link{OneOfstringarray} \link{optional} An array of strings describing the intended use of the dataset.}

\item{\code{file}}{data.frame The data to be used for the creation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatasetFromFile-new}{\code{DatasetFromFile$new()}}
\item \href{#method-DatasetFromFile-validate}{\code{DatasetFromFile$validate()}}
\item \href{#method-DatasetFromFile-toJSON}{\code{DatasetFromFile$toJSON()}}
\item \href{#method-DatasetFromFile-fromJSON}{\code{DatasetFromFile$fromJSON()}}
\item \href{#method-DatasetFromFile-clone}{\code{DatasetFromFile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetFromFile-new"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetFromFile-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetFromFile$new(
  file = NULL,
  categories = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{The data to be used for the creation.}

\item{\code{categories}}{An array of strings describing the intended use of the dataset.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetFromFile-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetFromFile-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetFromFile$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetFromFile-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetFromFile-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetFromFile$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetFromFile-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetFromFile-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetFromFile$fromJSON(DatasetFromFileJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatasetFromFileJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetFromFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetFromFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetFromFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
