% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_refresh_execution_result.R
\docType{class}
\name{DatasetRefreshExecutionResult}
\alias{DatasetRefreshExecutionResult}
\title{DatasetRefreshExecutionResult}
\format{
An \code{R6Class} generator object
}
\description{
DatasetRefreshExecutionResult Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{completedAt}}{character UTC completion date, in RFC-3339 format.}

\item{\code{datasetId}}{character Dataset ID associated with this result.}

\item{\code{datasetVersionId}}{character Dataset version ID associated with this result}

\item{\code{executionId}}{character Result ID.}

\item{\code{jobId}}{character Job ID associated with this result.}

\item{\code{message}}{character Current status of execution.}

\item{\code{startedAt}}{character UTC start date, in RFC-3339 format.}

\item{\code{status}}{character Status of this dataset refresh.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DatasetRefreshExecutionResult$new()}}
\item \href{#method-validate}{\code{DatasetRefreshExecutionResult$validate()}}
\item \href{#method-toJSON}{\code{DatasetRefreshExecutionResult$toJSON()}}
\item \href{#method-fromJSON}{\code{DatasetRefreshExecutionResult$fromJSON()}}
\item \href{#method-clone}{\code{DatasetRefreshExecutionResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRefreshExecutionResult$new(
  completedAt = NULL,
  datasetId = NULL,
  datasetVersionId = NULL,
  executionId = NULL,
  jobId = NULL,
  message = NULL,
  startedAt = NULL,
  status = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{completedAt}}{UTC completion date, in RFC-3339 format.}

\item{\code{datasetId}}{Dataset ID associated with this result.}

\item{\code{datasetVersionId}}{Dataset version ID associated with this result}

\item{\code{executionId}}{Result ID.}

\item{\code{jobId}}{Job ID associated with this result.}

\item{\code{message}}{Current status of execution.}

\item{\code{startedAt}}{UTC start date, in RFC-3339 format.}

\item{\code{status}}{Status of this dataset refresh.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRefreshExecutionResult$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRefreshExecutionResult$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRefreshExecutionResult$fromJSON(
  DatasetRefreshExecutionResultJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatasetRefreshExecutionResultJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRefreshExecutionResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
