% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_refresh_job_response.R
\docType{class}
\name{DatasetRefreshJobResponse}
\alias{DatasetRefreshJobResponse}
\title{DatasetRefreshJobResponse}
\format{
An \code{R6Class} generator object
}
\description{
DatasetRefreshJobResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{categories}}{\link{OneOfstringarray} An array of strings describing the intended use of the dataset.}

\item{\code{createdBy}}{character The user who created this dataset refresh job.}

\item{\code{credentialId}}{character ID used to validate with Kerberos authentication service if Kerberos is enabled.}

\item{\code{credentials}}{character A JSON string describing the data engine queries credentials to use when refreshing.}

\item{\code{datasetId}}{character ID of the dataset the user scheduled job applies to.}

\item{\code{enabled}}{character Indicates whether the scheduled job is active (true) or inactive(false).}

\item{\code{jobId}}{character The scheduled job ID.}

\item{\code{name}}{character The scheduled job's name.}

\item{\code{schedule}}{\link{DatasetRefresh}}

\item{\code{scheduleReferenceDate}}{character The UTC reference date in RFC-3339 format of when the schedule starts from. Can be used to help build a more intuitive schedule picker.}

\item{\code{updatedAt}}{character The UTC date in RFC-3339 format of when the job was last updated.}

\item{\code{updatedBy}}{character The user who last modified this dataset refresh job.}

\item{\code{useKerberos}}{character Boolean (true) if the Kerberos authentication service is needed when refreshing a job.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatasetRefreshJobResponse-new}{\code{DatasetRefreshJobResponse$new()}}
\item \href{#method-DatasetRefreshJobResponse-validate}{\code{DatasetRefreshJobResponse$validate()}}
\item \href{#method-DatasetRefreshJobResponse-toJSON}{\code{DatasetRefreshJobResponse$toJSON()}}
\item \href{#method-DatasetRefreshJobResponse-fromJSON}{\code{DatasetRefreshJobResponse$fromJSON()}}
\item \href{#method-DatasetRefreshJobResponse-clone}{\code{DatasetRefreshJobResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetRefreshJobResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetRefreshJobResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRefreshJobResponse$new(
  categories = NULL,
  createdBy = NULL,
  credentialId = NULL,
  credentials = NULL,
  datasetId = NULL,
  enabled = NULL,
  jobId = NULL,
  name = NULL,
  schedule = NULL,
  scheduleReferenceDate = NULL,
  updatedAt = NULL,
  updatedBy = NULL,
  useKerberos = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{categories}}{An array of strings describing the intended use of the dataset.}

\item{\code{createdBy}}{The user who created this dataset refresh job.}

\item{\code{credentialId}}{ID used to validate with Kerberos authentication service if Kerberos is enabled.}

\item{\code{credentials}}{A JSON string describing the data engine queries credentials to use when refreshing.}

\item{\code{datasetId}}{ID of the dataset the user scheduled job applies to.}

\item{\code{enabled}}{Indicates whether the scheduled job is active (true) or inactive(false).}

\item{\code{jobId}}{The scheduled job ID.}

\item{\code{name}}{The scheduled job's name.}

\item{\code{scheduleReferenceDate}}{The UTC reference date in RFC-3339 format of when the schedule starts from. Can be used to help build a more intuitive schedule picker.}

\item{\code{updatedAt}}{The UTC date in RFC-3339 format of when the job was last updated.}

\item{\code{updatedBy}}{The user who last modified this dataset refresh job.}

\item{\code{useKerberos}}{Boolean (true) if the Kerberos authentication service is needed when refreshing a job.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetRefreshJobResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetRefreshJobResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRefreshJobResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetRefreshJobResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetRefreshJobResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRefreshJobResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetRefreshJobResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetRefreshJobResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRefreshJobResponse$fromJSON(
  DatasetRefreshJobResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatasetRefreshJobResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetRefreshJobResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetRefreshJobResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRefreshJobResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
