% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_refresh_job_update.R
\docType{class}
\name{DatasetRefreshJobUpdate}
\alias{DatasetRefreshJobUpdate}
\title{DatasetRefreshJobUpdate}
\format{
An \code{R6Class} generator object
}
\description{
DatasetRefreshJobUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{categories}}{\link{OneOfstringarray} \link{optional} An array of strings describing the intended use of the dataset. The supported options are ``TRAINING``, and ``PREDICTION``.}

\item{\code{credentialId}}{character \link{optional} The ID of the set of credentials to use to run the scheduled job when the Kerberos authentication service is utilized. Required when useKerberos is true.}

\item{\code{credentials}}{character \link{optional} A JSON string describing the data engine queries credentials to use when refreshing.}

\item{\code{enabled}}{character \link{optional} Boolean for whether the scheduled job is active (true) or inactive (false).}

\item{\code{name}}{character \link{optional} Scheduled job name.}

\item{\code{schedule}}{\link{DatasetRefresh} \link{optional}}

\item{\code{scheduleReferenceDate}}{character \link{optional} The UTC reference date in RFC-3339 format of when the schedule starts from. This value is returned in `/api/v2/datasets/(datasetId)/refreshJobs/(jobId)/` to help build a more intuitive schedule picker. Required when ``schedule`` is being updated. The default is the current time.}

\item{\code{useKerberos}}{character \link{optional} If true, the Kerberos authentication system is used in conjunction with a credential ID.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DatasetRefreshJobUpdate$new()}}
\item \href{#method-validate}{\code{DatasetRefreshJobUpdate$validate()}}
\item \href{#method-toJSON}{\code{DatasetRefreshJobUpdate$toJSON()}}
\item \href{#method-fromJSON}{\code{DatasetRefreshJobUpdate$fromJSON()}}
\item \href{#method-clone}{\code{DatasetRefreshJobUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRefreshJobUpdate$new(
  categories = NULL,
  credentialId = NULL,
  credentials = NULL,
  enabled = NULL,
  name = NULL,
  schedule = NULL,
  scheduleReferenceDate = NULL,
  useKerberos = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{categories}}{An array of strings describing the intended use of the dataset. The supported options are ``TRAINING``, and ``PREDICTION``.}

\item{\code{credentialId}}{The ID of the set of credentials to use to run the scheduled job when the Kerberos authentication service is utilized. Required when useKerberos is true.}

\item{\code{credentials}}{A JSON string describing the data engine queries credentials to use when refreshing.}

\item{\code{enabled}}{Boolean for whether the scheduled job is active (true) or inactive (false).}

\item{\code{name}}{Scheduled job name.}

\item{\code{scheduleReferenceDate}}{The UTC reference date in RFC-3339 format of when the schedule starts from. This value is returned in `/api/v2/datasets/(datasetId)/refreshJobs/(jobId)/` to help build a more intuitive schedule picker. Required when ``schedule`` is being updated. The default is the current time.}

\item{\code{useKerberos}}{If true, the Kerberos authentication system is used in conjunction with a credential ID.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRefreshJobUpdate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRefreshJobUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRefreshJobUpdate$fromJSON(
  DatasetRefreshJobUpdateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatasetRefreshJobUpdateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRefreshJobUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
