% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_relationship_update.R
\docType{class}
\name{DatasetRelationshipUpdate}
\alias{DatasetRelationshipUpdate}
\title{DatasetRelationshipUpdate}
\format{
An \code{R6Class} generator object
}
\description{
DatasetRelationshipUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{linkedDatasetId}}{character \link{optional} id of another dataset with which to create relationships}

\item{\code{linkedFeatures}}{list( character ) \link{optional} list of features belonging to the linked dataset}

\item{\code{sourceFeatures}}{list( character ) \link{optional} list of features belonging to the source dataset}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatasetRelationshipUpdate-new}{\code{DatasetRelationshipUpdate$new()}}
\item \href{#method-DatasetRelationshipUpdate-validate}{\code{DatasetRelationshipUpdate$validate()}}
\item \href{#method-DatasetRelationshipUpdate-toJSON}{\code{DatasetRelationshipUpdate$toJSON()}}
\item \href{#method-DatasetRelationshipUpdate-fromJSON}{\code{DatasetRelationshipUpdate$fromJSON()}}
\item \href{#method-DatasetRelationshipUpdate-clone}{\code{DatasetRelationshipUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetRelationshipUpdate-new"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetRelationshipUpdate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRelationshipUpdate$new(
  linkedDatasetId = NULL,
  linkedFeatures = NULL,
  sourceFeatures = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{linkedDatasetId}}{id of another dataset with which to create relationships}

\item{\code{linkedFeatures}}{list of features belonging to the linked dataset}

\item{\code{sourceFeatures}}{list of features belonging to the source dataset}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetRelationshipUpdate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetRelationshipUpdate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRelationshipUpdate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetRelationshipUpdate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetRelationshipUpdate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRelationshipUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetRelationshipUpdate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetRelationshipUpdate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRelationshipUpdate$fromJSON(
  DatasetRelationshipUpdateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatasetRelationshipUpdateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetRelationshipUpdate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetRelationshipUpdate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRelationshipUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
