% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_roles_with_id.R
\docType{class}
\name{DatasetRolesWithId}
\alias{DatasetRolesWithId}
\title{DatasetRolesWithId}
\format{
An \code{R6Class} generator object
}
\description{
DatasetRolesWithId Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{canShare}}{character \link{optional} Whether the org/group/user should be able to share with others. If true, the org/group/user will be able to grant any role up to and including their own to other orgs/groups/user. If `role` is `NO_ROLE` `canShare` is ignored.}

\item{\code{canUseData}}{character \link{optional} Whether the user/group/org should be able to view, download and process data (use to create projects, predictions, etc). For OWNER `canUseData` is always True. If `role` is empty `canUseData` is ignored.}

\item{\code{id}}{character The org/group/user ID.}

\item{\code{role}}{character The role of the org/group/user on this dataset or \"NO_ROLE\" for removing access when used with route to modify access.}

\item{\code{shareRecipientType}}{character It describes the recipient type.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DatasetRolesWithId$new()}}
\item \href{#method-validate}{\code{DatasetRolesWithId$validate()}}
\item \href{#method-toJSON}{\code{DatasetRolesWithId$toJSON()}}
\item \href{#method-fromJSON}{\code{DatasetRolesWithId$fromJSON()}}
\item \href{#method-clone}{\code{DatasetRolesWithId$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRolesWithId$new(
  id = NULL,
  role = NULL,
  shareRecipientType = NULL,
  canShare = FALSE,
  canUseData = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The org/group/user ID.}

\item{\code{role}}{The role of the org/group/user on this dataset or \"NO_ROLE\" for removing access when used with route to modify access.}

\item{\code{shareRecipientType}}{It describes the recipient type.}

\item{\code{canShare}}{Whether the org/group/user should be able to share with others. If true, the org/group/user will be able to grant any role up to and including their own to other orgs/groups/user. If `role` is `NO_ROLE` `canShare` is ignored.}

\item{\code{canUseData}}{Whether the user/group/org should be able to view, download and process data (use to create projects, predictions, etc). For OWNER `canUseData` is always True. If `role` is empty `canUseData` is ignored.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRolesWithId$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRolesWithId$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRolesWithId$fromJSON(DatasetRolesWithIdJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatasetRolesWithIdJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRolesWithId$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
