% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_transform_response.R
\docType{class}
\name{DatasetTransformResponse}
\alias{DatasetTransformResponse}
\title{DatasetTransformResponse}
\format{
An \code{R6Class} generator object
}
\description{
DatasetTransformResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dateExtraction}}{character The value to extract from the date column.}

\item{\code{name}}{character The feature name.}

\item{\code{parentName}}{character The name of the parent feature.}

\item{\code{replacement}}{\link{AnyOfstringbooleannumberinteger} The replacement in case of a failed transformation.}

\item{\code{variableType}}{character The type of the transform.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatasetTransformResponse-new}{\code{DatasetTransformResponse$new()}}
\item \href{#method-DatasetTransformResponse-validate}{\code{DatasetTransformResponse$validate()}}
\item \href{#method-DatasetTransformResponse-toJSON}{\code{DatasetTransformResponse$toJSON()}}
\item \href{#method-DatasetTransformResponse-fromJSON}{\code{DatasetTransformResponse$fromJSON()}}
\item \href{#method-DatasetTransformResponse-clone}{\code{DatasetTransformResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetTransformResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetTransformResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetTransformResponse$new(
  dateExtraction = NULL,
  name = NULL,
  parentName = NULL,
  replacement = NULL,
  variableType = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dateExtraction}}{The value to extract from the date column.}

\item{\code{name}}{The feature name.}

\item{\code{parentName}}{The name of the parent feature.}

\item{\code{replacement}}{The replacement in case of a failed transformation.}

\item{\code{variableType}}{The type of the transform.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetTransformResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetTransformResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetTransformResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetTransformResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetTransformResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetTransformResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetTransformResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetTransformResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetTransformResponse$fromJSON(
  DatasetTransformResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatasetTransformResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetTransformResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetTransformResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetTransformResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
