% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source_describe_permissions_response.R
\docType{class}
\name{DataSourceDescribePermissionsResponse}
\alias{DataSourceDescribePermissionsResponse}
\title{DataSourceDescribePermissionsResponse}
\format{
An \code{R6Class} generator object
}
\description{
DataSourceDescribePermissionsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{canCreatePredictions}}{character True if the user can create predictions from the data source.}

\item{\code{canCreateProject}}{character True if the user can create project from the data source.}

\item{\code{canDelete}}{character True if the user can delete the data source.}

\item{\code{canEdit}}{character True if the user can edit data source info.}

\item{\code{canSetRoles}}{list( character ) Roles the user can grant or revoke from other users, groups, or organizations.}

\item{\code{canShare}}{character True if the user can share the data source.}

\item{\code{canView}}{character True if the user can view data source info.}

\item{\code{dataSourceId}}{character The ID of the data source.}

\item{\code{userId}}{character The ID of the user identified by username.}

\item{\code{username}}{character `username` of a user with access to this data source.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataSourceDescribePermissionsResponse$new()}}
\item \href{#method-validate}{\code{DataSourceDescribePermissionsResponse$validate()}}
\item \href{#method-toJSON}{\code{DataSourceDescribePermissionsResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{DataSourceDescribePermissionsResponse$fromJSON()}}
\item \href{#method-clone}{\code{DataSourceDescribePermissionsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSourceDescribePermissionsResponse$new(
  canCreatePredictions = NULL,
  canCreateProject = NULL,
  canDelete = NULL,
  canEdit = NULL,
  canSetRoles = NULL,
  canShare = NULL,
  canView = NULL,
  dataSourceId = NULL,
  userId = NULL,
  username = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{canCreatePredictions}}{True if the user can create predictions from the data source.}

\item{\code{canCreateProject}}{True if the user can create project from the data source.}

\item{\code{canDelete}}{True if the user can delete the data source.}

\item{\code{canEdit}}{True if the user can edit data source info.}

\item{\code{canSetRoles}}{Roles the user can grant or revoke from other users, groups, or organizations.}

\item{\code{canShare}}{True if the user can share the data source.}

\item{\code{canView}}{True if the user can view data source info.}

\item{\code{dataSourceId}}{The ID of the data source.}

\item{\code{userId}}{The ID of the user identified by username.}

\item{\code{username}}{`username` of a user with access to this data source.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSourceDescribePermissionsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSourceDescribePermissionsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSourceDescribePermissionsResponse$fromJSON(
  DataSourceDescribePermissionsResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DataSourceDescribePermissionsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSourceDescribePermissionsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
