% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source_in_list.R
\docType{class}
\name{DataSourceInList}
\alias{DataSourceInList}
\title{DataSourceInList}
\format{
An \code{R6Class} generator object
}
\description{
DataSourceInList Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{canDelete}}{character True if the user can delete the data source.}

\item{\code{canShare}}{character True if the user can share the data source.}

\item{\code{canonicalName}}{character Data source canonical name.}

\item{\code{creator}}{character ID of the user who created the data source.}

\item{\code{id}}{character Data source ID.}

\item{\code{params}}{\link{OneOfJDBCTableDataSourceJDBCQueryDataSourceDRConnectorV1DataSource} Data source configuration.}

\item{\code{role}}{character Role of the user making the request on this data source.}

\item{\code{type}}{character Data source type.}

\item{\code{updated}}{character ISO 8601-formatted date/time of the data source update.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataSourceInList-new}{\code{DataSourceInList$new()}}
\item \href{#method-DataSourceInList-validate}{\code{DataSourceInList$validate()}}
\item \href{#method-DataSourceInList-toJSON}{\code{DataSourceInList$toJSON()}}
\item \href{#method-DataSourceInList-fromJSON}{\code{DataSourceInList$fromJSON()}}
\item \href{#method-DataSourceInList-clone}{\code{DataSourceInList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSourceInList-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataSourceInList-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSourceInList$new(
  canDelete = NULL,
  canShare = NULL,
  canonicalName = NULL,
  creator = NULL,
  id = NULL,
  params = NULL,
  role = NULL,
  type = NULL,
  updated = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{canDelete}}{True if the user can delete the data source.}

\item{\code{canShare}}{True if the user can share the data source.}

\item{\code{canonicalName}}{Data source canonical name.}

\item{\code{creator}}{ID of the user who created the data source.}

\item{\code{id}}{Data source ID.}

\item{\code{params}}{Data source configuration.}

\item{\code{role}}{Role of the user making the request on this data source.}

\item{\code{type}}{Data source type.}

\item{\code{updated}}{ISO 8601-formatted date/time of the data source update.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSourceInList-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DataSourceInList-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSourceInList$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSourceInList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataSourceInList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSourceInList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSourceInList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataSourceInList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSourceInList$fromJSON(DataSourceInListJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DataSourceInListJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSourceInList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataSourceInList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSourceInList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
