% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_store_column_response.R
\docType{class}
\name{DataStoreColumnResponse}
\alias{DataStoreColumnResponse}
\title{DataStoreColumnResponse}
\format{
An \code{R6Class} generator object
}
\description{
DataStoreColumnResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dataType}}{character DataType of the column.}

\item{\code{name}}{character Name of the column.}

\item{\code{precision}}{character Precision of the column.}

\item{\code{scale}}{character Scale of the column.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataStoreColumnResponse-new}{\code{DataStoreColumnResponse$new()}}
\item \href{#method-DataStoreColumnResponse-validate}{\code{DataStoreColumnResponse$validate()}}
\item \href{#method-DataStoreColumnResponse-toJSON}{\code{DataStoreColumnResponse$toJSON()}}
\item \href{#method-DataStoreColumnResponse-fromJSON}{\code{DataStoreColumnResponse$fromJSON()}}
\item \href{#method-DataStoreColumnResponse-clone}{\code{DataStoreColumnResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreColumnResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreColumnResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreColumnResponse$new(
  dataType = NULL,
  name = NULL,
  precision = NULL,
  scale = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataType}}{DataType of the column.}

\item{\code{name}}{Name of the column.}

\item{\code{precision}}{Precision of the column.}

\item{\code{scale}}{Scale of the column.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreColumnResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreColumnResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreColumnResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreColumnResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreColumnResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreColumnResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreColumnResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreColumnResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreColumnResponse$fromJSON(
  DataStoreColumnResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DataStoreColumnResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreColumnResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreColumnResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreColumnResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
