% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_store_columns_list.R
\docType{class}
\name{DataStoreColumnsList}
\alias{DataStoreColumnsList}
\title{DataStoreColumnsList}
\format{
An \code{R6Class} generator object
}
\description{
DataStoreColumnsList Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalog}}{character \link{optional} Name of specified database catalog.}

\item{\code{credentialId}}{character \link{optional} Id of credential mapping.}

\item{\code{password}}{character \link{optional} Password for the db connection.}

\item{\code{query}}{character \link{optional} Schema query.}

\item{\code{schema}}{character \link{optional} Schema name.}

\item{\code{table}}{character \link{optional} Table name.}

\item{\code{user}}{character \link{optional} Username for the db connection.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataStoreColumnsList-new}{\code{DataStoreColumnsList$new()}}
\item \href{#method-DataStoreColumnsList-validate}{\code{DataStoreColumnsList$validate()}}
\item \href{#method-DataStoreColumnsList-toJSON}{\code{DataStoreColumnsList$toJSON()}}
\item \href{#method-DataStoreColumnsList-fromJSON}{\code{DataStoreColumnsList$fromJSON()}}
\item \href{#method-DataStoreColumnsList-clone}{\code{DataStoreColumnsList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreColumnsList-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreColumnsList-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreColumnsList$new(
  catalog = NULL,
  credentialId = NULL,
  password = NULL,
  query = NULL,
  schema = NULL,
  table = NULL,
  user = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalog}}{Name of specified database catalog.}

\item{\code{credentialId}}{Id of credential mapping.}

\item{\code{password}}{Password for the db connection.}

\item{\code{query}}{Schema query.}

\item{\code{schema}}{Schema name.}

\item{\code{table}}{Table name.}

\item{\code{user}}{Username for the db connection.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreColumnsList-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreColumnsList-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreColumnsList$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreColumnsList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreColumnsList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreColumnsList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreColumnsList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreColumnsList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreColumnsList$fromJSON(DataStoreColumnsListJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DataStoreColumnsListJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreColumnsList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreColumnsList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreColumnsList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
