% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_store_describe_permissions_response.R
\docType{class}
\name{DataStoreDescribePermissionsResponse}
\alias{DataStoreDescribePermissionsResponse}
\title{DataStoreDescribePermissionsResponse}
\format{
An \code{R6Class} generator object
}
\description{
DataStoreDescribePermissionsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{canCreateDataSource}}{character True if the user can create data source from this data store.}

\item{\code{canDelete}}{character True if the user can delete the data store.}

\item{\code{canEdit}}{character True if the user can edit data store info.}

\item{\code{canScanDatabase}}{character True if the user can scan data store database.}

\item{\code{canSetRoles}}{list( character ) Roles the user can grant or revoke from other users, groups, or organizations.}

\item{\code{canShare}}{character True if the user can share the data store.}

\item{\code{canTestConnection}}{character True if the user can test data store database connection.}

\item{\code{canView}}{character True if the user can view data store info.}

\item{\code{dataStoreId}}{character The ID of the data store.}

\item{\code{userId}}{character The ID of the user identified by username.}

\item{\code{username}}{character `username` of a user with access to this data store.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataStoreDescribePermissionsResponse$new()}}
\item \href{#method-validate}{\code{DataStoreDescribePermissionsResponse$validate()}}
\item \href{#method-toJSON}{\code{DataStoreDescribePermissionsResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{DataStoreDescribePermissionsResponse$fromJSON()}}
\item \href{#method-clone}{\code{DataStoreDescribePermissionsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreDescribePermissionsResponse$new(
  canCreateDataSource = NULL,
  canDelete = NULL,
  canEdit = NULL,
  canScanDatabase = NULL,
  canSetRoles = NULL,
  canShare = NULL,
  canTestConnection = NULL,
  canView = NULL,
  dataStoreId = NULL,
  userId = NULL,
  username = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{canCreateDataSource}}{True if the user can create data source from this data store.}

\item{\code{canDelete}}{True if the user can delete the data store.}

\item{\code{canEdit}}{True if the user can edit data store info.}

\item{\code{canScanDatabase}}{True if the user can scan data store database.}

\item{\code{canSetRoles}}{Roles the user can grant or revoke from other users, groups, or organizations.}

\item{\code{canShare}}{True if the user can share the data store.}

\item{\code{canTestConnection}}{True if the user can test data store database connection.}

\item{\code{canView}}{True if the user can view data store info.}

\item{\code{dataStoreId}}{The ID of the data store.}

\item{\code{userId}}{The ID of the user identified by username.}

\item{\code{username}}{`username` of a user with access to this data store.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreDescribePermissionsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreDescribePermissionsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreDescribePermissionsResponse$fromJSON(
  DataStoreDescribePermissionsResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DataStoreDescribePermissionsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreDescribePermissionsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
