% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_store_retrieve_response.R
\docType{class}
\name{DataStoreRetrieveResponse}
\alias{DataStoreRetrieveResponse}
\title{DataStoreRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
DataStoreRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{associatedAuthTypes}}{list( character ) \link{optional} Supported authentication types for the JDBC configuration.}

\item{\code{canonicalName}}{character The user-friendly name of the data store.}

\item{\code{creator}}{character ID of the user who created the data store.}

\item{\code{id}}{character Data store ID.}

\item{\code{params}}{\link{OneOfJDBCDataStoreDetailsDRConnectorV1Details} Data store configuration.}

\item{\code{role}}{character Role of the user making the request on this data store.}

\item{\code{type}}{character Data store type.}

\item{\code{updated}}{character ISO 8601-formatted date/time of the data store update.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataStoreRetrieveResponse-new}{\code{DataStoreRetrieveResponse$new()}}
\item \href{#method-DataStoreRetrieveResponse-validate}{\code{DataStoreRetrieveResponse$validate()}}
\item \href{#method-DataStoreRetrieveResponse-toJSON}{\code{DataStoreRetrieveResponse$toJSON()}}
\item \href{#method-DataStoreRetrieveResponse-fromJSON}{\code{DataStoreRetrieveResponse$fromJSON()}}
\item \href{#method-DataStoreRetrieveResponse-clone}{\code{DataStoreRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreRetrieveResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreRetrieveResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreRetrieveResponse$new(
  canonicalName = NULL,
  creator = NULL,
  id = NULL,
  params = NULL,
  role = NULL,
  type = NULL,
  updated = NULL,
  associatedAuthTypes = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{canonicalName}}{The user-friendly name of the data store.}

\item{\code{creator}}{ID of the user who created the data store.}

\item{\code{id}}{Data store ID.}

\item{\code{params}}{Data store configuration.}

\item{\code{role}}{Role of the user making the request on this data store.}

\item{\code{type}}{Data store type.}

\item{\code{updated}}{ISO 8601-formatted date/time of the data store update.}

\item{\code{associatedAuthTypes}}{Supported authentication types for the JDBC configuration.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreRetrieveResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreRetrieveResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreRetrieveResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreRetrieveResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreRetrieveResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreRetrieveResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreRetrieveResponse$fromJSON(
  DataStoreRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DataStoreRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreRetrieveResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreRetrieveResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
