% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_store_schemas_list.R
\docType{class}
\name{DataStoreSchemasList}
\alias{DataStoreSchemasList}
\title{DataStoreSchemasList}
\format{
An \code{R6Class} generator object
}
\description{
DataStoreSchemasList Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalog}}{character Name of the catalog associated with the schemas. If more than one catalog is queried, the returned value will be the catalog name for the first record returned.}

\item{\code{catalogs}}{list( character ) List of catalogs associated with each retrieved schema. If no catalog is found, this value is null.}

\item{\code{schemas}}{list( character ) List of schemas available in the data store.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataStoreSchemasList-new}{\code{DataStoreSchemasList$new()}}
\item \href{#method-DataStoreSchemasList-validate}{\code{DataStoreSchemasList$validate()}}
\item \href{#method-DataStoreSchemasList-toJSON}{\code{DataStoreSchemasList$toJSON()}}
\item \href{#method-DataStoreSchemasList-fromJSON}{\code{DataStoreSchemasList$fromJSON()}}
\item \href{#method-DataStoreSchemasList-clone}{\code{DataStoreSchemasList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreSchemasList-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreSchemasList-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreSchemasList$new(
  catalog = NULL,
  catalogs = NULL,
  schemas = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalog}}{Name of the catalog associated with the schemas. If more than one catalog is queried, the returned value will be the catalog name for the first record returned.}

\item{\code{catalogs}}{List of catalogs associated with each retrieved schema. If no catalog is found, this value is null.}

\item{\code{schemas}}{List of schemas available in the data store.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreSchemasList-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreSchemasList-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreSchemasList$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreSchemasList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreSchemasList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreSchemasList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreSchemasList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreSchemasList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreSchemasList$fromJSON(DataStoreSchemasListJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DataStoreSchemasListJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreSchemasList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreSchemasList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreSchemasList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
