% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_store_sql_verify.R
\docType{class}
\name{DataStoreSQLVerify}
\alias{DataStoreSQLVerify}
\title{DataStoreSQLVerify}
\format{
An \code{R6Class} generator object
}
\description{
DataStoreSQLVerify Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{credentialId}}{character \link{optional} ID of the set of credentials to use instead of username and password.}

\item{\code{maxRows}}{integer Maximum number of rows of data to return if successful.}

\item{\code{password}}{character \link{optional} Password for data store authentication.}

\item{\code{query}}{character SQL query to verify.}

\item{\code{useKerberos}}{character \link{optional} Whether to use Kerberos for data store authentication.}

\item{\code{user}}{character \link{optional} Username for data store authentication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataStoreSQLVerify-new}{\code{DataStoreSQLVerify$new()}}
\item \href{#method-DataStoreSQLVerify-validate}{\code{DataStoreSQLVerify$validate()}}
\item \href{#method-DataStoreSQLVerify-toJSON}{\code{DataStoreSQLVerify$toJSON()}}
\item \href{#method-DataStoreSQLVerify-fromJSON}{\code{DataStoreSQLVerify$fromJSON()}}
\item \href{#method-DataStoreSQLVerify-clone}{\code{DataStoreSQLVerify$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreSQLVerify-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreSQLVerify-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreSQLVerify$new(
  maxRows = NULL,
  query = NULL,
  credentialId = NULL,
  password = NULL,
  useKerberos = FALSE,
  user = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maxRows}}{Maximum number of rows of data to return if successful.}

\item{\code{query}}{SQL query to verify.}

\item{\code{credentialId}}{ID of the set of credentials to use instead of username and password.}

\item{\code{password}}{Password for data store authentication.}

\item{\code{useKerberos}}{Whether to use Kerberos for data store authentication.}

\item{\code{user}}{Username for data store authentication.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreSQLVerify-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreSQLVerify-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreSQLVerify$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreSQLVerify-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreSQLVerify-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreSQLVerify$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreSQLVerify-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreSQLVerify-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreSQLVerify$fromJSON(DataStoreSQLVerifyJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DataStoreSQLVerifyJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreSQLVerify-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreSQLVerify-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreSQLVerify$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
