% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_store_sql_verify_response.R
\docType{class}
\name{DataStoreSQLVerifyResponse}
\alias{DataStoreSQLVerifyResponse}
\title{DataStoreSQLVerifyResponse}
\format{
An \code{R6Class} generator object
}
\description{
DataStoreSQLVerifyResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{columns}}{list( character ) List of columns for the returned set of records. Column order matches the order of values in a record.}

\item{\code{records}}{list( \link{array[character]} ) List of records output by the query.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataStoreSQLVerifyResponse-new}{\code{DataStoreSQLVerifyResponse$new()}}
\item \href{#method-DataStoreSQLVerifyResponse-validate}{\code{DataStoreSQLVerifyResponse$validate()}}
\item \href{#method-DataStoreSQLVerifyResponse-toJSON}{\code{DataStoreSQLVerifyResponse$toJSON()}}
\item \href{#method-DataStoreSQLVerifyResponse-fromJSON}{\code{DataStoreSQLVerifyResponse$fromJSON()}}
\item \href{#method-DataStoreSQLVerifyResponse-clone}{\code{DataStoreSQLVerifyResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreSQLVerifyResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreSQLVerifyResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreSQLVerifyResponse$new(
  columns = NULL,
  records = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{List of columns for the returned set of records. Column order matches the order of values in a record.}

\item{\code{records}}{List of records output by the query.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreSQLVerifyResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreSQLVerifyResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreSQLVerifyResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreSQLVerifyResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreSQLVerifyResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreSQLVerifyResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreSQLVerifyResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreSQLVerifyResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreSQLVerifyResponse$fromJSON(
  DataStoreSQLVerifyResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DataStoreSQLVerifyResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStoreSQLVerifyResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataStoreSQLVerifyResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreSQLVerifyResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
