% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_store_tables.R
\docType{class}
\name{DataStoreTables}
\alias{DataStoreTables}
\title{DataStoreTables}
\format{
An \code{R6Class} generator object
}
\description{
DataStoreTables Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalog}}{character \link{optional} Only show tables in this catalog.}

\item{\code{credentialId}}{character \link{optional} ID of the set of credentials to use instead of username and password.}

\item{\code{password}}{character \link{optional} Password for data store authentication.}

\item{\code{schema}}{character \link{optional} Only show tables in this schema.}

\item{\code{useKerberos}}{character \link{optional} Whether to use Kerberos for data store authentication.}

\item{\code{user}}{character \link{optional} Username for data store authentication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataStoreTables$new()}}
\item \href{#method-validate}{\code{DataStoreTables$validate()}}
\item \href{#method-toJSON}{\code{DataStoreTables$toJSON()}}
\item \href{#method-fromJSON}{\code{DataStoreTables$fromJSON()}}
\item \href{#method-clone}{\code{DataStoreTables$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreTables$new(
  catalog = NULL,
  credentialId = NULL,
  password = NULL,
  schema = NULL,
  useKerberos = FALSE,
  user = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalog}}{Only show tables in this catalog.}

\item{\code{credentialId}}{ID of the set of credentials to use instead of username and password.}

\item{\code{password}}{Password for data store authentication.}

\item{\code{schema}}{Only show tables in this schema.}

\item{\code{useKerberos}}{Whether to use Kerberos for data store authentication.}

\item{\code{user}}{Username for data store authentication.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreTables$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreTables$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreTables$fromJSON(DataStoreTablesJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DataStoreTablesJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStoreTables$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
