% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_partitioning_data_for_open_api.R
\docType{class}
\name{DatetimePartitioningDataForOpenApi}
\alias{DatetimePartitioningDataForOpenApi}
\title{DatetimePartitioningDataForOpenApi}
\format{
An \code{R6Class} generator object
}
\description{
DatetimePartitioningDataForOpenApi Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aggregationType}}{character \link{optional} For multiseries projects only. The aggregation type to apply when creating cross-series features.}

\item{\code{allowPartialHistoryTimeSeriesPredictions}}{character \link{optional} Specifies whether the time series predictions can use partial historical data.}

\item{\code{autopilotClusterList}}{list( integer ) \link{optional} A list of integers where each value will be used as the number of clusters in Autopilot model(s) for unsupervised clustering projects. Cannot be specified unless `unsupervisedMode` is true and `unsupervisedType` is set to `clustering`.}

\item{\code{autopilotDataSelectionMethod}}{character \link{optional} The Data Selection method to be used by autopilot when creating models for datetime-partitioned datasets.}

\item{\code{backtests}}{list( \link{OneOfBacktestOldMethodForOpenApiBacktestNewMethodForOpenApi} ) \link{optional} An array specifying individual backtests.}

\item{\code{calendarId}}{character \link{optional} The ID of the calendar to be used in this project.}

\item{\code{crossSeriesGroupByColumns}}{list( character ) \link{optional} For multiseries projects with cross-series features enabled only. List of columns (currently of length 1). Setting that indicates how to further split series into related groups. For example, if every series is sales of an individual product, the series group-by could be the product category with values like \"men's clothing\", \"sports equipment\", etc.}

\item{\code{datetimePartitionColumn}}{character The date column that will be used as a datetime partition column.}

\item{\code{defaultToAPriori}}{character \link{optional} Renamed to `defaultToKnownInAdvance`.}

\item{\code{defaultToDoNotDerive}}{character \link{optional} For time series projects only. Sets whether all features default to being treated as do-not-derive features, excluding them from feature derivation. Individual features can be set to a value different than the default by using the `featureSettings` parameter.}

\item{\code{defaultToKnownInAdvance}}{character \link{optional} For time series projects only. Sets whether all features default to being treated as known in advance features, which are features that are known into the future. Features marked as known in advance must be specified into the future when making predictions. The default is false, all features are not known in advance. Individual features can be set to a value different than the default using the `featureSettings` parameter. See the :ref:`Time Series Overview <time_series_overview>` for more context.}

\item{\code{differencingMethod}}{character \link{optional} For time series projects only. Used to specify which differencing method to apply if the data is stationary. For classification problems `simple` and `seasonal` are not allowed. Parameter `periodicities` must be specified if `seasonal` is chosen. Defaults to `auto`.}

\item{\code{disableHoldout}}{character \link{optional} Whether to suppress allocating a holdout fold. If `disableHoldout` is set to true, `holdoutStartDate` and `holdoutDuration` must not be set.}

\item{\code{featureDerivationWindowEnd}}{integer \link{optional} For time series projects only. How many timeUnits of the datetimePartitionColumn into the past relative to the forecast point the feature derivation window should end.}

\item{\code{featureDerivationWindowStart}}{integer \link{optional} For time series projects only. How many timeUnits of the datetimePartitionColumn into the past relative to the forecast point the feature derivation window should begin.}

\item{\code{featureSettings}}{list( \link{FeatureSetting} ) \link{optional} An array specifying per feature settings. Features can be left unspecified.}

\item{\code{forecastWindowEnd}}{integer \link{optional} For time series projects only. How many timeUnits of the datetimePartitionColumn into the future relative to the forecast point the forecast window should end.}

\item{\code{forecastWindowStart}}{integer \link{optional} For time series projects only. How many timeUnits of the datetimePartitionColumn into the future relative to the forecast point the forecast window should start.}

\item{\code{gapDuration}}{character \link{optional} The duration of the gap between holdout training and holdout scoring data. For time series projects, defaults to the duration of the gap between the end of the feature derivation window and the beginning of the forecast window. For OTV projects, defaults to a zero duration (P0Y0M0D).}

\item{\code{holdoutDuration}}{character \link{optional} The duration of holdout scoring data. When specifying `holdoutDuration`, `holdoutStartDate` must also be specified. This attribute cannot be specified when `disableHoldout` is true.}

\item{\code{holdoutEndDate}}{character \link{optional} The end date of holdout scoring data. When specifying `holdoutEndDate`, `holdoutStartDate` must also be specified. This attribute cannot be specified when `disableHoldout` is true.}

\item{\code{holdoutStartDate}}{character \link{optional} The start date of holdout scoring data. When specifying `holdoutStartDate`, one of `holdoutEndDate` or `holdoutDuration` must also be specified. This attribute cannot be specified when `disableHoldout` is true.}

\item{\code{isHoldoutModified}}{character \link{optional} A boolean value indicating whether holdout settings (start/end dates) have been modified by user.}

\item{\code{modelSplits}}{integer \link{optional} Sets the cap on the number of jobs per model used when building models to control number of jobs in the queue. Higher number of modelSplits will allow for less downsampling leading to the use of more post-processed data.}

\item{\code{multiseriesIdColumns}}{list( character ) \link{optional} May be used only with time series projects. An array of the column names identifying  the series to which each row of the dataset belongs. Currently only one multiseries ID column is supported. See the :ref:`multiseries <multiseries>` section of the time series documentation for more context.}

\item{\code{numberOfBacktests}}{integer \link{optional} The number of backtests to use. If omitted, defaults to a positive value selected by the server based on the validation and gap durations.}

\item{\code{periodicities}}{list( \link{Periodicity} ) \link{optional} A list of periodicities for time series projects only. For classification problems periodicities are not allowed. If this is provided, parameter 'differencing_method' will default to 'seasonal' if not provided or 'auto'.}

\item{\code{treatAsExponential}}{character \link{optional} For time series projects only. Used to specify whether to treat data as exponential trend and apply transformations like log-transform. For classification problems `always` is not allowed. Defaults to `auto`.}

\item{\code{unsupervisedMode}}{character \link{optional} A boolean value indicating whether an unsupervised project should be created.}

\item{\code{unsupervisedType}}{character \link{optional} The type of unsupervised project. Only valid when `unsupervisedMode` is true. If `unsupervisedMode`, defaults to `anomaly`.}

\item{\code{useCrossSeriesFeatures}}{character \link{optional} For multiseries projects only. Indicating whether to use cross-series features.}

\item{\code{useSupervisedFeatureReduction}}{character \link{optional} When true, during feature generation DataRobot runs a supervised algorithm that identifies those features with predictive impact on the target and builds feature lists using only qualifying features. Setting false can severely impact autopilot duration, especially for datasets with many features.}

\item{\code{useTimeSeries}}{character \link{optional} A boolean value indicating whether a time series project should be created instead of a regular project which uses datetime partitioning.}

\item{\code{validationDuration}}{character \link{optional} The default validation duration for all backtests. If the primary date/time feature in a time series project is irregular, you cannot set a default validation length. Instead, set each duration individually. For an OTV project setting the validation duration will always use regular partitioning. Omitting it will use irregular partitioning if the date/time feature is irregular.}

\item{\code{windowsBasisUnit}}{character \link{optional} For time series projects only. Indicates which unit is basis for feature derivation window and forecast window. Valid options are detected time unit or `ROW`. If omitted, the default value is detected time unit.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DatetimePartitioningDataForOpenApi$new()}}
\item \href{#method-validate}{\code{DatetimePartitioningDataForOpenApi$validate()}}
\item \href{#method-toJSON}{\code{DatetimePartitioningDataForOpenApi$toJSON()}}
\item \href{#method-fromJSON}{\code{DatetimePartitioningDataForOpenApi$fromJSON()}}
\item \href{#method-clone}{\code{DatetimePartitioningDataForOpenApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatetimePartitioningDataForOpenApi$new(
  datetimePartitionColumn = NULL,
  aggregationType = NULL,
  allowPartialHistoryTimeSeriesPredictions = FALSE,
  autopilotClusterList = NULL,
  autopilotDataSelectionMethod = NULL,
  backtests = NULL,
  calendarId = NULL,
  crossSeriesGroupByColumns = NULL,
  defaultToAPriori = NULL,
  defaultToDoNotDerive = NULL,
  defaultToKnownInAdvance = NULL,
  differencingMethod = NULL,
  disableHoldout = FALSE,
  featureDerivationWindowEnd = NULL,
  featureDerivationWindowStart = NULL,
  featureSettings = NULL,
  forecastWindowEnd = NULL,
  forecastWindowStart = NULL,
  gapDuration = NULL,
  holdoutDuration = NULL,
  holdoutEndDate = NULL,
  holdoutStartDate = NULL,
  isHoldoutModified = NULL,
  modelSplits = 5,
  multiseriesIdColumns = NULL,
  numberOfBacktests = NULL,
  periodicities = NULL,
  treatAsExponential = NULL,
  unsupervisedMode = FALSE,
  unsupervisedType = NULL,
  useCrossSeriesFeatures = FALSE,
  useSupervisedFeatureReduction = TRUE,
  useTimeSeries = FALSE,
  validationDuration = NULL,
  windowsBasisUnit = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datetimePartitionColumn}}{The date column that will be used as a datetime partition column.}

\item{\code{aggregationType}}{For multiseries projects only. The aggregation type to apply when creating cross-series features.}

\item{\code{allowPartialHistoryTimeSeriesPredictions}}{Specifies whether the time series predictions can use partial historical data.}

\item{\code{autopilotClusterList}}{A list of integers where each value will be used as the number of clusters in Autopilot model(s) for unsupervised clustering projects. Cannot be specified unless `unsupervisedMode` is true and `unsupervisedType` is set to `clustering`.}

\item{\code{autopilotDataSelectionMethod}}{The Data Selection method to be used by autopilot when creating models for datetime-partitioned datasets.}

\item{\code{backtests}}{An array specifying individual backtests.}

\item{\code{calendarId}}{The ID of the calendar to be used in this project.}

\item{\code{crossSeriesGroupByColumns}}{For multiseries projects with cross-series features enabled only. List of columns (currently of length 1). Setting that indicates how to further split series into related groups. For example, if every series is sales of an individual product, the series group-by could be the product category with values like \"men's clothing\", \"sports equipment\", etc.}

\item{\code{defaultToAPriori}}{Renamed to `defaultToKnownInAdvance`.}

\item{\code{defaultToDoNotDerive}}{For time series projects only. Sets whether all features default to being treated as do-not-derive features, excluding them from feature derivation. Individual features can be set to a value different than the default by using the `featureSettings` parameter.}

\item{\code{defaultToKnownInAdvance}}{For time series projects only. Sets whether all features default to being treated as known in advance features, which are features that are known into the future. Features marked as known in advance must be specified into the future when making predictions. The default is false, all features are not known in advance. Individual features can be set to a value different than the default using the `featureSettings` parameter. See the :ref:`Time Series Overview <time_series_overview>` for more context.}

\item{\code{differencingMethod}}{For time series projects only. Used to specify which differencing method to apply if the data is stationary. For classification problems `simple` and `seasonal` are not allowed. Parameter `periodicities` must be specified if `seasonal` is chosen. Defaults to `auto`.}

\item{\code{disableHoldout}}{Whether to suppress allocating a holdout fold. If `disableHoldout` is set to true, `holdoutStartDate` and `holdoutDuration` must not be set.}

\item{\code{featureDerivationWindowEnd}}{For time series projects only. How many timeUnits of the datetimePartitionColumn into the past relative to the forecast point the feature derivation window should end.}

\item{\code{featureDerivationWindowStart}}{For time series projects only. How many timeUnits of the datetimePartitionColumn into the past relative to the forecast point the feature derivation window should begin.}

\item{\code{featureSettings}}{An array specifying per feature settings. Features can be left unspecified.}

\item{\code{forecastWindowEnd}}{For time series projects only. How many timeUnits of the datetimePartitionColumn into the future relative to the forecast point the forecast window should end.}

\item{\code{forecastWindowStart}}{For time series projects only. How many timeUnits of the datetimePartitionColumn into the future relative to the forecast point the forecast window should start.}

\item{\code{gapDuration}}{The duration of the gap between holdout training and holdout scoring data. For time series projects, defaults to the duration of the gap between the end of the feature derivation window and the beginning of the forecast window. For OTV projects, defaults to a zero duration (P0Y0M0D).}

\item{\code{holdoutDuration}}{The duration of holdout scoring data. When specifying `holdoutDuration`, `holdoutStartDate` must also be specified. This attribute cannot be specified when `disableHoldout` is true.}

\item{\code{holdoutEndDate}}{The end date of holdout scoring data. When specifying `holdoutEndDate`, `holdoutStartDate` must also be specified. This attribute cannot be specified when `disableHoldout` is true.}

\item{\code{holdoutStartDate}}{The start date of holdout scoring data. When specifying `holdoutStartDate`, one of `holdoutEndDate` or `holdoutDuration` must also be specified. This attribute cannot be specified when `disableHoldout` is true.}

\item{\code{isHoldoutModified}}{A boolean value indicating whether holdout settings (start/end dates) have been modified by user.}

\item{\code{modelSplits}}{Sets the cap on the number of jobs per model used when building models to control number of jobs in the queue. Higher number of modelSplits will allow for less downsampling leading to the use of more post-processed data.}

\item{\code{multiseriesIdColumns}}{May be used only with time series projects. An array of the column names identifying  the series to which each row of the dataset belongs. Currently only one multiseries ID column is supported. See the :ref:`multiseries <multiseries>` section of the time series documentation for more context.}

\item{\code{numberOfBacktests}}{The number of backtests to use. If omitted, defaults to a positive value selected by the server based on the validation and gap durations.}

\item{\code{periodicities}}{A list of periodicities for time series projects only. For classification problems periodicities are not allowed. If this is provided, parameter 'differencing_method' will default to 'seasonal' if not provided or 'auto'.}

\item{\code{treatAsExponential}}{For time series projects only. Used to specify whether to treat data as exponential trend and apply transformations like log-transform. For classification problems `always` is not allowed. Defaults to `auto`.}

\item{\code{unsupervisedMode}}{A boolean value indicating whether an unsupervised project should be created.}

\item{\code{unsupervisedType}}{The type of unsupervised project. Only valid when `unsupervisedMode` is true. If `unsupervisedMode`, defaults to `anomaly`.}

\item{\code{useCrossSeriesFeatures}}{For multiseries projects only. Indicating whether to use cross-series features.}

\item{\code{useSupervisedFeatureReduction}}{When true, during feature generation DataRobot runs a supervised algorithm that identifies those features with predictive impact on the target and builds feature lists using only qualifying features. Setting false can severely impact autopilot duration, especially for datasets with many features.}

\item{\code{useTimeSeries}}{A boolean value indicating whether a time series project should be created instead of a regular project which uses datetime partitioning.}

\item{\code{validationDuration}}{The default validation duration for all backtests. If the primary date/time feature in a time series project is irregular, you cannot set a default validation length. Instead, set each duration individually. For an OTV project setting the validation duration will always use regular partitioning. Omitting it will use irregular partitioning if the date/time feature is irregular.}

\item{\code{windowsBasisUnit}}{For time series projects only. Indicates which unit is basis for feature derivation window and forecast window. Valid options are detected time unit or `ROW`. If omitted, the default value is detected time unit.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatetimePartitioningDataForOpenApi$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatetimePartitioningDataForOpenApi$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatetimePartitioningDataForOpenApi$fromJSON(
  DatetimePartitioningDataForOpenApiJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatetimePartitioningDataForOpenApiJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatetimePartitioningDataForOpenApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
