% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_trend_plots_create.R
\docType{class}
\name{DatetimeTrendPlotsCreate}
\alias{DatetimeTrendPlotsCreate}
\title{DatetimeTrendPlotsCreate}
\format{
An \code{R6Class} generator object
}
\description{
DatetimeTrendPlotsCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{backtest}}{\link{OneOfintegerstring} \link{optional} Compute plots for a specific backtest (use the backtest index starting from zero) or `holdout`. If not specified the first backtest (backtest index 0) will be used.}

\item{\code{forecastDistanceEnd}}{integer \link{optional} The end of forecast distance range (forecast window) to compute. If not specified, the last forecast distance for this project will be used. Forecast distance specifies the number of time steps between the predicted point and the origin point. Only available for time series supervised projects.}

\item{\code{forecastDistanceStart}}{integer \link{optional} The start of forecast distance range (forecast window) to compute. If not specified, the first forecast distance for this project will be used. Forecast distance specifies the number of time steps between the predicted point and the origin point. Only available for time series supervised projects.}

\item{\code{fullAverage}}{character \link{optional} Whether to compute an average plot for all series. Only available for time series multiseries projects.}

\item{\code{seriesIds}}{list( character ) \link{optional} Only available for time series multiseries projects. Each element should be a name of a single series in a multiseries project. It is possible to compute a maximum of 1000 series per one request. If not specified the first 1000 series in alphabetical order will be computed. It is not possible to specify `fullAverage: true` while also setting `seriesIds`. This parameter can only be specified after first 1000 series in alphabetical order are computed.}

\item{\code{source}}{character \link{optional} The source of the data for the backtest/holdout.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DatetimeTrendPlotsCreate$new()}}
\item \href{#method-validate}{\code{DatetimeTrendPlotsCreate$validate()}}
\item \href{#method-toJSON}{\code{DatetimeTrendPlotsCreate$toJSON()}}
\item \href{#method-fromJSON}{\code{DatetimeTrendPlotsCreate$fromJSON()}}
\item \href{#method-clone}{\code{DatetimeTrendPlotsCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatetimeTrendPlotsCreate$new(
  backtest = NULL,
  forecastDistanceEnd = NULL,
  forecastDistanceStart = NULL,
  fullAverage = FALSE,
  seriesIds = NULL,
  source = "validation",
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{backtest}}{Compute plots for a specific backtest (use the backtest index starting from zero) or `holdout`. If not specified the first backtest (backtest index 0) will be used.}

\item{\code{forecastDistanceEnd}}{The end of forecast distance range (forecast window) to compute. If not specified, the last forecast distance for this project will be used. Forecast distance specifies the number of time steps between the predicted point and the origin point. Only available for time series supervised projects.}

\item{\code{forecastDistanceStart}}{The start of forecast distance range (forecast window) to compute. If not specified, the first forecast distance for this project will be used. Forecast distance specifies the number of time steps between the predicted point and the origin point. Only available for time series supervised projects.}

\item{\code{fullAverage}}{Whether to compute an average plot for all series. Only available for time series multiseries projects.}

\item{\code{seriesIds}}{Only available for time series multiseries projects. Each element should be a name of a single series in a multiseries project. It is possible to compute a maximum of 1000 series per one request. If not specified the first 1000 series in alphabetical order will be computed. It is not possible to specify `fullAverage: true` while also setting `seriesIds`. This parameter can only be specified after first 1000 series in alphabetical order are computed.}

\item{\code{source}}{The source of the data for the backtest/holdout.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatetimeTrendPlotsCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatetimeTrendPlotsCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatetimeTrendPlotsCreate$fromJSON(
  DatetimeTrendPlotsCreateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatetimeTrendPlotsCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatetimeTrendPlotsCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
