% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_trend_plots_metadata_start_end_dates.R
\docType{class}
\name{DatetimeTrendPlotsMetadataStartEndDates}
\alias{DatetimeTrendPlotsMetadataStartEndDates}
\title{DatetimeTrendPlotsMetadataStartEndDates}
\format{
An \code{R6Class} generator object
}
\description{
DatetimeTrendPlotsMetadataStartEndDates Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{endDate}}{character The datetime of the end of the chart data (exclusive). Null if chart data is not computed.}

\item{\code{startDate}}{character The datetime of the start of the chart data (inclusive). Null if chart data is not computed.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatetimeTrendPlotsMetadataStartEndDates-new}{\code{DatetimeTrendPlotsMetadataStartEndDates$new()}}
\item \href{#method-DatetimeTrendPlotsMetadataStartEndDates-validate}{\code{DatetimeTrendPlotsMetadataStartEndDates$validate()}}
\item \href{#method-DatetimeTrendPlotsMetadataStartEndDates-toJSON}{\code{DatetimeTrendPlotsMetadataStartEndDates$toJSON()}}
\item \href{#method-DatetimeTrendPlotsMetadataStartEndDates-fromJSON}{\code{DatetimeTrendPlotsMetadataStartEndDates$fromJSON()}}
\item \href{#method-DatetimeTrendPlotsMetadataStartEndDates-clone}{\code{DatetimeTrendPlotsMetadataStartEndDates$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatetimeTrendPlotsMetadataStartEndDates-new"></a>}}
\if{latex}{\out{\hypertarget{method-DatetimeTrendPlotsMetadataStartEndDates-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatetimeTrendPlotsMetadataStartEndDates$new(
  endDate = NULL,
  startDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endDate}}{The datetime of the end of the chart data (exclusive). Null if chart data is not computed.}

\item{\code{startDate}}{The datetime of the start of the chart data (inclusive). Null if chart data is not computed.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatetimeTrendPlotsMetadataStartEndDates-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DatetimeTrendPlotsMetadataStartEndDates-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatetimeTrendPlotsMetadataStartEndDates$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatetimeTrendPlotsMetadataStartEndDates-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatetimeTrendPlotsMetadataStartEndDates-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatetimeTrendPlotsMetadataStartEndDates$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatetimeTrendPlotsMetadataStartEndDates-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatetimeTrendPlotsMetadataStartEndDates-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatetimeTrendPlotsMetadataStartEndDates$fromJSON(
  DatetimeTrendPlotsMetadataStartEndDatesJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatetimeTrendPlotsMetadataStartEndDatesJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatetimeTrendPlotsMetadataStartEndDates-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DatetimeTrendPlotsMetadataStartEndDates-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatetimeTrendPlotsMetadataStartEndDates$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
