% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{DeleteAnomalyAssessmentRecord}
\alias{DeleteAnomalyAssessmentRecord}
\title{DeleteAnomalyAssessmentRecord}
\usage{
DeleteAnomalyAssessmentRecord(projectId, recordId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete the anomaly assessment record.
}
\details{
Delete the anomaly assessment record.

This method invokes \verb{DELETE /projects/\{projectId\}/anomalyAssessmentRecords/\{recordId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsAnomalyAssessmentRecordsDelete in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
