% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_apiWrapper.R
\name{DeleteAutomatedDocuments}
\alias{DeleteAutomatedDocuments}
\title{DeleteAutomatedDocuments}
\usage{
DeleteAutomatedDocuments(documentId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete automated document.
}
\details{
Delete a document using its ID. Example request:  > \verb{text > DELETE https://app.datarobot.com/api/v2/automatedDocuments/5ec4ea7e41054c158c5b002f/ HTTP/1.1 > Authorization: Bearer DnwzBUSTOtKBO5Sp1hoUByG4YgZwCCw4 >}

This method invokes \verb{DELETE /automatedDocuments/\{documentId\}/} in the DataRobot Public API.
}
\seealso{
The method AutomatedDocumentsDelete in \link{DocumentationApi}, which invokes the same underlying API endpoint.
}
