% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{DeleteBatchPredictions}
\alias{DeleteBatchPredictions}
\title{DeleteBatchPredictions}
\usage{
DeleteBatchPredictions(predictionJobId, partNumber = 0, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Cancel a Batch Prediction job
}
\details{
If the job is running, it will be aborted. Then it will be removed, meaning all underlying data will be deleted and the job is removed from the list of jobs.

This method invokes \verb{DELETE /batchPredictions/\{predictionJobId\}/} in the DataRobot Public API.
}
\seealso{
The method BatchPredictionsDelete in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
