% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{DeleteCalendar}
\alias{DeleteCalendar}
\title{DeleteCalendar}
\usage{
DeleteCalendar(calendarId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete a calendar.
}
\details{
Delete a calendar. This can only be done if all projects and deployments using the calendar have been deleted.

This method invokes \verb{DELETE /calendars/\{calendarId\}/} in the DataRobot Public API.
}
\seealso{
The method CalendarsDelete in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
