% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_apiWrapper.R
\name{DeleteCredentials}
\alias{DeleteCredentials}
\title{DeleteCredentials}
\usage{
DeleteCredentials(credentialId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete the credentials set.
}
\details{
Delete the credential set matching the specified ID if it is not used by any data connection or batch prediction job.

This method invokes \verb{DELETE /credentials/\{credentialId\}/} in the DataRobot Public API.
}
\seealso{
The method CredentialsDelete in \link{CredentialsApi}, which invokes the same underlying API endpoint.
}
