% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{DeleteCustomModelsVersionsConversions}
\alias{DeleteCustomModelsVersionsConversions}
\title{DeleteCustomModelsVersionsConversions}
\usage{
DeleteCustomModelsVersionsConversions(
  customModelId,
  conversionId,
  customModelVersionId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Stop a given custom model conversion.
}
\details{
Stop a running conversion for given model and model version.

This method invokes \verb{DELETE /customModels/\{customModelId\}/versions/\{customModelVersionId\}/conversions/\{conversionId\}/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsVersionsConversionsDelete in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
