% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{DeleteCustomModelTests}
\alias{DeleteCustomModelTests}
\title{DeleteCustomModelTests}
\usage{
DeleteCustomModelTests(customModelTestId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Cancel custom model test.
}
\details{
Cancel custom inference model testing.

This method invokes \verb{DELETE /customModelTests/\{customModelTestId\}/} in the DataRobot Public API.
}
\seealso{
The method CustomModelTestsDelete in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
