% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{DeleteDatasetsFeaturelists}
\alias{DeleteDatasetsFeaturelists}
\title{DeleteDatasetsFeaturelists}
\usage{
DeleteDatasetsFeaturelists(datasetId, featurelistId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete dataset featurelist
}
\details{
Deletes the indicated featurelist of the dataset with given ID.

This method invokes \verb{DELETE /datasets/\{datasetId\}/featurelists/\{featurelistId\}/} in the DataRobot Public API.
}
\seealso{
The method DatasetsFeaturelistsDelete in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
