% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{DeleteDatasetsRefreshJobs}
\alias{DeleteDatasetsRefreshJobs}
\title{DeleteDatasetsRefreshJobs}
\usage{
DeleteDatasetsRefreshJobs(jobId, datasetId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Deletes an existing dataset refresh job
}
\details{
Deletes an existing dataset refresh job.

This method invokes \verb{DELETE /datasets/\{datasetId\}/refreshJobs/\{jobId\}/} in the DataRobot Public API.
}
\seealso{
The method DatasetsRefreshJobsDelete in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
