% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{DeleteDatasetsVersions}
\alias{DeleteDatasetsVersions}
\title{DeleteDatasetsVersions}
\usage{
DeleteDatasetsVersions(datasetId, datasetVersionId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete dataset version
}
\details{
Marks the dataset version with the given ID as deleted.

This method invokes \verb{DELETE /datasets/\{datasetId\}/versions/\{datasetVersionId\}/} in the DataRobot Public API.
}
\seealso{
The method DatasetsVersionsDelete in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
