% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{DeleteDataSource}
\alias{DeleteDataSource}
\title{DeleteDataSource}
\usage{
DeleteDataSource(dataSourceId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete the data source.
}
\details{
Delete the data source with given ID if it is not used by any dataset.

This method invokes \verb{DELETE /externalDataSources/\{dataSourceId\}/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataSourcesDelete in \link{DataConnectivityApi}, which invokes the same underlying API endpoint.
}
