% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{DeleteDataStore}
\alias{DeleteDataStore}
\title{DeleteDataStore}
\usage{
DeleteDataStore(dataStoreId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete the data store.
}
\details{
Delete the data store with given ID if it is not used by any data source.

This method invokes \verb{DELETE /externalDataStores/\{dataStoreId\}/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataStoresDelete in \link{DataConnectivityApi}, which invokes the same underlying API endpoint.
}
