% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{DeleteDeployment}
\alias{DeleteDeployment}
\title{DeleteDeployment}
\usage{
DeleteDeployment(deploymentId, ignoreManagementAgent = "false", ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete deployment
}
\details{
Delete a deployment.

This method invokes \verb{DELETE /deployments/\{deploymentId\}/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsDelete in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
