% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleted_project_count_response.R
\docType{class}
\name{DeletedProjectCountResponse}
\alias{DeletedProjectCountResponse}
\title{DeletedProjectCountResponse}
\format{
An \code{R6Class} generator object
}
\description{
DeletedProjectCountResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{deletedProjectsCount}}{integer Amount of soft-deleted projects. The value is limited by projectCountLimit}

\item{\code{projectCountLimit}}{integer Deleted projects counting limit value. Stop counting above this limit}

\item{\code{valueExceedsLimit}}{character If an actual number of soft-deleted projects exceeds counting limit}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeletedProjectCountResponse-new}{\code{DeletedProjectCountResponse$new()}}
\item \href{#method-DeletedProjectCountResponse-validate}{\code{DeletedProjectCountResponse$validate()}}
\item \href{#method-DeletedProjectCountResponse-toJSON}{\code{DeletedProjectCountResponse$toJSON()}}
\item \href{#method-DeletedProjectCountResponse-fromJSON}{\code{DeletedProjectCountResponse$fromJSON()}}
\item \href{#method-DeletedProjectCountResponse-clone}{\code{DeletedProjectCountResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeletedProjectCountResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeletedProjectCountResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeletedProjectCountResponse$new(
  deletedProjectsCount = NULL,
  projectCountLimit = NULL,
  valueExceedsLimit = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deletedProjectsCount}}{Amount of soft-deleted projects. The value is limited by projectCountLimit}

\item{\code{projectCountLimit}}{Deleted projects counting limit value. Stop counting above this limit}

\item{\code{valueExceedsLimit}}{If an actual number of soft-deleted projects exceeds counting limit}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeletedProjectCountResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DeletedProjectCountResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeletedProjectCountResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeletedProjectCountResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeletedProjectCountResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeletedProjectCountResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeletedProjectCountResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeletedProjectCountResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeletedProjectCountResponse$fromJSON(
  DeletedProjectCountResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeletedProjectCountResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeletedProjectCountResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeletedProjectCountResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeletedProjectCountResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
