% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleted_project_response.R
\docType{class}
\name{DeletedProjectResponse}
\alias{DeletedProjectResponse}
\title{DeletedProjectResponse}
\format{
An \code{R6Class} generator object
}
\description{
DeletedProjectResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{createdBy}}{\link{DeletedProjectUser}}

\item{\code{deletedBy}}{\link{DeletedProjectUser}}

\item{\code{deletionTime}}{character ISO-8601 formatted date when project was deleted}

\item{\code{fileName}}{character The name of the file uploaded for the project dataset}

\item{\code{id}}{character The ID of the project}

\item{\code{organization}}{\link{DeletedProjectOrganization}}

\item{\code{projectName}}{character The name of the project}

\item{\code{scheduledForDeletion}}{character Whether project permanent deletion has already been scheduled}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeletedProjectResponse-new}{\code{DeletedProjectResponse$new()}}
\item \href{#method-DeletedProjectResponse-validate}{\code{DeletedProjectResponse$validate()}}
\item \href{#method-DeletedProjectResponse-toJSON}{\code{DeletedProjectResponse$toJSON()}}
\item \href{#method-DeletedProjectResponse-fromJSON}{\code{DeletedProjectResponse$fromJSON()}}
\item \href{#method-DeletedProjectResponse-clone}{\code{DeletedProjectResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeletedProjectResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeletedProjectResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeletedProjectResponse$new(
  createdBy = NULL,
  deletedBy = NULL,
  deletionTime = NULL,
  fileName = NULL,
  id = NULL,
  organization = NULL,
  projectName = NULL,
  scheduledForDeletion = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deletionTime}}{ISO-8601 formatted date when project was deleted}

\item{\code{fileName}}{The name of the file uploaded for the project dataset}

\item{\code{id}}{The ID of the project}

\item{\code{projectName}}{The name of the project}

\item{\code{scheduledForDeletion}}{Whether project permanent deletion has already been scheduled}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeletedProjectResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DeletedProjectResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeletedProjectResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeletedProjectResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeletedProjectResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeletedProjectResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeletedProjectResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeletedProjectResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeletedProjectResponse$fromJSON(
  DeletedProjectResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeletedProjectResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeletedProjectResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeletedProjectResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeletedProjectResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
