% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{DeleteGroups}
\alias{DeleteGroups}
\title{DeleteGroups}
\usage{
DeleteGroups(groupId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete user group
}
\details{
Delete the user group.

This method invokes \verb{DELETE /groups/\{groupId\}/} in the DataRobot Public API.
}
\seealso{
The method GroupsDelete in \link{UserManagementApi}, which invokes the same underlying API endpoint.
}
