% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{DeleteManyCustomTasksVersionsDependencyBuild}
\alias{DeleteManyCustomTasksVersionsDependencyBuild}
\title{DeleteManyCustomTasksVersionsDependencyBuild}
\usage{
DeleteManyCustomTasksVersionsDependencyBuild(
  customTaskVersionId,
  customTaskId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Cancel dependency build.
}
\details{
Cancel the custom task version's dependency build.

This method invokes \verb{DELETE /customTasks/\{customTaskId\}/versions/\{customTaskVersionId\}/dependencyBuild/} in the DataRobot Public API.
}
\seealso{
The method CustomTasksVersionsDependencyBuildDeleteMany in \link{CustomTasksApi}, which invokes the same underlying API endpoint.
}
