% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{DeleteManyGroups}
\alias{DeleteManyGroups}
\title{DeleteManyGroups}
\usage{
DeleteManyGroups(groups, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete multiple user groups
}
\details{
Delete the user groups.

This method invokes \verb{DELETE /groups/} in the DataRobot Public API.
}
\seealso{
The method GroupsDeleteMany in \link{UserManagementApi}, which invokes the same underlying API endpoint.
}
