% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_apiWrapper.R
\name{DeleteManyUsersRateLimitUsage}
\alias{DeleteManyUsersRateLimitUsage}
\title{DeleteManyUsersRateLimitUsage}
\usage{
DeleteManyUsersRateLimitUsage(userId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Reset resource usage for all resources
}
\details{
Reset specified user's rate limit resource usage to zero for all resources. When windows roll over, all limits are automatically reset. Use this route to reset usage sooner.

This method invokes \verb{DELETE /users/\{userId\}/rateLimitUsage/} in the DataRobot Public API.
}
\seealso{
The method UsersRateLimitUsageDeleteMany in \link{AnalyticsApi}, which invokes the same underlying API endpoint.
}
