% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{DeleteModelJob}
\alias{DeleteModelJob}
\title{DeleteModelJob}
\usage{
DeleteModelJob(project, modelJobId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Cancel a modeling job.
}
\details{
Cancel a modeling job

This method invokes \verb{DELETE /projects/\{projectId\}/modelJobs/\{jobId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelJobsDelete in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
