% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{DeleteNotificationPolicies}
\alias{DeleteNotificationPolicies}
\title{DeleteNotificationPolicies}
\usage{
DeleteNotificationPolicies(policyId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete notification policy
}
\details{
Delete the notification policy.

This method invokes \verb{DELETE /notificationPolicies/\{policyId\}/} in the DataRobot Public API.
}
\seealso{
The method NotificationPoliciesDelete in \link{NotificationsApi}, which invokes the same underlying API endpoint.
}
