% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{DeletePredictionDataset}
\alias{DeletePredictionDataset}
\title{DeletePredictionDataset}
\usage{
DeletePredictionDataset(project, datasetId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete a dataset that was uploaded for prediction.
}
\details{
Delete a dataset that was uploaded for prediction.

This method invokes \verb{DELETE /projects/\{projectId\}/predictionDatasets/\{datasetId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPredictionDatasetsDelete in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
