% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{DeletePredictJob}
\alias{DeletePredictJob}
\title{DeletePredictJob}
\usage{
DeletePredictJob(project, predictJobId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Cancel a queued prediction job
}
\details{
Cancel a queued prediction job

This method invokes \verb{DELETE /projects/\{projectId\}/predictJobs/\{jobId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPredictJobsDelete in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
