% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{DeleteProject}
\alias{DeleteProject}
\title{DeleteProject}
\usage{
DeleteProject(project, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete a project
}
\details{
Delete a project

This method invokes \verb{DELETE /projects/\{projectId\}/} in the DataRobot Public API.
}
\seealso{
The method Delete in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
