% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{DeleteProjectCleanupJobs}
\alias{DeleteProjectCleanupJobs}
\title{DeleteProjectCleanupJobs}
\usage{
DeleteProjectCleanupJobs(statusId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Cancel Scheduled Project Permadelete Job
}
\details{
Stop permadelete job, if possible

This method invokes \verb{DELETE /projectCleanupJobs/\{statusId\}/} in the DataRobot Public API.
}
\seealso{
The method CleanupJobsDelete in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
