% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{DeleteProjectsSecondaryDatasetsConfigurations}
\alias{DeleteProjectsSecondaryDatasetsConfigurations}
\title{DeleteProjectsSecondaryDatasetsConfigurations}
\usage{
DeleteProjectsSecondaryDatasetsConfigurations(
  secondaryDatasetConfigId,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Soft deletes a secondary dataset configuration.
}
\details{
Soft deletes a secondary dataset configuration.

This method invokes \verb{DELETE /projects/\{projectId\}/secondaryDatasetsConfigurations/\{secondaryDatasetConfigId\}/} in the DataRobot Public API.
}
\seealso{
The method SecondaryDatasetsConfigurationsDelete in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
