% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{DeleteUseCases}
\alias{DeleteUseCases}
\title{DeleteUseCases}
\usage{
DeleteUseCases(useCaseId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete a use case
}
\details{
Delete a use case

This method invokes \verb{DELETE /useCases/\{useCaseId\}/} in the DataRobot Public API.
}
\seealso{
The method UseCasesDelete in \link{UseCaseApi}, which invokes the same underlying API endpoint.
}
