% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R,
%   R/blueprints_apiWrapper.R
\name{DeleteUserBlueprints}
\alias{DeleteUserBlueprints}
\title{DeleteUserBlueprints}
\usage{
DeleteUserBlueprints(userBlueprintId, ...)

DeleteUserBlueprints(userBlueprintId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete a user blueprint.

Delete a user blueprint.
}
\details{
Delete a user blueprint, specified by the userBlueprintId.

This method invokes \verb{DELETE /userBlueprints/\{userBlueprintId\}/} in the DataRobot Public API.

Delete a user blueprint, specified by the userBlueprintId.

This method invokes \verb{DELETE /userBlueprints/\{userBlueprintId\}/} in the DataRobot Public API.
}
\seealso{
The method UserBlueprintsDelete in \link{AiCatalogApi}, which invokes the same underlying API endpoint.

The method UserBlueprintsDelete in \link{BlueprintsApi}, which invokes the same underlying API endpoint.
}
