% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{DeleteUserCleanupJobs}
\alias{DeleteUserCleanupJobs}
\title{DeleteUserCleanupJobs}
\usage{
DeleteUserCleanupJobs(statusId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Cancel users perma-deletion.
}
\details{
Cancel permanent deletion of the selected users.

This method invokes \verb{DELETE /userCleanupJobs/\{statusId\}/} in the DataRobot Public API.
}
\seealso{
The method UserCleanupJobsDelete in \link{UserManagementApi}, which invokes the same underlying API endpoint.
}
