% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{DeleteUserCleanupSummaries}
\alias{DeleteUserCleanupSummaries}
\title{DeleteUserCleanupSummaries}
\usage{
DeleteUserCleanupSummaries(reportId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete users permanent delete report.
}
\details{
Delete users permanent delete report.

This method invokes \verb{DELETE /userCleanupSummaries/\{reportId\}/} in the DataRobot Public API.
}
\seealso{
The method UserCleanupSummariesDelete in \link{UserManagementApi}, which invokes the same underlying API endpoint.
}
