% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{DeleteUserNotifications}
\alias{DeleteUserNotifications}
\title{DeleteUserNotifications}
\usage{
DeleteUserNotifications(userNotificationId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Delete user notification
}
\details{
Delete one notification associated with the user.

This method invokes \verb{DELETE /userNotifications/\{userNotificationId\}/} in the DataRobot Public API.
}
\seealso{
The method UserNotificationsDelete in \link{NotificationsApi}, which invokes the same underlying API endpoint.
}
