% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_response.R
\docType{class}
\name{DependencyResponse}
\alias{DependencyResponse}
\title{DependencyResponse}
\format{
An \code{R6Class} generator object
}
\description{
DependencyResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{constraints}}{list( \link{DependencyConstraint} ) Constraints that should be applied to the dependency when installed.}

\item{\code{line}}{character The original line from the requirements.txt file.}

\item{\code{lineNumber}}{integer The line number the requirement was on in requirements.txt.}

\item{\code{packageName}}{character The dependency's package name.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DependencyResponse-new}{\code{DependencyResponse$new()}}
\item \href{#method-DependencyResponse-validate}{\code{DependencyResponse$validate()}}
\item \href{#method-DependencyResponse-toJSON}{\code{DependencyResponse$toJSON()}}
\item \href{#method-DependencyResponse-fromJSON}{\code{DependencyResponse$fromJSON()}}
\item \href{#method-DependencyResponse-clone}{\code{DependencyResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DependencyResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DependencyResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyResponse$new(
  constraints = NULL,
  line = NULL,
  lineNumber = NULL,
  packageName = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{constraints}}{Constraints that should be applied to the dependency when installed.}

\item{\code{line}}{The original line from the requirements.txt file.}

\item{\code{lineNumber}}{The line number the requirement was on in requirements.txt.}

\item{\code{packageName}}{The dependency's package name.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DependencyResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DependencyResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DependencyResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DependencyResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DependencyResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DependencyResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyResponse$fromJSON(DependencyResponseJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DependencyResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DependencyResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DependencyResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DependencyResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
