% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_capabilities_response.R
\docType{class}
\name{DeploymentCapabilitiesResponse}
\alias{DeploymentCapabilitiesResponse}
\title{DeploymentCapabilitiesResponse}
\format{
An \code{R6Class} generator object
}
\description{
DeploymentCapabilitiesResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{supportsAutomaticActuals}}{character \link{optional} With this feature enabled, DataRobot infers actual values from time series history data and automatically feeds them back for accuracy estimation.}

\item{\code{supportsChallengerModels}}{character If the deployment supports challenger models.}

\item{\code{supportsFeatureDriftTracking}}{character If the deployment supports feature drift tracking.}

\item{\code{supportsFeatureTypeInfoValidation}}{character \link{optional} If the deployment supports feature type information validation. This will be True for deployments created from imported model packages whose feature data contains feature type information.}

\item{\code{supportsHumilityRecommendedRules}}{character If the deployment supports calculating values for humility recommended rules.}

\item{\code{supportsHumilityRules}}{character If the deployment supports humble ai.}

\item{\code{supportsHumilityRulesDefaultCalculations}}{character If the deployment supports calculating default values for humility rules.}

\item{\code{supportsModelReplacement}}{character If the deployment supports model replacement.}

\item{\code{supportsPortablePredictionServer}}{character \link{optional} If the deployment supports portable prediction servers. this enables to downloada dedicated model package to be deployed on external prediction environments}

\item{\code{supportsPredictionIntervals}}{character if the deployment supports prediction intervals.}

\item{\code{supportsPredictionWarning}}{character If the deployment supports prediction warnings.}

\item{\code{supportsRetraining}}{character \link{optional} Whether deployment supports retraining.}

\item{\code{supportsScoringCodeDownload}}{character \link{optional} If the currently deployed model supports scoring code download.}

\item{\code{supportsSecondaryDatasets}}{character If the deployments supports secondary datasets.}

\item{\code{supportsSegmentedAnalysisDriftAndAccuracy}}{character \link{optional} If the deployment supports tracking features in training and predictions data for segmented analysis of drift and accuracy.}

\item{\code{supportsShapBasedPredictionExplanations}}{character \link{optional} If the currently deployed model supports shap based prediction explanations.}

\item{\code{supportsTargetDriftTracking}}{character If the deployment supports target drift tracking.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeploymentCapabilitiesResponse-new}{\code{DeploymentCapabilitiesResponse$new()}}
\item \href{#method-DeploymentCapabilitiesResponse-validate}{\code{DeploymentCapabilitiesResponse$validate()}}
\item \href{#method-DeploymentCapabilitiesResponse-toJSON}{\code{DeploymentCapabilitiesResponse$toJSON()}}
\item \href{#method-DeploymentCapabilitiesResponse-fromJSON}{\code{DeploymentCapabilitiesResponse$fromJSON()}}
\item \href{#method-DeploymentCapabilitiesResponse-clone}{\code{DeploymentCapabilitiesResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentCapabilitiesResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentCapabilitiesResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentCapabilitiesResponse$new(
  supportsChallengerModels = NULL,
  supportsFeatureDriftTracking = NULL,
  supportsHumilityRecommendedRules = NULL,
  supportsHumilityRules = NULL,
  supportsHumilityRulesDefaultCalculations = NULL,
  supportsModelReplacement = NULL,
  supportsPredictionIntervals = NULL,
  supportsPredictionWarning = NULL,
  supportsSecondaryDatasets = NULL,
  supportsTargetDriftTracking = NULL,
  supportsAutomaticActuals = NULL,
  supportsFeatureTypeInfoValidation = NULL,
  supportsPortablePredictionServer = NULL,
  supportsRetraining = NULL,
  supportsScoringCodeDownload = NULL,
  supportsSegmentedAnalysisDriftAndAccuracy = NULL,
  supportsShapBasedPredictionExplanations = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{supportsChallengerModels}}{If the deployment supports challenger models.}

\item{\code{supportsFeatureDriftTracking}}{If the deployment supports feature drift tracking.}

\item{\code{supportsHumilityRecommendedRules}}{If the deployment supports calculating values for humility recommended rules.}

\item{\code{supportsHumilityRules}}{If the deployment supports humble ai.}

\item{\code{supportsHumilityRulesDefaultCalculations}}{If the deployment supports calculating default values for humility rules.}

\item{\code{supportsModelReplacement}}{If the deployment supports model replacement.}

\item{\code{supportsPredictionIntervals}}{if the deployment supports prediction intervals.}

\item{\code{supportsPredictionWarning}}{If the deployment supports prediction warnings.}

\item{\code{supportsSecondaryDatasets}}{If the deployments supports secondary datasets.}

\item{\code{supportsTargetDriftTracking}}{If the deployment supports target drift tracking.}

\item{\code{supportsAutomaticActuals}}{With this feature enabled, DataRobot infers actual values from time series history data and automatically feeds them back for accuracy estimation.}

\item{\code{supportsFeatureTypeInfoValidation}}{If the deployment supports feature type information validation. This will be True for deployments created from imported model packages whose feature data contains feature type information.}

\item{\code{supportsPortablePredictionServer}}{If the deployment supports portable prediction servers. this enables to downloada dedicated model package to be deployed on external prediction environments}

\item{\code{supportsRetraining}}{Whether deployment supports retraining.}

\item{\code{supportsScoringCodeDownload}}{If the currently deployed model supports scoring code download.}

\item{\code{supportsSegmentedAnalysisDriftAndAccuracy}}{If the deployment supports tracking features in training and predictions data for segmented analysis of drift and accuracy.}

\item{\code{supportsShapBasedPredictionExplanations}}{If the currently deployed model supports shap based prediction explanations.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentCapabilitiesResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentCapabilitiesResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentCapabilitiesResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentCapabilitiesResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentCapabilitiesResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentCapabilitiesResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentCapabilitiesResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentCapabilitiesResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentCapabilitiesResponse$fromJSON(
  DeploymentCapabilitiesResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeploymentCapabilitiesResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentCapabilitiesResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentCapabilitiesResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentCapabilitiesResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
