% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_create_from_learning_model.R
\docType{class}
\name{DeploymentCreateFromLearningModel}
\alias{DeploymentCreateFromLearningModel}
\title{DeploymentCreateFromLearningModel}
\format{
An \code{R6Class} generator object
}
\description{
DeploymentCreateFromLearningModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{defaultPredictionServerId}}{character \link{optional} ID of the default prediction server for deployment. Required for DataRobot Cloud, must be omitted for Enterprise installations.}

\item{\code{description}}{character A description for the deployment.}

\item{\code{importance}}{character \link{optional} Used only when `Model Deployment Approvals Workflow` is enabled. The parameter to show how important the deployment is. If `Model Deployment Approvals Workflow` is enabled and ``materiality`` is not specified defaults to ``MODERATE``.}

\item{\code{label}}{character A human-readable name for the deployment.}

\item{\code{modelId}}{character ID of the model to be deployed.}

\item{\code{predictionThreshold}}{numeric \link{optional} Prediction threshold used for binary classification. If not specified, model default prediction threshold will be used.}

\item{\code{status}}{character \link{optional} Status of the deployment}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeploymentCreateFromLearningModel-new}{\code{DeploymentCreateFromLearningModel$new()}}
\item \href{#method-DeploymentCreateFromLearningModel-validate}{\code{DeploymentCreateFromLearningModel$validate()}}
\item \href{#method-DeploymentCreateFromLearningModel-toJSON}{\code{DeploymentCreateFromLearningModel$toJSON()}}
\item \href{#method-DeploymentCreateFromLearningModel-fromJSON}{\code{DeploymentCreateFromLearningModel$fromJSON()}}
\item \href{#method-DeploymentCreateFromLearningModel-clone}{\code{DeploymentCreateFromLearningModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentCreateFromLearningModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentCreateFromLearningModel-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentCreateFromLearningModel$new(
  description = NULL,
  label = NULL,
  modelId = NULL,
  defaultPredictionServerId = NULL,
  importance = NULL,
  predictionThreshold = NULL,
  status = "active",
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{A description for the deployment.}

\item{\code{label}}{A human-readable name for the deployment.}

\item{\code{modelId}}{ID of the model to be deployed.}

\item{\code{defaultPredictionServerId}}{ID of the default prediction server for deployment. Required for DataRobot Cloud, must be omitted for Enterprise installations.}

\item{\code{importance}}{Used only when `Model Deployment Approvals Workflow` is enabled. The parameter to show how important the deployment is. If `Model Deployment Approvals Workflow` is enabled and ``materiality`` is not specified defaults to ``MODERATE``.}

\item{\code{predictionThreshold}}{Prediction threshold used for binary classification. If not specified, model default prediction threshold will be used.}

\item{\code{status}}{Status of the deployment}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentCreateFromLearningModel-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentCreateFromLearningModel-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentCreateFromLearningModel$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentCreateFromLearningModel-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentCreateFromLearningModel-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentCreateFromLearningModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentCreateFromLearningModel-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentCreateFromLearningModel-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentCreateFromLearningModel$fromJSON(
  DeploymentCreateFromLearningModelJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeploymentCreateFromLearningModelJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentCreateFromLearningModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentCreateFromLearningModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentCreateFromLearningModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
