% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/deployment_create_from_project_recommended_model.R
\docType{class}
\name{DeploymentCreateFromProjectRecommendedModel}
\alias{DeploymentCreateFromProjectRecommendedModel}
\title{DeploymentCreateFromProjectRecommendedModel}
\format{
An \code{R6Class} generator object
}
\description{
DeploymentCreateFromProjectRecommendedModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{defaultPredictionServerId}}{character \link{optional} ID of the default prediction server for deployment. Required for DataRobot Cloud, must be omitted for Enterprise installations.}

\item{\code{description}}{character A description for the deployment.}

\item{\code{importance}}{character \link{optional} Used only when `Model Deployment Approvals Workflow` is enabled. The parameter to show how important the deployment is. If `Model Deployment Approvals Workflow` is enabled and ``materiality`` is not specified defaults to ``MODERATE``.}

\item{\code{label}}{character A human-readable name for the deployment.}

\item{\code{projectId}}{character ID of the project entry to deploy recommended model from.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DeploymentCreateFromProjectRecommendedModel$new()}}
\item \href{#method-validate}{\code{DeploymentCreateFromProjectRecommendedModel$validate()}}
\item \href{#method-toJSON}{\code{DeploymentCreateFromProjectRecommendedModel$toJSON()}}
\item \href{#method-fromJSON}{\code{DeploymentCreateFromProjectRecommendedModel$fromJSON()}}
\item \href{#method-clone}{\code{DeploymentCreateFromProjectRecommendedModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentCreateFromProjectRecommendedModel$new(
  description = NULL,
  label = NULL,
  projectId = NULL,
  defaultPredictionServerId = NULL,
  importance = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{A description for the deployment.}

\item{\code{label}}{A human-readable name for the deployment.}

\item{\code{projectId}}{ID of the project entry to deploy recommended model from.}

\item{\code{defaultPredictionServerId}}{ID of the default prediction server for deployment. Required for DataRobot Cloud, must be omitted for Enterprise installations.}

\item{\code{importance}}{Used only when `Model Deployment Approvals Workflow` is enabled. The parameter to show how important the deployment is. If `Model Deployment Approvals Workflow` is enabled and ``materiality`` is not specified defaults to ``MODERATE``.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentCreateFromProjectRecommendedModel$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentCreateFromProjectRecommendedModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentCreateFromProjectRecommendedModel$fromJSON(
  DeploymentCreateFromProjectRecommendedModelJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeploymentCreateFromProjectRecommendedModelJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentCreateFromProjectRecommendedModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
