% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_dataset_create.R
\docType{class}
\name{DeploymentDatasetCreate}
\alias{DeploymentDatasetCreate}
\title{DeploymentDatasetCreate}
\format{
An \code{R6Class} generator object
}
\description{
DeploymentDatasetCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actualValueColumn}}{character Column name that contains actual values.}

\item{\code{associationIdColumn}}{character Column name that contains unique identifiers used with a predicted rows.}

\item{\code{datasetId}}{character The ID of dataset from catalog.}

\item{\code{datasetVersionId}}{character \link{optional} Version of dataset to retrieve.}

\item{\code{keepActualsWithoutPredictions}}{character \link{optional} Indicates if actual without predictions are kept.}

\item{\code{password}}{character \link{optional} The password for database authentication.}

\item{\code{timestampColumn}}{character \link{optional} Column name that contain datetime when actual values were obtained.}

\item{\code{user}}{character \link{optional} The username for database authentication.}

\item{\code{wasActedOnColumn}}{character \link{optional} Column name that contains boolean values if any action was made based on predictions data.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeploymentDatasetCreate-new}{\code{DeploymentDatasetCreate$new()}}
\item \href{#method-DeploymentDatasetCreate-validate}{\code{DeploymentDatasetCreate$validate()}}
\item \href{#method-DeploymentDatasetCreate-toJSON}{\code{DeploymentDatasetCreate$toJSON()}}
\item \href{#method-DeploymentDatasetCreate-fromJSON}{\code{DeploymentDatasetCreate$fromJSON()}}
\item \href{#method-DeploymentDatasetCreate-clone}{\code{DeploymentDatasetCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentDatasetCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentDatasetCreate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentDatasetCreate$new(
  actualValueColumn = NULL,
  associationIdColumn = NULL,
  datasetId = NULL,
  datasetVersionId = NULL,
  keepActualsWithoutPredictions = TRUE,
  password = NULL,
  timestampColumn = NULL,
  user = NULL,
  wasActedOnColumn = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{actualValueColumn}}{Column name that contains actual values.}

\item{\code{associationIdColumn}}{Column name that contains unique identifiers used with a predicted rows.}

\item{\code{datasetId}}{The ID of dataset from catalog.}

\item{\code{datasetVersionId}}{Version of dataset to retrieve.}

\item{\code{keepActualsWithoutPredictions}}{Indicates if actual without predictions are kept.}

\item{\code{password}}{The password for database authentication.}

\item{\code{timestampColumn}}{Column name that contain datetime when actual values were obtained.}

\item{\code{user}}{The username for database authentication.}

\item{\code{wasActedOnColumn}}{Column name that contains boolean values if any action was made based on predictions data.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentDatasetCreate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentDatasetCreate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentDatasetCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentDatasetCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentDatasetCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentDatasetCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentDatasetCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentDatasetCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentDatasetCreate$fromJSON(
  DeploymentDatasetCreateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeploymentDatasetCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentDatasetCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentDatasetCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentDatasetCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
