% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_fairness_health_response.R
\docType{class}
\name{DeploymentFairnessHealthResponse}
\alias{DeploymentFairnessHealthResponse}
\title{DeploymentFairnessHealthResponse}
\format{
An \code{R6Class} generator object
}
\description{
DeploymentFairnessHealthResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{endDate}}{character End date of fairness health period.}

\item{\code{influentialFairnessMetric}}{character Fairness metric with the most failing classes}

\item{\code{messages}}{list( character ) List of fairness status explanations}

\item{\code{protectedFeaturesCount}}{integer Count of the configured protected features for the deployment.}

\item{\code{startDate}}{character Start date of fairness health period.}

\item{\code{status}}{character              Overall status of the deployment health based of calculated fairness scores.                * ``unknown`` - used when the deployment fairness status was never calculated.               * ``passing`` - used when all fairness criteria are met.               * ``warning`` - used when certain criteria are not met                               but those are not critical for the user.                               It is an intermediate status between ``passing`` and ``failing``.               * ``failing`` - used when critical criteria are not met,                               indicating that the deployment is failing fairness checks.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeploymentFairnessHealthResponse-new}{\code{DeploymentFairnessHealthResponse$new()}}
\item \href{#method-DeploymentFairnessHealthResponse-validate}{\code{DeploymentFairnessHealthResponse$validate()}}
\item \href{#method-DeploymentFairnessHealthResponse-toJSON}{\code{DeploymentFairnessHealthResponse$toJSON()}}
\item \href{#method-DeploymentFairnessHealthResponse-fromJSON}{\code{DeploymentFairnessHealthResponse$fromJSON()}}
\item \href{#method-DeploymentFairnessHealthResponse-clone}{\code{DeploymentFairnessHealthResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentFairnessHealthResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentFairnessHealthResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentFairnessHealthResponse$new(
  endDate = NULL,
  influentialFairnessMetric = NULL,
  messages = NULL,
  protectedFeaturesCount = NULL,
  startDate = NULL,
  status = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endDate}}{End date of fairness health period.}

\item{\code{influentialFairnessMetric}}{Fairness metric with the most failing classes}

\item{\code{messages}}{List of fairness status explanations}

\item{\code{protectedFeaturesCount}}{Count of the configured protected features for the deployment.}

\item{\code{startDate}}{Start date of fairness health period.}

\item{\code{status}}{Overall status of the deployment health based of calculated fairness scores.                * ``unknown`` - used when the deployment fairness status was never calculated.               * ``passing`` - used when all fairness criteria are met.               * ``warning`` - used when certain criteria are not met                               but those are not critical for the user.                               It is an intermediate status between ``passing`` and ``failing``.               * ``failing`` - used when critical criteria are not met,                               indicating that the deployment is failing fairness checks.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentFairnessHealthResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentFairnessHealthResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentFairnessHealthResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentFairnessHealthResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentFairnessHealthResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentFairnessHealthResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentFairnessHealthResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentFairnessHealthResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentFairnessHealthResponse$fromJSON(
  DeploymentFairnessHealthResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeploymentFairnessHealthResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentFairnessHealthResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentFairnessHealthResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentFairnessHealthResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
