% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_feature_drift_object.R
\docType{class}
\name{DeploymentFeatureDriftObject}
\alias{DeploymentFeatureDriftObject}
\title{DeploymentFeatureDriftObject}
\format{
An \code{R6Class} generator object
}
\description{
DeploymentFeatureDriftObject Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{baselineSampleSize}}{integer The sample size of the training data.}

\item{\code{driftScore}}{numeric The drift score for this feature.}

\item{\code{featureImpact}}{numeric The feature impact score for this feature.}

\item{\code{name}}{character The name of the feature.}

\item{\code{sampleSize}}{integer The number of predictions used to compute the drift score.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DeploymentFeatureDriftObject$new()}}
\item \href{#method-validate}{\code{DeploymentFeatureDriftObject$validate()}}
\item \href{#method-toJSON}{\code{DeploymentFeatureDriftObject$toJSON()}}
\item \href{#method-fromJSON}{\code{DeploymentFeatureDriftObject$fromJSON()}}
\item \href{#method-clone}{\code{DeploymentFeatureDriftObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentFeatureDriftObject$new(
  baselineSampleSize = NULL,
  driftScore = NULL,
  featureImpact = NULL,
  name = NULL,
  sampleSize = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{baselineSampleSize}}{The sample size of the training data.}

\item{\code{driftScore}}{The drift score for this feature.}

\item{\code{featureImpact}}{The feature impact score for this feature.}

\item{\code{name}}{The name of the feature.}

\item{\code{sampleSize}}{The number of predictions used to compute the drift score.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentFeatureDriftObject$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentFeatureDriftObject$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentFeatureDriftObject$fromJSON(
  DeploymentFeatureDriftObjectJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeploymentFeatureDriftObjectJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentFeatureDriftObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
