% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_feature_drift_response.R
\docType{class}
\name{DeploymentFeatureDriftResponse}
\alias{DeploymentFeatureDriftResponse}
\title{DeploymentFeatureDriftResponse}
\format{
An \code{R6Class} generator object
}
\description{
DeploymentFeatureDriftResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer \link{optional} The number of items returned on this page.}

\item{\code{data}}{list( \link{DeploymentFeatureDriftObject} ) An array \link{DriftObject}, each in the form described below}

\item{\code{metric}}{character \link{optional} Metric used to calculate drift score.}

\item{\code{modelId}}{character The id of the model for which the features drift is being retrieved.}

\item{\code{next_}}{character \link{optional} A URL pointing to the next page (if null, there is no next page)}

\item{\code{period}}{\link{TimeRange}}

\item{\code{previous}}{character \link{optional} A URL pointing to the previous page (if null, there is no previous page)}

\item{\code{segmentAttribute}}{character \link{optional} The name of the segment on which segment analysis is being performed.}

\item{\code{segmentValue}}{character \link{optional} The value of the `segmentAttribute` to segment on.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeploymentFeatureDriftResponse-new}{\code{DeploymentFeatureDriftResponse$new()}}
\item \href{#method-DeploymentFeatureDriftResponse-validate}{\code{DeploymentFeatureDriftResponse$validate()}}
\item \href{#method-DeploymentFeatureDriftResponse-toJSON}{\code{DeploymentFeatureDriftResponse$toJSON()}}
\item \href{#method-DeploymentFeatureDriftResponse-fromJSON}{\code{DeploymentFeatureDriftResponse$fromJSON()}}
\item \href{#method-DeploymentFeatureDriftResponse-clone}{\code{DeploymentFeatureDriftResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentFeatureDriftResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentFeatureDriftResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentFeatureDriftResponse$new(
  data = NULL,
  modelId = NULL,
  period = NULL,
  count = NULL,
  metric = NULL,
  next_ = NULL,
  previous = NULL,
  segmentAttribute = NULL,
  segmentValue = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{An array \link{DriftObject}, each in the form described below}

\item{\code{modelId}}{The id of the model for which the features drift is being retrieved.}

\item{\code{count}}{The number of items returned on this page.}

\item{\code{metric}}{Metric used to calculate drift score.}

\item{\code{next_}}{A URL pointing to the next page (if null, there is no next page)}

\item{\code{previous}}{A URL pointing to the previous page (if null, there is no previous page)}

\item{\code{segmentAttribute}}{The name of the segment on which segment analysis is being performed.}

\item{\code{segmentValue}}{The value of the `segmentAttribute` to segment on.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentFeatureDriftResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentFeatureDriftResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentFeatureDriftResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentFeatureDriftResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentFeatureDriftResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentFeatureDriftResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentFeatureDriftResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentFeatureDriftResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentFeatureDriftResponse$fromJSON(
  DeploymentFeatureDriftResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeploymentFeatureDriftResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentFeatureDriftResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentFeatureDriftResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentFeatureDriftResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
