% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_governance_response.R
\docType{class}
\name{DeploymentGovernanceResponse}
\alias{DeploymentGovernanceResponse}
\title{DeploymentGovernanceResponse}
\format{
An \code{R6Class} generator object
}
\description{
DeploymentGovernanceResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{approvalStatus}}{character Status to show whether the deployment was approved or not. It also shows up as a part of metadata within the prediction response.}

\item{\code{hasOpenedChangeRequests}}{character If there are change request related to this deployment with `PENDING` status.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeploymentGovernanceResponse-new}{\code{DeploymentGovernanceResponse$new()}}
\item \href{#method-DeploymentGovernanceResponse-validate}{\code{DeploymentGovernanceResponse$validate()}}
\item \href{#method-DeploymentGovernanceResponse-toJSON}{\code{DeploymentGovernanceResponse$toJSON()}}
\item \href{#method-DeploymentGovernanceResponse-fromJSON}{\code{DeploymentGovernanceResponse$fromJSON()}}
\item \href{#method-DeploymentGovernanceResponse-clone}{\code{DeploymentGovernanceResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentGovernanceResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentGovernanceResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentGovernanceResponse$new(
  approvalStatus = NULL,
  hasOpenedChangeRequests = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{approvalStatus}}{Status to show whether the deployment was approved or not. It also shows up as a part of metadata within the prediction response.}

\item{\code{hasOpenedChangeRequests}}{If there are change request related to this deployment with `PENDING` status.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentGovernanceResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentGovernanceResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentGovernanceResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentGovernanceResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentGovernanceResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentGovernanceResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentGovernanceResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentGovernanceResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentGovernanceResponse$fromJSON(
  DeploymentGovernanceResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeploymentGovernanceResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentGovernanceResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentGovernanceResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentGovernanceResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
