% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_model_custom_model_image_response.R
\docType{class}
\name{DeploymentModelCustomModelImageResponse}
\alias{DeploymentModelCustomModelImageResponse}
\title{DeploymentModelCustomModelImageResponse}
\format{
An \code{R6Class} generator object
}
\description{
DeploymentModelCustomModelImageResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{customModelId}}{character ID of the custom model.}

\item{\code{customModelName}}{character Name of the custom model.}

\item{\code{customModelVersionId}}{character ID of the custom model version.}

\item{\code{customModelVersionLabel}}{character Label of the custom model version.}

\item{\code{executionEnvironmentId}}{character ID of the execution environment.}

\item{\code{executionEnvironmentName}}{character Name of the execution environment.}

\item{\code{executionEnvironmentVersionId}}{character ID of the execution environment version.}

\item{\code{executionEnvironmentVersionLabel}}{character Label of the execution environment version.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DeploymentModelCustomModelImageResponse$new()}}
\item \href{#method-validate}{\code{DeploymentModelCustomModelImageResponse$validate()}}
\item \href{#method-toJSON}{\code{DeploymentModelCustomModelImageResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{DeploymentModelCustomModelImageResponse$fromJSON()}}
\item \href{#method-clone}{\code{DeploymentModelCustomModelImageResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentModelCustomModelImageResponse$new(
  customModelId = NULL,
  customModelName = NULL,
  customModelVersionId = NULL,
  customModelVersionLabel = NULL,
  executionEnvironmentId = NULL,
  executionEnvironmentName = NULL,
  executionEnvironmentVersionId = NULL,
  executionEnvironmentVersionLabel = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{customModelId}}{ID of the custom model.}

\item{\code{customModelName}}{Name of the custom model.}

\item{\code{customModelVersionId}}{ID of the custom model version.}

\item{\code{customModelVersionLabel}}{Label of the custom model version.}

\item{\code{executionEnvironmentId}}{ID of the execution environment.}

\item{\code{executionEnvironmentName}}{Name of the execution environment.}

\item{\code{executionEnvironmentVersionId}}{ID of the execution environment version.}

\item{\code{executionEnvironmentVersionLabel}}{Label of the execution environment version.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentModelCustomModelImageResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentModelCustomModelImageResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentModelCustomModelImageResponse$fromJSON(
  DeploymentModelCustomModelImageResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeploymentModelCustomModelImageResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentModelCustomModelImageResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
