% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_model_health_response.R
\docType{class}
\name{DeploymentModelHealthResponse}
\alias{DeploymentModelHealthResponse}
\title{DeploymentModelHealthResponse}
\format{
An \code{R6Class} generator object
}
\description{
DeploymentModelHealthResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{endDate}}{character End date of model health period.}

\item{\code{message}}{character A message providing more detail on the status.}

\item{\code{startDate}}{character Start date of model health period.}

\item{\code{status}}{character Model health status.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeploymentModelHealthResponse-new}{\code{DeploymentModelHealthResponse$new()}}
\item \href{#method-DeploymentModelHealthResponse-validate}{\code{DeploymentModelHealthResponse$validate()}}
\item \href{#method-DeploymentModelHealthResponse-toJSON}{\code{DeploymentModelHealthResponse$toJSON()}}
\item \href{#method-DeploymentModelHealthResponse-fromJSON}{\code{DeploymentModelHealthResponse$fromJSON()}}
\item \href{#method-DeploymentModelHealthResponse-clone}{\code{DeploymentModelHealthResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentModelHealthResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentModelHealthResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentModelHealthResponse$new(
  endDate = NULL,
  message = NULL,
  startDate = NULL,
  status = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endDate}}{End date of model health period.}

\item{\code{message}}{A message providing more detail on the status.}

\item{\code{startDate}}{Start date of model health period.}

\item{\code{status}}{Model health status.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentModelHealthResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentModelHealthResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentModelHealthResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentModelHealthResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentModelHealthResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentModelHealthResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentModelHealthResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentModelHealthResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentModelHealthResponse$fromJSON(
  DeploymentModelHealthResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeploymentModelHealthResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentModelHealthResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentModelHealthResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentModelHealthResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
