% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_model_response.R
\docType{class}
\name{DeploymentModelResponse}
\alias{DeploymentModelResponse}
\title{DeploymentModelResponse}
\format{
An \code{R6Class} generator object
}
\description{
DeploymentModelResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{buildEnvironmentType}}{character \link{optional} Build environment type of the current model.}

\item{\code{customModelImage}}{\link{DeploymentModelCustomModelImageResponse} \link{optional}}

\item{\code{decisionFlowId}}{character \link{optional} The ID of the decision flow if the deployment has one.}

\item{\code{decisionFlowVersionId}}{character \link{optional} The ID of the decision flow version if the deployment has one.}

\item{\code{deployedAt}}{character \link{optional} Timestamp of when current model was deployed.}

\item{\code{hasAutomodel}}{character \link{optional} Whether the deployment's current model has an associated automodel.}

\item{\code{hasDecisionFlow}}{character \link{optional} True if this deployment contains a Decision Flow}

\item{\code{id}}{character ID of the current model.}

\item{\code{projectId}}{character \link{optional} Project ID of the current model.}

\item{\code{projectName}}{character Project name of the current model.}

\item{\code{targetName}}{character Target name of the current model.}

\item{\code{targetType}}{character Target type of the current model.}

\item{\code{type}}{character Type of the current model.}

\item{\code{unsupervisedMode}}{character Whether the current model's project is unsupervised.}

\item{\code{unsupervisedType}}{character \link{optional} Only valid when unsupervisedMode is True. The type of unsupervised project, anomaly or clustering. If unsupervisedMode, defaults to 'anomaly'.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DeploymentModelResponse$new()}}
\item \href{#method-validate}{\code{DeploymentModelResponse$validate()}}
\item \href{#method-toJSON}{\code{DeploymentModelResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{DeploymentModelResponse$fromJSON()}}
\item \href{#method-clone}{\code{DeploymentModelResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentModelResponse$new(
  id = NULL,
  projectName = NULL,
  targetName = NULL,
  targetType = NULL,
  type = NULL,
  unsupervisedMode = NULL,
  buildEnvironmentType = NULL,
  customModelImage = NULL,
  decisionFlowId = NULL,
  decisionFlowVersionId = NULL,
  deployedAt = NULL,
  hasAutomodel = NULL,
  hasDecisionFlow = NULL,
  projectId = NULL,
  unsupervisedType = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{ID of the current model.}

\item{\code{projectName}}{Project name of the current model.}

\item{\code{targetName}}{Target name of the current model.}

\item{\code{targetType}}{Target type of the current model.}

\item{\code{type}}{Type of the current model.}

\item{\code{unsupervisedMode}}{Whether the current model's project is unsupervised.}

\item{\code{buildEnvironmentType}}{Build environment type of the current model.}

\item{\code{decisionFlowId}}{The ID of the decision flow if the deployment has one.}

\item{\code{decisionFlowVersionId}}{The ID of the decision flow version if the deployment has one.}

\item{\code{deployedAt}}{Timestamp of when current model was deployed.}

\item{\code{hasAutomodel}}{Whether the deployment's current model has an associated automodel.}

\item{\code{hasDecisionFlow}}{True if this deployment contains a Decision Flow}

\item{\code{projectId}}{Project ID of the current model.}

\item{\code{unsupervisedType}}{Only valid when unsupervisedMode is True. The type of unsupervised project, anomaly or clustering. If unsupervisedMode, defaults to 'anomaly'.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentModelResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentModelResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentModelResponse$fromJSON(
  DeploymentModelResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeploymentModelResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentModelResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
