% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_owner_response.R
\docType{class}
\name{DeploymentOwnerResponse}
\alias{DeploymentOwnerResponse}
\title{DeploymentOwnerResponse}
\format{
An \code{R6Class} generator object
}
\description{
DeploymentOwnerResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{email}}{character Email address of the owner.}

\item{\code{firstName}}{character First name of the owner.}

\item{\code{id}}{character ID of the owner.}

\item{\code{lastName}}{character Last name of the owner.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeploymentOwnerResponse-new}{\code{DeploymentOwnerResponse$new()}}
\item \href{#method-DeploymentOwnerResponse-validate}{\code{DeploymentOwnerResponse$validate()}}
\item \href{#method-DeploymentOwnerResponse-toJSON}{\code{DeploymentOwnerResponse$toJSON()}}
\item \href{#method-DeploymentOwnerResponse-fromJSON}{\code{DeploymentOwnerResponse$fromJSON()}}
\item \href{#method-DeploymentOwnerResponse-clone}{\code{DeploymentOwnerResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentOwnerResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentOwnerResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentOwnerResponse$new(
  email = NULL,
  firstName = NULL,
  id = NULL,
  lastName = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{Email address of the owner.}

\item{\code{firstName}}{First name of the owner.}

\item{\code{id}}{ID of the owner.}

\item{\code{lastName}}{Last name of the owner.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentOwnerResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentOwnerResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentOwnerResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentOwnerResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentOwnerResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentOwnerResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentOwnerResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentOwnerResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentOwnerResponse$fromJSON(
  DeploymentOwnerResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeploymentOwnerResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentOwnerResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentOwnerResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentOwnerResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
