% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_prediction_environment_response.R
\docType{class}
\name{DeploymentPredictionEnvironmentResponse}
\alias{DeploymentPredictionEnvironmentResponse}
\title{DeploymentPredictionEnvironmentResponse}
\format{
An \code{R6Class} generator object
}
\description{
DeploymentPredictionEnvironmentResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character ID of the PredictionEnvironment.}

\item{\code{isManagedByManagementAgent}}{character True if PredictionEnvironment is using Management Agent.}

\item{\code{name}}{character Name of the PredictionEnvironment.}

\item{\code{platform}}{character Platform of the PredictionEnvironment.}

\item{\code{plugin}}{character \link{optional} Plugin name of the PredictionEnvironment.}

\item{\code{supportedModelFormats}}{list( character ) \link{optional} Model formats that the PredictionEnvironment supports.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DeploymentPredictionEnvironmentResponse$new()}}
\item \href{#method-validate}{\code{DeploymentPredictionEnvironmentResponse$validate()}}
\item \href{#method-toJSON}{\code{DeploymentPredictionEnvironmentResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{DeploymentPredictionEnvironmentResponse$fromJSON()}}
\item \href{#method-clone}{\code{DeploymentPredictionEnvironmentResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentPredictionEnvironmentResponse$new(
  id = NULL,
  isManagedByManagementAgent = NULL,
  name = NULL,
  platform = NULL,
  plugin = NULL,
  supportedModelFormats = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{ID of the PredictionEnvironment.}

\item{\code{isManagedByManagementAgent}}{True if PredictionEnvironment is using Management Agent.}

\item{\code{name}}{Name of the PredictionEnvironment.}

\item{\code{platform}}{Platform of the PredictionEnvironment.}

\item{\code{plugin}}{Plugin name of the PredictionEnvironment.}

\item{\code{supportedModelFormats}}{Model formats that the PredictionEnvironment supports.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentPredictionEnvironmentResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentPredictionEnvironmentResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentPredictionEnvironmentResponse$fromJSON(
  DeploymentPredictionEnvironmentResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeploymentPredictionEnvironmentResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentPredictionEnvironmentResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
