% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_retrieve_response.R
\docType{class}
\name{DeploymentRetrieveResponse}
\alias{DeploymentRetrieveResponse}
\title{DeploymentRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
DeploymentRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{accuracyHealth}}{\link{DeploymentAccuracyHealthResponse}}

\item{\code{approvalStatus}}{character \link{optional} Status to show whether the deployment was approved or not. It also shows up as a part of metadata within the prediction response.}

\item{\code{capabilities}}{\link{DeploymentCapabilitiesResponse}}

\item{\code{createdAt}}{character The date and time of when the deployment was created, in ISO 8601 format.}

\item{\code{defaultPredictionServer}}{\link{DeploymentDefaultPredictionServerResponse}}

\item{\code{description}}{character Description of the deployment.}

\item{\code{governance}}{\link{DeploymentGovernanceResponse} \link{optional}}

\item{\code{id}}{character ID of the deployment.}

\item{\code{importance}}{character \link{optional} Shows how important this deployment is.}

\item{\code{label}}{character Label of the deployment.}

\item{\code{model}}{\link{DeploymentModelResponse}}

\item{\code{modelHealth}}{\link{DeploymentModelHealthResponse}}

\item{\code{modelPackage}}{\link{DeploymentModelPackageResponse} \link{optional}}

\item{\code{modelPackageInitialDownload}}{\link{DeploymentModelPackageInitialDownloadResponse} \link{optional}}

\item{\code{openedChangeRequests}}{list( character ) \link{optional} An array of the change request IDs related to this deployment that have.}

\item{\code{owners}}{\link{DeploymentOwnersResponse} \link{optional}}

\item{\code{permissions}}{list( character ) Permissions that the user making the request has on the deployment.}

\item{\code{predictionEnvironment}}{\link{DeploymentPredictionEnvironmentResponse} \link{optional}}

\item{\code{predictionUsage}}{\link{DeploymentPredictionUsageResponse}}

\item{\code{scoringCodeInitialDownload}}{\link{DeploymentScoringCodeInitialDownloadResponse} \link{optional}}

\item{\code{serviceHealth}}{\link{DeploymentServiceHealthResponse}}

\item{\code{settings}}{\link{DeploymentSettingsInfo}}

\item{\code{status}}{character Displays current deployment status.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DeploymentRetrieveResponse$new()}}
\item \href{#method-validate}{\code{DeploymentRetrieveResponse$validate()}}
\item \href{#method-toJSON}{\code{DeploymentRetrieveResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{DeploymentRetrieveResponse$fromJSON()}}
\item \href{#method-clone}{\code{DeploymentRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentRetrieveResponse$new(
  accuracyHealth = NULL,
  capabilities = NULL,
  createdAt = NULL,
  defaultPredictionServer = NULL,
  description = NULL,
  id = NULL,
  label = NULL,
  model = NULL,
  modelHealth = NULL,
  permissions = NULL,
  predictionUsage = NULL,
  serviceHealth = NULL,
  settings = NULL,
  status = NULL,
  approvalStatus = NULL,
  governance = NULL,
  importance = NULL,
  modelPackage = NULL,
  modelPackageInitialDownload = NULL,
  openedChangeRequests = NULL,
  owners = NULL,
  predictionEnvironment = NULL,
  scoringCodeInitialDownload = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{createdAt}}{The date and time of when the deployment was created, in ISO 8601 format.}

\item{\code{description}}{Description of the deployment.}

\item{\code{id}}{ID of the deployment.}

\item{\code{label}}{Label of the deployment.}

\item{\code{permissions}}{Permissions that the user making the request has on the deployment.}

\item{\code{status}}{Displays current deployment status.}

\item{\code{approvalStatus}}{Status to show whether the deployment was approved or not. It also shows up as a part of metadata within the prediction response.}

\item{\code{importance}}{Shows how important this deployment is.}

\item{\code{openedChangeRequests}}{An array of the change request IDs related to this deployment that have.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentRetrieveResponse$fromJSON(
  DeploymentRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeploymentRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
