% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_secondary_dataset.R
\docType{class}
\name{DeploymentSecondaryDataset}
\alias{DeploymentSecondaryDataset}
\title{DeploymentSecondaryDataset}
\format{
An \code{R6Class} generator object
}
\description{
DeploymentSecondaryDataset Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalogId}}{character ID of the catalog item}

\item{\code{catalogVersionId}}{character ID of the catalog version}

\item{\code{identifier}}{character Short name of this table (used directly as part of generated feature names)}

\item{\code{snapshotPolicy}}{character Policy to use by a dataset while making prediction}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeploymentSecondaryDataset-new}{\code{DeploymentSecondaryDataset$new()}}
\item \href{#method-DeploymentSecondaryDataset-validate}{\code{DeploymentSecondaryDataset$validate()}}
\item \href{#method-DeploymentSecondaryDataset-toJSON}{\code{DeploymentSecondaryDataset$toJSON()}}
\item \href{#method-DeploymentSecondaryDataset-fromJSON}{\code{DeploymentSecondaryDataset$fromJSON()}}
\item \href{#method-DeploymentSecondaryDataset-clone}{\code{DeploymentSecondaryDataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentSecondaryDataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentSecondaryDataset-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentSecondaryDataset$new(
  catalogId = NULL,
  catalogVersionId = NULL,
  identifier = NULL,
  snapshotPolicy = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalogId}}{ID of the catalog item}

\item{\code{catalogVersionId}}{ID of the catalog version}

\item{\code{identifier}}{Short name of this table (used directly as part of generated feature names)}

\item{\code{snapshotPolicy}}{Policy to use by a dataset while making prediction}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentSecondaryDataset-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentSecondaryDataset-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentSecondaryDataset$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentSecondaryDataset-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentSecondaryDataset-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentSecondaryDataset$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentSecondaryDataset-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentSecondaryDataset-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentSecondaryDataset$fromJSON(
  DeploymentSecondaryDatasetJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeploymentSecondaryDatasetJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentSecondaryDataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentSecondaryDataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentSecondaryDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
