% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_service_health_response.R
\docType{class}
\name{DeploymentServiceHealthResponse}
\alias{DeploymentServiceHealthResponse}
\title{DeploymentServiceHealthResponse}
\format{
An \code{R6Class} generator object
}
\description{
DeploymentServiceHealthResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{endDate}}{character End date of model service period.}

\item{\code{startDate}}{character Start date of service health period.}

\item{\code{status}}{character Service health status.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeploymentServiceHealthResponse-new}{\code{DeploymentServiceHealthResponse$new()}}
\item \href{#method-DeploymentServiceHealthResponse-validate}{\code{DeploymentServiceHealthResponse$validate()}}
\item \href{#method-DeploymentServiceHealthResponse-toJSON}{\code{DeploymentServiceHealthResponse$toJSON()}}
\item \href{#method-DeploymentServiceHealthResponse-fromJSON}{\code{DeploymentServiceHealthResponse$fromJSON()}}
\item \href{#method-DeploymentServiceHealthResponse-clone}{\code{DeploymentServiceHealthResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentServiceHealthResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentServiceHealthResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentServiceHealthResponse$new(
  endDate = NULL,
  startDate = NULL,
  status = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endDate}}{End date of model service period.}

\item{\code{startDate}}{Start date of service health period.}

\item{\code{status}}{Service health status.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentServiceHealthResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentServiceHealthResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentServiceHealthResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentServiceHealthResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentServiceHealthResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentServiceHealthResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentServiceHealthResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentServiceHealthResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentServiceHealthResponse$fromJSON(
  DeploymentServiceHealthResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeploymentServiceHealthResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentServiceHealthResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentServiceHealthResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentServiceHealthResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
