% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_scoring_code_build_payload.R
\docType{class}
\name{DeploymentsScoringCodeBuildPayload}
\alias{DeploymentsScoringCodeBuildPayload}
\title{DeploymentsScoringCodeBuildPayload}
\format{
An \code{R6Class} generator object
}
\description{
DeploymentsScoringCodeBuildPayload Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{includeAgent}}{character \link{optional} Whether the Scoring Code built will include tracking agent}

\item{\code{includePredictionExplanations}}{character \link{optional} Whether the Scoring Code built will include prediction explanations}

\item{\code{includePredictionIntervals}}{character \link{optional} Whether the Scoring Code built will include prediction intervals}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeploymentsScoringCodeBuildPayload-new}{\code{DeploymentsScoringCodeBuildPayload$new()}}
\item \href{#method-DeploymentsScoringCodeBuildPayload-validate}{\code{DeploymentsScoringCodeBuildPayload$validate()}}
\item \href{#method-DeploymentsScoringCodeBuildPayload-toJSON}{\code{DeploymentsScoringCodeBuildPayload$toJSON()}}
\item \href{#method-DeploymentsScoringCodeBuildPayload-fromJSON}{\code{DeploymentsScoringCodeBuildPayload$fromJSON()}}
\item \href{#method-DeploymentsScoringCodeBuildPayload-clone}{\code{DeploymentsScoringCodeBuildPayload$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentsScoringCodeBuildPayload-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentsScoringCodeBuildPayload-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentsScoringCodeBuildPayload$new(
  includeAgent = NULL,
  includePredictionExplanations = NULL,
  includePredictionIntervals = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{includeAgent}}{Whether the Scoring Code built will include tracking agent}

\item{\code{includePredictionExplanations}}{Whether the Scoring Code built will include prediction explanations}

\item{\code{includePredictionIntervals}}{Whether the Scoring Code built will include prediction intervals}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentsScoringCodeBuildPayload-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentsScoringCodeBuildPayload-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentsScoringCodeBuildPayload$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentsScoringCodeBuildPayload-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentsScoringCodeBuildPayload-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentsScoringCodeBuildPayload$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentsScoringCodeBuildPayload-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentsScoringCodeBuildPayload-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentsScoringCodeBuildPayload$fromJSON(
  DeploymentsScoringCodeBuildPayloadJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeploymentsScoringCodeBuildPayloadJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentsScoringCodeBuildPayload-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentsScoringCodeBuildPayload-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentsScoringCodeBuildPayload$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
