% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_target_drift_response.R
\docType{class}
\name{DeploymentTargetDriftResponse}
\alias{DeploymentTargetDriftResponse}
\title{DeploymentTargetDriftResponse}
\format{
An \code{R6Class} generator object
}
\description{
DeploymentTargetDriftResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{baselineSampleSize}}{integer sample size of the training data.}

\item{\code{driftScore}}{numeric drift score for the target.}

\item{\code{metric}}{character \link{optional} Metric used to calculate drift score.}

\item{\code{modelId}}{character id of the model for which data drift is being retrieved.}

\item{\code{period}}{\link{TimeRange}}

\item{\code{sampleSize}}{integer number of predictions used to compute the drift score.}

\item{\code{segmentAttribute}}{character \link{optional} The name of the segment on which segment analysis is being performed.}

\item{\code{segmentValue}}{character \link{optional} The value of the `segmentAttribute` to segment on.}

\item{\code{targetName}}{character name of the target feature.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DeploymentTargetDriftResponse$new()}}
\item \href{#method-validate}{\code{DeploymentTargetDriftResponse$validate()}}
\item \href{#method-toJSON}{\code{DeploymentTargetDriftResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{DeploymentTargetDriftResponse$fromJSON()}}
\item \href{#method-clone}{\code{DeploymentTargetDriftResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentTargetDriftResponse$new(
  baselineSampleSize = NULL,
  driftScore = NULL,
  modelId = NULL,
  period = NULL,
  sampleSize = NULL,
  targetName = NULL,
  metric = NULL,
  segmentAttribute = NULL,
  segmentValue = "",
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{baselineSampleSize}}{sample size of the training data.}

\item{\code{driftScore}}{drift score for the target.}

\item{\code{modelId}}{id of the model for which data drift is being retrieved.}

\item{\code{sampleSize}}{number of predictions used to compute the drift score.}

\item{\code{targetName}}{name of the target feature.}

\item{\code{metric}}{Metric used to calculate drift score.}

\item{\code{segmentAttribute}}{The name of the segment on which segment analysis is being performed.}

\item{\code{segmentValue}}{The value of the `segmentAttribute` to segment on.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentTargetDriftResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentTargetDriftResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentTargetDriftResponse$fromJSON(
  DeploymentTargetDriftResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeploymentTargetDriftResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentTargetDriftResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
