% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_training_data_assignment.R
\docType{class}
\name{DeprecatedTrainingDataAssignment}
\alias{DeprecatedTrainingDataAssignment}
\title{DeprecatedTrainingDataAssignment}
\format{
An \code{R6Class} generator object
}
\description{
DeprecatedTrainingDataAssignment Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datasetId}}{character \link{optional} The ID of the dataset.}

\item{\code{datasetName}}{character \link{optional} A user-friendly name for the dataset.}

\item{\code{datasetVersionId}}{character \link{optional} The ID of the dataset version.}

\item{\code{partitionColumn}}{character \link{optional} The name of the column containing the partition assignments.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeprecatedTrainingDataAssignment-new}{\code{DeprecatedTrainingDataAssignment$new()}}
\item \href{#method-DeprecatedTrainingDataAssignment-validate}{\code{DeprecatedTrainingDataAssignment$validate()}}
\item \href{#method-DeprecatedTrainingDataAssignment-toJSON}{\code{DeprecatedTrainingDataAssignment$toJSON()}}
\item \href{#method-DeprecatedTrainingDataAssignment-fromJSON}{\code{DeprecatedTrainingDataAssignment$fromJSON()}}
\item \href{#method-DeprecatedTrainingDataAssignment-clone}{\code{DeprecatedTrainingDataAssignment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeprecatedTrainingDataAssignment-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeprecatedTrainingDataAssignment-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeprecatedTrainingDataAssignment$new(
  datasetId = NULL,
  datasetName = NULL,
  datasetVersionId = NULL,
  partitionColumn = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasetId}}{The ID of the dataset.}

\item{\code{datasetName}}{A user-friendly name for the dataset.}

\item{\code{datasetVersionId}}{The ID of the dataset version.}

\item{\code{partitionColumn}}{The name of the column containing the partition assignments.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeprecatedTrainingDataAssignment-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DeprecatedTrainingDataAssignment-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeprecatedTrainingDataAssignment$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeprecatedTrainingDataAssignment-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeprecatedTrainingDataAssignment-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeprecatedTrainingDataAssignment$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeprecatedTrainingDataAssignment-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeprecatedTrainingDataAssignment-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeprecatedTrainingDataAssignment$fromJSON(
  DeprecatedTrainingDataAssignmentJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeprecatedTrainingDataAssignmentJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeprecatedTrainingDataAssignment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeprecatedTrainingDataAssignment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeprecatedTrainingDataAssignment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
