% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_features.R
\docType{class}
\name{DerivedFeatures}
\alias{DerivedFeatures}
\title{DerivedFeatures}
\format{
An \code{R6Class} generator object
}
\description{
DerivedFeatures Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{coefficient}}{numeric The coefficient for this feature.}

\item{\code{derivedFeature}}{character The name of the derived feature.}

\item{\code{originalFeature}}{character The name of the feature used to derive this feature.}

\item{\code{stageCoefficients}}{list( \link{StageCoefficients} ) An array of json objects describing separate coefficients for every stage of model (empty for single stage models).}

\item{\code{transformations}}{list( \link{Transformations} ) An array of json objects describing the transformations applied to create this derived feature.}

\item{\code{type}}{character The type of this feature.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DerivedFeatures-new}{\code{DerivedFeatures$new()}}
\item \href{#method-DerivedFeatures-validate}{\code{DerivedFeatures$validate()}}
\item \href{#method-DerivedFeatures-toJSON}{\code{DerivedFeatures$toJSON()}}
\item \href{#method-DerivedFeatures-fromJSON}{\code{DerivedFeatures$fromJSON()}}
\item \href{#method-DerivedFeatures-clone}{\code{DerivedFeatures$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DerivedFeatures-new"></a>}}
\if{latex}{\out{\hypertarget{method-DerivedFeatures-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DerivedFeatures$new(
  coefficient = NULL,
  derivedFeature = NULL,
  originalFeature = NULL,
  stageCoefficients = NULL,
  transformations = NULL,
  type = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coefficient}}{The coefficient for this feature.}

\item{\code{derivedFeature}}{The name of the derived feature.}

\item{\code{originalFeature}}{The name of the feature used to derive this feature.}

\item{\code{stageCoefficients}}{An array of json objects describing separate coefficients for every stage of model (empty for single stage models).}

\item{\code{transformations}}{An array of json objects describing the transformations applied to create this derived feature.}

\item{\code{type}}{The type of this feature.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DerivedFeatures-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DerivedFeatures-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DerivedFeatures$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DerivedFeatures-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DerivedFeatures-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DerivedFeatures$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DerivedFeatures-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DerivedFeatures-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DerivedFeatures$fromJSON(DerivedFeaturesJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DerivedFeaturesJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DerivedFeatures-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DerivedFeatures-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DerivedFeatures$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
