% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discarded_features_response.R
\docType{class}
\name{DiscardedFeaturesResponse}
\alias{DiscardedFeaturesResponse}
\title{DiscardedFeaturesResponse}
\format{
An \code{R6Class} generator object
}
\description{
DiscardedFeaturesResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer Discarded features count.}

\item{\code{features}}{list( character ) Discarded features.}

\item{\code{remainingRestoreLimit}}{integer The remaining available number of the features which can be restored in this project.}

\item{\code{totalRestoreLimit}}{integer The total limit indicating how many features can be restored in this project.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DiscardedFeaturesResponse$new()}}
\item \href{#method-validate}{\code{DiscardedFeaturesResponse$validate()}}
\item \href{#method-toJSON}{\code{DiscardedFeaturesResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{DiscardedFeaturesResponse$fromJSON()}}
\item \href{#method-clone}{\code{DiscardedFeaturesResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscardedFeaturesResponse$new(
  count = NULL,
  features = NULL,
  remainingRestoreLimit = NULL,
  totalRestoreLimit = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{Discarded features count.}

\item{\code{features}}{Discarded features.}

\item{\code{remainingRestoreLimit}}{The remaining available number of the features which can be restored in this project.}

\item{\code{totalRestoreLimit}}{The total limit indicating how many features can be restored in this project.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscardedFeaturesResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscardedFeaturesResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscardedFeaturesResponse$fromJSON(
  DiscardedFeaturesResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DiscardedFeaturesResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscardedFeaturesResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
