% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ReturnResponse}
\alias{.ReturnResponse}
\title{A helper method to return the content of an API response.}
\usage{
.ReturnResponse(apiResponse.content)
}
\value{
\code{apiResponse.content} as an R6 object or an S3 list, dependent on the value of the
option \code{datarobot.apicore.returnS3}
}
\description{
A helper method to return the content of an API response.
}
