% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.setComplexProperty}
\alias{.setComplexProperty}
\title{Validate and set complex types}
\usage{
.setComplexProperty(typeList, propertyData)
}
\arguments{
\item{typeList}{An object vector of R6 class generators.}

\item{propertyData}{A deserialized JSON string, OR an R6 object instance.}
}
\value{
The R6 input object if it's one of the listed types in \code{typeList}, or a newly instantiated
R6 object based on the input map if that map can be converted into one of the types in \code{typeList}, or NULL if input is NULL.
}
\description{
A helper method to validate complex types for anyOf and oneOf properties in an R6 object.
}
