% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.toS3}
\alias{.toS3}
\title{A helper method to return an arbitrary object as an S3 object or as is.}
\usage{
.toS3(potential.R6.object)
}
\description{
If the object parameter is an R6 object, it will be converted to a list and all of the R6
functions dropped. The R6 class name will be preserved.
}
\details{
This requires that the R6 object have the property \code{lock_objects = FALSE}; otherwise errors will
be thrown.
}
