% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{DownloadPrimeCode}
\alias{DownloadPrimeCode}
\title{DownloadPrimeCode}
\usage{
DownloadPrimeCode(project, primeFileId, filepath, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Download Code
}
\details{
Download code from an existing Prime file

This method invokes \verb{GET /projects/\{projectId\}/primeFiles/\{primeFileId\}/download/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPrimeFilesDownloadList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
