% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_response.R
\docType{class}
\name{DriverResponse}
\alias{DriverResponse}
\title{DriverResponse}
\format{
An \code{R6Class} generator object
}
\description{
DriverResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{associatedAuthTypes}}{list( character ) \link{optional} Authentication types associated with this connector configuration.}

\item{\code{baseNames}}{list( character ) \link{optional} Original file name(s) of the uploaded JAR file(s). If there are multiple JAR files required for the driver, each of the original file names should be present in the list in the same order as found in 'localJarUrls'}

\item{\code{canonicalName}}{character \link{optional} User-friendly driver name.}

\item{\code{className}}{character \link{optional} Driver class name. For example 'com.amazon.redshift.jdbc.Driver')}

\item{\code{configurationId}}{character \link{optional} Driver configuration ID if it was provided during driver upload.}

\item{\code{creator}}{character ID of the user who uploaded the driver.}

\item{\code{id}}{character Driver ID.}

\item{\code{version}}{character \link{optional} Driver version, which is used to construct the canonical name if driver configuration ID was provided during driver upload.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DriverResponse-new}{\code{DriverResponse$new()}}
\item \href{#method-DriverResponse-validate}{\code{DriverResponse$validate()}}
\item \href{#method-DriverResponse-toJSON}{\code{DriverResponse$toJSON()}}
\item \href{#method-DriverResponse-fromJSON}{\code{DriverResponse$fromJSON()}}
\item \href{#method-DriverResponse-clone}{\code{DriverResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DriverResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DriverResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DriverResponse$new(
  creator = NULL,
  id = NULL,
  associatedAuthTypes = NULL,
  baseNames = NULL,
  canonicalName = NULL,
  className = NULL,
  configurationId = NULL,
  version = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{creator}}{ID of the user who uploaded the driver.}

\item{\code{id}}{Driver ID.}

\item{\code{associatedAuthTypes}}{Authentication types associated with this connector configuration.}

\item{\code{baseNames}}{Original file name(s) of the uploaded JAR file(s). If there are multiple JAR files required for the driver, each of the original file names should be present in the list in the same order as found in 'localJarUrls'}

\item{\code{canonicalName}}{User-friendly driver name.}

\item{\code{className}}{Driver class name. For example 'com.amazon.redshift.jdbc.Driver')}

\item{\code{configurationId}}{Driver configuration ID if it was provided during driver upload.}

\item{\code{version}}{Driver version, which is used to construct the canonical name if driver configuration ID was provided during driver upload.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DriverResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DriverResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DriverResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DriverResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DriverResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DriverResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DriverResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DriverResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DriverResponse$fromJSON(DriverResponseJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DriverResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DriverResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DriverResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DriverResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
