% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_count_by_type_response.R
\docType{class}
\name{EntityCountByTypeResponse}
\alias{EntityCountByTypeResponse}
\title{EntityCountByTypeResponse}
\format{
An \code{R6Class} generator object
}
\description{
EntityCountByTypeResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{numCalendars}}{integer The number of calendars that use the dataset}

\item{\code{numExternalModelPackages}}{integer The number of external model packages that use the dataset}

\item{\code{numFeatureDiscoveryConfigs}}{integer The number of feature discovery configs that use the dataset}

\item{\code{numPredictionDatasets}}{integer The number of prediction datasets that use the dataset}

\item{\code{numProjects}}{integer The number of projects that use the dataset}

\item{\code{numSparkSqlQueries}}{integer The number of spark sql queries that use the dataset}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EntityCountByTypeResponse-new}{\code{EntityCountByTypeResponse$new()}}
\item \href{#method-EntityCountByTypeResponse-validate}{\code{EntityCountByTypeResponse$validate()}}
\item \href{#method-EntityCountByTypeResponse-toJSON}{\code{EntityCountByTypeResponse$toJSON()}}
\item \href{#method-EntityCountByTypeResponse-fromJSON}{\code{EntityCountByTypeResponse$fromJSON()}}
\item \href{#method-EntityCountByTypeResponse-clone}{\code{EntityCountByTypeResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntityCountByTypeResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-EntityCountByTypeResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityCountByTypeResponse$new(
  numCalendars = NULL,
  numExternalModelPackages = NULL,
  numFeatureDiscoveryConfigs = NULL,
  numPredictionDatasets = NULL,
  numProjects = NULL,
  numSparkSqlQueries = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{numCalendars}}{The number of calendars that use the dataset}

\item{\code{numExternalModelPackages}}{The number of external model packages that use the dataset}

\item{\code{numFeatureDiscoveryConfigs}}{The number of feature discovery configs that use the dataset}

\item{\code{numPredictionDatasets}}{The number of prediction datasets that use the dataset}

\item{\code{numProjects}}{The number of projects that use the dataset}

\item{\code{numSparkSqlQueries}}{The number of spark sql queries that use the dataset}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntityCountByTypeResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-EntityCountByTypeResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityCountByTypeResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntityCountByTypeResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-EntityCountByTypeResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityCountByTypeResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntityCountByTypeResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-EntityCountByTypeResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityCountByTypeResponse$fromJSON(
  EntityCountByTypeResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{EntityCountByTypeResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntityCountByTypeResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EntityCountByTypeResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityCountByTypeResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
