% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eureqa_distribution_detail_response.R
\docType{class}
\name{EureqaDistributionDetailResponse}
\alias{EureqaDistributionDetailResponse}
\title{EureqaDistributionDetailResponse}
\format{
An \code{R6Class} generator object
}
\description{
EureqaDistributionDetailResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bins}}{list( \link{ClassificationBinDataResponse} ) The distribution plot data.}

\item{\code{complexity}}{integer The complexity score for this solution. Complexity score is a function of the mathematical operators used in the current solution. The complexity calculation can be tuned via model hyperparameters.}

\item{\code{error}}{numeric The error for the current solution, as computed by eureqa using the `errorMetric` error metric. None if Eureqa model refitted existing solutions.}

\item{\code{errorMetric}}{character The Eureqa error metric identifier used to compute error metrics for this search. Note that Eureqa error metrics do not correspond 1:1 with DataRobot error metrics - the available metrics are not the same, and even equivalent metrics may be computed slightly differently.}

\item{\code{eureqaSolutionId}}{character The ID of the solution.}

\item{\code{expression}}{character The eureqa \"solution string\". This is a mathematical expression; human-readable but with strict syntax specifications defined by Eureqa.}

\item{\code{expressionAnnotated}}{character The `expression`, rendered with additional tags to assist in automatic parsing.}

\item{\code{threshold}}{numeric Classifier threshold selected by the backend, used to determine which model values are binned as positive and which are binned as negative. Must have a value between the `binStart` of the first bin and `binEnd` of the last bin.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EureqaDistributionDetailResponse-new}{\code{EureqaDistributionDetailResponse$new()}}
\item \href{#method-EureqaDistributionDetailResponse-validate}{\code{EureqaDistributionDetailResponse$validate()}}
\item \href{#method-EureqaDistributionDetailResponse-toJSON}{\code{EureqaDistributionDetailResponse$toJSON()}}
\item \href{#method-EureqaDistributionDetailResponse-fromJSON}{\code{EureqaDistributionDetailResponse$fromJSON()}}
\item \href{#method-EureqaDistributionDetailResponse-clone}{\code{EureqaDistributionDetailResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EureqaDistributionDetailResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-EureqaDistributionDetailResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EureqaDistributionDetailResponse$new(
  bins = NULL,
  complexity = NULL,
  error = NULL,
  errorMetric = NULL,
  eureqaSolutionId = NULL,
  expression = NULL,
  expressionAnnotated = NULL,
  threshold = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bins}}{The distribution plot data.}

\item{\code{complexity}}{The complexity score for this solution. Complexity score is a function of the mathematical operators used in the current solution. The complexity calculation can be tuned via model hyperparameters.}

\item{\code{error}}{The error for the current solution, as computed by eureqa using the `errorMetric` error metric. None if Eureqa model refitted existing solutions.}

\item{\code{errorMetric}}{The Eureqa error metric identifier used to compute error metrics for this search. Note that Eureqa error metrics do not correspond 1:1 with DataRobot error metrics - the available metrics are not the same, and even equivalent metrics may be computed slightly differently.}

\item{\code{eureqaSolutionId}}{The ID of the solution.}

\item{\code{expression}}{The eureqa \"solution string\". This is a mathematical expression; human-readable but with strict syntax specifications defined by Eureqa.}

\item{\code{expressionAnnotated}}{The `expression`, rendered with additional tags to assist in automatic parsing.}

\item{\code{threshold}}{Classifier threshold selected by the backend, used to determine which model values are binned as positive and which are binned as negative. Must have a value between the `binStart` of the first bin and `binEnd` of the last bin.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EureqaDistributionDetailResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-EureqaDistributionDetailResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EureqaDistributionDetailResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EureqaDistributionDetailResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-EureqaDistributionDetailResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EureqaDistributionDetailResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EureqaDistributionDetailResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-EureqaDistributionDetailResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EureqaDistributionDetailResponse$fromJSON(
  EureqaDistributionDetailResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{EureqaDistributionDetailResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EureqaDistributionDetailResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EureqaDistributionDetailResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EureqaDistributionDetailResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
