% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eureqa_model_detail_response.R
\docType{class}
\name{EureqaModelDetailResponse}
\alias{EureqaModelDetailResponse}
\title{EureqaModelDetailResponse}
\format{
An \code{R6Class} generator object
}
\description{
EureqaModelDetailResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{complexity}}{integer The complexity score for this solution. Complexity score is a function of the mathematical operators used in the current solution. The complexity calculation can be tuned via model hyperparameters.}

\item{\code{error}}{numeric The error for the current solution, as computed by eureqa using the `errorMetric` error metric. None if Eureqa model refitted existing solutions.}

\item{\code{errorMetric}}{character The Eureqa error metric identifier used to compute error metrics for this search. Note that Eureqa error metrics do not correspond 1:1 with DataRobot error metrics - the available metrics are not the same, and even equivalent metrics may be computed slightly differently.}

\item{\code{eureqaSolutionId}}{character The ID of the solution.}

\item{\code{expression}}{character The eureqa \"solution string\". This is a mathematical expression; human-readable but with strict syntax specifications defined by Eureqa.}

\item{\code{expressionAnnotated}}{character The `expression`, rendered with additional tags to assist in automatic parsing.}

\item{\code{plotData}}{list( \link{PlotDataResponse} ) The plot data.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EureqaModelDetailResponse$new()}}
\item \href{#method-validate}{\code{EureqaModelDetailResponse$validate()}}
\item \href{#method-toJSON}{\code{EureqaModelDetailResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{EureqaModelDetailResponse$fromJSON()}}
\item \href{#method-clone}{\code{EureqaModelDetailResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EureqaModelDetailResponse$new(
  complexity = NULL,
  error = NULL,
  errorMetric = NULL,
  eureqaSolutionId = NULL,
  expression = NULL,
  expressionAnnotated = NULL,
  plotData = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{complexity}}{The complexity score for this solution. Complexity score is a function of the mathematical operators used in the current solution. The complexity calculation can be tuned via model hyperparameters.}

\item{\code{error}}{The error for the current solution, as computed by eureqa using the `errorMetric` error metric. None if Eureqa model refitted existing solutions.}

\item{\code{errorMetric}}{The Eureqa error metric identifier used to compute error metrics for this search. Note that Eureqa error metrics do not correspond 1:1 with DataRobot error metrics - the available metrics are not the same, and even equivalent metrics may be computed slightly differently.}

\item{\code{eureqaSolutionId}}{The ID of the solution.}

\item{\code{expression}}{The eureqa \"solution string\". This is a mathematical expression; human-readable but with strict syntax specifications defined by Eureqa.}

\item{\code{expressionAnnotated}}{The `expression`, rendered with additional tags to assist in automatic parsing.}

\item{\code{plotData}}{The plot data.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EureqaModelDetailResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EureqaModelDetailResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EureqaModelDetailResponse$fromJSON(
  EureqaModelDetailResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{EureqaModelDetailResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EureqaModelDetailResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
