% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_group.R
\docType{class}
\name{EventGroup}
\alias{EventGroup}
\title{EventGroup}
\format{
An \code{R6Class} generator object
}
\description{
EventGroup Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{events}}{list( \link{EventType} ) The events included in this group.}

\item{\code{id}}{character The id of the event group.}

\item{\code{label}}{character The display name of the event group.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EventGroup-new}{\code{EventGroup$new()}}
\item \href{#method-EventGroup-validate}{\code{EventGroup$validate()}}
\item \href{#method-EventGroup-toJSON}{\code{EventGroup$toJSON()}}
\item \href{#method-EventGroup-fromJSON}{\code{EventGroup$fromJSON()}}
\item \href{#method-EventGroup-clone}{\code{EventGroup$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventGroup-new"></a>}}
\if{latex}{\out{\hypertarget{method-EventGroup-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventGroup$new(
  events = NULL,
  id = NULL,
  label = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{events}}{The events included in this group.}

\item{\code{id}}{The id of the event group.}

\item{\code{label}}{The display name of the event group.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventGroup-validate"></a>}}
\if{latex}{\out{\hypertarget{method-EventGroup-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventGroup$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventGroup-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-EventGroup-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventGroup$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventGroup-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-EventGroup-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventGroup$fromJSON(EventGroupJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{EventGroupJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventGroup-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EventGroup-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
