% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_group_response.R
\docType{class}
\name{EventGroupResponse}
\alias{EventGroupResponse}
\title{EventGroupResponse}
\format{
An \code{R6Class} generator object
}
\description{
EventGroupResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{events}}{list( character ) Event types belonging to the group}

\item{\code{id}}{character Group ID}

\item{\code{label}}{character Group name for display}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EventGroupResponse-new}{\code{EventGroupResponse$new()}}
\item \href{#method-EventGroupResponse-validate}{\code{EventGroupResponse$validate()}}
\item \href{#method-EventGroupResponse-toJSON}{\code{EventGroupResponse$toJSON()}}
\item \href{#method-EventGroupResponse-fromJSON}{\code{EventGroupResponse$fromJSON()}}
\item \href{#method-EventGroupResponse-clone}{\code{EventGroupResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventGroupResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-EventGroupResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventGroupResponse$new(
  events = NULL,
  id = NULL,
  label = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{events}}{Event types belonging to the group}

\item{\code{id}}{Group ID}

\item{\code{label}}{Group name for display}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventGroupResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-EventGroupResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventGroupResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventGroupResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-EventGroupResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventGroupResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventGroupResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-EventGroupResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventGroupResponse$fromJSON(EventGroupResponseJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{EventGroupResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventGroupResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EventGroupResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventGroupResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
