% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_scores_list_data.R
\docType{class}
\name{ExternalScoresListData}
\alias{ExternalScoresListData}
\title{ExternalScoresListData}
\format{
An \code{R6Class} generator object
}
\description{
ExternalScoresListData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actualValueColumn}}{character The name of the column with actuals that was used to calculate the scores.}

\item{\code{datasetId}}{character The datasetId the data comes from.}

\item{\code{modelId}}{character The modelId for the scores.}

\item{\code{projectId}}{character The projectId for the scores.}

\item{\code{scores}}{list( \link{ExternalScoresListDataScore} ) A json array of the computed scores.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExternalScoresListData-new}{\code{ExternalScoresListData$new()}}
\item \href{#method-ExternalScoresListData-validate}{\code{ExternalScoresListData$validate()}}
\item \href{#method-ExternalScoresListData-toJSON}{\code{ExternalScoresListData$toJSON()}}
\item \href{#method-ExternalScoresListData-fromJSON}{\code{ExternalScoresListData$fromJSON()}}
\item \href{#method-ExternalScoresListData-clone}{\code{ExternalScoresListData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalScoresListData-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalScoresListData-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalScoresListData$new(
  actualValueColumn = NULL,
  datasetId = NULL,
  modelId = NULL,
  projectId = NULL,
  scores = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{actualValueColumn}}{The name of the column with actuals that was used to calculate the scores.}

\item{\code{datasetId}}{The datasetId the data comes from.}

\item{\code{modelId}}{The modelId for the scores.}

\item{\code{projectId}}{The projectId for the scores.}

\item{\code{scores}}{A json array of the computed scores.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalScoresListData-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalScoresListData-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalScoresListData$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalScoresListData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalScoresListData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalScoresListData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalScoresListData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalScoresListData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalScoresListData$fromJSON(
  ExternalScoresListDataJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ExternalScoresListDataJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalScoresListData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalScoresListData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalScoresListData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
