% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_ts_baseline_payload.R
\docType{class}
\name{ExternalTSBaselinePayload}
\alias{ExternalTSBaselinePayload}
\title{ExternalTSBaselinePayload}
\format{
An \code{R6Class} generator object
}
\description{
ExternalTSBaselinePayload Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{backtests}}{list( \link{Backtests} ) \link{optional} An array of the configured backtests.}

\item{\code{catalogVersionId}}{character The version id of the external baseline data item in the AI catalog.}

\item{\code{datetimePartitionColumn}}{character The date column that will be used as the datetime partition column for the specified project.}

\item{\code{forecastWindowEnd}}{integer For time series projects only. How many timeUnits of the datetimePartitionColumn into the future relative to the forecast point the forecast window should end.}

\item{\code{forecastWindowStart}}{integer For time series projects only. How many timeUnits of the datetimePartitionColumn into the future relative to the forecast point the forecast window should start.}

\item{\code{holdoutEndDate}}{character \link{optional} The end date of holdout scoring data.}

\item{\code{holdoutStartDate}}{character \link{optional} The start date of holdout scoring data.}

\item{\code{multiseriesIdColumns}}{list( character ) \link{optional} An array of column names identifying the multiseries ID column(s)to use to identify series within the data. Must match the multiseries ID column(s) for the specified project. Currently, only one multiseries ID column may be specified.}

\item{\code{target}}{character The selected target of the specified project.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExternalTSBaselinePayload-new}{\code{ExternalTSBaselinePayload$new()}}
\item \href{#method-ExternalTSBaselinePayload-validate}{\code{ExternalTSBaselinePayload$validate()}}
\item \href{#method-ExternalTSBaselinePayload-toJSON}{\code{ExternalTSBaselinePayload$toJSON()}}
\item \href{#method-ExternalTSBaselinePayload-fromJSON}{\code{ExternalTSBaselinePayload$fromJSON()}}
\item \href{#method-ExternalTSBaselinePayload-clone}{\code{ExternalTSBaselinePayload$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalTSBaselinePayload-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalTSBaselinePayload-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalTSBaselinePayload$new(
  catalogVersionId = NULL,
  datetimePartitionColumn = NULL,
  forecastWindowEnd = NULL,
  forecastWindowStart = NULL,
  target = NULL,
  backtests = NULL,
  holdoutEndDate = NULL,
  holdoutStartDate = NULL,
  multiseriesIdColumns = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalogVersionId}}{The version id of the external baseline data item in the AI catalog.}

\item{\code{datetimePartitionColumn}}{The date column that will be used as the datetime partition column for the specified project.}

\item{\code{forecastWindowEnd}}{For time series projects only. How many timeUnits of the datetimePartitionColumn into the future relative to the forecast point the forecast window should end.}

\item{\code{forecastWindowStart}}{For time series projects only. How many timeUnits of the datetimePartitionColumn into the future relative to the forecast point the forecast window should start.}

\item{\code{target}}{The selected target of the specified project.}

\item{\code{backtests}}{An array of the configured backtests.}

\item{\code{holdoutEndDate}}{The end date of holdout scoring data.}

\item{\code{holdoutStartDate}}{The start date of holdout scoring data.}

\item{\code{multiseriesIdColumns}}{An array of column names identifying the multiseries ID column(s)to use to identify series within the data. Must match the multiseries ID column(s) for the specified project. Currently, only one multiseries ID column may be specified.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalTSBaselinePayload-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalTSBaselinePayload-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalTSBaselinePayload$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalTSBaselinePayload-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalTSBaselinePayload-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalTSBaselinePayload$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalTSBaselinePayload-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalTSBaselinePayload-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalTSBaselinePayload$fromJSON(
  ExternalTSBaselinePayloadJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ExternalTSBaselinePayloadJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalTSBaselinePayload-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalTSBaselinePayload-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalTSBaselinePayload$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
