% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_ts_baseline_response.R
\docType{class}
\name{ExternalTSBaselineResponse}
\alias{ExternalTSBaselineResponse}
\title{ExternalTSBaselineResponse}
\format{
An \code{R6Class} generator object
}
\description{
ExternalTSBaselineResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{backtests}}{list( \link{Backtests} ) \link{optional} An array of the configured backtests.}

\item{\code{baselineValidationJobId}}{character The id of the validation job.}

\item{\code{catalogVersionId}}{character The version id of the external baseline data item in the AI catalog.}

\item{\code{datetimePartitionColumn}}{character The date column that will be used as the datetime partition column for the specified project.}

\item{\code{forecastWindowEnd}}{integer \link{optional} For time series projects only. How many timeUnits of the datetimePartitionColumn into the future relative to the forecast point the forecast window should end.}

\item{\code{forecastWindowStart}}{integer \link{optional} For time series projects only. How many timeUnits of the datetimePartitionColumn into the future relative to the forecast point the forecast window should start.}

\item{\code{holdoutEndDate}}{character \link{optional} The end date of holdout scoring data.}

\item{\code{holdoutStartDate}}{character \link{optional} The start date of holdout scoring data.}

\item{\code{isExternalBaselineDatasetValid}}{character Indicates whether the external dataset has pass the validation check or not.}

\item{\code{message}}{character A message providing mode detail on the validation result.}

\item{\code{multiseriesIdColumns}}{list( character ) \link{optional} An array of column names identifying the multiseries ID column(s)to use to identify series within the data. Must match the multiseries ID column(s) for the specified project. Currently, only one multiseries ID column may be specified.}

\item{\code{projectId}}{character The project id of the external baseline data item.}

\item{\code{target}}{character The selected target of the specified project.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ExternalTSBaselineResponse$new()}}
\item \href{#method-validate}{\code{ExternalTSBaselineResponse$validate()}}
\item \href{#method-toJSON}{\code{ExternalTSBaselineResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{ExternalTSBaselineResponse$fromJSON()}}
\item \href{#method-clone}{\code{ExternalTSBaselineResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalTSBaselineResponse$new(
  baselineValidationJobId = NULL,
  catalogVersionId = NULL,
  datetimePartitionColumn = NULL,
  isExternalBaselineDatasetValid = NULL,
  message = NULL,
  projectId = NULL,
  target = NULL,
  backtests = NULL,
  forecastWindowEnd = NULL,
  forecastWindowStart = NULL,
  holdoutEndDate = NULL,
  holdoutStartDate = NULL,
  multiseriesIdColumns = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{baselineValidationJobId}}{The id of the validation job.}

\item{\code{catalogVersionId}}{The version id of the external baseline data item in the AI catalog.}

\item{\code{datetimePartitionColumn}}{The date column that will be used as the datetime partition column for the specified project.}

\item{\code{isExternalBaselineDatasetValid}}{Indicates whether the external dataset has pass the validation check or not.}

\item{\code{message}}{A message providing mode detail on the validation result.}

\item{\code{projectId}}{The project id of the external baseline data item.}

\item{\code{target}}{The selected target of the specified project.}

\item{\code{backtests}}{An array of the configured backtests.}

\item{\code{forecastWindowEnd}}{For time series projects only. How many timeUnits of the datetimePartitionColumn into the future relative to the forecast point the forecast window should end.}

\item{\code{forecastWindowStart}}{For time series projects only. How many timeUnits of the datetimePartitionColumn into the future relative to the forecast point the forecast window should start.}

\item{\code{holdoutEndDate}}{The end date of holdout scoring data.}

\item{\code{holdoutStartDate}}{The start date of holdout scoring data.}

\item{\code{multiseriesIdColumns}}{An array of column names identifying the multiseries ID column(s)to use to identify series within the data. Must match the multiseries ID column(s) for the specified project. Currently, only one multiseries ID column may be specified.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalTSBaselineResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalTSBaselineResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalTSBaselineResponse$fromJSON(
  ExternalTSBaselineResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ExternalTSBaselineResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalTSBaselineResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
