% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairness_insight.R
\docType{class}
\name{FairnessInsight}
\alias{FairnessInsight}
\title{FairnessInsight}
\format{
An \code{R6Class} generator object
}
\description{
FairnessInsight Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fairnessMetric}}{character The fairness metric used to calculate the fairness scores.}

\item{\code{fairnessThreshold}}{numeric Value of the fairness threshold, defined in project options.}

\item{\code{modelId}}{character ID of the model fairness was measured for.}

\item{\code{perClassFairness}}{list( \link{PerClassFairness} ) An array of calculated fairness scores for each protected feature class.}

\item{\code{predictionThreshold}}{numeric \link{optional} Model's prediction threshold used when insight was calculated. ``null`` if prediction threshold is not required for the fairness metric calculations.}

\item{\code{protectedFeature}}{character Name of the protected feature the fairness calculation is made for.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FairnessInsight$new()}}
\item \href{#method-validate}{\code{FairnessInsight$validate()}}
\item \href{#method-toJSON}{\code{FairnessInsight$toJSON()}}
\item \href{#method-fromJSON}{\code{FairnessInsight$fromJSON()}}
\item \href{#method-clone}{\code{FairnessInsight$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessInsight$new(
  fairnessMetric = NULL,
  fairnessThreshold = NULL,
  modelId = NULL,
  perClassFairness = NULL,
  protectedFeature = NULL,
  predictionThreshold = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fairnessMetric}}{The fairness metric used to calculate the fairness scores.}

\item{\code{fairnessThreshold}}{Value of the fairness threshold, defined in project options.}

\item{\code{modelId}}{ID of the model fairness was measured for.}

\item{\code{perClassFairness}}{An array of calculated fairness scores for each protected feature class.}

\item{\code{protectedFeature}}{Name of the protected feature the fairness calculation is made for.}

\item{\code{predictionThreshold}}{Model's prediction threshold used when insight was calculated. ``null`` if prediction threshold is not required for the fairness metric calculations.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessInsight$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessInsight$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessInsight$fromJSON(FairnessInsightJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FairnessInsightJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessInsight$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
