% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairness_insights_start_calculation_payload.R
\docType{class}
\name{FairnessInsightsStartCalculationPayload}
\alias{FairnessInsightsStartCalculationPayload}
\title{FairnessInsightsStartCalculationPayload}
\format{
An \code{R6Class} generator object
}
\description{
FairnessInsightsStartCalculationPayload Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fairnessMetricsSet}}{character \link{optional} Metric to use for calculating fairness. Can be one of ``proportionalParity``, ``equalParity``, ``predictionBalance``, ``trueFavorableAndUnfavorableRateParity`` or ``FavorableAndUnfavorablePredictiveValueParity``. Used and required only if \emph{Bias & Fairness in AutoML} feature is enabled.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FairnessInsightsStartCalculationPayload-new}{\code{FairnessInsightsStartCalculationPayload$new()}}
\item \href{#method-FairnessInsightsStartCalculationPayload-validate}{\code{FairnessInsightsStartCalculationPayload$validate()}}
\item \href{#method-FairnessInsightsStartCalculationPayload-toJSON}{\code{FairnessInsightsStartCalculationPayload$toJSON()}}
\item \href{#method-FairnessInsightsStartCalculationPayload-fromJSON}{\code{FairnessInsightsStartCalculationPayload$fromJSON()}}
\item \href{#method-FairnessInsightsStartCalculationPayload-clone}{\code{FairnessInsightsStartCalculationPayload$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FairnessInsightsStartCalculationPayload-new"></a>}}
\if{latex}{\out{\hypertarget{method-FairnessInsightsStartCalculationPayload-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessInsightsStartCalculationPayload$new(
  fairnessMetricsSet = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fairnessMetricsSet}}{Metric to use for calculating fairness. Can be one of ``proportionalParity``, ``equalParity``, ``predictionBalance``, ``trueFavorableAndUnfavorableRateParity`` or ``FavorableAndUnfavorablePredictiveValueParity``. Used and required only if \emph{Bias & Fairness in AutoML} feature is enabled.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FairnessInsightsStartCalculationPayload-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FairnessInsightsStartCalculationPayload-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessInsightsStartCalculationPayload$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FairnessInsightsStartCalculationPayload-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FairnessInsightsStartCalculationPayload-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessInsightsStartCalculationPayload$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FairnessInsightsStartCalculationPayload-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FairnessInsightsStartCalculationPayload-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessInsightsStartCalculationPayload$fromJSON(
  FairnessInsightsStartCalculationPayloadJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FairnessInsightsStartCalculationPayloadJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FairnessInsightsStartCalculationPayload-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FairnessInsightsStartCalculationPayload-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FairnessInsightsStartCalculationPayload$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
