% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/feature_association_details_retrieve_controller_response.R
\docType{class}
\name{FeatureAssociationDetailsRetrieveControllerResponse}
\alias{FeatureAssociationDetailsRetrieveControllerResponse}
\title{FeatureAssociationDetailsRetrieveControllerResponse}
\format{
An \code{R6Class} generator object
}
\description{
FeatureAssociationDetailsRetrieveControllerResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{chartType}}{character Which type of plotting the pair of features gets in the UI, e.g. `SCATTER`}

\item{\code{features}}{list( character ) The name of `feature1` and `feature2`.}

\item{\code{types}}{list( character ) The type of `feature1` and `feature2`. Possible values: `CATEGORICAL`, `NUMERIC`.}

\item{\code{values}}{list( \link{array[AnyOfstringnumber]} ) The data triplets for pairwise plotting.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FeatureAssociationDetailsRetrieveControllerResponse$new()}}
\item \href{#method-validate}{\code{FeatureAssociationDetailsRetrieveControllerResponse$validate()}}
\item \href{#method-toJSON}{\code{FeatureAssociationDetailsRetrieveControllerResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{FeatureAssociationDetailsRetrieveControllerResponse$fromJSON()}}
\item \href{#method-clone}{\code{FeatureAssociationDetailsRetrieveControllerResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationDetailsRetrieveControllerResponse$new(
  chartType = NULL,
  features = NULL,
  types = NULL,
  values = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chartType}}{Which type of plotting the pair of features gets in the UI, e.g. `SCATTER`}

\item{\code{features}}{The name of `feature1` and `feature2`.}

\item{\code{types}}{The type of `feature1` and `feature2`. Possible values: `CATEGORICAL`, `NUMERIC`.}

\item{\code{values}}{The data triplets for pairwise plotting.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationDetailsRetrieveControllerResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationDetailsRetrieveControllerResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationDetailsRetrieveControllerResponse$fromJSON(
  FeatureAssociationDetailsRetrieveControllerResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureAssociationDetailsRetrieveControllerResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationDetailsRetrieveControllerResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
