% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_association_list.R
\docType{class}
\name{FeatureAssociationList}
\alias{FeatureAssociationList}
\title{FeatureAssociationList}
\format{
An \code{R6Class} generator object
}
\description{
FeatureAssociationList Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{featurelistId}}{character The featurelist Id.}

\item{\code{hasFam}}{character Whether Feature Association Matrix is calculated for featurelist.}

\item{\code{title}}{character The name of featurelist.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureAssociationList-new}{\code{FeatureAssociationList$new()}}
\item \href{#method-FeatureAssociationList-validate}{\code{FeatureAssociationList$validate()}}
\item \href{#method-FeatureAssociationList-toJSON}{\code{FeatureAssociationList$toJSON()}}
\item \href{#method-FeatureAssociationList-fromJSON}{\code{FeatureAssociationList$fromJSON()}}
\item \href{#method-FeatureAssociationList-clone}{\code{FeatureAssociationList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAssociationList-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAssociationList-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationList$new(
  featurelistId = NULL,
  hasFam = NULL,
  title = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{featurelistId}}{The featurelist Id.}

\item{\code{hasFam}}{Whether Feature Association Matrix is calculated for featurelist.}

\item{\code{title}}{The name of featurelist.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAssociationList-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAssociationList-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationList$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAssociationList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAssociationList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAssociationList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAssociationList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationList$fromJSON(
  FeatureAssociationListJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureAssociationListJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAssociationList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAssociationList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
