% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/feature_association_retrieve_controller_response.R
\docType{class}
\name{FeatureAssociationRetrieveControllerResponse}
\alias{FeatureAssociationRetrieveControllerResponse}
\title{FeatureAssociationRetrieveControllerResponse}
\format{
An \code{R6Class} generator object
}
\description{
FeatureAssociationRetrieveControllerResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{features}}{list( \link{FeatureAssociationRetrieveFeatures} ) Metadata for each feature and where it goes in the matrix as structured below.}

\item{\code{strengths}}{list( \link{FeatureAssociationRetrieveStrengths} ) Pairwise statistics for the available features as structured below.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureAssociationRetrieveControllerResponse-new}{\code{FeatureAssociationRetrieveControllerResponse$new()}}
\item \href{#method-FeatureAssociationRetrieveControllerResponse-validate}{\code{FeatureAssociationRetrieveControllerResponse$validate()}}
\item \href{#method-FeatureAssociationRetrieveControllerResponse-toJSON}{\code{FeatureAssociationRetrieveControllerResponse$toJSON()}}
\item \href{#method-FeatureAssociationRetrieveControllerResponse-fromJSON}{\code{FeatureAssociationRetrieveControllerResponse$fromJSON()}}
\item \href{#method-FeatureAssociationRetrieveControllerResponse-clone}{\code{FeatureAssociationRetrieveControllerResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAssociationRetrieveControllerResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAssociationRetrieveControllerResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationRetrieveControllerResponse$new(
  features = NULL,
  strengths = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{features}}{Metadata for each feature and where it goes in the matrix as structured below.}

\item{\code{strengths}}{Pairwise statistics for the available features as structured below.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAssociationRetrieveControllerResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAssociationRetrieveControllerResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationRetrieveControllerResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAssociationRetrieveControllerResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAssociationRetrieveControllerResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationRetrieveControllerResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAssociationRetrieveControllerResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAssociationRetrieveControllerResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationRetrieveControllerResponse$fromJSON(
  FeatureAssociationRetrieveControllerResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureAssociationRetrieveControllerResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAssociationRetrieveControllerResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAssociationRetrieveControllerResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationRetrieveControllerResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
