% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_association_retrieve_features.R
\docType{class}
\name{FeatureAssociationRetrieveFeatures}
\alias{FeatureAssociationRetrieveFeatures}
\title{FeatureAssociationRetrieveFeatures}
\format{
An \code{R6Class} generator object
}
\description{
FeatureAssociationRetrieveFeatures Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{alphabeticSortIndex}}{integer A number representing the alphabetical order of this feature compared to the other features in this dataset.}

\item{\code{clusterId}}{integer The ID of the feature cluster this feature belongs to.}

\item{\code{clusterName}}{character The name of feature cluster.}

\item{\code{clusterSortIndex}}{integer A number representing the ordering of the feature across all feature clusters. Features in the same cluster always have adjacent indices.}

\item{\code{feature}}{character The name of the feature.}

\item{\code{importanceSortIndex}}{integer A number ranking the importance of this feature compared to the other features in this dataset.}

\item{\code{strengthSortIndex}}{integer A number ranking the strength of this feature compared to the other features in this dataset.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureAssociationRetrieveFeatures-new}{\code{FeatureAssociationRetrieveFeatures$new()}}
\item \href{#method-FeatureAssociationRetrieveFeatures-validate}{\code{FeatureAssociationRetrieveFeatures$validate()}}
\item \href{#method-FeatureAssociationRetrieveFeatures-toJSON}{\code{FeatureAssociationRetrieveFeatures$toJSON()}}
\item \href{#method-FeatureAssociationRetrieveFeatures-fromJSON}{\code{FeatureAssociationRetrieveFeatures$fromJSON()}}
\item \href{#method-FeatureAssociationRetrieveFeatures-clone}{\code{FeatureAssociationRetrieveFeatures$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAssociationRetrieveFeatures-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAssociationRetrieveFeatures-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationRetrieveFeatures$new(
  alphabeticSortIndex = NULL,
  clusterId = NULL,
  clusterName = NULL,
  clusterSortIndex = NULL,
  feature = NULL,
  importanceSortIndex = NULL,
  strengthSortIndex = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alphabeticSortIndex}}{A number representing the alphabetical order of this feature compared to the other features in this dataset.}

\item{\code{clusterId}}{The ID of the feature cluster this feature belongs to.}

\item{\code{clusterName}}{The name of feature cluster.}

\item{\code{clusterSortIndex}}{A number representing the ordering of the feature across all feature clusters. Features in the same cluster always have adjacent indices.}

\item{\code{feature}}{The name of the feature.}

\item{\code{importanceSortIndex}}{A number ranking the importance of this feature compared to the other features in this dataset.}

\item{\code{strengthSortIndex}}{A number ranking the strength of this feature compared to the other features in this dataset.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAssociationRetrieveFeatures-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAssociationRetrieveFeatures-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationRetrieveFeatures$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAssociationRetrieveFeatures-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAssociationRetrieveFeatures-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationRetrieveFeatures$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAssociationRetrieveFeatures-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAssociationRetrieveFeatures-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationRetrieveFeatures$fromJSON(
  FeatureAssociationRetrieveFeaturesJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureAssociationRetrieveFeaturesJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAssociationRetrieveFeatures-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAssociationRetrieveFeatures-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationRetrieveFeatures$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
