% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_association_retrieve_strengths.R
\docType{class}
\name{FeatureAssociationRetrieveStrengths}
\alias{FeatureAssociationRetrieveStrengths}
\title{FeatureAssociationRetrieveStrengths}
\format{
An \code{R6Class} generator object
}
\description{
FeatureAssociationRetrieveStrengths Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{feature1}}{character The name of the first feature.}

\item{\code{feature2}}{character The name of the second feature.}

\item{\code{statistic}}{numeric Feature association statistics for `feature1` and `feature2`. For features with no pairwise statistics available the value is `null`.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureAssociationRetrieveStrengths-new}{\code{FeatureAssociationRetrieveStrengths$new()}}
\item \href{#method-FeatureAssociationRetrieveStrengths-validate}{\code{FeatureAssociationRetrieveStrengths$validate()}}
\item \href{#method-FeatureAssociationRetrieveStrengths-toJSON}{\code{FeatureAssociationRetrieveStrengths$toJSON()}}
\item \href{#method-FeatureAssociationRetrieveStrengths-fromJSON}{\code{FeatureAssociationRetrieveStrengths$fromJSON()}}
\item \href{#method-FeatureAssociationRetrieveStrengths-clone}{\code{FeatureAssociationRetrieveStrengths$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAssociationRetrieveStrengths-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAssociationRetrieveStrengths-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationRetrieveStrengths$new(
  feature1 = NULL,
  feature2 = NULL,
  statistic = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature1}}{The name of the first feature.}

\item{\code{feature2}}{The name of the second feature.}

\item{\code{statistic}}{Feature association statistics for `feature1` and `feature2`. For features with no pairwise statistics available the value is `null`.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAssociationRetrieveStrengths-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAssociationRetrieveStrengths-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationRetrieveStrengths$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAssociationRetrieveStrengths-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAssociationRetrieveStrengths-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationRetrieveStrengths$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAssociationRetrieveStrengths-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAssociationRetrieveStrengths-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationRetrieveStrengths$fromJSON(
  FeatureAssociationRetrieveStrengthsJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureAssociationRetrieveStrengthsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAssociationRetrieveStrengths-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAssociationRetrieveStrengths-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAssociationRetrieveStrengths$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
