% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_discovery_log_list_response.R
\docType{class}
\name{FeatureDiscoveryLogListResponse}
\alias{FeatureDiscoveryLogListResponse}
\title{FeatureDiscoveryLogListResponse}
\format{
An \code{R6Class} generator object
}
\description{
FeatureDiscoveryLogListResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer Number of items returned on this page}

\item{\code{featureDiscoveryLog}}{list( character ) List of lines retrieved from the feature discovery log}

\item{\code{next_}}{character URL pointing to the next page (if null, there is no next page)}

\item{\code{previous}}{character URL pointing to the previous page (if null, there is no previous page)}

\item{\code{totalLogLines}}{integer total number of lines in feature derivation log}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureDiscoveryLogListResponse-new}{\code{FeatureDiscoveryLogListResponse$new()}}
\item \href{#method-FeatureDiscoveryLogListResponse-validate}{\code{FeatureDiscoveryLogListResponse$validate()}}
\item \href{#method-FeatureDiscoveryLogListResponse-toJSON}{\code{FeatureDiscoveryLogListResponse$toJSON()}}
\item \href{#method-FeatureDiscoveryLogListResponse-fromJSON}{\code{FeatureDiscoveryLogListResponse$fromJSON()}}
\item \href{#method-FeatureDiscoveryLogListResponse-clone}{\code{FeatureDiscoveryLogListResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureDiscoveryLogListResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureDiscoveryLogListResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureDiscoveryLogListResponse$new(
  count = NULL,
  featureDiscoveryLog = NULL,
  next_ = NULL,
  previous = NULL,
  totalLogLines = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{Number of items returned on this page}

\item{\code{featureDiscoveryLog}}{List of lines retrieved from the feature discovery log}

\item{\code{next_}}{URL pointing to the next page (if null, there is no next page)}

\item{\code{previous}}{URL pointing to the previous page (if null, there is no previous page)}

\item{\code{totalLogLines}}{total number of lines in feature derivation log}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureDiscoveryLogListResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureDiscoveryLogListResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureDiscoveryLogListResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureDiscoveryLogListResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureDiscoveryLogListResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureDiscoveryLogListResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureDiscoveryLogListResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureDiscoveryLogListResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureDiscoveryLogListResponse$fromJSON(
  FeatureDiscoveryLogListResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureDiscoveryLogListResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureDiscoveryLogListResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureDiscoveryLogListResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureDiscoveryLogListResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
