% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_drift_over_time_bucket.R
\docType{class}
\name{FeatureDriftOverTimeBucket}
\alias{FeatureDriftOverTimeBucket}
\title{FeatureDriftOverTimeBucket}
\format{
An \code{R6Class} generator object
}
\description{
FeatureDriftOverTimeBucket Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{baselineSampleSize}}{integer The sample size of the training data used during model creation}

\item{\code{driftScore}}{numeric The aggregated drift score for the target.}

\item{\code{featureImpact}}{numeric The feature impact score of the feature.}

\item{\code{featureName}}{character Name of the feature.}

\item{\code{period}}{\link{TimeRange}}

\item{\code{sampleSize}}{integer The sample size in the query period used to calculate drift score.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureDriftOverTimeBucket-new}{\code{FeatureDriftOverTimeBucket$new()}}
\item \href{#method-FeatureDriftOverTimeBucket-validate}{\code{FeatureDriftOverTimeBucket$validate()}}
\item \href{#method-FeatureDriftOverTimeBucket-toJSON}{\code{FeatureDriftOverTimeBucket$toJSON()}}
\item \href{#method-FeatureDriftOverTimeBucket-fromJSON}{\code{FeatureDriftOverTimeBucket$fromJSON()}}
\item \href{#method-FeatureDriftOverTimeBucket-clone}{\code{FeatureDriftOverTimeBucket$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureDriftOverTimeBucket-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureDriftOverTimeBucket-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureDriftOverTimeBucket$new(
  baselineSampleSize = NULL,
  driftScore = NULL,
  featureImpact = NULL,
  featureName = NULL,
  period = NULL,
  sampleSize = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{baselineSampleSize}}{The sample size of the training data used during model creation}

\item{\code{driftScore}}{The aggregated drift score for the target.}

\item{\code{featureImpact}}{The feature impact score of the feature.}

\item{\code{featureName}}{Name of the feature.}

\item{\code{sampleSize}}{The sample size in the query period used to calculate drift score.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureDriftOverTimeBucket-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureDriftOverTimeBucket-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureDriftOverTimeBucket$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureDriftOverTimeBucket-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureDriftOverTimeBucket-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureDriftOverTimeBucket$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureDriftOverTimeBucket-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureDriftOverTimeBucket-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureDriftOverTimeBucket$fromJSON(
  FeatureDriftOverTimeBucketJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureDriftOverTimeBucketJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureDriftOverTimeBucket-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureDriftOverTimeBucket-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureDriftOverTimeBucket$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
