% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_effects_datetime_response.R
\docType{class}
\name{FeatureEffectsDatetimeResponse}
\alias{FeatureEffectsDatetimeResponse}
\title{FeatureEffectsDatetimeResponse}
\format{
An \code{R6Class} generator object
}
\description{
FeatureEffectsDatetimeResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{backtestIndex}}{character The backtest index. For example: `0`, `1`, ..., `20`, `holdout`, `startstop`.}

\item{\code{featureEffects}}{list( \link{FeatureEffects} ) Feature Effects computational results for each feature.}

\item{\code{modelId}}{character The model ID}

\item{\code{projectId}}{character The project ID}

\item{\code{source}}{character Models data source.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FeatureEffectsDatetimeResponse$new()}}
\item \href{#method-validate}{\code{FeatureEffectsDatetimeResponse$validate()}}
\item \href{#method-toJSON}{\code{FeatureEffectsDatetimeResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{FeatureEffectsDatetimeResponse$fromJSON()}}
\item \href{#method-clone}{\code{FeatureEffectsDatetimeResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureEffectsDatetimeResponse$new(
  backtestIndex = NULL,
  featureEffects = NULL,
  modelId = NULL,
  projectId = NULL,
  source = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{backtestIndex}}{The backtest index. For example: `0`, `1`, ..., `20`, `holdout`, `startstop`.}

\item{\code{featureEffects}}{Feature Effects computational results for each feature.}

\item{\code{modelId}}{The model ID}

\item{\code{projectId}}{The project ID}

\item{\code{source}}{Models data source.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureEffectsDatetimeResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureEffectsDatetimeResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureEffectsDatetimeResponse$fromJSON(
  FeatureEffectsDatetimeResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureEffectsDatetimeResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureEffectsDatetimeResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
