% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_fit_datetime_response.R
\docType{class}
\name{FeatureFitDatetimeResponse}
\alias{FeatureFitDatetimeResponse}
\title{FeatureFitDatetimeResponse}
\format{
An \code{R6Class} generator object
}
\description{
FeatureFitDatetimeResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{backtestIndex}}{character The backtest index. For example: `0`, `1`, ..., `20`, `holdout`, `startstop`.}

\item{\code{featureFit}}{list( \link{FeatureFit} ) Feature Fit computational results for each feature.}

\item{\code{modelId}}{character The model ID}

\item{\code{projectId}}{character The project ID}

\item{\code{source}}{character Models data source.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureFitDatetimeResponse-new}{\code{FeatureFitDatetimeResponse$new()}}
\item \href{#method-FeatureFitDatetimeResponse-validate}{\code{FeatureFitDatetimeResponse$validate()}}
\item \href{#method-FeatureFitDatetimeResponse-toJSON}{\code{FeatureFitDatetimeResponse$toJSON()}}
\item \href{#method-FeatureFitDatetimeResponse-fromJSON}{\code{FeatureFitDatetimeResponse$fromJSON()}}
\item \href{#method-FeatureFitDatetimeResponse-clone}{\code{FeatureFitDatetimeResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureFitDatetimeResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureFitDatetimeResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureFitDatetimeResponse$new(
  backtestIndex = NULL,
  featureFit = NULL,
  modelId = NULL,
  projectId = NULL,
  source = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{backtestIndex}}{The backtest index. For example: `0`, `1`, ..., `20`, `holdout`, `startstop`.}

\item{\code{featureFit}}{Feature Fit computational results for each feature.}

\item{\code{modelId}}{The model ID}

\item{\code{projectId}}{The project ID}

\item{\code{source}}{Models data source.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureFitDatetimeResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureFitDatetimeResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureFitDatetimeResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureFitDatetimeResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureFitDatetimeResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureFitDatetimeResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureFitDatetimeResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureFitDatetimeResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureFitDatetimeResponse$fromJSON(
  FeatureFitDatetimeResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureFitDatetimeResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureFitDatetimeResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureFitDatetimeResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureFitDatetimeResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
