% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_histogram_plot_response.R
\docType{class}
\name{FeatureHistogramPlotResponse}
\alias{FeatureHistogramPlotResponse}
\title{FeatureHistogramPlotResponse}
\format{
An \code{R6Class} generator object
}
\description{
FeatureHistogramPlotResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{numeric number of values in the bin (or weights if project is weighted)}

\item{\code{label}}{character bin start for numerical/uncapped, or string value for categorical. The bin `==Missing==` is created for rows that did not have the feature.}

\item{\code{target}}{numeric Average value of the target feature values for the bin. For regression projects, it will be null, if the feature was deemed as low informative or project target has not been selected yet or AIM processing has not finished yet. You can use :http:get:`/api/v2/projects/(projectId)/features/` endpoint to find more about low informative features. For binary classification, the same conditions apply as above, but the value should be treated as the ratio of total positives in the bin to bin's total size (`count`). For multiclass projects the value is always null.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FeatureHistogramPlotResponse$new()}}
\item \href{#method-validate}{\code{FeatureHistogramPlotResponse$validate()}}
\item \href{#method-toJSON}{\code{FeatureHistogramPlotResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{FeatureHistogramPlotResponse$fromJSON()}}
\item \href{#method-clone}{\code{FeatureHistogramPlotResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureHistogramPlotResponse$new(
  count = NULL,
  label = NULL,
  target = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{number of values in the bin (or weights if project is weighted)}

\item{\code{label}}{bin start for numerical/uncapped, or string value for categorical. The bin `==Missing==` is created for rows that did not have the feature.}

\item{\code{target}}{Average value of the target feature values for the bin. For regression projects, it will be null, if the feature was deemed as low informative or project target has not been selected yet or AIM processing has not finished yet. You can use :http:get:`/api/v2/projects/(projectId)/features/` endpoint to find more about low informative features. For binary classification, the same conditions apply as above, but the value should be treated as the ratio of total positives in the bin to bin's total size (`count`). For multiclass projects the value is always null.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureHistogramPlotResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureHistogramPlotResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureHistogramPlotResponse$fromJSON(
  FeatureHistogramPlotResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureHistogramPlotResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureHistogramPlotResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
