% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_impact_response.R
\docType{class}
\name{FeatureImpactResponse}
\alias{FeatureImpactResponse}
\title{FeatureImpactResponse}
\format{
An \code{R6Class} generator object
}
\description{
FeatureImpactResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer Number of feature impact records in a given batch.}

\item{\code{featureImpacts}}{list( \link{FeatureImpactItem} ) A list which contains feature impact scores for each feature used by a model. If the model has more than 1000 features, the most important 1000 features are returned.}

\item{\code{next_}}{character URL for the next page of results or `null`.}

\item{\code{previous}}{character URL for the next page of results or `null`.}

\item{\code{ranRedundancyDetection}}{character Indicates whether redundant feature identification was run while calculating this feature impact.}

\item{\code{rowCount}}{integer The number of rows that was used to calculate feature impact. For the feature impact calculated with the default logic, without specifying the ``rowCount``, we return ``null`` here.}

\item{\code{shapBased}}{character Indicates whether feature impact was calculated using Shapley values. True for anomaly detection models when the project is unsupervised, as permutation approach is not applicable. Note that supervised projects must use an alternative route for SHAP impact: /api/v2/projects/(projectId)/models/(modelId)/shapImpact/}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FeatureImpactResponse$new()}}
\item \href{#method-validate}{\code{FeatureImpactResponse$validate()}}
\item \href{#method-toJSON}{\code{FeatureImpactResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{FeatureImpactResponse$fromJSON()}}
\item \href{#method-clone}{\code{FeatureImpactResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImpactResponse$new(
  count = NULL,
  featureImpacts = NULL,
  next_ = NULL,
  previous = NULL,
  ranRedundancyDetection = NULL,
  rowCount = NULL,
  shapBased = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{Number of feature impact records in a given batch.}

\item{\code{featureImpacts}}{A list which contains feature impact scores for each feature used by a model. If the model has more than 1000 features, the most important 1000 features are returned.}

\item{\code{next_}}{URL for the next page of results or `null`.}

\item{\code{previous}}{URL for the next page of results or `null`.}

\item{\code{ranRedundancyDetection}}{Indicates whether redundant feature identification was run while calculating this feature impact.}

\item{\code{rowCount}}{The number of rows that was used to calculate feature impact. For the feature impact calculated with the default logic, without specifying the ``rowCount``, we return ``null`` here.}

\item{\code{shapBased}}{Indicates whether feature impact was calculated using Shapley values. True for anomaly detection models when the project is unsupervised, as permutation approach is not applicable. Note that supervised projects must use an alternative route for SHAP impact: /api/v2/projects/(projectId)/models/(modelId)/shapImpact/}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImpactResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImpactResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImpactResponse$fromJSON(
  FeatureImpactResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureImpactResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImpactResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
