% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/feature_key_summary_details_response_validator_summarized_categorical.R
\docType{class}
\name{FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical}
\alias{FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical}
\title{FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical}
\format{
An \code{R6Class} generator object
}
\description{
FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dataQualities}}{character The indicator of data quality assessment of the feature.}

\item{\code{max}}{numeric Maximum value of the key.}

\item{\code{mean}}{numeric Mean value of the key.}

\item{\code{median}}{numeric Median value of the key.}

\item{\code{min}}{numeric Minimum value of the key.}

\item{\code{pctRows}}{numeric Percentage occurrence of key in the EDA sample of the feature.}

\item{\code{stdDev}}{numeric Standard deviation of the key.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical-new}{\code{FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical$new()}}
\item \href{#method-FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical-validate}{\code{FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical$validate()}}
\item \href{#method-FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical-toJSON}{\code{FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical$toJSON()}}
\item \href{#method-FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical-fromJSON}{\code{FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical$fromJSON()}}
\item \href{#method-FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical-clone}{\code{FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical$new(
  dataQualities = NULL,
  max = NULL,
  mean = NULL,
  median = NULL,
  min = NULL,
  pctRows = NULL,
  stdDev = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataQualities}}{The indicator of data quality assessment of the feature.}

\item{\code{max}}{Maximum value of the key.}

\item{\code{mean}}{Mean value of the key.}

\item{\code{median}}{Median value of the key.}

\item{\code{min}}{Minimum value of the key.}

\item{\code{pctRows}}{Percentage occurrence of key in the EDA sample of the feature.}

\item{\code{stdDev}}{Standard deviation of the key.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical$fromJSON(
  FeatureKeySummaryDetailsResponseValidatorSummarizedCategoricalJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureKeySummaryDetailsResponseValidatorSummarizedCategoricalJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureKeySummaryDetailsResponseValidatorSummarizedCategorical$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
