% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_lineage_step.R
\docType{class}
\name{FeatureLineageStep}
\alias{FeatureLineageStep}
\title{FeatureLineageStep}
\format{
An \code{R6Class} generator object
}
\description{
FeatureLineageStep Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{arguments}}{object \link{optional} Generic key-value pairs to describe \strong{action} step additional parameters.}

\item{\code{catalogId}}{character \link{optional} id of the catalog for a \strong{data} step.}

\item{\code{catalogVersionId}}{character \link{optional} id of the catalog version for a \strong{data} step.}

\item{\code{description}}{character \link{optional} Description of the step.}

\item{\code{groupBy}}{list( character ) \link{optional} List of columns which this \strong{action} step aggregated.}

\item{\code{id}}{integer Step id starting with 0.}

\item{\code{isTimeAware}}{character \link{optional} Indicator of step being time aware. Mandatory only for \strong{action} and \strong{join} steps. \strong{action} step provides additional information about feature derivation window in the `timeInfo` field.}

\item{\code{joinInfo}}{\link{FeatureLineageJoin} \link{optional}}

\item{\code{name}}{character \link{optional} Name of the step.}

\item{\code{parents}}{list( integer ) `id` of steps which use this step output as their input.}

\item{\code{stepType}}{character One of four types of a step. \strong{data} - source features. \strong{action} - data aggregation or trasformation. \strong{join} - SQL JOIN. \strong{generatedData} - final feature. There is always one \strong{generatedData} step and at least one \strong{data} step.}

\item{\code{timeInfo}}{\link{FeatureLineageTimeInfo} \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureLineageStep-new}{\code{FeatureLineageStep$new()}}
\item \href{#method-FeatureLineageStep-validate}{\code{FeatureLineageStep$validate()}}
\item \href{#method-FeatureLineageStep-toJSON}{\code{FeatureLineageStep$toJSON()}}
\item \href{#method-FeatureLineageStep-fromJSON}{\code{FeatureLineageStep$fromJSON()}}
\item \href{#method-FeatureLineageStep-clone}{\code{FeatureLineageStep$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureLineageStep-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureLineageStep-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureLineageStep$new(
  id = NULL,
  parents = NULL,
  stepType = NULL,
  arguments = NULL,
  catalogId = NULL,
  catalogVersionId = NULL,
  description = NULL,
  groupBy = NULL,
  isTimeAware = NULL,
  joinInfo = NULL,
  name = NULL,
  timeInfo = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Step id starting with 0.}

\item{\code{parents}}{`id` of steps which use this step output as their input.}

\item{\code{stepType}}{One of four types of a step. \strong{data} - source features. \strong{action} - data aggregation or trasformation. \strong{join} - SQL JOIN. \strong{generatedData} - final feature. There is always one \strong{generatedData} step and at least one \strong{data} step.}

\item{\code{arguments}}{Generic key-value pairs to describe \strong{action} step additional parameters.}

\item{\code{catalogId}}{id of the catalog for a \strong{data} step.}

\item{\code{catalogVersionId}}{id of the catalog version for a \strong{data} step.}

\item{\code{description}}{Description of the step.}

\item{\code{groupBy}}{List of columns which this \strong{action} step aggregated.}

\item{\code{isTimeAware}}{Indicator of step being time aware. Mandatory only for \strong{action} and \strong{join} steps. \strong{action} step provides additional information about feature derivation window in the `timeInfo` field.}

\item{\code{name}}{Name of the step.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureLineageStep-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureLineageStep-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureLineageStep$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureLineageStep-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureLineageStep-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureLineageStep$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureLineageStep-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureLineageStep-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureLineageStep$fromJSON(FeatureLineageStepJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureLineageStepJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureLineageStep-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureLineageStep-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureLineageStep$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
