% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_list_create.R
\docType{class}
\name{FeatureListCreate}
\alias{FeatureListCreate}
\title{FeatureListCreate}
\format{
An \code{R6Class} generator object
}
\description{
FeatureListCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{description}}{character \link{optional} Description of the featurelist}

\item{\code{features}}{list( character ) List of names of features to be included in the new featurelist, all features listed must be part of the universe. All features for this dataset for the request to succeed.}

\item{\code{name}}{character Name of the featurelist to be created}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureListCreate-new}{\code{FeatureListCreate$new()}}
\item \href{#method-FeatureListCreate-validate}{\code{FeatureListCreate$validate()}}
\item \href{#method-FeatureListCreate-toJSON}{\code{FeatureListCreate$toJSON()}}
\item \href{#method-FeatureListCreate-fromJSON}{\code{FeatureListCreate$fromJSON()}}
\item \href{#method-FeatureListCreate-clone}{\code{FeatureListCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureListCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureListCreate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureListCreate$new(
  features = NULL,
  name = NULL,
  description = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{features}}{List of names of features to be included in the new featurelist, all features listed must be part of the universe. All features for this dataset for the request to succeed.}

\item{\code{name}}{Name of the featurelist to be created}

\item{\code{description}}{Description of the featurelist}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureListCreate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureListCreate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureListCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureListCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureListCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureListCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureListCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureListCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureListCreate$fromJSON(FeatureListCreateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureListCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureListCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureListCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureListCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
