% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurelist_destroy_response.R
\docType{class}
\name{FeaturelistDestroyResponse}
\alias{FeaturelistDestroyResponse}
\title{FeaturelistDestroyResponse}
\format{
An \code{R6Class} generator object
}
\description{
FeaturelistDestroyResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{canDelete}}{character Whether the featurelist can be deleted.}

\item{\code{deletionBlockedReason}}{character If the featurelist can't be deleted, this explains why.}

\item{\code{dryRun}}{character Whether this was a dry-run or the featurelist was actually deleted.}

\item{\code{numAffectedJobs}}{integer Number of incomplete jobs using this featurelist.}

\item{\code{numAffectedModels}}{integer Number of models using this featurelist.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeaturelistDestroyResponse-new}{\code{FeaturelistDestroyResponse$new()}}
\item \href{#method-FeaturelistDestroyResponse-validate}{\code{FeaturelistDestroyResponse$validate()}}
\item \href{#method-FeaturelistDestroyResponse-toJSON}{\code{FeaturelistDestroyResponse$toJSON()}}
\item \href{#method-FeaturelistDestroyResponse-fromJSON}{\code{FeaturelistDestroyResponse$fromJSON()}}
\item \href{#method-FeaturelistDestroyResponse-clone}{\code{FeaturelistDestroyResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeaturelistDestroyResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeaturelistDestroyResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeaturelistDestroyResponse$new(
  canDelete = NULL,
  deletionBlockedReason = NULL,
  dryRun = NULL,
  numAffectedJobs = NULL,
  numAffectedModels = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{canDelete}}{Whether the featurelist can be deleted.}

\item{\code{deletionBlockedReason}}{If the featurelist can't be deleted, this explains why.}

\item{\code{dryRun}}{Whether this was a dry-run or the featurelist was actually deleted.}

\item{\code{numAffectedJobs}}{Number of incomplete jobs using this featurelist.}

\item{\code{numAffectedModels}}{Number of models using this featurelist.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeaturelistDestroyResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FeaturelistDestroyResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeaturelistDestroyResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeaturelistDestroyResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeaturelistDestroyResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeaturelistDestroyResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeaturelistDestroyResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeaturelistDestroyResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeaturelistDestroyResponse$fromJSON(
  FeaturelistDestroyResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeaturelistDestroyResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeaturelistDestroyResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeaturelistDestroyResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeaturelistDestroyResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
