% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurelist_response.R
\docType{class}
\name{FeaturelistResponse}
\alias{FeaturelistResponse}
\title{FeaturelistResponse}
\format{
An \code{R6Class} generator object
}
\description{
FeaturelistResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{created}}{character A :ref:`timestamp <time_format>` string specifying when the featurelist was created.}

\item{\code{description}}{character User-friendly description of the featurelist, which can be updated by users.}

\item{\code{features}}{list( character ) Names of features included in the featurelist.}

\item{\code{id}}{character Featurelist ID.}

\item{\code{isUserCreated}}{character Whether the featurelist was created manually by a user or by DataRobot automation.}

\item{\code{name}}{character the name of the featurelist}

\item{\code{numModels}}{integer The number of models that currently use this featurelist. A model is considered to use a featurelist if it is used to train the model or as a monotonic constraint featurelist, or if the model is a blender with at least one component model using the featurelist.}

\item{\code{projectId}}{character Project ID the featurelist belongs to.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeaturelistResponse-new}{\code{FeaturelistResponse$new()}}
\item \href{#method-FeaturelistResponse-validate}{\code{FeaturelistResponse$validate()}}
\item \href{#method-FeaturelistResponse-toJSON}{\code{FeaturelistResponse$toJSON()}}
\item \href{#method-FeaturelistResponse-fromJSON}{\code{FeaturelistResponse$fromJSON()}}
\item \href{#method-FeaturelistResponse-clone}{\code{FeaturelistResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeaturelistResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeaturelistResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeaturelistResponse$new(
  created = NULL,
  description = NULL,
  features = NULL,
  id = NULL,
  isUserCreated = NULL,
  name = NULL,
  numModels = NULL,
  projectId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{created}}{A :ref:`timestamp <time_format>` string specifying when the featurelist was created.}

\item{\code{description}}{User-friendly description of the featurelist, which can be updated by users.}

\item{\code{features}}{Names of features included in the featurelist.}

\item{\code{id}}{Featurelist ID.}

\item{\code{isUserCreated}}{Whether the featurelist was created manually by a user or by DataRobot automation.}

\item{\code{name}}{the name of the featurelist}

\item{\code{numModels}}{The number of models that currently use this featurelist. A model is considered to use a featurelist if it is used to train the model or as a monotonic constraint featurelist, or if the model is a blender with at least one component model using the featurelist.}

\item{\code{projectId}}{Project ID the featurelist belongs to.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeaturelistResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FeaturelistResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeaturelistResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeaturelistResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeaturelistResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeaturelistResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeaturelistResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeaturelistResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeaturelistResponse$fromJSON(FeaturelistResponseJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeaturelistResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeaturelistResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeaturelistResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeaturelistResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
