% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_metric_details_response.R
\docType{class}
\name{FeatureMetricDetailsResponse}
\alias{FeatureMetricDetailsResponse}
\title{FeatureMetricDetailsResponse}
\format{
An \code{R6Class} generator object
}
\description{
FeatureMetricDetailsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ascending}}{character Should the metric be sorted in ascending order}

\item{\code{metricName}}{character Name of the metric}

\item{\code{supportsBinary}}{character This metric is valid for binary classifciaton}

\item{\code{supportsMulticlass}}{character This metric is valid for mutliclass classifciaton}

\item{\code{supportsRegression}}{character This metric is valid for regression}

\item{\code{supportsTimeseries}}{character This metric is valid for timeseries}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureMetricDetailsResponse-new}{\code{FeatureMetricDetailsResponse$new()}}
\item \href{#method-FeatureMetricDetailsResponse-validate}{\code{FeatureMetricDetailsResponse$validate()}}
\item \href{#method-FeatureMetricDetailsResponse-toJSON}{\code{FeatureMetricDetailsResponse$toJSON()}}
\item \href{#method-FeatureMetricDetailsResponse-fromJSON}{\code{FeatureMetricDetailsResponse$fromJSON()}}
\item \href{#method-FeatureMetricDetailsResponse-clone}{\code{FeatureMetricDetailsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureMetricDetailsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureMetricDetailsResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureMetricDetailsResponse$new(
  ascending = NULL,
  metricName = NULL,
  supportsBinary = NULL,
  supportsMulticlass = NULL,
  supportsRegression = NULL,
  supportsTimeseries = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ascending}}{Should the metric be sorted in ascending order}

\item{\code{metricName}}{Name of the metric}

\item{\code{supportsBinary}}{This metric is valid for binary classifciaton}

\item{\code{supportsMulticlass}}{This metric is valid for mutliclass classifciaton}

\item{\code{supportsRegression}}{This metric is valid for regression}

\item{\code{supportsTimeseries}}{This metric is valid for timeseries}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureMetricDetailsResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureMetricDetailsResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureMetricDetailsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureMetricDetailsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureMetricDetailsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureMetricDetailsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureMetricDetailsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureMetricDetailsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureMetricDetailsResponse$fromJSON(
  FeatureMetricDetailsResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureMetricDetailsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureMetricDetailsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureMetricDetailsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureMetricDetailsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
