% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_setting.R
\docType{class}
\name{FeatureSetting}
\alias{FeatureSetting}
\title{FeatureSetting}
\format{
An \code{R6Class} generator object
}
\description{
FeatureSetting Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aPriori}}{character \link{optional} Renamed to `knownInAdvance`.}

\item{\code{doNotDerive}}{character \link{optional} For time series projects only. Sets whether the feature is do-not-derive, i.e., is excluded from feature derivation. If not specified, the feature uses the value from the `defaultToDoNotDerive` flag.}

\item{\code{featureName}}{character The name of the feature being specified.}

\item{\code{knownInAdvance}}{character \link{optional} For time series projects only. Sets whether the feature is known in advance, i.e., values for future dates are known at prediction time. If not specified, the feature uses the value from the `defaultToKnownInAdvance` flag.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureSetting-new}{\code{FeatureSetting$new()}}
\item \href{#method-FeatureSetting-validate}{\code{FeatureSetting$validate()}}
\item \href{#method-FeatureSetting-toJSON}{\code{FeatureSetting$toJSON()}}
\item \href{#method-FeatureSetting-fromJSON}{\code{FeatureSetting$fromJSON()}}
\item \href{#method-FeatureSetting-clone}{\code{FeatureSetting$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureSetting-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureSetting-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetting$new(
  featureName = NULL,
  aPriori = NULL,
  doNotDerive = NULL,
  knownInAdvance = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{featureName}}{The name of the feature being specified.}

\item{\code{aPriori}}{Renamed to `knownInAdvance`.}

\item{\code{doNotDerive}}{For time series projects only. Sets whether the feature is do-not-derive, i.e., is excluded from feature derivation. If not specified, the feature uses the value from the `defaultToDoNotDerive` flag.}

\item{\code{knownInAdvance}}{For time series projects only. Sets whether the feature is known in advance, i.e., values for future dates are known at prediction time. If not specified, the feature uses the value from the `defaultToKnownInAdvance` flag.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureSetting-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureSetting-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetting$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureSetting-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureSetting-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetting$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureSetting-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureSetting-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetting$fromJSON(FeatureSettingJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureSettingJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureSetting-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureSetting-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSetting$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
