% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_transform.R
\docType{class}
\name{FeatureTransform}
\alias{FeatureTransform}
\title{FeatureTransform}
\format{
An \code{R6Class} generator object
}
\description{
FeatureTransform Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dateExtraction}}{character \link{optional} The value to extract from the date column, of these options: `\link{year|yearDay|month|monthDay|week|weekDay}`. Required for transformation of a date column. Otherwise must not be provided.}

\item{\code{name}}{character The name of the new feature. Must not be the same as any existing features for this project. Must not contain '/' character.}

\item{\code{parentName}}{character The name of the parent feature.}

\item{\code{replacement}}{\link{AnyOfstringbooleannumberinteger} \link{optional} The replacement in case of a failed transformation.}

\item{\code{variableType}}{character The type of the new feature. Must be one of `text`, `categorical` (Deprecated in version v2.21), `numeric`, or `categoricalInt`. See the description of this method for more information.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FeatureTransform$new()}}
\item \href{#method-validate}{\code{FeatureTransform$validate()}}
\item \href{#method-toJSON}{\code{FeatureTransform$toJSON()}}
\item \href{#method-fromJSON}{\code{FeatureTransform$fromJSON()}}
\item \href{#method-clone}{\code{FeatureTransform$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTransform$new(
  name = NULL,
  parentName = NULL,
  variableType = NULL,
  dateExtraction = NULL,
  replacement = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the new feature. Must not be the same as any existing features for this project. Must not contain '/' character.}

\item{\code{parentName}}{The name of the parent feature.}

\item{\code{variableType}}{The type of the new feature. Must be one of `text`, `categorical` (Deprecated in version v2.21), `numeric`, or `categoricalInt`. See the description of this method for more information.}

\item{\code{dateExtraction}}{The value to extract from the date column, of these options: `\link{year|yearDay|month|monthDay|week|weekDay}`. Required for transformation of a date column. Otherwise must not be provided.}

\item{\code{replacement}}{The replacement in case of a failed transformation.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTransform$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTransform$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTransform$fromJSON(FeatureTransformJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureTransformJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTransform$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
