% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/float_list.R
\docType{class}
\name{FloatList}
\alias{FloatList}
\title{FloatList}
\format{
An \code{R6Class} generator object
}
\description{
FloatList Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{maxLength}}{integer Maximum permitted length of the list.}

\item{\code{maxVal}}{numeric Maximum permitted value.}

\item{\code{minLength}}{integer Minimum permitted length of the list.}

\item{\code{minVal}}{numeric Minimum permitted value.}

\item{\code{supportsGridSearch}}{character When True, Grid Search is supported for this parameter.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FloatList-new}{\code{FloatList$new()}}
\item \href{#method-FloatList-validate}{\code{FloatList$validate()}}
\item \href{#method-FloatList-toJSON}{\code{FloatList$toJSON()}}
\item \href{#method-FloatList-fromJSON}{\code{FloatList$fromJSON()}}
\item \href{#method-FloatList-clone}{\code{FloatList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FloatList-new"></a>}}
\if{latex}{\out{\hypertarget{method-FloatList-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FloatList$new(
  maxLength = NULL,
  maxVal = NULL,
  minLength = NULL,
  minVal = NULL,
  supportsGridSearch = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maxLength}}{Maximum permitted length of the list.}

\item{\code{maxVal}}{Maximum permitted value.}

\item{\code{minLength}}{Minimum permitted length of the list.}

\item{\code{minVal}}{Minimum permitted value.}

\item{\code{supportsGridSearch}}{When True, Grid Search is supported for this parameter.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FloatList-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FloatList-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FloatList$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FloatList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FloatList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FloatList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FloatList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FloatList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FloatList$fromJSON(FloatListJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FloatListJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FloatList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FloatList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FloatList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
