% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_date.R
\docType{class}
\name{ForecastDate}
\alias{ForecastDate}
\title{ForecastDate}
\format{
An \code{R6Class} generator object
}
\description{
ForecastDate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{columnName}}{character \link{optional} The column name in prediction datasets to be used as forecast date}

\item{\code{datetimeFormat}}{character \link{optional} The datetime format of the forecast date column in prediction datasets.}

\item{\code{enabled}}{character Indicates whether predictions by forecast dates is enabled for the deployment.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ForecastDate-new}{\code{ForecastDate$new()}}
\item \href{#method-ForecastDate-validate}{\code{ForecastDate$validate()}}
\item \href{#method-ForecastDate-toJSON}{\code{ForecastDate$toJSON()}}
\item \href{#method-ForecastDate-fromJSON}{\code{ForecastDate$fromJSON()}}
\item \href{#method-ForecastDate-clone}{\code{ForecastDate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForecastDate-new"></a>}}
\if{latex}{\out{\hypertarget{method-ForecastDate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForecastDate$new(
  enabled = NULL,
  columnName = NULL,
  datetimeFormat = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enabled}}{Indicates whether predictions by forecast dates is enabled for the deployment.}

\item{\code{columnName}}{The column name in prediction datasets to be used as forecast date}

\item{\code{datetimeFormat}}{The datetime format of the forecast date column in prediction datasets.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForecastDate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ForecastDate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForecastDate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForecastDate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ForecastDate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForecastDate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForecastDate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ForecastDate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForecastDate$fromJSON(ForecastDateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ForecastDateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForecastDate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ForecastDate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForecastDate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
