% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_distance_plot_data_entry_response.R
\docType{class}
\name{ForecastDistancePlotDataEntryResponse}
\alias{ForecastDistancePlotDataEntryResponse}
\title{ForecastDistancePlotDataEntryResponse}
\format{
An \code{R6Class} generator object
}
\description{
ForecastDistancePlotDataEntryResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{backtestingScore}}{numeric Backtesting score for this forecast distance. If backtesting has not been run for this model, this score will be `null`.}

\item{\code{forecastDistance}}{integer The number of time units the scored rows are away from the forecast point.}

\item{\code{holdoutScore}}{numeric Holdout set score for this forecast distance. If holdout is locked for the project, this score will be `null`.}

\item{\code{validationScore}}{numeric Validation set score for this forecast distance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ForecastDistancePlotDataEntryResponse-new}{\code{ForecastDistancePlotDataEntryResponse$new()}}
\item \href{#method-ForecastDistancePlotDataEntryResponse-validate}{\code{ForecastDistancePlotDataEntryResponse$validate()}}
\item \href{#method-ForecastDistancePlotDataEntryResponse-toJSON}{\code{ForecastDistancePlotDataEntryResponse$toJSON()}}
\item \href{#method-ForecastDistancePlotDataEntryResponse-fromJSON}{\code{ForecastDistancePlotDataEntryResponse$fromJSON()}}
\item \href{#method-ForecastDistancePlotDataEntryResponse-clone}{\code{ForecastDistancePlotDataEntryResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForecastDistancePlotDataEntryResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ForecastDistancePlotDataEntryResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForecastDistancePlotDataEntryResponse$new(
  backtestingScore = NULL,
  forecastDistance = NULL,
  holdoutScore = NULL,
  validationScore = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{backtestingScore}}{Backtesting score for this forecast distance. If backtesting has not been run for this model, this score will be `null`.}

\item{\code{forecastDistance}}{The number of time units the scored rows are away from the forecast point.}

\item{\code{holdoutScore}}{Holdout set score for this forecast distance. If holdout is locked for the project, this score will be `null`.}

\item{\code{validationScore}}{Validation set score for this forecast distance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForecastDistancePlotDataEntryResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ForecastDistancePlotDataEntryResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForecastDistancePlotDataEntryResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForecastDistancePlotDataEntryResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ForecastDistancePlotDataEntryResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForecastDistancePlotDataEntryResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForecastDistancePlotDataEntryResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ForecastDistancePlotDataEntryResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForecastDistancePlotDataEntryResponse$fromJSON(
  ForecastDistancePlotDataEntryResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ForecastDistancePlotDataEntryResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForecastDistancePlotDataEntryResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ForecastDistancePlotDataEntryResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForecastDistancePlotDataEntryResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
