% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_vs_actual_plots_bins.R
\docType{class}
\name{ForecastVsActualPlotsBins}
\alias{ForecastVsActualPlotsBins}
\title{ForecastVsActualPlotsBins}
\format{
An \code{R6Class} generator object
}
\description{
ForecastVsActualPlotsBins Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual}}{numeric Average actual value of the target in the bin. `null` if there are no entries in the bin.}

\item{\code{endDate}}{character The datetime of the end of the bin (exclusive).}

\item{\code{error}}{numeric Average absolute residual value of the bin. `null` if there are no entries in the bin.}

\item{\code{forecasts}}{list( numeric ) An array of average forecasts for the model for each forecast distance. Empty if there are no forecasts in the bin. Each index in the `forecasts` array maps to `forecastDistances` array index.}

\item{\code{frequency}}{integer Indicates number of values averaged in bin in case of a resolution change.}

\item{\code{normalizedError}}{numeric Normalized average absolute residual value of the bin. `null` if there are no entries in the bin.}

\item{\code{startDate}}{character The datetime of the start of the bin (inclusive).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ForecastVsActualPlotsBins-new}{\code{ForecastVsActualPlotsBins$new()}}
\item \href{#method-ForecastVsActualPlotsBins-validate}{\code{ForecastVsActualPlotsBins$validate()}}
\item \href{#method-ForecastVsActualPlotsBins-toJSON}{\code{ForecastVsActualPlotsBins$toJSON()}}
\item \href{#method-ForecastVsActualPlotsBins-fromJSON}{\code{ForecastVsActualPlotsBins$fromJSON()}}
\item \href{#method-ForecastVsActualPlotsBins-clone}{\code{ForecastVsActualPlotsBins$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForecastVsActualPlotsBins-new"></a>}}
\if{latex}{\out{\hypertarget{method-ForecastVsActualPlotsBins-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForecastVsActualPlotsBins$new(
  actual = NULL,
  endDate = NULL,
  error = NULL,
  forecasts = NULL,
  frequency = NULL,
  normalizedError = NULL,
  startDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{actual}}{Average actual value of the target in the bin. `null` if there are no entries in the bin.}

\item{\code{endDate}}{The datetime of the end of the bin (exclusive).}

\item{\code{error}}{Average absolute residual value of the bin. `null` if there are no entries in the bin.}

\item{\code{forecasts}}{An array of average forecasts for the model for each forecast distance. Empty if there are no forecasts in the bin. Each index in the `forecasts` array maps to `forecastDistances` array index.}

\item{\code{frequency}}{Indicates number of values averaged in bin in case of a resolution change.}

\item{\code{normalizedError}}{Normalized average absolute residual value of the bin. `null` if there are no entries in the bin.}

\item{\code{startDate}}{The datetime of the start of the bin (inclusive).}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForecastVsActualPlotsBins-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ForecastVsActualPlotsBins-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForecastVsActualPlotsBins$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForecastVsActualPlotsBins-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ForecastVsActualPlotsBins-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForecastVsActualPlotsBins$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForecastVsActualPlotsBins-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ForecastVsActualPlotsBins-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForecastVsActualPlotsBins$fromJSON(
  ForecastVsActualPlotsBinsJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ForecastVsActualPlotsBinsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForecastVsActualPlotsBins-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ForecastVsActualPlotsBins-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForecastVsActualPlotsBins$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
