% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_vs_actual_plots_metadata_response.R
\docType{class}
\name{ForecastVsActualPlotsMetadataResponse}
\alias{ForecastVsActualPlotsMetadataResponse}
\title{ForecastVsActualPlotsMetadataResponse}
\format{
An \code{R6Class} generator object
}
\description{
ForecastVsActualPlotsMetadataResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{backtestMetadata}}{list( \link{DatetimeTrendPlotsBacktestMetadata} ) An array of metadata information for each backtest. The array index of metadata object is the backtest index.}

\item{\code{backtestStatuses}}{list( \link{ForecastVsActualPlotsStatus} ) An array of status information for each backtest. The array index of status object is the backtest index.}

\item{\code{holdoutMetadata}}{\link{DatetimeTrendPlotsBacktestMetadata}}

\item{\code{holdoutStatuses}}{\link{ForecastVsActualPlotsStatus}}

\item{\code{resolutions}}{list( character ) An array of available time resolutions for which plots can be retrieved.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ForecastVsActualPlotsMetadataResponse$new()}}
\item \href{#method-validate}{\code{ForecastVsActualPlotsMetadataResponse$validate()}}
\item \href{#method-toJSON}{\code{ForecastVsActualPlotsMetadataResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{ForecastVsActualPlotsMetadataResponse$fromJSON()}}
\item \href{#method-clone}{\code{ForecastVsActualPlotsMetadataResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForecastVsActualPlotsMetadataResponse$new(
  backtestMetadata = NULL,
  backtestStatuses = NULL,
  holdoutMetadata = NULL,
  holdoutStatuses = NULL,
  resolutions = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{backtestMetadata}}{An array of metadata information for each backtest. The array index of metadata object is the backtest index.}

\item{\code{backtestStatuses}}{An array of status information for each backtest. The array index of status object is the backtest index.}

\item{\code{resolutions}}{An array of available time resolutions for which plots can be retrieved.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForecastVsActualPlotsMetadataResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForecastVsActualPlotsMetadataResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForecastVsActualPlotsMetadataResponse$fromJSON(
  ForecastVsActualPlotsMetadataResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ForecastVsActualPlotsMetadataResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForecastVsActualPlotsMetadataResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
