% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequent_value_data.R
\docType{class}
\name{FrequentValueData}
\alias{FrequentValueData}
\title{FrequentValueData}
\format{
An \code{R6Class} generator object
}
\description{
FrequentValueData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer Count of specified frequent value in the sample, weighted by exposure or weights}

\item{\code{dataQuality}}{character Any data quality issue associated with this particularvalue of the feature. Possible data quality types include 'excess_zero', 'inlier', 'disguised_missing_value', and 'no_issues_found' and the relevant statistics.}

\item{\code{target}}{numeric Average target value for the specified frequent value if the target is binary or numeric. With weights or exposure, this becomes a weighted average. If the target is not set, it returns None.}

\item{\code{value}}{\link{AnyOfnumberstring} Specified frequent value, either a float or a string, like `=All Others+`}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FrequentValueData-new}{\code{FrequentValueData$new()}}
\item \href{#method-FrequentValueData-validate}{\code{FrequentValueData$validate()}}
\item \href{#method-FrequentValueData-toJSON}{\code{FrequentValueData$toJSON()}}
\item \href{#method-FrequentValueData-fromJSON}{\code{FrequentValueData$fromJSON()}}
\item \href{#method-FrequentValueData-clone}{\code{FrequentValueData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrequentValueData-new"></a>}}
\if{latex}{\out{\hypertarget{method-FrequentValueData-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrequentValueData$new(
  count = NULL,
  dataQuality = NULL,
  target = NULL,
  value = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{Count of specified frequent value in the sample, weighted by exposure or weights}

\item{\code{dataQuality}}{Any data quality issue associated with this particularvalue of the feature. Possible data quality types include 'excess_zero', 'inlier', 'disguised_missing_value', and 'no_issues_found' and the relevant statistics.}

\item{\code{target}}{Average target value for the specified frequent value if the target is binary or numeric. With weights or exposure, this becomes a weighted average. If the target is not set, it returns None.}

\item{\code{value}}{Specified frequent value, either a float or a string, like `=All Others+`}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrequentValueData-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FrequentValueData-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrequentValueData$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrequentValueData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FrequentValueData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrequentValueData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrequentValueData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FrequentValueData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrequentValueData$fromJSON(FrequentValueDataJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FrequentValueDataJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrequentValueData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FrequentValueData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrequentValueData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
