% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequent_values_response.R
\docType{class}
\name{FrequentValuesResponse}
\alias{FrequentValuesResponse}
\title{FrequentValuesResponse}
\format{
An \code{R6Class} generator object
}
\description{
FrequentValuesResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{frequentValues}}{list( \link{FrequentValueData} ) List of frequent value and data quality information}

\item{\code{name}}{character Name of the feature}

\item{\code{numRows}}{integer Number of rows in the sample used to determine frequent values}

\item{\code{projectId}}{character Project Id}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FrequentValuesResponse-new}{\code{FrequentValuesResponse$new()}}
\item \href{#method-FrequentValuesResponse-validate}{\code{FrequentValuesResponse$validate()}}
\item \href{#method-FrequentValuesResponse-toJSON}{\code{FrequentValuesResponse$toJSON()}}
\item \href{#method-FrequentValuesResponse-fromJSON}{\code{FrequentValuesResponse$fromJSON()}}
\item \href{#method-FrequentValuesResponse-clone}{\code{FrequentValuesResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrequentValuesResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-FrequentValuesResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrequentValuesResponse$new(
  frequentValues = NULL,
  name = NULL,
  numRows = NULL,
  projectId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{frequentValues}}{List of frequent value and data quality information}

\item{\code{name}}{Name of the feature}

\item{\code{numRows}}{Number of rows in the sample used to determine frequent values}

\item{\code{projectId}}{Project Id}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrequentValuesResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FrequentValuesResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrequentValuesResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrequentValuesResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FrequentValuesResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrequentValuesResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrequentValuesResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FrequentValuesResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrequentValuesResponse$fromJSON(
  FrequentValuesResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FrequentValuesResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrequentValuesResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FrequentValuesResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrequentValuesResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
