% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frozen_model_create.R
\docType{class}
\name{FrozenModelCreate}
\alias{FrozenModelCreate}
\title{FrozenModelCreate}
\format{
An \code{R6Class} generator object
}
\description{
FrozenModelCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{modelId}}{character the ID of an existing model to use as a source of training parameters.}

\item{\code{nClusters}}{integer \link{optional} The number of clusters to use in the specified unsupervised clustering model. Only valid in unsupervised clustering projects.}

\item{\code{samplePct}}{numeric \link{optional} the percentage of the dataset to use with the model. Only one of `samplePct` and `trainingRowCount` should be specified. The specified percentage should be between 0.0 and 100.0.}

\item{\code{trainingRowCount}}{integer \link{optional} the integer number of rows of the dataset to use with the model. Only one of `samplePct` and `trainingRowCount` should be specified.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FrozenModelCreate-new}{\code{FrozenModelCreate$new()}}
\item \href{#method-FrozenModelCreate-validate}{\code{FrozenModelCreate$validate()}}
\item \href{#method-FrozenModelCreate-toJSON}{\code{FrozenModelCreate$toJSON()}}
\item \href{#method-FrozenModelCreate-fromJSON}{\code{FrozenModelCreate$fromJSON()}}
\item \href{#method-FrozenModelCreate-clone}{\code{FrozenModelCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrozenModelCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-FrozenModelCreate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrozenModelCreate$new(
  modelId = NULL,
  nClusters = NULL,
  samplePct = NULL,
  trainingRowCount = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{modelId}}{the ID of an existing model to use as a source of training parameters.}

\item{\code{nClusters}}{The number of clusters to use in the specified unsupervised clustering model. Only valid in unsupervised clustering projects.}

\item{\code{samplePct}}{the percentage of the dataset to use with the model. Only one of `samplePct` and `trainingRowCount` should be specified. The specified percentage should be between 0.0 and 100.0.}

\item{\code{trainingRowCount}}{the integer number of rows of the dataset to use with the model. Only one of `samplePct` and `trainingRowCount` should be specified.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrozenModelCreate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FrozenModelCreate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrozenModelCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrozenModelCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FrozenModelCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrozenModelCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrozenModelCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FrozenModelCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrozenModelCreate$fromJSON(FrozenModelCreateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FrozenModelCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrozenModelCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FrozenModelCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrozenModelCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
