% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frozen_model_retrieve_response.R
\docType{class}
\name{FrozenModelRetrieveResponse}
\alias{FrozenModelRetrieveResponse}
\title{FrozenModelRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
FrozenModelRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{blueprintId}}{character The blueprint used to construct the model - note this is not an ObjectId.}

\item{\code{featurelistId}}{character The ID of the featurelist used by the model.}

\item{\code{featurelistName}}{character The name of the featurelist used by the model.}

\item{\code{id}}{character The ID of the model.}

\item{\code{isFrozen}}{character When True, indicates that the model is frozen, i.e., uses tuning parameters from a parent model}

\item{\code{isStarred}}{character \link{optional} Whether the model has been starred.}

\item{\code{metrics}}{object The performance of the model according to various metrics.}

\item{\code{modelCategory}}{character Indicates what kind of model it is - will be `prime` for DataRobot Prime models, `blend` for blender models, and `model` for all other models.}

\item{\code{modelNumber}}{integer The model number from the leaderboard.}

\item{\code{modelType}}{character Identifies the model, e.g. `Nystroem Kernel SVM Regressor`.}

\item{\code{monotonicDecreasingFeaturelistId}}{character The ID of the featurelist that defines the set of features with a monotonically decreasing relationship to the target. If null, no such constraints are enforced.}

\item{\code{monotonicIncreasingFeaturelistId}}{character The ID of the featurelist that defines the set of features with a monotonically increasing relationship to the target. If null, no such constraints are enforced.}

\item{\code{parentModelId}}{character The model that tuning parameters are derived from.}

\item{\code{predictionThreshold}}{numeric Threshold used for binary classification in predictions.}

\item{\code{predictionThresholdReadOnly}}{character Indicates whether modification of a predictions threshold is forbidden. Since v2.22 threshold modification is allowed.}

\item{\code{processes}}{list( character ) List of processes used by the model.}

\item{\code{projectId}}{character The ID of the project to which the model belongs.}

\item{\code{samplePct}}{numeric The percentage of the dataset used in training the model.}

\item{\code{supportsComposableMl}}{character Indicates whether this blueprint is supported in Composable ML.}

\item{\code{supportsMonotonicConstraints}}{character When True, the model supports enforcing montonic constraints}

\item{\code{trainingDuration}}{character The duration spanned by the dates in the partition column for the data used to train the model.}

\item{\code{trainingEndDate}}{character The end date of the dates in the partition column for the data used to train the model.}

\item{\code{trainingRowCount}}{integer The number of rows used to train the model.}

\item{\code{trainingStartDate}}{character The start date of the dates in the partition column for the data used to train the model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FrozenModelRetrieveResponse$new()}}
\item \href{#method-validate}{\code{FrozenModelRetrieveResponse$validate()}}
\item \href{#method-toJSON}{\code{FrozenModelRetrieveResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{FrozenModelRetrieveResponse$fromJSON()}}
\item \href{#method-clone}{\code{FrozenModelRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrozenModelRetrieveResponse$new(
  blueprintId = NULL,
  featurelistId = NULL,
  featurelistName = NULL,
  id = NULL,
  isFrozen = NULL,
  metrics = NULL,
  modelCategory = NULL,
  modelNumber = NULL,
  modelType = NULL,
  monotonicDecreasingFeaturelistId = NULL,
  monotonicIncreasingFeaturelistId = NULL,
  parentModelId = NULL,
  predictionThreshold = NULL,
  predictionThresholdReadOnly = NULL,
  processes = NULL,
  projectId = NULL,
  samplePct = NULL,
  supportsComposableMl = NULL,
  supportsMonotonicConstraints = NULL,
  trainingDuration = NULL,
  trainingEndDate = NULL,
  trainingRowCount = NULL,
  trainingStartDate = NULL,
  isStarred = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blueprintId}}{The blueprint used to construct the model - note this is not an ObjectId.}

\item{\code{featurelistId}}{The ID of the featurelist used by the model.}

\item{\code{featurelistName}}{The name of the featurelist used by the model.}

\item{\code{id}}{The ID of the model.}

\item{\code{isFrozen}}{When True, indicates that the model is frozen, i.e., uses tuning parameters from a parent model}

\item{\code{metrics}}{The performance of the model according to various metrics.}

\item{\code{modelCategory}}{Indicates what kind of model it is - will be `prime` for DataRobot Prime models, `blend` for blender models, and `model` for all other models.}

\item{\code{modelNumber}}{The model number from the leaderboard.}

\item{\code{modelType}}{Identifies the model, e.g. `Nystroem Kernel SVM Regressor`.}

\item{\code{monotonicDecreasingFeaturelistId}}{The ID of the featurelist that defines the set of features with a monotonically decreasing relationship to the target. If null, no such constraints are enforced.}

\item{\code{monotonicIncreasingFeaturelistId}}{The ID of the featurelist that defines the set of features with a monotonically increasing relationship to the target. If null, no such constraints are enforced.}

\item{\code{parentModelId}}{The model that tuning parameters are derived from.}

\item{\code{predictionThreshold}}{Threshold used for binary classification in predictions.}

\item{\code{predictionThresholdReadOnly}}{Indicates whether modification of a predictions threshold is forbidden. Since v2.22 threshold modification is allowed.}

\item{\code{processes}}{List of processes used by the model.}

\item{\code{projectId}}{The ID of the project to which the model belongs.}

\item{\code{samplePct}}{The percentage of the dataset used in training the model.}

\item{\code{supportsComposableMl}}{Indicates whether this blueprint is supported in Composable ML.}

\item{\code{supportsMonotonicConstraints}}{When True, the model supports enforcing montonic constraints}

\item{\code{trainingDuration}}{The duration spanned by the dates in the partition column for the data used to train the model.}

\item{\code{trainingEndDate}}{The end date of the dates in the partition column for the data used to train the model.}

\item{\code{trainingRowCount}}{The number of rows used to train the model.}

\item{\code{trainingStartDate}}{The start date of the dates in the partition column for the data used to train the model.}

\item{\code{isStarred}}{Whether the model has been starred.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrozenModelRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrozenModelRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrozenModelRetrieveResponse$fromJSON(
  FrozenModelRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FrozenModelRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrozenModelRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
