% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_dataset_details_response.R
\docType{class}
\name{FullDatasetDetailsResponse}
\alias{FullDatasetDetailsResponse}
\title{FullDatasetDetailsResponse}
\format{
An \code{R6Class} generator object
}
\description{
FullDatasetDetailsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{categories}}{list( character ) An array of strings describing the intended use of the dataset.}

\item{\code{createdBy}}{character Username of the user who created the dataset.}

\item{\code{creationDate}}{character The date when the dataset was created.}

\item{\code{dataEngineQueryId}}{character ID of the source data engine query.}

\item{\code{dataPersisted}}{character If true, user is allowed to view extended data profile (which includes data statistics like min/max/median/mean, histogram, etc.) and download data. If false, download is not allowed and only the data schema (feature names and types) will be available.}

\item{\code{dataSourceId}}{character ID of the datasource used as the source of the dataset.}

\item{\code{dataSourceType}}{character The type of the datasource that was used as the source of the dataset.}

\item{\code{datasetId}}{character The ID of this dataset.}

\item{\code{datasetSize}}{integer The size of the dataset as a CSV in bytes.}

\item{\code{description}}{character The description of the dataset.}

\item{\code{eda1ModificationDate}}{character The ISO 8601 formatted date and time when the EDA1 for the dataset was updated.}

\item{\code{eda1ModifierFullName}}{character The user who was the last to update EDA1 for the dataset.}

\item{\code{entityCountByType}}{\link{EntityCountByTypeResponse} \link{optional}}

\item{\code{error}}{character Details of exception raised during ingestion process, if any.}

\item{\code{featureCount}}{integer Total number of features in the dataset.}

\item{\code{featureCountByType}}{list( \link{FeatureCountByTypeResponse} ) Number of features in the dataset grouped by feature type.}

\item{\code{isDataEngineEligible}}{character Whether this dataset can be a data source of a data engine query.}

\item{\code{isLatestVersion}}{character Whether this dataset version is the latest version of this dataset.}

\item{\code{isSnapshot}}{character Whether the dataset is an immutable snapshot of data which has previously been retrieved and saved to DataRobot.}

\item{\code{lastModificationDate}}{character The ISO 8601 formatted date and time when the dataset was last modified.}

\item{\code{lastModifierFullName}}{character Full name of user who was the last to modify the dataset.}

\item{\code{name}}{character The name of this dataset in the catalog.}

\item{\code{processingState}}{character Current ingestion process state of dataset.}

\item{\code{rowCount}}{integer The number of rows in the dataset.}

\item{\code{sampleSize}}{\link{SampleSize} \link{optional}}

\item{\code{tags}}{list( character ) List of tags attached to the item.}

\item{\code{timeSeriesProperties}}{\link{TimeSeriesProperties}}

\item{\code{uri}}{character The URI to datasource. For example, `file_name.csv`, or `jdbc:DATA_SOURCE_GIVEN_NAME/SCHEMA.TABLE_NAME`, or `jdbc:DATA_SOURCE_GIVEN_NAME/<query>` for `query` based datasources, or`https://s3.amazonaws.com/datarobot_test/kickcars-sample-200.csv`, etc.}

\item{\code{versionId}}{character The object ID of the catalog_version the dataset belongs to.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FullDatasetDetailsResponse$new()}}
\item \href{#method-validate}{\code{FullDatasetDetailsResponse$validate()}}
\item \href{#method-toJSON}{\code{FullDatasetDetailsResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{FullDatasetDetailsResponse$fromJSON()}}
\item \href{#method-clone}{\code{FullDatasetDetailsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullDatasetDetailsResponse$new(
  categories = NULL,
  createdBy = NULL,
  creationDate = NULL,
  dataEngineQueryId = NULL,
  dataPersisted = NULL,
  dataSourceId = NULL,
  dataSourceType = NULL,
  datasetId = NULL,
  datasetSize = NULL,
  description = NULL,
  eda1ModificationDate = NULL,
  eda1ModifierFullName = NULL,
  error = NULL,
  featureCount = NULL,
  featureCountByType = NULL,
  isDataEngineEligible = NULL,
  isLatestVersion = NULL,
  isSnapshot = NULL,
  lastModificationDate = NULL,
  lastModifierFullName = NULL,
  name = NULL,
  processingState = NULL,
  rowCount = NULL,
  tags = NULL,
  timeSeriesProperties = NULL,
  uri = NULL,
  versionId = NULL,
  entityCountByType = NULL,
  sampleSize = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{categories}}{An array of strings describing the intended use of the dataset.}

\item{\code{createdBy}}{Username of the user who created the dataset.}

\item{\code{creationDate}}{The date when the dataset was created.}

\item{\code{dataEngineQueryId}}{ID of the source data engine query.}

\item{\code{dataPersisted}}{If true, user is allowed to view extended data profile (which includes data statistics like min/max/median/mean, histogram, etc.) and download data. If false, download is not allowed and only the data schema (feature names and types) will be available.}

\item{\code{dataSourceId}}{ID of the datasource used as the source of the dataset.}

\item{\code{dataSourceType}}{The type of the datasource that was used as the source of the dataset.}

\item{\code{datasetId}}{The ID of this dataset.}

\item{\code{datasetSize}}{The size of the dataset as a CSV in bytes.}

\item{\code{description}}{The description of the dataset.}

\item{\code{eda1ModificationDate}}{The ISO 8601 formatted date and time when the EDA1 for the dataset was updated.}

\item{\code{eda1ModifierFullName}}{The user who was the last to update EDA1 for the dataset.}

\item{\code{error}}{Details of exception raised during ingestion process, if any.}

\item{\code{featureCount}}{Total number of features in the dataset.}

\item{\code{featureCountByType}}{Number of features in the dataset grouped by feature type.}

\item{\code{isDataEngineEligible}}{Whether this dataset can be a data source of a data engine query.}

\item{\code{isLatestVersion}}{Whether this dataset version is the latest version of this dataset.}

\item{\code{isSnapshot}}{Whether the dataset is an immutable snapshot of data which has previously been retrieved and saved to DataRobot.}

\item{\code{lastModificationDate}}{The ISO 8601 formatted date and time when the dataset was last modified.}

\item{\code{lastModifierFullName}}{Full name of user who was the last to modify the dataset.}

\item{\code{name}}{The name of this dataset in the catalog.}

\item{\code{processingState}}{Current ingestion process state of dataset.}

\item{\code{rowCount}}{The number of rows in the dataset.}

\item{\code{tags}}{List of tags attached to the item.}

\item{\code{uri}}{The URI to datasource. For example, `file_name.csv`, or `jdbc:DATA_SOURCE_GIVEN_NAME/SCHEMA.TABLE_NAME`, or `jdbc:DATA_SOURCE_GIVEN_NAME/<query>` for `query` based datasources, or`https://s3.amazonaws.com/datarobot_test/kickcars-sample-200.csv`, etc.}

\item{\code{versionId}}{The object ID of the catalog_version the dataset belongs to.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullDatasetDetailsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullDatasetDetailsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullDatasetDetailsResponse$fromJSON(
  FullDatasetDetailsResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FullDatasetDetailsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullDatasetDetailsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
