% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp_key.R
\docType{class}
\name{GCPKey}
\alias{GCPKey}
\title{GCPKey}
\format{
An \code{R6Class} generator object
}
\description{
GCPKey Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{authProviderX509CertUrl}}{character \link{optional} Auth provider X509 certificate URL.}

\item{\code{authUri}}{character \link{optional} Auth URI.}

\item{\code{clientEmail}}{character \link{optional} Client email address.}

\item{\code{clientId}}{character \link{optional} Client ID.}

\item{\code{clientX509CertUrl}}{character \link{optional} Client X509 certificate URL.}

\item{\code{privateKey}}{character \link{optional} Private key.}

\item{\code{privateKeyId}}{character \link{optional} Private key ID}

\item{\code{projectId}}{character \link{optional} Project ID.}

\item{\code{tokenUri}}{character \link{optional} Token URI.}

\item{\code{type}}{character GCP account type.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GCPKey$new()}}
\item \href{#method-validate}{\code{GCPKey$validate()}}
\item \href{#method-toJSON}{\code{GCPKey$toJSON()}}
\item \href{#method-fromJSON}{\code{GCPKey$fromJSON()}}
\item \href{#method-clone}{\code{GCPKey$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCPKey$new(
  type = NULL,
  authProviderX509CertUrl = NULL,
  authUri = NULL,
  clientEmail = NULL,
  clientId = NULL,
  clientX509CertUrl = NULL,
  privateKey = NULL,
  privateKeyId = NULL,
  projectId = NULL,
  tokenUri = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{GCP account type.}

\item{\code{authProviderX509CertUrl}}{Auth provider X509 certificate URL.}

\item{\code{authUri}}{Auth URI.}

\item{\code{clientEmail}}{Client email address.}

\item{\code{clientId}}{Client ID.}

\item{\code{clientX509CertUrl}}{Client X509 certificate URL.}

\item{\code{privateKey}}{Private key.}

\item{\code{privateKeyId}}{Private key ID}

\item{\code{projectId}}{Project ID.}

\item{\code{tokenUri}}{Token URI.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCPKey$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCPKey$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCPKey$fromJSON(GCPKeyJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{GCPKeyJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCPKey$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
