% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_metadata.R
\docType{class}
\name{GeneratedMetadata}
\alias{GeneratedMetadata}
\title{GeneratedMetadata}
\format{
An \code{R6Class} generator object
}
\description{
GeneratedMetadata Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{outputColumns}}{list( \link{array[character]} ) A list of column lists that are associated with the output datasets from the custom model conversion process.}

\item{\code{outputDatasets}}{list( character ) A list of output datasets from the custom model conversion process.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeneratedMetadata-new}{\code{GeneratedMetadata$new()}}
\item \href{#method-GeneratedMetadata-validate}{\code{GeneratedMetadata$validate()}}
\item \href{#method-GeneratedMetadata-toJSON}{\code{GeneratedMetadata$toJSON()}}
\item \href{#method-GeneratedMetadata-fromJSON}{\code{GeneratedMetadata$fromJSON()}}
\item \href{#method-GeneratedMetadata-clone}{\code{GeneratedMetadata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneratedMetadata-new"></a>}}
\if{latex}{\out{\hypertarget{method-GeneratedMetadata-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneratedMetadata$new(
  outputColumns = NULL,
  outputDatasets = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outputColumns}}{A list of column lists that are associated with the output datasets from the custom model conversion process.}

\item{\code{outputDatasets}}{A list of output datasets from the custom model conversion process.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneratedMetadata-validate"></a>}}
\if{latex}{\out{\hypertarget{method-GeneratedMetadata-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneratedMetadata$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneratedMetadata-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GeneratedMetadata-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneratedMetadata$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneratedMetadata-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GeneratedMetadata-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneratedMetadata$fromJSON(GeneratedMetadataJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{GeneratedMetadataJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneratedMetadata-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeneratedMetadata-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneratedMetadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
