% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator_settings.R
\docType{class}
\name{GeneratorSettings}
\alias{GeneratorSettings}
\title{GeneratorSettings}
\format{
An \code{R6Class} generator object
}
\description{
GeneratorSettings Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datetimePartitionColumn}}{character The date column that will be used as a datetime partition column in time series project.}

\item{\code{defaultCategoricalAggregationMethod}}{character Default aggregation method used for categorical feature.}

\item{\code{defaultNumericAggregationMethod}}{character Default aggregation method used for numeric feature.}

\item{\code{defaultTextAggregationMethod}}{character \link{optional} Default aggregation method used for text feature.}

\item{\code{endToSeriesMaxDatetime}}{character \link{optional} A boolean value indicating whether generates post-aggregated series up to series maximum datetime or global maximum datetime.}

\item{\code{multiseriesIdColumns}}{list( character ) \link{optional} An array with the names of columns identifying the series to which row of the output dataset belongs. Currently, only one multiseries ID column is supported.}

\item{\code{startFromSeriesMinDatetime}}{character \link{optional} A boolean value indicating whether post-aggregated series starts from series minimum datetime or global minimum datetime.}

\item{\code{target}}{character \link{optional} The name of target for the output dataset.}

\item{\code{timeStep}}{integer Number of time steps for the output dataset.}

\item{\code{timeUnit}}{character Indicates which unit is a basis for time steps of the output dataset.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GeneratorSettings$new()}}
\item \href{#method-validate}{\code{GeneratorSettings$validate()}}
\item \href{#method-toJSON}{\code{GeneratorSettings$toJSON()}}
\item \href{#method-fromJSON}{\code{GeneratorSettings$fromJSON()}}
\item \href{#method-clone}{\code{GeneratorSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneratorSettings$new(
  datetimePartitionColumn = NULL,
  defaultCategoricalAggregationMethod = NULL,
  defaultNumericAggregationMethod = NULL,
  timeStep = NULL,
  timeUnit = NULL,
  defaultTextAggregationMethod = NULL,
  endToSeriesMaxDatetime = TRUE,
  multiseriesIdColumns = NULL,
  startFromSeriesMinDatetime = TRUE,
  target = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datetimePartitionColumn}}{The date column that will be used as a datetime partition column in time series project.}

\item{\code{defaultCategoricalAggregationMethod}}{Default aggregation method used for categorical feature.}

\item{\code{defaultNumericAggregationMethod}}{Default aggregation method used for numeric feature.}

\item{\code{timeStep}}{Number of time steps for the output dataset.}

\item{\code{timeUnit}}{Indicates which unit is a basis for time steps of the output dataset.}

\item{\code{defaultTextAggregationMethod}}{Default aggregation method used for text feature.}

\item{\code{endToSeriesMaxDatetime}}{A boolean value indicating whether generates post-aggregated series up to series maximum datetime or global maximum datetime.}

\item{\code{multiseriesIdColumns}}{An array with the names of columns identifying the series to which row of the output dataset belongs. Currently, only one multiseries ID column is supported.}

\item{\code{startFromSeriesMinDatetime}}{A boolean value indicating whether post-aggregated series starts from series minimum datetime or global minimum datetime.}

\item{\code{target}}{The name of target for the output dataset.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneratorSettings$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneratorSettings$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneratorSettings$fromJSON(GeneratorSettingsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{GeneratorSettingsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneratorSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
