% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_feature_plot_retrieve_response.R
\docType{class}
\name{GeometryFeaturePlotRetrieveResponse}
\alias{GeometryFeaturePlotRetrieveResponse}
\title{GeometryFeaturePlotRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
GeometryFeaturePlotRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{feature}}{character Name of location feature to plot on map.}

\item{\code{plotData}}{\link{GeoFeaturePlotData}}

\item{\code{projectId}}{character The project to select a location feature from.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeometryFeaturePlotRetrieveResponse-new}{\code{GeometryFeaturePlotRetrieveResponse$new()}}
\item \href{#method-GeometryFeaturePlotRetrieveResponse-validate}{\code{GeometryFeaturePlotRetrieveResponse$validate()}}
\item \href{#method-GeometryFeaturePlotRetrieveResponse-toJSON}{\code{GeometryFeaturePlotRetrieveResponse$toJSON()}}
\item \href{#method-GeometryFeaturePlotRetrieveResponse-fromJSON}{\code{GeometryFeaturePlotRetrieveResponse$fromJSON()}}
\item \href{#method-GeometryFeaturePlotRetrieveResponse-clone}{\code{GeometryFeaturePlotRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeometryFeaturePlotRetrieveResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-GeometryFeaturePlotRetrieveResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeometryFeaturePlotRetrieveResponse$new(
  feature = NULL,
  plotData = NULL,
  projectId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{Name of location feature to plot on map.}

\item{\code{projectId}}{The project to select a location feature from.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeometryFeaturePlotRetrieveResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-GeometryFeaturePlotRetrieveResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeometryFeaturePlotRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeometryFeaturePlotRetrieveResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GeometryFeaturePlotRetrieveResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeometryFeaturePlotRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeometryFeaturePlotRetrieveResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GeometryFeaturePlotRetrieveResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeometryFeaturePlotRetrieveResponse$fromJSON(
  GeometryFeaturePlotRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{GeometryFeaturePlotRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeometryFeaturePlotRetrieveResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeometryFeaturePlotRetrieveResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeometryFeaturePlotRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
