% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial_insights.R
\docType{class}
\name{GeospatialInsights}
\alias{GeospatialInsights}
\title{GeospatialInsights}
\format{
An \code{R6Class} generator object
}
\description{
GeospatialInsights Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{insightName}}{character Insight name.}

\item{\code{perCluster}}{list( \link{PerClusterGeospatial} ) Statistic for different feature values in this cluster.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeospatialInsights-new}{\code{GeospatialInsights$new()}}
\item \href{#method-GeospatialInsights-validate}{\code{GeospatialInsights$validate()}}
\item \href{#method-GeospatialInsights-toJSON}{\code{GeospatialInsights$toJSON()}}
\item \href{#method-GeospatialInsights-fromJSON}{\code{GeospatialInsights$fromJSON()}}
\item \href{#method-GeospatialInsights-clone}{\code{GeospatialInsights$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeospatialInsights-new"></a>}}
\if{latex}{\out{\hypertarget{method-GeospatialInsights-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeospatialInsights$new(
  insightName = NULL,
  perCluster = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{insightName}}{Insight name.}

\item{\code{perCluster}}{Statistic for different feature values in this cluster.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeospatialInsights-validate"></a>}}
\if{latex}{\out{\hypertarget{method-GeospatialInsights-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeospatialInsights$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeospatialInsights-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GeospatialInsights-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeospatialInsights$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeospatialInsights-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GeospatialInsights-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeospatialInsights$fromJSON(GeospatialInsightsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{GeospatialInsightsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeospatialInsights-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeospatialInsights-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeospatialInsights$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
