% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{GetAnomalyAssessmentExplanations}
\alias{GetAnomalyAssessmentExplanations}
\title{GetAnomalyAssessmentExplanations}
\usage{
GetAnomalyAssessmentExplanations(
  projectId,
  recordId,
  startDate = NULL,
  endDate = NULL,
  pointsCount = NULL,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve anomaly assessment record.
}
\details{
Retrieve anomaly assessment record. Two out of three parameters: startDate, endDate or pointsCount must be specified.

This method invokes \verb{GET /projects/\{projectId\}/anomalyAssessmentRecords/\{recordId\}/explanations/} in the DataRobot Public API.
}
\seealso{
The method ProjectsAnomalyAssessmentRecordsExplanationsList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
