% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{GetAnomalyAssessmentPredictionsPreview}
\alias{GetAnomalyAssessmentPredictionsPreview}
\title{GetAnomalyAssessmentPredictionsPreview}
\usage{
GetAnomalyAssessmentPredictionsPreview(projectId, recordId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve predictions preview for the anomaly assessment record.
}
\details{
Retrieve predictions preview for the anomaly assessment record.

This method invokes \verb{GET /projects/\{projectId\}/anomalyAssessmentRecords/\{recordId\}/predictionsPreview/} in the DataRobot Public API.
}
\seealso{
The method ProjectsAnomalyAssessmentRecordsPredictionsPreviewList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
