% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{GetBlenderModel}
\alias{GetBlenderModel}
\title{GetBlenderModel}
\usage{
GetBlenderModel(project, modelId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a blender.
}
\details{
Retrieve a blender. Blenders are a special type of models, so the response includes all attributes that would be in a response to :\url{http:get}:/api/v2/projects/(projectId)/models/(modelId)/ as well as some additional ones.

This method invokes \verb{GET /projects/\{projectId\}/blenderModels/\{modelId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsBlenderModelsRetrieve in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
