% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints_apiWrapper.R
\name{GetBlueprint}
\alias{GetBlueprint}
\title{GetBlueprint}
\usage{
GetBlueprint(project, blueprintId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a blueprint by its ID.
}
\details{
Retrieve a blueprint by its ID.

This method invokes \verb{GET /projects/\{projectId\}/blueprints/\{blueprintId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsBlueprintsRetrieve in \link{BlueprintsApi}, which invokes the same underlying API endpoint.
}
