% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints_apiWrapper.R
\name{GetBlueprintChart}
\alias{GetBlueprintChart}
\title{GetBlueprintChart}
\usage{
GetBlueprintChart(project, blueprintId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a blueprint chart by blueprint id.
}
\details{
Retrieve a blueprint chart by blueprint id.

This method invokes \verb{GET /projects/\{projectId\}/blueprints/\{blueprintId\}/blueprintChart/} in the DataRobot Public API.
}
\seealso{
The method ProjectsBlueprintsBlueprintChartList in \link{BlueprintsApi}, which invokes the same underlying API endpoint.
}
