% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints_apiWrapper.R
\name{GetBlueprintDocumentation}
\alias{GetBlueprintDocumentation}
\title{GetBlueprintDocumentation}
\usage{
GetBlueprintDocumentation(project, blueprintId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve blueprint tasks documentation.
}
\details{
Retrieve blueprint tasks documentation.

This method invokes \verb{GET /projects/\{projectId\}/blueprints/\{blueprintId\}/blueprintDocs/} in the DataRobot Public API.
}
\seealso{
The method ProjectsBlueprintsBlueprintDocsList in \link{BlueprintsApi}, which invokes the same underlying API endpoint.
}
