% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{GetCalendar}
\alias{GetCalendar}
\title{GetCalendar}
\usage{
GetCalendar(calendarId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve information about a calendar.
}
\details{
List all the information about a calendar such as the total number of event dates, the earliest calendar event date, the IDs of projects currently using this calendar and the others.

This method invokes \verb{GET /calendars/\{calendarId\}/} in the DataRobot Public API.
}
\seealso{
The method CalendarsRetrieve in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
