% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{GetDataSource}
\alias{GetDataSource}
\title{GetDataSource}
\usage{
GetDataSource(dataSourceId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Data source details.
}
\details{
Return details of the existing data source with given ID, including SQL query or selected table and schema data, which fully describe which data to extract and from which location.

This method invokes \verb{GET /externalDataSources/\{dataSourceId\}/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataSourcesRetrieve in \link{DataConnectivityApi}, which invokes the same underlying API endpoint.
}
