% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{GetDataStore}
\alias{GetDataStore}
\title{GetDataStore}
\usage{
GetDataStore(dataStoreId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Data store details.
}
\details{
A configured connection to a database - it has a name and a specified driver. The driver may be specified by a JDBC URL or connection parameters if the driver was created with the parameter configuration.

This method invokes \verb{GET /externalDataStores/\{dataStoreId\}/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataStoresRetrieve in \link{DataConnectivityApi}, which invokes the same underlying API endpoint.
}
