% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{GetDataStoreSchemas}
\alias{GetDataStoreSchemas}
\title{GetDataStoreSchemas}
\usage{
GetDataStoreSchemas(dataStoreId, username, password, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieves a data store's data schemas.
}
\details{
Retrieves a data store's data schemas.

This method invokes \verb{POST /externalDataStores/\{dataStoreId\}/schemas/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataStoresSchemasCreate in \link{DataConnectivityApi}, which invokes the same underlying API endpoint.
}
