% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{GetDataStoreTables}
\alias{GetDataStoreTables}
\title{GetDataStoreTables}
\usage{
GetDataStoreTables(dataStoreId, username, password, schema = NULL, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieves a data store's database tables (including views).
}
\details{
Retrieves a data store's database tables (including views).

This method invokes \verb{POST /externalDataStores/\{dataStoreId\}/tables/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataStoresTablesCreate in \link{DataConnectivityApi}, which invokes the same underlying API endpoint.
}
