% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{GetDatetimeModel}
\alias{GetDatetimeModel}
\title{GetDatetimeModel}
\usage{
GetDatetimeModel(project, modelId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get datetime model
}
\details{
Look up a particular datetime model All durations and datetimes are specified in accordance with \verb{timestamp and duration formatting rules <time_format>}.

This method invokes \verb{GET /projects/\{projectId\}/datetimeModels/\{modelId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsDatetimeModelsRetrieve in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
