% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{GetDeployment}
\alias{GetDeployment}
\title{GetDeployment}
\usage{
GetDeployment(deploymentId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve deployment
}
\details{
Retrieve a deployment.

This method invokes \verb{GET /deployments/\{deploymentId\}/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsRetrieve in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
