% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{GetDeploymentServiceStats}
\alias{GetDeploymentServiceStats}
\title{GetDeploymentServiceStats}
\usage{
GetDeploymentServiceStats(
  deploymentId,
  modelId = NULL,
  start = NULL,
  end = NULL,
  executionTimeQuantile = NULL,
  responseTimeQuantile = NULL,
  slowRequestsThreshold = NULL,
  segmentAttribute = NULL,
  segmentValue = NULL,
  bucketSize = NULL,
  monitoringType = "serviceHealth",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve service statistics
}
\details{
Retrieve deployment service health statistics overview.

This method invokes \verb{GET /deployments/\{deploymentId\}/serviceStats/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsServiceStatsList in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
