% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{GetDeploymentSettings}
\alias{GetDeploymentSettings}
\title{GetDeploymentSettings}
\usage{
GetDeploymentSettings(deployment, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get deployment settings
}
\details{
Retrieve deployment settings.

This method invokes \verb{GET /deployments/\{deploymentId\}/settings/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsSettingsList in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
