% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{GetDriver}
\alias{GetDriver}
\title{GetDriver}
\usage{
GetDriver(driverId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve driver details.
}
\details{
Retrieve driver details by ID.

This method invokes \verb{GET /externalDataDrivers/\{driverId\}/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataDriversRetrieve in \link{DataConnectivityApi}, which invokes the same underlying API endpoint.
}
