% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{GetFeatureDiscoveryRelationships}
\alias{GetFeatureDiscoveryRelationships}
\title{GetFeatureDiscoveryRelationships}
\usage{
GetFeatureDiscoveryRelationships(project, configId = NULL, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve relationships configuration for a project
}
\details{
Retrieve relationships configuration for a project

This method invokes \verb{GET /projects/\{projectId\}/relationshipsConfiguration/} in the DataRobot Public API.
}
\seealso{
The method RelationshipsConfigurationList in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
