% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{GetFeatureHistogram}
\alias{GetFeatureHistogram}
\title{GetFeatureHistogram}
\usage{
GetFeatureHistogram(project, featureName, binLimit = NULL, key = NULL, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get feature histogram
}
\details{
Get histogram chart data for a specific feature. Information that can be used to build histogram charts. Plot data returned is based on raw data that is calculated during initial project creation and updated after the project's target variable has been selected. The number of bins in the histogram is no greater than the requested limit.

This method invokes \verb{GET /projects/\{projectId\}/featureHistograms/\{featureName\}/} in the DataRobot Public API.
}
\seealso{
The method FeatureHistogramsRetrieve in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
