% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{GetFeaturelist}
\alias{GetFeaturelist}
\title{GetFeaturelist}
\usage{
GetFeaturelist(project, featurelistId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a feature list
}
\details{
Retrieve a single known feature list.

This method invokes \verb{GET /projects/\{projectId\}/featurelists/\{featurelistId\}/} in the DataRobot Public API.
}
\seealso{
The method FeaturelistsRetrieve in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
