% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{GetJob}
\alias{GetJob}
\title{GetJob}
\usage{
GetJob(project, jobId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get job
}
\details{
Retrieve details for a job that has been started but has not yet completed.

This method invokes \verb{GET /projects/\{projectId\}/jobs/\{jobId\}/} in the DataRobot Public API.
}
\seealso{
The method JobsRetrieve in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
