% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{GetMissingValuesReport}
\alias{GetMissingValuesReport}
\title{GetMissingValuesReport}
\usage{
GetMissingValuesReport(project, modelId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a summary of how the model's subtasks handle missing values.
}
\details{
Retrieve a summary of how the model's subtasks handle missing values   Only models built after the missing value report feature was added will have reports, and only models with at least one imputation or encoding task, e.g. ordinal encoding, missing value imputation. Blenders and scaleout models do not support Missing Value reports.  The report will describe how each feature's missing values were treated, and report how many missing values were present in the training data. Features which were not processed by a given blueprint task will not mention it: for instance, a categorical feature with many unique values may not be considered eligible for processing by a One-Hot Encoding  Report is collected for those features which are considered eligible by given blueprint task. For instance, categorical feature with a lot of unique values may not be considered as eligible in One-Hot Encoding task.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/missingReport/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsMissingReportList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
