% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{GetModel}
\alias{GetModel}
\title{GetModel}
\usage{
GetModel(project, modelId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get model
}
\details{
Look up a particular model.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsRetrieve in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
