% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints_apiWrapper.R
\name{GetModelBlueprintChart}
\alias{GetModelBlueprintChart}
\title{GetModelBlueprintChart}
\usage{
GetModelBlueprintChart(project, modelId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a reduced model blueprint chart by model id.
}
\details{
Retrieve a reduced model blueprint chart by model id. The model blueprint charts are reduced from the full blueprint charts to show only those sections of the blueprint that were actually used in the model, given the selected featurelist.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/blueprintChart/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsBlueprintChartList in \link{BlueprintsApi}, which invokes the same underlying API endpoint.
}
