% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints_apiWrapper.R
\name{GetModelBlueprintDocumentation}
\alias{GetModelBlueprintDocumentation}
\title{GetModelBlueprintDocumentation}
\usage{
GetModelBlueprintDocumentation(project, modelId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve task documentation for a reduced model blueprint.
}
\details{
Retrieve task documentation for a reduced model blueprint. The model blueprint is reduced from the full blueprint to show only those sections of the blueprint that were actually used in the model, given the selected featurelist.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/blueprintDocs/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsBlueprintDocsList in \link{BlueprintsApi}, which invokes the same underlying API endpoint.
}
