% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{GetModelingFeaturelist}
\alias{GetModelingFeaturelist}
\title{GetModelingFeaturelist}
\usage{
GetModelingFeaturelist(project, featurelistId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a single modeling featurelist by ID
}
\details{
Retrieve a single modeling featurelist by ID. When reporting the number of models that \"use\" a featurelist, a model is considered to use a featurelist if it is used as to train the model or as a monotonic constraint featurelist, or if the model is a blender with component models that use the featurelist. This route will only become available after the target and partitioning options have been set for a project. Modeling featurelists are featurelists of modeling features, and are the appropriate featurelists to use when creating models or restarting the autopilot. In a time series project, these will be distinct from those returned from :\url{http:get}:/api/v2/projects/(projectId)/featurelists/ while in other projects these will be identical.

This method invokes \verb{GET /projects/\{projectId\}/modelingFeaturelists/\{featurelistId\}/} in the DataRobot Public API.
}
\seealso{
The method ModelingFeaturelistsRetrieve in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
