% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{GetModelJob}
\alias{GetModelJob}
\title{GetModelJob}
\usage{
GetModelJob(project, modelJobId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Look up a specific modeling job
}
\details{
Look up a particular modeling job

This method invokes \verb{GET /projects/\{projectId\}/modelJobs/\{jobId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelJobsRetrieve in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
