% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R,
%   R/predictions_apiWrapper.R
\name{GetPredictionExplanationsMetadata}
\alias{GetPredictionExplanationsMetadata}
\title{GetPredictionExplanationsMetadata}
\usage{
GetPredictionExplanationsMetadata(
  project,
  predictionExplanationId,
  datasetId = NULL,
  finishTime = NULL,
  id = NULL,
  maxExplanations = NULL,
  modelId = NULL,
  numColumns = NULL,
  predictionExplanationsLocation = NULL,
  predictionThreshold = NULL,
  thresholdHigh = NULL,
  thresholdLow = NULL,
  ...
)

GetPredictionExplanationsMetadata(
  project,
  predictionExplanationId,
  datasetId = NULL,
  finishTime = NULL,
  id = NULL,
  maxExplanations = NULL,
  modelId = NULL,
  numColumns = NULL,
  predictionExplanationsLocation = NULL,
  predictionThreshold = NULL,
  thresholdHigh = NULL,
  thresholdLow = NULL,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a PredictionExplanationsRecord object.

Retrieve a PredictionExplanationsRecord object.
}
\details{
Retrieve a PredictionExplanationsRecord object. A PredictionExplanationsRecord contains metadata about the computed prediction explanations and the location at which the PredictionExplanations can be retrieved.

This method invokes \verb{GET /projects/\{projectId\}/predictionExplanationsRecords/\{predictionExplanationsId\}/} in the DataRobot Public API.

Retrieve a PredictionExplanationsRecord object. A PredictionExplanationsRecord contains metadata about the computed prediction explanations and the location at which the PredictionExplanations can be retrieved.

This method invokes \verb{GET /projects/\{projectId\}/predictionExplanationsRecords/\{predictionExplanationsId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPredictionExplanationsRecordsRetrieve in \link{InsightsApi}, which invokes the same underlying API endpoint.

The method ProjectsPredictionExplanationsRecordsRetrieve in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
