% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{GetPredictJob}
\alias{GetPredictJob}
\title{GetPredictJob}
\usage{
GetPredictJob(project, predictJobId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Look up a particular prediction job
}
\details{
Look up a particular prediction job

This method invokes \verb{GET /projects/\{projectId\}/predictJobs/\{jobId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPredictJobsRetrieve in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
