% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{GetPredictJobs}
\alias{GetPredictJobs}
\title{GetPredictJobs}
\usage{
GetPredictJobs(project, status = NULL, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List all prediction jobs for a project
}
\details{
List all prediction jobs for a project

This method invokes \verb{GET /projects/\{projectId\}/predictJobs/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPredictJobsList in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
