% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{GetPrimeEligibility}
\alias{GetPrimeEligibility}
\title{GetPrimeEligibility}
\usage{
GetPrimeEligibility(project, modelId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Check a Model for Prime Eligibility
}
\details{
Check if a model can be approximated by DataRobot Prime

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/primeInfo/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsPrimeInfoList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
