% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{GetPrimeFile}
\alias{GetPrimeFile}
\title{GetPrimeFile}
\usage{
GetPrimeFile(project, primeFileId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve metadata about a DataRobot Prime file
}
\details{
Retrieve metadata about a DataRobot Prime file available for download

This method invokes \verb{GET /projects/\{projectId\}/primeFiles/\{primeFileId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPrimeFilesRetrieve in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
