% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{GetProject}
\alias{GetProject}
\title{GetProject}
\usage{
GetProject(project, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get project.
}
\details{
Look up a particular project

This method invokes \verb{GET /projects/\{projectId\}/} in the DataRobot Public API.
}
\seealso{
The method Retrieve in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
