% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{GetProjectStatus}
\alias{GetProjectStatus}
\title{GetProjectStatus}
\usage{
GetProjectStatus(project, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Check project status
}
\details{
Check the status of a project

This method invokes \verb{GET /projects/\{projectId\}/status/} in the DataRobot Public API.
}
\seealso{
The method StatusList in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
