% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{GetRatingTable}
\alias{GetRatingTable}
\title{GetRatingTable}
\usage{
GetRatingTable(project, ratingTableId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve Rating Table Information
}
\details{
Retrieves a rating table

This method invokes \verb{GET /projects/\{projectId\}/ratingTables/\{ratingTableId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsRatingTablesRetrieve in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
