% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{GetRatingTableModel}
\alias{GetRatingTableModel}
\title{GetRatingTableModel}
\usage{
GetRatingTableModel(project, modelId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve Rating Table Model
}
\details{
Look up a particular rating table model

This method invokes \verb{GET /projects/\{projectId\}/ratingTableModels/\{modelId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsRatingTableModelsRetrieve in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
