% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure_apiWrapper.R
\name{GetServerVersion}
\alias{GetServerVersion}
\title{GetServerVersion}
\usage{
GetServerVersion(...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve version information.
}
\details{
Provides the version information for the current version of the API.

This method invokes \verb{GET /version/} in the DataRobot Public API.
}
\seealso{
The method VersionList in \link{InfrastructureApi}, which invokes the same underlying API endpoint.
}
