% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{GetValidMetrics}
\alias{GetValidMetrics}
\title{GetValidMetrics}
\usage{
GetValidMetrics(project, target, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List feature metrics
}
\details{
List the appropriate metrics if a feature were chosen as the target. The metrics listed will include both weighted and unweighted metrics - which are appropriate will depend on whether a weights column is used.

This method invokes \verb{GET /projects/\{projectId\}/features/metrics/} in the DataRobot Public API.
}
\seealso{
The method FeaturesMetricsList in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
