% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{GetWordCloud}
\alias{GetWordCloud}
\title{GetWordCloud}
\usage{
GetWordCloud(project, modelId, excludeStopWords = FALSE, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve word cloud data for a model.
}
\details{
Retrieve word cloud data for a model. Not all models will have word cloud data available, even when they use text features.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/wordCloud/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsWordCloudList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
