% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grant_access_control_with_username.R
\docType{class}
\name{GrantAccessControlWithUsername}
\alias{GrantAccessControlWithUsername}
\title{GrantAccessControlWithUsername}
\format{
An \code{R6Class} generator object
}
\description{
GrantAccessControlWithUsername Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{role}}{character The role of the recipient on this entity. One of OWNER, USER, OBSERVER.}

\item{\code{shareRecipientType}}{character Describes the recipient type, either user, group, or organization.}

\item{\code{username}}{character Username of the user to update the access role for.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GrantAccessControlWithUsername-new}{\code{GrantAccessControlWithUsername$new()}}
\item \href{#method-GrantAccessControlWithUsername-validate}{\code{GrantAccessControlWithUsername$validate()}}
\item \href{#method-GrantAccessControlWithUsername-toJSON}{\code{GrantAccessControlWithUsername$toJSON()}}
\item \href{#method-GrantAccessControlWithUsername-fromJSON}{\code{GrantAccessControlWithUsername$fromJSON()}}
\item \href{#method-GrantAccessControlWithUsername-clone}{\code{GrantAccessControlWithUsername$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GrantAccessControlWithUsername-new"></a>}}
\if{latex}{\out{\hypertarget{method-GrantAccessControlWithUsername-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GrantAccessControlWithUsername$new(
  role = NULL,
  shareRecipientType = NULL,
  username = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{The role of the recipient on this entity. One of OWNER, USER, OBSERVER.}

\item{\code{shareRecipientType}}{Describes the recipient type, either user, group, or organization.}

\item{\code{username}}{Username of the user to update the access role for.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GrantAccessControlWithUsername-validate"></a>}}
\if{latex}{\out{\hypertarget{method-GrantAccessControlWithUsername-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GrantAccessControlWithUsername$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GrantAccessControlWithUsername-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GrantAccessControlWithUsername-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GrantAccessControlWithUsername$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GrantAccessControlWithUsername-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GrantAccessControlWithUsername-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GrantAccessControlWithUsername$fromJSON(
  GrantAccessControlWithUsernameJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{GrantAccessControlWithUsernameJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GrantAccessControlWithUsername-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GrantAccessControlWithUsername-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GrantAccessControlWithUsername$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
