% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grant_access_control_with_username_with_grant.R
\docType{class}
\name{GrantAccessControlWithUsernameWithGrant}
\alias{GrantAccessControlWithUsernameWithGrant}
\title{GrantAccessControlWithUsernameWithGrant}
\format{
An \code{R6Class} generator object
}
\description{
GrantAccessControlWithUsernameWithGrant Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{canShare}}{character \link{optional} Whether the org/group/user should be able to share with others.If true, the org/group/user will be able to grant any role up to and includingtheir own to other orgs/groups/user. If `role` is `NO_ROLE` `canShare` is ignored.}

\item{\code{role}}{character The role of the recipient on this entity. One of OWNER, USER, OBSERVER.}

\item{\code{shareRecipientType}}{character Describes the recipient type, either user, group, or organization.}

\item{\code{username}}{character Username of the user to update the access role for.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GrantAccessControlWithUsernameWithGrant$new()}}
\item \href{#method-validate}{\code{GrantAccessControlWithUsernameWithGrant$validate()}}
\item \href{#method-toJSON}{\code{GrantAccessControlWithUsernameWithGrant$toJSON()}}
\item \href{#method-fromJSON}{\code{GrantAccessControlWithUsernameWithGrant$fromJSON()}}
\item \href{#method-clone}{\code{GrantAccessControlWithUsernameWithGrant$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GrantAccessControlWithUsernameWithGrant$new(
  role = NULL,
  shareRecipientType = NULL,
  username = NULL,
  canShare = TRUE,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{The role of the recipient on this entity. One of OWNER, USER, OBSERVER.}

\item{\code{shareRecipientType}}{Describes the recipient type, either user, group, or organization.}

\item{\code{username}}{Username of the user to update the access role for.}

\item{\code{canShare}}{Whether the org/group/user should be able to share with others.If true, the org/group/user will be able to grant any role up to and includingtheir own to other orgs/groups/user. If `role` is `NO_ROLE` `canShare` is ignored.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GrantAccessControlWithUsernameWithGrant$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GrantAccessControlWithUsernameWithGrant$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GrantAccessControlWithUsernameWithGrant$fromJSON(
  GrantAccessControlWithUsernameWithGrantJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{GrantAccessControlWithUsernameWithGrantJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GrantAccessControlWithUsernameWithGrant$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
