% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfs_project_create.R
\docType{class}
\name{HdfsProjectCreate}
\alias{HdfsProjectCreate}
\title{HdfsProjectCreate}
\format{
An \code{R6Class} generator object
}
\description{
HdfsProjectCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{password}}{character \link{optional} Password for authenticating to HDFS using Kerberos. The password will be encrypted on the server side in scope of HTTP request and never saved or stored.}

\item{\code{port}}{integer \link{optional} Port of the WebHDFS Namenode server. If not specified, defaults to HDFS default port 50070.}

\item{\code{projectName}}{character \link{optional} Name of the project to be created. If not specified, project name will be based on the file name.}

\item{\code{url}}{character URL of the WebHDFS resource. Represent the file using the `hdfs://` protocol marker (for example,  `hdfs:///tmp/somedataset.csv`).}

\item{\code{user}}{character \link{optional} Username for authenticating to HDFS using Kerberos}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HdfsProjectCreate$new()}}
\item \href{#method-validate}{\code{HdfsProjectCreate$validate()}}
\item \href{#method-toJSON}{\code{HdfsProjectCreate$toJSON()}}
\item \href{#method-fromJSON}{\code{HdfsProjectCreate$fromJSON()}}
\item \href{#method-clone}{\code{HdfsProjectCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HdfsProjectCreate$new(
  url = NULL,
  password = NULL,
  port = NULL,
  projectName = NULL,
  user = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{URL of the WebHDFS resource. Represent the file using the `hdfs://` protocol marker (for example,  `hdfs:///tmp/somedataset.csv`).}

\item{\code{password}}{Password for authenticating to HDFS using Kerberos. The password will be encrypted on the server side in scope of HTTP request and never saved or stored.}

\item{\code{port}}{Port of the WebHDFS Namenode server. If not specified, defaults to HDFS default port 50070.}

\item{\code{projectName}}{Name of the project to be created. If not specified, project name will be based on the file name.}

\item{\code{user}}{Username for authenticating to HDFS using Kerberos}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HdfsProjectCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HdfsProjectCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HdfsProjectCreate$fromJSON(HdfsProjectCreateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{HdfsProjectCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HdfsProjectCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
