% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_output_adaptor.R
\docType{class}
\name{HttpOutputAdaptor}
\alias{HttpOutputAdaptor}
\title{HttpOutputAdaptor}
\format{
An \code{R6Class} generator object
}
\description{
HttpOutputAdaptor Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{headers}}{object \link{optional} Extra headers to send with the request}

\item{\code{method}}{character Method to use when saving the CSV file}

\item{\code{type}}{character Type name for this output type}

\item{\code{url}}{character URL for the CSV file}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HttpOutputAdaptor-new}{\code{HttpOutputAdaptor$new()}}
\item \href{#method-HttpOutputAdaptor-validate}{\code{HttpOutputAdaptor$validate()}}
\item \href{#method-HttpOutputAdaptor-toJSON}{\code{HttpOutputAdaptor$toJSON()}}
\item \href{#method-HttpOutputAdaptor-fromJSON}{\code{HttpOutputAdaptor$fromJSON()}}
\item \href{#method-HttpOutputAdaptor-clone}{\code{HttpOutputAdaptor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpOutputAdaptor-new"></a>}}
\if{latex}{\out{\hypertarget{method-HttpOutputAdaptor-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpOutputAdaptor$new(
  method = NULL,
  type = NULL,
  url = NULL,
  headers = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Method to use when saving the CSV file}

\item{\code{type}}{Type name for this output type}

\item{\code{url}}{URL for the CSV file}

\item{\code{headers}}{Extra headers to send with the request}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpOutputAdaptor-validate"></a>}}
\if{latex}{\out{\hypertarget{method-HttpOutputAdaptor-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpOutputAdaptor$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpOutputAdaptor-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-HttpOutputAdaptor-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpOutputAdaptor$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpOutputAdaptor-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-HttpOutputAdaptor-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpOutputAdaptor$fromJSON(HttpOutputAdaptorJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{HttpOutputAdaptorJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpOutputAdaptor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HttpOutputAdaptor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpOutputAdaptor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
