% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humility_stats_over_time_retrieve_response.R
\docType{class}
\name{HumilityStatsOverTimeRetrieveResponse}
\alias{HumilityStatsOverTimeRetrieveResponse}
\title{HumilityStatsOverTimeRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
HumilityStatsOverTimeRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{buckets}}{list( \link{HumilityStatsBucket} ) An array of `bucket` objects, representing service health stats of the deployment over time}

\item{\code{modelId}}{character \link{optional} The id of the model for which metrics are being retrieved.}

\item{\code{segmentAttribute}}{character \link{optional} The name of the segment on which segment analysis is being performed.}

\item{\code{segmentValue}}{character \link{optional} The value of the `segmentAttribute` to segment on.}

\item{\code{summary}}{\link{HumilityStatsBucket}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HumilityStatsOverTimeRetrieveResponse$new()}}
\item \href{#method-validate}{\code{HumilityStatsOverTimeRetrieveResponse$validate()}}
\item \href{#method-toJSON}{\code{HumilityStatsOverTimeRetrieveResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{HumilityStatsOverTimeRetrieveResponse$fromJSON()}}
\item \href{#method-clone}{\code{HumilityStatsOverTimeRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HumilityStatsOverTimeRetrieveResponse$new(
  buckets = NULL,
  summary = NULL,
  modelId = NULL,
  segmentAttribute = NULL,
  segmentValue = "",
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{buckets}}{An array of `bucket` objects, representing service health stats of the deployment over time}

\item{\code{modelId}}{The id of the model for which metrics are being retrieved.}

\item{\code{segmentAttribute}}{The name of the segment on which segment analysis is being performed.}

\item{\code{segmentValue}}{The value of the `segmentAttribute` to segment on.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HumilityStatsOverTimeRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HumilityStatsOverTimeRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HumilityStatsOverTimeRetrieveResponse$fromJSON(
  HumilityStatsOverTimeRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{HumilityStatsOverTimeRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HumilityStatsOverTimeRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
