% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_augmentation_list_create_response.R
\docType{class}
\name{ImageAugmentationListCreateResponse}
\alias{ImageAugmentationListCreateResponse}
\title{ImageAugmentationListCreateResponse}
\format{
An \code{R6Class} generator object
}
\description{
ImageAugmentationListCreateResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{augmentationListId}}{character                  Id of the newly created augmentation list which can be used to:                 - retrieve the full augmentation list details;                 - retrieve augmentation previews;                 - to set the list id of a model in advanced tuning;                 - to delete or rename the list.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImageAugmentationListCreateResponse-new}{\code{ImageAugmentationListCreateResponse$new()}}
\item \href{#method-ImageAugmentationListCreateResponse-validate}{\code{ImageAugmentationListCreateResponse$validate()}}
\item \href{#method-ImageAugmentationListCreateResponse-toJSON}{\code{ImageAugmentationListCreateResponse$toJSON()}}
\item \href{#method-ImageAugmentationListCreateResponse-fromJSON}{\code{ImageAugmentationListCreateResponse$fromJSON()}}
\item \href{#method-ImageAugmentationListCreateResponse-clone}{\code{ImageAugmentationListCreateResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageAugmentationListCreateResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImageAugmentationListCreateResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationListCreateResponse$new(
  augmentationListId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{augmentationListId}}{Id of the newly created augmentation list which can be used to:                 - retrieve the full augmentation list details;                 - retrieve augmentation previews;                 - to set the list id of a model in advanced tuning;                 - to delete or rename the list.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageAugmentationListCreateResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ImageAugmentationListCreateResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationListCreateResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageAugmentationListCreateResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ImageAugmentationListCreateResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationListCreateResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageAugmentationListCreateResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ImageAugmentationListCreateResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationListCreateResponse$fromJSON(
  ImageAugmentationListCreateResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ImageAugmentationListCreateResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageAugmentationListCreateResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImageAugmentationListCreateResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationListCreateResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
