% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_augmentation_list_patch_param.R
\docType{class}
\name{ImageAugmentationListPatchParam}
\alias{ImageAugmentationListPatchParam}
\title{ImageAugmentationListPatchParam}
\format{
An \code{R6Class} generator object
}
\description{
ImageAugmentationListPatchParam Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{featureName}}{character \link{optional} The name of the image feature containing the data to be augmented}

\item{\code{inUse}}{character \link{optional} This is set to true when the Augmentation List has been used to train a model}

\item{\code{initialList}}{character \link{optional} Whether this list will be used during autopilot to perform image augmentation}

\item{\code{name}}{character \link{optional} The name of the image augmentation list}

\item{\code{numberOfNewImages}}{integer \link{optional} Number of new rows to add for each existing row}

\item{\code{projectId}}{character \link{optional} The project containing the image data to be augmented}

\item{\code{transformationProbability}}{numeric \link{optional} Probability that each enabled transformation will be applied to an image. This does not apply to Horizontal or Vertical Flip, which are set to 50\% always.}

\item{\code{transformations}}{list( \link{Transformation} ) \link{optional} List of Transformations to possibly apply to each image}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ImageAugmentationListPatchParam$new()}}
\item \href{#method-validate}{\code{ImageAugmentationListPatchParam$validate()}}
\item \href{#method-toJSON}{\code{ImageAugmentationListPatchParam$toJSON()}}
\item \href{#method-fromJSON}{\code{ImageAugmentationListPatchParam$fromJSON()}}
\item \href{#method-clone}{\code{ImageAugmentationListPatchParam$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationListPatchParam$new(
  featureName = NULL,
  inUse = NULL,
  initialList = NULL,
  name = NULL,
  numberOfNewImages = NULL,
  projectId = NULL,
  transformationProbability = NULL,
  transformations = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{featureName}}{The name of the image feature containing the data to be augmented}

\item{\code{inUse}}{This is set to true when the Augmentation List has been used to train a model}

\item{\code{initialList}}{Whether this list will be used during autopilot to perform image augmentation}

\item{\code{name}}{The name of the image augmentation list}

\item{\code{numberOfNewImages}}{Number of new rows to add for each existing row}

\item{\code{projectId}}{The project containing the image data to be augmented}

\item{\code{transformationProbability}}{Probability that each enabled transformation will be applied to an image. This does not apply to Horizontal or Vertical Flip, which are set to 50\% always.}

\item{\code{transformations}}{List of Transformations to possibly apply to each image}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationListPatchParam$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationListPatchParam$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationListPatchParam$fromJSON(
  ImageAugmentationListPatchParamJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ImageAugmentationListPatchParamJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationListPatchParam$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
