% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_augmentation_options_response.R
\docType{class}
\name{ImageAugmentationOptionsResponse}
\alias{ImageAugmentationOptionsResponse}
\title{ImageAugmentationOptionsResponse}
\format{
An \code{R6Class} generator object
}
\description{
ImageAugmentationOptionsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{currentNumberOfNewImages}}{integer Number of new images to be created for each original image during training}

\item{\code{currentTransformationProbability}}{numeric Probability that each transformation included in an augmentation list will be applied to an image, if `affectedByTransformationProbability` for that transformation is True}

\item{\code{featureName}}{character \link{optional} Name of the image feature that the augmentation list is associated with}

\item{\code{id}}{character Augmentation list id}

\item{\code{maxNumberOfNewImages}}{integer Maximum number of new images per original image to be generated during training}

\item{\code{maxTransformationProbability}}{numeric Maximum probability that each enabled augmentation will be applied to an image}

\item{\code{minNumberOfNewImages}}{integer Minimum number of new images per original image to be generated during training}

\item{\code{minTransformationProbability}}{numeric Minimum probability that each enabled augmentation will be applied to an image}

\item{\code{name}}{character The name of the augmentation list}

\item{\code{projectId}}{character The id of the project containing the image data to be augmented}

\item{\code{transformations}}{list( \link{ImageAugmentationOptionsTransformation} ) List of Transformations to possibly apply to each image}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImageAugmentationOptionsResponse-new}{\code{ImageAugmentationOptionsResponse$new()}}
\item \href{#method-ImageAugmentationOptionsResponse-validate}{\code{ImageAugmentationOptionsResponse$validate()}}
\item \href{#method-ImageAugmentationOptionsResponse-toJSON}{\code{ImageAugmentationOptionsResponse$toJSON()}}
\item \href{#method-ImageAugmentationOptionsResponse-fromJSON}{\code{ImageAugmentationOptionsResponse$fromJSON()}}
\item \href{#method-ImageAugmentationOptionsResponse-clone}{\code{ImageAugmentationOptionsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageAugmentationOptionsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImageAugmentationOptionsResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationOptionsResponse$new(
  currentNumberOfNewImages = NULL,
  currentTransformationProbability = NULL,
  id = NULL,
  maxNumberOfNewImages = NULL,
  maxTransformationProbability = NULL,
  minNumberOfNewImages = NULL,
  minTransformationProbability = NULL,
  name = NULL,
  projectId = NULL,
  transformations = NULL,
  featureName = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{currentNumberOfNewImages}}{Number of new images to be created for each original image during training}

\item{\code{currentTransformationProbability}}{Probability that each transformation included in an augmentation list will be applied to an image, if `affectedByTransformationProbability` for that transformation is True}

\item{\code{id}}{Augmentation list id}

\item{\code{maxNumberOfNewImages}}{Maximum number of new images per original image to be generated during training}

\item{\code{maxTransformationProbability}}{Maximum probability that each enabled augmentation will be applied to an image}

\item{\code{minNumberOfNewImages}}{Minimum number of new images per original image to be generated during training}

\item{\code{minTransformationProbability}}{Minimum probability that each enabled augmentation will be applied to an image}

\item{\code{name}}{The name of the augmentation list}

\item{\code{projectId}}{The id of the project containing the image data to be augmented}

\item{\code{transformations}}{List of Transformations to possibly apply to each image}

\item{\code{featureName}}{Name of the image feature that the augmentation list is associated with}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageAugmentationOptionsResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ImageAugmentationOptionsResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationOptionsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageAugmentationOptionsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ImageAugmentationOptionsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationOptionsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageAugmentationOptionsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ImageAugmentationOptionsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationOptionsResponse$fromJSON(
  ImageAugmentationOptionsResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ImageAugmentationOptionsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageAugmentationOptionsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImageAugmentationOptionsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationOptionsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
