% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_augmentation_options_transformation.R
\docType{class}
\name{ImageAugmentationOptionsTransformation}
\alias{ImageAugmentationOptionsTransformation}
\title{ImageAugmentationOptionsTransformation}
\format{
An \code{R6Class} generator object
}
\description{
ImageAugmentationOptionsTransformation Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{affectedByTransformationProbability}}{character If true, whenever this transformation is included in an augmentation list, this transformation will be applied to each image with probability set by the Transformation Probability. If false, whenever this transformation is included in an augmentation list, this transformation will be applied to each image with probability described in the Platform Documentation here: https://app.datarobot.com/docs/modeling/special-workflows/visual-ai/tti-augment/ttia-lists.html#transformations}

\item{\code{enabledByDefault}}{character Determines if the parameter should be default selected in the UI}

\item{\code{name}}{character The name of the transformation}

\item{\code{params}}{list( \link{ImageAugmentationOptionsTransformationParam} ) The list of parameters that control the transformation}

\item{\code{translatedName}}{character Translated name of the transformation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ImageAugmentationOptionsTransformation$new()}}
\item \href{#method-validate}{\code{ImageAugmentationOptionsTransformation$validate()}}
\item \href{#method-toJSON}{\code{ImageAugmentationOptionsTransformation$toJSON()}}
\item \href{#method-fromJSON}{\code{ImageAugmentationOptionsTransformation$fromJSON()}}
\item \href{#method-clone}{\code{ImageAugmentationOptionsTransformation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationOptionsTransformation$new(
  affectedByTransformationProbability = NULL,
  enabledByDefault = NULL,
  name = NULL,
  params = NULL,
  translatedName = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{affectedByTransformationProbability}}{If true, whenever this transformation is included in an augmentation list, this transformation will be applied to each image with probability set by the Transformation Probability. If false, whenever this transformation is included in an augmentation list, this transformation will be applied to each image with probability described in the Platform Documentation here: https://app.datarobot.com/docs/modeling/special-workflows/visual-ai/tti-augment/ttia-lists.html#transformations}

\item{\code{enabledByDefault}}{Determines if the parameter should be default selected in the UI}

\item{\code{name}}{The name of the transformation}

\item{\code{params}}{The list of parameters that control the transformation}

\item{\code{translatedName}}{Translated name of the transformation}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationOptionsTransformation$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationOptionsTransformation$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationOptionsTransformation$fromJSON(
  ImageAugmentationOptionsTransformationJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ImageAugmentationOptionsTransformationJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationOptionsTransformation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
