% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/image_augmentation_options_transformation_param.R
\docType{class}
\name{ImageAugmentationOptionsTransformationParam}
\alias{ImageAugmentationOptionsTransformationParam}
\title{ImageAugmentationOptionsTransformationParam}
\format{
An \code{R6Class} generator object
}
\description{
ImageAugmentationOptionsTransformationParam Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{currentValue}}{\link{AnyOfintegernumber} \link{optional} Current transformation value}

\item{\code{maxValue}}{\link{AnyOfintegernumber} \link{optional} Max transformation value}

\item{\code{minValue}}{\link{AnyOfintegernumber} \link{optional} Min transformation value}

\item{\code{name}}{character Transformation param name}

\item{\code{translatedName}}{character Translated name of the parameter}

\item{\code{type}}{character The type of the parameter (int, float, etc...)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImageAugmentationOptionsTransformationParam-new}{\code{ImageAugmentationOptionsTransformationParam$new()}}
\item \href{#method-ImageAugmentationOptionsTransformationParam-validate}{\code{ImageAugmentationOptionsTransformationParam$validate()}}
\item \href{#method-ImageAugmentationOptionsTransformationParam-toJSON}{\code{ImageAugmentationOptionsTransformationParam$toJSON()}}
\item \href{#method-ImageAugmentationOptionsTransformationParam-fromJSON}{\code{ImageAugmentationOptionsTransformationParam$fromJSON()}}
\item \href{#method-ImageAugmentationOptionsTransformationParam-clone}{\code{ImageAugmentationOptionsTransformationParam$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageAugmentationOptionsTransformationParam-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImageAugmentationOptionsTransformationParam-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationOptionsTransformationParam$new(
  name = NULL,
  translatedName = NULL,
  type = NULL,
  currentValue = NULL,
  maxValue = NULL,
  minValue = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Transformation param name}

\item{\code{translatedName}}{Translated name of the parameter}

\item{\code{type}}{The type of the parameter (int, float, etc...)}

\item{\code{currentValue}}{Current transformation value}

\item{\code{maxValue}}{Max transformation value}

\item{\code{minValue}}{Min transformation value}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageAugmentationOptionsTransformationParam-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ImageAugmentationOptionsTransformationParam-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationOptionsTransformationParam$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageAugmentationOptionsTransformationParam-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ImageAugmentationOptionsTransformationParam-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationOptionsTransformationParam$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageAugmentationOptionsTransformationParam-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ImageAugmentationOptionsTransformationParam-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationOptionsTransformationParam$fromJSON(
  ImageAugmentationOptionsTransformationParamJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ImageAugmentationOptionsTransformationParamJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageAugmentationOptionsTransformationParam-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImageAugmentationOptionsTransformationParam-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationOptionsTransformationParam$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
