% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_augmentation_retrieve_samples_item.R
\docType{class}
\name{ImageAugmentationRetrieveSamplesItem}
\alias{ImageAugmentationRetrieveSamplesItem}
\title{ImageAugmentationRetrieveSamplesItem}
\format{
An \code{R6Class} generator object
}
\description{
ImageAugmentationRetrieveSamplesItem Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{height}}{integer Height of the image in pixels}

\item{\code{imageId}}{character Id of the image. The augmented image file can be retrieved with :http:get:`/api/v2/projects/(projectId)/images/(imageId)/file/`}

\item{\code{originalImageId}}{character Id of the original image that was transformed to produce the augmented image. If this is an original image (from the original training dataset) this value will be null. The id can be used to retrieve the original image file with: :http:get:`/api/v2/projects/(projectId)/images/(imageId)/file/`}

\item{\code{projectId}}{character Project id}

\item{\code{width}}{integer Width of the image in pixels}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImageAugmentationRetrieveSamplesItem-new}{\code{ImageAugmentationRetrieveSamplesItem$new()}}
\item \href{#method-ImageAugmentationRetrieveSamplesItem-validate}{\code{ImageAugmentationRetrieveSamplesItem$validate()}}
\item \href{#method-ImageAugmentationRetrieveSamplesItem-toJSON}{\code{ImageAugmentationRetrieveSamplesItem$toJSON()}}
\item \href{#method-ImageAugmentationRetrieveSamplesItem-fromJSON}{\code{ImageAugmentationRetrieveSamplesItem$fromJSON()}}
\item \href{#method-ImageAugmentationRetrieveSamplesItem-clone}{\code{ImageAugmentationRetrieveSamplesItem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageAugmentationRetrieveSamplesItem-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImageAugmentationRetrieveSamplesItem-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationRetrieveSamplesItem$new(
  height = NULL,
  imageId = NULL,
  originalImageId = NULL,
  projectId = NULL,
  width = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{height}}{Height of the image in pixels}

\item{\code{imageId}}{Id of the image. The augmented image file can be retrieved with :http:get:`/api/v2/projects/(projectId)/images/(imageId)/file/`}

\item{\code{originalImageId}}{Id of the original image that was transformed to produce the augmented image. If this is an original image (from the original training dataset) this value will be null. The id can be used to retrieve the original image file with: :http:get:`/api/v2/projects/(projectId)/images/(imageId)/file/`}

\item{\code{projectId}}{Project id}

\item{\code{width}}{Width of the image in pixels}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageAugmentationRetrieveSamplesItem-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ImageAugmentationRetrieveSamplesItem-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationRetrieveSamplesItem$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageAugmentationRetrieveSamplesItem-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ImageAugmentationRetrieveSamplesItem-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationRetrieveSamplesItem$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageAugmentationRetrieveSamplesItem-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ImageAugmentationRetrieveSamplesItem-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationRetrieveSamplesItem$fromJSON(
  ImageAugmentationRetrieveSamplesItemJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ImageAugmentationRetrieveSamplesItemJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageAugmentationRetrieveSamplesItem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImageAugmentationRetrieveSamplesItem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageAugmentationRetrieveSamplesItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
