% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_embedding.R
\docType{class}
\name{ImageEmbedding}
\alias{ImageEmbedding}
\title{ImageEmbedding}
\format{
An \code{R6Class} generator object
}
\description{
ImageEmbedding Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actualTargetValue}}{\link{OneOfstringintegerarray} Actual target value of the dataset row.}

\item{\code{imageId}}{character Id of the image. The actual image file can be retrieved with :http:get:`/api/v2/projects/(projectId)/images/(imageId)/file/`.}

\item{\code{positionX}}{numeric x coordinate of the image in the embedding space.}

\item{\code{positionY}}{numeric y coordinate of the image in the embedding space.}

\item{\code{prediction}}{\link{InsightsPredictionField}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImageEmbedding-new}{\code{ImageEmbedding$new()}}
\item \href{#method-ImageEmbedding-validate}{\code{ImageEmbedding$validate()}}
\item \href{#method-ImageEmbedding-toJSON}{\code{ImageEmbedding$toJSON()}}
\item \href{#method-ImageEmbedding-fromJSON}{\code{ImageEmbedding$fromJSON()}}
\item \href{#method-ImageEmbedding-clone}{\code{ImageEmbedding$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageEmbedding-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImageEmbedding-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageEmbedding$new(
  actualTargetValue = NULL,
  imageId = NULL,
  positionX = NULL,
  positionY = NULL,
  prediction = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{actualTargetValue}}{Actual target value of the dataset row.}

\item{\code{imageId}}{Id of the image. The actual image file can be retrieved with :http:get:`/api/v2/projects/(projectId)/images/(imageId)/file/`.}

\item{\code{positionX}}{x coordinate of the image in the embedding space.}

\item{\code{positionY}}{y coordinate of the image in the embedding space.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageEmbedding-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ImageEmbedding-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageEmbedding$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageEmbedding-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ImageEmbedding-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageEmbedding$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageEmbedding-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ImageEmbedding-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageEmbedding$fromJSON(ImageEmbeddingJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ImageEmbeddingJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageEmbedding-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImageEmbedding-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageEmbedding$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
