% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_feature.R
\docType{class}
\name{ImageFeature}
\alias{ImageFeature}
\title{ImageFeature}
\format{
An \code{R6Class} generator object
}
\description{
ImageFeature Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{featureImpact}}{numeric \link{optional} Feature Impact score.}

\item{\code{featureName}}{character Feature name.}

\item{\code{featureType}}{character Feature Type.}

\item{\code{insights}}{list( \link{ImageInsights} ) A list of Cluster Insights for an image feature.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImageFeature-new}{\code{ImageFeature$new()}}
\item \href{#method-ImageFeature-validate}{\code{ImageFeature$validate()}}
\item \href{#method-ImageFeature-toJSON}{\code{ImageFeature$toJSON()}}
\item \href{#method-ImageFeature-fromJSON}{\code{ImageFeature$fromJSON()}}
\item \href{#method-ImageFeature-clone}{\code{ImageFeature$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageFeature-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImageFeature-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageFeature$new(
  featureName = NULL,
  featureType = NULL,
  insights = NULL,
  featureImpact = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{featureName}}{Feature name.}

\item{\code{featureType}}{Feature Type.}

\item{\code{insights}}{A list of Cluster Insights for an image feature.}

\item{\code{featureImpact}}{Feature Impact score.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageFeature-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ImageFeature-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageFeature$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageFeature-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ImageFeature-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageFeature$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageFeature-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ImageFeature-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageFeature$fromJSON(ImageFeatureJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ImageFeatureJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageFeature-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImageFeature-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageFeature$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
