% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_metadata_with_bins.R
\docType{class}
\name{ImageMetadataWithBins}
\alias{ImageMetadataWithBins}
\title{ImageMetadataWithBins}
\format{
An \code{R6Class} generator object
}
\description{
ImageMetadataWithBins Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{height}}{integer Height of the image in pixels}

\item{\code{imageId}}{character Id of the image. The actual image file can be retrieved with :http:get:`/api/v2/projects/(projectId)/images/(imageId)/file/`}

\item{\code{targetBinEnd}}{integer \link{optional} Target value for bin end for regression, null for classification}

\item{\code{targetBinRowCount}}{integer Number of rows in this target bin}

\item{\code{targetBinStart}}{integer \link{optional} Target value for bin start for regression, null for classification}

\item{\code{targetValue}}{numeric \link{optional} Target value}

\item{\code{width}}{integer Width of the image in pixels}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ImageMetadataWithBins$new()}}
\item \href{#method-validate}{\code{ImageMetadataWithBins$validate()}}
\item \href{#method-toJSON}{\code{ImageMetadataWithBins$toJSON()}}
\item \href{#method-fromJSON}{\code{ImageMetadataWithBins$fromJSON()}}
\item \href{#method-clone}{\code{ImageMetadataWithBins$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageMetadataWithBins$new(
  height = NULL,
  imageId = NULL,
  targetBinRowCount = NULL,
  width = NULL,
  targetBinEnd = NULL,
  targetBinStart = NULL,
  targetValue = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{height}}{Height of the image in pixels}

\item{\code{imageId}}{Id of the image. The actual image file can be retrieved with :http:get:`/api/v2/projects/(projectId)/images/(imageId)/file/`}

\item{\code{targetBinRowCount}}{Number of rows in this target bin}

\item{\code{width}}{Width of the image in pixels}

\item{\code{targetBinEnd}}{Target value for bin end for regression, null for classification}

\item{\code{targetBinStart}}{Target value for bin start for regression, null for classification}

\item{\code{targetValue}}{Target value}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageMetadataWithBins$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageMetadataWithBins$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageMetadataWithBins$fromJSON(
  ImageMetadataWithBinsJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ImageMetadataWithBinsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageMetadataWithBins$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
