% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported_model_retrieve_response.R
\docType{class}
\name{ImportedModelRetrieveResponse}
\alias{ImportedModelRetrieveResponse}
\title{ImportedModelRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
ImportedModelRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{createdById}}{character ID of the user who originally created the model}

\item{\code{createdByUsername}}{character Name of the user who originally created the model}

\item{\code{datasetName}}{character Filename of the dataset used to create the project}

\item{\code{displayName}}{character Human-readable model import name}

\item{\code{featurelistName}}{character Name of the featurelist used to train the model}

\item{\code{id}}{character Imported model id}

\item{\code{importedAt}}{character datetime the model was imported, RFC 3339 Z Date and Time}

\item{\code{importedById}}{character ID of the user who imported the model}

\item{\code{importedByUsername}}{character Name of the user who imported the model}

\item{\code{isTimeSeries}}{character True if this is a timeseries model, otherwise - false}

\item{\code{modelId}}{character Original model ID}

\item{\code{modelName}}{character Human-readable model type}

\item{\code{note}}{character Manually added note about this model}

\item{\code{originUrl}}{character URL of the application the model was exported from}

\item{\code{projectId}}{character Original project ID the models belongs to}

\item{\code{projectName}}{character Original project name}

\item{\code{pythonVersion}}{character Version of python runtime used to process this model}

\item{\code{target}}{character Name of the projects target}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ImportedModelRetrieveResponse$new()}}
\item \href{#method-validate}{\code{ImportedModelRetrieveResponse$validate()}}
\item \href{#method-toJSON}{\code{ImportedModelRetrieveResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{ImportedModelRetrieveResponse$fromJSON()}}
\item \href{#method-clone}{\code{ImportedModelRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportedModelRetrieveResponse$new(
  createdById = NULL,
  createdByUsername = NULL,
  datasetName = NULL,
  displayName = NULL,
  featurelistName = NULL,
  id = NULL,
  importedAt = NULL,
  importedById = NULL,
  importedByUsername = NULL,
  isTimeSeries = NULL,
  modelId = NULL,
  modelName = NULL,
  note = NULL,
  originUrl = NULL,
  projectId = NULL,
  projectName = NULL,
  pythonVersion = NULL,
  target = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{createdById}}{ID of the user who originally created the model}

\item{\code{createdByUsername}}{Name of the user who originally created the model}

\item{\code{datasetName}}{Filename of the dataset used to create the project}

\item{\code{displayName}}{Human-readable model import name}

\item{\code{featurelistName}}{Name of the featurelist used to train the model}

\item{\code{id}}{Imported model id}

\item{\code{importedAt}}{datetime the model was imported, RFC 3339 Z Date and Time}

\item{\code{importedById}}{ID of the user who imported the model}

\item{\code{importedByUsername}}{Name of the user who imported the model}

\item{\code{isTimeSeries}}{True if this is a timeseries model, otherwise - false}

\item{\code{modelId}}{Original model ID}

\item{\code{modelName}}{Human-readable model type}

\item{\code{note}}{Manually added note about this model}

\item{\code{originUrl}}{URL of the application the model was exported from}

\item{\code{projectId}}{Original project ID the models belongs to}

\item{\code{projectName}}{Original project name}

\item{\code{pythonVersion}}{Version of python runtime used to process this model}

\item{\code{target}}{Name of the projects target}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportedModelRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportedModelRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportedModelRetrieveResponse$fromJSON(
  ImportedModelRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ImportedModelRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportedModelRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
