% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure_api.R
\docType{class}
\name{InfrastructureApi}
\alias{InfrastructureApi}
\title{Infrastructure operations}
\format{
An \code{R6Class} generator object
}
\description{
datarobot.apicore.Infrastructure
}
\examples{

## ------------------------------------------------
## Method `InfrastructureApi$ClusterLicenseList`
## ------------------------------------------------

\dontrun{
library(datarobot.apicore)

api.instance <- InfrastructureApi$new()
result <- api.instance$ClusterLicenseList()
}

## ------------------------------------------------
## Method `InfrastructureApi$ClusterLicensePutMany`
## ------------------------------------------------

\dontrun{
library(datarobot.apicore)
clusterLicenseUpdate <- ClusterLicenseUpdate$new() # ClusterLicenseUpdate | 

api.instance <- InfrastructureApi$new()
result <- api.instance$ClusterLicensePutMany(clusterLicenseUpdate=clusterLicenseUpdate)
}

## ------------------------------------------------
## Method `InfrastructureApi$ClusterLicenseValidationCreate`
## ------------------------------------------------

\dontrun{
library(datarobot.apicore)
clusterLicenseUpdate <- ClusterLicenseUpdate$new() # ClusterLicenseUpdate | 

api.instance <- InfrastructureApi$new()
result <- api.instance$ClusterLicenseValidationCreate(clusterLicenseUpdate=clusterLicenseUpdate)
}

## ------------------------------------------------
## Method `InfrastructureApi$VersionList`
## ------------------------------------------------

\dontrun{
library(datarobot.apicore)

api.instance <- InfrastructureApi$new()
result <- api.instance$VersionList()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InfrastructureApi-new}{\code{InfrastructureApi$new()}}
\item \href{#method-InfrastructureApi-ClusterLicenseList}{\code{InfrastructureApi$ClusterLicenseList()}}
\item \href{#method-InfrastructureApi-ClusterLicensePutMany}{\code{InfrastructureApi$ClusterLicensePutMany()}}
\item \href{#method-InfrastructureApi-ClusterLicenseValidationCreate}{\code{InfrastructureApi$ClusterLicenseValidationCreate()}}
\item \href{#method-InfrastructureApi-VersionList}{\code{InfrastructureApi$VersionList()}}
\item \href{#method-InfrastructureApi-clone}{\code{InfrastructureApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfrastructureApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-InfrastructureApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfrastructureApi$new(apiClient)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{apiClient}}{A configurable \code{ApiClient} instance. If none provided, a new client with default configuration will be created.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfrastructureApi-ClusterLicenseList"></a>}}
\if{latex}{\out{\hypertarget{method-InfrastructureApi-ClusterLicenseList}{}}}
\subsection{Method \code{ClusterLicenseList()}}{
Retrieve the information about cluster license
Produces: "application/json"
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfrastructureApi$ClusterLicenseList(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional. Additional named parameters to be passed downward.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Retrieve the information about the currently deployed cluster license.


This method invokes \verb{GET /clusterLicense/} in the DataRobot Public API.


Response status codes, messages, and headers:
\itemize{
\item \strong{\code{200}}
\itemize{
}
}
}

\subsection{Returns}{
\link{ClusterLicenseRetrieveResponse}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
library(datarobot.apicore)

api.instance <- InfrastructureApi$new()
result <- api.instance$ClusterLicenseList()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfrastructureApi-ClusterLicensePutMany"></a>}}
\if{latex}{\out{\hypertarget{method-InfrastructureApi-ClusterLicensePutMany}{}}}
\subsection{Method \code{ClusterLicensePutMany()}}{
Create or replace license
Produces: NA
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfrastructureApi$ClusterLicensePutMany(clusterLicenseUpdate = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clusterLicenseUpdate}}{\link{ClusterLicenseUpdate}.}

\item{\code{...}}{Optional. Additional named parameters to be passed downward.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create or replace the cluster license.


This method invokes \verb{PUT /clusterLicense/} in the DataRobot Public API.


Response status codes, messages, and headers:
\itemize{
\item \strong{\code{200}}
\itemize{
}
\item \strong{\code{422}} License is invalid or has expired.
\itemize{
}
}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
library(datarobot.apicore)
clusterLicenseUpdate <- ClusterLicenseUpdate$new() # ClusterLicenseUpdate | 

api.instance <- InfrastructureApi$new()
result <- api.instance$ClusterLicensePutMany(clusterLicenseUpdate=clusterLicenseUpdate)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfrastructureApi-ClusterLicenseValidationCreate"></a>}}
\if{latex}{\out{\hypertarget{method-InfrastructureApi-ClusterLicenseValidationCreate}{}}}
\subsection{Method \code{ClusterLicenseValidationCreate()}}{
Check if a license is valid
Produces: "application/json"
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfrastructureApi$ClusterLicenseValidationCreate(
  clusterLicenseUpdate = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clusterLicenseUpdate}}{\link{ClusterLicenseUpdate}.}

\item{\code{...}}{Optional. Additional named parameters to be passed downward.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if a cluster license is valid.


This method invokes \verb{POST /clusterLicenseValidation/} in the DataRobot Public API.


Response status codes, messages, and headers:
\itemize{
\item \strong{\code{200}}
\itemize{
}
\item \strong{\code{422}} License is invalid or has expired.
\itemize{
}
}
}

\subsection{Returns}{
\link{ClusterLicenseValidationResponse}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
library(datarobot.apicore)
clusterLicenseUpdate <- ClusterLicenseUpdate$new() # ClusterLicenseUpdate | 

api.instance <- InfrastructureApi$new()
result <- api.instance$ClusterLicenseValidationCreate(clusterLicenseUpdate=clusterLicenseUpdate)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfrastructureApi-VersionList"></a>}}
\if{latex}{\out{\hypertarget{method-InfrastructureApi-VersionList}{}}}
\subsection{Method \code{VersionList()}}{
Retrieve version information.
Produces: "application/json"
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfrastructureApi$VersionList(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional. Additional named parameters to be passed downward.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Provides the version information for the current version of the API.


This method invokes \verb{GET /version/} in the DataRobot Public API.


Response status codes, messages, and headers:
\itemize{
\item \strong{\code{200}}
\itemize{
}
}
}

\subsection{Returns}{
\link{VersionRetrieveResponse}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
library(datarobot.apicore)

api.instance <- InfrastructureApi$new()
result <- api.instance$VersionList()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InfrastructureApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InfrastructureApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfrastructureApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
