% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{InitializeAnomalyAssessment}
\alias{InitializeAnomalyAssessment}
\title{InitializeAnomalyAssessment}
\usage{
InitializeAnomalyAssessment(
  projectId,
  modelId,
  backtest,
  source,
  seriesId = NULL,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Calculate the anomaly assessment insight.
}
\details{
Initialize the anomaly assessment insight and calculate Shapley explanations for the most anomalous points in the subset. The insight is available for anomaly detection models in time series unsupervised projects which also support calculation of Shapley values.

This method invokes \verb{POST /projects/\{projectId\}/models/\{modelId\}/anomalyAssessmentInitialization/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsAnomalyAssessmentInitializationCreate in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
