% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isa}
\alias{isa}
\title{Type check}
\usage{
isa(x, what)
}
\arguments{
\item{x}{a \R object}
}
\description{
This function is a backport of the \code{\link[base:class]{base::isa()}} function in R 4.x to permit 3.x compatibility.
}
