% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{IsBlenderEligible}
\alias{IsBlenderEligible}
\title{IsBlenderEligible}
\usage{
IsBlenderEligible(project, modelIds, blendMethod, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Check if models can be blended.
}
\details{
Check if models can be blended.

This method invokes \verb{POST /projects/\{projectId\}/blenderModels/blendCheck/} in the DataRobot Public API.
}
\seealso{
The method ProjectsBlenderModelsBlendCheckCreate in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
