% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc_field_schemas.R
\docType{class}
\name{JDBCFieldSchemas}
\alias{JDBCFieldSchemas}
\title{JDBCFieldSchemas}
\format{
An \code{R6Class} generator object
}
\description{
JDBCFieldSchemas Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{choices}}{list( character ) If non-empty, a list of all possible values for this parameter.}

\item{\code{default}}{character Default value of the JDBC parameter.}

\item{\code{description}}{character Description of this parameter.}

\item{\code{index}}{integer \link{optional} Sort order within one `kind`.}

\item{\code{kind}}{character Use of this parameter in constructing the JDBC URL.}

\item{\code{name}}{character Name of the JDBC parameter.}

\item{\code{required}}{character Whether or not the parameter is required for a connection.}

\item{\code{visibleByDefault}}{character Whether or not the parameter should be shown in the UI by default.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{JDBCFieldSchemas$new()}}
\item \href{#method-validate}{\code{JDBCFieldSchemas$validate()}}
\item \href{#method-toJSON}{\code{JDBCFieldSchemas$toJSON()}}
\item \href{#method-fromJSON}{\code{JDBCFieldSchemas$fromJSON()}}
\item \href{#method-clone}{\code{JDBCFieldSchemas$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JDBCFieldSchemas$new(
  choices = NULL,
  default = NULL,
  description = NULL,
  kind = NULL,
  name = NULL,
  required = NULL,
  visibleByDefault = NULL,
  index = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{choices}}{If non-empty, a list of all possible values for this parameter.}

\item{\code{default}}{Default value of the JDBC parameter.}

\item{\code{description}}{Description of this parameter.}

\item{\code{kind}}{Use of this parameter in constructing the JDBC URL.}

\item{\code{name}}{Name of the JDBC parameter.}

\item{\code{required}}{Whether or not the parameter is required for a connection.}

\item{\code{visibleByDefault}}{Whether or not the parameter should be shown in the UI by default.}

\item{\code{index}}{Sort order within one `kind`.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JDBCFieldSchemas$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JDBCFieldSchemas$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JDBCFieldSchemas$fromJSON(JDBCFieldSchemasJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{JDBCFieldSchemasJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JDBCFieldSchemas$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
