% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc_table_data_source.R
\docType{class}
\name{JDBCTableDataSource}
\alias{JDBCTableDataSource}
\title{JDBCTableDataSource}
\format{
An \code{R6Class} generator object
}
\description{
JDBCTableDataSource Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalog}}{character \link{optional} Catalog name in the database if supported.}

\item{\code{dataStoreId}}{character Data store ID for this data source.}

\item{\code{fetchSize}}{integer \link{optional} User-specified fetch size.}

\item{\code{partitionColumn}}{character \link{optional} The name of the partition column. It is needed to allow parallel execution for the 10GB+ projects.}

\item{\code{schema}}{character \link{optional} Schema associated with the table or view in the database if the data source is not query based.}

\item{\code{table}}{character Table or view name in the database if the data source is not query based.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JDBCTableDataSource-new}{\code{JDBCTableDataSource$new()}}
\item \href{#method-JDBCTableDataSource-validate}{\code{JDBCTableDataSource$validate()}}
\item \href{#method-JDBCTableDataSource-toJSON}{\code{JDBCTableDataSource$toJSON()}}
\item \href{#method-JDBCTableDataSource-fromJSON}{\code{JDBCTableDataSource$fromJSON()}}
\item \href{#method-JDBCTableDataSource-clone}{\code{JDBCTableDataSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JDBCTableDataSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-JDBCTableDataSource-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JDBCTableDataSource$new(
  dataStoreId = NULL,
  table = NULL,
  catalog = NULL,
  fetchSize = NULL,
  partitionColumn = NULL,
  schema = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataStoreId}}{Data store ID for this data source.}

\item{\code{table}}{Table or view name in the database if the data source is not query based.}

\item{\code{catalog}}{Catalog name in the database if supported.}

\item{\code{fetchSize}}{User-specified fetch size.}

\item{\code{partitionColumn}}{The name of the partition column. It is needed to allow parallel execution for the 10GB+ projects.}

\item{\code{schema}}{Schema associated with the table or view in the database if the data source is not query based.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JDBCTableDataSource-validate"></a>}}
\if{latex}{\out{\hypertarget{method-JDBCTableDataSource-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JDBCTableDataSource$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JDBCTableDataSource-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JDBCTableDataSource-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JDBCTableDataSource$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JDBCTableDataSource-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JDBCTableDataSource-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JDBCTableDataSource$fromJSON(JDBCTableDataSourceJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{JDBCTableDataSourceJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JDBCTableDataSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JDBCTableDataSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JDBCTableDataSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
