% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_details_response.R
\docType{class}
\name{JobDetailsResponse}
\alias{JobDetailsResponse}
\title{JobDetailsResponse}
\format{
An \code{R6Class} generator object
}
\description{
JobDetailsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character The job ID.}

\item{\code{isBlocked}}{character True if the job is waiting for its dependencies to be resolved first.}

\item{\code{jobType}}{character The job type.}

\item{\code{message}}{character Error message in case of failure.}

\item{\code{modelId}}{character The model this job is associated with.}

\item{\code{projectId}}{character The project the job belongs to.}

\item{\code{status}}{character The job status.}

\item{\code{url}}{character A URL that can be used to request details about the job.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{JobDetailsResponse$new()}}
\item \href{#method-validate}{\code{JobDetailsResponse$validate()}}
\item \href{#method-toJSON}{\code{JobDetailsResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{JobDetailsResponse$fromJSON()}}
\item \href{#method-clone}{\code{JobDetailsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsResponse$new(
  id = NULL,
  isBlocked = NULL,
  jobType = NULL,
  message = NULL,
  modelId = NULL,
  projectId = NULL,
  status = NULL,
  url = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The job ID.}

\item{\code{isBlocked}}{True if the job is waiting for its dependencies to be resolved first.}

\item{\code{jobType}}{The job type.}

\item{\code{message}}{Error message in case of failure.}

\item{\code{modelId}}{The model this job is associated with.}

\item{\code{projectId}}{The project the job belongs to.}

\item{\code{status}}{The job status.}

\item{\code{url}}{A URL that can be used to request details about the job.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsResponse$fromJSON(JobDetailsResponseJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{JobDetailsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
