% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_run_details.R
\docType{class}
\name{JobRunDetails}
\alias{JobRunDetails}
\title{JobRunDetails}
\format{
An \code{R6Class} generator object
}
\description{
JobRunDetails Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{completed}}{character The time this job instance completed in ISO UTC format (e.g. '2020-02-20T18:00:02.218000Z')}

\item{\code{created}}{character The time this job instance was created in ISO UTC format (e.g. '2020-02-20T18:00:02.218000Z')}

\item{\code{duration}}{integer The number of seconds this job took to be completed}

\item{\code{failed}}{character The time this job instance failed in ISO UTC format (e.g. '2020-02-20T18:00:02.218000Z')}

\item{\code{failedRows}}{integer The number of rows that failed to score correctly}

\item{\code{jobRunId}}{character The ID of the job run. Used for accessing log information}

\item{\code{jobStatus}}{character The current status of the job}

\item{\code{scoredRows}}{integer The number of rows that were successfully scored}

\item{\code{totalRows}}{integer The number of rows that were processed}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{JobRunDetails$new()}}
\item \href{#method-validate}{\code{JobRunDetails$validate()}}
\item \href{#method-toJSON}{\code{JobRunDetails$toJSON()}}
\item \href{#method-fromJSON}{\code{JobRunDetails$fromJSON()}}
\item \href{#method-clone}{\code{JobRunDetails$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobRunDetails$new(
  completed = NULL,
  created = NULL,
  duration = NULL,
  failed = NULL,
  failedRows = NULL,
  jobRunId = NULL,
  jobStatus = NULL,
  scoredRows = NULL,
  totalRows = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{completed}}{The time this job instance completed in ISO UTC format (e.g. '2020-02-20T18:00:02.218000Z')}

\item{\code{created}}{The time this job instance was created in ISO UTC format (e.g. '2020-02-20T18:00:02.218000Z')}

\item{\code{duration}}{The number of seconds this job took to be completed}

\item{\code{failed}}{The time this job instance failed in ISO UTC format (e.g. '2020-02-20T18:00:02.218000Z')}

\item{\code{failedRows}}{The number of rows that failed to score correctly}

\item{\code{jobRunId}}{The ID of the job run. Used for accessing log information}

\item{\code{jobStatus}}{The current status of the job}

\item{\code{scoredRows}}{The number of rows that were successfully scored}

\item{\code{totalRows}}{The number of rows that were processed}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobRunDetails$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobRunDetails$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobRunDetails$fromJSON(JobRunDetailsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{JobRunDetailsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobRunDetails$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
