% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_run_time_based_forecast_point_policy.R
\docType{class}
\name{JobRunTimeBasedForecastPointPolicy}
\alias{JobRunTimeBasedForecastPointPolicy}
\title{JobRunTimeBasedForecastPointPolicy}
\format{
An \code{R6Class} generator object
}
\description{
JobRunTimeBasedForecastPointPolicy Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{configuration}}{\link{JobRunTimeBasedForecastPointPolicySettings} \link{optional}}

\item{\code{type}}{character Type of the forecast point policy. Forecast point will be based on the scheduled run time of the job or the current moment in UTC if job was launched manually. Run time can be adjusted backwards or forwards.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JobRunTimeBasedForecastPointPolicy-new}{\code{JobRunTimeBasedForecastPointPolicy$new()}}
\item \href{#method-JobRunTimeBasedForecastPointPolicy-validate}{\code{JobRunTimeBasedForecastPointPolicy$validate()}}
\item \href{#method-JobRunTimeBasedForecastPointPolicy-toJSON}{\code{JobRunTimeBasedForecastPointPolicy$toJSON()}}
\item \href{#method-JobRunTimeBasedForecastPointPolicy-fromJSON}{\code{JobRunTimeBasedForecastPointPolicy$fromJSON()}}
\item \href{#method-JobRunTimeBasedForecastPointPolicy-clone}{\code{JobRunTimeBasedForecastPointPolicy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobRunTimeBasedForecastPointPolicy-new"></a>}}
\if{latex}{\out{\hypertarget{method-JobRunTimeBasedForecastPointPolicy-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobRunTimeBasedForecastPointPolicy$new(
  type = NULL,
  configuration = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Type of the forecast point policy. Forecast point will be based on the scheduled run time of the job or the current moment in UTC if job was launched manually. Run time can be adjusted backwards or forwards.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobRunTimeBasedForecastPointPolicy-validate"></a>}}
\if{latex}{\out{\hypertarget{method-JobRunTimeBasedForecastPointPolicy-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobRunTimeBasedForecastPointPolicy$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobRunTimeBasedForecastPointPolicy-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobRunTimeBasedForecastPointPolicy-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobRunTimeBasedForecastPointPolicy$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobRunTimeBasedForecastPointPolicy-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobRunTimeBasedForecastPointPolicy-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobRunTimeBasedForecastPointPolicy$fromJSON(
  JobRunTimeBasedForecastPointPolicyJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{JobRunTimeBasedForecastPointPolicyJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobRunTimeBasedForecastPointPolicy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JobRunTimeBasedForecastPointPolicy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobRunTimeBasedForecastPointPolicy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
