% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/job_run_time_based_forecast_point_policy_settings.R
\docType{class}
\name{JobRunTimeBasedForecastPointPolicySettings}
\alias{JobRunTimeBasedForecastPointPolicySettings}
\title{JobRunTimeBasedForecastPointPolicySettings}
\format{
An \code{R6Class} generator object
}
\description{
JobRunTimeBasedForecastPointPolicySettings Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{offset}}{character Offset to shift job run time in a ISO-8601 format. Example of the positive offset 'P2DT5H3M', example of the negative offset  '-P2DT5H4M'}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JobRunTimeBasedForecastPointPolicySettings-new}{\code{JobRunTimeBasedForecastPointPolicySettings$new()}}
\item \href{#method-JobRunTimeBasedForecastPointPolicySettings-validate}{\code{JobRunTimeBasedForecastPointPolicySettings$validate()}}
\item \href{#method-JobRunTimeBasedForecastPointPolicySettings-toJSON}{\code{JobRunTimeBasedForecastPointPolicySettings$toJSON()}}
\item \href{#method-JobRunTimeBasedForecastPointPolicySettings-fromJSON}{\code{JobRunTimeBasedForecastPointPolicySettings$fromJSON()}}
\item \href{#method-JobRunTimeBasedForecastPointPolicySettings-clone}{\code{JobRunTimeBasedForecastPointPolicySettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobRunTimeBasedForecastPointPolicySettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-JobRunTimeBasedForecastPointPolicySettings-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobRunTimeBasedForecastPointPolicySettings$new(
  offset = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{offset}}{Offset to shift job run time in a ISO-8601 format. Example of the positive offset 'P2DT5H3M', example of the negative offset  '-P2DT5H4M'}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobRunTimeBasedForecastPointPolicySettings-validate"></a>}}
\if{latex}{\out{\hypertarget{method-JobRunTimeBasedForecastPointPolicySettings-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobRunTimeBasedForecastPointPolicySettings$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobRunTimeBasedForecastPointPolicySettings-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobRunTimeBasedForecastPointPolicySettings-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobRunTimeBasedForecastPointPolicySettings$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobRunTimeBasedForecastPointPolicySettings-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobRunTimeBasedForecastPointPolicySettings-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobRunTimeBasedForecastPointPolicySettings$fromJSON(
  JobRunTimeBasedForecastPointPolicySettingsJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{JobRunTimeBasedForecastPointPolicySettingsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobRunTimeBasedForecastPointPolicySettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JobRunTimeBasedForecastPointPolicySettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobRunTimeBasedForecastPointPolicySettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
