% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_relevance_plot.R
\docType{class}
\name{LabelRelevancePlot}
\alias{LabelRelevancePlot}
\title{LabelRelevancePlot}
\format{
An \code{R6Class} generator object
}
\description{
LabelRelevancePlot Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{labelRelevance}}{integer Label relevance value.}

\item{\code{rowCount}}{integer Number of rows for which the label has the given relevance.}

\item{\code{rowPct}}{numeric Percentage of rows for which the label has the given relevance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LabelRelevancePlot-new}{\code{LabelRelevancePlot$new()}}
\item \href{#method-LabelRelevancePlot-validate}{\code{LabelRelevancePlot$validate()}}
\item \href{#method-LabelRelevancePlot-toJSON}{\code{LabelRelevancePlot$toJSON()}}
\item \href{#method-LabelRelevancePlot-fromJSON}{\code{LabelRelevancePlot$fromJSON()}}
\item \href{#method-LabelRelevancePlot-clone}{\code{LabelRelevancePlot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelRelevancePlot-new"></a>}}
\if{latex}{\out{\hypertarget{method-LabelRelevancePlot-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelRelevancePlot$new(
  labelRelevance = NULL,
  rowCount = NULL,
  rowPct = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{labelRelevance}}{Label relevance value.}

\item{\code{rowCount}}{Number of rows for which the label has the given relevance.}

\item{\code{rowPct}}{Percentage of rows for which the label has the given relevance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelRelevancePlot-validate"></a>}}
\if{latex}{\out{\hypertarget{method-LabelRelevancePlot-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelRelevancePlot$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelRelevancePlot-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LabelRelevancePlot-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelRelevancePlot$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelRelevancePlot-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LabelRelevancePlot-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelRelevancePlot$fromJSON(LabelRelevancePlotJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{LabelRelevancePlotJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelRelevancePlot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LabelRelevancePlot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelRelevancePlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
