% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelwise_lift_chart.R
\docType{class}
\name{LabelwiseLiftChart}
\alias{LabelwiseLiftChart}
\title{LabelwiseLiftChart}
\format{
An \code{R6Class} generator object
}
\description{
LabelwiseLiftChart Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{labelBins}}{list( \link{LabelwiseLiftChartItem} ) Lift charts for the given data source.}

\item{\code{labels}}{list( character ) All available target labels for this insight.}

\item{\code{modelId}}{character Model Id.}

\item{\code{projectId}}{character Project Id.}

\item{\code{source}}{character Data source of Lift charts.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LabelwiseLiftChart-new}{\code{LabelwiseLiftChart$new()}}
\item \href{#method-LabelwiseLiftChart-validate}{\code{LabelwiseLiftChart$validate()}}
\item \href{#method-LabelwiseLiftChart-toJSON}{\code{LabelwiseLiftChart$toJSON()}}
\item \href{#method-LabelwiseLiftChart-fromJSON}{\code{LabelwiseLiftChart$fromJSON()}}
\item \href{#method-LabelwiseLiftChart-clone}{\code{LabelwiseLiftChart$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelwiseLiftChart-new"></a>}}
\if{latex}{\out{\hypertarget{method-LabelwiseLiftChart-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelwiseLiftChart$new(
  labelBins = NULL,
  labels = NULL,
  modelId = NULL,
  projectId = NULL,
  source = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{labelBins}}{Lift charts for the given data source.}

\item{\code{labels}}{All available target labels for this insight.}

\item{\code{modelId}}{Model Id.}

\item{\code{projectId}}{Project Id.}

\item{\code{source}}{Data source of Lift charts.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelwiseLiftChart-validate"></a>}}
\if{latex}{\out{\hypertarget{method-LabelwiseLiftChart-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelwiseLiftChart$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelwiseLiftChart-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LabelwiseLiftChart-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelwiseLiftChart$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelwiseLiftChart-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LabelwiseLiftChart-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelwiseLiftChart$fromJSON(LabelwiseLiftChartJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{LabelwiseLiftChartJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelwiseLiftChart-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LabelwiseLiftChart-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelwiseLiftChart$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
