% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelwise_roc.R
\docType{class}
\name{LabelwiseROC}
\alias{LabelwiseROC}
\title{LabelwiseROC}
\format{
An \code{R6Class} generator object
}
\description{
LabelwiseROC Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{averageModelMetrics}}{\link{AverageModelMetricsField}}

\item{\code{charts}}{list( \link{LabelwiseROCItem} ) ROC data for all labels from one data source.}

\item{\code{count}}{integer \link{optional} Number of labels returned on this page.}

\item{\code{labels}}{list( character ) All available target labels for this insight.}

\item{\code{next_}}{character URL pointing to the next page (if null, there is no next page).}

\item{\code{previous}}{character URL pointing to the previous page (if null, there is no previous page).}

\item{\code{rocType}}{character Type of ROC.}

\item{\code{totalCount}}{integer Total number of labels across all pages.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LabelwiseROC-new}{\code{LabelwiseROC$new()}}
\item \href{#method-LabelwiseROC-validate}{\code{LabelwiseROC$validate()}}
\item \href{#method-LabelwiseROC-toJSON}{\code{LabelwiseROC$toJSON()}}
\item \href{#method-LabelwiseROC-fromJSON}{\code{LabelwiseROC$fromJSON()}}
\item \href{#method-LabelwiseROC-clone}{\code{LabelwiseROC$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelwiseROC-new"></a>}}
\if{latex}{\out{\hypertarget{method-LabelwiseROC-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelwiseROC$new(
  averageModelMetrics = NULL,
  charts = NULL,
  labels = NULL,
  next_ = NULL,
  previous = NULL,
  rocType = NULL,
  totalCount = NULL,
  count = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{charts}}{ROC data for all labels from one data source.}

\item{\code{labels}}{All available target labels for this insight.}

\item{\code{next_}}{URL pointing to the next page (if null, there is no next page).}

\item{\code{previous}}{URL pointing to the previous page (if null, there is no previous page).}

\item{\code{rocType}}{Type of ROC.}

\item{\code{totalCount}}{Total number of labels across all pages.}

\item{\code{count}}{Number of labels returned on this page.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelwiseROC-validate"></a>}}
\if{latex}{\out{\hypertarget{method-LabelwiseROC-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelwiseROC$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelwiseROC-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LabelwiseROC-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelwiseROC$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelwiseROC-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LabelwiseROC-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelwiseROC$fromJSON(LabelwiseROCJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{LabelwiseROCJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelwiseROC-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LabelwiseROC-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelwiseROC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
