% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelwise_roc_point.R
\docType{class}
\name{LabelwiseROCPoint}
\alias{LabelwiseROCPoint}
\title{LabelwiseROCPoint}
\format{
An \code{R6Class} generator object
}
\description{
LabelwiseROCPoint Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{accuracy}}{numeric Accuracy.}

\item{\code{f1Score}}{numeric F1 score.}

\item{\code{falseNegativeScore}}{integer False negative score.}

\item{\code{falsePositiveRate}}{numeric False positive rate.}

\item{\code{falsePositiveScore}}{integer False positive score.}

\item{\code{fractionPredictedAsNegative}}{numeric Fraction of negative predictions.}

\item{\code{fractionPredictedAsPositive}}{numeric Fraction of positive predictions.}

\item{\code{liftNegative}}{numeric Negative lift.}

\item{\code{liftPositive}}{numeric Positive lift.}

\item{\code{matthewsCorrelationCoefficient}}{numeric Matthews correlation coefficient.}

\item{\code{negativePredictiveValue}}{numeric Negative predictive value.}

\item{\code{positivePredictiveValue}}{numeric Positive predictive value.}

\item{\code{threshold}}{numeric Threshold.}

\item{\code{trueNegativeRate}}{numeric True negative rate.}

\item{\code{trueNegativeScore}}{integer True negative score.}

\item{\code{truePositiveRate}}{numeric True positive rate.}

\item{\code{truePositiveScore}}{integer True positive score.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LabelwiseROCPoint$new()}}
\item \href{#method-validate}{\code{LabelwiseROCPoint$validate()}}
\item \href{#method-toJSON}{\code{LabelwiseROCPoint$toJSON()}}
\item \href{#method-fromJSON}{\code{LabelwiseROCPoint$fromJSON()}}
\item \href{#method-clone}{\code{LabelwiseROCPoint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelwiseROCPoint$new(
  accuracy = NULL,
  f1Score = NULL,
  falseNegativeScore = NULL,
  falsePositiveRate = NULL,
  falsePositiveScore = NULL,
  fractionPredictedAsNegative = NULL,
  fractionPredictedAsPositive = NULL,
  liftNegative = NULL,
  liftPositive = NULL,
  matthewsCorrelationCoefficient = NULL,
  negativePredictiveValue = NULL,
  positivePredictiveValue = NULL,
  threshold = NULL,
  trueNegativeRate = NULL,
  trueNegativeScore = NULL,
  truePositiveRate = NULL,
  truePositiveScore = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{accuracy}}{Accuracy.}

\item{\code{f1Score}}{F1 score.}

\item{\code{falseNegativeScore}}{False negative score.}

\item{\code{falsePositiveRate}}{False positive rate.}

\item{\code{falsePositiveScore}}{False positive score.}

\item{\code{fractionPredictedAsNegative}}{Fraction of negative predictions.}

\item{\code{fractionPredictedAsPositive}}{Fraction of positive predictions.}

\item{\code{liftNegative}}{Negative lift.}

\item{\code{liftPositive}}{Positive lift.}

\item{\code{matthewsCorrelationCoefficient}}{Matthews correlation coefficient.}

\item{\code{negativePredictiveValue}}{Negative predictive value.}

\item{\code{positivePredictiveValue}}{Positive predictive value.}

\item{\code{threshold}}{Threshold.}

\item{\code{trueNegativeRate}}{True negative rate.}

\item{\code{trueNegativeScore}}{True negative score.}

\item{\code{truePositiveRate}}{True positive rate.}

\item{\code{truePositiveScore}}{True positive score.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelwiseROCPoint$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelwiseROCPoint$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelwiseROCPoint$fromJSON(LabelwiseROCPointJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{LabelwiseROCPointJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelwiseROCPoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
