% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListAnomalyAssessmentRecords}
\alias{ListAnomalyAssessmentRecords}
\title{ListAnomalyAssessmentRecords}
\usage{
ListAnomalyAssessmentRecords(
  projectId,
  modelId,
  backtest = NULL,
  source = NULL,
  seriesId = NULL,
  limit = 100,
  offset = 0,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve anomaly assessment records.
}
\details{
Retrieve anomaly assessment records.

This method invokes \verb{GET /projects/\{projectId\}/anomalyAssessmentRecords/} in the DataRobot Public API.
}
\seealso{
The method ProjectsAnomalyAssessmentRecordsList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
