% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{ListApprovalPolicyTriggers}
\alias{ListApprovalPolicyTriggers}
\title{ListApprovalPolicyTriggers}
\usage{
ListApprovalPolicyTriggers(...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get a list of available policy triggers.
}
\details{
Get a list of available policy triggers.

This method invokes \verb{GET /approvalPolicyTriggers/} in the DataRobot Public API.
}
\seealso{
The method ApprovalPolicyTriggersList in \link{GovernanceApi}, which invokes the same underlying API endpoint.
}
