% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_apiWrapper.R
\name{ListAutomatedDocuments}
\alias{ListAutomatedDocuments}
\title{ListAutomatedDocuments}
\usage{
ListAutomatedDocuments(
  documentType,
  entityId,
  locale,
  outputFormat,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List all generated documents.
}
\details{
Get information about all previously generated documents available for your account. The information includes document ID and type, ID of the entity it was generated for, time of creation, and other information.  Example request to get a list of all documents:  \verb{text GET https://app.datarobot.com/api/v2/automatedDocuments/ HTTP/1.1 Authorization: Bearer DnwzBUSTOtKBO5Sp1hoUByG4YgZwCCw4}  Query for specific documents. For example, request a list of documents generated for a specific model in \code{docx} and \code{html} formats:  \verb{text GET https://app.datarobot.com/api/v2/automatedDocuments?entityId= 5ec4ea7e41054c158c5b002f&outputFormat=docx&outputFormat=html HTTP/1.1 Authorization: Bearer DnwzBUSTOtKBO5Sp1hoUByG4YgZwCCw4}  In response body, you will get a page with queried documents. This is an example response with one document returned:  \verb{json \{     \\"totalCount\\": 1,     \\"count\\": 1,     \\"previous\\": null,     \\"next\\": null,     \\"data\\": [         \{             \\"id\\": \\"5ebdb5e911a5fb85edff2b3c\\",             \\"documentType\\": \\"MODEL_COMPLIANCE\\",             \\"entityId\\": \\"5ebbb5e7d9d7b96e3d48e3b5\\",             \\"templateId\\": \\"5bd812e5f750edd392fa880f\\",             \\"locale\\": \\"EN_US\\",             \\"outputFormat\\": \\"DOCX\\",             \\"createdAt\\": \\"2019-11-07T11:12:13.141516Z\\"         \}     ] \}}  If there are no matching documents, you will get a page with an empty data array.

This method invokes \verb{GET /automatedDocuments/} in the DataRobot Public API.
}
\seealso{
The method AutomatedDocumentsList in \link{DocumentationApi}, which invokes the same underlying API endpoint.
}
