% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{ListBatchPredictionJobDefinitions}
\alias{ListBatchPredictionJobDefinitions}
\title{ListBatchPredictionJobDefinitions}
\usage{
ListBatchPredictionJobDefinitions(deploymentId, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List Batch Prediction job definitions
}
\details{
List all Batch Prediction jobs definitions available

This method invokes \verb{GET /batchPredictionJobDefinitions/} in the DataRobot Public API.
}
\seealso{
The method BatchPredictionJobDefinitionsList in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
