% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{ListBatchPredictionJobDefinitionsPortable}
\alias{ListBatchPredictionJobDefinitionsPortable}
\title{ListBatchPredictionJobDefinitionsPortable}
\usage{
ListBatchPredictionJobDefinitionsPortable(jobDefinitionId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve job definition snippet for PBP
}
\details{
Retrieve a Batch Prediction job definition for Portable Batch Predictions

This method invokes \verb{GET /batchPredictionJobDefinitions/\{jobDefinitionId\}/portable/} in the DataRobot Public API.
}
\seealso{
The method BatchPredictionJobDefinitionsPortableList in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
