% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{ListBatchPredictions}
\alias{ListBatchPredictions}
\title{ListBatchPredictions}
\usage{
ListBatchPredictions(
  modelId,
  outputType,
  orderBy,
  source,
  endDateTime,
  jobId,
  hostname,
  startDateTime,
  deploymentId,
  cutoffHours,
  batchPredictionJobDefinitionId,
  intakeType,
  status,
  offset = 0,
  allJobs = FALSE,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List batch prediction jobs
}
\details{
Get a collection of batch prediction jobs by statuses

This method invokes \verb{GET /batchPredictions/} in the DataRobot Public API.
}
\seealso{
The method BatchPredictionsList in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
