% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints_apiWrapper.R
\name{ListBlueprints}
\alias{ListBlueprints}
\title{ListBlueprints}
\usage{
ListBlueprints(project, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List blueprints
}
\details{
List appropriate blueprints for the project.

This method invokes \verb{GET /projects/\{projectId\}/blueprints/} in the DataRobot Public API.
}
\seealso{
The method ProjectsBlueprintsList in \link{BlueprintsApi}, which invokes the same underlying API endpoint.
}
