% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{ListCalendarCountryCodes}
\alias{ListCalendarCountryCodes}
\title{ListCalendarCountryCodes}
\usage{
ListCalendarCountryCodes(offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the list of allowed country codes to request preloaded calendars generation for.
}
\details{
Retrieve the list of allowed country codes to request preloaded calendars generation for.

This method invokes \verb{GET /calendarCountryCodes/} in the DataRobot Public API.
}
\seealso{
The method CalendarCountryCodesList in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
