% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{ListCalendarsAccessControl}
\alias{ListCalendarsAccessControl}
\title{ListCalendarsAccessControl}
\usage{
ListCalendarsAccessControl(
  calendarId,
  userId,
  username,
  offset = 0,
  limit = 0,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get a list of users who have access to this calendar and their roles on the calendar.
}
\details{
Get a list of users who have access to this calendar and their roles on the calendar.

This method invokes \verb{GET /calendars/\{calendarId\}/accessControl/} in the DataRobot Public API.
}
\seealso{
The method CalendarsAccessControlList in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
