% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{ListChangeRequestsSuggestedReviewers}
\alias{ListChangeRequestsSuggestedReviewers}
\title{ListChangeRequestsSuggestedReviewers}
\usage{
ListChangeRequestsSuggestedReviewers(
  changeRequestId,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List suggested reviewers.
}
\details{
List users, suggested to review the Change Request.

This method invokes \verb{GET /changeRequests/\{changeRequestId\}/suggestedReviewers/} in the DataRobot Public API.
}
\seealso{
The method ChangeRequestsSuggestedReviewersList in \link{GovernanceApi}, which invokes the same underlying API endpoint.
}
