% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure_apiWrapper.R
\name{ListClusterLicense}
\alias{ListClusterLicense}
\title{ListClusterLicense}
\usage{
ListClusterLicense(...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the information about cluster license
}
\details{
Retrieve the information about the currently deployed cluster license.

This method invokes \verb{GET /clusterLicense/} in the DataRobot Public API.
}
\seealso{
The method ClusterLicenseList in \link{InfrastructureApi}, which invokes the same underlying API endpoint.
}
