% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_apiWrapper.R
\name{ListComplianceDocTemplatesDefault}
\alias{ListComplianceDocTemplatesDefault}
\title{ListComplianceDocTemplatesDefault}
\usage{
ListComplianceDocTemplatesDefault(type = "normal", ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the default documentation template
}
\details{
Retrieve the default documentation template.

This method invokes \verb{GET /complianceDocTemplates/default/} in the DataRobot Public API.
}
\seealso{
The method ComplianceDocTemplatesDefaultList in \link{DocumentationApi}, which invokes the same underlying API endpoint.
}
