% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_apiWrapper.R
\name{ListCredentials}
\alias{ListCredentials}
\title{ListCredentials}
\usage{
ListCredentials(types, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List credentials.
}
\details{
List all sets of credentials available for given user.

This method invokes \verb{GET /credentials/} in the DataRobot Public API.
}
\seealso{
The method CredentialsList in \link{CredentialsApi}, which invokes the same underlying API endpoint.
}
