% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListCustomModels}
\alias{ListCustomModels}
\title{ListCustomModels}
\usage{
ListCustomModels(
  isDeployed,
  orderBy,
  customModelType,
  searchFor,
  offset = 0,
  limit = 1000,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List custom models.
}
\details{
Retrieve metadata for all custom models the user has access to.

This method invokes \verb{GET /customModels/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
