% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListCustomModelsAccessControl}
\alias{ListCustomModelsAccessControl}
\title{ListCustomModelsAccessControl}
\usage{
ListCustomModelsAccessControl(customModelId, offset = 0, limit = 1000, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get a list of users who have access to this custom model and their roles on it.
}
\details{
Get a list of users who have access to this custom model and their roles on it.

This method invokes \verb{GET /customModels/\{customModelId\}/accessControl/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsAccessControlList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
