% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListCustomModelsVersionsDependencyBuildLog}
\alias{ListCustomModelsVersionsDependencyBuildLog}
\title{ListCustomModelsVersionsDependencyBuildLog}
\usage{
ListCustomModelsVersionsDependencyBuildLog(
  customModelId,
  customModelVersionId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the custom model version's dependency build log.
}
\details{
Retrieve the custom model version's dependency build log.

This method invokes \verb{GET /customModels/\{customModelId\}/versions/\{customModelVersionId\}/dependencyBuildLog/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsVersionsDependencyBuildLogList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
