% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListCustomModelTestsLog}
\alias{ListCustomModelTestsLog}
\title{ListCustomModelTestsLog}
\usage{
ListCustomModelTestsLog(customModelTestId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get custom model test log.
}
\details{
Retrieve the logs from a model testing attempt.

This method invokes \verb{GET /customModelTests/\{customModelTestId\}/log/} in the DataRobot Public API.
}
\seealso{
The method CustomModelTestsLogList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
