% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListCustomModelTestsTail}
\alias{ListCustomModelTestsTail}
\title{ListCustomModelTestsTail}
\usage{
ListCustomModelTestsTail(customModelTestId, lines = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get custom model test log tail.
}
\details{
Retrieve the last N lines of logs from a model testing attempt.

This method invokes \verb{GET /customModelTests/\{customModelTestId\}/tail/} in the DataRobot Public API.
}
\seealso{
The method CustomModelTestsTailList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
