% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{ListCustomTasks}
\alias{ListCustomTasks}
\title{ListCustomTasks}
\usage{
ListCustomTasks(orderBy, searchFor, offset = 0, limit = 1000, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List custom tasks.
}
\details{
Retrieve metadata for all custom tasks the user has access to.

This method invokes \verb{GET /customTasks/} in the DataRobot Public API.
}
\seealso{
The method CustomTasksList in \link{CustomTasksApi}, which invokes the same underlying API endpoint.
}
