% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{ListCustomTasksVersionsDownload}
\alias{ListCustomTasksVersionsDownload}
\title{ListCustomTasksVersionsDownload}
\usage{
ListCustomTasksVersionsDownload(customTaskVersionId, customTaskId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Download custom task version content.
}
\details{
Download a specific item bundle from a custom task as a zip compressed archive.

This method invokes \verb{GET /customTasks/\{customTaskId\}/versions/\{customTaskVersionId\}/download/} in the DataRobot Public API.
}
\seealso{
The method CustomTasksVersionsDownloadList in \link{CustomTasksApi}, which invokes the same underlying API endpoint.
}
