% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{ListDatasetsFeaturelists}
\alias{ListDatasetsFeaturelists}
\title{ListDatasetsFeaturelists}
\usage{
ListDatasetsFeaturelists(
  datasetId,
  offset = 0,
  limit = 100,
  orderBy = "name",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve dataset featurelists
}
\details{
Retrieves the featurelists of the dataset with given ID and the latest dataset version.

This method invokes \verb{GET /datasets/\{datasetId\}/featurelists/} in the DataRobot Public API.
}
\seealso{
The method DatasetsFeaturelistsList in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
