% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{ListDatasetsFile}
\alias{ListDatasetsFile}
\title{ListDatasetsFile}
\usage{
ListDatasetsFile(datasetId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve original dataset data
}
\details{
Retrieve all the originally uploaded data, in CSV form.

This method invokes \verb{GET /datasets/\{datasetId\}/file/} in the DataRobot Public API.
}
\seealso{
The method DatasetsFileList in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
