% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R
\name{ListDatasetsImagesDataQualityLogFile}
\alias{ListDatasetsImagesDataQualityLogFile}
\title{ListDatasetsImagesDataQualityLogFile}
\usage{
ListDatasetsImagesDataQualityLogFile(datasetId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a text file containing the images data quality log.
}
\details{
Retrieve a text file containing the images data quality log.

This method invokes \verb{GET /datasets/\{datasetId\}/imagesDataQualityLog/file/} in the DataRobot Public API.
}
\seealso{
The method DatasetsImagesDataQualityLogFileList in \link{ImagesApi}, which invokes the same underlying API endpoint.
}
