% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{ListDatasetsProjects}
\alias{ListDatasetsProjects}
\title{ListDatasetsProjects}
\usage{
ListDatasetsProjects(datasetId, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get dataset projects
}
\details{
Retrieves a dataset's projects by dataset ID.

This method invokes \verb{GET /datasets/\{datasetId\}/projects/} in the DataRobot Public API.
}
\seealso{
The method DatasetsProjectsList in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
