% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{ListDatasetsSharedRoles}
\alias{ListDatasetsSharedRoles}
\title{ListDatasetsSharedRoles}
\usage{
ListDatasetsSharedRoles(
  shareRecipientType,
  name,
  datasetId,
  id,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List dataset shared roles
}
\details{
Get a list of users, groups and organizations who have access to this dataset and their roles.

This method invokes \verb{GET /datasets/\{datasetId\}/sharedRoles/} in the DataRobot Public API.
}
\seealso{
The method DatasetsSharedRolesList in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
