% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{ListDatasetsVersionsFile}
\alias{ListDatasetsVersionsFile}
\title{ListDatasetsVersionsFile}
\usage{
ListDatasetsVersionsFile(datasetId, datasetVersionId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve original dataset data
}
\details{
Retrieve all the originally uploaded data, in CSV form.

This method invokes \verb{GET /datasets/\{datasetId\}/versions/\{datasetVersionId\}/file/} in the DataRobot Public API.
}
\seealso{
The method DatasetsVersionsFileList in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
