% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{ListDatasetsVersionsProjects}
\alias{ListDatasetsVersionsProjects}
\title{ListDatasetsVersionsProjects}
\usage{
ListDatasetsVersionsProjects(
  datasetId,
  datasetVersionId,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get dataset projects by version
}
\details{
Retrieves a dataset's projects for the specified catalog dataset and dataset version id.

This method invokes \verb{GET /datasets/\{datasetId\}/versions/\{datasetVersionId\}/projects/} in the DataRobot Public API.
}
\seealso{
The method DatasetsVersionsProjectsList in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
