% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{ListDataStores}
\alias{ListDataStores}
\title{ListDataStores}
\usage{
ListDataStores(
  offset = 0,
  limit = 100,
  type = "jdbc",
  showHidden = "False",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List data stores.
}
\details{
Return list with details of the existing data stores available for the user.

This method invokes \verb{GET /externalDataStores/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataStoresList in \link{DataConnectivityApi}, which invokes the same underlying API endpoint.
}
