% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeployments}
\alias{ListDeployments}
\title{ListDeployments}
\usage{
ListDeployments(
  orderBy = NULL,
  search = NULL,
  offset = 0,
  limit = 20,
  serviceHealth = NULL,
  modelHealth = NULL,
  accuracyHealth = NULL,
  role = NULL,
  status = NULL,
  importance = NULL,
  lastPredictionTimestampStart = NULL,
  lastPredictionTimestampEnd = NULL,
  predictionUsageDailyAvgGreaterThan = NULL,
  predictionUsageDailyAvgLessThan = NULL,
  defaultPredictionServerId = NULL,
  buildEnvironmentType = NULL,
  executionEnvironmentType = NULL,
  predictionEnvironmentPlatform = NULL,
  hasDecisionFlow = NULL,
  createdByMe = NULL,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List deployments
}
\details{
List deployments a user can view.

This method invokes \verb{GET /deployments/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsList in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
