% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeploymentsChallengers}
\alias{ListDeploymentsChallengers}
\title{ListDeploymentsChallengers}
\usage{
ListDeploymentsChallengers(deploymentId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List challenger models
}
\details{
List challenger models for deployment.

This method invokes \verb{GET /deployments/\{deploymentId\}/challengers/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsChallengersList in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
