% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeploymentsFeatures}
\alias{ListDeploymentsFeatures}
\title{ListDeploymentsFeatures}
\usage{
ListDeploymentsFeatures(
  search,
  deploymentId,
  offset = 0,
  includeNonPredictionFeatures = "false",
  limit = 50,
  forSegmentedAnalysis = "false",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get deployment features
}
\details{
Retrieve features contained within the universe dataset associated with a specific deployment. By default, this returns all raw features required for predictions.

This method invokes \verb{GET /deployments/\{deploymentId\}/features/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsFeaturesList in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
