% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeploymentsHumilityStats}
\alias{ListDeploymentsHumilityStats}
\title{ListDeploymentsHumilityStats}
\usage{
ListDeploymentsHumilityStats(
  segmentAttribute,
  modelId,
  deploymentId,
  start,
  end,
  bucketSize,
  segmentValue = "",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve humility stats
}
\details{
Retrieve humility rule service triggers statistics overview.

This method invokes \verb{GET /deployments/\{deploymentId\}/humilityStats/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsHumilityStatsList in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
