% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeploymentsHumilityStatsOverTime}
\alias{ListDeploymentsHumilityStatsOverTime}
\title{ListDeploymentsHumilityStatsOverTime}
\usage{
ListDeploymentsHumilityStatsOverTime(
  segmentAttribute,
  modelId,
  deploymentId,
  start,
  end,
  bucketSize,
  segmentValue = "",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve humility stats over time
}
\details{
Retrieve humility service statistics over time.

This method invokes \verb{GET /deployments/\{deploymentId\}/humilityStatsOverTime/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsHumilityStatsOverTimeList in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
