% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeploymentsModelSecondaryDatasetConfigurationHistory}
\alias{ListDeploymentsModelSecondaryDatasetConfigurationHistory}
\title{ListDeploymentsModelSecondaryDatasetConfigurationHistory}
\usage{
ListDeploymentsModelSecondaryDatasetConfigurationHistory(
  deploymentId,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List the secondary datasets configuration history for a deployment
}
\details{
List all the secondary datasets configuration used by a given Feature Discovery deployment

This method invokes \verb{GET /deployments/\{deploymentId\}/model/secondaryDatasetConfigurationHistory/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsModelSecondaryDatasetConfigurationHistoryList in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
