% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeploymentsPredictionResults}
\alias{ListDeploymentsPredictionResults}
\title{ListDeploymentsPredictionResults}
\usage{
ListDeploymentsPredictionResults(
  actualsPresent,
  modelId,
  deploymentId,
  start,
  end,
  accept,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve predictions results.
}
\details{
Retrieve predictions results of the deployment.

This method invokes \verb{GET /deployments/\{deploymentId\}/predictionResults/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsPredictionResultsList in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
