% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeploymentsRetrainingPolicies}
\alias{ListDeploymentsRetrainingPolicies}
\title{ListDeploymentsRetrainingPolicies}
\usage{
ListDeploymentsRetrainingPolicies(deploymentId, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Endpoint for fetching a list of deployment retraining policies.
}
\details{
Retrieve a list of deployment retraining policies.

This method invokes \verb{GET /deployments/\{deploymentId\}/retrainingPolicies/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsRetrainingPoliciesList in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
