% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeploymentsServiceStatsOverTime}
\alias{ListDeploymentsServiceStatsOverTime}
\title{ListDeploymentsServiceStatsOverTime}
\usage{
ListDeploymentsServiceStatsOverTime(
  segmentAttribute,
  modelId,
  quantile,
  deploymentId,
  start,
  threshold,
  end,
  bucketSize,
  metric = "totalPredictions",
  segmentValue = "",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve service statistics over time
}
\details{
Retrieve deployment service health statistics over time data for one single metric.

This method invokes \verb{GET /deployments/\{deploymentId\}/serviceStatsOverTime/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsServiceStatsOverTimeList in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
