% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListDeploymentsSharedRoles}
\alias{ListDeploymentsSharedRoles}
\title{ListDeploymentsSharedRoles}
\usage{
ListDeploymentsSharedRoles(
  shareRecipientType,
  deploymentId,
  name,
  id,
  offset = 0,
  limit = 10,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get model deployment's access control list
}
\details{
Get a list of users, groups and organizations who have access to this deployment and their roles.

This method invokes \verb{GET /deployments/\{deploymentId\}/sharedRoles/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsSharedRolesList in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
