% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_apiWrapper.R
\name{ListEventLogsEvents}
\alias{ListEventLogsEvents}
\title{ListEventLogsEvents}
\usage{
ListEventLogsEvents(...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve all the available events. DEPRECATED API.
}
\details{
Retrieve all the available events. DEPRECATED API.

This method invokes \verb{GET /eventLogs/events/} in the DataRobot Public API.
}
\seealso{
The method EventLogsEventsList in \link{AnalyticsApi}, which invokes the same underlying API endpoint.
}
