% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{ListExternalDataDriversConfiguration}
\alias{ListExternalDataDriversConfiguration}
\title{ListExternalDataDriversConfiguration}
\usage{
ListExternalDataDriversConfiguration(driverId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Driver configuration details.
}
\details{
Retrieve the driver configuration with the specified ID.

This method invokes \verb{GET /externalDataDrivers/\{driverId\}/configuration/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataDriversConfigurationList in \link{DataConnectivityApi}, which invokes the same underlying API endpoint.
}
