% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{ListExternalDriverConfigurations}
\alias{ListExternalDriverConfigurations}
\title{ListExternalDriverConfigurations}
\usage{
ListExternalDriverConfigurations(
  showHidden = "False",
  offset = 0,
  limit = 100,
  type = "jdbc",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List available driver configurations.
}
\details{
Retrieve matching driver configurations based on query.

This method invokes \verb{GET /externalDriverConfigurations/} in the DataRobot Public API.
}
\seealso{
The method ExternalDriverConfigurationsList in \link{DataConnectivityApi}, which invokes the same underlying API endpoint.
}
