% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{ListFeatureInfo}
\alias{ListFeatureInfo}
\title{ListFeatureInfo}
\usage{
ListFeatureInfo(project, forSegmentedAnalysis = "false", ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List project features
}
\details{
List the features from a project with descriptive information.

This method invokes \verb{GET /projects/\{projectId\}/features/} in the DataRobot Public API.
}
\seealso{
The method FeaturesList in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
