% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{ListFeaturelists}
\alias{ListFeaturelists}
\title{ListFeaturelists}
\usage{
ListFeaturelists(project, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List featurelists
}
\details{
List all featurelists for a project.

This method invokes \verb{GET /projects/\{projectId\}/featurelists/} in the DataRobot Public API.
}
\seealso{
The method FeaturelistsList in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
