% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_apiWrapper.R
\name{ListForCredentialCredentialsAssociations}
\alias{ListForCredentialCredentialsAssociations}
\title{ListForCredentialCredentialsAssociations}
\usage{
ListForCredentialCredentialsAssociations(
  types,
  credentialId,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List all objects associated with the specified credentials.
}
\details{
List all objects associated with specified credentials for the current user.

This method invokes \verb{GET /credentials/\{credentialId\}/associations/} in the DataRobot Public API.
}
\seealso{
The method CredentialsAssociationsListForCredential in \link{CredentialsApi}, which invokes the same underlying API endpoint.
}
