% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{ListGroupsUsers}
\alias{ListGroupsUsers}
\title{ListGroupsUsers}
\usage{
ListGroupsUsers(
  groupId,
  orderBy,
  isAdmin,
  isActive,
  offset = 0,
  limit = 50,
  namePart = "",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List users in group
}
\details{
List users in the group.

This method invokes \verb{GET /groups/\{groupId\}/users/} in the DataRobot Public API.
}
\seealso{
The method GroupsUsersList in \link{UserManagementApi}, which invokes the same underlying API endpoint.
}
