% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{ListJobs}
\alias{ListJobs}
\title{ListJobs}
\usage{
ListJobs(project, status = NULL, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List project jobs
}
\details{
List the project's jobs.

This method invokes \verb{GET /projects/\{projectId\}/jobs/} in the DataRobot Public API.
}
\seealso{
The method JobsList in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
