% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{ListModelingFeaturelists}
\alias{ListModelingFeaturelists}
\title{ListModelingFeaturelists}
\usage{
ListModelingFeaturelists(project, offset = 0, limit = 0, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List all modeling featurelists from a project
}
\details{
List all modeling featurelists from the project requested by ID. This route will only become available after the target and partitioning options have been set for a project. Modeling featurelists are featurelists of modeling features, and are the correct featurelists to use when creating models or restarting the autopilot. In a time series project, these will differ from those returned from :\url{http:get}:/api/v2/projects/(projectId)/featurelists/ while in other projects these will be identical. See the \verb{documentation <input_vs_modeling>} for more information on the distinction between input and modeling data in time series projects.

This method invokes \verb{GET /projects/\{projectId\}/modelingFeaturelists/} in the DataRobot Public API.
}
\seealso{
The method ModelingFeaturelistsList in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
