% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListModelPackagesCapabilities}
\alias{ListModelPackagesCapabilities}
\title{ListModelPackagesCapabilities}
\usage{
ListModelPackagesCapabilities(modelPackageId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve capabilities.
}
\details{
Retrieve the capabilities for the model package.

This method invokes \verb{GET /modelPackages/\{modelPackageId\}/capabilities/} in the DataRobot Public API.
}
\seealso{
The method ModelPackagesCapabilitiesList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
