% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListModelPackagesSharedRoles}
\alias{ListModelPackagesSharedRoles}
\title{ListModelPackagesSharedRoles}
\usage{
ListModelPackagesSharedRoles(
  shareRecipientType,
  name,
  id,
  modelPackageId,
  offset = 0,
  limit = 10,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get model package's access control list
}
\details{
Get a list of users, groups and organizations who have access to this template and their roles on the template.

This method invokes \verb{GET /modelPackages/\{modelPackageId\}/sharedRoles/} in the DataRobot Public API.
}
\seealso{
The method ModelPackagesSharedRolesList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
