% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListModelRecommendations}
\alias{ListModelRecommendations}
\title{ListModelRecommendations}
\usage{
ListModelRecommendations(project, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List recommended models for the project
}
\details{
Retrieves all of the current recommended models for the project

This method invokes \verb{GET /projects/\{projectId\}/recommendedModels/} in the DataRobot Public API.
}
\seealso{
The method ProjectsRecommendedModelsList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
