% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListModels}
\alias{ListModels}
\title{ListModels}
\usage{
ListModels(
  project,
  orderBy = NULL,
  filter = NULL,
  bulkOperationId = NULL,
  name = NULL,
  samplePct = NULL,
  withMetric = NULL,
  isStarred = NULL,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List project models
}
\details{
Lists all the models from a project.

This method invokes \verb{GET /projects/\{projectId\}/models/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
