% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListMultilabelInsightsHistogram}
\alias{ListMultilabelInsightsHistogram}
\title{ListMultilabelInsightsHistogram}
\usage{
ListMultilabelInsightsHistogram(multilabelInsightsKey, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve multicategorical feature histogram.
}
\details{
Retrieve multicategorical feature histogram.

This method invokes \verb{GET /multilabelInsights/\{multilabelInsightsKey\}/histogram/} in the DataRobot Public API.
}
\seealso{
The method MultilabelInsightsHistogramList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
