% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListMultilabelInsightsPairwiseStatistics}
\alias{ListMultilabelInsightsPairwiseStatistics}
\title{ListMultilabelInsightsPairwiseStatistics}
\usage{
ListMultilabelInsightsPairwiseStatistics(
  statisticType,
  multilabelInsightsKey,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve pairwise statistics for the given multilabel insights key.
}
\details{
Retrieve multilabel specific pairwise label statistics for the given multilabel insights key: correlation, joint probability and conditional probability.

This method invokes \verb{GET /multilabelInsights/\{multilabelInsightsKey\}/pairwiseStatistics/} in the DataRobot Public API.
}
\seealso{
The method MultilabelInsightsPairwiseStatisticsList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
