% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{ListNotificationChannels}
\alias{ListNotificationChannels}
\title{ListNotificationChannels}
\usage{
ListNotificationChannels(namePart, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List notification channels
}
\details{
List the notification channels according to the query.

This method invokes \verb{GET /notificationChannels/} in the DataRobot Public API.
}
\seealso{
The method NotificationChannelsList in \link{NotificationsApi}, which invokes the same underlying API endpoint.
}
