% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{ListNotificationEvents}
\alias{ListNotificationEvents}
\title{ListNotificationEvents}
\usage{
ListNotificationEvents(...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List event types and groups the user can include in notification policies.
}
\details{
List event types and groups the user can include in notification policies. Events and groups are filtered by user permissions and event properties. It is not a complete list of all defined events; instead, it is a list of the events available to the user.

This method invokes \verb{GET /notificationEvents/} in the DataRobot Public API.
}
\seealso{
The method NotificationEventsList in \link{NotificationsApi}, which invokes the same underlying API endpoint.
}
