% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{ListNotificationLogs}
\alias{ListNotificationLogs}
\title{ListNotificationLogs}
\usage{
ListNotificationLogs(policyId, channelId, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List the notification logs
}
\details{
List the notification logs that correspond to provided conditions. Default ordering is desc by notification log timestamp

This method invokes \verb{GET /notificationLogs/} in the DataRobot Public API.
}
\seealso{
The method NotificationLogsList in \link{NotificationsApi}, which invokes the same underlying API endpoint.
}
