% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{ListOrganizations}
\alias{ListOrganizations}
\title{ListOrganizations}
\usage{
ListOrganizations(offset = 0, limit = 100, namePart = "", ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List organizations
}
\details{
List organizations available in the system.

This method invokes \verb{GET /organizations/} in the DataRobot Public API.
}
\seealso{
The method OrganizationsList in \link{UserManagementApi}, which invokes the same underlying API endpoint.
}
