% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{ListPredictionDatasets}
\alias{ListPredictionDatasets}
\title{ListPredictionDatasets}
\usage{
ListPredictionDatasets(project, offset = 0, limit = 0, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List predictions datasets uploaded to a project.
}
\details{
List predictions datasets uploaded to a project.

This method invokes \verb{GET /projects/\{projectId\}/predictionDatasets/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPredictionDatasetsList in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
