% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R,
%   R/predictions_apiWrapper.R
\name{ListPredictionExplanationsMetadata}
\alias{ListPredictionExplanationsMetadata}
\title{ListPredictionExplanationsMetadata}
\usage{
ListPredictionExplanationsMetadata(
  project,
  modelId = NULL,
  limit = NULL,
  offset = NULL,
  ...
)

ListPredictionExplanationsMetadata(
  project,
  modelId = NULL,
  limit = NULL,
  offset = NULL,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List PredictionExplanationsRecord objects for a project.

List PredictionExplanationsRecord objects for a project.
}
\details{
List PredictionExplanationsRecord objects for a project. These contain metadata about the computed prediction explanations and the location at which the PredictionExplanations can be retrieved.

This method invokes \verb{GET /projects/\{projectId\}/predictionExplanationsRecords/} in the DataRobot Public API.

List PredictionExplanationsRecord objects for a project. These contain metadata about the computed prediction explanations and the location at which the PredictionExplanations can be retrieved.

This method invokes \verb{GET /projects/\{projectId\}/predictionExplanationsRecords/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPredictionExplanationsRecordsList in \link{InsightsApi}, which invokes the same underlying API endpoint.

The method ProjectsPredictionExplanationsRecordsList in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
