% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{ListPredictions}
\alias{ListPredictions}
\title{ListPredictions}
\usage{
ListPredictions(
  project,
  modelId = NULL,
  datasetId = NULL,
  offset = 0,
  limit = 1000,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get a list of prediction records.
}
\details{
Get a list of prediction records.  v2.21 Use :\url{http:get}:/api/v2/projects/(projectId)/predictionsMetadata/ instead. The only difference is that parameter datasetId is renamed to predictionDatasetId both in request and response.

This method invokes \verb{GET /projects/\{projectId\}/predictions/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPredictionsList in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
