% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ListPredictionServers}
\alias{ListPredictionServers}
\title{ListPredictionServers}
\usage{
ListPredictionServers(offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List prediction servers.
}
\details{
List prediction servers available for a user.

This method invokes \verb{GET /predictionServers/} in the DataRobot Public API.
}
\seealso{
The method PredictionServersList in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
