% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListPrimeFiles}
\alias{ListPrimeFiles}
\title{ListPrimeFiles}
\usage{
ListPrimeFiles(
  project,
  parentModelId = NULL,
  modelId = NULL,
  offset = 0,
  limit = 0,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get Prime Files
}
\details{
List all DataRobot Prime files available for download

This method invokes \verb{GET /projects/\{projectId\}/primeFiles/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPrimeFilesList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
