% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{ListProjectCleanupJobsDownload}
\alias{ListProjectCleanupJobsDownload}
\title{ListProjectCleanupJobsDownload}
\usage{
ListProjectCleanupJobsDownload(statusId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Download a projects permadeletion report.
}
\details{
Get a file containing a per-project report of permanent deletion.

This method invokes \verb{GET /projectCleanupJobs/\{statusId\}/download/} in the DataRobot Public API.
}
\seealso{
The method CleanupJobsDownloadList in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
