% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{ListProjectsAccessControl}
\alias{ListProjectsAccessControl}
\title{ListProjectsAccessControl}
\usage{
ListProjectsAccessControl(
  userId,
  projectId,
  username,
  offset = 0,
  limit = 0,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get project's access control list
}
\details{
Get a list of users who have access to this project and their roles on the project.

This method invokes \verb{GET /projects/\{projectId\}/accessControl/} in the DataRobot Public API.
}
\seealso{
The method AccessControlList in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
