% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListProjectsBiasMitigationFeatureInfo}
\alias{ListProjectsBiasMitigationFeatureInfo}
\title{ListProjectsBiasMitigationFeatureInfo}
\usage{
ListProjectsBiasMitigationFeatureInfo(featureName, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get bias mitigation data quality information for a given projectId and featureName.
}
\details{
Get bias mitigation data quality information for a given projectId and featureName.

This method invokes \verb{GET /projects/\{projectId\}/biasMitigationFeatureInfo/} in the DataRobot Public API.
}
\seealso{
The method ProjectsBiasMitigationFeatureInfoList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
