% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsBiasVsAccuracyInsights}
\alias{ListProjectsBiasVsAccuracyInsights}
\title{ListProjectsBiasVsAccuracyInsights}
\usage{
ListProjectsBiasVsAccuracyInsights(
  protectedFeature,
  accuracyMetric,
  projectId,
  fairnessMetric,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List Bias vs Accuracy insights.
}
\details{
Retrieve a list of Bias vs Accuracy insights for the model.

This method invokes \verb{GET /projects/\{projectId\}/biasVsAccuracyInsights/} in the DataRobot Public API.
}
\seealso{
The method ProjectsBiasVsAccuracyInsightsList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
