% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListProjectsBlenderModels}
\alias{ListProjectsBlenderModels}
\title{ListProjectsBlenderModels}
\usage{
ListProjectsBlenderModels(projectId, offset = 0, limit = 0, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List all blenders in a project.
}
\details{
List all blenders in a project.

This method invokes \verb{GET /projects/\{projectId\}/blenderModels/} in the DataRobot Public API.
}
\seealso{
The method ProjectsBlenderModelsList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
