% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{ListProjectsCalendarEvents}
\alias{ListProjectsCalendarEvents}
\title{ListProjectsCalendarEvents}
\usage{
ListProjectsCalendarEvents(
  endDate,
  projectId,
  startDate,
  seriesId,
  offset = 0,
  limit = 1000,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List available calendar events for the project.
}
\details{
List available calendar events for the project.

This method invokes \verb{GET /projects/\{projectId\}/calendarEvents/} in the DataRobot Public API.
}
\seealso{
The method CalendarEventsList in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
