% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListProjectsCombinedModels}
\alias{ListProjectsCombinedModels}
\title{ListProjectsCombinedModels}
\usage{
ListProjectsCombinedModels(projectId, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve all existing combined models for this project.
}
\details{
Retrieve all existing combined models for this project. .. note:      To retrieve information on the segments for a combined model, retrieve the combined model using :http:get:\verb{/api/v2/projects/(projectId)/combinedModels/(combinedModelId)/}

This method invokes \verb{GET /projects/\{projectId\}/combinedModels/} in the DataRobot Public API.
}
\seealso{
The method ProjectsCombinedModelsList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
