% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListProjectsCombinedModelsSegmentsDownload}
\alias{ListProjectsCombinedModelsSegmentsDownload}
\title{ListProjectsCombinedModelsSegmentsDownload}
\usage{
ListProjectsCombinedModelsSegmentsDownload(combinedModelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Download Combined Model segments info.
}
\details{
Download Combined Model segments info (name, related project & model details) as a CSV.

This method invokes \verb{GET /projects/\{projectId\}/combinedModels/\{combinedModelId\}/segments/download/} in the DataRobot Public API.
}
\seealso{
The method ProjectsCombinedModelsSegmentsDownloadList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
