% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsDatetimeModelsFeatureEffects}
\alias{ListProjectsDatetimeModelsFeatureEffects}
\title{ListProjectsDatetimeModelsFeatureEffects}
\usage{
ListProjectsDatetimeModelsFeatureEffects(
  backtestIndex,
  modelId,
  projectId,
  includeIcePlots,
  source = "training",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve Feature Effects for a model backtest.
}
\details{
Retrieve Feature Effects for a model backtest. Feature Effects provides partial dependence and predicted vs actual values for the top 500 features, ordered by feature impact score. The partial dependence shows marginal effect of a feature on the target variable after accounting for the average effects of all other predictive features. It indicates how, holding all other variables except the feature of interest as they were, the value of this feature affects your prediction. If a Feature Effects job was previously submitted for a given backtest, this endpoint will return a response structured as {\"message\": \\<message\\>, \"jobId\": \\<jobId\\>} where jobId is the ID of the job. Retrieve the job with :\url{http:get}:/api/v2/projects/(projectId)/jobs/(jobId)/

This method invokes \verb{GET /projects/\{projectId\}/datetimeModels/\{modelId\}/featureEffects/} in the DataRobot Public API.
}
\seealso{
The method ProjectsDatetimeModelsFeatureEffectsList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
