% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsDatetimeModelsFeatureEffectsMetadata}
\alias{ListProjectsDatetimeModelsFeatureEffectsMetadata}
\title{ListProjectsDatetimeModelsFeatureEffectsMetadata}
\usage{
ListProjectsDatetimeModelsFeatureEffectsMetadata(modelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve Feature Effects metadata for each backtest. Response contains status and available sources for each backtest of the model.
}
\details{
Retrieve Feature Effects metadata for each backtest. Response contains status and available sources for each backtest of the model. One of the provided backtestIndex indexes used for submitting the compute request and retrieving Feature Effects. \\* Start/stop models contain a single backtestIndex response value of startstop. \\* Other models contain backtestIndex of 0, 1, ..., holdout. One of the provided source parameters used for retrieving Feature Effects. \\* Each backtest source can be, at a minimum, training or validation. If holdout is configured for the project, backtestIndex also includes holdout with sources training and holdout. \\* Source value of training is always available. (versions prior to v2.17 support validation only) \\* When a start/stop model is trained into validation or holdout without stacked predictions (i.e., no out-of-sample predictions in validation or holdout), validation and holdout sources are not available. \\* Source holdout is not available when there is no holdout configured for the project.

This method invokes \verb{GET /projects/\{projectId\}/datetimeModels/\{modelId\}/featureEffectsMetadata/} in the DataRobot Public API.
}
\seealso{
The method ProjectsDatetimeModelsFeatureEffectsMetadataList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
