% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsDatetimeModelsForecastDistanceStabilityPlot}
\alias{ListProjectsDatetimeModelsForecastDistanceStabilityPlot}
\title{ListProjectsDatetimeModelsForecastDistanceStabilityPlot}
\usage{
ListProjectsDatetimeModelsForecastDistanceStabilityPlot(
  metricName,
  modelId,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a plot displaying the stability of the time series model across different forecast distances.
}
\details{
Retrieve a plot displaying the stability of the time series model across different forecast distances. .. note:: All durations and datetimes are specified in accordance with the \verb{timestamp and duration formatting rules <time_format>}.

This method invokes \verb{GET /projects/\{projectId\}/datetimeModels/\{modelId\}/forecastDistanceStabilityPlot/} in the DataRobot Public API.
}
\seealso{
The method ProjectsDatetimeModelsForecastDistanceStabilityPlotList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
