% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R
\name{ListProjectsDuplicateImages}
\alias{ListProjectsDuplicateImages}
\title{ListProjectsDuplicateImages}
\usage{
ListProjectsDuplicateImages(column, projectId, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get a list of duplicate images containing the number of occurrences of each image
}
\details{
Get a list of duplicate images containing the number of occurrences of each image

This method invokes \verb{GET /projects/\{projectId\}/duplicateImages/\{column\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsDuplicateImagesList in \link{ImagesApi}, which invokes the same underlying API endpoint.
}
