% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsFeatureAssociationFeaturelists}
\alias{ListProjectsFeatureAssociationFeaturelists}
\title{ListProjectsFeatureAssociationFeaturelists}
\usage{
ListProjectsFeatureAssociationFeaturelists(projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List all featurelists with feature association matrix availability flags for a project.
}
\details{
List all featurelists with feature association matrix availability flags for a project.

This method invokes \verb{GET /projects/\{projectId\}/featureAssociationFeaturelists/} in the DataRobot Public API.
}
\seealso{
The method ProjectsFeatureAssociationFeaturelistsList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
