% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{ListProjectsFeatureDiscoveryLogs}
\alias{ListProjectsFeatureDiscoveryLogs}
\title{ListProjectsFeatureDiscoveryLogs}
\usage{
ListProjectsFeatureDiscoveryLogs(projectId, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the feature discovery log content and log length
}
\details{
Retrieve the feature discovery log content and log length for a feature discovery project. This route is only supported for feature discovery projects that have finished partitioning

This method invokes \verb{GET /projects/\{projectId\}/featureDiscoveryLogs/} in the DataRobot Public API.
}
\seealso{
The method FeatureDiscoveryLogsList in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
