% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{ListProjectsFeatureDiscoveryRecipeSQLsDownload}
\alias{ListProjectsFeatureDiscoveryRecipeSQLsDownload}
\title{ListProjectsFeatureDiscoveryRecipeSQLsDownload}
\usage{
ListProjectsFeatureDiscoveryRecipeSQLsDownload(
  modelId,
  projectId,
  statusOnly = "false",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Download feature discovery SQL recipe
}
\details{
Download feature discovery SQL recipe for a project

This method invokes \verb{GET /projects/\{projectId\}/featureDiscoveryRecipeSQLs/download/} in the DataRobot Public API.
}
\seealso{
The method FeatureDiscoveryRecipeSQLsDownloadList in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
