% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListProjectsFrozenModels}
\alias{ListProjectsFrozenModels}
\title{ListProjectsFrozenModels}
\usage{
ListProjectsFrozenModels(withMetric, projectId, offset = 0, limit = 0, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List all frozen models from a project.
}
\details{
List all frozen models from a project.

This method invokes \verb{GET /projects/\{projectId\}/frozenModels/} in the DataRobot Public API.
}
\seealso{
The method ProjectsFrozenModelsList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
