% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R
\name{ListProjectsImagesFile}
\alias{ListProjectsImagesFile}
\title{ListProjectsImagesFile}
\usage{
ListProjectsImagesFile(imageId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Returns a file for a single image
}
\details{
Returns a file for a single image

This method invokes \verb{GET /projects/\{projectId\}/images/\{imageId\}/file/} in the DataRobot Public API.
}
\seealso{
The method ProjectsImagesFileList in \link{ImagesApi}, which invokes the same underlying API endpoint.
}
