% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsClusterInsights}
\alias{ListProjectsModelsClusterInsights}
\title{ListProjectsModelsClusterInsights}
\usage{
ListProjectsModelsClusterInsights(
  modelId,
  orderBy,
  searchFor,
  projectId,
  offset = 0,
  limit = 15,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve Cluster Insights for all features.
}
\details{
Retrieve all computed Cluster Insights for a clustering project model.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/clusterInsights/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsClusterInsightsList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
