% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListProjectsModelsClusterNames}
\alias{ListProjectsModelsClusterNames}
\title{ListProjectsModelsClusterNames}
\usage{
ListProjectsModelsClusterNames(modelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve cluster names assigned to an unsupervised cluster model
}
\details{
Retrieve all cluster names assigned to an unsupervised cluster model

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/clusterNames/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsClusterNamesList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
