% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsConfusionChartsClassDetails}
\alias{ListProjectsModelsConfusionChartsClassDetails}
\title{ListProjectsModelsConfusionChartsClassDetails}
\usage{
ListProjectsModelsConfusionChartsClassDetails(
  modelId,
  className,
  source,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Calculates and sends frequency of class in distributed among other classes for actual and predicted data.
}
\details{
Calculates and sends frequency of class in distributed among other > classes for actual and predicted data. A confusion chart class details > for given class gives stats of misclassification done by model for > given class for actual and predicted data. .. note:: Available for > multiclass projects only.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/confusionCharts/\{source\}/classDetails/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsConfusionChartsClassDetailsList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
