% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsDataDisparityInsights}
\alias{ListProjectsModelsDataDisparityInsights}
\title{ListProjectsModelsDataDisparityInsights}
\usage{
ListProjectsModelsDataDisparityInsights(
  className1,
  feature,
  modelId,
  className2,
  projectId,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get Cross Class Data Disparity results.
}
\details{
Retrieve a list of Cross Class Data Disparity insights for the model.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/dataDisparityInsights/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsDataDisparityInsightsList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
