% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsDatasetResidualsCharts}
\alias{ListProjectsModelsDatasetResidualsCharts}
\title{ListProjectsModelsDatasetResidualsCharts}
\usage{
ListProjectsModelsDatasetResidualsCharts(
  modelId,
  datasetId,
  projectId,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List of residuals charts objects on prediction datasets.
}
\details{
List of residuals charts objects on prediction datasets for a project with filtering option by dataset. Prediction dataset may have residuals chart computed if it contained a column with actual values and predictions were computed on this dataset. Residuals charts are supported for regression projects only.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/datasetResidualsCharts/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsDatasetResidualsChartsList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
