% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsDatasetRocCurves}
\alias{ListProjectsModelsDatasetRocCurves}
\title{ListProjectsModelsDatasetRocCurves}
\usage{
ListProjectsModelsDatasetRocCurves(
  modelId,
  datasetId,
  projectId,
  offset = 0,
  limit = 100,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List of ROC curve objects on prediction datasets for a project with filtering option by dataset.
}
\details{
List of ROC curve objects on prediction datasets for a project with filtering option by dataset.  Prediction dataset may have ROC curve computed if it contained a column with actual values and predictions were computed on this dataset. Each ROC curve object includes an array of points showing the performance of the model at different thresholds for classification, and arrays of sample predictions for both the positive and negative classes.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/datasetRocCurves/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsDatasetRocCurvesList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
