% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsFeatureFitMetadata}
\alias{ListProjectsModelsFeatureFitMetadata}
\title{ListProjectsModelsFeatureFitMetadata}
\usage{
ListProjectsModelsFeatureFitMetadata(modelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve Feature Fit metadata. Response contains status and available sources.
}
\details{
NOTE: Starting from API v2.29 this endpoint will be deprecated. Retrieve Feature Fit metadata. Response contains status and available sources. One of the provided source parameters used for retrieving Feature Fit. \\* Source can be, at a minimum, training or validation. If holdout is configured for the project, source also includes holdout. \\* Source value of training is always available.(versions prior to v2.17 support validation only) \\* When a model is trained into validation or holdout without stacked predictions (i.e., no out-of-sample predictions in validation or holdout), validation and holdout sources are not available.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/featureFitMetadata/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsFeatureFitMetadataList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
