% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsFeatureListsClusterInsights}
\alias{ListProjectsModelsFeatureListsClusterInsights}
\title{ListProjectsModelsFeatureListsClusterInsights}
\usage{
ListProjectsModelsFeatureListsClusterInsights(
  modelId,
  orderBy,
  datasetId,
  searchFor,
  projectId,
  offset = 0,
  limit = 15,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve Cluster Insights for a single featurelist
}
\details{
Retrieve computed Cluster Insights for a clustering project model on a single featurelist.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/featureLists/\{datasetId\}/clusterInsights/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsFeatureListsClusterInsightsList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
