% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R, R/insights_apiWrapper.R
\name{ListProjectsModelsImageEmbeddings}
\alias{ListProjectsModelsImageEmbeddings}
\title{ListProjectsModelsImageEmbeddings}
\usage{
ListProjectsModelsImageEmbeddings(modelId, featureName, projectId, ...)

ListProjectsModelsImageEmbeddings(modelId, featureName, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve ImageEmbeddings for a feature of a model.

Retrieve ImageEmbeddings for a feature of a model.
}
\details{
Retrieve ImageEmbeddings for a feature of a model.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/imageEmbeddings/} in the DataRobot Public API.

Retrieve ImageEmbeddings for a feature of a model.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/imageEmbeddings/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsImageEmbeddingsList in \link{ImagesApi}, which invokes the same underlying API endpoint.

The method ProjectsModelsImageEmbeddingsList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
