% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListProjectsModelsNumIterationsTrained}
\alias{ListProjectsModelsNumIterationsTrained}
\title{ListProjectsModelsNumIterationsTrained}
\usage{
ListProjectsModelsNumIterationsTrained(modelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get number of iterations trained
}
\details{
Retrieve the actual number of iterations or estimators trained by a tree-based early stopping model.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/numIterationsTrained/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsNumIterationsTrainedList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
