% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsResiduals}
\alias{ListProjectsModelsResiduals}
\title{ListProjectsModelsResiduals}
\usage{
ListProjectsModelsResiduals(modelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve all residuals charts for a model.
}
\details{
Retrieve all residuals charts for a model.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/residuals/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsResidualsList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
