% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsModelsRocCurve}
\alias{ListProjectsModelsRocCurve}
\title{ListProjectsModelsRocCurve}
\usage{
ListProjectsModelsRocCurve(modelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve all available ROC curves for model.
}
\details{
Retrieve all available ROC curves for model. The response will include a json list of all available ROC curves, in the same format as the response from :\url{http:get}:/api/v2/projects/(projectId)/models/(modelId)/rocCurve/(source)/.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/rocCurve/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsRocCurveList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
