% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsMulticategoricalInvalidFormatFile}
\alias{ListProjectsMulticategoricalInvalidFormatFile}
\title{ListProjectsMulticategoricalInvalidFormatFile}
\usage{
ListProjectsMulticategoricalInvalidFormatFile(projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get file with format errors of potential multicategorical features.
}
\details{
Get file with format errors of potential multicategorical features.

This method invokes \verb{GET /projects/\{projectId\}/multicategoricalInvalidFormat/file/} in the DataRobot Public API.
}
\seealso{
The method ProjectsMulticategoricalInvalidFormatFileList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
