% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{ListProjectsMultiseriesIdsCrossSeriesProperties}
\alias{ListProjectsMultiseriesIdsCrossSeriesProperties}
\title{ListProjectsMultiseriesIdsCrossSeriesProperties}
\usage{
ListProjectsMultiseriesIdsCrossSeriesProperties(
  multiseriesId,
  crossSeriesGroupByColumns,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve eligible cross-series group-by columns.
}
\details{
Retrieve eligible cross-series group-by columns.  Note that validation will have to have been triggered via :\url{http:post}:/api/v2/projects/(projectId)/crossSeriesProperties/ in order for results to appear here.

This method invokes \verb{GET /projects/\{projectId\}/multiseriesIds/\{multiseriesId\}/crossSeriesProperties/} in the DataRobot Public API.
}
\seealso{
The method MultiseriesIdsCrossSeriesPropertiesList in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
