% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsPayoffMatrices}
\alias{ListProjectsPayoffMatrices}
\title{ListProjectsPayoffMatrices}
\usage{
ListProjectsPayoffMatrices(projectId, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List of all payoff matrices for a project.
}
\details{
List of all payoff matrices for a project.

This method invokes \verb{GET /projects/\{projectId\}/payoffMatrices/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPayoffMatricesList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
