% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{ListProjectsSegmentationTasksMappings}
\alias{ListProjectsSegmentationTasksMappings}
\title{ListProjectsSegmentationTasksMappings}
\usage{
ListProjectsSegmentationTasksMappings(
  segmentationTaskId,
  projectId,
  offset = 0,
  limit = 10,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve seriesId to segmentId mappings.
}
\details{
Retrieve the seriesId to segmentId mappings for a Segmentation Task.

This method invokes \verb{GET /projects/\{projectId\}/segmentationTasks/\{segmentationTaskId\}/mappings/} in the DataRobot Public API.
}
\seealso{
The method SegmentationTasksMappingsList in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
