% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{ListProjectsShapMatrices}
\alias{ListProjectsShapMatrices}
\title{ListProjectsShapMatrices}
\usage{
ListProjectsShapMatrices(projectId, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List SHAP matrix records.
}
\details{
Get a list of available SHAP matrix records.

This method invokes \verb{GET /projects/\{projectId\}/shapMatrices/} in the DataRobot Public API.
}
\seealso{
The method ProjectsShapMatricesList in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
