% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListRatingTableModels}
\alias{ListRatingTableModels}
\title{ListRatingTableModels}
\usage{
ListRatingTableModels(
  project,
  bulkOperationId = NULL,
  orderBy = NULL,
  name = NULL,
  samplePct = NULL,
  withMetric = NULL,
  isStarred = NULL,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List Rating Table Models
}
\details{
Lists all the models from a project that have rating tables

This method invokes \verb{GET /projects/\{projectId\}/ratingTableModels/} in the DataRobot Public API.
}
\seealso{
The method ProjectsRatingTableModelsList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
