% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{ListRatingTables}
\alias{ListRatingTables}
\title{ListRatingTables}
\usage{
ListRatingTables(
  project,
  parentModelId = NULL,
  modelId = NULL,
  offset = 0,
  limit = 0,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List Rating Tables For The Project
}
\details{
List RatingTable objects for a project.   These contain metadata about the rating table and the location at which the corresponding rating   table file can be retrieved.

This method invokes \verb{GET /projects/\{projectId\}/ratingTables/} in the DataRobot Public API.
}
\seealso{
The method ProjectsRatingTablesList in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
