% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{ListScheduledJobs}
\alias{ListScheduledJobs}
\title{ListScheduledJobs}
\usage{
ListScheduledJobs(
  integrationTypeName,
  deploymentId,
  filterEnabled,
  search = "",
  offset = 0,
  queryByUser = "createdBy",
  limit = 20,
  orderBy = "name",
  typeId = "predictionIntegration",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List scheduled deployment batch prediction jobs a user can view
}
\details{
Get a list of scheduled batch prediction jobs a user can view

This method invokes \verb{GET /scheduledJobs/} in the DataRobot Public API.
}
\seealso{
The method ScheduledJobsList in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
