% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{ListTrainingPredictions}
\alias{ListTrainingPredictions}
\title{ListTrainingPredictions}
\usage{
ListTrainingPredictions(project, offset = 0, limit = 0, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List training prediction jobs
}
\details{
Get a list of training prediction records

This method invokes \verb{GET /projects/\{projectId\}/trainingPredictions/} in the DataRobot Public API.
}
\seealso{
The method TrainingPredictionsList in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
