% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_apiWrapper.R
\name{ListUsageDataExportsSupportedEvents}
\alias{ListUsageDataExportsSupportedEvents}
\title{ListUsageDataExportsSupportedEvents}
\usage{
ListUsageDataExportsSupportedEvents(...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Describe supported available audit events with which to filter result data.
}
\details{
Describe supported available audit events with which to filter result data.

This method invokes \verb{GET /usageDataExports/supportedEvents/} in the DataRobot Public API.
}
\seealso{
The method UsageDataExportsSupportedEventsList in \link{AnalyticsApi}, which invokes the same underlying API endpoint.
}
