% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{ListUseCases}
\alias{ListUseCases}
\title{ListUseCases}
\usage{
ListUseCases(stage, orderBy, namePart, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List use cases the requesting user has access to.
}
\details{
List use cases the requesting user has access to.

This method invokes \verb{GET /useCases/} in the DataRobot Public API.
}
\seealso{
The method UseCasesList in \link{UseCaseApi}, which invokes the same underlying API endpoint.
}
