% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{ListUseCasesAttachments}
\alias{ListUseCasesAttachments}
\title{ListUseCasesAttachments}
\usage{
ListUseCasesAttachments(useCaseId, type, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get a list of resources attached to this use case.
}
\details{
Get a list of resources attached to this use case.

This method invokes \verb{GET /useCases/\{useCaseId\}/attachments/} in the DataRobot Public API.
}
\seealso{
The method UseCasesAttachmentsList in \link{UseCaseApi}, which invokes the same underlying API endpoint.
}
