% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{ListUseCaseValueTemplates}
\alias{ListUseCaseValueTemplates}
\title{ListUseCaseValueTemplates}
\usage{
ListUseCaseValueTemplates(...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List available use cases templates.
}
\details{
List available use cases templates.

This method invokes \verb{GET /useCaseValueTemplates/} in the DataRobot Public API.
}
\seealso{
The method UseCaseValueTemplatesList in \link{UseCaseApi}, which invokes the same underlying API endpoint.
}
