% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R,
%   R/blueprints_apiWrapper.R
\name{ListUserBlueprints}
\alias{ListUserBlueprints}
\title{ListUserBlueprints}
\usage{
ListUserBlueprints(projectId, offset = 0, limit = 100, ...)

ListUserBlueprints(projectId, offset = 0, limit = 100, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List user blueprints.

List user blueprints.
}
\details{
Fetch a list of the user blueprints the current user has access to

This method invokes \verb{GET /userBlueprints/} in the DataRobot Public API.

Fetch a list of the user blueprints the current user has access to

This method invokes \verb{GET /userBlueprints/} in the DataRobot Public API.
}
\seealso{
The method UserBlueprintsList in \link{AiCatalogApi}, which invokes the same underlying API endpoint.

The method UserBlueprintsList in \link{BlueprintsApi}, which invokes the same underlying API endpoint.
}
