% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R,
%   R/blueprints_apiWrapper.R
\name{ListUserBlueprintsInputTypes}
\alias{ListUserBlueprintsInputTypes}
\title{ListUserBlueprintsInputTypes}
\usage{
ListUserBlueprintsInputTypes(...)

ListUserBlueprintsInputTypes(...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve input types.

Retrieve input types.
}
\details{
Retrieve the input types which can be used with User Blueprints.

This method invokes \verb{GET /userBlueprintsInputTypes/} in the DataRobot Public API.

Retrieve the input types which can be used with User Blueprints.

This method invokes \verb{GET /userBlueprintsInputTypes/} in the DataRobot Public API.
}
\seealso{
The method UserBlueprintsInputTypesList in \link{AiCatalogApi}, which invokes the same underlying API endpoint.

The method UserBlueprintsInputTypesList in \link{BlueprintsApi}, which invokes the same underlying API endpoint.
}
