% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R,
%   R/blueprints_apiWrapper.R
\name{ListUserBlueprintsSharedRoles}
\alias{ListUserBlueprintsSharedRoles}
\title{ListUserBlueprintsSharedRoles}
\usage{
ListUserBlueprintsSharedRoles(
  userBlueprintId,
  shareRecipientType,
  name,
  id,
  offset = 0,
  limit = 10,
  ...
)

ListUserBlueprintsSharedRoles(
  userBlueprintId,
  shareRecipientType,
  name,
  id,
  offset = 0,
  limit = 10,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get a list of users, groups and organizations that have an access to this user blueprint

Get a list of users, groups and organizations that have an access to this user blueprint
}
\details{
Get a list of users, groups and organizations that have an access to this user blueprint

This method invokes \verb{GET /userBlueprints/\{userBlueprintId\}/sharedRoles/} in the DataRobot Public API.

Get a list of users, groups and organizations that have an access to this user blueprint

This method invokes \verb{GET /userBlueprints/\{userBlueprintId\}/sharedRoles/} in the DataRobot Public API.
}
\seealso{
The method UserBlueprintsSharedRolesList in \link{AiCatalogApi}, which invokes the same underlying API endpoint.

The method UserBlueprintsSharedRolesList in \link{BlueprintsApi}, which invokes the same underlying API endpoint.
}
