% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R,
%   R/blueprints_apiWrapper.R
\name{ListUserBlueprintsTasks}
\alias{ListUserBlueprintsTasks}
\title{ListUserBlueprintsTasks}
\usage{
ListUserBlueprintsTasks(userBlueprintId, blueprintId, projectId, ...)

ListUserBlueprintsTasks(userBlueprintId, blueprintId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve tasks for blueprint construction.

Retrieve tasks for blueprint construction.
}
\details{
Retrieve the available tasks, organized into categories, which can be used to create or modify User Blueprints.

This method invokes \verb{GET /userBlueprintsTasks/} in the DataRobot Public API.

Retrieve the available tasks, organized into categories, which can be used to create or modify User Blueprints.

This method invokes \verb{GET /userBlueprintsTasks/} in the DataRobot Public API.
}
\seealso{
The method UserBlueprintsTasksList in \link{AiCatalogApi}, which invokes the same underlying API endpoint.

The method UserBlueprintsTasksList in \link{BlueprintsApi}, which invokes the same underlying API endpoint.
}
