% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{ListUserCleanupPreviewsContent}
\alias{ListUserCleanupPreviewsContent}
\title{ListUserCleanupPreviewsContent}
\usage{
ListUserCleanupPreviewsContent(reportId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Users permanent delete extended preview.
}
\details{
Retrieve users permanent delete extended preview.

This method invokes \verb{GET /userCleanupPreviews/\{reportId\}/content/} in the DataRobot Public API.
}
\seealso{
The method UserCleanupPreviewsContentList in \link{UserManagementApi}, which invokes the same underlying API endpoint.
}
