% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{ListUserCleanupSummariesContent}
\alias{ListUserCleanupSummariesContent}
\title{ListUserCleanupSummariesContent}
\usage{
ListUserCleanupSummariesContent(reportId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Users permanent delete extended summary.
}
\details{
Retrieve users permanent delete extended summary.

This method invokes \verb{GET /userCleanupSummaries/\{reportId\}/content/} in the DataRobot Public API.
}
\seealso{
The method UserCleanupSummariesContentList in \link{UserManagementApi}, which invokes the same underlying API endpoint.
}
