% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlops_api.R
\docType{class}
\name{MlopsApi}
\alias{MlopsApi}
\title{Mlops operations}
\format{
An \code{R6Class} generator object
}
\description{
datarobot.apicore.Mlops
}
\examples{

## ------------------------------------------------
## Method `MlopsApi$MlopsPortablePredictionServerImageList`
## ------------------------------------------------

\dontrun{
library(datarobot.apicore)

api.instance <- MlopsApi$new()
result <- api.instance$MlopsPortablePredictionServerImageList()
}

## ------------------------------------------------
## Method `MlopsApi$MlopsPortablePredictionServerImageMetadataList`
## ------------------------------------------------

\dontrun{
library(datarobot.apicore)

api.instance <- MlopsApi$new()
result <- api.instance$MlopsPortablePredictionServerImageMetadataList()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MlopsApi$new()}}
\item \href{#method-MlopsPortablePredictionServerImageList}{\code{MlopsApi$MlopsPortablePredictionServerImageList()}}
\item \href{#method-MlopsPortablePredictionServerImageMetadataList}{\code{MlopsApi$MlopsPortablePredictionServerImageMetadataList()}}
\item \href{#method-clone}{\code{MlopsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MlopsApi$new(apiClient)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{apiClient}}{A configurable \code{ApiClient} instance. If none provided, a new client with default configuration will be created.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MlopsPortablePredictionServerImageList"></a>}}
\if{latex}{\out{\hypertarget{method-MlopsPortablePredictionServerImageList}{}}}
\subsection{Method \code{MlopsPortablePredictionServerImageList()}}{
Downloads the latest Portable Prediction Server (PPS) Docker image
Produces: NA
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MlopsApi$MlopsPortablePredictionServerImageList(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional. Additional named parameters to be passed downward.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Fetches the latest Portable Prediction Server (PPS) Docker image. The resulting image can be docker-loaded. Since the image can be quite large (14GB+) consider querying its metadata to check the image size in advance and content hash to verify the downloaded image afterwards. In some environments it can HTTP redirect to some other service (like S3 or GCP) using pre-signed URLs.


Response status codes, messages, and headers:
\itemize{
\item \strong{\code{200}} Download the latest available Portable Prediction Server (PPS) Docker image
\itemize{
}
\item \strong{\code{302}} Redirect to another service for more efficient content download using pre-signed URL
\itemize{
}
\item \strong{\code{404}} No PPS images found in the system
\itemize{
}
}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
library(datarobot.apicore)

api.instance <- MlopsApi$new()
result <- api.instance$MlopsPortablePredictionServerImageList()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MlopsPortablePredictionServerImageMetadataList"></a>}}
\if{latex}{\out{\hypertarget{method-MlopsPortablePredictionServerImageMetadataList}{}}}
\subsection{Method \code{MlopsPortablePredictionServerImageMetadataList()}}{
Fetches currently active PPS Docker image metadata
Produces: "application/json"
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MlopsApi$MlopsPortablePredictionServerImageMetadataList(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional. Additional named parameters to be passed downward.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Fetches currently active PPS Docker image metadata


Response status codes, messages, and headers:
\itemize{
\item \strong{\code{200}} Available Portable Prediction Server (PPS) image metadata
\itemize{
}
\item \strong{\code{404}} No PPS images found in the system
\itemize{
}
}
}

\subsection{Returns}{
\link{PPSImageMetadataResponse}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
library(datarobot.apicore)

api.instance <- MlopsApi$new()
result <- api.instance$MlopsPortablePredictionServerImageMetadataList()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MlopsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
