% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/model_compliance_docs_initializations_response.R
\docType{class}
\name{ModelComplianceDocsInitializationsResponse}
\alias{ModelComplianceDocsInitializationsResponse}
\title{ModelComplianceDocsInitializationsResponse}
\format{
An \code{R6Class} generator object
}
\description{
ModelComplianceDocsInitializationsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{initialized}}{character Whether compliance documentation pre-preprocessing is initialized for the model}

\item{\code{status}}{character Compliance documentation pre-processing initialization status}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelComplianceDocsInitializationsResponse-new}{\code{ModelComplianceDocsInitializationsResponse$new()}}
\item \href{#method-ModelComplianceDocsInitializationsResponse-validate}{\code{ModelComplianceDocsInitializationsResponse$validate()}}
\item \href{#method-ModelComplianceDocsInitializationsResponse-toJSON}{\code{ModelComplianceDocsInitializationsResponse$toJSON()}}
\item \href{#method-ModelComplianceDocsInitializationsResponse-fromJSON}{\code{ModelComplianceDocsInitializationsResponse$fromJSON()}}
\item \href{#method-ModelComplianceDocsInitializationsResponse-clone}{\code{ModelComplianceDocsInitializationsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelComplianceDocsInitializationsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelComplianceDocsInitializationsResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelComplianceDocsInitializationsResponse$new(
  initialized = NULL,
  status = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initialized}}{Whether compliance documentation pre-preprocessing is initialized for the model}

\item{\code{status}}{Compliance documentation pre-processing initialization status}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelComplianceDocsInitializationsResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelComplianceDocsInitializationsResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelComplianceDocsInitializationsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelComplianceDocsInitializationsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelComplianceDocsInitializationsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelComplianceDocsInitializationsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelComplianceDocsInitializationsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelComplianceDocsInitializationsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelComplianceDocsInitializationsResponse$fromJSON(
  ModelComplianceDocsInitializationsResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ModelComplianceDocsInitializationsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelComplianceDocsInitializationsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelComplianceDocsInitializationsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelComplianceDocsInitializationsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
