% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/model_confusion_chart_class_details_for_dataset_retrieve.R
\docType{class}
\name{ModelConfusionChartClassDetailsForDatasetRetrieve}
\alias{ModelConfusionChartClassDetailsForDatasetRetrieve}
\title{ModelConfusionChartClassDetailsForDatasetRetrieve}
\format{
An \code{R6Class} generator object
}
\description{
ModelConfusionChartClassDetailsForDatasetRetrieve Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actualFrequency}}{list( \link{ActualFrequency} ) One vs All actual percentage and count in a format specified below sorted by percentage in decreasing order}

\item{\code{className}}{character Name of a class for which distribution frequency is requested}

\item{\code{datasetId}}{character The dataset to retrieve a Confusion chart from.}

\item{\code{modelId}}{character The model to retrieve a Confusion chart from.}

\item{\code{predictedFrequency}}{list( \link{PredictedFrequency} ) One vs All predicted percentage and count in a format specified below sorted by percentage in decreasing order}

\item{\code{projectId}}{character The project to retrieve a Confusion chart from.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelConfusionChartClassDetailsForDatasetRetrieve$new()}}
\item \href{#method-validate}{\code{ModelConfusionChartClassDetailsForDatasetRetrieve$validate()}}
\item \href{#method-toJSON}{\code{ModelConfusionChartClassDetailsForDatasetRetrieve$toJSON()}}
\item \href{#method-fromJSON}{\code{ModelConfusionChartClassDetailsForDatasetRetrieve$fromJSON()}}
\item \href{#method-clone}{\code{ModelConfusionChartClassDetailsForDatasetRetrieve$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfusionChartClassDetailsForDatasetRetrieve$new(
  actualFrequency = NULL,
  className = NULL,
  datasetId = NULL,
  modelId = NULL,
  predictedFrequency = NULL,
  projectId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{actualFrequency}}{One vs All actual percentage and count in a format specified below sorted by percentage in decreasing order}

\item{\code{className}}{Name of a class for which distribution frequency is requested}

\item{\code{datasetId}}{The dataset to retrieve a Confusion chart from.}

\item{\code{modelId}}{The model to retrieve a Confusion chart from.}

\item{\code{predictedFrequency}}{One vs All predicted percentage and count in a format specified below sorted by percentage in decreasing order}

\item{\code{projectId}}{The project to retrieve a Confusion chart from.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfusionChartClassDetailsForDatasetRetrieve$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfusionChartClassDetailsForDatasetRetrieve$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfusionChartClassDetailsForDatasetRetrieve$fromJSON(
  ModelConfusionChartClassDetailsForDatasetRetrieveJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ModelConfusionChartClassDetailsForDatasetRetrieveJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfusionChartClassDetailsForDatasetRetrieve$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
