% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/model_confusion_chart_class_details_retrieve_reponse_controller.R
\docType{class}
\name{ModelConfusionChartClassDetailsRetrieveReponseController}
\alias{ModelConfusionChartClassDetailsRetrieveReponseController}
\title{ModelConfusionChartClassDetailsRetrieveReponseController}
\format{
An \code{R6Class} generator object
}
\description{
ModelConfusionChartClassDetailsRetrieveReponseController Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actualFrequency}}{list( \link{ActualFrequency} ) One vs all actual percentage and count in a format specified below sorted by percentage in decreasing order}

\item{\code{className}}{character Name of a class for which distribution frequency is requested.}

\item{\code{predictedFrequency}}{list( \link{PredictedFrequency} ) One vs all predicted percentage and count in a format specified below sorted by percentage in decreasing order}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelConfusionChartClassDetailsRetrieveReponseController-new}{\code{ModelConfusionChartClassDetailsRetrieveReponseController$new()}}
\item \href{#method-ModelConfusionChartClassDetailsRetrieveReponseController-validate}{\code{ModelConfusionChartClassDetailsRetrieveReponseController$validate()}}
\item \href{#method-ModelConfusionChartClassDetailsRetrieveReponseController-toJSON}{\code{ModelConfusionChartClassDetailsRetrieveReponseController$toJSON()}}
\item \href{#method-ModelConfusionChartClassDetailsRetrieveReponseController-fromJSON}{\code{ModelConfusionChartClassDetailsRetrieveReponseController$fromJSON()}}
\item \href{#method-ModelConfusionChartClassDetailsRetrieveReponseController-clone}{\code{ModelConfusionChartClassDetailsRetrieveReponseController$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfusionChartClassDetailsRetrieveReponseController-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfusionChartClassDetailsRetrieveReponseController-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfusionChartClassDetailsRetrieveReponseController$new(
  actualFrequency = NULL,
  className = NULL,
  predictedFrequency = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{actualFrequency}}{One vs all actual percentage and count in a format specified below sorted by percentage in decreasing order}

\item{\code{className}}{Name of a class for which distribution frequency is requested.}

\item{\code{predictedFrequency}}{One vs all predicted percentage and count in a format specified below sorted by percentage in decreasing order}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfusionChartClassDetailsRetrieveReponseController-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfusionChartClassDetailsRetrieveReponseController-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfusionChartClassDetailsRetrieveReponseController$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfusionChartClassDetailsRetrieveReponseController-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfusionChartClassDetailsRetrieveReponseController-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfusionChartClassDetailsRetrieveReponseController$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfusionChartClassDetailsRetrieveReponseController-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfusionChartClassDetailsRetrieveReponseController-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfusionChartClassDetailsRetrieveReponseController$fromJSON(
  ModelConfusionChartClassDetailsRetrieveReponseControllerJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ModelConfusionChartClassDetailsRetrieveReponseControllerJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfusionChartClassDetailsRetrieveReponseController-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfusionChartClassDetailsRetrieveReponseController-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfusionChartClassDetailsRetrieveReponseController$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
