% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_confusion_chart_retrieve_response.R
\docType{class}
\name{ModelConfusionChartRetrieveResponse}
\alias{ModelConfusionChartRetrieveResponse}
\title{ModelConfusionChartRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
ModelConfusionChartRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{columns}}{list( integer ) \link{colStart, colEnd} column dimension of confusion matrix in response}

\item{\code{data}}{\link{ConfusionChartData}}

\item{\code{numberOfClasses}}{integer count of classes in full confusion matrix.}

\item{\code{rows}}{list( integer ) \link{rowStart, rowEnd} row dimension of confusion matrix in response}

\item{\code{source}}{character source of the chart}

\item{\code{totalMatrixSum}}{integer sum of all values in full confusion matrix}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelConfusionChartRetrieveResponse$new()}}
\item \href{#method-validate}{\code{ModelConfusionChartRetrieveResponse$validate()}}
\item \href{#method-toJSON}{\code{ModelConfusionChartRetrieveResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{ModelConfusionChartRetrieveResponse$fromJSON()}}
\item \href{#method-clone}{\code{ModelConfusionChartRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfusionChartRetrieveResponse$new(
  columns = NULL,
  data = NULL,
  numberOfClasses = NULL,
  rows = NULL,
  source = NULL,
  totalMatrixSum = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{\link{colStart, colEnd} column dimension of confusion matrix in response}

\item{\code{numberOfClasses}}{count of classes in full confusion matrix.}

\item{\code{rows}}{\link{rowStart, rowEnd} row dimension of confusion matrix in response}

\item{\code{source}}{source of the chart}

\item{\code{totalMatrixSum}}{sum of all values in full confusion matrix}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfusionChartRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfusionChartRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfusionChartRetrieveResponse$fromJSON(
  ModelConfusionChartRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ModelConfusionChartRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfusionChartRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
