% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_package_capabilities.R
\docType{class}
\name{ModelPackageCapabilities}
\alias{ModelPackageCapabilities}
\title{ModelPackageCapabilities}
\format{
An \code{R6Class} generator object
}
\description{
ModelPackageCapabilities Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{supportsAutomaticActuals}}{character \link{optional} Whether inferring actual values from time series history data and automatically feeding them back for accuracy estimation is supported by this model package.}

\item{\code{supportsChallengerModels}}{character Whether Challenger Models are supported by this model package.}

\item{\code{supportsFeatureDriftTracking}}{character Whether Feature Drift is supported by this model package.}

\item{\code{supportsHumilityRecommendedRules}}{character Whether calculating values for recommended Humility Rules is supported by this model package.}

\item{\code{supportsHumilityRules}}{character Whether Humility Rules are supported by this model package.}

\item{\code{supportsHumilityRulesDefaultCalculations}}{character Whether calculating default values for Humility Rules is supported by this model package.}

\item{\code{supportsPredictionWarning}}{character Whether Prediction Warnings are supported by this model package.}

\item{\code{supportsRetraining}}{character \link{optional} Whether deployment supports retraining.}

\item{\code{supportsScoringCodeDownload}}{character \link{optional} Whether scoring code download is supported by this model package.}

\item{\code{supportsSecondaryDatasets}}{character If the deployments supports secondary datasets.}

\item{\code{supportsSegmentedAnalysisDriftAndAccuracy}}{character Whether tracking features in training and predictions data for segmented analysis is supported by this model package.}

\item{\code{supportsShapBasedPredictionExplanations}}{character Whether shap-based prediction explanations are supported by this model package.}

\item{\code{supportsTargetDriftTracking}}{character Whether Target Drift is supported by this model package.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelPackageCapabilities$new()}}
\item \href{#method-validate}{\code{ModelPackageCapabilities$validate()}}
\item \href{#method-toJSON}{\code{ModelPackageCapabilities$toJSON()}}
\item \href{#method-fromJSON}{\code{ModelPackageCapabilities$fromJSON()}}
\item \href{#method-clone}{\code{ModelPackageCapabilities$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageCapabilities$new(
  supportsChallengerModels = NULL,
  supportsFeatureDriftTracking = NULL,
  supportsHumilityRecommendedRules = NULL,
  supportsHumilityRules = NULL,
  supportsHumilityRulesDefaultCalculations = NULL,
  supportsPredictionWarning = NULL,
  supportsSecondaryDatasets = NULL,
  supportsSegmentedAnalysisDriftAndAccuracy = NULL,
  supportsShapBasedPredictionExplanations = NULL,
  supportsTargetDriftTracking = NULL,
  supportsAutomaticActuals = NULL,
  supportsRetraining = NULL,
  supportsScoringCodeDownload = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{supportsChallengerModels}}{Whether Challenger Models are supported by this model package.}

\item{\code{supportsFeatureDriftTracking}}{Whether Feature Drift is supported by this model package.}

\item{\code{supportsHumilityRecommendedRules}}{Whether calculating values for recommended Humility Rules is supported by this model package.}

\item{\code{supportsHumilityRules}}{Whether Humility Rules are supported by this model package.}

\item{\code{supportsHumilityRulesDefaultCalculations}}{Whether calculating default values for Humility Rules is supported by this model package.}

\item{\code{supportsPredictionWarning}}{Whether Prediction Warnings are supported by this model package.}

\item{\code{supportsSecondaryDatasets}}{If the deployments supports secondary datasets.}

\item{\code{supportsSegmentedAnalysisDriftAndAccuracy}}{Whether tracking features in training and predictions data for segmented analysis is supported by this model package.}

\item{\code{supportsShapBasedPredictionExplanations}}{Whether shap-based prediction explanations are supported by this model package.}

\item{\code{supportsTargetDriftTracking}}{Whether Target Drift is supported by this model package.}

\item{\code{supportsAutomaticActuals}}{Whether inferring actual values from time series history data and automatically feeding them back for accuracy estimation is supported by this model package.}

\item{\code{supportsRetraining}}{Whether deployment supports retraining.}

\item{\code{supportsScoringCodeDownload}}{Whether scoring code download is supported by this model package.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageCapabilities$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageCapabilities$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageCapabilities$fromJSON(
  ModelPackageCapabilitiesJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ModelPackageCapabilitiesJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageCapabilities$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
