% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_package_capability.R
\docType{class}
\name{ModelPackageCapability}
\alias{ModelPackageCapability}
\title{ModelPackageCapability}
\format{
An \code{R6Class} generator object
}
\description{
ModelPackageCapability Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{messages}}{list( character ) Messages explaining why the capability is supported or not supported.}

\item{\code{name}}{character The name of the capability.}

\item{\code{supported}}{character If the capability is supported.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelPackageCapability-new}{\code{ModelPackageCapability$new()}}
\item \href{#method-ModelPackageCapability-validate}{\code{ModelPackageCapability$validate()}}
\item \href{#method-ModelPackageCapability-toJSON}{\code{ModelPackageCapability$toJSON()}}
\item \href{#method-ModelPackageCapability-fromJSON}{\code{ModelPackageCapability$fromJSON()}}
\item \href{#method-ModelPackageCapability-clone}{\code{ModelPackageCapability$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelPackageCapability-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelPackageCapability-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageCapability$new(
  messages = NULL,
  name = NULL,
  supported = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{messages}}{Messages explaining why the capability is supported or not supported.}

\item{\code{name}}{The name of the capability.}

\item{\code{supported}}{If the capability is supported.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelPackageCapability-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelPackageCapability-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageCapability$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelPackageCapability-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelPackageCapability-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageCapability$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelPackageCapability-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelPackageCapability-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageCapability$fromJSON(
  ModelPackageCapabilityJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ModelPackageCapabilityJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelPackageCapability-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelPackageCapability-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageCapability$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
