% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_package_model_description.R
\docType{class}
\name{ModelPackageModelDescription}
\alias{ModelPackageModelDescription}
\title{ModelPackageModelDescription}
\format{
An \code{R6Class} generator object
}
\description{
ModelPackageModelDescription Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{buildEnvironmentType}}{character build environment type of the model}

\item{\code{description}}{character a description of the model}

\item{\code{location}}{character location of the model}

\item{\code{modelName}}{character \link{optional} model name}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelPackageModelDescription-new}{\code{ModelPackageModelDescription$new()}}
\item \href{#method-ModelPackageModelDescription-validate}{\code{ModelPackageModelDescription$validate()}}
\item \href{#method-ModelPackageModelDescription-toJSON}{\code{ModelPackageModelDescription$toJSON()}}
\item \href{#method-ModelPackageModelDescription-fromJSON}{\code{ModelPackageModelDescription$fromJSON()}}
\item \href{#method-ModelPackageModelDescription-clone}{\code{ModelPackageModelDescription$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelPackageModelDescription-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelPackageModelDescription-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageModelDescription$new(
  buildEnvironmentType = NULL,
  description = NULL,
  location = NULL,
  modelName = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{buildEnvironmentType}}{build environment type of the model}

\item{\code{description}}{a description of the model}

\item{\code{location}}{location of the model}

\item{\code{modelName}}{model name}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelPackageModelDescription-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelPackageModelDescription-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageModelDescription$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelPackageModelDescription-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelPackageModelDescription-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageModelDescription$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelPackageModelDescription-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelPackageModelDescription-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageModelDescription$fromJSON(
  ModelPackageModelDescriptionJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ModelPackageModelDescriptionJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelPackageModelDescription-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelPackageModelDescription-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageModelDescription$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
