% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_package_model_kind.R
\docType{class}
\name{ModelPackageModelKind}
\alias{ModelPackageModelKind}
\title{ModelPackageModelKind}
\format{
An \code{R6Class} generator object
}
\description{
ModelPackageModelKind Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{isAnomalyDetectionModel}}{character true if this is an anomaly detection model}

\item{\code{isCombinedModel}}{character true if model is a combined model}

\item{\code{isDecisionFlow}}{character true if this is a decision flow}

\item{\code{isFeatureDiscovery}}{character true if this model uses the Feature Discovery feature}

\item{\code{isMultiseries}}{character true if model is multiseries}

\item{\code{isTimeSeries}}{character true if model is time series}

\item{\code{isUnsupervisedLearning}}{character true if model used unsupervised learning}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelPackageModelKind-new}{\code{ModelPackageModelKind$new()}}
\item \href{#method-ModelPackageModelKind-validate}{\code{ModelPackageModelKind$validate()}}
\item \href{#method-ModelPackageModelKind-toJSON}{\code{ModelPackageModelKind$toJSON()}}
\item \href{#method-ModelPackageModelKind-fromJSON}{\code{ModelPackageModelKind$fromJSON()}}
\item \href{#method-ModelPackageModelKind-clone}{\code{ModelPackageModelKind$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelPackageModelKind-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelPackageModelKind-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageModelKind$new(
  isAnomalyDetectionModel = NULL,
  isCombinedModel = NULL,
  isDecisionFlow = NULL,
  isFeatureDiscovery = NULL,
  isMultiseries = NULL,
  isTimeSeries = NULL,
  isUnsupervisedLearning = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{isAnomalyDetectionModel}}{true if this is an anomaly detection model}

\item{\code{isCombinedModel}}{true if model is a combined model}

\item{\code{isDecisionFlow}}{true if this is a decision flow}

\item{\code{isFeatureDiscovery}}{true if this model uses the Feature Discovery feature}

\item{\code{isMultiseries}}{true if model is multiseries}

\item{\code{isTimeSeries}}{true if model is time series}

\item{\code{isUnsupervisedLearning}}{true if model used unsupervised learning}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelPackageModelKind-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelPackageModelKind-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageModelKind$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelPackageModelKind-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelPackageModelKind-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageModelKind$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelPackageModelKind-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelPackageModelKind-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageModelKind$fromJSON(
  ModelPackageModelKindJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ModelPackageModelKindJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelPackageModelKind-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelPackageModelKind-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageModelKind$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
