% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_package_retrieve_response.R
\docType{class}
\name{ModelPackageRetrieveResponse}
\alias{ModelPackageRetrieveResponse}
\title{ModelPackageRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
ModelPackageRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{activeDeploymentCount}}{integer Number of deployments currently using this model package}

\item{\code{buildStatus}}{character \link{optional} Model package build status}

\item{\code{capabilities}}{\link{ModelPackageCapabilities}}

\item{\code{datasets}}{\link{ModelPackageDatasets}}

\item{\code{id}}{character ID of the Model package}

\item{\code{importMeta}}{\link{ModelPackageImportMeta}}

\item{\code{isArchived}}{character Whether the model package is permanently archived (cannot be used in deployment or replacement)}

\item{\code{isDeprecated}}{character Whether the model package is deprecated. eg. python2 models are deprecated.}

\item{\code{modelDescription}}{\link{ModelPackageModelDescription}}

\item{\code{modelExecutionType}}{character Type of model package. `dedicated` (native DataRobot models) and `custom_inference_model` (user added inference models) both execute on DataRobot prediction servers, `external` do not}

\item{\code{modelId}}{character ID of the model}

\item{\code{modelKind}}{\link{ModelPackageModelKind}}

\item{\code{name}}{character Model package name}

\item{\code{permissions}}{list( character ) List of action permissions the user making the request has on the model package}

\item{\code{sourceMeta}}{\link{ModelPackageSourceMeta}}

\item{\code{target}}{\link{ModelPackageTarget}}

\item{\code{timeseries}}{\link{ModelPackageTimeseries}}

\item{\code{userProvidedId}}{character \link{optional} A user-provided unique ID associated with the given custom inference model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelPackageRetrieveResponse-new}{\code{ModelPackageRetrieveResponse$new()}}
\item \href{#method-ModelPackageRetrieveResponse-validate}{\code{ModelPackageRetrieveResponse$validate()}}
\item \href{#method-ModelPackageRetrieveResponse-toJSON}{\code{ModelPackageRetrieveResponse$toJSON()}}
\item \href{#method-ModelPackageRetrieveResponse-fromJSON}{\code{ModelPackageRetrieveResponse$fromJSON()}}
\item \href{#method-ModelPackageRetrieveResponse-clone}{\code{ModelPackageRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelPackageRetrieveResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelPackageRetrieveResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageRetrieveResponse$new(
  activeDeploymentCount = NULL,
  capabilities = NULL,
  datasets = NULL,
  id = NULL,
  importMeta = NULL,
  isArchived = NULL,
  isDeprecated = NULL,
  modelDescription = NULL,
  modelExecutionType = NULL,
  modelId = NULL,
  modelKind = NULL,
  name = NULL,
  permissions = NULL,
  sourceMeta = NULL,
  target = NULL,
  timeseries = NULL,
  buildStatus = NULL,
  userProvidedId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{activeDeploymentCount}}{Number of deployments currently using this model package}

\item{\code{id}}{ID of the Model package}

\item{\code{isArchived}}{Whether the model package is permanently archived (cannot be used in deployment or replacement)}

\item{\code{isDeprecated}}{Whether the model package is deprecated. eg. python2 models are deprecated.}

\item{\code{modelExecutionType}}{Type of model package. `dedicated` (native DataRobot models) and `custom_inference_model` (user added inference models) both execute on DataRobot prediction servers, `external` do not}

\item{\code{modelId}}{ID of the model}

\item{\code{name}}{Model package name}

\item{\code{permissions}}{List of action permissions the user making the request has on the model package}

\item{\code{buildStatus}}{Model package build status}

\item{\code{userProvidedId}}{A user-provided unique ID associated with the given custom inference model.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelPackageRetrieveResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelPackageRetrieveResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelPackageRetrieveResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelPackageRetrieveResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelPackageRetrieveResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelPackageRetrieveResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageRetrieveResponse$fromJSON(
  ModelPackageRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ModelPackageRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelPackageRetrieveResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelPackageRetrieveResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
