% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_package_target.R
\docType{class}
\name{ModelPackageTarget}
\alias{ModelPackageTarget}
\title{ModelPackageTarget}
\format{
An \code{R6Class} generator object
}
\description{
ModelPackageTarget Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{classCount}}{integer Number of classes for classification models.}

\item{\code{classNames}}{list( character ) Class names for prediction results. When target type is Binary, two class names are returned. The first element is the minority (positive) class and the second element is the majority (negative) class. Limited to 100 returned for Multiclass.}

\item{\code{name}}{character name of the target column}

\item{\code{predictionProbabilitiesColumn}}{character Field or column name containing prediction probabilities}

\item{\code{predictionThreshold}}{numeric Prediction threshold used for binary classification models}

\item{\code{type}}{character Target type of the model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelPackageTarget$new()}}
\item \href{#method-validate}{\code{ModelPackageTarget$validate()}}
\item \href{#method-toJSON}{\code{ModelPackageTarget$toJSON()}}
\item \href{#method-fromJSON}{\code{ModelPackageTarget$fromJSON()}}
\item \href{#method-clone}{\code{ModelPackageTarget$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageTarget$new(
  classCount = NULL,
  classNames = NULL,
  name = NULL,
  predictionProbabilitiesColumn = NULL,
  predictionThreshold = NULL,
  type = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{classCount}}{Number of classes for classification models.}

\item{\code{classNames}}{Class names for prediction results. When target type is Binary, two class names are returned. The first element is the minority (positive) class and the second element is the majority (negative) class. Limited to 100 returned for Multiclass.}

\item{\code{name}}{name of the target column}

\item{\code{predictionProbabilitiesColumn}}{Field or column name containing prediction probabilities}

\item{\code{predictionThreshold}}{Prediction threshold used for binary classification models}

\item{\code{type}}{Target type of the model.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageTarget$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageTarget$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageTarget$fromJSON(ModelPackageTargetJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ModelPackageTargetJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackageTarget$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
