% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_residuals_chart_data.R
\docType{class}
\name{ModelResidualsChartData}
\alias{ModelResidualsChartData}
\title{ModelResidualsChartData}
\format{
An \code{R6Class} generator object
}
\description{
ModelResidualsChartData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{coefficientOfDetermination}}{numeric Also known as the r-squared value. This value is calculated over the downsampled dataset, not the full input}

\item{\code{data}}{list( \link{array[numeric]} ) The rows of chart data in \link{actual, predicted, residual, row number} form. If the row number was not available at the time of model creation, the row number will be null. \strong{NOTE}: In DataRobot v5.2, the row number will not be included.}

\item{\code{histogram}}{list( \link{ModelResidualsHistogram} ) Data to plot a histogram of residual values}

\item{\code{residualMean}}{numeric The arithmetic mean of the predicted value minus the actual value over the downsampled dataset}

\item{\code{standardDeviation}}{numeric A measure of deviation from the group as a whole}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelResidualsChartData-new}{\code{ModelResidualsChartData$new()}}
\item \href{#method-ModelResidualsChartData-validate}{\code{ModelResidualsChartData$validate()}}
\item \href{#method-ModelResidualsChartData-toJSON}{\code{ModelResidualsChartData$toJSON()}}
\item \href{#method-ModelResidualsChartData-fromJSON}{\code{ModelResidualsChartData$fromJSON()}}
\item \href{#method-ModelResidualsChartData-clone}{\code{ModelResidualsChartData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelResidualsChartData-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelResidualsChartData-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelResidualsChartData$new(
  coefficientOfDetermination = NULL,
  data = NULL,
  histogram = NULL,
  residualMean = NULL,
  standardDeviation = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coefficientOfDetermination}}{Also known as the r-squared value. This value is calculated over the downsampled dataset, not the full input}

\item{\code{data}}{The rows of chart data in \link{actual, predicted, residual, row number} form. If the row number was not available at the time of model creation, the row number will be null. \strong{NOTE}: In DataRobot v5.2, the row number will not be included.}

\item{\code{histogram}}{Data to plot a histogram of residual values}

\item{\code{residualMean}}{The arithmetic mean of the predicted value minus the actual value over the downsampled dataset}

\item{\code{standardDeviation}}{A measure of deviation from the group as a whole}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelResidualsChartData-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelResidualsChartData-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelResidualsChartData$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelResidualsChartData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelResidualsChartData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelResidualsChartData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelResidualsChartData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelResidualsChartData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelResidualsChartData$fromJSON(
  ModelResidualsChartDataJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ModelResidualsChartDataJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelResidualsChartData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelResidualsChartData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelResidualsChartData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
