% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_response.R
\docType{class}
\name{ModelResponse}
\alias{ModelResponse}
\title{ModelResponse}
\format{
An \code{R6Class} generator object
}
\description{
ModelResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datasetName}}{character Name of dataset used to train challenger model}

\item{\code{description}}{character Description of the model.}

\item{\code{executionType}}{character Type of the current model.}

\item{\code{id}}{character ID of the current model.}

\item{\code{isDeprecated}}{character Whether the current model is deprecated model. eg. python2 based model.}

\item{\code{name}}{character Name of the model.}

\item{\code{projectId}}{character Project ID of the current model.}

\item{\code{projectName}}{character Project name of the current model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelResponse-new}{\code{ModelResponse$new()}}
\item \href{#method-ModelResponse-validate}{\code{ModelResponse$validate()}}
\item \href{#method-ModelResponse-toJSON}{\code{ModelResponse$toJSON()}}
\item \href{#method-ModelResponse-fromJSON}{\code{ModelResponse$fromJSON()}}
\item \href{#method-ModelResponse-clone}{\code{ModelResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelResponse$new(
  datasetName = NULL,
  description = NULL,
  executionType = NULL,
  id = NULL,
  isDeprecated = NULL,
  name = NULL,
  projectId = NULL,
  projectName = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasetName}}{Name of dataset used to train challenger model}

\item{\code{description}}{Description of the model.}

\item{\code{executionType}}{Type of the current model.}

\item{\code{id}}{ID of the current model.}

\item{\code{isDeprecated}}{Whether the current model is deprecated model. eg. python2 based model.}

\item{\code{name}}{Name of the model.}

\item{\code{projectId}}{Project ID of the current model.}

\item{\code{projectName}}{Project name of the current model.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelResponse$fromJSON(ModelResponseJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ModelResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
