% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_roc_curve_response.R
\docType{class}
\name{ModelRocCurveResponse}
\alias{ModelRocCurveResponse}
\title{ModelRocCurveResponse}
\format{
An \code{R6Class} generator object
}
\description{
ModelRocCurveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{auc}}{numeric AUC value}

\item{\code{kolmogorovSmirnovMetric}}{numeric Kolmogorov-Smirnov metric value}

\item{\code{negativeClassPredictions}}{list( numeric ) List of example predictions for the negative class.}

\item{\code{positiveClassPredictions}}{list( numeric ) List of example predictions for the positive class.}

\item{\code{rocPoints}}{list( \link{RocPointsResponse} ) The ROC curve data for that source, as specified below.}

\item{\code{source}}{character Source of the data.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelRocCurveResponse$new()}}
\item \href{#method-validate}{\code{ModelRocCurveResponse$validate()}}
\item \href{#method-toJSON}{\code{ModelRocCurveResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{ModelRocCurveResponse$fromJSON()}}
\item \href{#method-clone}{\code{ModelRocCurveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelRocCurveResponse$new(
  auc = NULL,
  kolmogorovSmirnovMetric = NULL,
  negativeClassPredictions = NULL,
  positiveClassPredictions = NULL,
  rocPoints = NULL,
  source = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{auc}}{AUC value}

\item{\code{kolmogorovSmirnovMetric}}{Kolmogorov-Smirnov metric value}

\item{\code{negativeClassPredictions}}{List of example predictions for the negative class.}

\item{\code{positiveClassPredictions}}{List of example predictions for the positive class.}

\item{\code{rocPoints}}{The ROC curve data for that source, as specified below.}

\item{\code{source}}{Source of the data.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelRocCurveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelRocCurveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelRocCurveResponse$fromJSON(
  ModelRocCurveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ModelRocCurveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelRocCurveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
