% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_update.R
\docType{class}
\name{ModelUpdate}
\alias{ModelUpdate}
\title{ModelUpdate}
\format{
An \code{R6Class} generator object
}
\description{
ModelUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{isStarred}}{character \link{optional} Mark model either as starred or unstarred.}

\item{\code{predictionThreshold}}{numeric \link{optional} Threshold used for binary classification in predictions. Default value is 0.5.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelUpdate-new}{\code{ModelUpdate$new()}}
\item \href{#method-ModelUpdate-validate}{\code{ModelUpdate$validate()}}
\item \href{#method-ModelUpdate-toJSON}{\code{ModelUpdate$toJSON()}}
\item \href{#method-ModelUpdate-fromJSON}{\code{ModelUpdate$fromJSON()}}
\item \href{#method-ModelUpdate-clone}{\code{ModelUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelUpdate-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelUpdate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelUpdate$new(
  isStarred = NULL,
  predictionThreshold = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{isStarred}}{Mark model either as starred or unstarred.}

\item{\code{predictionThreshold}}{Threshold used for binary classification in predictions. Default value is 0.5.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelUpdate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelUpdate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelUpdate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelUpdate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelUpdate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelUpdate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelUpdate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelUpdate$fromJSON(ModelUpdateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ModelUpdateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelUpdate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelUpdate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
