% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicategorical_invalid_format_error_list.R
\docType{class}
\name{MulticategoricalInvalidFormatErrorList}
\alias{MulticategoricalInvalidFormatErrorList}
\title{MulticategoricalInvalidFormatErrorList}
\format{
An \code{R6Class} generator object
}
\description{
MulticategoricalInvalidFormatErrorList Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{error}}{character Error type.}

\item{\code{feature}}{character Feature name.}

\item{\code{rowData}}{character Content of the row containing format error.}

\item{\code{rowIndex}}{integer Row index of the row containing format error.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MulticategoricalInvalidFormatErrorList-new}{\code{MulticategoricalInvalidFormatErrorList$new()}}
\item \href{#method-MulticategoricalInvalidFormatErrorList-validate}{\code{MulticategoricalInvalidFormatErrorList$validate()}}
\item \href{#method-MulticategoricalInvalidFormatErrorList-toJSON}{\code{MulticategoricalInvalidFormatErrorList$toJSON()}}
\item \href{#method-MulticategoricalInvalidFormatErrorList-fromJSON}{\code{MulticategoricalInvalidFormatErrorList$fromJSON()}}
\item \href{#method-MulticategoricalInvalidFormatErrorList-clone}{\code{MulticategoricalInvalidFormatErrorList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MulticategoricalInvalidFormatErrorList-new"></a>}}
\if{latex}{\out{\hypertarget{method-MulticategoricalInvalidFormatErrorList-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticategoricalInvalidFormatErrorList$new(
  error = NULL,
  feature = NULL,
  rowData = NULL,
  rowIndex = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{error}}{Error type.}

\item{\code{feature}}{Feature name.}

\item{\code{rowData}}{Content of the row containing format error.}

\item{\code{rowIndex}}{Row index of the row containing format error.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MulticategoricalInvalidFormatErrorList-validate"></a>}}
\if{latex}{\out{\hypertarget{method-MulticategoricalInvalidFormatErrorList-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticategoricalInvalidFormatErrorList$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MulticategoricalInvalidFormatErrorList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MulticategoricalInvalidFormatErrorList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticategoricalInvalidFormatErrorList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MulticategoricalInvalidFormatErrorList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MulticategoricalInvalidFormatErrorList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticategoricalInvalidFormatErrorList$fromJSON(
  MulticategoricalInvalidFormatErrorListJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MulticategoricalInvalidFormatErrorListJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MulticategoricalInvalidFormatErrorList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MulticategoricalInvalidFormatErrorList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticategoricalInvalidFormatErrorList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
