% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_datetime_feature_effects_response.R
\docType{class}
\name{MulticlassDatetimeFeatureEffectsResponse}
\alias{MulticlassDatetimeFeatureEffectsResponse}
\title{MulticlassDatetimeFeatureEffectsResponse}
\format{
An \code{R6Class} generator object
}
\description{
MulticlassDatetimeFeatureEffectsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{backtestIndex}}{character The backtest index. For example: `0`, `1`, ..., `20`, `holdout`, `startstop`.}

\item{\code{count}}{integer \link{optional} Number of items returned on this page.}

\item{\code{data}}{list( \link{MulticlassFeatureEffects} ) A list of feature effects scores for each class in multiclass project.}

\item{\code{modelId}}{character The model ID}

\item{\code{next_}}{character URL pointing to the next page (if null, there is no next page).}

\item{\code{previous}}{character URL pointing to the previous page (if null, there is no previous page).}

\item{\code{projectId}}{character The project ID}

\item{\code{source}}{character Models data source.}

\item{\code{totalCount}}{integer The total number of items across all pages.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MulticlassDatetimeFeatureEffectsResponse$new()}}
\item \href{#method-validate}{\code{MulticlassDatetimeFeatureEffectsResponse$validate()}}
\item \href{#method-toJSON}{\code{MulticlassDatetimeFeatureEffectsResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{MulticlassDatetimeFeatureEffectsResponse$fromJSON()}}
\item \href{#method-clone}{\code{MulticlassDatetimeFeatureEffectsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassDatetimeFeatureEffectsResponse$new(
  backtestIndex = NULL,
  data = NULL,
  modelId = NULL,
  next_ = NULL,
  previous = NULL,
  projectId = NULL,
  source = NULL,
  totalCount = NULL,
  count = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{backtestIndex}}{The backtest index. For example: `0`, `1`, ..., `20`, `holdout`, `startstop`.}

\item{\code{data}}{A list of feature effects scores for each class in multiclass project.}

\item{\code{modelId}}{The model ID}

\item{\code{next_}}{URL pointing to the next page (if null, there is no next page).}

\item{\code{previous}}{URL pointing to the previous page (if null, there is no previous page).}

\item{\code{projectId}}{The project ID}

\item{\code{source}}{Models data source.}

\item{\code{totalCount}}{The total number of items across all pages.}

\item{\code{count}}{Number of items returned on this page.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassDatetimeFeatureEffectsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassDatetimeFeatureEffectsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassDatetimeFeatureEffectsResponse$fromJSON(
  MulticlassDatetimeFeatureEffectsResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MulticlassDatetimeFeatureEffectsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassDatetimeFeatureEffectsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
