% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_feature_effects.R
\docType{class}
\name{MulticlassFeatureEffects}
\alias{MulticlassFeatureEffects}
\title{MulticlassFeatureEffects}
\format{
An \code{R6Class} generator object
}
\description{
MulticlassFeatureEffects Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{class}}{character Target class label.}

\item{\code{featureImpactScore}}{numeric Feature impact score.}

\item{\code{featureName}}{character Name of the feature.}

\item{\code{featureType}}{character Feature type, either numeric or categorical.}

\item{\code{isBinnable}}{character Whether values can be grouped into bins.}

\item{\code{isScalable}}{character Whether numeric feature values can be reported on a log scale.}

\item{\code{partialDependence}}{\link{PartialDependence} \link{optional}}

\item{\code{predictedVsActual}}{\link{PredictedVsActual} \link{optional}}

\item{\code{weightLabel}}{character Weight label if weight was configured for the project.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MulticlassFeatureEffects$new()}}
\item \href{#method-validate}{\code{MulticlassFeatureEffects$validate()}}
\item \href{#method-toJSON}{\code{MulticlassFeatureEffects$toJSON()}}
\item \href{#method-fromJSON}{\code{MulticlassFeatureEffects$fromJSON()}}
\item \href{#method-clone}{\code{MulticlassFeatureEffects$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureEffects$new(
  class = NULL,
  featureImpactScore = NULL,
  featureName = NULL,
  featureType = NULL,
  isBinnable = NULL,
  isScalable = NULL,
  weightLabel = NULL,
  partialDependence = NULL,
  predictedVsActual = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class}}{Target class label.}

\item{\code{featureImpactScore}}{Feature impact score.}

\item{\code{featureName}}{Name of the feature.}

\item{\code{featureType}}{Feature type, either numeric or categorical.}

\item{\code{isBinnable}}{Whether values can be grouped into bins.}

\item{\code{isScalable}}{Whether numeric feature values can be reported on a log scale.}

\item{\code{weightLabel}}{Weight label if weight was configured for the project.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureEffects$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureEffects$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureEffects$fromJSON(
  MulticlassFeatureEffectsJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MulticlassFeatureEffectsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureEffects$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
