% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_feature_effects_response.R
\docType{class}
\name{MulticlassFeatureEffectsResponse}
\alias{MulticlassFeatureEffectsResponse}
\title{MulticlassFeatureEffectsResponse}
\format{
An \code{R6Class} generator object
}
\description{
MulticlassFeatureEffectsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer \link{optional} Number of items returned on this page.}

\item{\code{data}}{list( \link{MulticlassFeatureEffects} ) A list of feature effects scores for each class in multiclass project.}

\item{\code{modelId}}{character The model ID}

\item{\code{next_}}{character URL pointing to the next page (if null, there is no next page).}

\item{\code{previous}}{character URL pointing to the previous page (if null, there is no previous page).}

\item{\code{projectId}}{character The project ID}

\item{\code{source}}{character Models data source.}

\item{\code{totalCount}}{integer The total number of items across all pages.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MulticlassFeatureEffectsResponse-new}{\code{MulticlassFeatureEffectsResponse$new()}}
\item \href{#method-MulticlassFeatureEffectsResponse-validate}{\code{MulticlassFeatureEffectsResponse$validate()}}
\item \href{#method-MulticlassFeatureEffectsResponse-toJSON}{\code{MulticlassFeatureEffectsResponse$toJSON()}}
\item \href{#method-MulticlassFeatureEffectsResponse-fromJSON}{\code{MulticlassFeatureEffectsResponse$fromJSON()}}
\item \href{#method-MulticlassFeatureEffectsResponse-clone}{\code{MulticlassFeatureEffectsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MulticlassFeatureEffectsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-MulticlassFeatureEffectsResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureEffectsResponse$new(
  data = NULL,
  modelId = NULL,
  next_ = NULL,
  previous = NULL,
  projectId = NULL,
  source = NULL,
  totalCount = NULL,
  count = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A list of feature effects scores for each class in multiclass project.}

\item{\code{modelId}}{The model ID}

\item{\code{next_}}{URL pointing to the next page (if null, there is no next page).}

\item{\code{previous}}{URL pointing to the previous page (if null, there is no previous page).}

\item{\code{projectId}}{The project ID}

\item{\code{source}}{Models data source.}

\item{\code{totalCount}}{The total number of items across all pages.}

\item{\code{count}}{Number of items returned on this page.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MulticlassFeatureEffectsResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-MulticlassFeatureEffectsResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureEffectsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MulticlassFeatureEffectsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MulticlassFeatureEffectsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureEffectsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MulticlassFeatureEffectsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MulticlassFeatureEffectsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureEffectsResponse$fromJSON(
  MulticlassFeatureEffectsResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MulticlassFeatureEffectsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MulticlassFeatureEffectsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MulticlassFeatureEffectsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureEffectsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
