% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_feature_impact.R
\docType{class}
\name{MulticlassFeatureImpact}
\alias{MulticlassFeatureImpact}
\title{MulticlassFeatureImpact}
\format{
An \code{R6Class} generator object
}
\description{
MulticlassFeatureImpact Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{class}}{character Target class label.}

\item{\code{featureImpacts}}{list( \link{FeatureImpactItem} ) A list which contains feature impact scores for each feature used by a model. If the model has more than 1000 features, the most important 1000 features are returned.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MulticlassFeatureImpact-new}{\code{MulticlassFeatureImpact$new()}}
\item \href{#method-MulticlassFeatureImpact-validate}{\code{MulticlassFeatureImpact$validate()}}
\item \href{#method-MulticlassFeatureImpact-toJSON}{\code{MulticlassFeatureImpact$toJSON()}}
\item \href{#method-MulticlassFeatureImpact-fromJSON}{\code{MulticlassFeatureImpact$fromJSON()}}
\item \href{#method-MulticlassFeatureImpact-clone}{\code{MulticlassFeatureImpact$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MulticlassFeatureImpact-new"></a>}}
\if{latex}{\out{\hypertarget{method-MulticlassFeatureImpact-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureImpact$new(
  class = NULL,
  featureImpacts = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class}}{Target class label.}

\item{\code{featureImpacts}}{A list which contains feature impact scores for each feature used by a model. If the model has more than 1000 features, the most important 1000 features are returned.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MulticlassFeatureImpact-validate"></a>}}
\if{latex}{\out{\hypertarget{method-MulticlassFeatureImpact-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureImpact$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MulticlassFeatureImpact-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MulticlassFeatureImpact-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureImpact$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MulticlassFeatureImpact-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MulticlassFeatureImpact-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureImpact$fromJSON(
  MulticlassFeatureImpactJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MulticlassFeatureImpactJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MulticlassFeatureImpact-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MulticlassFeatureImpact-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureImpact$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
