% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_feature_impact_response.R
\docType{class}
\name{MulticlassFeatureImpactResponse}
\alias{MulticlassFeatureImpactResponse}
\title{MulticlassFeatureImpactResponse}
\format{
An \code{R6Class} generator object
}
\description{
MulticlassFeatureImpactResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{classFeatureImpacts}}{list( \link{MulticlassFeatureImpact} ) A list of feature importance scores for each class in multiclass project.}

\item{\code{ranRedundancyDetection}}{character Indicates whether redundant feature identification was run while calculating this feature impact. Currently always False, as redundant feature identification isn't supported for multiclass in DataRobot.}

\item{\code{shapBased}}{character Indicates whether feature impact was calculated using Shapley values. Currently always `False`, as SHAP isn't supported for multiclass in DataRobot.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MulticlassFeatureImpactResponse$new()}}
\item \href{#method-validate}{\code{MulticlassFeatureImpactResponse$validate()}}
\item \href{#method-toJSON}{\code{MulticlassFeatureImpactResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{MulticlassFeatureImpactResponse$fromJSON()}}
\item \href{#method-clone}{\code{MulticlassFeatureImpactResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureImpactResponse$new(
  classFeatureImpacts = NULL,
  ranRedundancyDetection = NULL,
  shapBased = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{classFeatureImpacts}}{A list of feature importance scores for each class in multiclass project.}

\item{\code{ranRedundancyDetection}}{Indicates whether redundant feature identification was run while calculating this feature impact. Currently always False, as redundant feature identification isn't supported for multiclass in DataRobot.}

\item{\code{shapBased}}{Indicates whether feature impact was calculated using Shapley values. Currently always `False`, as SHAP isn't supported for multiclass in DataRobot.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureImpactResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureImpactResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureImpactResponse$fromJSON(
  MulticlassFeatureImpactResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MulticlassFeatureImpactResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureImpactResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
