% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_lift_data.R
\docType{class}
\name{MulticlassLiftData}
\alias{MulticlassLiftData}
\title{MulticlassLiftData}
\format{
An \code{R6Class} generator object
}
\description{
MulticlassLiftData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{classBins}}{list( \link{MulticlassLiftDataClassBins} ) List of lift chart data for each target class.}

\item{\code{datasetId}}{character The dataset id of dataset which was used to compute Lift chart.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MulticlassLiftData-new}{\code{MulticlassLiftData$new()}}
\item \href{#method-MulticlassLiftData-validate}{\code{MulticlassLiftData$validate()}}
\item \href{#method-MulticlassLiftData-toJSON}{\code{MulticlassLiftData$toJSON()}}
\item \href{#method-MulticlassLiftData-fromJSON}{\code{MulticlassLiftData$fromJSON()}}
\item \href{#method-MulticlassLiftData-clone}{\code{MulticlassLiftData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MulticlassLiftData-new"></a>}}
\if{latex}{\out{\hypertarget{method-MulticlassLiftData-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassLiftData$new(
  classBins = NULL,
  datasetId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{classBins}}{List of lift chart data for each target class.}

\item{\code{datasetId}}{The dataset id of dataset which was used to compute Lift chart.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MulticlassLiftData-validate"></a>}}
\if{latex}{\out{\hypertarget{method-MulticlassLiftData-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassLiftData$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MulticlassLiftData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MulticlassLiftData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassLiftData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MulticlassLiftData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MulticlassLiftData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassLiftData$fromJSON(MulticlassLiftDataJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MulticlassLiftDataJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MulticlassLiftData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MulticlassLiftData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassLiftData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
