% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiseries_histograms_bin.R
\docType{class}
\name{MultiseriesHistogramsBin}
\alias{MultiseriesHistogramsBin}
\title{MultiseriesHistogramsBin}
\format{
An \code{R6Class} generator object
}
\description{
MultiseriesHistogramsBin Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer The value count of the bin}

\item{\code{left}}{\link{OneOfnumberstring} The inclusive left boundary of the bin.}

\item{\code{right}}{\link{OneOfnumberstring} The exclusive right boundary of the bin. The last bin has an inclusive right boundary.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultiseriesHistogramsBin-new}{\code{MultiseriesHistogramsBin$new()}}
\item \href{#method-MultiseriesHistogramsBin-validate}{\code{MultiseriesHistogramsBin$validate()}}
\item \href{#method-MultiseriesHistogramsBin-toJSON}{\code{MultiseriesHistogramsBin$toJSON()}}
\item \href{#method-MultiseriesHistogramsBin-fromJSON}{\code{MultiseriesHistogramsBin$fromJSON()}}
\item \href{#method-MultiseriesHistogramsBin-clone}{\code{MultiseriesHistogramsBin$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiseriesHistogramsBin-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultiseriesHistogramsBin-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiseriesHistogramsBin$new(
  count = NULL,
  left = NULL,
  right = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{The value count of the bin}

\item{\code{left}}{The inclusive left boundary of the bin.}

\item{\code{right}}{The exclusive right boundary of the bin. The last bin has an inclusive right boundary.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiseriesHistogramsBin-validate"></a>}}
\if{latex}{\out{\hypertarget{method-MultiseriesHistogramsBin-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiseriesHistogramsBin$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiseriesHistogramsBin-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MultiseriesHistogramsBin-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiseriesHistogramsBin$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiseriesHistogramsBin-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MultiseriesHistogramsBin-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiseriesHistogramsBin$fromJSON(
  MultiseriesHistogramsBinJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MultiseriesHistogramsBinJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiseriesHistogramsBin-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultiseriesHistogramsBin-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiseriesHistogramsBin$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
