% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiseries_names_controller_response.R
\docType{class}
\name{MultiseriesNamesControllerResponse}
\alias{MultiseriesNamesControllerResponse}
\title{MultiseriesNamesControllerResponse}
\format{
An \code{R6Class} generator object
}
\description{
MultiseriesNamesControllerResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer Total number of series items in the response}

\item{\code{data}}{\link{MultiseriesNamesControllerDataRecord}}

\item{\code{next_}}{character A URL pointing to the next page (if `null`, there is no next page).}

\item{\code{previous}}{character A URL pointing to the previous page (if `null`, there is no previous page).}

\item{\code{totalSeriesCount}}{integer Total number of series items}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultiseriesNamesControllerResponse-new}{\code{MultiseriesNamesControllerResponse$new()}}
\item \href{#method-MultiseriesNamesControllerResponse-validate}{\code{MultiseriesNamesControllerResponse$validate()}}
\item \href{#method-MultiseriesNamesControllerResponse-toJSON}{\code{MultiseriesNamesControllerResponse$toJSON()}}
\item \href{#method-MultiseriesNamesControllerResponse-fromJSON}{\code{MultiseriesNamesControllerResponse$fromJSON()}}
\item \href{#method-MultiseriesNamesControllerResponse-clone}{\code{MultiseriesNamesControllerResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiseriesNamesControllerResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultiseriesNamesControllerResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiseriesNamesControllerResponse$new(
  count = NULL,
  data = NULL,
  next_ = NULL,
  previous = NULL,
  totalSeriesCount = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{Total number of series items in the response}

\item{\code{next_}}{A URL pointing to the next page (if `null`, there is no next page).}

\item{\code{previous}}{A URL pointing to the previous page (if `null`, there is no previous page).}

\item{\code{totalSeriesCount}}{Total number of series items}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiseriesNamesControllerResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-MultiseriesNamesControllerResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiseriesNamesControllerResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiseriesNamesControllerResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MultiseriesNamesControllerResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiseriesNamesControllerResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiseriesNamesControllerResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MultiseriesNamesControllerResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiseriesNamesControllerResponse$fromJSON(
  MultiseriesNamesControllerResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MultiseriesNamesControllerResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiseriesNamesControllerResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultiseriesNamesControllerResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiseriesNamesControllerResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
