% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiseries_payload.R
\docType{class}
\name{MultiseriesPayload}
\alias{MultiseriesPayload}
\title{MultiseriesPayload}
\format{
An \code{R6Class} generator object
}
\description{
MultiseriesPayload Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datetimePartitionColumn}}{character The date column that will be used to perform detection and validation for.}

\item{\code{multiseriesIdColumns}}{list( character ) \link{optional} List of one or more names of potential multiseries id columns.  If not provided, all numerical and categorical columns are used.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultiseriesPayload-new}{\code{MultiseriesPayload$new()}}
\item \href{#method-MultiseriesPayload-validate}{\code{MultiseriesPayload$validate()}}
\item \href{#method-MultiseriesPayload-toJSON}{\code{MultiseriesPayload$toJSON()}}
\item \href{#method-MultiseriesPayload-fromJSON}{\code{MultiseriesPayload$fromJSON()}}
\item \href{#method-MultiseriesPayload-clone}{\code{MultiseriesPayload$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiseriesPayload-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultiseriesPayload-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiseriesPayload$new(
  datetimePartitionColumn = NULL,
  multiseriesIdColumns = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datetimePartitionColumn}}{The date column that will be used to perform detection and validation for.}

\item{\code{multiseriesIdColumns}}{List of one or more names of potential multiseries id columns.  If not provided, all numerical and categorical columns are used.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiseriesPayload-validate"></a>}}
\if{latex}{\out{\hypertarget{method-MultiseriesPayload-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiseriesPayload$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiseriesPayload-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MultiseriesPayload-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiseriesPayload$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiseriesPayload-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MultiseriesPayload-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiseriesPayload$fromJSON(MultiseriesPayloadJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MultiseriesPayloadJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiseriesPayload-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultiseriesPayload-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiseriesPayload$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
