% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notification_channel.R
\docType{class}
\name{NotificationChannel}
\alias{NotificationChannel}
\title{NotificationChannel}
\format{
An \code{R6Class} generator object
}
\description{
NotificationChannel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{channelType}}{character The type of the new notification channel.}

\item{\code{contentType}}{character The content type of the messages of the new notification channel.}

\item{\code{createdAt}}{character The date of the notification channel creation.}

\item{\code{customHeaders}}{list( \link{CustomerHeader} ) Custom headers and their values to be sent in the new notification channel.}

\item{\code{emailAddress}}{character The email address to be used in the new notification channel.}

\item{\code{id}}{character The id of the notification channel.}

\item{\code{languageCode}}{character The preferred language code.}

\item{\code{name}}{character The name of the new notification channel.}

\item{\code{orgId}}{character The id of organization that notification channel belongs to.}

\item{\code{payloadUrl}}{character The payload URL of the new notification channel.}

\item{\code{secretToken}}{character Secret token to be used for new notification channel.}

\item{\code{uid}}{character The identifier of the user who created the channel.}

\item{\code{validateSsl}}{character Defines if  validate ssl or not in the notification channel.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NotificationChannel-new}{\code{NotificationChannel$new()}}
\item \href{#method-NotificationChannel-validate}{\code{NotificationChannel$validate()}}
\item \href{#method-NotificationChannel-toJSON}{\code{NotificationChannel$toJSON()}}
\item \href{#method-NotificationChannel-fromJSON}{\code{NotificationChannel$fromJSON()}}
\item \href{#method-NotificationChannel-clone}{\code{NotificationChannel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationChannel-new"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationChannel-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationChannel$new(
  channelType = NULL,
  contentType = NULL,
  createdAt = NULL,
  customHeaders = NULL,
  emailAddress = NULL,
  id = NULL,
  languageCode = NULL,
  name = NULL,
  orgId = NULL,
  payloadUrl = NULL,
  secretToken = NULL,
  uid = NULL,
  validateSsl = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{channelType}}{The type of the new notification channel.}

\item{\code{contentType}}{The content type of the messages of the new notification channel.}

\item{\code{createdAt}}{The date of the notification channel creation.}

\item{\code{customHeaders}}{Custom headers and their values to be sent in the new notification channel.}

\item{\code{emailAddress}}{The email address to be used in the new notification channel.}

\item{\code{id}}{The id of the notification channel.}

\item{\code{languageCode}}{The preferred language code.}

\item{\code{name}}{The name of the new notification channel.}

\item{\code{orgId}}{The id of organization that notification channel belongs to.}

\item{\code{payloadUrl}}{The payload URL of the new notification channel.}

\item{\code{secretToken}}{Secret token to be used for new notification channel.}

\item{\code{uid}}{The identifier of the user who created the channel.}

\item{\code{validateSsl}}{Defines if  validate ssl or not in the notification channel.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationChannel-validate"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationChannel-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationChannel$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationChannel-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationChannel-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationChannel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationChannel-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationChannel-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationChannel$fromJSON(NotificationChannelJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{NotificationChannelJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationChannel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationChannel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationChannel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
