% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notification_channels_list_response.R
\docType{class}
\name{NotificationChannelsListResponse}
\alias{NotificationChannelsListResponse}
\title{NotificationChannelsListResponse}
\format{
An \code{R6Class} generator object
}
\description{
NotificationChannelsListResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer The number of channels returned.}

\item{\code{data}}{list( \link{NotificationChannelResponse} ) The notification channels.}

\item{\code{next_}}{character URL pointing to the next page.}

\item{\code{previous}}{character URL pointing to the previous page.}

\item{\code{totalCount}}{integer The total number of channels that satisfy the query.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NotificationChannelsListResponse-new}{\code{NotificationChannelsListResponse$new()}}
\item \href{#method-NotificationChannelsListResponse-validate}{\code{NotificationChannelsListResponse$validate()}}
\item \href{#method-NotificationChannelsListResponse-toJSON}{\code{NotificationChannelsListResponse$toJSON()}}
\item \href{#method-NotificationChannelsListResponse-fromJSON}{\code{NotificationChannelsListResponse$fromJSON()}}
\item \href{#method-NotificationChannelsListResponse-clone}{\code{NotificationChannelsListResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationChannelsListResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationChannelsListResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationChannelsListResponse$new(
  count = NULL,
  data = NULL,
  next_ = NULL,
  previous = NULL,
  totalCount = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{The number of channels returned.}

\item{\code{data}}{The notification channels.}

\item{\code{next_}}{URL pointing to the next page.}

\item{\code{previous}}{URL pointing to the previous page.}

\item{\code{totalCount}}{The total number of channels that satisfy the query.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationChannelsListResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationChannelsListResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationChannelsListResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationChannelsListResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationChannelsListResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationChannelsListResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationChannelsListResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationChannelsListResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationChannelsListResponse$fromJSON(
  NotificationChannelsListResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{NotificationChannelsListResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationChannelsListResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationChannelsListResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationChannelsListResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
