% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notification_channel_update.R
\docType{class}
\name{NotificationChannelUpdate}
\alias{NotificationChannelUpdate}
\title{NotificationChannelUpdate}
\format{
An \code{R6Class} generator object
}
\description{
NotificationChannelUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{contentType}}{character \link{optional} The content type of the messages of the new notification channel.}

\item{\code{customHeaders}}{list( \link{CustomerHeader} ) \link{optional} Custom headers and their values to be sent in the new notification channel.}

\item{\code{emailAddress}}{character \link{optional} The email address to be used in the new notification channel.}

\item{\code{languageCode}}{character \link{optional} The preferred language code.}

\item{\code{name}}{character \link{optional} The name of the new notification channel.}

\item{\code{payloadUrl}}{character \link{optional} The payload URL of the new notification channel.}

\item{\code{secretToken}}{character \link{optional} Secret token to be used for new notification channel.}

\item{\code{validateSsl}}{character \link{optional} Defines if  validate ssl or not in the notification channel.}

\item{\code{verificationCode}}{character \link{optional} Required if the channel type is Email.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NotificationChannelUpdate-new}{\code{NotificationChannelUpdate$new()}}
\item \href{#method-NotificationChannelUpdate-validate}{\code{NotificationChannelUpdate$validate()}}
\item \href{#method-NotificationChannelUpdate-toJSON}{\code{NotificationChannelUpdate$toJSON()}}
\item \href{#method-NotificationChannelUpdate-fromJSON}{\code{NotificationChannelUpdate$fromJSON()}}
\item \href{#method-NotificationChannelUpdate-clone}{\code{NotificationChannelUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationChannelUpdate-new"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationChannelUpdate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationChannelUpdate$new(
  contentType = NULL,
  customHeaders = NULL,
  emailAddress = NULL,
  languageCode = NULL,
  name = NULL,
  payloadUrl = NULL,
  secretToken = NULL,
  validateSsl = NULL,
  verificationCode = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{contentType}}{The content type of the messages of the new notification channel.}

\item{\code{customHeaders}}{Custom headers and their values to be sent in the new notification channel.}

\item{\code{emailAddress}}{The email address to be used in the new notification channel.}

\item{\code{languageCode}}{The preferred language code.}

\item{\code{name}}{The name of the new notification channel.}

\item{\code{payloadUrl}}{The payload URL of the new notification channel.}

\item{\code{secretToken}}{Secret token to be used for new notification channel.}

\item{\code{validateSsl}}{Defines if  validate ssl or not in the notification channel.}

\item{\code{verificationCode}}{Required if the channel type is Email.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationChannelUpdate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationChannelUpdate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationChannelUpdate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationChannelUpdate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationChannelUpdate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationChannelUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationChannelUpdate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationChannelUpdate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationChannelUpdate$fromJSON(
  NotificationChannelUpdateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{NotificationChannelUpdateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationChannelUpdate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationChannelUpdate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationChannelUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
