% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notification_email_channel_verification.R
\docType{class}
\name{NotificationEmailChannelVerification}
\alias{NotificationEmailChannelVerification}
\title{NotificationEmailChannelVerification}
\format{
An \code{R6Class} generator object
}
\description{
NotificationEmailChannelVerification Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{channelType}}{character The type of the new notification channel}

\item{\code{emailAddress}}{character The email address of the recipient}

\item{\code{name}}{character The name of the new notification channel}

\item{\code{orgId}}{character \link{optional} The id of organization that notification channel belongs to}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NotificationEmailChannelVerification-new}{\code{NotificationEmailChannelVerification$new()}}
\item \href{#method-NotificationEmailChannelVerification-validate}{\code{NotificationEmailChannelVerification$validate()}}
\item \href{#method-NotificationEmailChannelVerification-toJSON}{\code{NotificationEmailChannelVerification$toJSON()}}
\item \href{#method-NotificationEmailChannelVerification-fromJSON}{\code{NotificationEmailChannelVerification$fromJSON()}}
\item \href{#method-NotificationEmailChannelVerification-clone}{\code{NotificationEmailChannelVerification$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationEmailChannelVerification-new"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationEmailChannelVerification-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationEmailChannelVerification$new(
  channelType = NULL,
  emailAddress = NULL,
  name = NULL,
  orgId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{channelType}}{The type of the new notification channel}

\item{\code{emailAddress}}{The email address of the recipient}

\item{\code{name}}{The name of the new notification channel}

\item{\code{orgId}}{The id of organization that notification channel belongs to}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationEmailChannelVerification-validate"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationEmailChannelVerification-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationEmailChannelVerification$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationEmailChannelVerification-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationEmailChannelVerification-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationEmailChannelVerification$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationEmailChannelVerification-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationEmailChannelVerification-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationEmailChannelVerification$fromJSON(
  NotificationEmailChannelVerificationJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{NotificationEmailChannelVerificationJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationEmailChannelVerification-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationEmailChannelVerification-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationEmailChannelVerification$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
