% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notification_log.R
\docType{class}
\name{NotificationLog}
\alias{NotificationLog}
\title{NotificationLog}
\format{
An \code{R6Class} generator object
}
\description{
NotificationLog Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{channelId}}{character The id of the channel that was used to send the notification.}

\item{\code{emailSubject}}{character The email subject of the notification.}

\item{\code{id}}{character The id of the notification log.}

\item{\code{parentNotificationId}}{character The id of the parent notification.}

\item{\code{policyId}}{character The id of the policy that was used to send the notification.}

\item{\code{request}}{character The request that was sent in the notification.}

\item{\code{response}}{character The response that was received after sending the notification.}

\item{\code{retryCount}}{integer The number of attempts to send the notification.}

\item{\code{status}}{character The status of the notification.}

\item{\code{timestamp}}{character The date when the notification was sent.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NotificationLog$new()}}
\item \href{#method-validate}{\code{NotificationLog$validate()}}
\item \href{#method-toJSON}{\code{NotificationLog$toJSON()}}
\item \href{#method-fromJSON}{\code{NotificationLog$fromJSON()}}
\item \href{#method-clone}{\code{NotificationLog$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationLog$new(
  channelId = NULL,
  emailSubject = NULL,
  id = NULL,
  parentNotificationId = NULL,
  policyId = NULL,
  request = NULL,
  response = NULL,
  retryCount = NULL,
  status = NULL,
  timestamp = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{channelId}}{The id of the channel that was used to send the notification.}

\item{\code{emailSubject}}{The email subject of the notification.}

\item{\code{id}}{The id of the notification log.}

\item{\code{parentNotificationId}}{The id of the parent notification.}

\item{\code{policyId}}{The id of the policy that was used to send the notification.}

\item{\code{request}}{The request that was sent in the notification.}

\item{\code{response}}{The response that was received after sending the notification.}

\item{\code{retryCount}}{The number of attempts to send the notification.}

\item{\code{status}}{The status of the notification.}

\item{\code{timestamp}}{The date when the notification was sent.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationLog$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationLog$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationLog$fromJSON(NotificationLogJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{NotificationLogJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationLog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
