% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notification_log_list_data_response.R
\docType{class}
\name{NotificationLogListDataResponse}
\alias{NotificationLogListDataResponse}
\title{NotificationLogListDataResponse}
\format{
An \code{R6Class} generator object
}
\description{
NotificationLogListDataResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{channelId}}{character The id of the channel that was used to send the notification}

\item{\code{emailSubject}}{character The email subject}

\item{\code{id}}{character The id of the notification log}

\item{\code{parentNotificationId}}{character The ID of the parent notification}

\item{\code{policyId}}{character the id of the policy that was used to send the notification}

\item{\code{request}}{\link{RequestNotification}}

\item{\code{response}}{\link{ResponseNotification}}

\item{\code{retryCount}}{integer The count of the retries while sending the notification}

\item{\code{status}}{character The status of the notification}

\item{\code{timestamp}}{character The date and time when the notification was sent}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NotificationLogListDataResponse-new}{\code{NotificationLogListDataResponse$new()}}
\item \href{#method-NotificationLogListDataResponse-validate}{\code{NotificationLogListDataResponse$validate()}}
\item \href{#method-NotificationLogListDataResponse-toJSON}{\code{NotificationLogListDataResponse$toJSON()}}
\item \href{#method-NotificationLogListDataResponse-fromJSON}{\code{NotificationLogListDataResponse$fromJSON()}}
\item \href{#method-NotificationLogListDataResponse-clone}{\code{NotificationLogListDataResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationLogListDataResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationLogListDataResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationLogListDataResponse$new(
  channelId = NULL,
  emailSubject = NULL,
  id = NULL,
  parentNotificationId = NULL,
  policyId = NULL,
  request = NULL,
  response = NULL,
  retryCount = NULL,
  status = NULL,
  timestamp = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{channelId}}{The id of the channel that was used to send the notification}

\item{\code{emailSubject}}{The email subject}

\item{\code{id}}{The id of the notification log}

\item{\code{parentNotificationId}}{The ID of the parent notification}

\item{\code{policyId}}{the id of the policy that was used to send the notification}

\item{\code{retryCount}}{The count of the retries while sending the notification}

\item{\code{status}}{The status of the notification}

\item{\code{timestamp}}{The date and time when the notification was sent}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationLogListDataResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationLogListDataResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationLogListDataResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationLogListDataResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationLogListDataResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationLogListDataResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationLogListDataResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationLogListDataResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationLogListDataResponse$fromJSON(
  NotificationLogListDataResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{NotificationLogListDataResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationLogListDataResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationLogListDataResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationLogListDataResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
