% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notification_policy_mute_create.R
\docType{class}
\name{NotificationPolicyMuteCreate}
\alias{NotificationPolicyMuteCreate}
\title{NotificationPolicyMuteCreate}
\format{
An \code{R6Class} generator object
}
\description{
NotificationPolicyMuteCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entityId}}{character The id of the entity to mute the notification for}

\item{\code{policyId}}{character The id of the policy to mute notification for}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NotificationPolicyMuteCreate-new}{\code{NotificationPolicyMuteCreate$new()}}
\item \href{#method-NotificationPolicyMuteCreate-validate}{\code{NotificationPolicyMuteCreate$validate()}}
\item \href{#method-NotificationPolicyMuteCreate-toJSON}{\code{NotificationPolicyMuteCreate$toJSON()}}
\item \href{#method-NotificationPolicyMuteCreate-fromJSON}{\code{NotificationPolicyMuteCreate$fromJSON()}}
\item \href{#method-NotificationPolicyMuteCreate-clone}{\code{NotificationPolicyMuteCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationPolicyMuteCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationPolicyMuteCreate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationPolicyMuteCreate$new(
  entityId = NULL,
  policyId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entityId}}{The id of the entity to mute the notification for}

\item{\code{policyId}}{The id of the policy to mute notification for}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationPolicyMuteCreate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationPolicyMuteCreate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationPolicyMuteCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationPolicyMuteCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationPolicyMuteCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationPolicyMuteCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationPolicyMuteCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationPolicyMuteCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationPolicyMuteCreate$fromJSON(
  NotificationPolicyMuteCreateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{NotificationPolicyMuteCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationPolicyMuteCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationPolicyMuteCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationPolicyMuteCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
