% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notification_policy_mute_list_response.R
\docType{class}
\name{NotificationPolicyMuteListResponse}
\alias{NotificationPolicyMuteListResponse}
\title{NotificationPolicyMuteListResponse}
\format{
An \code{R6Class} generator object
}
\description{
NotificationPolicyMuteListResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer The number of items returned on this page}

\item{\code{data}}{list( \link{NotificationPolicyMuteResponse} ) A list of Notification Policy Mutes}

\item{\code{next_}}{character URL pointing to the next page (if null, there is no next page)}

\item{\code{previous}}{character URL pointing to the previous page (if null, there is no previous page)}

\item{\code{totalCount}}{integer The total number of items returned by the query}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NotificationPolicyMuteListResponse-new}{\code{NotificationPolicyMuteListResponse$new()}}
\item \href{#method-NotificationPolicyMuteListResponse-validate}{\code{NotificationPolicyMuteListResponse$validate()}}
\item \href{#method-NotificationPolicyMuteListResponse-toJSON}{\code{NotificationPolicyMuteListResponse$toJSON()}}
\item \href{#method-NotificationPolicyMuteListResponse-fromJSON}{\code{NotificationPolicyMuteListResponse$fromJSON()}}
\item \href{#method-NotificationPolicyMuteListResponse-clone}{\code{NotificationPolicyMuteListResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationPolicyMuteListResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationPolicyMuteListResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationPolicyMuteListResponse$new(
  count = NULL,
  data = NULL,
  next_ = NULL,
  previous = NULL,
  totalCount = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{The number of items returned on this page}

\item{\code{data}}{A list of Notification Policy Mutes}

\item{\code{next_}}{URL pointing to the next page (if null, there is no next page)}

\item{\code{previous}}{URL pointing to the previous page (if null, there is no previous page)}

\item{\code{totalCount}}{The total number of items returned by the query}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationPolicyMuteListResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationPolicyMuteListResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationPolicyMuteListResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationPolicyMuteListResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationPolicyMuteListResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationPolicyMuteListResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationPolicyMuteListResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationPolicyMuteListResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationPolicyMuteListResponse$fromJSON(
  NotificationPolicyMuteListResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{NotificationPolicyMuteListResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationPolicyMuteListResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationPolicyMuteListResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationPolicyMuteListResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
