% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notification_policy_update.R
\docType{class}
\name{NotificationPolicyUpdate}
\alias{NotificationPolicyUpdate}
\title{NotificationPolicyUpdate}
\format{
An \code{R6Class} generator object
}
\description{
NotificationPolicyUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{active}}{character \link{optional} Defines if the notification policy is active or not.}

\item{\code{channelId}}{character \link{optional} The id of the notification channel to be used to send the notification.}

\item{\code{eventGroup}}{character \link{optional} The group of the event that trigger the notification.}

\item{\code{eventType}}{character \link{optional} The type of the event that triggers the notification.}

\item{\code{name}}{character \link{optional} The name of the notification policy.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NotificationPolicyUpdate$new()}}
\item \href{#method-validate}{\code{NotificationPolicyUpdate$validate()}}
\item \href{#method-toJSON}{\code{NotificationPolicyUpdate$toJSON()}}
\item \href{#method-fromJSON}{\code{NotificationPolicyUpdate$fromJSON()}}
\item \href{#method-clone}{\code{NotificationPolicyUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationPolicyUpdate$new(
  active = NULL,
  channelId = NULL,
  eventGroup = NULL,
  eventType = NULL,
  name = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{active}}{Defines if the notification policy is active or not.}

\item{\code{channelId}}{The id of the notification channel to be used to send the notification.}

\item{\code{eventGroup}}{The group of the event that trigger the notification.}

\item{\code{eventType}}{The type of the event that triggers the notification.}

\item{\code{name}}{The name of the notification policy.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationPolicyUpdate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationPolicyUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationPolicyUpdate$fromJSON(
  NotificationPolicyUpdateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{NotificationPolicyUpdateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationPolicyUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
