% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o_auth_credentials.R
\docType{class}
\name{OAuthCredentials}
\alias{OAuthCredentials}
\title{OAuthCredentials}
\format{
An \code{R6Class} generator object
}
\description{
OAuthCredentials Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{credentialType}}{character The type of these credentials, 'oauth' here.}

\item{\code{oauthAccessToken}}{character \link{optional} The oauth access token.}

\item{\code{oauthClientId}}{character \link{optional} The oauth client ID.}

\item{\code{oauthClientSecret}}{character \link{optional} The oauth client secret.}

\item{\code{oauthRefreshToken}}{character The oauth refresh token.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OAuthCredentials$new()}}
\item \href{#method-validate}{\code{OAuthCredentials$validate()}}
\item \href{#method-toJSON}{\code{OAuthCredentials$toJSON()}}
\item \href{#method-fromJSON}{\code{OAuthCredentials$fromJSON()}}
\item \href{#method-clone}{\code{OAuthCredentials$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthCredentials$new(
  credentialType = NULL,
  oauthRefreshToken = NULL,
  oauthAccessToken = "null",
  oauthClientId = "null",
  oauthClientSecret = "null",
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{credentialType}}{The type of these credentials, 'oauth' here.}

\item{\code{oauthRefreshToken}}{The oauth refresh token.}

\item{\code{oauthAccessToken}}{The oauth access token.}

\item{\code{oauthClientId}}{The oauth client ID.}

\item{\code{oauthClientSecret}}{The oauth client secret.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthCredentials$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthCredentials$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthCredentials$fromJSON(OAuthCredentialsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{OAuthCredentialsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthCredentials$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
