% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_message_info.R
\docType{class}
\name{OneMessageInfo}
\alias{OneMessageInfo}
\title{OneMessageInfo}
\format{
An \code{R6Class} generator object
}
\description{
OneMessageInfo Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{additionalInfo}}{list( character ) \link{optional} Zero or more text strings for secondary display after user clicks for more information.}

\item{\code{messageLevel}}{character Message severity level.}

\item{\code{messageText}}{character Text for primary display in UI.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OneMessageInfo-new}{\code{OneMessageInfo$new()}}
\item \href{#method-OneMessageInfo-validate}{\code{OneMessageInfo$validate()}}
\item \href{#method-OneMessageInfo-toJSON}{\code{OneMessageInfo$toJSON()}}
\item \href{#method-OneMessageInfo-fromJSON}{\code{OneMessageInfo$fromJSON()}}
\item \href{#method-OneMessageInfo-clone}{\code{OneMessageInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OneMessageInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-OneMessageInfo-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneMessageInfo$new(
  messageLevel = NULL,
  messageText = NULL,
  additionalInfo = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{messageLevel}}{Message severity level.}

\item{\code{messageText}}{Text for primary display in UI.}

\item{\code{additionalInfo}}{Zero or more text strings for secondary display after user clicks for more information.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OneMessageInfo-validate"></a>}}
\if{latex}{\out{\hypertarget{method-OneMessageInfo-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneMessageInfo$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OneMessageInfo-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OneMessageInfo-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneMessageInfo$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OneMessageInfo-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OneMessageInfo-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneMessageInfo$fromJSON(OneMessageInfoJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{OneMessageInfoJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OneMessageInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OneMessageInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneMessageInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
