% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimized_datetime_partitioning_list.R
\docType{class}
\name{OptimizedDatetimePartitioningList}
\alias{OptimizedDatetimePartitioningList}
\title{OptimizedDatetimePartitioningList}
\format{
An \code{R6Class} generator object
}
\description{
OptimizedDatetimePartitioningList Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datetimePartitionColumn}}{character The date column that will be used as a datetime partition column.}

\item{\code{id}}{character The ID of the datetime partitioning.}

\item{\code{partitionData}}{\link{DatetimePartitioningResponse}}

\item{\code{target}}{character The name of the target column.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptimizedDatetimePartitioningList-new}{\code{OptimizedDatetimePartitioningList$new()}}
\item \href{#method-OptimizedDatetimePartitioningList-validate}{\code{OptimizedDatetimePartitioningList$validate()}}
\item \href{#method-OptimizedDatetimePartitioningList-toJSON}{\code{OptimizedDatetimePartitioningList$toJSON()}}
\item \href{#method-OptimizedDatetimePartitioningList-fromJSON}{\code{OptimizedDatetimePartitioningList$fromJSON()}}
\item \href{#method-OptimizedDatetimePartitioningList-clone}{\code{OptimizedDatetimePartitioningList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizedDatetimePartitioningList-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizedDatetimePartitioningList-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizedDatetimePartitioningList$new(
  datetimePartitionColumn = NULL,
  id = NULL,
  partitionData = NULL,
  target = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datetimePartitionColumn}}{The date column that will be used as a datetime partition column.}

\item{\code{id}}{The ID of the datetime partitioning.}

\item{\code{target}}{The name of the target column.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizedDatetimePartitioningList-validate"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizedDatetimePartitioningList-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizedDatetimePartitioningList$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizedDatetimePartitioningList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizedDatetimePartitioningList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizedDatetimePartitioningList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizedDatetimePartitioningList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizedDatetimePartitioningList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizedDatetimePartitioningList$fromJSON(
  OptimizedDatetimePartitioningListJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{OptimizedDatetimePartitioningListJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizedDatetimePartitioningList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizedDatetimePartitioningList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizedDatetimePartitioningList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
