% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimized_datetime_partitioning_list_response.R
\docType{class}
\name{OptimizedDatetimePartitioningListResponse}
\alias{OptimizedDatetimePartitioningListResponse}
\title{OptimizedDatetimePartitioningListResponse}
\format{
An \code{R6Class} generator object
}
\description{
OptimizedDatetimePartitioningListResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer \link{optional} Number of items returned on this page.}

\item{\code{data}}{list( \link{OptimizedDatetimePartitioningList} ) A list of datetime partitionings returned in order of creation.}

\item{\code{next_}}{character URL pointing to the next page (if null, there is no next page).}

\item{\code{previous}}{character URL pointing to the previous page (if null, there is no previous page).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptimizedDatetimePartitioningListResponse-new}{\code{OptimizedDatetimePartitioningListResponse$new()}}
\item \href{#method-OptimizedDatetimePartitioningListResponse-validate}{\code{OptimizedDatetimePartitioningListResponse$validate()}}
\item \href{#method-OptimizedDatetimePartitioningListResponse-toJSON}{\code{OptimizedDatetimePartitioningListResponse$toJSON()}}
\item \href{#method-OptimizedDatetimePartitioningListResponse-fromJSON}{\code{OptimizedDatetimePartitioningListResponse$fromJSON()}}
\item \href{#method-OptimizedDatetimePartitioningListResponse-clone}{\code{OptimizedDatetimePartitioningListResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizedDatetimePartitioningListResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizedDatetimePartitioningListResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizedDatetimePartitioningListResponse$new(
  data = NULL,
  next_ = NULL,
  previous = NULL,
  count = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A list of datetime partitionings returned in order of creation.}

\item{\code{next_}}{URL pointing to the next page (if null, there is no next page).}

\item{\code{previous}}{URL pointing to the previous page (if null, there is no previous page).}

\item{\code{count}}{Number of items returned on this page.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizedDatetimePartitioningListResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizedDatetimePartitioningListResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizedDatetimePartitioningListResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizedDatetimePartitioningListResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizedDatetimePartitioningListResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizedDatetimePartitioningListResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizedDatetimePartitioningListResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizedDatetimePartitioningListResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizedDatetimePartitioningListResponse$fromJSON(
  OptimizedDatetimePartitioningListResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{OptimizedDatetimePartitioningListResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizedDatetimePartitioningListResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizedDatetimePartitioningListResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizedDatetimePartitioningListResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
