% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization_retrieve.R
\docType{class}
\name{OrganizationRetrieve}
\alias{OrganizationRetrieve}
\title{OrganizationRetrieve}
\format{
An \code{R6Class} generator object
}
\description{
OrganizationRetrieve Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{accountPermissions}}{named list( character ) \link{optional} Account permissions available to this organization. This will only be present if the corresponding feature flag is set.}

\item{\code{agreementStatus}}{character The status of the organization's agreement to the terms of DataRobot.}

\item{\code{datasetRefreshJobLimit}}{integer How many enabled dataset refresh jobs are allowed per-dataset for this organization.}

\item{\code{datasetRefreshJobUserLimit}}{integer How many enabled dataset refresh jobs are allowed per-user for this organization.}

\item{\code{defaultUserMaxGpuWorkers}}{integer Maximum number of concurrent GPU workers assigned to a newly created user of this organization.}

\item{\code{defaultUserMaxWorkers}}{integer Maximum number of concurrent workers assigned to a newly created user of this organization.}

\item{\code{desiredCustomModelContainerSize}}{integer \link{optional} The desired custom-model memory size. This will only be present if the corresponding feature flag is set.}

\item{\code{enableSso}}{character \link{optional} If the SSO is enabled for this organization. This will only be present if the corresponding feature flag is set.}

\item{\code{groupsCount}}{integer The number of user groups belonging to this organization.}

\item{\code{id}}{character The organization identifier.}

\item{\code{inactiveUsersCount}}{integer The number of inactive users for this organization.}

\item{\code{maxCustomModelContainerSize}}{integer \link{optional} The maximum memory that might be allocated by the custom-model. This will only be present if the corresponding feature flag is set.}

\item{\code{maxCustomModelReplicasPerDeployment}}{integer \link{optional} A fixed number of replicas that will be set for the given custom-model. This will only be present if the corresponding feature flag is set.}

\item{\code{maxCustomModelReplicasPerDeploymentForBatchPredictions}}{integer \link{optional} The maximum custom inference model replicas per deployment for batch predictions. This will only be present if the corresponding feature flag is set.}

\item{\code{maxDeploymentLimit}}{integer The absolute limit on the number of deployments an organization is allowed to create. A value of zero means unlimited.}

\item{\code{maxEdaWorkers}}{integer Maximum number of EDA workers assigned available to this organization.}

\item{\code{maxGpuWorkers}}{integer Maximum number of concurrent GPU workers available to this organization.}

\item{\code{maxPipelineModuleRuntimes}}{integer Maximum number of Pipeline module runtimes available to this organization.}

\item{\code{maxUploadSize}}{integer \link{optional} Maximum size for file uploads (MB). This will only be present if the corresponding feature flag is set.}

\item{\code{maxUploadSizeCatalog}}{integer \link{optional} Maximum size for catalog file uploads (MB). This will only be present if the corresponding feature flag is set.}

\item{\code{maxWorkers}}{integer Maximum number of concurrent workers available to this organization.}

\item{\code{maximumActiveUsers}}{integer The limit for active users for this organization. A value of zero means unlimited.}

\item{\code{membersCount}}{integer The number of members in this organization.}

\item{\code{mlopsEventStorageRetentionDays}}{integer \link{optional} The number of days to keep MLOps events in storage. Events with older timestamps will be removed.}

\item{\code{name}}{character The name of the organization.}

\item{\code{ormVersion}}{character On-demand Resource Manager (prediction service provisioning) version.}

\item{\code{prepaidDeploymentLimit}}{integer The number of deployments an organization can create under their contract. When it is reached more deployments can be made, but the user will be warned that this will result in a higher billing. A value of zero means unlimited.}

\item{\code{restrictedSharing}}{character \link{optional} Whether sharing is only allowed within the organization. This will only be present if the corresponding feature flag is set.}

\item{\code{snapshotLimit}}{integer The number of snapshots allowed for a dataset for this organization.}

\item{\code{supportEmail}}{character The support email of the organization.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrganizationRetrieve-new}{\code{OrganizationRetrieve$new()}}
\item \href{#method-OrganizationRetrieve-validate}{\code{OrganizationRetrieve$validate()}}
\item \href{#method-OrganizationRetrieve-toJSON}{\code{OrganizationRetrieve$toJSON()}}
\item \href{#method-OrganizationRetrieve-fromJSON}{\code{OrganizationRetrieve$fromJSON()}}
\item \href{#method-OrganizationRetrieve-clone}{\code{OrganizationRetrieve$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrganizationRetrieve-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrganizationRetrieve-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationRetrieve$new(
  agreementStatus = NULL,
  datasetRefreshJobLimit = NULL,
  datasetRefreshJobUserLimit = NULL,
  defaultUserMaxGpuWorkers = NULL,
  defaultUserMaxWorkers = NULL,
  groupsCount = NULL,
  id = NULL,
  inactiveUsersCount = NULL,
  maxDeploymentLimit = NULL,
  maxEdaWorkers = NULL,
  maxGpuWorkers = NULL,
  maxPipelineModuleRuntimes = NULL,
  maxWorkers = NULL,
  maximumActiveUsers = NULL,
  membersCount = NULL,
  name = NULL,
  ormVersion = NULL,
  prepaidDeploymentLimit = NULL,
  snapshotLimit = NULL,
  supportEmail = NULL,
  accountPermissions = NULL,
  desiredCustomModelContainerSize = NULL,
  enableSso = NULL,
  maxCustomModelContainerSize = NULL,
  maxCustomModelReplicasPerDeployment = NULL,
  maxCustomModelReplicasPerDeploymentForBatchPredictions = NULL,
  maxUploadSize = NULL,
  maxUploadSizeCatalog = NULL,
  mlopsEventStorageRetentionDays = 0,
  restrictedSharing = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{agreementStatus}}{The status of the organization's agreement to the terms of DataRobot.}

\item{\code{datasetRefreshJobLimit}}{How many enabled dataset refresh jobs are allowed per-dataset for this organization.}

\item{\code{datasetRefreshJobUserLimit}}{How many enabled dataset refresh jobs are allowed per-user for this organization.}

\item{\code{defaultUserMaxGpuWorkers}}{Maximum number of concurrent GPU workers assigned to a newly created user of this organization.}

\item{\code{defaultUserMaxWorkers}}{Maximum number of concurrent workers assigned to a newly created user of this organization.}

\item{\code{groupsCount}}{The number of user groups belonging to this organization.}

\item{\code{id}}{The organization identifier.}

\item{\code{inactiveUsersCount}}{The number of inactive users for this organization.}

\item{\code{maxDeploymentLimit}}{The absolute limit on the number of deployments an organization is allowed to create. A value of zero means unlimited.}

\item{\code{maxEdaWorkers}}{Maximum number of EDA workers assigned available to this organization.}

\item{\code{maxGpuWorkers}}{Maximum number of concurrent GPU workers available to this organization.}

\item{\code{maxPipelineModuleRuntimes}}{Maximum number of Pipeline module runtimes available to this organization.}

\item{\code{maxWorkers}}{Maximum number of concurrent workers available to this organization.}

\item{\code{maximumActiveUsers}}{The limit for active users for this organization. A value of zero means unlimited.}

\item{\code{membersCount}}{The number of members in this organization.}

\item{\code{name}}{The name of the organization.}

\item{\code{ormVersion}}{On-demand Resource Manager (prediction service provisioning) version.}

\item{\code{prepaidDeploymentLimit}}{The number of deployments an organization can create under their contract. When it is reached more deployments can be made, but the user will be warned that this will result in a higher billing. A value of zero means unlimited.}

\item{\code{snapshotLimit}}{The number of snapshots allowed for a dataset for this organization.}

\item{\code{supportEmail}}{The support email of the organization.}

\item{\code{accountPermissions}}{Account permissions available to this organization. This will only be present if the corresponding feature flag is set.}

\item{\code{desiredCustomModelContainerSize}}{The desired custom-model memory size. This will only be present if the corresponding feature flag is set.}

\item{\code{enableSso}}{If the SSO is enabled for this organization. This will only be present if the corresponding feature flag is set.}

\item{\code{maxCustomModelContainerSize}}{The maximum memory that might be allocated by the custom-model. This will only be present if the corresponding feature flag is set.}

\item{\code{maxCustomModelReplicasPerDeployment}}{A fixed number of replicas that will be set for the given custom-model. This will only be present if the corresponding feature flag is set.}

\item{\code{maxCustomModelReplicasPerDeploymentForBatchPredictions}}{The maximum custom inference model replicas per deployment for batch predictions. This will only be present if the corresponding feature flag is set.}

\item{\code{maxUploadSize}}{Maximum size for file uploads (MB). This will only be present if the corresponding feature flag is set.}

\item{\code{maxUploadSizeCatalog}}{Maximum size for catalog file uploads (MB). This will only be present if the corresponding feature flag is set.}

\item{\code{mlopsEventStorageRetentionDays}}{The number of days to keep MLOps events in storage. Events with older timestamps will be removed.}

\item{\code{restrictedSharing}}{Whether sharing is only allowed within the organization. This will only be present if the corresponding feature flag is set.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrganizationRetrieve-validate"></a>}}
\if{latex}{\out{\hypertarget{method-OrganizationRetrieve-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationRetrieve$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrganizationRetrieve-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrganizationRetrieve-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationRetrieve$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrganizationRetrieve-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrganizationRetrieve-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationRetrieve$fromJSON(OrganizationRetrieveJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{OrganizationRetrieveJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrganizationRetrieve-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrganizationRetrieve-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationRetrieve$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
