% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization_user.R
\docType{class}
\name{OrganizationUser}
\alias{OrganizationUser}
\title{OrganizationUser}
\format{
An \code{R6Class} generator object
}
\description{
OrganizationUser Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{accessRoleId}}{character \link{optional} \link{DEPRECATED - Use accessRoleIds instead} ID of the access role assigned to the user being created}

\item{\code{accessRoleIds}}{list( character ) \link{optional} List of access role ids assigned to the user being created}

\item{\code{create}}{character \link{optional} If the user should be created if not found}

\item{\code{firstName}}{character \link{optional} First name of the user being created}

\item{\code{language}}{character \link{optional} Language of the user being created}

\item{\code{lastName}}{character \link{optional} Last name of the user being created}

\item{\code{orgAdmin}}{character \link{optional} Is this user should be marked as an organizational admin}

\item{\code{password}}{character \link{optional} Password for the user being created. Should be specified if `create` set to `true`}

\item{\code{requireClickthrough}}{character \link{optional} Require the user being created to agree to a clickthrough}

\item{\code{username}}{character The username of the user to add to the organization}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrganizationUser-new}{\code{OrganizationUser$new()}}
\item \href{#method-OrganizationUser-validate}{\code{OrganizationUser$validate()}}
\item \href{#method-OrganizationUser-toJSON}{\code{OrganizationUser$toJSON()}}
\item \href{#method-OrganizationUser-fromJSON}{\code{OrganizationUser$fromJSON()}}
\item \href{#method-OrganizationUser-clone}{\code{OrganizationUser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrganizationUser-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrganizationUser-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUser$new(
  username = NULL,
  accessRoleId = NULL,
  accessRoleIds = NULL,
  create = FALSE,
  firstName = NULL,
  language = NULL,
  lastName = NULL,
  orgAdmin = FALSE,
  password = NULL,
  requireClickthrough = FALSE,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{The username of the user to add to the organization}

\item{\code{accessRoleId}}{\link{DEPRECATED - Use accessRoleIds instead} ID of the access role assigned to the user being created}

\item{\code{accessRoleIds}}{List of access role ids assigned to the user being created}

\item{\code{create}}{If the user should be created if not found}

\item{\code{firstName}}{First name of the user being created}

\item{\code{language}}{Language of the user being created}

\item{\code{lastName}}{Last name of the user being created}

\item{\code{orgAdmin}}{Is this user should be marked as an organizational admin}

\item{\code{password}}{Password for the user being created. Should be specified if `create` set to `true`}

\item{\code{requireClickthrough}}{Require the user being created to agree to a clickthrough}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrganizationUser-validate"></a>}}
\if{latex}{\out{\hypertarget{method-OrganizationUser-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUser$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrganizationUser-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrganizationUser-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUser$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrganizationUser-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OrganizationUser-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUser$fromJSON(OrganizationUserJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{OrganizationUserJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrganizationUser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrganizationUser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
