% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization_user_response.R
\docType{class}
\name{OrganizationUserResponse}
\alias{OrganizationUserResponse}
\title{OrganizationUserResponse}
\format{
An \code{R6Class} generator object
}
\description{
OrganizationUserResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{accessRoleIds}}{list( character ) \link{optional} List of access role ids assigned to the user}

\item{\code{activated}}{character Whether the organization user is activated.}

\item{\code{expirationDate}}{character User expiration date}

\item{\code{firstName}}{character First name of the user being created}

\item{\code{id}}{character The organization user identifier.}

\item{\code{lastName}}{character Last name of the user being created}

\item{\code{maxWorkers}}{integer Maximum number of concurrent workers available to this user.}

\item{\code{orgAdmin}}{character Is this user should be marked as an organizational admin}

\item{\code{organizationId}}{character The organization identifier.}

\item{\code{scheduledForDeletion}}{character \link{optional} Whether the user is scheduled for deletion. Only set when a specific feature flag is configured.}

\item{\code{username}}{character The username of the user to add to the organization}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OrganizationUserResponse$new()}}
\item \href{#method-validate}{\code{OrganizationUserResponse$validate()}}
\item \href{#method-toJSON}{\code{OrganizationUserResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{OrganizationUserResponse$fromJSON()}}
\item \href{#method-clone}{\code{OrganizationUserResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUserResponse$new(
  activated = NULL,
  expirationDate = NULL,
  firstName = NULL,
  id = NULL,
  lastName = NULL,
  maxWorkers = NULL,
  orgAdmin = NULL,
  organizationId = NULL,
  username = NULL,
  accessRoleIds = NULL,
  scheduledForDeletion = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{activated}}{Whether the organization user is activated.}

\item{\code{expirationDate}}{User expiration date}

\item{\code{firstName}}{First name of the user being created}

\item{\code{id}}{The organization user identifier.}

\item{\code{lastName}}{Last name of the user being created}

\item{\code{maxWorkers}}{Maximum number of concurrent workers available to this user.}

\item{\code{orgAdmin}}{Is this user should be marked as an organizational admin}

\item{\code{organizationId}}{The organization identifier.}

\item{\code{username}}{The username of the user to add to the organization}

\item{\code{accessRoleIds}}{List of access role ids assigned to the user}

\item{\code{scheduledForDeletion}}{Whether the user is scheduled for deletion. Only set when a specific feature flag is configured.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUserResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUserResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUserResponse$fromJSON(
  OrganizationUserResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{OrganizationUserResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUserResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
