% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_manual_selection_create_payload.R
\docType{class}
\name{PairwiseManualSelectionCreatePayload}
\alias{PairwiseManualSelectionCreatePayload}
\title{PairwiseManualSelectionCreatePayload}
\format{
An \code{R6Class} generator object
}
\description{
PairwiseManualSelectionCreatePayload Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{columnLabels}}{list( character ) Manually selected column labels.}

\item{\code{featureName}}{character Name of the feature the request is related to}

\item{\code{multilabelInsightsKey}}{character Key for multilabel insights, unique per project, feature, and EDA stage. The most recent key can be retrieved via :http:get:`/api/v2/projects/(projectId)/features/` or :http:get:`/api/v2/projects/(projectId)/features/(featurename:featureName)/`}

\item{\code{name}}{character Name for the set of manually selected labels.}

\item{\code{projectId}}{character ID of the project the request is related to.}

\item{\code{rowLabels}}{list( character ) Manually selected row labels.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PairwiseManualSelectionCreatePayload-new}{\code{PairwiseManualSelectionCreatePayload$new()}}
\item \href{#method-PairwiseManualSelectionCreatePayload-validate}{\code{PairwiseManualSelectionCreatePayload$validate()}}
\item \href{#method-PairwiseManualSelectionCreatePayload-toJSON}{\code{PairwiseManualSelectionCreatePayload$toJSON()}}
\item \href{#method-PairwiseManualSelectionCreatePayload-fromJSON}{\code{PairwiseManualSelectionCreatePayload$fromJSON()}}
\item \href{#method-PairwiseManualSelectionCreatePayload-clone}{\code{PairwiseManualSelectionCreatePayload$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseManualSelectionCreatePayload-new"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseManualSelectionCreatePayload-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseManualSelectionCreatePayload$new(
  columnLabels = NULL,
  featureName = NULL,
  multilabelInsightsKey = NULL,
  name = NULL,
  projectId = NULL,
  rowLabels = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columnLabels}}{Manually selected column labels.}

\item{\code{featureName}}{Name of the feature the request is related to}

\item{\code{multilabelInsightsKey}}{Key for multilabel insights, unique per project, feature, and EDA stage. The most recent key can be retrieved via :http:get:`/api/v2/projects/(projectId)/features/` or :http:get:`/api/v2/projects/(projectId)/features/(featurename:featureName)/`}

\item{\code{name}}{Name for the set of manually selected labels.}

\item{\code{projectId}}{ID of the project the request is related to.}

\item{\code{rowLabels}}{Manually selected row labels.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseManualSelectionCreatePayload-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseManualSelectionCreatePayload-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseManualSelectionCreatePayload$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseManualSelectionCreatePayload-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseManualSelectionCreatePayload-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseManualSelectionCreatePayload$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseManualSelectionCreatePayload-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseManualSelectionCreatePayload-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseManualSelectionCreatePayload$fromJSON(
  PairwiseManualSelectionCreatePayloadJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PairwiseManualSelectionCreatePayloadJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseManualSelectionCreatePayload-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseManualSelectionCreatePayload-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseManualSelectionCreatePayload$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
