% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_manual_selections_retrieve_response.R
\docType{class}
\name{PairwiseManualSelectionsRetrieveResponse}
\alias{PairwiseManualSelectionsRetrieveResponse}
\title{PairwiseManualSelectionsRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
PairwiseManualSelectionsRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{PairwiseManualSelectionCreatedItem} ) List of manually selected labels sets}

\item{\code{featureName}}{character Name of the feature the request is related to}

\item{\code{multilabelInsightsKey}}{character Key for multilabel insights, unique per project, feature, and EDA stage. The most recent key can be retrieved via :http:get:`/api/v2/projects/(projectId)/features/` or :http:get:`/api/v2/projects/(projectId)/features/(featurename:featureName)/`}

\item{\code{projectId}}{character ID of the project the request is related to.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PairwiseManualSelectionsRetrieveResponse-new}{\code{PairwiseManualSelectionsRetrieveResponse$new()}}
\item \href{#method-PairwiseManualSelectionsRetrieveResponse-validate}{\code{PairwiseManualSelectionsRetrieveResponse$validate()}}
\item \href{#method-PairwiseManualSelectionsRetrieveResponse-toJSON}{\code{PairwiseManualSelectionsRetrieveResponse$toJSON()}}
\item \href{#method-PairwiseManualSelectionsRetrieveResponse-fromJSON}{\code{PairwiseManualSelectionsRetrieveResponse$fromJSON()}}
\item \href{#method-PairwiseManualSelectionsRetrieveResponse-clone}{\code{PairwiseManualSelectionsRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseManualSelectionsRetrieveResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseManualSelectionsRetrieveResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseManualSelectionsRetrieveResponse$new(
  data = NULL,
  featureName = NULL,
  multilabelInsightsKey = NULL,
  projectId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{List of manually selected labels sets}

\item{\code{featureName}}{Name of the feature the request is related to}

\item{\code{multilabelInsightsKey}}{Key for multilabel insights, unique per project, feature, and EDA stage. The most recent key can be retrieved via :http:get:`/api/v2/projects/(projectId)/features/` or :http:get:`/api/v2/projects/(projectId)/features/(featurename:featureName)/`}

\item{\code{projectId}}{ID of the project the request is related to.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseManualSelectionsRetrieveResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseManualSelectionsRetrieveResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseManualSelectionsRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseManualSelectionsRetrieveResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseManualSelectionsRetrieveResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseManualSelectionsRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseManualSelectionsRetrieveResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseManualSelectionsRetrieveResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseManualSelectionsRetrieveResponse$fromJSON(
  PairwiseManualSelectionsRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PairwiseManualSelectionsRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairwiseManualSelectionsRetrieveResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PairwiseManualSelectionsRetrieveResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairwiseManualSelectionsRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
