% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_front_response.R
\docType{class}
\name{ParetoFrontResponse}
\alias{ParetoFrontResponse}
\title{ParetoFrontResponse}
\format{
An \code{R6Class} generator object
}
\description{
ParetoFrontResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{errorMetric}}{character The Eureqa error metric identifier used to compute error metrics for this search. Note that Eureqa error metrics do not correspond 1:1 with DataRobot error metrics - the available metrics are not the same, and even equivalent metrics may be computed slightly differently.}

\item{\code{hyperparameters}}{\link{OneOfHyperparametersResponse} The hyperparameters used by this run of the Eureqa blueprint.}

\item{\code{projectId}}{character The project ID of the Eureqa model.}

\item{\code{solutions}}{list( \link{SolutionResponse} ) The Eureqa model solutions.}

\item{\code{targetType}}{character The type of the target variable.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ParetoFrontResponse-new}{\code{ParetoFrontResponse$new()}}
\item \href{#method-ParetoFrontResponse-validate}{\code{ParetoFrontResponse$validate()}}
\item \href{#method-ParetoFrontResponse-toJSON}{\code{ParetoFrontResponse$toJSON()}}
\item \href{#method-ParetoFrontResponse-fromJSON}{\code{ParetoFrontResponse$fromJSON()}}
\item \href{#method-ParetoFrontResponse-clone}{\code{ParetoFrontResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParetoFrontResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ParetoFrontResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParetoFrontResponse$new(
  errorMetric = NULL,
  hyperparameters = NULL,
  projectId = NULL,
  solutions = NULL,
  targetType = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{errorMetric}}{The Eureqa error metric identifier used to compute error metrics for this search. Note that Eureqa error metrics do not correspond 1:1 with DataRobot error metrics - the available metrics are not the same, and even equivalent metrics may be computed slightly differently.}

\item{\code{hyperparameters}}{The hyperparameters used by this run of the Eureqa blueprint.}

\item{\code{projectId}}{The project ID of the Eureqa model.}

\item{\code{solutions}}{The Eureqa model solutions.}

\item{\code{targetType}}{The type of the target variable.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParetoFrontResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ParetoFrontResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParetoFrontResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParetoFrontResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ParetoFrontResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParetoFrontResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParetoFrontResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ParetoFrontResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParetoFrontResponse$fromJSON(ParetoFrontResponseJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ParetoFrontResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParetoFrontResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ParetoFrontResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParetoFrontResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
