% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ParseRFC3339Timestamp}
\alias{ParseRFC3339Timestamp}
\title{ParseRFC3339Timestamp}
\usage{
ParseRFC3339Timestamp(timestampstring)
}
\arguments{
\item{timestampstring}{character. Timestamp in RFC 3339 format.}
}
\value{
The input timestamp as a POSIXt
}
\description{
The DataRobot APIs returns dates in RFC 3339 format.
}
\seealso{
Other API datetime functions: 
\code{\link{RFC3339DateTimeFormat}}
}
\concept{API datetime functions}
