% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioning_single_extended_warning.R
\docType{class}
\name{PartitioningSingleExtendedWarning}
\alias{PartitioningSingleExtendedWarning}
\title{PartitioningSingleExtendedWarning}
\format{
An \code{R6Class} generator object
}
\description{
PartitioningSingleExtendedWarning Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{message}}{character Warning message.}

\item{\code{title}}{character Warning short title.}

\item{\code{type}}{character Warning severity type.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PartitioningSingleExtendedWarning-new}{\code{PartitioningSingleExtendedWarning$new()}}
\item \href{#method-PartitioningSingleExtendedWarning-validate}{\code{PartitioningSingleExtendedWarning$validate()}}
\item \href{#method-PartitioningSingleExtendedWarning-toJSON}{\code{PartitioningSingleExtendedWarning$toJSON()}}
\item \href{#method-PartitioningSingleExtendedWarning-fromJSON}{\code{PartitioningSingleExtendedWarning$fromJSON()}}
\item \href{#method-PartitioningSingleExtendedWarning-clone}{\code{PartitioningSingleExtendedWarning$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitioningSingleExtendedWarning-new"></a>}}
\if{latex}{\out{\hypertarget{method-PartitioningSingleExtendedWarning-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitioningSingleExtendedWarning$new(
  message = NULL,
  title = NULL,
  type = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{Warning message.}

\item{\code{title}}{Warning short title.}

\item{\code{type}}{Warning severity type.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitioningSingleExtendedWarning-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PartitioningSingleExtendedWarning-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitioningSingleExtendedWarning$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitioningSingleExtendedWarning-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PartitioningSingleExtendedWarning-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitioningSingleExtendedWarning$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitioningSingleExtendedWarning-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PartitioningSingleExtendedWarning-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitioningSingleExtendedWarning$fromJSON(
  PartitioningSingleExtendedWarningJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PartitioningSingleExtendedWarningJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitioningSingleExtendedWarning-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PartitioningSingleExtendedWarning-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitioningSingleExtendedWarning$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
