% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioning_warning.R
\docType{class}
\name{PartitioningWarning}
\alias{PartitioningWarning}
\title{PartitioningWarning}
\format{
An \code{R6Class} generator object
}
\description{
PartitioningWarning Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{backtestIndex}}{integer Backtest index. Null if the warning does not correspond to a single backtest.}

\item{\code{partition}}{character Partition name.}

\item{\code{warnings}}{list( character ) A list of strings representing warnings for the specified partition}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PartitioningWarning-new}{\code{PartitioningWarning$new()}}
\item \href{#method-PartitioningWarning-validate}{\code{PartitioningWarning$validate()}}
\item \href{#method-PartitioningWarning-toJSON}{\code{PartitioningWarning$toJSON()}}
\item \href{#method-PartitioningWarning-fromJSON}{\code{PartitioningWarning$fromJSON()}}
\item \href{#method-PartitioningWarning-clone}{\code{PartitioningWarning$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitioningWarning-new"></a>}}
\if{latex}{\out{\hypertarget{method-PartitioningWarning-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitioningWarning$new(
  backtestIndex = NULL,
  partition = NULL,
  warnings = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{backtestIndex}}{Backtest index. Null if the warning does not correspond to a single backtest.}

\item{\code{partition}}{Partition name.}

\item{\code{warnings}}{A list of strings representing warnings for the specified partition}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitioningWarning-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PartitioningWarning-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitioningWarning$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitioningWarning-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PartitioningWarning-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitioningWarning$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitioningWarning-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PartitioningWarning-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitioningWarning$fromJSON(PartitioningWarningJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PartitioningWarningJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitioningWarning-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PartitioningWarning-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitioningWarning$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
