% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{PatchBatchPredictionJobDefinitions}
\alias{PatchBatchPredictionJobDefinitions}
\title{PatchBatchPredictionJobDefinitions}
\usage{
PatchBatchPredictionJobDefinitions(
  columnNamesRemapping,
  timeseriesSettings,
  modelId,
  passthroughColumns,
  outputSettings,
  explanationNumTopClasses,
  enabled,
  predictionWarningEnabled,
  intakeSettings,
  deploymentId,
  predictionInstance,
  csvSettings,
  jobDefinitionId,
  chunkSize,
  explanationClassNames,
  modelPackageId,
  schedule,
  numConcurrent,
  passthroughColumnsSet,
  explanationAlgorithm,
  name,
  thresholdHigh,
  thresholdLow,
  pinnedModelId,
  includeProbabilitiesClasses = c(),
  includePredictionStatus = FALSE,
  skipDriftTracking = FALSE,
  disableRowLevelErrorHandling = FALSE,
  abortOnError = TRUE,
  maxExplanations = 0,
  includeProbabilities = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update Batch Prediction job definition
}
\details{
Update a Batch Prediction job definition

This method invokes \verb{PATCH /batchPredictionJobDefinitions/\{jobDefinitionId\}/} in the DataRobot Public API.
}
\seealso{
The method BatchPredictionJobDefinitionsPatch in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
