% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{PatchCustomModelsVersions}
\alias{PatchCustomModelsVersions}
\title{PatchCustomModelsVersions}
\usage{
PatchCustomModelsVersions(
  requiredMetadata,
  customModelId,
  description,
  requiredMetadataValues,
  customModelVersionId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update custom model version.
}
\details{
Edit metadata of a specific model version.

This method invokes \verb{PATCH /customModels/\{customModelId\}/versions/\{customModelVersionId\}/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsVersionsPatch in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
