% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_dataset.R
\docType{class}
\name{PatchDataset}
\alias{PatchDataset}
\title{PatchDataset}
\format{
An \code{R6Class} generator object
}
\description{
PatchDataset Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{categories}}{\link{OneOfstringarray} \link{optional} An array of strings describing the intended use of the dataset. If any categories were previously specified for the dataset, they will be overwritten.}

\item{\code{name}}{character \link{optional} The new name of the dataset.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PatchDataset-new}{\code{PatchDataset$new()}}
\item \href{#method-PatchDataset-validate}{\code{PatchDataset$validate()}}
\item \href{#method-PatchDataset-toJSON}{\code{PatchDataset$toJSON()}}
\item \href{#method-PatchDataset-fromJSON}{\code{PatchDataset$fromJSON()}}
\item \href{#method-PatchDataset-clone}{\code{PatchDataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PatchDataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-PatchDataset-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PatchDataset$new(categories = NULL, name = NULL, validateParams = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{categories}}{An array of strings describing the intended use of the dataset. If any categories were previously specified for the dataset, they will be overwritten.}

\item{\code{name}}{The new name of the dataset.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PatchDataset-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PatchDataset-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PatchDataset$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PatchDataset-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PatchDataset-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PatchDataset$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PatchDataset-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PatchDataset-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PatchDataset$fromJSON(PatchDatasetJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PatchDatasetJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PatchDataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PatchDataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PatchDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
