% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{PatchDatasets}
\alias{PatchDatasets}
\title{PatchDatasets}
\usage{
PatchDatasets(name, datasetId, categories, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Modify dataset
}
\details{
Modifies the specified dataset.

This method invokes \verb{PATCH /datasets/\{datasetId\}/} in the DataRobot Public API.
}
\seealso{
The method DatasetsPatch in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
