% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{PatchDatasetsFeaturelists}
\alias{PatchDatasetsFeaturelists}
\title{PatchDatasetsFeaturelists}
\usage{
PatchDatasetsFeaturelists(name, datasetId, featurelistId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update dataset featurelist
}
\details{
Modifies the indicated featurelist of the dataset with given ID.

This method invokes \verb{PATCH /datasets/\{datasetId\}/featurelists/\{featurelistId\}/} in the DataRobot Public API.
}
\seealso{
The method DatasetsFeaturelistsPatch in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
