% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{PatchDatasetsRelationships}
\alias{PatchDatasetsRelationships}
\title{PatchDatasetsRelationships}
\usage{
PatchDatasetsRelationships(
  linkedFeatures,
  datasetRelationshipId,
  linkedDatasetId,
  datasetId,
  sourceFeatures,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update dataset relationship.
}
\details{
Update a dataset relationship.

This method invokes \verb{PATCH /datasets/\{datasetId\}/relationships/\{datasetRelationshipId\}/} in the DataRobot Public API.
}
\seealso{
The method DatasetsRelationshipsPatch in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
