% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{PatchDeletedProjects}
\alias{PatchDeletedProjects}
\title{PatchDeletedProjects}
\usage{
PatchDeletedProjects(action, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Recover soft-deleted project
}
\details{
Recover (undelete) soft-deleted project

This method invokes \verb{PATCH /deletedProjects/\{projectId\}/} in the DataRobot Public API.
}
\seealso{
The method DeletedProjectsPatch in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
