% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{PatchDeployments}
\alias{PatchDeployments}
\title{PatchDeployments}
\usage{
PatchDeployments(importance, deploymentId, description, label, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update deployment
}
\details{
Update a deployment's label and description.

This method invokes \verb{PATCH /deployments/\{deploymentId\}/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsPatch in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
