% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{PatchDeploymentsRetrainingPolicies}
\alias{PatchDeploymentsRetrainingPolicies}
\title{PatchDeploymentsRetrainingPolicies}
\usage{
PatchDeploymentsRetrainingPolicies(
  projectOptionsStrategy,
  projectOptions,
  autopilotOptions,
  modelSelectionStrategy,
  retrainingPolicyId,
  deploymentId,
  name,
  description,
  action,
  trigger,
  timeSeriesOptions,
  featureListStrategy,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Endpoint for updating a deployment retraining policy.
}
\details{
Update a deployment retraining policy.

This method invokes \verb{PATCH /deployments/\{deploymentId\}/retrainingPolicies/\{retrainingPolicyId\}/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsRetrainingPoliciesPatch in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
