% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{PatchExternalDataDrivers}
\alias{PatchExternalDataDrivers}
\title{PatchExternalDataDrivers}
\usage{
PatchExternalDataDrivers(
  skipConfigVerification,
  baseNames,
  driverId,
  localJarUrls,
  className,
  configurationId,
  version,
  removeConfig,
  canonicalName,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update properties of an existing JDBC Driver.
}
\details{
Update properties of an existing JDBC driver. To change the canonicalName and className, you must first remove the driver configuration, if it exists, as its properties will otherwise override name changes.

This method invokes \verb{PATCH /externalDataDrivers/\{driverId\}/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataDriversPatch in \link{DataConnectivityApi}, which invokes the same underlying API endpoint.
}
