% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R
\name{PatchImageAugmentationLists}
\alias{PatchImageAugmentationLists}
\title{PatchImageAugmentationLists}
\usage{
PatchImageAugmentationLists(
  transformationProbability,
  initialList,
  featureName,
  augmentationId,
  transformations,
  name,
  inUse,
  numberOfNewImages,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update and existing augmentation list, with passed in values.
}
\details{
Update and existing augmentation list, with passed in values.

This method invokes \verb{PATCH /imageAugmentationLists/\{augmentationId\}/} in the DataRobot Public API.
}
\seealso{
The method ImageAugmentationListsPatch in \link{ImagesApi}, which invokes the same underlying API endpoint.
}
