% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{PatchManyChangeRequestsStatus}
\alias{PatchManyChangeRequestsStatus}
\title{PatchManyChangeRequestsStatus}
\usage{
PatchManyChangeRequestsStatus(changeRequestId, status, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Resolve or Cancel the Change Request.
}
\details{
Resolve or Cancel the Change Request.

This method invokes \verb{PATCH /changeRequests/\{changeRequestId\}/status/} in the DataRobot Public API.
}
\seealso{
The method ChangeRequestsStatusPatchMany in \link{GovernanceApi}, which invokes the same underlying API endpoint.
}
