% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{PatchManyCustomModelsAccessControl}
\alias{PatchManyCustomModelsAccessControl}
\title{PatchManyCustomModelsAccessControl}
\usage{
PatchManyCustomModelsAccessControl(data, customModelId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Grant access or update roles for users on this custom model and appropriate learning data.
}
\details{
Grant access or update roles for users on this custom model and appropriate learning data. Up to 100 user roles may be set in a single request.

This method invokes \verb{PATCH /customModels/\{customModelId\}/accessControl/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsAccessControlPatchMany in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
