% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{PatchManyDatasets}
\alias{PatchManyDatasets}
\title{PatchManyDatasets}
\usage{
PatchManyDatasets(payload, datasetIds, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Execute bulk dataset action
}
\details{
Execute the specified bulk action on multiple datasets.

This method invokes \verb{PATCH /datasets/} in the DataRobot Public API.
}
\seealso{
The method DatasetsPatchMany in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
