% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{PatchManyDatasetsAccessControl}
\alias{PatchManyDatasetsAccessControl}
\title{PatchManyDatasetsAccessControl}
\usage{
PatchManyDatasetsAccessControl(
  data,
  datasetId,
  applyGrantToLinkedObjects = FALSE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Modify dataset access
}
\details{
Grant access to the dataset at the specified role level, or remove access to the dataset.

This method invokes \verb{PATCH /datasets/\{datasetId\}/accessControl/} in the DataRobot Public API.
}
\seealso{
The method DatasetsAccessControlPatchMany in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
