% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{PatchManyDatasetsDeleted}
\alias{PatchManyDatasetsDeleted}
\title{PatchManyDatasetsDeleted}
\usage{
PatchManyDatasetsDeleted(datasetId, body, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Recover deleted dataset
}
\details{
Recover the dataset item with given datasetId from deleted.

This method invokes \verb{PATCH /datasets/\{datasetId\}/deleted/} in the DataRobot Public API.
}
\seealso{
The method DatasetsDeletedPatchMany in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
