% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{PatchManyDatasetsSharedRoles}
\alias{PatchManyDatasetsSharedRoles}
\title{PatchManyDatasetsSharedRoles}
\usage{
PatchManyDatasetsSharedRoles(
  roles,
  datasetId,
  operation,
  applyGrantToLinkedObjects = FALSE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Modify dataset shared roles
}
\details{
Grant access, remove access or update roles for organizations, groups or users on this dataset. Up to 100 roles may be set per array in a single request.

This method invokes \verb{PATCH /datasets/\{datasetId\}/sharedRoles/} in the DataRobot Public API.
}
\seealso{
The method DatasetsSharedRolesPatchMany in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
