% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{PatchManyDatasetsVersionsDeleted}
\alias{PatchManyDatasetsVersionsDeleted}
\title{PatchManyDatasetsVersionsDeleted}
\usage{
PatchManyDatasetsVersionsDeleted(datasetId, body, datasetVersionId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Recover deleted dataset version
}
\details{
Recover the dataset version item with given datasetId and datasetVersionId from deleted.

This method invokes \verb{PATCH /datasets/\{datasetId\}/versions/\{datasetVersionId\}/deleted/} in the DataRobot Public API.
}
\seealso{
The method DatasetsVersionsDeletedPatchMany in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
