% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{PatchManyDeletedDeployments}
\alias{PatchManyDeletedDeployments}
\title{PatchManyDeletedDeployments}
\usage{
PatchManyDeletedDeployments(action, deploymentIds, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Erase deleted deployments
}
\details{
Permanently erase data for deleted deployments. Only available as part of an on-premise or private/hybrid cloud deployment. Requires a CAN_DELETE_APP_PROJECTS permission to execute.

This method invokes \verb{PATCH /deletedDeployments/} in the DataRobot Public API.
}
\seealso{
The method DeletedDeploymentsPatchMany in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
