% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{PatchManyDeploymentsModelSecondaryDatasetConfiguration}
\alias{PatchManyDeploymentsModelSecondaryDatasetConfiguration}
\title{PatchManyDeploymentsModelSecondaryDatasetConfiguration}
\usage{
PatchManyDeploymentsModelSecondaryDatasetConfiguration(
  secondaryDatasetConfigId,
  credentialsIds,
  deploymentId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update the secondary datasets configuration for the deployed model.
}
\details{
Update the secondary datasets configuration used by the deployed feature discovery model.

This method invokes \verb{PATCH /deployments/\{deploymentId\}/model/secondaryDatasetConfiguration/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsModelSecondaryDatasetConfigurationPatchMany in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
