% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{PatchManyDeploymentsRetrainingSettings}
\alias{PatchManyDeploymentsRetrainingSettings}
\title{PatchManyDeploymentsRetrainingSettings}
\usage{
PatchManyDeploymentsRetrainingSettings(
  deploymentId,
  datasetId,
  credentialId,
  predictionEnvironmentId,
  retrainingUserId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Endpoint for updating deployment retraining settings.
}
\details{
Update deployment retraining settings.

This method invokes \verb{PATCH /deployments/\{deploymentId\}/retrainingSettings/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsRetrainingSettingsPatchMany in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
