% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{PatchManyExternalDataSourcesAccessControl}
\alias{PatchManyExternalDataSourcesAccessControl}
\title{PatchManyExternalDataSourcesAccessControl}
\usage{
PatchManyExternalDataSourcesAccessControl(dataSourceId, data, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update data source's access controls
}
\details{
Set roles for users on this data source. Note that when granting access to a data source, access to the corresponding data store as a \"CONSUMER\" will also be granted.

This method invokes \verb{PATCH /externalDataSources/\{dataSourceId\}/accessControl/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataSourcesAccessControlPatchMany in \link{DataConnectivityApi}, which invokes the same underlying API endpoint.
}
