% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{PatchManyExternalDataStoresAccessControl}
\alias{PatchManyExternalDataStoresAccessControl}
\title{PatchManyExternalDataStoresAccessControl}
\usage{
PatchManyExternalDataStoresAccessControl(data, dataStoreId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update data store's controls
}
\details{
Set roles for users on this data store.

This method invokes \verb{PATCH /externalDataStores/\{dataStoreId\}/accessControl/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataStoresAccessControlPatchMany in \link{DataConnectivityApi}, which invokes the same underlying API endpoint.
}
