% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{PatchManyProjectsAccessControl}
\alias{PatchManyProjectsAccessControl}
\title{PatchManyProjectsAccessControl}
\usage{
PatchManyProjectsAccessControl(
  data,
  projectId,
  sendNotification = TRUE,
  includeFeatureDiscoveryEntities = FALSE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update project's access controls
}
\details{
Set roles for users on this project.

This method invokes \verb{PATCH /projects/\{projectId\}/accessControl/} in the DataRobot Public API.
}
\seealso{
The method AccessControlPatchMany in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
