% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{PatchManyUseCasesSharedRoles}
\alias{PatchManyUseCasesSharedRoles}
\title{PatchManyUseCasesSharedRoles}
\usage{
PatchManyUseCasesSharedRoles(useCaseId, roles, operation, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Share a use case with a user, group or organization.
}
\details{
Share a use case with a user, group or organization.

This method invokes \verb{PATCH /useCases/\{useCaseId\}/sharedRoles/} in the DataRobot Public API.
}
\seealso{
The method UseCasesSharedRolesPatchMany in \link{UseCaseApi}, which invokes the same underlying API endpoint.
}
