% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{PatchManyUserNotifications}
\alias{PatchManyUserNotifications}
\title{PatchManyUserNotifications}
\usage{
PatchManyUserNotifications(...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Mark all as read
}
\details{
Mark all associated notifications with the user as read.

This method invokes \verb{PATCH /userNotifications/} in the DataRobot Public API.
}
\seealso{
The method UserNotificationsPatchMany in \link{NotificationsApi}, which invokes the same underlying API endpoint.
}
