% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{PatchMultilabelInsightsPairwiseManualSelections}
\alias{PatchMultilabelInsightsPairwiseManualSelections}
\title{PatchMultilabelInsightsPairwiseManualSelections}
\usage{
PatchMultilabelInsightsPairwiseManualSelections(
  name,
  manualSelectionListId,
  multilabelInsightsKey,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update label list's name.
}
\details{
Update label list's name.

This method invokes \verb{PATCH /multilabelInsights/\{multilabelInsightsKey\}/pairwiseManualSelections/\{manualSelectionListId\}/} in the DataRobot Public API.
}
\seealso{
The method MultilabelInsightsPairwiseManualSelectionsPatch in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
