% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{PatchScheduledJobs}
\alias{PatchScheduledJobs}
\title{PatchScheduledJobs}
\usage{
PatchScheduledJobs(jobId, status, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Run or stop a previously created scheduled integration job
}
\details{
Run or stop a previously created scheduled integration job

This method invokes \verb{PATCH /scheduledJobs/\{jobId\}/} in the DataRobot Public API.
}
\seealso{
The method ScheduledJobsPatch in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
