% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R,
%   R/blueprints_apiWrapper.R
\name{PatchUserBlueprints}
\alias{PatchUserBlueprints}
\title{PatchUserBlueprints}
\usage{
PatchUserBlueprints(
  blueprint,
  userBlueprintId,
  description,
  modelType,
  projectId,
  isInplaceEditor = FALSE,
  getDynamicLabels = FALSE,
  decompressedBlueprint = FALSE,
  saveToCatalog = TRUE,
  ...
)

PatchUserBlueprints(
  blueprint,
  userBlueprintId,
  description,
  modelType,
  projectId,
  isInplaceEditor = FALSE,
  getDynamicLabels = FALSE,
  decompressedBlueprint = FALSE,
  saveToCatalog = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update a user blueprint.

Update a user blueprint.
}
\details{
Update a user blueprint

This method invokes \verb{PATCH /userBlueprints/\{userBlueprintId\}/} in the DataRobot Public API.

Update a user blueprint

This method invokes \verb{PATCH /userBlueprints/\{userBlueprintId\}/} in the DataRobot Public API.
}
\seealso{
The method UserBlueprintsPatch in \link{AiCatalogApi}, which invokes the same underlying API endpoint.

The method UserBlueprintsPatch in \link{BlueprintsApi}, which invokes the same underlying API endpoint.
}
