% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{PatchUserNotifications}
\alias{PatchUserNotifications}
\title{PatchUserNotifications}
\usage{
PatchUserNotifications(userNotificationId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Mark as read
}
\details{
Mark one associated notification with the user as read.

This method invokes \verb{PATCH /userNotifications/\{userNotificationId\}/} in the DataRobot Public API.
}
\seealso{
The method UserNotificationsPatch in \link{NotificationsApi}, which invokes the same underlying API endpoint.
}
