% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{PauseQueue}
\alias{PauseQueue}
\title{PauseQueue}
\usage{
PauseQueue(project, command = NULL, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Pause or unpause Autopilot
}
\details{
Pause or unpause the autopilot for a project.

This method invokes \verb{POST /projects/\{projectId\}/autopilot/} in the DataRobot Public API.
}
\seealso{
The method AutopilotCreate in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
