% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoff_matrices_response.R
\docType{class}
\name{PayoffMatricesResponse}
\alias{PayoffMatricesResponse}
\title{PayoffMatricesResponse}
\format{
An \code{R6Class} generator object
}
\description{
PayoffMatricesResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{falseNegativeValue}}{numeric Payoff value for false negatives used in profit curve calculation.}

\item{\code{falsePositiveValue}}{numeric Payoff value for false positives used in profit curve calculation.}

\item{\code{id}}{character ObjectId of the payoff matrix.}

\item{\code{name}}{character Label for the payoff matrix.}

\item{\code{projectId}}{character ID of the project associated with the payoff matrix.}

\item{\code{trueNegativeValue}}{numeric Payoff value for true negatives used in profit curve calculation.}

\item{\code{truePositiveValue}}{numeric Payoff value for true positives used in profit curve calculation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PayoffMatricesResponse-new}{\code{PayoffMatricesResponse$new()}}
\item \href{#method-PayoffMatricesResponse-validate}{\code{PayoffMatricesResponse$validate()}}
\item \href{#method-PayoffMatricesResponse-toJSON}{\code{PayoffMatricesResponse$toJSON()}}
\item \href{#method-PayoffMatricesResponse-fromJSON}{\code{PayoffMatricesResponse$fromJSON()}}
\item \href{#method-PayoffMatricesResponse-clone}{\code{PayoffMatricesResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PayoffMatricesResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-PayoffMatricesResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PayoffMatricesResponse$new(
  falseNegativeValue = NULL,
  falsePositiveValue = NULL,
  id = NULL,
  name = NULL,
  projectId = NULL,
  trueNegativeValue = NULL,
  truePositiveValue = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{falseNegativeValue}}{Payoff value for false negatives used in profit curve calculation.}

\item{\code{falsePositiveValue}}{Payoff value for false positives used in profit curve calculation.}

\item{\code{id}}{ObjectId of the payoff matrix.}

\item{\code{name}}{Label for the payoff matrix.}

\item{\code{projectId}}{ID of the project associated with the payoff matrix.}

\item{\code{trueNegativeValue}}{Payoff value for true negatives used in profit curve calculation.}

\item{\code{truePositiveValue}}{Payoff value for true positives used in profit curve calculation.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PayoffMatricesResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PayoffMatricesResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PayoffMatricesResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PayoffMatricesResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PayoffMatricesResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PayoffMatricesResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PayoffMatricesResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PayoffMatricesResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PayoffMatricesResponse$fromJSON(
  PayoffMatricesResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PayoffMatricesResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PayoffMatricesResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PayoffMatricesResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PayoffMatricesResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
