% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_cluster_text.R
\docType{class}
\name{PerClusterText}
\alias{PerClusterText}
\title{PerClusterText}
\format{
An \code{R6Class} generator object
}
\description{
PerClusterText Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{clusterName}}{character Cluster name.}

\item{\code{missingRowsPercent}}{numeric \link{optional} A percentage of all rows that have a missing value for this feature.}

\item{\code{perValueStatistics}}{list( \link{PerValueStatisticTextListItem} ) Statistic value for feature values in all data or a cluster.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PerClusterText-new}{\code{PerClusterText$new()}}
\item \href{#method-PerClusterText-validate}{\code{PerClusterText$validate()}}
\item \href{#method-PerClusterText-toJSON}{\code{PerClusterText$toJSON()}}
\item \href{#method-PerClusterText-fromJSON}{\code{PerClusterText$fromJSON()}}
\item \href{#method-PerClusterText-clone}{\code{PerClusterText$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerClusterText-new"></a>}}
\if{latex}{\out{\hypertarget{method-PerClusterText-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerClusterText$new(
  clusterName = NULL,
  perValueStatistics = NULL,
  missingRowsPercent = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clusterName}}{Cluster name.}

\item{\code{perValueStatistics}}{Statistic value for feature values in all data or a cluster.}

\item{\code{missingRowsPercent}}{A percentage of all rows that have a missing value for this feature.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerClusterText-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PerClusterText-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerClusterText$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerClusterText-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PerClusterText-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerClusterText$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerClusterText-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PerClusterText-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerClusterText$fromJSON(PerClusterTextJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PerClusterTextJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerClusterText-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PerClusterText-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerClusterText$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
