% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_feature_missing_report.R
\docType{class}
\name{PerFeatureMissingReport}
\alias{PerFeatureMissingReport}
\title{PerFeatureMissingReport}
\format{
An \code{R6Class} generator object
}
\description{
PerFeatureMissingReport Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{feature}}{character The name of the feature}

\item{\code{missingCount}}{integer The number of missing values in the training data}

\item{\code{missingPercentage}}{numeric The percentage of missing values in the training data}

\item{\code{tasks}}{named list( \link{PerFeatureTaskMissingReport} ) Information on individual tasks of the model which were used to process the feature. The names of properties will be task ids (which correspond to the ids used in the blueprint chart endpoints like :http:get:`/api/v2/projects/(projectId)/blueprints/(blueprintId)/blueprintChart/`) The corresponding value for each task will be of the form `task` described.}

\item{\code{type}}{character The type of the feature, e.g. `Categorical` or `Numeric`}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PerFeatureMissingReport$new()}}
\item \href{#method-validate}{\code{PerFeatureMissingReport$validate()}}
\item \href{#method-toJSON}{\code{PerFeatureMissingReport$toJSON()}}
\item \href{#method-fromJSON}{\code{PerFeatureMissingReport$fromJSON()}}
\item \href{#method-clone}{\code{PerFeatureMissingReport$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerFeatureMissingReport$new(
  feature = NULL,
  missingCount = NULL,
  missingPercentage = NULL,
  tasks = NULL,
  type = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{The name of the feature}

\item{\code{missingCount}}{The number of missing values in the training data}

\item{\code{missingPercentage}}{The percentage of missing values in the training data}

\item{\code{tasks}}{Information on individual tasks of the model which were used to process the feature. The names of properties will be task ids (which correspond to the ids used in the blueprint chart endpoints like :http:get:`/api/v2/projects/(projectId)/blueprints/(blueprintId)/blueprintChart/`) The corresponding value for each task will be of the form `task` described.}

\item{\code{type}}{The type of the feature, e.g. `Categorical` or `Numeric`}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerFeatureMissingReport$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerFeatureMissingReport$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerFeatureMissingReport$fromJSON(
  PerFeatureMissingReportJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PerFeatureMissingReportJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerFeatureMissingReport$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
