% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_feature_task_missing_report.R
\docType{class}
\name{PerFeatureTaskMissingReport}
\alias{PerFeatureTaskMissingReport}
\title{PerFeatureTaskMissingReport}
\format{
An \code{R6Class} generator object
}
\description{
PerFeatureTaskMissingReport Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{descriptions}}{list( character ) Human readable aggregated information about how the task handles missing values. The following descriptions may be present: what value is imputed for missing values, whether the feature being missing is treated as a feature by the task, whether missing values are treated as infrequent values, whether infrequent values are treated as missing values, and whether missing values are ignored.}

\item{\code{name}}{character Task name, e.g. 'Ordinal encoding of categorical variables'}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PerFeatureTaskMissingReport-new}{\code{PerFeatureTaskMissingReport$new()}}
\item \href{#method-PerFeatureTaskMissingReport-validate}{\code{PerFeatureTaskMissingReport$validate()}}
\item \href{#method-PerFeatureTaskMissingReport-toJSON}{\code{PerFeatureTaskMissingReport$toJSON()}}
\item \href{#method-PerFeatureTaskMissingReport-fromJSON}{\code{PerFeatureTaskMissingReport$fromJSON()}}
\item \href{#method-PerFeatureTaskMissingReport-clone}{\code{PerFeatureTaskMissingReport$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerFeatureTaskMissingReport-new"></a>}}
\if{latex}{\out{\hypertarget{method-PerFeatureTaskMissingReport-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerFeatureTaskMissingReport$new(
  descriptions = NULL,
  name = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{descriptions}}{Human readable aggregated information about how the task handles missing values. The following descriptions may be present: what value is imputed for missing values, whether the feature being missing is treated as a feature by the task, whether missing values are treated as infrequent values, whether infrequent values are treated as missing values, and whether missing values are ignored.}

\item{\code{name}}{Task name, e.g. 'Ordinal encoding of categorical variables'}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerFeatureTaskMissingReport-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PerFeatureTaskMissingReport-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerFeatureTaskMissingReport$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerFeatureTaskMissingReport-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PerFeatureTaskMissingReport-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerFeatureTaskMissingReport$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerFeatureTaskMissingReport-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PerFeatureTaskMissingReport-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerFeatureTaskMissingReport$fromJSON(
  PerFeatureTaskMissingReportJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PerFeatureTaskMissingReportJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerFeatureTaskMissingReport-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PerFeatureTaskMissingReport-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerFeatureTaskMissingReport$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
