% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_ngram_text_explanations.R
\docType{class}
\name{PerNgramTextExplanations}
\alias{PerNgramTextExplanations}
\title{PerNgramTextExplanations}
\format{
An \code{R6Class} generator object
}
\description{
PerNgramTextExplanations Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{isUnknown}}{character Whether the ngram is identifiable by the blueprint or not.}

\item{\code{ngrams}}{list( \link{PredictionExplanationsPredictionValues} ) List of JSON objects with the ngram starting index, ngram ending index and unknown ngram information.}

\item{\code{qualitativateStrength}}{character A human-readable description of how strongly these ngrams's affected the prediction(e.g. '+++', '--', '+', '<+', '<-').}

\item{\code{strength}}{numeric The amount these ngrams's affected the prediction.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PerNgramTextExplanations-new}{\code{PerNgramTextExplanations$new()}}
\item \href{#method-PerNgramTextExplanations-validate}{\code{PerNgramTextExplanations$validate()}}
\item \href{#method-PerNgramTextExplanations-toJSON}{\code{PerNgramTextExplanations$toJSON()}}
\item \href{#method-PerNgramTextExplanations-fromJSON}{\code{PerNgramTextExplanations$fromJSON()}}
\item \href{#method-PerNgramTextExplanations-clone}{\code{PerNgramTextExplanations$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerNgramTextExplanations-new"></a>}}
\if{latex}{\out{\hypertarget{method-PerNgramTextExplanations-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerNgramTextExplanations$new(
  isUnknown = NULL,
  ngrams = NULL,
  qualitativateStrength = NULL,
  strength = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{isUnknown}}{Whether the ngram is identifiable by the blueprint or not.}

\item{\code{ngrams}}{List of JSON objects with the ngram starting index, ngram ending index and unknown ngram information.}

\item{\code{qualitativateStrength}}{A human-readable description of how strongly these ngrams's affected the prediction(e.g. '+++', '--', '+', '<+', '<-').}

\item{\code{strength}}{The amount these ngrams's affected the prediction.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerNgramTextExplanations-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PerNgramTextExplanations-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerNgramTextExplanations$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerNgramTextExplanations-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PerNgramTextExplanations-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerNgramTextExplanations$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerNgramTextExplanations-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PerNgramTextExplanations-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerNgramTextExplanations$fromJSON(
  PerNgramTextExplanationsJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PerNgramTextExplanationsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerNgramTextExplanations-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PerNgramTextExplanations-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerNgramTextExplanations$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
