% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_response.R
\docType{class}
\name{PlotDataResponse}
\alias{PlotDataResponse}
\title{PlotDataResponse}
\format{
An \code{R6Class} generator object
}
\description{
PlotDataResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual}}{numeric The actual value of the target variable for the specified row.}

\item{\code{predicted}}{numeric The predicted value of the target by the solution for the specified row.}

\item{\code{row}}{integer The row number from the raw source data. Used as the X axis for the plot when rendered in the web application.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PlotDataResponse-new}{\code{PlotDataResponse$new()}}
\item \href{#method-PlotDataResponse-validate}{\code{PlotDataResponse$validate()}}
\item \href{#method-PlotDataResponse-toJSON}{\code{PlotDataResponse$toJSON()}}
\item \href{#method-PlotDataResponse-fromJSON}{\code{PlotDataResponse$fromJSON()}}
\item \href{#method-PlotDataResponse-clone}{\code{PlotDataResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataResponse$new(
  actual = NULL,
  predicted = NULL,
  row = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{actual}}{The actual value of the target variable for the specified row.}

\item{\code{predicted}}{The predicted value of the target by the solution for the specified row.}

\item{\code{row}}{The row number from the raw source data. Used as the X axis for the plot when rendered in the web application.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataResponse$fromJSON(PlotDataResponseJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PlotDataResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
