% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pps_image_metadata_response.R
\docType{class}
\name{PPSImageMetadataResponse}
\alias{PPSImageMetadataResponse}
\title{PPSImageMetadataResponse}
\format{
An \code{R6Class} generator object
}
\description{
PPSImageMetadataResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{baseImageId}}{character Internal base image entity id for troubleshooting purposes}

\item{\code{created}}{character ISO formatted image upload date}

\item{\code{datarobotRuntimeImageTag}}{character \link{optional} For internal use only.}

\item{\code{dockerImageId}}{character A Docker image id (immutable, content-based) hash associated with the given image}

\item{\code{filename}}{character The name of the file when the download requested}

\item{\code{hash}}{character Hash of the image content, supposed to be used for verifying content after the download. The algorithm used for hashing is specified in `hashAlgorithm` field. Note that hash is calculated over compressed image data.}

\item{\code{hashAlgorithm}}{character An algorithm name used for calculating content `hash`}

\item{\code{imageSize}}{integer Size in bytes of the compressed PPS image data}

\item{\code{shortDockerImageId}}{character A 12-chars shortened version of the `dockerImageId` as shown in 'docker images' command line command output}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PPSImageMetadataResponse-new}{\code{PPSImageMetadataResponse$new()}}
\item \href{#method-PPSImageMetadataResponse-validate}{\code{PPSImageMetadataResponse$validate()}}
\item \href{#method-PPSImageMetadataResponse-toJSON}{\code{PPSImageMetadataResponse$toJSON()}}
\item \href{#method-PPSImageMetadataResponse-fromJSON}{\code{PPSImageMetadataResponse$fromJSON()}}
\item \href{#method-PPSImageMetadataResponse-clone}{\code{PPSImageMetadataResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PPSImageMetadataResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-PPSImageMetadataResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PPSImageMetadataResponse$new(
  baseImageId = NULL,
  created = NULL,
  dockerImageId = NULL,
  filename = NULL,
  hash = NULL,
  hashAlgorithm = NULL,
  imageSize = NULL,
  shortDockerImageId = NULL,
  datarobotRuntimeImageTag = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{baseImageId}}{Internal base image entity id for troubleshooting purposes}

\item{\code{created}}{ISO formatted image upload date}

\item{\code{dockerImageId}}{A Docker image id (immutable, content-based) hash associated with the given image}

\item{\code{filename}}{The name of the file when the download requested}

\item{\code{hash}}{Hash of the image content, supposed to be used for verifying content after the download. The algorithm used for hashing is specified in `hashAlgorithm` field. Note that hash is calculated over compressed image data.}

\item{\code{hashAlgorithm}}{An algorithm name used for calculating content `hash`}

\item{\code{imageSize}}{Size in bytes of the compressed PPS image data}

\item{\code{shortDockerImageId}}{A 12-chars shortened version of the `dockerImageId` as shown in 'docker images' command line command output}

\item{\code{datarobotRuntimeImageTag}}{For internal use only.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PPSImageMetadataResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PPSImageMetadataResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PPSImageMetadataResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PPSImageMetadataResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PPSImageMetadataResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PPSImageMetadataResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PPSImageMetadataResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PPSImageMetadataResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PPSImageMetadataResponse$fromJSON(
  PPSImageMetadataResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PPSImageMetadataResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PPSImageMetadataResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PPSImageMetadataResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PPSImageMetadataResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
