% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicted_vs_actual.R
\docType{class}
\name{PredictedVsActual}
\alias{PredictedVsActual}
\title{PredictedVsActual}
\format{
An \code{R6Class} generator object
}
\description{
PredictedVsActual Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{PredictedVsActualData} ) Predicted versus actual results.}

\item{\code{isCapped}}{character Indicates whether the data for computation is capped.}

\item{\code{logScaledData}}{list( \link{PredictedVsActualData} ) Predicted versus actual results in a log scale.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictedVsActual-new}{\code{PredictedVsActual$new()}}
\item \href{#method-PredictedVsActual-validate}{\code{PredictedVsActual$validate()}}
\item \href{#method-PredictedVsActual-toJSON}{\code{PredictedVsActual$toJSON()}}
\item \href{#method-PredictedVsActual-fromJSON}{\code{PredictedVsActual$fromJSON()}}
\item \href{#method-PredictedVsActual-clone}{\code{PredictedVsActual$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictedVsActual-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictedVsActual-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictedVsActual$new(
  data = NULL,
  isCapped = NULL,
  logScaledData = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Predicted versus actual results.}

\item{\code{isCapped}}{Indicates whether the data for computation is capped.}

\item{\code{logScaledData}}{Predicted versus actual results in a log scale.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictedVsActual-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PredictedVsActual-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictedVsActual$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictedVsActual-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictedVsActual-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictedVsActual$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictedVsActual-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictedVsActual-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictedVsActual$fromJSON(PredictedVsActualJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictedVsActualJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictedVsActual-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictedVsActual-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictedVsActual$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
