% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_dataset_retrieve_response.R
\docType{class}
\name{PredictionDatasetRetrieveResponse}
\alias{PredictionDatasetRetrieveResponse}
\title{PredictionDatasetRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
PredictionDatasetRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actualValueColumn}}{character \link{optional} Optional, only available for unsupervised projects, in case dataset was uploaded with actual value column specified. Name of the column which will be used to calculate the classification metrics and insights.}

\item{\code{catalogId}}{character The ID of the AI catalog entry used to create the prediction, dataset or None if not created from the AI catalog.}

\item{\code{catalogVersionId}}{character The ID of the AI catalog version used to create the prediction dataset, or None if not created from the AI catalog.}

\item{\code{containsTargetValues}}{character \link{optional} If True, dataset contains target values and can be used to calculate the classification metrics and insights. Only applies for supervised projects.}

\item{\code{created}}{character The date string of when the dataset was created, of the format`YYYY-mm-ddTHH:MM:SS.ssssssZ`, like ``2016-06-09T11:32:34.170338Z``.}

\item{\code{dataEndDate}}{character \link{optional} Only available for time series projects, a date string representing the maximum primary date of the prediction dataset.}

\item{\code{dataQualityWarnings}}{\link{DataQualityWarningsRecord}}

\item{\code{dataStartDate}}{character \link{optional} Only available for time series projects, a date string representing the minimum primary date of the prediction dataset.}

\item{\code{detectedActualValueColumns}}{list( \link{ActualValueColumnInfo} ) \link{optional} Only available for unsupervised projects, a list of detected `actualValueColumnInfo` objects which can be used to calculate the classification metrics and insights.}

\item{\code{forecastPoint}}{character The date string of the forecastPoint of this prediction dataset. Only non-null for time series projects.}

\item{\code{forecastPointRange}}{list( character ) \link{optional} Only available for time series projects, the start and end of the range of dates available for use as the forecast point, detected based on the uploaded prediction dataset.}

\item{\code{id}}{character The ID of this dataset.}

\item{\code{maxForecastDate}}{character \link{optional} Only available for time series projects, a date string representing the maximum forecast date of this prediction dataset.}

\item{\code{name}}{character The name of the dataset when it was uploaded.}

\item{\code{numColumns}}{integer The number of columns in this dataset.}

\item{\code{numRows}}{integer The number of rows in this dataset.}

\item{\code{predictionsEndDate}}{character The date string of the prediction end date of this prediction dataset. Used for bulk predictions. Note that this parameter is for generating historical predictions using the training data. Only non-null for time series projects.}

\item{\code{predictionsStartDate}}{character The date string of the prediction start date of this prediction dataset. Used for bulk predictions. Note that this parameter is for generating historical predictions using the training data. Only non-null for time series projects.}

\item{\code{projectId}}{character The project ID that owns this dataset.}

\item{\code{secondaryDatasetsConfigId}}{character \link{optional} Only available for Feature discovery projects. Id of the secondary dataset config used by the dataset for the prediction.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictionDatasetRetrieveResponse-new}{\code{PredictionDatasetRetrieveResponse$new()}}
\item \href{#method-PredictionDatasetRetrieveResponse-validate}{\code{PredictionDatasetRetrieveResponse$validate()}}
\item \href{#method-PredictionDatasetRetrieveResponse-toJSON}{\code{PredictionDatasetRetrieveResponse$toJSON()}}
\item \href{#method-PredictionDatasetRetrieveResponse-fromJSON}{\code{PredictionDatasetRetrieveResponse$fromJSON()}}
\item \href{#method-PredictionDatasetRetrieveResponse-clone}{\code{PredictionDatasetRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionDatasetRetrieveResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionDatasetRetrieveResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionDatasetRetrieveResponse$new(
  catalogId = NULL,
  catalogVersionId = NULL,
  created = NULL,
  dataQualityWarnings = NULL,
  forecastPoint = NULL,
  id = NULL,
  name = NULL,
  numColumns = NULL,
  numRows = NULL,
  predictionsEndDate = NULL,
  predictionsStartDate = NULL,
  projectId = NULL,
  actualValueColumn = NULL,
  containsTargetValues = NULL,
  dataEndDate = NULL,
  dataStartDate = NULL,
  detectedActualValueColumns = NULL,
  forecastPointRange = NULL,
  maxForecastDate = NULL,
  secondaryDatasetsConfigId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalogId}}{The ID of the AI catalog entry used to create the prediction, dataset or None if not created from the AI catalog.}

\item{\code{catalogVersionId}}{The ID of the AI catalog version used to create the prediction dataset, or None if not created from the AI catalog.}

\item{\code{created}}{The date string of when the dataset was created, of the format`YYYY-mm-ddTHH:MM:SS.ssssssZ`, like ``2016-06-09T11:32:34.170338Z``.}

\item{\code{forecastPoint}}{The date string of the forecastPoint of this prediction dataset. Only non-null for time series projects.}

\item{\code{id}}{The ID of this dataset.}

\item{\code{name}}{The name of the dataset when it was uploaded.}

\item{\code{numColumns}}{The number of columns in this dataset.}

\item{\code{numRows}}{The number of rows in this dataset.}

\item{\code{predictionsEndDate}}{The date string of the prediction end date of this prediction dataset. Used for bulk predictions. Note that this parameter is for generating historical predictions using the training data. Only non-null for time series projects.}

\item{\code{predictionsStartDate}}{The date string of the prediction start date of this prediction dataset. Used for bulk predictions. Note that this parameter is for generating historical predictions using the training data. Only non-null for time series projects.}

\item{\code{projectId}}{The project ID that owns this dataset.}

\item{\code{actualValueColumn}}{Optional, only available for unsupervised projects, in case dataset was uploaded with actual value column specified. Name of the column which will be used to calculate the classification metrics and insights.}

\item{\code{containsTargetValues}}{If True, dataset contains target values and can be used to calculate the classification metrics and insights. Only applies for supervised projects.}

\item{\code{dataEndDate}}{Only available for time series projects, a date string representing the maximum primary date of the prediction dataset.}

\item{\code{dataStartDate}}{Only available for time series projects, a date string representing the minimum primary date of the prediction dataset.}

\item{\code{detectedActualValueColumns}}{Only available for unsupervised projects, a list of detected `actualValueColumnInfo` objects which can be used to calculate the classification metrics and insights.}

\item{\code{forecastPointRange}}{Only available for time series projects, the start and end of the range of dates available for use as the forecast point, detected based on the uploaded prediction dataset.}

\item{\code{maxForecastDate}}{Only available for time series projects, a date string representing the maximum forecast date of this prediction dataset.}

\item{\code{secondaryDatasetsConfigId}}{Only available for Feature discovery projects. Id of the secondary dataset config used by the dataset for the prediction.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionDatasetRetrieveResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionDatasetRetrieveResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionDatasetRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionDatasetRetrieveResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionDatasetRetrieveResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionDatasetRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionDatasetRetrieveResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionDatasetRetrieveResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionDatasetRetrieveResponse$fromJSON(
  PredictionDatasetRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionDatasetRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionDatasetRetrieveResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionDatasetRetrieveResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionDatasetRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
