% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_data_source.R
\docType{class}
\name{PredictionDataSource}
\alias{PredictionDataSource}
\title{PredictionDataSource}
\format{
An \code{R6Class} generator object
}
\description{
PredictionDataSource Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actualValueColumn}}{character \link{optional} The actual value column name, valid for the prediction files if the project is unsupervised and the dataset is considered as bulk predictions dataset.}

\item{\code{credentialData}}{\link{OneOfBasicCredentialsS3CredentialsOAuthCredentials} \link{optional} The credentials to authenticate with the database, to use instead of user/password or credential ID.}

\item{\code{credentialId}}{character \link{optional} The credential ID to use for database authentication.}

\item{\code{credentials}}{list( \link{OneOfPasswordCredentialsCredentialId} ) \link{optional} A list of credentials for the secondary datasets used in feature discovery project.}

\item{\code{dataSourceId}}{character The ID of ``DataSource``.}

\item{\code{forecastPoint}}{character \link{optional} For time series projects only. The time in the dataset relative to which predictions are generated. This value is optional. If not specified the default value is the value in the row with the latest specified timestamp. Specifying this value for a project that is not a time series project will result in an error.}

\item{\code{password}}{character \link{optional} The password (in cleartext) for database authentication. The password will be encrypted on the server side in scope of HTTP request and never saved or stored. DEPRECATED: please use ``credentialId`` or ``credentialData`` instead.}

\item{\code{predictionsEndDate}}{character \link{optional} The end date for bulk predictions, exclusive. Used for time series projects only. Note that this parameter is used for generating historical predictions using the training data, not for future predictions. If not specified, the dataset is not considered as a bulk predictions dataset. This parameter should be provided in conjunction with a ``predictionsStartDate``, and cannot be provided with the ``forecastPoint`` parameter.}

\item{\code{predictionsStartDate}}{character \link{optional} The start date for bulk predictions. Used for time series projects only. Note that this parameter is used for generating historical predictions using the training data, not for future predictions. If not specified, the dataset is not considered as a bulk predictions dataset. This parameter should be provided in conjunction with a ``predictionsEndDate``, and cannot be provided with the ``forecastPoint`` parameter.}

\item{\code{relaxKnownInAdvanceFeaturesCheck}}{character \link{optional} For time series projects only. If true, missing values in the known in advance features are allowed in the forecast window at the prediction time. This value is optional. If omitted or false, missing values are not allowed.}

\item{\code{secondaryDatasetsConfigId}}{character \link{optional} For feature discovery projects only. The ID of the alternative secondary dataset config to use during prediction.}

\item{\code{useKerberos}}{character \link{optional} If true, use kerberos authentication for database authentication. Default is false.}

\item{\code{user}}{character \link{optional} The username for database authentication. DEPRECATED: please use ``credentialId`` or ``credentialData`` instead.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PredictionDataSource$new()}}
\item \href{#method-validate}{\code{PredictionDataSource$validate()}}
\item \href{#method-toJSON}{\code{PredictionDataSource$toJSON()}}
\item \href{#method-fromJSON}{\code{PredictionDataSource$fromJSON()}}
\item \href{#method-clone}{\code{PredictionDataSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionDataSource$new(
  dataSourceId = NULL,
  actualValueColumn = NULL,
  credentialData = NULL,
  credentialId = NULL,
  credentials = NULL,
  forecastPoint = NULL,
  password = NULL,
  predictionsEndDate = NULL,
  predictionsStartDate = NULL,
  relaxKnownInAdvanceFeaturesCheck = NULL,
  secondaryDatasetsConfigId = NULL,
  useKerberos = FALSE,
  user = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataSourceId}}{The ID of ``DataSource``.}

\item{\code{actualValueColumn}}{The actual value column name, valid for the prediction files if the project is unsupervised and the dataset is considered as bulk predictions dataset.}

\item{\code{credentialData}}{The credentials to authenticate with the database, to use instead of user/password or credential ID.}

\item{\code{credentialId}}{The credential ID to use for database authentication.}

\item{\code{credentials}}{A list of credentials for the secondary datasets used in feature discovery project.}

\item{\code{forecastPoint}}{For time series projects only. The time in the dataset relative to which predictions are generated. This value is optional. If not specified the default value is the value in the row with the latest specified timestamp. Specifying this value for a project that is not a time series project will result in an error.}

\item{\code{password}}{The password (in cleartext) for database authentication. The password will be encrypted on the server side in scope of HTTP request and never saved or stored. DEPRECATED: please use ``credentialId`` or ``credentialData`` instead.}

\item{\code{predictionsEndDate}}{The end date for bulk predictions, exclusive. Used for time series projects only. Note that this parameter is used for generating historical predictions using the training data, not for future predictions. If not specified, the dataset is not considered as a bulk predictions dataset. This parameter should be provided in conjunction with a ``predictionsStartDate``, and cannot be provided with the ``forecastPoint`` parameter.}

\item{\code{predictionsStartDate}}{The start date for bulk predictions. Used for time series projects only. Note that this parameter is used for generating historical predictions using the training data, not for future predictions. If not specified, the dataset is not considered as a bulk predictions dataset. This parameter should be provided in conjunction with a ``predictionsEndDate``, and cannot be provided with the ``forecastPoint`` parameter.}

\item{\code{relaxKnownInAdvanceFeaturesCheck}}{For time series projects only. If true, missing values in the known in advance features are allowed in the forecast window at the prediction time. This value is optional. If omitted or false, missing values are not allowed.}

\item{\code{secondaryDatasetsConfigId}}{For feature discovery projects only. The ID of the alternative secondary dataset config to use during prediction.}

\item{\code{useKerberos}}{If true, use kerberos authentication for database authentication. Default is false.}

\item{\code{user}}{The username for database authentication. DEPRECATED: please use ``credentialId`` or ``credentialData`` instead.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionDataSource$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionDataSource$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionDataSource$fromJSON(PredictionDataSourceJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionDataSourceJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionDataSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
