% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_explanations_create.R
\docType{class}
\name{PredictionExplanationsCreate}
\alias{PredictionExplanationsCreate}
\title{PredictionExplanationsCreate}
\format{
An \code{R6Class} generator object
}
\description{
PredictionExplanationsCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{classNames}}{list( character ) \link{optional} List of class names that will be explained for each row for multiclass. Mutually exclusive with numTopClasses. If neither specified - we assume numTopClasses=1.}

\item{\code{datasetId}}{character The dataset ID.}

\item{\code{maxExplanations}}{integer \link{optional} The maximum number of prediction explanations to supply per row of the dataset.}

\item{\code{modelId}}{character The model ID.}

\item{\code{numTopClasses}}{integer \link{optional} Number of top predicted classes for each row that will be explained for multiclass. Mutually exclusive with classNames. If neither specified - we assume numTopClasses=1.}

\item{\code{thresholdHigh}}{numeric \link{optional} The high threshold, above which a prediction must score in order for prediction explanations to be computed. If neither thresholdHigh nor thresholdLow is specified, prediction explanations will be computed for all rows.}

\item{\code{thresholdLow}}{numeric \link{optional} The lower threshold, below which a prediction must score in order for prediction explanations to be computed for a row in the dataset. If neither thresholdHigh nor thresholdLow is specified, prediction explanations will be computed for all rows.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictionExplanationsCreate-new}{\code{PredictionExplanationsCreate$new()}}
\item \href{#method-PredictionExplanationsCreate-validate}{\code{PredictionExplanationsCreate$validate()}}
\item \href{#method-PredictionExplanationsCreate-toJSON}{\code{PredictionExplanationsCreate$toJSON()}}
\item \href{#method-PredictionExplanationsCreate-fromJSON}{\code{PredictionExplanationsCreate$fromJSON()}}
\item \href{#method-PredictionExplanationsCreate-clone}{\code{PredictionExplanationsCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsCreate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsCreate$new(
  datasetId = NULL,
  modelId = NULL,
  classNames = NULL,
  maxExplanations = 3,
  numTopClasses = NULL,
  thresholdHigh = NULL,
  thresholdLow = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasetId}}{The dataset ID.}

\item{\code{modelId}}{The model ID.}

\item{\code{classNames}}{List of class names that will be explained for each row for multiclass. Mutually exclusive with numTopClasses. If neither specified - we assume numTopClasses=1.}

\item{\code{maxExplanations}}{The maximum number of prediction explanations to supply per row of the dataset.}

\item{\code{numTopClasses}}{Number of top predicted classes for each row that will be explained for multiclass. Mutually exclusive with classNames. If neither specified - we assume numTopClasses=1.}

\item{\code{thresholdHigh}}{The high threshold, above which a prediction must score in order for prediction explanations to be computed. If neither thresholdHigh nor thresholdLow is specified, prediction explanations will be computed for all rows.}

\item{\code{thresholdLow}}{The lower threshold, below which a prediction must score in order for prediction explanations to be computed for a row in the dataset. If neither thresholdHigh nor thresholdLow is specified, prediction explanations will be computed for all rows.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsCreate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsCreate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsCreate$fromJSON(
  PredictionExplanationsCreateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionExplanationsCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
