% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_explanations_initialization_create.R
\docType{class}
\name{PredictionExplanationsInitializationCreate}
\alias{PredictionExplanationsInitializationCreate}
\title{PredictionExplanationsInitializationCreate}
\format{
An \code{R6Class} generator object
}
\description{
PredictionExplanationsInitializationCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{maxExplanations}}{integer \link{optional} The maximum number of prediction explanations to supply per row of the dataset.}

\item{\code{thresholdHigh}}{numeric \link{optional} The high threshold, above which a prediction must score in order for prediction explanations to be computed. If neither thresholdHigh nor thresholdLow is specified, prediction explanations will be computed for all rows.}

\item{\code{thresholdLow}}{numeric \link{optional} The lower threshold, below which a prediction must score in order for prediction explanations to be computed for a row in the dataset. If neither thresholdHigh nor thresholdLow is specified, prediction explanations will be computed for all rows.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictionExplanationsInitializationCreate-new}{\code{PredictionExplanationsInitializationCreate$new()}}
\item \href{#method-PredictionExplanationsInitializationCreate-validate}{\code{PredictionExplanationsInitializationCreate$validate()}}
\item \href{#method-PredictionExplanationsInitializationCreate-toJSON}{\code{PredictionExplanationsInitializationCreate$toJSON()}}
\item \href{#method-PredictionExplanationsInitializationCreate-fromJSON}{\code{PredictionExplanationsInitializationCreate$fromJSON()}}
\item \href{#method-PredictionExplanationsInitializationCreate-clone}{\code{PredictionExplanationsInitializationCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsInitializationCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsInitializationCreate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsInitializationCreate$new(
  maxExplanations = 3,
  thresholdHigh = NULL,
  thresholdLow = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maxExplanations}}{The maximum number of prediction explanations to supply per row of the dataset.}

\item{\code{thresholdHigh}}{The high threshold, above which a prediction must score in order for prediction explanations to be computed. If neither thresholdHigh nor thresholdLow is specified, prediction explanations will be computed for all rows.}

\item{\code{thresholdLow}}{The lower threshold, below which a prediction must score in order for prediction explanations to be computed for a row in the dataset. If neither thresholdHigh nor thresholdLow is specified, prediction explanations will be computed for all rows.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsInitializationCreate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsInitializationCreate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsInitializationCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsInitializationCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsInitializationCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsInitializationCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsInitializationCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsInitializationCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsInitializationCreate$fromJSON(
  PredictionExplanationsInitializationCreateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionExplanationsInitializationCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsInitializationCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsInitializationCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsInitializationCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
