% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prediction_explanations_initialization_retrieve.R
\docType{class}
\name{PredictionExplanationsInitializationRetrieve}
\alias{PredictionExplanationsInitializationRetrieve}
\title{PredictionExplanationsInitializationRetrieve}
\format{
An \code{R6Class} generator object
}
\description{
PredictionExplanationsInitializationRetrieve Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{modelId}}{character The model ID.}

\item{\code{predictionExplanationsSample}}{list( \link{PredictionExplanationsRow} ) Each is a PredictionExplanationsRow. They represent a small sample of prediction explanations that could be generated for a particular dataset. They will have the same schema as the `data` array in the response from :http:get:`/api/v2/projects/(projectId)/predictionExplanations/(predictionExplanationsId)/`. As of v2.21 only difference is that there is no forecastPoint in response for time series projects.}

\item{\code{projectId}}{character The project ID.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictionExplanationsInitializationRetrieve-new}{\code{PredictionExplanationsInitializationRetrieve$new()}}
\item \href{#method-PredictionExplanationsInitializationRetrieve-validate}{\code{PredictionExplanationsInitializationRetrieve$validate()}}
\item \href{#method-PredictionExplanationsInitializationRetrieve-toJSON}{\code{PredictionExplanationsInitializationRetrieve$toJSON()}}
\item \href{#method-PredictionExplanationsInitializationRetrieve-fromJSON}{\code{PredictionExplanationsInitializationRetrieve$fromJSON()}}
\item \href{#method-PredictionExplanationsInitializationRetrieve-clone}{\code{PredictionExplanationsInitializationRetrieve$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsInitializationRetrieve-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsInitializationRetrieve-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsInitializationRetrieve$new(
  modelId = NULL,
  predictionExplanationsSample = NULL,
  projectId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{modelId}}{The model ID.}

\item{\code{predictionExplanationsSample}}{Each is a PredictionExplanationsRow. They represent a small sample of prediction explanations that could be generated for a particular dataset. They will have the same schema as the `data` array in the response from :http:get:`/api/v2/projects/(projectId)/predictionExplanations/(predictionExplanationsId)/`. As of v2.21 only difference is that there is no forecastPoint in response for time series projects.}

\item{\code{projectId}}{The project ID.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsInitializationRetrieve-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsInitializationRetrieve-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsInitializationRetrieve$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsInitializationRetrieve-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsInitializationRetrieve-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsInitializationRetrieve$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsInitializationRetrieve-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsInitializationRetrieve-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsInitializationRetrieve$fromJSON(
  PredictionExplanationsInitializationRetrieveJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionExplanationsInitializationRetrieveJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsInitializationRetrieve-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsInitializationRetrieve-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsInitializationRetrieve$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
