% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_explanations_prediction_values.R
\docType{class}
\name{PredictionExplanationsPredictionValues}
\alias{PredictionExplanationsPredictionValues}
\title{PredictionExplanationsPredictionValues}
\format{
An \code{R6Class} generator object
}
\description{
PredictionExplanationsPredictionValues Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{label}}{character Describes what this model output corresponds to. For regression projects, it is the name of the target feature. For classification projects, it is a level from the target feature. For Anomaly Detection models it is an Anomaly Score.}

\item{\code{value}}{numeric The output of the prediction. For regression projects, it is the predicted value of the target. For classification projects, it is the predicted probability the row belongs to the class identified by the label.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictionExplanationsPredictionValues-new}{\code{PredictionExplanationsPredictionValues$new()}}
\item \href{#method-PredictionExplanationsPredictionValues-validate}{\code{PredictionExplanationsPredictionValues$validate()}}
\item \href{#method-PredictionExplanationsPredictionValues-toJSON}{\code{PredictionExplanationsPredictionValues$toJSON()}}
\item \href{#method-PredictionExplanationsPredictionValues-fromJSON}{\code{PredictionExplanationsPredictionValues$fromJSON()}}
\item \href{#method-PredictionExplanationsPredictionValues-clone}{\code{PredictionExplanationsPredictionValues$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsPredictionValues-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsPredictionValues-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsPredictionValues$new(
  label = NULL,
  value = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{Describes what this model output corresponds to. For regression projects, it is the name of the target feature. For classification projects, it is a level from the target feature. For Anomaly Detection models it is an Anomaly Score.}

\item{\code{value}}{The output of the prediction. For regression projects, it is the predicted value of the target. For classification projects, it is the predicted probability the row belongs to the class identified by the label.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsPredictionValues-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsPredictionValues-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsPredictionValues$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsPredictionValues-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsPredictionValues-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsPredictionValues$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsPredictionValues-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsPredictionValues-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsPredictionValues$fromJSON(
  PredictionExplanationsPredictionValuesJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionExplanationsPredictionValuesJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsPredictionValues-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsPredictionValues-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsPredictionValues$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
