% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_explanations_record.R
\docType{class}
\name{PredictionExplanationsRecord}
\alias{PredictionExplanationsRecord}
\title{PredictionExplanationsRecord}
\format{
An \code{R6Class} generator object
}
\description{
PredictionExplanationsRecord Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datasetId}}{character The dataset ID.}

\item{\code{finishTime}}{numeric Timestamp referencing when computation for these prediction explanations finished.}

\item{\code{id}}{character The PredictionExplanationsRecord ID.}

\item{\code{maxExplanations}}{integer The maximum number of codes generated per prediction.}

\item{\code{modelId}}{character The model ID.}

\item{\code{numColumns}}{integer The number of columns prediction explanations were computed for.}

\item{\code{predictionExplanationsLocation}}{character Where to retrieve the prediction explanations.}

\item{\code{predictionThreshold}}{numeric The threshold value used for binary classification prediction.}

\item{\code{projectId}}{character The project ID.}

\item{\code{thresholdHigh}}{numeric The prediction explanation high threshold. Predictions must be above this value (or below the thresholdLow value) to have PredictionExplanations computed.}

\item{\code{thresholdLow}}{numeric The prediction explanation low threshold. Predictions must be below this value (or above the thresholdHigh value) to have PredictionExplanations computed.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictionExplanationsRecord-new}{\code{PredictionExplanationsRecord$new()}}
\item \href{#method-PredictionExplanationsRecord-validate}{\code{PredictionExplanationsRecord$validate()}}
\item \href{#method-PredictionExplanationsRecord-toJSON}{\code{PredictionExplanationsRecord$toJSON()}}
\item \href{#method-PredictionExplanationsRecord-fromJSON}{\code{PredictionExplanationsRecord$fromJSON()}}
\item \href{#method-PredictionExplanationsRecord-clone}{\code{PredictionExplanationsRecord$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsRecord-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsRecord-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsRecord$new(
  datasetId = NULL,
  finishTime = NULL,
  id = NULL,
  maxExplanations = NULL,
  modelId = NULL,
  numColumns = NULL,
  predictionExplanationsLocation = NULL,
  predictionThreshold = NULL,
  projectId = NULL,
  thresholdHigh = NULL,
  thresholdLow = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasetId}}{The dataset ID.}

\item{\code{finishTime}}{Timestamp referencing when computation for these prediction explanations finished.}

\item{\code{id}}{The PredictionExplanationsRecord ID.}

\item{\code{maxExplanations}}{The maximum number of codes generated per prediction.}

\item{\code{modelId}}{The model ID.}

\item{\code{numColumns}}{The number of columns prediction explanations were computed for.}

\item{\code{predictionExplanationsLocation}}{Where to retrieve the prediction explanations.}

\item{\code{predictionThreshold}}{The threshold value used for binary classification prediction.}

\item{\code{projectId}}{The project ID.}

\item{\code{thresholdHigh}}{The prediction explanation high threshold. Predictions must be above this value (or below the thresholdLow value) to have PredictionExplanations computed.}

\item{\code{thresholdLow}}{The prediction explanation low threshold. Predictions must be below this value (or above the thresholdHigh value) to have PredictionExplanations computed.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsRecord-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsRecord-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsRecord$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsRecord-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsRecord-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsRecord$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsRecord-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsRecord-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsRecord$fromJSON(
  PredictionExplanationsRecordJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionExplanationsRecordJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsRecord-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsRecord-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsRecord$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
