% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_explanations_record_list.R
\docType{class}
\name{PredictionExplanationsRecordList}
\alias{PredictionExplanationsRecordList}
\title{PredictionExplanationsRecordList}
\format{
An \code{R6Class} generator object
}
\description{
PredictionExplanationsRecordList Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer The number of items returned on this page.}

\item{\code{data}}{list( \link{PredictionExplanationsRecord} ) Each has the same schema as if retrieving the prediction explanations individually from :http:get:`/api/v2/projects/(projectId)/predictionExplanationsRecords/(predictionExplanationsId)/`.}

\item{\code{next_}}{character A URL pointing to the next page (if `null`, there is no next page).}

\item{\code{previous}}{character A URL pointing to the previous page (if `null`, there is no previous page).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictionExplanationsRecordList-new}{\code{PredictionExplanationsRecordList$new()}}
\item \href{#method-PredictionExplanationsRecordList-validate}{\code{PredictionExplanationsRecordList$validate()}}
\item \href{#method-PredictionExplanationsRecordList-toJSON}{\code{PredictionExplanationsRecordList$toJSON()}}
\item \href{#method-PredictionExplanationsRecordList-fromJSON}{\code{PredictionExplanationsRecordList$fromJSON()}}
\item \href{#method-PredictionExplanationsRecordList-clone}{\code{PredictionExplanationsRecordList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsRecordList-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsRecordList-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsRecordList$new(
  count = NULL,
  data = NULL,
  next_ = NULL,
  previous = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{The number of items returned on this page.}

\item{\code{data}}{Each has the same schema as if retrieving the prediction explanations individually from :http:get:`/api/v2/projects/(projectId)/predictionExplanationsRecords/(predictionExplanationsId)/`.}

\item{\code{next_}}{A URL pointing to the next page (if `null`, there is no next page).}

\item{\code{previous}}{A URL pointing to the previous page (if `null`, there is no previous page).}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsRecordList-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsRecordList-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsRecordList$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsRecordList-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsRecordList-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsRecordList$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsRecordList-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsRecordList-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsRecordList$fromJSON(
  PredictionExplanationsRecordListJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionExplanationsRecordListJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsRecordList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsRecordList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsRecordList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
