% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_explanations_retrieve.R
\docType{class}
\name{PredictionExplanationsRetrieve}
\alias{PredictionExplanationsRetrieve}
\title{PredictionExplanationsRetrieve}
\format{
An \code{R6Class} generator object
}
\description{
PredictionExplanationsRetrieve Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{adjustmentMethod}}{character 'exposureNormalized' (for regression projects with exposure) or 'N/A' (for classification projects) The value of 'exposureNormalized' indicates that prediction outputs are adjusted (or divided) by exposure. The value of 'N/A' indicates that no adjustments are applied to the adjusted predictions and they are identical to the unadjusted predictions.}

\item{\code{count}}{integer How many rows of prediction explanations were returned.}

\item{\code{data}}{list( \link{PredictionExplanationsRow} ) Each is a PredictionExplanationsRow corresponding to one row of the prediction dataset.}

\item{\code{id}}{character The ID of this group of prediction explanations.}

\item{\code{next_}}{character URL pointing to the next page (if null, there is no next page).}

\item{\code{predictionExplanationsRecordLocation}}{character The URL of the PredictionExplanationsRecord associated with these prediction explanations.}

\item{\code{previous}}{character URL pointing to the previous page (if null, there is no previous page).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictionExplanationsRetrieve-new}{\code{PredictionExplanationsRetrieve$new()}}
\item \href{#method-PredictionExplanationsRetrieve-validate}{\code{PredictionExplanationsRetrieve$validate()}}
\item \href{#method-PredictionExplanationsRetrieve-toJSON}{\code{PredictionExplanationsRetrieve$toJSON()}}
\item \href{#method-PredictionExplanationsRetrieve-fromJSON}{\code{PredictionExplanationsRetrieve$fromJSON()}}
\item \href{#method-PredictionExplanationsRetrieve-clone}{\code{PredictionExplanationsRetrieve$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsRetrieve-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsRetrieve-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsRetrieve$new(
  adjustmentMethod = NULL,
  count = NULL,
  data = NULL,
  id = NULL,
  next_ = NULL,
  predictionExplanationsRecordLocation = NULL,
  previous = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{adjustmentMethod}}{'exposureNormalized' (for regression projects with exposure) or 'N/A' (for classification projects) The value of 'exposureNormalized' indicates that prediction outputs are adjusted (or divided) by exposure. The value of 'N/A' indicates that no adjustments are applied to the adjusted predictions and they are identical to the unadjusted predictions.}

\item{\code{count}}{How many rows of prediction explanations were returned.}

\item{\code{data}}{Each is a PredictionExplanationsRow corresponding to one row of the prediction dataset.}

\item{\code{id}}{The ID of this group of prediction explanations.}

\item{\code{next_}}{URL pointing to the next page (if null, there is no next page).}

\item{\code{predictionExplanationsRecordLocation}}{The URL of the PredictionExplanationsRecord associated with these prediction explanations.}

\item{\code{previous}}{URL pointing to the previous page (if null, there is no previous page).}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsRetrieve-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsRetrieve-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsRetrieve$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsRetrieve-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsRetrieve-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsRetrieve$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsRetrieve-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsRetrieve-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsRetrieve$fromJSON(
  PredictionExplanationsRetrieveJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionExplanationsRetrieveJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsRetrieve-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsRetrieve-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsRetrieve$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
