% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_explanations_values.R
\docType{class}
\name{PredictionExplanationsValues}
\alias{PredictionExplanationsValues}
\title{PredictionExplanationsValues}
\format{
An \code{R6Class} generator object
}
\description{
PredictionExplanationsValues Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{feature}}{character The name of the feature contributing to the prediction.}

\item{\code{featureValue}}{\link{OneOfstringnumber} The value the feature took on for this row.}

\item{\code{label}}{\link{OneOfstringnumber} Describes what output was driven by this prediction explanation. For regression projects, it is the name of the target feature. For classification projects, it is the class whose probability increasing would correspond to a positive strength of this prediction explanation.}

\item{\code{qualitativeStrength}}{character The feature strength as a string, a plus or minus indicator from `+++` to `---`.}

\item{\code{strength}}{numeric \link{optional} The feature strength as a float.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PredictionExplanationsValues$new()}}
\item \href{#method-validate}{\code{PredictionExplanationsValues$validate()}}
\item \href{#method-toJSON}{\code{PredictionExplanationsValues$toJSON()}}
\item \href{#method-fromJSON}{\code{PredictionExplanationsValues$fromJSON()}}
\item \href{#method-clone}{\code{PredictionExplanationsValues$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsValues$new(
  feature = NULL,
  featureValue = NULL,
  label = NULL,
  qualitativeStrength = NULL,
  strength = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{The name of the feature contributing to the prediction.}

\item{\code{featureValue}}{The value the feature took on for this row.}

\item{\code{label}}{Describes what output was driven by this prediction explanation. For regression projects, it is the name of the target feature. For classification projects, it is the class whose probability increasing would correspond to a positive strength of this prediction explanation.}

\item{\code{qualitativeStrength}}{The feature strength as a string, a plus or minus indicator from `+++` to `---`.}

\item{\code{strength}}{The feature strength as a float.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsValues$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsValues$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsValues$fromJSON(
  PredictionExplanationsValuesJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionExplanationsValuesJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsValues$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
