% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_file_upload.R
\docType{class}
\name{PredictionFileUpload}
\alias{PredictionFileUpload}
\title{PredictionFileUpload}
\format{
An \code{R6Class} generator object
}
\description{
PredictionFileUpload Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actualValueColumn}}{character \link{optional} Actual value column name, valid for the prediction files if the project is unsupervised and the dataset is considered as bulk predictions dataset.}

\item{\code{credentials}}{character \link{optional} A list of credentials for the secondary datasets used in feature discovery project}

\item{\code{file}}{data.frame The dataset file to upload for prediction.}

\item{\code{forecastPoint}}{character \link{optional} For time series projects only. The time in the dataset relative to which predictions are generated. If not specified the default value is the value in the row with the latest specified timestamp. Specifying this value for a project that is not a time series project will result in an error.}

\item{\code{predictionsEndDate}}{character \link{optional} Used for time series projects only. The end date for bulk predictions. Note that this parameter is used for generating historical predictions using the training data, not for future predictions. If not specified, the dataset is not considered as a bulk predictions dataset. This parameter should be provided in conjunction with a ``predictionsStartDate``, and cannot be provided with the ``forecastPoint`` parameter.}

\item{\code{predictionsStartDate}}{character \link{optional} Used for time series projects only. The start date for bulk predictions. Note that this parameter is used for generating historical predictions using the training data, not for future predictions. If not specified, the dataset is not considered as a bulk predictions dataset. This parameter should be provided in conjunction with a ``predictionsEndDate``, and cannot be provided with the ``forecastPoint`` parameter.}

\item{\code{relaxKnownInAdvanceFeaturesCheck}}{character \link{optional} A boolean flag. If true, missing values in the known in advance features are allowed in the forecast window at the prediction time. If omitted or false, missing values are not allowed. For time series projects only.}

\item{\code{secondaryDatasetsConfigId}}{character \link{optional} Optional, for feature discovery projects only. The Id of the alternative secondary dataset config to use during prediction.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PredictionFileUpload$new()}}
\item \href{#method-validate}{\code{PredictionFileUpload$validate()}}
\item \href{#method-toJSON}{\code{PredictionFileUpload$toJSON()}}
\item \href{#method-fromJSON}{\code{PredictionFileUpload$fromJSON()}}
\item \href{#method-clone}{\code{PredictionFileUpload$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionFileUpload$new(
  file = NULL,
  actualValueColumn = NULL,
  credentials = NULL,
  forecastPoint = NULL,
  predictionsEndDate = NULL,
  predictionsStartDate = NULL,
  relaxKnownInAdvanceFeaturesCheck = NULL,
  secondaryDatasetsConfigId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{The dataset file to upload for prediction.}

\item{\code{actualValueColumn}}{Actual value column name, valid for the prediction files if the project is unsupervised and the dataset is considered as bulk predictions dataset.}

\item{\code{credentials}}{A list of credentials for the secondary datasets used in feature discovery project}

\item{\code{forecastPoint}}{For time series projects only. The time in the dataset relative to which predictions are generated. If not specified the default value is the value in the row with the latest specified timestamp. Specifying this value for a project that is not a time series project will result in an error.}

\item{\code{predictionsEndDate}}{Used for time series projects only. The end date for bulk predictions. Note that this parameter is used for generating historical predictions using the training data, not for future predictions. If not specified, the dataset is not considered as a bulk predictions dataset. This parameter should be provided in conjunction with a ``predictionsStartDate``, and cannot be provided with the ``forecastPoint`` parameter.}

\item{\code{predictionsStartDate}}{Used for time series projects only. The start date for bulk predictions. Note that this parameter is used for generating historical predictions using the training data, not for future predictions. If not specified, the dataset is not considered as a bulk predictions dataset. This parameter should be provided in conjunction with a ``predictionsEndDate``, and cannot be provided with the ``forecastPoint`` parameter.}

\item{\code{relaxKnownInAdvanceFeaturesCheck}}{A boolean flag. If true, missing values in the known in advance features are allowed in the forecast window at the prediction time. If omitted or false, missing values are not allowed. For time series projects only.}

\item{\code{secondaryDatasetsConfigId}}{Optional, for feature discovery projects only. The Id of the alternative secondary dataset config to use during prediction.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionFileUpload$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionFileUpload$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionFileUpload$fromJSON(PredictionFileUploadJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionFileUploadJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionFileUpload$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
