% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_integration_job_create_payload.R
\docType{class}
\name{PredictionIntegrationJobCreatePayload}
\alias{PredictionIntegrationJobCreatePayload}
\title{PredictionIntegrationJobCreatePayload}
\format{
An \code{R6Class} generator object
}
\description{
PredictionIntegrationJobCreatePayload Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{deploymentId}}{character Id of the deployment used to run predictions}

\item{\code{integrationTypeId}}{character Type Id for this integration}

\item{\code{integrationTypeName}}{character Name of the integration}

\item{\code{name}}{character Scheduled Job name}

\item{\code{predictionOptions}}{\link{PredictionOptionsPayload}}

\item{\code{schedule}}{\link{ScheduledJobPayload}}

\item{\code{source}}{\link{OneOfUsernamePasswordCatalogDataSourcePayloadCredentialsCatalogDataSourcePayload} Object describing data source for prediction}

\item{\code{target}}{\link{OneOfCredentialsJdbcPredictionTargetPayloadUsernamePasswordJdbcPredictionTargetPayloadTableauServerPredictionTargetPayload} Settings for the predictions output destination}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PredictionIntegrationJobCreatePayload$new()}}
\item \href{#method-validate}{\code{PredictionIntegrationJobCreatePayload$validate()}}
\item \href{#method-toJSON}{\code{PredictionIntegrationJobCreatePayload$toJSON()}}
\item \href{#method-fromJSON}{\code{PredictionIntegrationJobCreatePayload$fromJSON()}}
\item \href{#method-clone}{\code{PredictionIntegrationJobCreatePayload$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionIntegrationJobCreatePayload$new(
  deploymentId = NULL,
  integrationTypeId = NULL,
  integrationTypeName = NULL,
  name = NULL,
  predictionOptions = NULL,
  schedule = NULL,
  source = NULL,
  target = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deploymentId}}{Id of the deployment used to run predictions}

\item{\code{integrationTypeId}}{Type Id for this integration}

\item{\code{integrationTypeName}}{Name of the integration}

\item{\code{name}}{Scheduled Job name}

\item{\code{predictionOptions}}{}

\item{\code{schedule}}{}

\item{\code{source}}{Object describing data source for prediction}

\item{\code{target}}{Settings for the predictions output destination}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionIntegrationJobCreatePayload$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionIntegrationJobCreatePayload$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionIntegrationJobCreatePayload$fromJSON(
  PredictionIntegrationJobCreatePayloadJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionIntegrationJobCreatePayloadJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionIntegrationJobCreatePayload$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
