% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_integration_job_response.R
\docType{class}
\name{PredictionIntegrationJobResponse}
\alias{PredictionIntegrationJobResponse}
\title{PredictionIntegrationJobResponse}
\format{
An \code{R6Class} generator object
}
\description{
PredictionIntegrationJobResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{deploymentId}}{character Id of the deployment used to run predictions}

\item{\code{id}}{character Unique identifier of the prediction integration job}

\item{\code{integrationTypeId}}{character Type Id for this integration}

\item{\code{integrationTypeName}}{character Name of the integration}

\item{\code{name}}{character Scheduled Job name}

\item{\code{predictionOptions}}{\link{PredictionOptionsPayload}}

\item{\code{schedule}}{\link{ScheduledJobPayload}}

\item{\code{source}}{\link{OneOfUsernamePasswordCatalogDataSourcePayloadCredentialsCatalogDataSourcePayload} Object describing data source for prediction}

\item{\code{target}}{\link{OneOfCredentialsJdbcPredictionTargetPayloadUsernamePasswordJdbcPredictionTargetPayloadTableauServerPredictionTargetResponse} Settings for the predictions output destination}

\item{\code{updatedAt}}{character Last updated date in format \%B \%d, \%Y}

\item{\code{updatedById}}{character User id of last modifier}

\item{\code{updatedByName}}{character Name of last modifier}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PredictionIntegrationJobResponse$new()}}
\item \href{#method-validate}{\code{PredictionIntegrationJobResponse$validate()}}
\item \href{#method-toJSON}{\code{PredictionIntegrationJobResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{PredictionIntegrationJobResponse$fromJSON()}}
\item \href{#method-clone}{\code{PredictionIntegrationJobResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionIntegrationJobResponse$new(
  deploymentId = NULL,
  id = NULL,
  integrationTypeId = NULL,
  integrationTypeName = NULL,
  name = NULL,
  predictionOptions = NULL,
  schedule = NULL,
  source = NULL,
  target = NULL,
  updatedAt = NULL,
  updatedById = NULL,
  updatedByName = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deploymentId}}{Id of the deployment used to run predictions}

\item{\code{id}}{Unique identifier of the prediction integration job}

\item{\code{integrationTypeId}}{Type Id for this integration}

\item{\code{integrationTypeName}}{Name of the integration}

\item{\code{name}}{Scheduled Job name}

\item{\code{predictionOptions}}{}

\item{\code{schedule}}{}

\item{\code{source}}{Object describing data source for prediction}

\item{\code{target}}{Settings for the predictions output destination}

\item{\code{updatedAt}}{Last updated date in format \%B \%d, \%Y}

\item{\code{updatedById}}{User id of last modifier}

\item{\code{updatedByName}}{Name of last modifier}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionIntegrationJobResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionIntegrationJobResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionIntegrationJobResponse$fromJSON(
  PredictionIntegrationJobResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionIntegrationJobResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionIntegrationJobResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
