% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_intervals_create.R
\docType{class}
\name{PredictionIntervalsCreate}
\alias{PredictionIntervalsCreate}
\title{PredictionIntervalsCreate}
\format{
An \code{R6Class} generator object
}
\description{
PredictionIntervalsCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{percentiles}}{list( integer ) The list of prediction intervals percentiles to calculate. Currently we only allow requesting one interval at a time.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictionIntervalsCreate-new}{\code{PredictionIntervalsCreate$new()}}
\item \href{#method-PredictionIntervalsCreate-validate}{\code{PredictionIntervalsCreate$validate()}}
\item \href{#method-PredictionIntervalsCreate-toJSON}{\code{PredictionIntervalsCreate$toJSON()}}
\item \href{#method-PredictionIntervalsCreate-fromJSON}{\code{PredictionIntervalsCreate$fromJSON()}}
\item \href{#method-PredictionIntervalsCreate-clone}{\code{PredictionIntervalsCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionIntervalsCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionIntervalsCreate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionIntervalsCreate$new(percentiles = NULL, validateParams = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{percentiles}}{The list of prediction intervals percentiles to calculate. Currently we only allow requesting one interval at a time.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionIntervalsCreate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionIntervalsCreate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionIntervalsCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionIntervalsCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionIntervalsCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionIntervalsCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionIntervalsCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionIntervalsCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionIntervalsCreate$fromJSON(
  PredictionIntervalsCreateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionIntervalsCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionIntervalsCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionIntervalsCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionIntervalsCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
