% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_options_payload.R
\docType{class}
\name{PredictionOptionsPayload}
\alias{PredictionOptionsPayload}
\title{PredictionOptionsPayload}
\format{
An \code{R6Class} generator object
}
\description{
PredictionOptionsPayload Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{errorLogColumn}}{character Toggle for saving errors}

\item{\code{overwrite}}{character Whether to overwrite the existing output or not}

\item{\code{passthroughColumns}}{list( character ) Array of strings of passthrough columns}

\item{\code{passthroughColumnsSchemas}}{list( \link{PassthroughColumnDefinitionPayload} ) \link{optional} (Only for Tableau) Array of passthrough column description objects with JDBC specific information for each of the users selected passthrough column, if any are included}

\item{\code{predictionExplanations}}{integer Quantity of prediction explanations}

\item{\code{timeseriesSettings}}{\link{TimeSeriesFieldsPayload} \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PredictionOptionsPayload$new()}}
\item \href{#method-validate}{\code{PredictionOptionsPayload$validate()}}
\item \href{#method-toJSON}{\code{PredictionOptionsPayload$toJSON()}}
\item \href{#method-fromJSON}{\code{PredictionOptionsPayload$fromJSON()}}
\item \href{#method-clone}{\code{PredictionOptionsPayload$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionOptionsPayload$new(
  errorLogColumn = NULL,
  overwrite = NULL,
  passthroughColumns = NULL,
  predictionExplanations = NULL,
  passthroughColumnsSchemas = NULL,
  timeseriesSettings = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{errorLogColumn}}{Toggle for saving errors}

\item{\code{overwrite}}{Whether to overwrite the existing output or not}

\item{\code{passthroughColumns}}{Array of strings of passthrough columns}

\item{\code{predictionExplanations}}{Quantity of prediction explanations}

\item{\code{passthroughColumnsSchemas}}{(Only for Tableau) Array of passthrough column description objects with JDBC specific information for each of the users selected passthrough column, if any are included}

\item{\code{timeseriesSettings}}{}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionOptionsPayload$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionOptionsPayload$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionOptionsPayload$fromJSON(
  PredictionOptionsPayloadJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionOptionsPayloadJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionOptionsPayload$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
