% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_proxy_data.R
\docType{class}
\name{PredictionProxyData}
\alias{PredictionProxyData}
\title{PredictionProxyData}
\format{
An \code{R6Class} generator object
}
\description{
PredictionProxyData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{deploymentApprovalStatus}}{\link{OneOfstringstring} \link{optional} The approval status of the deployment.}

\item{\code{forecastDistance}}{integer \link{optional} In the case of time series project, the number of time units this row is away from the forecastPoint. The unit of time is determined by the timeUnit of the datetime partition column.}

\item{\code{forecastPoint}}{character \link{optional} For time series projects only. The time in the dataset relative to which predictions were generated.}

\item{\code{isOutlierPrediction}}{character \link{optional} The outlier warning check result.}

\item{\code{passthroughValues}}{object \link{optional}}

\item{\code{prediction}}{\link{OneOfstringnumber} In the case of a regression, the value predicted for the target in that row. In the case of any classification type, the class of the target which the row is predicted to belong to.}

\item{\code{predictionExplanations}}{list( \link{PredictionExplanationsValues} ) \link{optional} Array contains `predictionExplanation` objects. Will be present only if `explanationAlgorithm` is not null (prediction explanations were requested).}

\item{\code{predictionIntervals}}{object \link{optional}}

\item{\code{predictionThreshold}}{numeric \link{optional} Threshold used for binary classification in predictions.}

\item{\code{predictionValues}}{list( \link{PredictionValues} ) Array of `predictionValue` objects for each class.}

\item{\code{rowId}}{integer The row in the prediction dataset this prediction corresponds to.}

\item{\code{seriesId}}{character \link{optional} The ID of the series value for a multiseries project. For time series projects that are not multiseries this will be a null.}

\item{\code{shapExplanationsMetadata}}{list( \link{ShapExplanationsMetadataValues} ) \link{optional} Array contains prediction explanation metadata.}

\item{\code{timestamp}}{character \link{optional} The timestamp of this row in the prediction dataset.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PredictionProxyData$new()}}
\item \href{#method-validate}{\code{PredictionProxyData$validate()}}
\item \href{#method-toJSON}{\code{PredictionProxyData$toJSON()}}
\item \href{#method-fromJSON}{\code{PredictionProxyData$fromJSON()}}
\item \href{#method-clone}{\code{PredictionProxyData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionProxyData$new(
  prediction = NULL,
  predictionValues = NULL,
  rowId = NULL,
  deploymentApprovalStatus = NULL,
  forecastDistance = NULL,
  forecastPoint = NULL,
  isOutlierPrediction = NULL,
  passthroughValues = NULL,
  predictionExplanations = NULL,
  predictionIntervals = NULL,
  predictionThreshold = NULL,
  seriesId = NULL,
  shapExplanationsMetadata = NULL,
  timestamp = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prediction}}{In the case of a regression, the value predicted for the target in that row. In the case of any classification type, the class of the target which the row is predicted to belong to.}

\item{\code{predictionValues}}{Array of `predictionValue` objects for each class.}

\item{\code{rowId}}{The row in the prediction dataset this prediction corresponds to.}

\item{\code{deploymentApprovalStatus}}{The approval status of the deployment.}

\item{\code{forecastDistance}}{In the case of time series project, the number of time units this row is away from the forecastPoint. The unit of time is determined by the timeUnit of the datetime partition column.}

\item{\code{forecastPoint}}{For time series projects only. The time in the dataset relative to which predictions were generated.}

\item{\code{isOutlierPrediction}}{The outlier warning check result.}

\item{\code{predictionExplanations}}{Array contains `predictionExplanation` objects. Will be present only if `explanationAlgorithm` is not null (prediction explanations were requested).}

\item{\code{predictionThreshold}}{Threshold used for binary classification in predictions.}

\item{\code{seriesId}}{The ID of the series value for a multiseries project. For time series projects that are not multiseries this will be a null.}

\item{\code{shapExplanationsMetadata}}{Array contains prediction explanation metadata.}

\item{\code{timestamp}}{The timestamp of this row in the prediction dataset.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionProxyData$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionProxyData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionProxyData$fromJSON(PredictionProxyDataJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionProxyDataJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionProxyData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
