% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_remote_event_data.R
\docType{class}
\name{PredictionRemoteEventData}
\alias{PredictionRemoteEventData}
\title{PredictionRemoteEventData}
\format{
An \code{R6Class} generator object
}
\description{
PredictionRemoteEventData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{errorUnderscorecode}}{character \link{optional} The error code if any.}

\item{\code{modelUnderscoreid}}{character \link{optional} The identifier of the model.}

\item{\code{responseUnderscorebody}}{character The response body message of the prediction event.}

\item{\code{statusUnderscorecode}}{integer The response status code of the prediction event.}

\item{\code{userUnderscoreid}}{character \link{optional} The identifier of the user that accesses the deployment.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PredictionRemoteEventData$new()}}
\item \href{#method-validate}{\code{PredictionRemoteEventData$validate()}}
\item \href{#method-toJSON}{\code{PredictionRemoteEventData$toJSON()}}
\item \href{#method-fromJSON}{\code{PredictionRemoteEventData$fromJSON()}}
\item \href{#method-clone}{\code{PredictionRemoteEventData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionRemoteEventData$new(
  responseUnderscorebody = NULL,
  statusUnderscorecode = NULL,
  errorUnderscorecode = NULL,
  modelUnderscoreid = NULL,
  userUnderscoreid = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{responseUnderscorebody}}{The response body message of the prediction event.}

\item{\code{statusUnderscorecode}}{The response status code of the prediction event.}

\item{\code{errorUnderscorecode}}{The error code if any.}

\item{\code{modelUnderscoreid}}{The identifier of the model.}

\item{\code{userUnderscoreid}}{The identifier of the user that accesses the deployment.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionRemoteEventData$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionRemoteEventData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionRemoteEventData$fromJSON(
  PredictionRemoteEventDataJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionRemoteEventDataJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionRemoteEventData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
