% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_result_response.R
\docType{class}
\name{PredictionResultResponse}
\alias{PredictionResultResponse}
\title{PredictionResultResponse}
\format{
An \code{R6Class} generator object
}
\description{
PredictionResultResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual}}{\link{OneOfintegernumberstring} Actual value of the prediction}

\item{\code{associationId}}{character Association id of the prediction, which is used to associate a prediction with its actual value and calculate accuracy.}

\item{\code{forecastDistance}}{integer \link{optional} The number of time units this prediction is away from the forecastPoint. The unit of time is determined by the timeUnit of the datetime partition column. For time series models only.}

\item{\code{modelId}}{character ID of the model used to make these predictions}

\item{\code{predicted}}{\link{OneOfintegernumberstring} Predicted value of the prediction}

\item{\code{timestamp}}{character When the prediction was made.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PredictionResultResponse$new()}}
\item \href{#method-validate}{\code{PredictionResultResponse$validate()}}
\item \href{#method-toJSON}{\code{PredictionResultResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{PredictionResultResponse$fromJSON()}}
\item \href{#method-clone}{\code{PredictionResultResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionResultResponse$new(
  actual = NULL,
  associationId = NULL,
  modelId = NULL,
  predicted = NULL,
  timestamp = NULL,
  forecastDistance = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{actual}}{Actual value of the prediction}

\item{\code{associationId}}{Association id of the prediction, which is used to associate a prediction with its actual value and calculate accuracy.}

\item{\code{modelId}}{ID of the model used to make these predictions}

\item{\code{predicted}}{Predicted value of the prediction}

\item{\code{timestamp}}{When the prediction was made.}

\item{\code{forecastDistance}}{The number of time units this prediction is away from the forecastPoint. The unit of time is determined by the timeUnit of the datetime partition column. For time series models only.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionResultResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionResultResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionResultResponse$fromJSON(
  PredictionResultResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionResultResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionResultResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
