% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_retrieve_response.R
\docType{class}
\name{PredictionRetrieveResponse}
\alias{PredictionRetrieveResponse}
\title{PredictionRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
PredictionRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actualValueColumn}}{character \link{optional} For time series unsupervised projects only. Will be present only if the prediction dataset has an actual value column. The name of the column with actuals that was used to calculate the scores and insights.}

\item{\code{explanationAlgorithm}}{character \link{optional} The selected algorithm to use for prediction explanations. At present, the only acceptable value is 'shap', which selects the SHapley Additive exPlanations (SHAP) explainer. Defaults to null (no prediction explanations).}

\item{\code{featureDerivationWindowCounts}}{integer \link{optional} For time series projects with partial history only. Indicates how many points were used during feature derivation in feature derivation window.}

\item{\code{includesPredictionIntervals}}{character \link{optional} For time series projects only. Indicates if prediction intervals will be part of the response. Defaults to False.}

\item{\code{maxExplanations}}{integer \link{optional} The maximum number of prediction explanations values to be returned with each row in the `predictions` json array. Null indicates 'no limit'. Will be present only if `explanationAlgorithm` was set.}

\item{\code{positiveClass}}{\link{OneOfstringintegernumber} For binary classification, the class of the target deemed the positive class. For all other project types this field will be null.}

\item{\code{predictionIntervalsSize}}{integer \link{optional} For time series projects only. Will be present only if `includePredictionIntervals` is True. Indicates the percentile used for prediction intervals calculation. Defaults to 80.}

\item{\code{predictions}}{list( \link{PredictionObject} ) The json array of predictions. The predictions in the response will have slightly different formats, depending on the project type.}

\item{\code{shapBaseValue}}{numeric \link{optional} Will be present only if `explanationAlgorithm` = 'shap'. The model's average prediction over the training data. SHAP values are deviations from the base value.}

\item{\code{shapWarnings}}{list( \link{ShapWarningValues} ) \link{optional} Will be present if `explanationAlgorithm` was set to `shap` and there were additivity failures during SHAP values calculation.}

\item{\code{task}}{character The prediction task.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PredictionRetrieveResponse$new()}}
\item \href{#method-validate}{\code{PredictionRetrieveResponse$validate()}}
\item \href{#method-toJSON}{\code{PredictionRetrieveResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{PredictionRetrieveResponse$fromJSON()}}
\item \href{#method-clone}{\code{PredictionRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionRetrieveResponse$new(
  positiveClass = NULL,
  predictions = NULL,
  task = NULL,
  actualValueColumn = NULL,
  explanationAlgorithm = NULL,
  featureDerivationWindowCounts = NULL,
  includesPredictionIntervals = NULL,
  maxExplanations = NULL,
  predictionIntervalsSize = NULL,
  shapBaseValue = NULL,
  shapWarnings = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{positiveClass}}{For binary classification, the class of the target deemed the positive class. For all other project types this field will be null.}

\item{\code{predictions}}{The json array of predictions. The predictions in the response will have slightly different formats, depending on the project type.}

\item{\code{task}}{The prediction task.}

\item{\code{actualValueColumn}}{For time series unsupervised projects only. Will be present only if the prediction dataset has an actual value column. The name of the column with actuals that was used to calculate the scores and insights.}

\item{\code{explanationAlgorithm}}{The selected algorithm to use for prediction explanations. At present, the only acceptable value is 'shap', which selects the SHapley Additive exPlanations (SHAP) explainer. Defaults to null (no prediction explanations).}

\item{\code{featureDerivationWindowCounts}}{For time series projects with partial history only. Indicates how many points were used during feature derivation in feature derivation window.}

\item{\code{includesPredictionIntervals}}{For time series projects only. Indicates if prediction intervals will be part of the response. Defaults to False.}

\item{\code{maxExplanations}}{The maximum number of prediction explanations values to be returned with each row in the `predictions` json array. Null indicates 'no limit'. Will be present only if `explanationAlgorithm` was set.}

\item{\code{predictionIntervalsSize}}{For time series projects only. Will be present only if `includePredictionIntervals` is True. Indicates the percentile used for prediction intervals calculation. Defaults to 80.}

\item{\code{shapBaseValue}}{Will be present only if `explanationAlgorithm` = 'shap'. The model's average prediction over the training data. SHAP values are deviations from the base value.}

\item{\code{shapWarnings}}{Will be present if `explanationAlgorithm` was set to `shap` and there were additivity failures during SHAP values calculation.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionRetrieveResponse$fromJSON(
  PredictionRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
