% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_server_response.R
\docType{class}
\name{PredictionServerResponse}
\alias{PredictionServerResponse}
\title{PredictionServerResponse}
\format{
An \code{R6Class} generator object
}
\description{
PredictionServerResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{batchPredictions}}{\link{BatchPredictionStatus}}

\item{\code{datarobotMinuskey}}{character The `datarobot-key` header used in requests to this prediction server.}

\item{\code{id}}{character ID of the prediction server.}

\item{\code{url}}{character URL of the prediction server.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictionServerResponse-new}{\code{PredictionServerResponse$new()}}
\item \href{#method-PredictionServerResponse-validate}{\code{PredictionServerResponse$validate()}}
\item \href{#method-PredictionServerResponse-toJSON}{\code{PredictionServerResponse$toJSON()}}
\item \href{#method-PredictionServerResponse-fromJSON}{\code{PredictionServerResponse$fromJSON()}}
\item \href{#method-PredictionServerResponse-clone}{\code{PredictionServerResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionServerResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionServerResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionServerResponse$new(
  batchPredictions = NULL,
  datarobotMinuskey = NULL,
  id = NULL,
  url = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datarobotMinuskey}}{The `datarobot-key` header used in requests to this prediction server.}

\item{\code{id}}{ID of the prediction server.}

\item{\code{url}}{URL of the prediction server.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionServerResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionServerResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionServerResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionServerResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionServerResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionServerResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionServerResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionServerResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionServerResponse$fromJSON(
  PredictionServerResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionServerResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionServerResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionServerResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionServerResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
