% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_usage_object_response.R
\docType{class}
\name{PredictionUsageObjectResponse}
\alias{PredictionUsageObjectResponse}
\title{PredictionUsageObjectResponse}
\format{
An \code{R6Class} generator object
}
\description{
PredictionUsageObjectResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{avgExecutionTime}}{numeric The average predictions execution time during `timestamp` hour}

\item{\code{blenderModelTypes}}{character \link{optional} The type of blender model.}

\item{\code{datasetId}}{character \link{optional} The id of the dataset.}

\item{\code{deploymentId}}{character \link{optional} The id of the deployment.}

\item{\code{deploymentType}}{character \link{optional} The type of deployment}

\item{\code{groupId}}{character \link{optional} The id of the group.}

\item{\code{groupName}}{character \link{optional} The name of the group.}

\item{\code{modelId}}{character \link{optional} The id of the model.}

\item{\code{modelType}}{character \link{optional} The type of the model.}

\item{\code{orgId}}{character \link{optional} The id of the organization.}

\item{\code{orgName}}{character \link{optional} The name of the organization.}

\item{\code{predictionExplanations}}{character If it contains prediction explanations.}

\item{\code{predictionMethod}}{character The prediction method that was used.}

\item{\code{predictionRowsCount}}{integer Predictions rows count during `timestamp` hour.}

\item{\code{projectId}}{character \link{optional} The id of the project.}

\item{\code{projectType}}{character \link{optional} The type of the project.}

\item{\code{pythonVersion}}{character \link{optional} The python version used to build models in the project}

\item{\code{recommendedModel}}{character \link{optional} If it is a recommended model.}

\item{\code{requestsCount}}{integer Request count during `timestamp` hour.}

\item{\code{serverErrorCount}}{integer Server error count during `timestamp` hour.}

\item{\code{timestamp}}{character The timestamp of the event.}

\item{\code{userErrorCount}}{integer User error count during `timestamp` hour.}

\item{\code{userId}}{character The id of the user.}

\item{\code{userProjectRole}}{character \link{optional} The role of the user to the related project.}

\item{\code{username}}{character \link{optional} The username of the user. Will be empty if ``includeIdentifyingFields`` is True.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PredictionUsageObjectResponse$new()}}
\item \href{#method-validate}{\code{PredictionUsageObjectResponse$validate()}}
\item \href{#method-toJSON}{\code{PredictionUsageObjectResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{PredictionUsageObjectResponse$fromJSON()}}
\item \href{#method-clone}{\code{PredictionUsageObjectResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionUsageObjectResponse$new(
  avgExecutionTime = NULL,
  predictionExplanations = NULL,
  predictionMethod = NULL,
  predictionRowsCount = NULL,
  requestsCount = NULL,
  serverErrorCount = NULL,
  timestamp = NULL,
  userErrorCount = NULL,
  userId = NULL,
  blenderModelTypes = NULL,
  datasetId = NULL,
  deploymentId = NULL,
  deploymentType = NULL,
  groupId = NULL,
  groupName = NULL,
  modelId = NULL,
  modelType = NULL,
  orgId = NULL,
  orgName = NULL,
  projectId = NULL,
  projectType = NULL,
  pythonVersion = NULL,
  recommendedModel = NULL,
  userProjectRole = NULL,
  username = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{avgExecutionTime}}{The average predictions execution time during `timestamp` hour}

\item{\code{predictionExplanations}}{If it contains prediction explanations.}

\item{\code{predictionMethod}}{The prediction method that was used.}

\item{\code{predictionRowsCount}}{Predictions rows count during `timestamp` hour.}

\item{\code{requestsCount}}{Request count during `timestamp` hour.}

\item{\code{serverErrorCount}}{Server error count during `timestamp` hour.}

\item{\code{timestamp}}{The timestamp of the event.}

\item{\code{userErrorCount}}{User error count during `timestamp` hour.}

\item{\code{userId}}{The id of the user.}

\item{\code{blenderModelTypes}}{The type of blender model.}

\item{\code{datasetId}}{The id of the dataset.}

\item{\code{deploymentId}}{The id of the deployment.}

\item{\code{deploymentType}}{The type of deployment}

\item{\code{groupId}}{The id of the group.}

\item{\code{groupName}}{The name of the group.}

\item{\code{modelId}}{The id of the model.}

\item{\code{modelType}}{The type of the model.}

\item{\code{orgId}}{The id of the organization.}

\item{\code{orgName}}{The name of the organization.}

\item{\code{projectId}}{The id of the project.}

\item{\code{projectType}}{The type of the project.}

\item{\code{pythonVersion}}{The python version used to build models in the project}

\item{\code{recommendedModel}}{If it is a recommended model.}

\item{\code{userProjectRole}}{The role of the user to the related project.}

\item{\code{username}}{The username of the user. Will be empty if ``includeIdentifyingFields`` is True.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionUsageObjectResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionUsageObjectResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionUsageObjectResponse$fromJSON(
  PredictionUsageObjectResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionUsageObjectResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionUsageObjectResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
