% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_values.R
\docType{class}
\name{PredictionValues}
\alias{PredictionValues}
\title{PredictionValues}
\format{
An \code{R6Class} generator object
}
\description{
PredictionValues Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{label}}{\link{OneOfstringnumber} In the case of classification, The label of the class.}

\item{\code{value}}{numeric In the case of classification, the predicted probability of the class identified by the label.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictionValues-new}{\code{PredictionValues$new()}}
\item \href{#method-PredictionValues-validate}{\code{PredictionValues$validate()}}
\item \href{#method-PredictionValues-toJSON}{\code{PredictionValues$toJSON()}}
\item \href{#method-PredictionValues-fromJSON}{\code{PredictionValues$fromJSON()}}
\item \href{#method-PredictionValues-clone}{\code{PredictionValues$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionValues-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionValues-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionValues$new(label = NULL, value = NULL, validateParams = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{In the case of classification, The label of the class.}

\item{\code{value}}{In the case of classification, the predicted probability of the class identified by the label.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionValues-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionValues-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionValues$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionValues-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionValues-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionValues$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionValues-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionValues-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionValues$fromJSON(PredictionValuesJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionValuesJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionValues-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionValues-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionValues$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
