% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_job_details_response.R
\docType{class}
\name{PredictJobDetailsResponse}
\alias{PredictJobDetailsResponse}
\title{PredictJobDetailsResponse}
\format{
An \code{R6Class} generator object
}
\description{
PredictJobDetailsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character the job ID of the job}

\item{\code{isBlocked}}{character True if a job is waiting for its dependencies to be resolved first.}

\item{\code{message}}{character An optional message about the job}

\item{\code{modelId}}{character The ID of the model}

\item{\code{projectId}}{character the project the job belongs to}

\item{\code{status}}{character the status of the job}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PredictJobDetailsResponse$new()}}
\item \href{#method-validate}{\code{PredictJobDetailsResponse$validate()}}
\item \href{#method-toJSON}{\code{PredictJobDetailsResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{PredictJobDetailsResponse$fromJSON()}}
\item \href{#method-clone}{\code{PredictJobDetailsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictJobDetailsResponse$new(
  id = NULL,
  isBlocked = NULL,
  message = NULL,
  modelId = NULL,
  projectId = NULL,
  status = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the job ID of the job}

\item{\code{isBlocked}}{True if a job is waiting for its dependencies to be resolved first.}

\item{\code{message}}{An optional message about the job}

\item{\code{modelId}}{The ID of the model}

\item{\code{projectId}}{the project the job belongs to}

\item{\code{status}}{the status of the job}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictJobDetailsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictJobDetailsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictJobDetailsResponse$fromJSON(
  PredictJobDetailsResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictJobDetailsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictJobDetailsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
