% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preloaded_calendar.R
\docType{class}
\name{PreloadedCalendar}
\alias{PreloadedCalendar}
\title{PreloadedCalendar}
\format{
An \code{R6Class} generator object
}
\description{
PreloadedCalendar Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{countryCode}}{character Code of the country for which holidays should be generated. Code needs to be uppercase and should belong to the list of countries which can be retrieved via :http:get:`/api/v2/calendarCountryCodes/`}

\item{\code{endDate}}{character Last date of the range of dates for which holidays are generated.}

\item{\code{startDate}}{character First date of the range of dates for which holidays are generated.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PreloadedCalendar$new()}}
\item \href{#method-validate}{\code{PreloadedCalendar$validate()}}
\item \href{#method-toJSON}{\code{PreloadedCalendar$toJSON()}}
\item \href{#method-fromJSON}{\code{PreloadedCalendar$fromJSON()}}
\item \href{#method-clone}{\code{PreloadedCalendar$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreloadedCalendar$new(
  countryCode = NULL,
  endDate = NULL,
  startDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{countryCode}}{Code of the country for which holidays should be generated. Code needs to be uppercase and should belong to the list of countries which can be retrieved via :http:get:`/api/v2/calendarCountryCodes/`}

\item{\code{endDate}}{Last date of the range of dates for which holidays are generated.}

\item{\code{startDate}}{First date of the range of dates for which holidays are generated.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreloadedCalendar$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreloadedCalendar$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreloadedCalendar$fromJSON(PreloadedCalendarJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PreloadedCalendarJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreloadedCalendar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
