% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preloaded_calendar_country_code_record.R
\docType{class}
\name{PreloadedCalendarCountryCodeRecord}
\alias{PreloadedCalendarCountryCodeRecord}
\title{PreloadedCalendarCountryCodeRecord}
\format{
An \code{R6Class} generator object
}
\description{
PreloadedCalendarCountryCodeRecord Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{code}}{character Country code that can be used for calendars generation.}

\item{\code{name}}{character Full name of the country.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PreloadedCalendarCountryCodeRecord-new}{\code{PreloadedCalendarCountryCodeRecord$new()}}
\item \href{#method-PreloadedCalendarCountryCodeRecord-validate}{\code{PreloadedCalendarCountryCodeRecord$validate()}}
\item \href{#method-PreloadedCalendarCountryCodeRecord-toJSON}{\code{PreloadedCalendarCountryCodeRecord$toJSON()}}
\item \href{#method-PreloadedCalendarCountryCodeRecord-fromJSON}{\code{PreloadedCalendarCountryCodeRecord$fromJSON()}}
\item \href{#method-PreloadedCalendarCountryCodeRecord-clone}{\code{PreloadedCalendarCountryCodeRecord$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PreloadedCalendarCountryCodeRecord-new"></a>}}
\if{latex}{\out{\hypertarget{method-PreloadedCalendarCountryCodeRecord-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreloadedCalendarCountryCodeRecord$new(
  code = NULL,
  name = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{Country code that can be used for calendars generation.}

\item{\code{name}}{Full name of the country.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PreloadedCalendarCountryCodeRecord-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PreloadedCalendarCountryCodeRecord-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreloadedCalendarCountryCodeRecord$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PreloadedCalendarCountryCodeRecord-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PreloadedCalendarCountryCodeRecord-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreloadedCalendarCountryCodeRecord$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PreloadedCalendarCountryCodeRecord-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PreloadedCalendarCountryCodeRecord-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreloadedCalendarCountryCodeRecord$fromJSON(
  PreloadedCalendarCountryCodeRecordJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PreloadedCalendarCountryCodeRecordJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PreloadedCalendarCountryCodeRecord-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PreloadedCalendarCountryCodeRecord-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreloadedCalendarCountryCodeRecord$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
