% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime_info_retrieve_response.R
\docType{class}
\name{PrimeInfoRetrieveResponse}
\alias{PrimeInfoRetrieveResponse}
\title{PrimeInfoRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
PrimeInfoRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{canMakePrime}}{character Indicating whether the requested model is a valid input for creating a Prime mode}

\item{\code{message}}{character May contain details about why a model is not eligible for DataRobotPrime}

\item{\code{messageId}}{integer An error code representing the reason the model cannot be approximated with DataRobot Prime; 0 for eligible models}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PrimeInfoRetrieveResponse-new}{\code{PrimeInfoRetrieveResponse$new()}}
\item \href{#method-PrimeInfoRetrieveResponse-validate}{\code{PrimeInfoRetrieveResponse$validate()}}
\item \href{#method-PrimeInfoRetrieveResponse-toJSON}{\code{PrimeInfoRetrieveResponse$toJSON()}}
\item \href{#method-PrimeInfoRetrieveResponse-fromJSON}{\code{PrimeInfoRetrieveResponse$fromJSON()}}
\item \href{#method-PrimeInfoRetrieveResponse-clone}{\code{PrimeInfoRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimeInfoRetrieveResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-PrimeInfoRetrieveResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimeInfoRetrieveResponse$new(
  canMakePrime = NULL,
  message = NULL,
  messageId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{canMakePrime}}{Indicating whether the requested model is a valid input for creating a Prime mode}

\item{\code{message}}{May contain details about why a model is not eligible for DataRobotPrime}

\item{\code{messageId}}{An error code representing the reason the model cannot be approximated with DataRobot Prime; 0 for eligible models}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimeInfoRetrieveResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PrimeInfoRetrieveResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimeInfoRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimeInfoRetrieveResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PrimeInfoRetrieveResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimeInfoRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimeInfoRetrieveResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PrimeInfoRetrieveResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimeInfoRetrieveResponse$fromJSON(
  PrimeInfoRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PrimeInfoRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimeInfoRetrieveResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PrimeInfoRetrieveResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimeInfoRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
