% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime_rulesets_list_response.R
\docType{class}
\name{PrimeRulesetsListResponse}
\alias{PrimeRulesetsListResponse}
\title{PrimeRulesetsListResponse}
\format{
An \code{R6Class} generator object
}
\description{
PrimeRulesetsListResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{modelId}}{character the ID of the Prime model using this ruleset (if it exists) or null}

\item{\code{parentModelId}}{character ID of the model this ruleset approximates}

\item{\code{projectId}}{character the project this ruleset belongs to}

\item{\code{ruleCount}}{integer the number of rules used by this ruleset}

\item{\code{rulesetId}}{integer the ID of the ruleset}

\item{\code{score}}{numeric the validation score of the ruleset}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PrimeRulesetsListResponse-new}{\code{PrimeRulesetsListResponse$new()}}
\item \href{#method-PrimeRulesetsListResponse-validate}{\code{PrimeRulesetsListResponse$validate()}}
\item \href{#method-PrimeRulesetsListResponse-toJSON}{\code{PrimeRulesetsListResponse$toJSON()}}
\item \href{#method-PrimeRulesetsListResponse-fromJSON}{\code{PrimeRulesetsListResponse$fromJSON()}}
\item \href{#method-PrimeRulesetsListResponse-clone}{\code{PrimeRulesetsListResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimeRulesetsListResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-PrimeRulesetsListResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimeRulesetsListResponse$new(
  modelId = NULL,
  parentModelId = NULL,
  projectId = NULL,
  ruleCount = NULL,
  rulesetId = NULL,
  score = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{modelId}}{the ID of the Prime model using this ruleset (if it exists) or null}

\item{\code{parentModelId}}{ID of the model this ruleset approximates}

\item{\code{projectId}}{the project this ruleset belongs to}

\item{\code{ruleCount}}{the number of rules used by this ruleset}

\item{\code{rulesetId}}{the ID of the ruleset}

\item{\code{score}}{the validation score of the ruleset}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimeRulesetsListResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PrimeRulesetsListResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimeRulesetsListResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimeRulesetsListResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PrimeRulesetsListResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimeRulesetsListResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimeRulesetsListResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PrimeRulesetsListResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimeRulesetsListResponse$fromJSON(
  PrimeRulesetsListResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PrimeRulesetsListResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimeRulesetsListResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PrimeRulesetsListResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimeRulesetsListResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
