% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_clone.R
\docType{class}
\name{ProjectClone}
\alias{ProjectClone}
\title{ProjectClone}
\format{
An \code{R6Class} generator object
}
\description{
ProjectClone Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{copyOptions}}{character \link{optional} Whether all project options should be copied to the cloned project.}

\item{\code{projectId}}{character The ID of the project to clone.}

\item{\code{projectName}}{character \link{optional} The name of the project to be created.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProjectClone-new}{\code{ProjectClone$new()}}
\item \href{#method-ProjectClone-validate}{\code{ProjectClone$validate()}}
\item \href{#method-ProjectClone-toJSON}{\code{ProjectClone$toJSON()}}
\item \href{#method-ProjectClone-fromJSON}{\code{ProjectClone$fromJSON()}}
\item \href{#method-ProjectClone-clone}{\code{ProjectClone$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectClone-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectClone-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectClone$new(
  projectId = NULL,
  copyOptions = FALSE,
  projectName = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectId}}{The ID of the project to clone.}

\item{\code{copyOptions}}{Whether all project options should be copied to the cloned project.}

\item{\code{projectName}}{The name of the project to be created.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectClone-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectClone-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectClone$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectClone-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectClone-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectClone$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectClone-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectClone-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectClone$fromJSON(ProjectCloneJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ProjectCloneJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectClone-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectClone-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectClone$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
