% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_create.R
\docType{class}
\name{ProjectCreate}
\alias{ProjectCreate}
\title{ProjectCreate}
\format{
An \code{R6Class} generator object
}
\description{
ProjectCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{credentialData}}{\link{OneOfBasicCredentialsS3CredentialsOAuthCredentials} \link{optional} The credentials to authenticate with the database, to be used instead of credential ID. Can only be used along with datasetId or dataSourceId.}

\item{\code{credentialId}}{character \link{optional} The ID of the set of credentials to authenticate with the database. Can only be used along with datasetId or dataSourceId.}

\item{\code{dataSourceId}}{character \link{optional} Identifier for the data source to retrieve.}

\item{\code{datasetId}}{character \link{optional} The ID of the dataset entry to use for the project dataset.}

\item{\code{datasetVersionId}}{character \link{optional} Only used when also providing a datasetId, and specifies the the ID of the dataset version to use for the project dataset. If not specified, the latest version associated with the dataset ID is used.}

\item{\code{password}}{character \link{optional} The password (in cleartext) for database authentication. The password will be encrypted on the server side as part of the HTTP request and never saved or stored. Can only be used along with datasetId or dataSourceId. DEPRECATED: please use credentialId or credentialData instead.}

\item{\code{projectName}}{character \link{optional} The name of the project to be created. If not specified, 'Untitled Project' will be used for database connections and file name will be used as the project name.}

\item{\code{url}}{character \link{optional} The URL to download the dataset used to create the project.}

\item{\code{useKerberos}}{character \link{optional} If true, use Kerberos authentication for database authentication. Default is false. Can only be used along with datasetId or dataSourceId.}

\item{\code{user}}{character \link{optional} The username for database authentication. Can only be used along with datasetId or dataSourceId. DEPRECATED: please use credentialId or credentialData instead.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProjectCreate$new()}}
\item \href{#method-validate}{\code{ProjectCreate$validate()}}
\item \href{#method-toJSON}{\code{ProjectCreate$toJSON()}}
\item \href{#method-fromJSON}{\code{ProjectCreate$fromJSON()}}
\item \href{#method-clone}{\code{ProjectCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectCreate$new(
  credentialData = NULL,
  credentialId = NULL,
  dataSourceId = NULL,
  datasetId = NULL,
  datasetVersionId = NULL,
  password = NULL,
  projectName = NULL,
  url = NULL,
  useKerberos = NULL,
  user = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{credentialData}}{The credentials to authenticate with the database, to be used instead of credential ID. Can only be used along with datasetId or dataSourceId.}

\item{\code{credentialId}}{The ID of the set of credentials to authenticate with the database. Can only be used along with datasetId or dataSourceId.}

\item{\code{dataSourceId}}{Identifier for the data source to retrieve.}

\item{\code{datasetId}}{The ID of the dataset entry to use for the project dataset.}

\item{\code{datasetVersionId}}{Only used when also providing a datasetId, and specifies the the ID of the dataset version to use for the project dataset. If not specified, the latest version associated with the dataset ID is used.}

\item{\code{password}}{The password (in cleartext) for database authentication. The password will be encrypted on the server side as part of the HTTP request and never saved or stored. Can only be used along with datasetId or dataSourceId. DEPRECATED: please use credentialId or credentialData instead.}

\item{\code{projectName}}{The name of the project to be created. If not specified, 'Untitled Project' will be used for database connections and file name will be used as the project name.}

\item{\code{url}}{The URL to download the dataset used to create the project.}

\item{\code{useKerberos}}{If true, use Kerberos authentication for database authentication. Default is false. Can only be used along with datasetId or dataSourceId.}

\item{\code{user}}{The username for database authentication. Can only be used along with datasetId or dataSourceId. DEPRECATED: please use credentialId or credentialData instead.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectCreate$fromJSON(ProjectCreateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ProjectCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
