% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_details_response.R
\docType{class}
\name{ProjectDetailsResponse}
\alias{ProjectDetailsResponse}
\title{ProjectDetailsResponse}
\format{
An \code{R6Class} generator object
}
\description{
ProjectDetailsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{advancedOptions}}{\link{ProjectAdvancedOptionsResponse}}

\item{\code{autopilotClusterList}}{list( integer ) \link{optional} Optional. A list of integers where each value will be used as the number of clusters in Autopilot model(s) for unsupervised clustering projects. Cannot be specified unless unsupervisedMode is true and unsupervisedType is set to 'clustering'.}

\item{\code{autopilotMode}}{integer The current autopilot mode, 0 for full autopilot, 2 for manual mode, 3 for quick mode, 4 for comprehensive mode}

\item{\code{created}}{character The time of project creation.}

\item{\code{featureEngineeringPredictionPoint}}{character \link{optional} The date column to be used as prediction point for time-based feature engineering.}

\item{\code{fileName}}{character The name of the dataset used to create the project.}

\item{\code{holdoutUnlocked}}{character whether the holdout has been unlocked}

\item{\code{id}}{character The ID of a project.}

\item{\code{maxClusters}}{integer \link{optional} Only valid when unsupervisedMode is True and unsupervisedType is 'clustering'. The maximum number of clusters allowed when training clustering models. If specified cannot be exceed the number of rows in a project's dataset divided by 50 and must be less than or equal to `minClusters`. If unsupervisedMode is True and unsupervisedType is 'clustering' then defaults to the number of rows in the project's dataset divided by 50 or 100 if that number of greater than 100.}

\item{\code{maxTrainPct}}{numeric the maximum percentage of the dataset that can be used to successfully train a model without going into the validation data.}

\item{\code{maxTrainRows}}{integer the maximum number of rows of the dataset that can be used to successfully train a model without going into the validation data}

\item{\code{metric}}{character the metric used to select the best-performing models.}

\item{\code{minClusters}}{integer \link{optional} Only valid when unsupervisedMode is True and unsupervisedType is 'clustering'. The minimum number of clusters allowed when training clustering models. If specified cannot be exceed the number of rows in a project's dataset divided by 50 and must be less than or equal to `maxClusters`. If unsupervisedMode is True and  unsupervisedType is 'clustering' then defaults to 2.}

\item{\code{partition}}{\link{ProjectPartitionResponse}}

\item{\code{positiveClass}}{numeric if the project uses binary classification, the class designated to be the positive class.  Otherwise, null.}

\item{\code{projectName}}{character The name of a project.}

\item{\code{stage}}{character the stage of the project - if modeling, then the target is successfully set, and modeling or predictions can proceed.}

\item{\code{target}}{character the target of the project, null if project is unsupervised.}

\item{\code{targetType}}{character The target type of the project.}

\item{\code{unsupervisedMode}}{character indicates whether a project is unsupervised.}

\item{\code{unsupervisedType}}{character \link{optional} Only valid when unsupervisedMode is True. The type of unsupervised project, anomaly or clustering. If unsupervisedMode, defaults to 'anomaly'.}

\item{\code{useFeatureDiscovery}}{character A boolean value indicating whether a feature discovery project was created as opposed to a regular project.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProjectDetailsResponse-new}{\code{ProjectDetailsResponse$new()}}
\item \href{#method-ProjectDetailsResponse-validate}{\code{ProjectDetailsResponse$validate()}}
\item \href{#method-ProjectDetailsResponse-toJSON}{\code{ProjectDetailsResponse$toJSON()}}
\item \href{#method-ProjectDetailsResponse-fromJSON}{\code{ProjectDetailsResponse$fromJSON()}}
\item \href{#method-ProjectDetailsResponse-clone}{\code{ProjectDetailsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectDetailsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectDetailsResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectDetailsResponse$new(
  advancedOptions = NULL,
  autopilotMode = NULL,
  created = NULL,
  fileName = NULL,
  holdoutUnlocked = NULL,
  id = NULL,
  maxTrainPct = NULL,
  maxTrainRows = NULL,
  metric = NULL,
  partition = NULL,
  positiveClass = NULL,
  projectName = NULL,
  stage = NULL,
  target = NULL,
  targetType = NULL,
  unsupervisedMode = NULL,
  useFeatureDiscovery = NULL,
  autopilotClusterList = NULL,
  featureEngineeringPredictionPoint = NULL,
  maxClusters = NULL,
  minClusters = NULL,
  unsupervisedType = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{autopilotMode}}{The current autopilot mode, 0 for full autopilot, 2 for manual mode, 3 for quick mode, 4 for comprehensive mode}

\item{\code{created}}{The time of project creation.}

\item{\code{fileName}}{The name of the dataset used to create the project.}

\item{\code{holdoutUnlocked}}{whether the holdout has been unlocked}

\item{\code{id}}{The ID of a project.}

\item{\code{maxTrainPct}}{the maximum percentage of the dataset that can be used to successfully train a model without going into the validation data.}

\item{\code{maxTrainRows}}{the maximum number of rows of the dataset that can be used to successfully train a model without going into the validation data}

\item{\code{metric}}{the metric used to select the best-performing models.}

\item{\code{positiveClass}}{if the project uses binary classification, the class designated to be the positive class.  Otherwise, null.}

\item{\code{projectName}}{The name of a project.}

\item{\code{stage}}{the stage of the project - if modeling, then the target is successfully set, and modeling or predictions can proceed.}

\item{\code{target}}{the target of the project, null if project is unsupervised.}

\item{\code{targetType}}{The target type of the project.}

\item{\code{unsupervisedMode}}{indicates whether a project is unsupervised.}

\item{\code{useFeatureDiscovery}}{A boolean value indicating whether a feature discovery project was created as opposed to a regular project.}

\item{\code{autopilotClusterList}}{Optional. A list of integers where each value will be used as the number of clusters in Autopilot model(s) for unsupervised clustering projects. Cannot be specified unless unsupervisedMode is true and unsupervisedType is set to 'clustering'.}

\item{\code{featureEngineeringPredictionPoint}}{The date column to be used as prediction point for time-based feature engineering.}

\item{\code{maxClusters}}{Only valid when unsupervisedMode is True and unsupervisedType is 'clustering'. The maximum number of clusters allowed when training clustering models. If specified cannot be exceed the number of rows in a project's dataset divided by 50 and must be less than or equal to `minClusters`. If unsupervisedMode is True and unsupervisedType is 'clustering' then defaults to the number of rows in the project's dataset divided by 50 or 100 if that number of greater than 100.}

\item{\code{minClusters}}{Only valid when unsupervisedMode is True and unsupervisedType is 'clustering'. The minimum number of clusters allowed when training clustering models. If specified cannot be exceed the number of rows in a project's dataset divided by 50 and must be less than or equal to `maxClusters`. If unsupervisedMode is True and  unsupervisedType is 'clustering' then defaults to 2.}

\item{\code{unsupervisedType}}{Only valid when unsupervisedMode is True. The type of unsupervised project, anomaly or clustering. If unsupervisedMode, defaults to 'anomaly'.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectDetailsResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectDetailsResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectDetailsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectDetailsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectDetailsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectDetailsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectDetailsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectDetailsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectDetailsResponse$fromJSON(
  ProjectDetailsResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ProjectDetailsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectDetailsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectDetailsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectDetailsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
