% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_feature_response.R
\docType{class}
\name{ProjectFeatureResponse}
\alias{ProjectFeatureResponse}
\title{ProjectFeatureResponse}
\format{
An \code{R6Class} generator object
}
\description{
ProjectFeatureResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dataQualities}}{character \link{optional} Data Quality Status}

\item{\code{dateFormat}}{character the date format string for how this feature was interpreted (or null if not a date feature).  If not null, it will be compatible with https://docs.python.org/2/library/time.html#time.strftime .}

\item{\code{featureLineageId}}{character id of a lineage for automatically generated features.}

\item{\code{featureType}}{character Feature type.}

\item{\code{id}}{integer the feature ID. (Note: Throughout the API, features are specified using their names, not this ID.)}

\item{\code{importance}}{numeric numeric measure of the strength of relationship between the feature and target (independent of any model or other features)}

\item{\code{isRestoredAfterReduction}}{character \link{optional} Whether feature is restored after feature reduction}

\item{\code{isZeroInflated}}{character \link{optional} Whether feature has an excessive number of zeros}

\item{\code{keySummary}}{\link{OneOfFeatureKeySummaryResponseValidatorSummarizedCategoricalarray} \link{optional} Per key summaries for Summarized Categorical or Multicategorical columns}

\item{\code{lowInformation}}{character whether feature has too few values to be informative}

\item{\code{max}}{\link{OneOfstringnumber} maximum value of the EDA sample of the feature.}

\item{\code{mean}}{\link{OneOfstringnumber} arithmetic mean of the EDA sample of the feature.}

\item{\code{median}}{\link{OneOfstringnumber} median of the EDA sample of the feature.}

\item{\code{min}}{\link{OneOfstringnumber} minimum value of the EDA sample of the feature.}

\item{\code{multilabelInsights}}{\link{MultilabelInsightsResponse} \link{optional}}

\item{\code{naCount}}{integer number of missing values}

\item{\code{name}}{character feature name}

\item{\code{parentFeatureNames}}{list( character ) \link{optional} an array of string feature names indicating which features in the input data were used to create this feature if the feature is a transformation.}

\item{\code{projectId}}{character the ID of the project the feature belongs to}

\item{\code{stdDev}}{\link{OneOfstringnumber} standard deviation of EDA sample of the feature.}

\item{\code{targetLeakage}}{character the detected level of risk for target leakage, if any. 'SKIPPED_DETECTION' indicates leakage detection was not run on the feature, 'FALSE' indicates no leakage, 'MODERATE_RISK' indicates a moderate risk of target leakage, and 'HIGH_RISK' indicates a high risk of target leakage.}

\item{\code{targetLeakageReason}}{character descriptive sentence explaining the reason for target leakage.}

\item{\code{timeSeriesEligibilityReason}}{character why the feature is ineligible for time series projects, or 'suitable' if it is eligible.}

\item{\code{timeSeriesEligible}}{character whether this feature can be used as a datetime partitioning feature for time series projects.  Only sufficiently regular date features can be selected as the datetime feature for time series projects.  Always false for non-date features.  Date features that cannot be used in datetime partitioning for a time series project may be eligible for an OTV project, which has less stringent requirements.}

\item{\code{timeStep}}{integer The minimum time step that can be used to specify time series windows.  The units for this value are the ``timeUnit``.  When specifying windows for time series projects, all windows must have durations that are integer multiples of this number.  Only present for date features that are eligible for time series projects and null otherwise.}

\item{\code{timeUnit}}{character the unit for the interval between values of this feature, e.g. DAY, MONTH, HOUR.  When specifying windows for time series projects, the windows are expressed in terms of this unit.  Only present for date features eligible for time series projects, and null otherwise.}

\item{\code{uniqueCount}}{integer \link{optional} number of unique values}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProjectFeatureResponse$new()}}
\item \href{#method-validate}{\code{ProjectFeatureResponse$validate()}}
\item \href{#method-toJSON}{\code{ProjectFeatureResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{ProjectFeatureResponse$fromJSON()}}
\item \href{#method-clone}{\code{ProjectFeatureResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectFeatureResponse$new(
  dateFormat = NULL,
  featureLineageId = NULL,
  featureType = NULL,
  id = NULL,
  importance = NULL,
  lowInformation = NULL,
  max = NULL,
  mean = NULL,
  median = NULL,
  min = NULL,
  naCount = NULL,
  name = NULL,
  projectId = NULL,
  stdDev = NULL,
  targetLeakage = NULL,
  targetLeakageReason = NULL,
  timeSeriesEligibilityReason = NULL,
  timeSeriesEligible = NULL,
  timeStep = NULL,
  timeUnit = NULL,
  dataQualities = NULL,
  isRestoredAfterReduction = NULL,
  isZeroInflated = NULL,
  keySummary = NULL,
  multilabelInsights = NULL,
  parentFeatureNames = NULL,
  uniqueCount = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dateFormat}}{the date format string for how this feature was interpreted (or null if not a date feature).  If not null, it will be compatible with https://docs.python.org/2/library/time.html#time.strftime .}

\item{\code{featureLineageId}}{id of a lineage for automatically generated features.}

\item{\code{featureType}}{Feature type.}

\item{\code{id}}{the feature ID. (Note: Throughout the API, features are specified using their names, not this ID.)}

\item{\code{importance}}{numeric measure of the strength of relationship between the feature and target (independent of any model or other features)}

\item{\code{lowInformation}}{whether feature has too few values to be informative}

\item{\code{max}}{maximum value of the EDA sample of the feature.}

\item{\code{mean}}{arithmetic mean of the EDA sample of the feature.}

\item{\code{median}}{median of the EDA sample of the feature.}

\item{\code{min}}{minimum value of the EDA sample of the feature.}

\item{\code{naCount}}{number of missing values}

\item{\code{name}}{feature name}

\item{\code{projectId}}{the ID of the project the feature belongs to}

\item{\code{stdDev}}{standard deviation of EDA sample of the feature.}

\item{\code{targetLeakage}}{the detected level of risk for target leakage, if any. 'SKIPPED_DETECTION' indicates leakage detection was not run on the feature, 'FALSE' indicates no leakage, 'MODERATE_RISK' indicates a moderate risk of target leakage, and 'HIGH_RISK' indicates a high risk of target leakage.}

\item{\code{targetLeakageReason}}{descriptive sentence explaining the reason for target leakage.}

\item{\code{timeSeriesEligibilityReason}}{why the feature is ineligible for time series projects, or 'suitable' if it is eligible.}

\item{\code{timeSeriesEligible}}{whether this feature can be used as a datetime partitioning feature for time series projects.  Only sufficiently regular date features can be selected as the datetime feature for time series projects.  Always false for non-date features.  Date features that cannot be used in datetime partitioning for a time series project may be eligible for an OTV project, which has less stringent requirements.}

\item{\code{timeStep}}{The minimum time step that can be used to specify time series windows.  The units for this value are the ``timeUnit``.  When specifying windows for time series projects, all windows must have durations that are integer multiples of this number.  Only present for date features that are eligible for time series projects and null otherwise.}

\item{\code{timeUnit}}{the unit for the interval between values of this feature, e.g. DAY, MONTH, HOUR.  When specifying windows for time series projects, the windows are expressed in terms of this unit.  Only present for date features eligible for time series projects, and null otherwise.}

\item{\code{dataQualities}}{Data Quality Status}

\item{\code{isRestoredAfterReduction}}{Whether feature is restored after feature reduction}

\item{\code{isZeroInflated}}{Whether feature has an excessive number of zeros}

\item{\code{keySummary}}{Per key summaries for Summarized Categorical or Multicategorical columns}

\item{\code{parentFeatureNames}}{an array of string feature names indicating which features in the input data were used to create this feature if the feature is a transformation.}

\item{\code{uniqueCount}}{number of unique values}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectFeatureResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectFeatureResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectFeatureResponse$fromJSON(
  ProjectFeatureResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ProjectFeatureResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectFeatureResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
