% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_nuke_job_status_summary_object.R
\docType{class}
\name{ProjectNukeJobStatusSummaryObject}
\alias{ProjectNukeJobStatusSummaryObject}
\title{ProjectNukeJobStatusSummaryObject}
\format{
An \code{R6Class} generator object
}
\description{
ProjectNukeJobStatusSummaryObject Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aborted}}{integer Number of project permadelete jobs with Aborted status.}

\item{\code{completed}}{integer Number of project permadelete jobs with Completed status.}

\item{\code{error}}{integer Number of project permadelete jobs with Error status.}

\item{\code{expired}}{integer Number of project permadelete jobs with Expired status.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProjectNukeJobStatusSummaryObject-new}{\code{ProjectNukeJobStatusSummaryObject$new()}}
\item \href{#method-ProjectNukeJobStatusSummaryObject-validate}{\code{ProjectNukeJobStatusSummaryObject$validate()}}
\item \href{#method-ProjectNukeJobStatusSummaryObject-toJSON}{\code{ProjectNukeJobStatusSummaryObject$toJSON()}}
\item \href{#method-ProjectNukeJobStatusSummaryObject-fromJSON}{\code{ProjectNukeJobStatusSummaryObject$fromJSON()}}
\item \href{#method-ProjectNukeJobStatusSummaryObject-clone}{\code{ProjectNukeJobStatusSummaryObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNukeJobStatusSummaryObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNukeJobStatusSummaryObject-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNukeJobStatusSummaryObject$new(
  aborted = NULL,
  completed = NULL,
  error = NULL,
  expired = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aborted}}{Number of project permadelete jobs with Aborted status.}

\item{\code{completed}}{Number of project permadelete jobs with Completed status.}

\item{\code{error}}{Number of project permadelete jobs with Error status.}

\item{\code{expired}}{Number of project permadelete jobs with Expired status.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNukeJobStatusSummaryObject-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNukeJobStatusSummaryObject-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNukeJobStatusSummaryObject$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNukeJobStatusSummaryObject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNukeJobStatusSummaryObject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNukeJobStatusSummaryObject$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNukeJobStatusSummaryObject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNukeJobStatusSummaryObject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNukeJobStatusSummaryObject$fromJSON(
  ProjectNukeJobStatusSummaryObjectJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ProjectNukeJobStatusSummaryObjectJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNukeJobStatusSummaryObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNukeJobStatusSummaryObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNukeJobStatusSummaryObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
