% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_options.R
\docType{class}
\name{ProjectOptions}
\alias{ProjectOptions}
\title{ProjectOptions}
\format{
An \code{R6Class} generator object
}
\description{
ProjectOptions Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cvMethod}}{character \link{optional} The partitioning method for projects used to build new models.}

\item{\code{holdoutPct}}{numeric \link{optional} The percentage of dataset to assign to holdout set in projects used to build new models.}

\item{\code{metric}}{character \link{optional} The model selection metric in projects used to build new models.}

\item{\code{reps}}{integer \link{optional} The number of cross validation folds to use for projects used to build new models.}

\item{\code{validationPct}}{numeric \link{optional} The percentage of dataset to assign to validation set in projects used to build new models.}

\item{\code{validationType}}{character \link{optional} The validation type for projects used to build new models.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProjectOptions-new}{\code{ProjectOptions$new()}}
\item \href{#method-ProjectOptions-validate}{\code{ProjectOptions$validate()}}
\item \href{#method-ProjectOptions-toJSON}{\code{ProjectOptions$toJSON()}}
\item \href{#method-ProjectOptions-fromJSON}{\code{ProjectOptions$fromJSON()}}
\item \href{#method-ProjectOptions-clone}{\code{ProjectOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectOptions-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectOptions$new(
  cvMethod = NULL,
  holdoutPct = NULL,
  metric = NULL,
  reps = NULL,
  validationPct = NULL,
  validationType = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cvMethod}}{The partitioning method for projects used to build new models.}

\item{\code{holdoutPct}}{The percentage of dataset to assign to holdout set in projects used to build new models.}

\item{\code{metric}}{The model selection metric in projects used to build new models.}

\item{\code{reps}}{The number of cross validation folds to use for projects used to build new models.}

\item{\code{validationPct}}{The percentage of dataset to assign to validation set in projects used to build new models.}

\item{\code{validationType}}{The validation type for projects used to build new models.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectOptions-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectOptions-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectOptions$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectOptions-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectOptions-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectOptions$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectOptions-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectOptions-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectOptions$fromJSON(ProjectOptionsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ProjectOptionsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
