% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_retrieve_response.R
\docType{class}
\name{ProjectRetrieveResponse}
\alias{ProjectRetrieveResponse}
\title{ProjectRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
ProjectRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{advancedOptions}}{\link{ProjectAdvancedOptionsResponse}}

\item{\code{autopilotClusterList}}{list( integer ) \link{optional} Optional. A list of integers where each value will be used as the number of clusters in Autopilot model(s) for unsupervised clustering projects. Cannot be specified unless unsupervisedMode is true and unsupervisedType is set to 'clustering'.}

\item{\code{autopilotMode}}{character The current autopilot mode. 0: Full Autopilot. 2: Manual Mode. 3: Quick Mode. 4: Comprehensive Autopilot. null: Mode not set.}

\item{\code{catalogId}}{character The ID of the AI catalog entry used to create the project, or null if not created from the AI catalog.}

\item{\code{catalogVersionId}}{character The ID of the AI catalog version used to create the project, or null if not created from the AI catalog.}

\item{\code{created}}{character The time of project creation.}

\item{\code{externalTimeSeriesBaselineDatasetMetadata}}{\link{ExternalTSBaselineMetadata} \link{optional}}

\item{\code{featureEngineeringPredictionPoint}}{character \link{optional} The date column to be used as prediction point for time-based feature engineering.}

\item{\code{fileName}}{character The name of the dataset used to create the project.}

\item{\code{holdoutUnlocked}}{character Whether the holdout has been unlocked.}

\item{\code{id}}{character The ID of the project.}

\item{\code{maxClusters}}{integer \link{optional} Only valid when unsupervisedMode is True and unsupervisedType is 'clustering'. The maximum number of clusters allowed when training clustering models. If specified cannot be exceed the number of rows in a project's dataset divided by 50 and must be less than or equal to `minClusters`. If unsupervisedMode is True and unsupervisedType is 'clustering' then defaults to the number of rows in the project's dataset divided by 50 or 100 if that number of greater than 100.}

\item{\code{maxTrainPct}}{numeric The maximum percentage of the dataset that can be used to successfully train a model without going into the validation data.}

\item{\code{maxTrainRows}}{integer The maximum number of rows of the dataset that can be used to successfully train a model without going into the validation data.}

\item{\code{metric}}{character The metric used to select the best-performing models.}

\item{\code{minClusters}}{integer \link{optional} Only valid when unsupervisedMode is True and unsupervisedType is 'clustering'. The minimum number of clusters allowed when training clustering models. If specified cannot be exceed the number of rows in a project's dataset divided by 50 and must be less than or equal to `maxClusters`. If unsupervisedMode is True and  unsupervisedType is 'clustering' then defaults to 2.}

\item{\code{partition}}{\link{Partition}}

\item{\code{positiveClass}}{\link{OneOfstringinteger} If the project uses binary classification, the class designated to be the positive class.}

\item{\code{primaryLocationColumn}}{character \link{optional} Primary location column name}

\item{\code{projectName}}{character The name of the project.}

\item{\code{queryGeneratorId}}{character The ID of the time series data prep query generator associated with the project, or null if there is no associated query generator.}

\item{\code{quickrun}}{character If the Autopilot mode is set to quick. DEPRECATED: look at autopilot_mode instead.}

\item{\code{relationshipsConfigurationId}}{character \link{optional} Relationships configuration id to be used for Feature Discovery projects.}

\item{\code{segmentation}}{\link{ProjectSegmentationInfoResponse} \link{optional}}

\item{\code{stage}}{character The stage of the project. If modeling, then the target is successfully set and modeling or predictions can proceed.}

\item{\code{target}}{character The target of the project, null if project is unsupervised.}

\item{\code{targetType}}{character The type of the selected target. Null if the project is unsupervised.}

\item{\code{unsupervisedMode}}{character Indicates whether a project is unsupervised.}

\item{\code{unsupervisedType}}{character \link{optional} Only valid when unsupervisedMode is True. The type of unsupervised project, anomaly or clustering. If unsupervisedMode, defaults to 'anomaly'.}

\item{\code{useFeatureDiscovery}}{character Indicates whether a feature discovery project was created as opposed to a regular project}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProjectRetrieveResponse-new}{\code{ProjectRetrieveResponse$new()}}
\item \href{#method-ProjectRetrieveResponse-validate}{\code{ProjectRetrieveResponse$validate()}}
\item \href{#method-ProjectRetrieveResponse-toJSON}{\code{ProjectRetrieveResponse$toJSON()}}
\item \href{#method-ProjectRetrieveResponse-fromJSON}{\code{ProjectRetrieveResponse$fromJSON()}}
\item \href{#method-ProjectRetrieveResponse-clone}{\code{ProjectRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectRetrieveResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectRetrieveResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectRetrieveResponse$new(
  advancedOptions = NULL,
  autopilotMode = NULL,
  catalogId = NULL,
  catalogVersionId = NULL,
  created = NULL,
  fileName = NULL,
  holdoutUnlocked = NULL,
  id = NULL,
  maxTrainPct = NULL,
  maxTrainRows = NULL,
  metric = NULL,
  partition = NULL,
  positiveClass = NULL,
  projectName = NULL,
  queryGeneratorId = NULL,
  quickrun = NULL,
  stage = NULL,
  target = NULL,
  targetType = NULL,
  unsupervisedMode = NULL,
  useFeatureDiscovery = NULL,
  autopilotClusterList = NULL,
  externalTimeSeriesBaselineDatasetMetadata = NULL,
  featureEngineeringPredictionPoint = NULL,
  maxClusters = NULL,
  minClusters = NULL,
  primaryLocationColumn = NULL,
  relationshipsConfigurationId = NULL,
  segmentation = NULL,
  unsupervisedType = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{autopilotMode}}{The current autopilot mode. 0: Full Autopilot. 2: Manual Mode. 3: Quick Mode. 4: Comprehensive Autopilot. null: Mode not set.}

\item{\code{catalogId}}{The ID of the AI catalog entry used to create the project, or null if not created from the AI catalog.}

\item{\code{catalogVersionId}}{The ID of the AI catalog version used to create the project, or null if not created from the AI catalog.}

\item{\code{created}}{The time of project creation.}

\item{\code{fileName}}{The name of the dataset used to create the project.}

\item{\code{holdoutUnlocked}}{Whether the holdout has been unlocked.}

\item{\code{id}}{The ID of the project.}

\item{\code{maxTrainPct}}{The maximum percentage of the dataset that can be used to successfully train a model without going into the validation data.}

\item{\code{maxTrainRows}}{The maximum number of rows of the dataset that can be used to successfully train a model without going into the validation data.}

\item{\code{metric}}{The metric used to select the best-performing models.}

\item{\code{positiveClass}}{If the project uses binary classification, the class designated to be the positive class.}

\item{\code{projectName}}{The name of the project.}

\item{\code{queryGeneratorId}}{The ID of the time series data prep query generator associated with the project, or null if there is no associated query generator.}

\item{\code{quickrun}}{If the Autopilot mode is set to quick. DEPRECATED: look at autopilot_mode instead.}

\item{\code{stage}}{The stage of the project. If modeling, then the target is successfully set and modeling or predictions can proceed.}

\item{\code{target}}{The target of the project, null if project is unsupervised.}

\item{\code{targetType}}{The type of the selected target. Null if the project is unsupervised.}

\item{\code{unsupervisedMode}}{Indicates whether a project is unsupervised.}

\item{\code{useFeatureDiscovery}}{Indicates whether a feature discovery project was created as opposed to a regular project}

\item{\code{autopilotClusterList}}{Optional. A list of integers where each value will be used as the number of clusters in Autopilot model(s) for unsupervised clustering projects. Cannot be specified unless unsupervisedMode is true and unsupervisedType is set to 'clustering'.}

\item{\code{featureEngineeringPredictionPoint}}{The date column to be used as prediction point for time-based feature engineering.}

\item{\code{maxClusters}}{Only valid when unsupervisedMode is True and unsupervisedType is 'clustering'. The maximum number of clusters allowed when training clustering models. If specified cannot be exceed the number of rows in a project's dataset divided by 50 and must be less than or equal to `minClusters`. If unsupervisedMode is True and unsupervisedType is 'clustering' then defaults to the number of rows in the project's dataset divided by 50 or 100 if that number of greater than 100.}

\item{\code{minClusters}}{Only valid when unsupervisedMode is True and unsupervisedType is 'clustering'. The minimum number of clusters allowed when training clustering models. If specified cannot be exceed the number of rows in a project's dataset divided by 50 and must be less than or equal to `maxClusters`. If unsupervisedMode is True and  unsupervisedType is 'clustering' then defaults to 2.}

\item{\code{primaryLocationColumn}}{Primary location column name}

\item{\code{relationshipsConfigurationId}}{Relationships configuration id to be used for Feature Discovery projects.}

\item{\code{unsupervisedType}}{Only valid when unsupervisedMode is True. The type of unsupervised project, anomaly or clustering. If unsupervisedMode, defaults to 'anomaly'.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectRetrieveResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectRetrieveResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectRetrieveResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectRetrieveResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectRetrieveResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectRetrieveResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectRetrieveResponse$fromJSON(
  ProjectRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ProjectRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectRetrieveResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectRetrieveResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
