% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{PutManyProjectsSegmentChampion}
\alias{PutManyProjectsSegmentChampion}
\title{PutManyProjectsSegmentChampion}
\usage{
PutManyProjectsSegmentChampion(modelId, projectId, clone_ = FALSE, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update champion model for a segment project.
}
\details{
Update champion model for a segment project.

This method invokes \verb{PUT /projects/\{projectId\}/segmentChampion/} in the DataRobot Public API.
}
\seealso{
The method ProjectsSegmentChampionPutMany in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
