% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limits.R
\docType{class}
\name{RateLimits}
\alias{RateLimits}
\title{RateLimits}
\format{
An \code{R6Class} generator object
}
\description{
RateLimits Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{limit}}{integer The number of instances of the resource allowed during the period}

\item{\code{resource}}{character The resource being limited}

\item{\code{usageTracked}}{character \link{optional} Whether or not a persistent record of usage of this resource is recorded}

\item{\code{window}}{integer The length of the window, in 'windowType' units.  For example 60 seconds, or 1 month}

\item{\code{windowType}}{character \link{optional} The unit of time to use for the rate limit window.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RateLimits-new}{\code{RateLimits$new()}}
\item \href{#method-RateLimits-validate}{\code{RateLimits$validate()}}
\item \href{#method-RateLimits-toJSON}{\code{RateLimits$toJSON()}}
\item \href{#method-RateLimits-fromJSON}{\code{RateLimits$fromJSON()}}
\item \href{#method-RateLimits-clone}{\code{RateLimits$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RateLimits-new"></a>}}
\if{latex}{\out{\hypertarget{method-RateLimits-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RateLimits$new(
  limit = NULL,
  resource = NULL,
  window = NULL,
  usageTracked = NULL,
  windowType = "seconds",
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{The number of instances of the resource allowed during the period}

\item{\code{resource}}{The resource being limited}

\item{\code{window}}{The length of the window, in 'windowType' units.  For example 60 seconds, or 1 month}

\item{\code{usageTracked}}{Whether or not a persistent record of usage of this resource is recorded}

\item{\code{windowType}}{The unit of time to use for the rate limit window.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RateLimits-validate"></a>}}
\if{latex}{\out{\hypertarget{method-RateLimits-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RateLimits$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RateLimits-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RateLimits-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RateLimits$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RateLimits-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RateLimits-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RateLimits$fromJSON(RateLimitsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{RateLimitsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RateLimits-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RateLimits-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RateLimits$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
