% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating_table_retrieve_response.R
\docType{class}
\name{RatingTableRetrieveResponse}
\alias{RatingTableRetrieveResponse}
\title{RatingTableRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
RatingTableRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{created}}{numeric ISO-8601 timestamp of when the rating table record was created.}

\item{\code{id}}{character the ID of the rating table record}

\item{\code{modelId}}{character the model ID of a model that was created from the rating table.         May be null if a model has not been created from the rating table.}

\item{\code{modelJobId}}{integer the qid of the job to create a model from this rating table.         Can be null if a model has not been created from the rating table.}

\item{\code{originalFilename}}{character the filename of the uploaded rating table file}

\item{\code{parentModelId}}{character the model ID of the model the rating table was modified from}

\item{\code{projectId}}{character the project ID of the rating table record}

\item{\code{ratingTableName}}{character the name of the rating table}

\item{\code{validationError}}{character rating table validation error messages. If the rating table         was validated successfully, it will be an empty string.}

\item{\code{validationJobId}}{character the qid of the created job to validate the rating table.         Can be null if the rating table has not been validated.}

\item{\code{validationWarnings}}{character rating table validation warning messages.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RatingTableRetrieveResponse$new()}}
\item \href{#method-validate}{\code{RatingTableRetrieveResponse$validate()}}
\item \href{#method-toJSON}{\code{RatingTableRetrieveResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{RatingTableRetrieveResponse$fromJSON()}}
\item \href{#method-clone}{\code{RatingTableRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RatingTableRetrieveResponse$new(
  created = NULL,
  id = NULL,
  modelId = NULL,
  modelJobId = NULL,
  originalFilename = NULL,
  parentModelId = NULL,
  projectId = NULL,
  ratingTableName = NULL,
  validationError = NULL,
  validationJobId = NULL,
  validationWarnings = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{created}}{ISO-8601 timestamp of when the rating table record was created.}

\item{\code{id}}{the ID of the rating table record}

\item{\code{modelId}}{the model ID of a model that was created from the rating table.         May be null if a model has not been created from the rating table.}

\item{\code{modelJobId}}{the qid of the job to create a model from this rating table.         Can be null if a model has not been created from the rating table.}

\item{\code{originalFilename}}{the filename of the uploaded rating table file}

\item{\code{parentModelId}}{the model ID of the model the rating table was modified from}

\item{\code{projectId}}{the project ID of the rating table record}

\item{\code{ratingTableName}}{the name of the rating table}

\item{\code{validationError}}{rating table validation error messages. If the rating table         was validated successfully, it will be an empty string.}

\item{\code{validationJobId}}{the qid of the created job to validate the rating table.         Can be null if the rating table has not been validated.}

\item{\code{validationWarnings}}{rating table validation warning messages.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RatingTableRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RatingTableRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RatingTableRetrieveResponse$fromJSON(
  RatingTableRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{RatingTableRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RatingTableRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
