% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasons.R
\docType{class}
\name{Reasons}
\alias{Reasons}
\title{Reasons}
\format{
An \code{R6Class} generator object
}
\description{
Reasons Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{supportsDocumentTextExtractionSampleInsight}}{character \link{optional} If present, the reason document text extraction sample insights are not supported for the model.}

\item{\code{supportsImageActivationMaps}}{character \link{optional} If present, the reason image activation maps are not supported for the model.}

\item{\code{supportsImageEmbedding}}{character \link{optional} If present, the reason image embeddings are not supported for the model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Reasons-new}{\code{Reasons$new()}}
\item \href{#method-Reasons-validate}{\code{Reasons$validate()}}
\item \href{#method-Reasons-toJSON}{\code{Reasons$toJSON()}}
\item \href{#method-Reasons-fromJSON}{\code{Reasons$fromJSON()}}
\item \href{#method-Reasons-clone}{\code{Reasons$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reasons-new"></a>}}
\if{latex}{\out{\hypertarget{method-Reasons-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reasons$new(
  supportsDocumentTextExtractionSampleInsight = NULL,
  supportsImageActivationMaps = NULL,
  supportsImageEmbedding = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{supportsDocumentTextExtractionSampleInsight}}{If present, the reason document text extraction sample insights are not supported for the model.}

\item{\code{supportsImageActivationMaps}}{If present, the reason image activation maps are not supported for the model.}

\item{\code{supportsImageEmbedding}}{If present, the reason image embeddings are not supported for the model.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reasons-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Reasons-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reasons$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reasons-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Reasons-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reasons$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reasons-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Reasons-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reasons$fromJSON(ReasonsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ReasonsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reasons-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Reasons-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reasons$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
