% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship.R
\docType{class}
\name{Relationship}
\alias{Relationship}
\title{Relationship}
\format{
An \code{R6Class} generator object
}
\description{
Relationship Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dataset1Identifier}}{character \link{optional} Identifier of the first dataset in the relationship. If this is not provided, it represents the primary dataset.}

\item{\code{dataset1Keys}}{list( character ) column(s) in the first dataset that are used to join to the second dataset.}

\item{\code{dataset2Identifier}}{character Identifier of the second dataset in the relationship.}

\item{\code{dataset2Keys}}{list( character ) column(s) in the second dataset that are used to join to the first dataset.}

\item{\code{featureDerivationWindowEnd}}{integer \link{optional} How many featureDerivationWindowUnits of each dataset's primary temporal key into the past relative to the datetimePartitionColumn the feature derivation window should end. Will be a non-positive integer, if present. If present, time-aware joins will be used. Only applicable when table1Identifier is not provided.}

\item{\code{featureDerivationWindowStart}}{integer \link{optional} How many featureDerivationWindowUnits of each dataset's primary temporal key into the past relative to the datetimePartitionColumn the feature derivation window should begin. Will be a negative integer, if present. If present, time-aware joins will be used. Only applicable when table1Identifier is not provided.}

\item{\code{featureDerivationWindowTimeUnit}}{character \link{optional} Time unit of the feature derivation window. Supported values are MILLISECOND, SECOND, MINUTE, HOUR, DAY, WEEK, MONTH, QUARTER, YEAR. If present, time-aware joins will be used. Only applicable when table1Identifier is not provided.}

\item{\code{featureDerivationWindows}}{list( \link{FeatureDerivationWindow} ) \link{optional} List of feature derivation window definitions that will be used.}

\item{\code{predictionPointRounding}}{integer \link{optional} Closest value of predictionPointRoundingTimeUnit to round the prediction point into the past when applying the feature derivation window. Will be a positive integer, if present. Only applicable when table1Identifier is not provided.}

\item{\code{predictionPointRoundingTimeUnit}}{character \link{optional} Time unit of the prediction point rounding. Supported values are MILLISECOND, SECOND, MINUTE, HOUR, DAY, WEEK, MONTH, QUARTER, YEAR. Only applicable when table1Identifier is not provided.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Relationship$new()}}
\item \href{#method-validate}{\code{Relationship$validate()}}
\item \href{#method-toJSON}{\code{Relationship$toJSON()}}
\item \href{#method-fromJSON}{\code{Relationship$fromJSON()}}
\item \href{#method-clone}{\code{Relationship$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$new(
  dataset1Keys = NULL,
  dataset2Identifier = NULL,
  dataset2Keys = NULL,
  dataset1Identifier = NULL,
  featureDerivationWindowEnd = NULL,
  featureDerivationWindowStart = NULL,
  featureDerivationWindowTimeUnit = NULL,
  featureDerivationWindows = NULL,
  predictionPointRounding = NULL,
  predictionPointRoundingTimeUnit = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset1Keys}}{column(s) in the first dataset that are used to join to the second dataset.}

\item{\code{dataset2Identifier}}{Identifier of the second dataset in the relationship.}

\item{\code{dataset2Keys}}{column(s) in the second dataset that are used to join to the first dataset.}

\item{\code{dataset1Identifier}}{Identifier of the first dataset in the relationship. If this is not provided, it represents the primary dataset.}

\item{\code{featureDerivationWindowEnd}}{How many featureDerivationWindowUnits of each dataset's primary temporal key into the past relative to the datetimePartitionColumn the feature derivation window should end. Will be a non-positive integer, if present. If present, time-aware joins will be used. Only applicable when table1Identifier is not provided.}

\item{\code{featureDerivationWindowStart}}{How many featureDerivationWindowUnits of each dataset's primary temporal key into the past relative to the datetimePartitionColumn the feature derivation window should begin. Will be a negative integer, if present. If present, time-aware joins will be used. Only applicable when table1Identifier is not provided.}

\item{\code{featureDerivationWindowTimeUnit}}{Time unit of the feature derivation window. Supported values are MILLISECOND, SECOND, MINUTE, HOUR, DAY, WEEK, MONTH, QUARTER, YEAR. If present, time-aware joins will be used. Only applicable when table1Identifier is not provided.}

\item{\code{featureDerivationWindows}}{List of feature derivation window definitions that will be used.}

\item{\code{predictionPointRounding}}{Closest value of predictionPointRoundingTimeUnit to round the prediction point into the past when applying the feature derivation window. Will be a positive integer, if present. Only applicable when table1Identifier is not provided.}

\item{\code{predictionPointRoundingTimeUnit}}{Time unit of the prediction point rounding. Supported values are MILLISECOND, SECOND, MINUTE, HOUR, DAY, WEEK, MONTH, QUARTER, YEAR. Only applicable when table1Identifier is not provided.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$fromJSON(RelationshipJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{RelationshipJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
