% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship_response.R
\docType{class}
\name{RelationshipResponse}
\alias{RelationshipResponse}
\title{RelationshipResponse}
\format{
An \code{R6Class} generator object
}
\description{
RelationshipResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dataset1Identifier}}{character Identifier of the first dataset in the relationship. If this is not provided, it represents the primary dataset.}

\item{\code{dataset1Keys}}{list( character ) column(s) in the first dataset that are used to join to the second dataset.}

\item{\code{dataset2Identifier}}{character Identifier of the second dataset in the relationship.}

\item{\code{dataset2Keys}}{list( character ) column(s) in the second dataset that are used to join to the first dataset.}

\item{\code{featureDerivationWindowEnd}}{integer \link{optional} How many featureDerivationWindowUnits of each dataset's primary temporal key into the past relative to the datetimePartitionColumn the feature derivation window should end. Will be a non-positive integer, if present. If present, time-aware joins will be used.}

\item{\code{featureDerivationWindowStart}}{integer \link{optional} How many featureDerivationWindowUnits of each dataset's primary temporal key into the past relative to the datetimePartitionColumn the feature derivation window should begin. Will be a negative integer, if present. If present, time-aware joins will be used.}

\item{\code{featureDerivationWindowTimeUnit}}{character \link{optional} Time unit of the feature derivation window. Supported values are MILLISECOND, SECOND, MINUTE, HOUR, DAY, WEEK, MONTH, QUARTER, YEAR. If present, time-aware joins will be used.}

\item{\code{featureDerivationWindows}}{list( \link{FeatureDerivationWindow} ) \link{optional} List of feature derivation window definitions that will be used.}

\item{\code{predictionPointRounding}}{integer \link{optional} Can be null, closest value of predictionPointRoundingTimeUnit to round the prediction point into the past when applying the feature derivation window. Will be a positive integer, if present.}

\item{\code{predictionPointRoundingTimeUnit}}{character \link{optional} Time unit of the prediction point rounding. Supported values are MILLISECOND, SECOND, MINUTE, HOUR, DAY, WEEK, MONTH, QUARTER, YEAR.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RelationshipResponse-new}{\code{RelationshipResponse$new()}}
\item \href{#method-RelationshipResponse-validate}{\code{RelationshipResponse$validate()}}
\item \href{#method-RelationshipResponse-toJSON}{\code{RelationshipResponse$toJSON()}}
\item \href{#method-RelationshipResponse-fromJSON}{\code{RelationshipResponse$fromJSON()}}
\item \href{#method-RelationshipResponse-clone}{\code{RelationshipResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RelationshipResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-RelationshipResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationshipResponse$new(
  dataset1Identifier = NULL,
  dataset1Keys = NULL,
  dataset2Identifier = NULL,
  dataset2Keys = NULL,
  featureDerivationWindowEnd = NULL,
  featureDerivationWindowStart = NULL,
  featureDerivationWindowTimeUnit = NULL,
  featureDerivationWindows = NULL,
  predictionPointRounding = NULL,
  predictionPointRoundingTimeUnit = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset1Identifier}}{Identifier of the first dataset in the relationship. If this is not provided, it represents the primary dataset.}

\item{\code{dataset1Keys}}{column(s) in the first dataset that are used to join to the second dataset.}

\item{\code{dataset2Identifier}}{Identifier of the second dataset in the relationship.}

\item{\code{dataset2Keys}}{column(s) in the second dataset that are used to join to the first dataset.}

\item{\code{featureDerivationWindowEnd}}{How many featureDerivationWindowUnits of each dataset's primary temporal key into the past relative to the datetimePartitionColumn the feature derivation window should end. Will be a non-positive integer, if present. If present, time-aware joins will be used.}

\item{\code{featureDerivationWindowStart}}{How many featureDerivationWindowUnits of each dataset's primary temporal key into the past relative to the datetimePartitionColumn the feature derivation window should begin. Will be a negative integer, if present. If present, time-aware joins will be used.}

\item{\code{featureDerivationWindowTimeUnit}}{Time unit of the feature derivation window. Supported values are MILLISECOND, SECOND, MINUTE, HOUR, DAY, WEEK, MONTH, QUARTER, YEAR. If present, time-aware joins will be used.}

\item{\code{featureDerivationWindows}}{List of feature derivation window definitions that will be used.}

\item{\code{predictionPointRounding}}{Can be null, closest value of predictionPointRoundingTimeUnit to round the prediction point into the past when applying the feature derivation window. Will be a positive integer, if present.}

\item{\code{predictionPointRoundingTimeUnit}}{Time unit of the prediction point rounding. Supported values are MILLISECOND, SECOND, MINUTE, HOUR, DAY, WEEK, MONTH, QUARTER, YEAR.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RelationshipResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-RelationshipResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationshipResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RelationshipResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RelationshipResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationshipResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RelationshipResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RelationshipResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationshipResponse$fromJSON(RelationshipResponseJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{RelationshipResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RelationshipResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RelationshipResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationshipResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
