% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_event_create.R
\docType{class}
\name{RemoteEventCreate}
\alias{RemoteEventCreate}
\title{RemoteEventCreate}
\format{
An \code{R6Class} generator object
}
\description{
RemoteEventCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{RemoteEventData} \link{optional}}

\item{\code{deploymentId}}{character \link{optional} The identifier of the deployment associated with the event.}

\item{\code{eventType}}{character The type of the event. Labels in all_lower_case are deprecated.}

\item{\code{message}}{character \link{optional} Descriptive message for health events.}

\item{\code{orgId}}{character \link{optional} The identifier of the organization associated with the event.}

\item{\code{predictionEnvironmentId}}{character \link{optional} The identifier of the prediction environment associated with the event.}

\item{\code{predictionRequestData}}{\link{PredictionRemoteEventData} \link{optional}}

\item{\code{spoolerChannelData}}{\link{SpoolerChannelEventData} \link{optional}}

\item{\code{timestamp}}{character The time when the event occurred.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RemoteEventCreate-new}{\code{RemoteEventCreate$new()}}
\item \href{#method-RemoteEventCreate-validate}{\code{RemoteEventCreate$validate()}}
\item \href{#method-RemoteEventCreate-toJSON}{\code{RemoteEventCreate$toJSON()}}
\item \href{#method-RemoteEventCreate-fromJSON}{\code{RemoteEventCreate$fromJSON()}}
\item \href{#method-RemoteEventCreate-clone}{\code{RemoteEventCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteEventCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteEventCreate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteEventCreate$new(
  eventType = NULL,
  timestamp = NULL,
  data = NULL,
  deploymentId = NULL,
  message = NULL,
  orgId = NULL,
  predictionEnvironmentId = NULL,
  predictionRequestData = NULL,
  spoolerChannelData = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{eventType}}{The type of the event. Labels in all_lower_case are deprecated.}

\item{\code{timestamp}}{The time when the event occurred.}

\item{\code{deploymentId}}{The identifier of the deployment associated with the event.}

\item{\code{message}}{Descriptive message for health events.}

\item{\code{orgId}}{The identifier of the organization associated with the event.}

\item{\code{predictionEnvironmentId}}{The identifier of the prediction environment associated with the event.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteEventCreate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteEventCreate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteEventCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteEventCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteEventCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteEventCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteEventCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteEventCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteEventCreate$fromJSON(RemoteEventCreateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{RemoteEventCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteEventCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteEventCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteEventCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
