% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{RenameRatingTable}
\alias{RenameRatingTable}
\title{RenameRatingTable}
\usage{
RenameRatingTable(project, ratingTableId, ratingTableName, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update an uploaded rating table
}
\details{
Rating tables may only be updated if they have not yet been used to create a model.

This method invokes \verb{PATCH /projects/\{projectId\}/ratingTables/\{ratingTableId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsRatingTablesPatch in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
