% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ReplaceDeployedModel}
\alias{ReplaceDeployedModel}
\title{ReplaceDeployedModel}
\usage{
ReplaceDeployedModel(
  deploymentId,
  newModelId,
  replacementReason,
  maxWait = 600,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Model Replacement.
}
\details{
Replace the model used to make predictions for the deployment. A validation process will be performed to make sure the new model is eligible as a replacement. If the validation fails, the model replacement will not occur.The Model Replacement Validation endpoint can be used to confirm the new model is eligible as a replacement.

This method invokes \verb{PATCH /deployments/\{deploymentId\}/model/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsModelPatchMany in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
