% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{RequestApproximation}
\alias{RequestApproximation}
\title{RequestApproximation}
\usage{
RequestApproximation(project, modelId, body = NULL, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create Rulesets
}
\details{
Approximate an existing model on the leaderboard with DataRobot Prime. A request body should be an empty JSON {}

This method invokes \verb{POST /projects/\{projectId\}/models/\{modelId\}/primeRulesets/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsPrimeRulesetsCreate in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
