% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{RequestBlender}
\alias{RequestBlender}
\title{RequestBlender}
\usage{
RequestBlender(project, modelsToBlend, blendMethod, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a blender from other models using a specified blender method.
}
\details{
Create a blender from other models using a specified blender method. Note: Time Series projects only allow the following blender methods: \"AVG\", \"MED\", \"FORECAST_DISTANCE_ENET\", and \"FORECAST_DISTANCE_AVG\".

This method invokes \verb{POST /projects/\{projectId\}/blenderModels/} in the DataRobot Public API.
}
\seealso{
The method ProjectsBlenderModelsCreate in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
