% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{RequestMultiSeriesDetection}
\alias{RequestMultiSeriesDetection}
\title{RequestMultiSeriesDetection}
\usage{
RequestMultiSeriesDetection(
  project,
  dateColumn,
  multiseriesIdColumns = NULL,
  maxWait = 600,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Detect multiseries properties
}
\details{
Analyze relationships between potential partition and multiseries ID columns. Time series projects require that each timestamp have at most one row corresponding to it. However, multiple series of data can be handled within a single project by designating a multiseries ID column that assigns each row to a particular series. See the \verb{multiseries <multiseries>} docs on time series projects for more information. A detection job analyzing the relationship between the multiseries ID column and the datetime partition column must be ran before it can be used. If the desired multiseries ID column(s) are known, it can be specified to limit the analysis to only those columns.

This method invokes \verb{POST /projects/\{projectId\}/multiseriesProperties/} in the DataRobot Public API.
}
\seealso{
The method MultiseriesPropertiesCreate in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
