% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{RequestNewModel}
\alias{RequestNewModel}
\title{RequestNewModel}
\usage{
RequestNewModel(
  project,
  blueprint,
  featurelist = NULL,
  samplePct = NULL,
  trainingRowCount = NULL,
  scoringType = NULL,
  monotonicIncreasingFeaturelistId = NULL,
  monotonicDecreasingFeaturelistId = NULL,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Train a new model
}
\details{
Train a new model. To specify the amount of data to use to train the model, use either samplePct to express a percentage of the rows of the dataset to use or trainingRowCount to express the number of rows to use. If neither samplePct or trainingRowCount is specified, the model will be trained on the maximum available training data that can be used to train an in-memory model. For projects using smart sampling, samplePct and trainingRowCount will be interpreted as a percent or number of rows of the minority class.   When configuring retraining sample sizes for models in projects with large row counts, DataRobot recommends requesting sample sizes using integer row counts instead of percentages. This is because percentages map to many actual possible row counts and only one of which is the actual sample size for up to validation. For example, if a project has 199,408 rows and you request a 64\% sample size, any number of rows between 126,625 rows and 128,618 rows maps to 64\% of the data. Using actual integer row counts (or project.max_training_rows) avoids ambiguity around how many rows of data you want the model to use.

This method invokes \verb{POST /projects/\{projectId\}/models/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsCreate in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
