% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{RequestNewRatingTableModel}
\alias{RequestNewRatingTableModel}
\title{RequestNewRatingTableModel}
\usage{
RequestNewRatingTableModel(project, ratingTableId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create New Models From A Rating Table
}
\details{
Create a new rating table model from a validated rating table record

This method invokes \verb{POST /projects/\{projectId\}/ratingTableModels/} in the DataRobot Public API.
}
\seealso{
The method ProjectsRatingTableModelsCreate in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
