% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{RequestPrimeModel}
\alias{RequestPrimeModel}
\title{RequestPrimeModel}
\usage{
RequestPrimeModel(project, ruleset, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a Prime Model from a Ruleset
}
\details{
Create a Prime model using a particular ruleset.  DataRobot Prime is not available for multiclass projects.  Once rulesets approximating a parent model have been created, using POST /api/v2/projects/(projectId)/models/(modelId)/primeRulesets/, this route will allow creation of a Prime model using one of those rulesets.  Available rulesets can be retrieved via GET /api/v2/projects/(projectId)/models/(modelId)/primeRulesets/.

This method invokes \verb{POST /projects/\{projectId\}/primeModels/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPrimeModelsCreate in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
