% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals_chart_for_datasets.R
\docType{class}
\name{ResidualsChartForDatasets}
\alias{ResidualsChartForDatasets}
\title{ResidualsChartForDatasets}
\format{
An \code{R6Class} generator object
}
\description{
ResidualsChartForDatasets Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{coefficientOfDetermination}}{numeric Also known as the r-squared value. This value is calculated over the downsampled dataset, not the full input.}

\item{\code{data}}{list( \link{array[numeric]} ) The rows of chart data in \link{actual, predicted, residual, row number} form.  If the row number was not available at the time of model creation, the row number will be `null`.}

\item{\code{datasetId}}{character The dataset ID.}

\item{\code{histogram}}{list( \link{Histogram} ) Data to plot a histogram of residual values. The object contains three keys, intervalStart, intervalEnd, and occurrences, the number of times the predicted value fits within that interval. For all but the last interval, the end value is exclusive.}

\item{\code{residualMean}}{numeric The arithmetic mean of the predicted value minus the actual value over the downsampled dataset.}

\item{\code{standardDeviation}}{numeric The Standard Deviation value measures variation in the dataset. A low value indicates that the data points tend to be close to the mean; a high value indicates that the data points are spread over a wider range of values.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResidualsChartForDatasets-new}{\code{ResidualsChartForDatasets$new()}}
\item \href{#method-ResidualsChartForDatasets-validate}{\code{ResidualsChartForDatasets$validate()}}
\item \href{#method-ResidualsChartForDatasets-toJSON}{\code{ResidualsChartForDatasets$toJSON()}}
\item \href{#method-ResidualsChartForDatasets-fromJSON}{\code{ResidualsChartForDatasets$fromJSON()}}
\item \href{#method-ResidualsChartForDatasets-clone}{\code{ResidualsChartForDatasets$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResidualsChartForDatasets-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResidualsChartForDatasets-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResidualsChartForDatasets$new(
  coefficientOfDetermination = NULL,
  data = NULL,
  datasetId = NULL,
  histogram = NULL,
  residualMean = NULL,
  standardDeviation = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coefficientOfDetermination}}{Also known as the r-squared value. This value is calculated over the downsampled dataset, not the full input.}

\item{\code{data}}{The rows of chart data in \link{actual, predicted, residual, row number} form.  If the row number was not available at the time of model creation, the row number will be `null`.}

\item{\code{datasetId}}{The dataset ID.}

\item{\code{histogram}}{Data to plot a histogram of residual values. The object contains three keys, intervalStart, intervalEnd, and occurrences, the number of times the predicted value fits within that interval. For all but the last interval, the end value is exclusive.}

\item{\code{residualMean}}{The arithmetic mean of the predicted value minus the actual value over the downsampled dataset.}

\item{\code{standardDeviation}}{The Standard Deviation value measures variation in the dataset. A low value indicates that the data points tend to be close to the mean; a high value indicates that the data points are spread over a wider range of values.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResidualsChartForDatasets-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ResidualsChartForDatasets-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResidualsChartForDatasets$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResidualsChartForDatasets-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResidualsChartForDatasets-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResidualsChartForDatasets$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResidualsChartForDatasets-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResidualsChartForDatasets-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResidualsChartForDatasets$fromJSON(
  ResidualsChartForDatasetsJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ResidualsChartForDatasetsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResidualsChartForDatasets-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResidualsChartForDatasets-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResidualsChartForDatasets$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
