% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_usage.R
\docType{class}
\name{ResourceUsage}
\alias{ResourceUsage}
\title{ResourceUsage}
\format{
An \code{R6Class} generator object
}
\description{
ResourceUsage Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{currentUsage}}{integer The usage accumulated within the current window.}

\item{\code{maxUsage}}{integer The maximum value currentUsage can take before requests using this resource are denied.}

\item{\code{resource}}{character The rate limit resource this usage applies to.}

\item{\code{timeToExpire}}{integer The number of seconds until this usage resets to zero.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResourceUsage-new}{\code{ResourceUsage$new()}}
\item \href{#method-ResourceUsage-validate}{\code{ResourceUsage$validate()}}
\item \href{#method-ResourceUsage-toJSON}{\code{ResourceUsage$toJSON()}}
\item \href{#method-ResourceUsage-fromJSON}{\code{ResourceUsage$fromJSON()}}
\item \href{#method-ResourceUsage-clone}{\code{ResourceUsage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUsage-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUsage-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUsage$new(
  currentUsage = NULL,
  maxUsage = NULL,
  resource = NULL,
  timeToExpire = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{currentUsage}}{The usage accumulated within the current window.}

\item{\code{maxUsage}}{The maximum value currentUsage can take before requests using this resource are denied.}

\item{\code{resource}}{The rate limit resource this usage applies to.}

\item{\code{timeToExpire}}{The number of seconds until this usage resets to zero.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUsage-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUsage-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUsage$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUsage-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUsage-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUsage$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUsage-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUsage-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUsage$fromJSON(ResourceUsageJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ResourceUsageJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUsage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUsage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUsage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
