% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_usage_response.R
\docType{class}
\name{ResourceUsageResponse}
\alias{ResourceUsageResponse}
\title{ResourceUsageResponse}
\format{
An \code{R6Class} generator object
}
\description{
ResourceUsageResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer The number of resource usage records.}

\item{\code{next_}}{character URL pointing to the next page.}

\item{\code{previous}}{character URL pointing to the previous page.}

\item{\code{usage}}{list( \link{ResourceUsage} ) The resource usage records.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResourceUsageResponse-new}{\code{ResourceUsageResponse$new()}}
\item \href{#method-ResourceUsageResponse-validate}{\code{ResourceUsageResponse$validate()}}
\item \href{#method-ResourceUsageResponse-toJSON}{\code{ResourceUsageResponse$toJSON()}}
\item \href{#method-ResourceUsageResponse-fromJSON}{\code{ResourceUsageResponse$fromJSON()}}
\item \href{#method-ResourceUsageResponse-clone}{\code{ResourceUsageResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUsageResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUsageResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUsageResponse$new(
  count = NULL,
  next_ = NULL,
  previous = NULL,
  usage = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{The number of resource usage records.}

\item{\code{next_}}{URL pointing to the next page.}

\item{\code{previous}}{URL pointing to the previous page.}

\item{\code{usage}}{The resource usage records.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUsageResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUsageResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUsageResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUsageResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUsageResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUsageResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUsageResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUsageResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUsageResponse$fromJSON(
  ResourceUsageResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ResourceUsageResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUsageResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUsageResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUsageResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
