% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_notification.R
\docType{class}
\name{ResponseNotification}
\alias{ResponseNotification}
\title{ResponseNotification}
\format{
An \code{R6Class} generator object
}
\description{
ResponseNotification Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{body}}{character The body of the response}

\item{\code{duration}}{integer The duration}

\item{\code{headers}}{character The headers of the response}

\item{\code{statusCode}}{character The status code of the response}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResponseNotification-new}{\code{ResponseNotification$new()}}
\item \href{#method-ResponseNotification-validate}{\code{ResponseNotification$validate()}}
\item \href{#method-ResponseNotification-toJSON}{\code{ResponseNotification$toJSON()}}
\item \href{#method-ResponseNotification-fromJSON}{\code{ResponseNotification$fromJSON()}}
\item \href{#method-ResponseNotification-clone}{\code{ResponseNotification$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseNotification-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseNotification-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseNotification$new(
  body = NULL,
  duration = NULL,
  headers = NULL,
  statusCode = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{body}}{The body of the response}

\item{\code{duration}}{The duration}

\item{\code{headers}}{The headers of the response}

\item{\code{statusCode}}{The status code of the response}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseNotification-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseNotification-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseNotification$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseNotification-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseNotification-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseNotification$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseNotification-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseNotification-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseNotification$fromJSON(ResponseNotificationJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ResponseNotificationJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseNotification-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseNotification-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseNotification$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
