% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{RestoreDiscardedFeatures}
\alias{RestoreDiscardedFeatures}
\title{RestoreDiscardedFeatures}
\usage{
RestoreDiscardedFeatures(projectId, featuresToRestore, maxWait = 60 * 10, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Restore discarded time series features.
}
\details{
Restore discarded time series features.

This method invokes \verb{POST /projects/\{projectId\}/modelingFeatures/fromDiscardedFeatures/} in the DataRobot Public API.
}
\seealso{
The method ModelingFeaturesFromDiscardedFeaturesCreate in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
