% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrain_datetime_model.R
\docType{class}
\name{RetrainDatetimeModel}
\alias{RetrainDatetimeModel}
\title{RetrainDatetimeModel}
\format{
An \code{R6Class} generator object
}
\description{
RetrainDatetimeModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{featurelistId}}{character \link{optional} If specified, the new model will be trained using this featurelist. Otherwise, the model will be trained on the same feature list as the source model.}

\item{\code{modelId}}{character The ID of an existing model to use as the source for the training parameters.}

\item{\code{monotonicDecreasingFeaturelistId}}{character \link{optional} The ID of the featurelist that defines the set of features with a monotonically decreasing relationship to the target. If ``null``, no such constraints are enforced.}

\item{\code{monotonicIncreasingFeaturelistId}}{character \link{optional} The ID of the featurelist that defines the set of features with a monotonically increasing relationship to the target. If ``null``, no such constraints are enforced.}

\item{\code{nClusters}}{integer \link{optional} The number of clusters to use in the specified unsupervised clustering model. Only valid in unsupervised clustering projects.}

\item{\code{samplingMethod}}{character \link{optional} Defines how training data is selected if subsampling is used (e.g., if `timeWindowSamplePct` is specified). Can be either ``random`` or ``latest``. If omitted, defaults to ``latest`` if `trainingRowCount` is used and ``random`` for other cases (e.g., if `trainingDuration` or `useProjectSettings` is specified). May only be specified for OTV projects.}

\item{\code{timeWindowSamplePct}}{integer \link{optional} An integer between 1 and 99 indicating the percentage of sampling within the time window. The points kept are determined by the value provided for the `samplingMethod` option. If specified, `trainingRowCount` may not be specified, and the specified model must either be a duration or selectedDateRange model, or one of `trainingDuration` or `trainingStartDate` and `trainingEndDate` must be specified.}

\item{\code{trainingDuration}}{character \link{optional} A duration string representing the training duration to use for training the new model. If specified, the model will be trained using the specified training duration. Otherwise, the original model's duration will be used. Only one of `trainingRowCount`, `trainingDuration`, `trainingStartDate` and `trainingEndDate`, or `useProjectSettings` may be specified.}

\item{\code{trainingEndDate}}{character \link{optional} A datetime string representing the end date of the data to use for training this model. Note that only one of `trainingDuration` or `trainingRowCount` or `trainingStartDate` and `trainingEndDate` should be specified. If `trainingStartDate` and `trainingEndDate` are specified, the source model must be frozen.}

\item{\code{trainingRowCount}}{integer \link{optional} The number of rows of data that should be used to train the model. If not specified, the original model's row count will be used. Only one of `trainingRowCount`, `trainingDuration`, `trainingStartDate` and `trainingEndDate`, or `useProjectSettings` may be specified.}

\item{\code{trainingStartDate}}{character \link{optional} A datetime string representing the start date of the data to use for training this model. Note that only one of `trainingDuration` or `trainingRowCount` or `trainingStartDate` and `trainingEndDate` should be specified. If `trainingStartDate` and `trainingEndDate` are specified, the source model must be frozen.}

\item{\code{useProjectSettings}}{character \link{optional} If ``True``, the model will be trained using the previously-specified custom backtest training settings. Only one of `trainingRowCount`, `trainingDuration`, `trainingStartDate` and `trainingEndDate`, or `useProjectSettings` may be specified.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RetrainDatetimeModel-new}{\code{RetrainDatetimeModel$new()}}
\item \href{#method-RetrainDatetimeModel-validate}{\code{RetrainDatetimeModel$validate()}}
\item \href{#method-RetrainDatetimeModel-toJSON}{\code{RetrainDatetimeModel$toJSON()}}
\item \href{#method-RetrainDatetimeModel-fromJSON}{\code{RetrainDatetimeModel$fromJSON()}}
\item \href{#method-RetrainDatetimeModel-clone}{\code{RetrainDatetimeModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RetrainDatetimeModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-RetrainDatetimeModel-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainDatetimeModel$new(
  modelId = NULL,
  featurelistId = NULL,
  monotonicDecreasingFeaturelistId = NULL,
  monotonicIncreasingFeaturelistId = NULL,
  nClusters = NULL,
  samplingMethod = NULL,
  timeWindowSamplePct = NULL,
  trainingDuration = NULL,
  trainingEndDate = NULL,
  trainingRowCount = NULL,
  trainingStartDate = NULL,
  useProjectSettings = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{modelId}}{The ID of an existing model to use as the source for the training parameters.}

\item{\code{featurelistId}}{If specified, the new model will be trained using this featurelist. Otherwise, the model will be trained on the same feature list as the source model.}

\item{\code{monotonicDecreasingFeaturelistId}}{The ID of the featurelist that defines the set of features with a monotonically decreasing relationship to the target. If ``null``, no such constraints are enforced.}

\item{\code{monotonicIncreasingFeaturelistId}}{The ID of the featurelist that defines the set of features with a monotonically increasing relationship to the target. If ``null``, no such constraints are enforced.}

\item{\code{nClusters}}{The number of clusters to use in the specified unsupervised clustering model. Only valid in unsupervised clustering projects.}

\item{\code{samplingMethod}}{Defines how training data is selected if subsampling is used (e.g., if `timeWindowSamplePct` is specified). Can be either ``random`` or ``latest``. If omitted, defaults to ``latest`` if `trainingRowCount` is used and ``random`` for other cases (e.g., if `trainingDuration` or `useProjectSettings` is specified). May only be specified for OTV projects.}

\item{\code{timeWindowSamplePct}}{An integer between 1 and 99 indicating the percentage of sampling within the time window. The points kept are determined by the value provided for the `samplingMethod` option. If specified, `trainingRowCount` may not be specified, and the specified model must either be a duration or selectedDateRange model, or one of `trainingDuration` or `trainingStartDate` and `trainingEndDate` must be specified.}

\item{\code{trainingDuration}}{A duration string representing the training duration to use for training the new model. If specified, the model will be trained using the specified training duration. Otherwise, the original model's duration will be used. Only one of `trainingRowCount`, `trainingDuration`, `trainingStartDate` and `trainingEndDate`, or `useProjectSettings` may be specified.}

\item{\code{trainingEndDate}}{A datetime string representing the end date of the data to use for training this model. Note that only one of `trainingDuration` or `trainingRowCount` or `trainingStartDate` and `trainingEndDate` should be specified. If `trainingStartDate` and `trainingEndDate` are specified, the source model must be frozen.}

\item{\code{trainingRowCount}}{The number of rows of data that should be used to train the model. If not specified, the original model's row count will be used. Only one of `trainingRowCount`, `trainingDuration`, `trainingStartDate` and `trainingEndDate`, or `useProjectSettings` may be specified.}

\item{\code{trainingStartDate}}{A datetime string representing the start date of the data to use for training this model. Note that only one of `trainingDuration` or `trainingRowCount` or `trainingStartDate` and `trainingEndDate` should be specified. If `trainingStartDate` and `trainingEndDate` are specified, the source model must be frozen.}

\item{\code{useProjectSettings}}{If ``True``, the model will be trained using the previously-specified custom backtest training settings. Only one of `trainingRowCount`, `trainingDuration`, `trainingStartDate` and `trainingEndDate`, or `useProjectSettings` may be specified.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RetrainDatetimeModel-validate"></a>}}
\if{latex}{\out{\hypertarget{method-RetrainDatetimeModel-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainDatetimeModel$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RetrainDatetimeModel-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RetrainDatetimeModel-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainDatetimeModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RetrainDatetimeModel-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RetrainDatetimeModel-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainDatetimeModel$fromJSON(RetrainDatetimeModelJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{RetrainDatetimeModelJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RetrainDatetimeModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RetrainDatetimeModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainDatetimeModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
