% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retraining_policy_create.R
\docType{class}
\name{RetrainingPolicyCreate}
\alias{RetrainingPolicyCreate}
\title{RetrainingPolicyCreate}
\format{
An \code{R6Class} generator object
}
\description{
RetrainingPolicyCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{action}}{character \link{optional} Configure the action to take on the resultant new model.}

\item{\code{autopilotOptions}}{\link{AutopilotOptions} \link{optional}}

\item{\code{description}}{character Description of the retraining policy.}

\item{\code{featureListStrategy}}{character \link{optional} Configure the feature list strategy used for modeling.}

\item{\code{modelSelectionStrategy}}{character \link{optional} Configure how new model is selected when the retraining policy runs.}

\item{\code{name}}{character Name of the retraining policy.}

\item{\code{projectOptions}}{\link{ProjectOptions} \link{optional}}

\item{\code{projectOptionsStrategy}}{character \link{optional} Configure the project option strategy used for modeling.}

\item{\code{timeSeriesOptions}}{\link{TimeSeriesOptions} \link{optional}}

\item{\code{trigger}}{\link{Trigger} \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RetrainingPolicyCreate-new}{\code{RetrainingPolicyCreate$new()}}
\item \href{#method-RetrainingPolicyCreate-validate}{\code{RetrainingPolicyCreate$validate()}}
\item \href{#method-RetrainingPolicyCreate-toJSON}{\code{RetrainingPolicyCreate$toJSON()}}
\item \href{#method-RetrainingPolicyCreate-fromJSON}{\code{RetrainingPolicyCreate$fromJSON()}}
\item \href{#method-RetrainingPolicyCreate-clone}{\code{RetrainingPolicyCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RetrainingPolicyCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-RetrainingPolicyCreate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainingPolicyCreate$new(
  description = NULL,
  name = NULL,
  action = NULL,
  autopilotOptions = NULL,
  featureListStrategy = NULL,
  modelSelectionStrategy = NULL,
  projectOptions = NULL,
  projectOptionsStrategy = NULL,
  timeSeriesOptions = NULL,
  trigger = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{Description of the retraining policy.}

\item{\code{name}}{Name of the retraining policy.}

\item{\code{action}}{Configure the action to take on the resultant new model.}

\item{\code{featureListStrategy}}{Configure the feature list strategy used for modeling.}

\item{\code{modelSelectionStrategy}}{Configure how new model is selected when the retraining policy runs.}

\item{\code{projectOptionsStrategy}}{Configure the project option strategy used for modeling.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RetrainingPolicyCreate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-RetrainingPolicyCreate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainingPolicyCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RetrainingPolicyCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RetrainingPolicyCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainingPolicyCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RetrainingPolicyCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RetrainingPolicyCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainingPolicyCreate$fromJSON(
  RetrainingPolicyCreateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{RetrainingPolicyCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RetrainingPolicyCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RetrainingPolicyCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainingPolicyCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
