% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retraining_policy_run_retrieve.R
\docType{class}
\name{RetrainingPolicyRunRetrieve}
\alias{RetrainingPolicyRunRetrieve}
\title{RetrainingPolicyRunRetrieve}
\format{
An \code{R6Class} generator object
}
\description{
RetrainingPolicyRunRetrieve Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{challengerId}}{character ID of the challenger created from this retraining policy run.}

\item{\code{errorMessage}}{character Error message of the retraining policy run.}

\item{\code{finishTime}}{character Finish time of the retraining policy run.}

\item{\code{id}}{character ID of the retraining policy run.}

\item{\code{modelPackageId}}{character ID of the model package created from this retraining policy run.}

\item{\code{projectId}}{character ID of the project created from this retraining policy run.}

\item{\code{startTime}}{character Start time of the retraining policy run.}

\item{\code{status}}{character Status of the retraining policy run.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RetrainingPolicyRunRetrieve-new}{\code{RetrainingPolicyRunRetrieve$new()}}
\item \href{#method-RetrainingPolicyRunRetrieve-validate}{\code{RetrainingPolicyRunRetrieve$validate()}}
\item \href{#method-RetrainingPolicyRunRetrieve-toJSON}{\code{RetrainingPolicyRunRetrieve$toJSON()}}
\item \href{#method-RetrainingPolicyRunRetrieve-fromJSON}{\code{RetrainingPolicyRunRetrieve$fromJSON()}}
\item \href{#method-RetrainingPolicyRunRetrieve-clone}{\code{RetrainingPolicyRunRetrieve$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RetrainingPolicyRunRetrieve-new"></a>}}
\if{latex}{\out{\hypertarget{method-RetrainingPolicyRunRetrieve-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainingPolicyRunRetrieve$new(
  challengerId = NULL,
  errorMessage = NULL,
  finishTime = NULL,
  id = NULL,
  modelPackageId = NULL,
  projectId = NULL,
  startTime = NULL,
  status = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{challengerId}}{ID of the challenger created from this retraining policy run.}

\item{\code{errorMessage}}{Error message of the retraining policy run.}

\item{\code{finishTime}}{Finish time of the retraining policy run.}

\item{\code{id}}{ID of the retraining policy run.}

\item{\code{modelPackageId}}{ID of the model package created from this retraining policy run.}

\item{\code{projectId}}{ID of the project created from this retraining policy run.}

\item{\code{startTime}}{Start time of the retraining policy run.}

\item{\code{status}}{Status of the retraining policy run.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RetrainingPolicyRunRetrieve-validate"></a>}}
\if{latex}{\out{\hypertarget{method-RetrainingPolicyRunRetrieve-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainingPolicyRunRetrieve$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RetrainingPolicyRunRetrieve-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RetrainingPolicyRunRetrieve-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainingPolicyRunRetrieve$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RetrainingPolicyRunRetrieve-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RetrainingPolicyRunRetrieve-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainingPolicyRunRetrieve$fromJSON(
  RetrainingPolicyRunRetrieveJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{RetrainingPolicyRunRetrieveJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RetrainingPolicyRunRetrieve-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RetrainingPolicyRunRetrieve-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainingPolicyRunRetrieve$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
