% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retraining_settings_update.R
\docType{class}
\name{RetrainingSettingsUpdate}
\alias{RetrainingSettingsUpdate}
\title{RetrainingSettingsUpdate}
\format{
An \code{R6Class} generator object
}
\description{
RetrainingSettingsUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{credentialId}}{character \link{optional} ID of the credential used to refresh retraining dataset.}

\item{\code{datasetId}}{character \link{optional} ID of the retraining dataset.}

\item{\code{predictionEnvironmentId}}{character \link{optional} ID of the prediction environment to associate with the challengers created by retraining policies.}

\item{\code{retrainingUserId}}{character \link{optional} ID of the retraining user.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RetrainingSettingsUpdate-new}{\code{RetrainingSettingsUpdate$new()}}
\item \href{#method-RetrainingSettingsUpdate-validate}{\code{RetrainingSettingsUpdate$validate()}}
\item \href{#method-RetrainingSettingsUpdate-toJSON}{\code{RetrainingSettingsUpdate$toJSON()}}
\item \href{#method-RetrainingSettingsUpdate-fromJSON}{\code{RetrainingSettingsUpdate$fromJSON()}}
\item \href{#method-RetrainingSettingsUpdate-clone}{\code{RetrainingSettingsUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RetrainingSettingsUpdate-new"></a>}}
\if{latex}{\out{\hypertarget{method-RetrainingSettingsUpdate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainingSettingsUpdate$new(
  credentialId = NULL,
  datasetId = NULL,
  predictionEnvironmentId = NULL,
  retrainingUserId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{credentialId}}{ID of the credential used to refresh retraining dataset.}

\item{\code{datasetId}}{ID of the retraining dataset.}

\item{\code{predictionEnvironmentId}}{ID of the prediction environment to associate with the challengers created by retraining policies.}

\item{\code{retrainingUserId}}{ID of the retraining user.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RetrainingSettingsUpdate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-RetrainingSettingsUpdate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainingSettingsUpdate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RetrainingSettingsUpdate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RetrainingSettingsUpdate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainingSettingsUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RetrainingSettingsUpdate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RetrainingSettingsUpdate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainingSettingsUpdate$fromJSON(
  RetrainingSettingsUpdateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{RetrainingSettingsUpdateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RetrainingSettingsUpdate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RetrainingSettingsUpdate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainingSettingsUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
