% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrain_model.R
\docType{class}
\name{RetrainModel}
\alias{RetrainModel}
\title{RetrainModel}
\format{
An \code{R6Class} generator object
}
\description{
RetrainModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{featurelistId}}{character \link{optional} If specified, the model will be trained using that featurelist, otherwise the model will be trained on the same feature list as before.}

\item{\code{modelId}}{character The model to be retrained}

\item{\code{monotonicDecreasingFeaturelistId}}{character \link{optional} The ID of the featurelist that defines the set of features with a monotonically decreasing relationship to the target. If null, no such constraints are enforced.}

\item{\code{monotonicIncreasingFeaturelistId}}{character \link{optional} The ID of the featurelist that defines the set of features with a monotonically increasing relationship to the target. If null, no such constraints are enforced.}

\item{\code{nClusters}}{integer \link{optional} The number of clusters to use in the specified unsupervised clustering model. Only valid in unsupervised clustering projects.}

\item{\code{samplePct}}{numeric \link{optional} The percentage of the dataset to use to use to train the model. The specified percentage should be between 0 and 100. If not specified, original model sample percent will be used.}

\item{\code{scoringType}}{character \link{optional} Validation is available for any partitioning. If the project uses cross validation, `crossValidation` may be used to indicate that all available training/validation combinations should be used.}

\item{\code{trainingRowCount}}{integer \link{optional} The number of rows to use to train the model. If not specified, original model training row count will be used.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RetrainModel$new()}}
\item \href{#method-validate}{\code{RetrainModel$validate()}}
\item \href{#method-toJSON}{\code{RetrainModel$toJSON()}}
\item \href{#method-fromJSON}{\code{RetrainModel$fromJSON()}}
\item \href{#method-clone}{\code{RetrainModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainModel$new(
  modelId = NULL,
  featurelistId = NULL,
  monotonicDecreasingFeaturelistId = NULL,
  monotonicIncreasingFeaturelistId = NULL,
  nClusters = NULL,
  samplePct = NULL,
  scoringType = NULL,
  trainingRowCount = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{modelId}}{The model to be retrained}

\item{\code{featurelistId}}{If specified, the model will be trained using that featurelist, otherwise the model will be trained on the same feature list as before.}

\item{\code{monotonicDecreasingFeaturelistId}}{The ID of the featurelist that defines the set of features with a monotonically decreasing relationship to the target. If null, no such constraints are enforced.}

\item{\code{monotonicIncreasingFeaturelistId}}{The ID of the featurelist that defines the set of features with a monotonically increasing relationship to the target. If null, no such constraints are enforced.}

\item{\code{nClusters}}{The number of clusters to use in the specified unsupervised clustering model. Only valid in unsupervised clustering projects.}

\item{\code{samplePct}}{The percentage of the dataset to use to use to train the model. The specified percentage should be between 0 and 100. If not specified, original model sample percent will be used.}

\item{\code{scoringType}}{Validation is available for any partitioning. If the project uses cross validation, `crossValidation` may be used to indicate that all available training/validation combinations should be used.}

\item{\code{trainingRowCount}}{The number of rows to use to train the model. If not specified, original model training row count will be used.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainModel$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainModel$fromJSON(RetrainModelJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{RetrainModelJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrainModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
