% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications_apiWrapper.R
\name{RetrieveApplications}
\alias{RetrieveApplications}
\title{RetrieveApplications}
\usage{
RetrieveApplications(applicationId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve an application
}
\details{
Retrieve an application

This method invokes \verb{GET /applications/\{applicationId\}/} in the DataRobot Public API.
}
\seealso{
The method ApplicationsRetrieve in \link{ApplicationsApi}, which invokes the same underlying API endpoint.
}
