% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications_apiWrapper.R
\name{RetrieveApplicationUserRole}
\alias{RetrieveApplicationUserRole}
\title{RetrieveApplicationUserRole}
\usage{
RetrieveApplicationUserRole(applicationId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get application user role
}
\details{
Get application user role

This method invokes \verb{GET /applications/\{applicationId\}/userRole/} in the DataRobot Public API.
}
\seealso{
The method ApplicationUserRoleRetrieve in \link{ApplicationsApi}, which invokes the same underlying API endpoint.
}
