% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{RetrieveBatchPredictionJobDefinitions}
\alias{RetrieveBatchPredictionJobDefinitions}
\title{RetrieveBatchPredictionJobDefinitions}
\usage{
RetrieveBatchPredictionJobDefinitions(jobDefinitionId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve Batch Prediction job definition
}
\details{
Retrieve a Batch Prediction job definition

This method invokes \verb{GET /batchPredictionJobDefinitions/\{jobDefinitionId\}/} in the DataRobot Public API.
}
\seealso{
The method BatchPredictionJobDefinitionsRetrieve in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
