% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{RetrieveCatalogItems}
\alias{RetrieveCatalogItems}
\title{RetrieveCatalogItems}
\usage{
RetrieveCatalogItems(catalogId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieves latest version information, by ID, for catalog items.
}
\details{
Retrieves latest version information, by ID, for catalog items.

This method invokes \verb{GET /catalogItems/\{catalogId\}/} in the DataRobot Public API.
}
\seealso{
The method CatalogItemsRetrieve in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
