% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{RetrieveChangeRequests}
\alias{RetrieveChangeRequests}
\title{RetrieveChangeRequests}
\usage{
RetrieveChangeRequests(changeRequestId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve Change Request.
}
\details{
Retrieve Change Request by ID.

This method invokes \verb{GET /changeRequests/\{changeRequestId\}/} in the DataRobot Public API.
}
\seealso{
The method ChangeRequestsRetrieve in \link{GovernanceApi}, which invokes the same underlying API endpoint.
}
