% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance_apiWrapper.R
\name{RetrieveChangeRequestsReviews}
\alias{RetrieveChangeRequestsReviews}
\title{RetrieveChangeRequestsReviews}
\usage{
RetrieveChangeRequestsReviews(changeRequestId, reviewId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve review.
}
\details{
Retrieve a review by ID.

This method invokes \verb{GET /changeRequests/\{changeRequestId\}/reviews/\{reviewId\}/} in the DataRobot Public API.
}
\seealso{
The method ChangeRequestsReviewsRetrieve in \link{GovernanceApi}, which invokes the same underlying API endpoint.
}
