% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_apiWrapper.R
\name{RetrieveCredentials}
\alias{RetrieveCredentials}
\title{RetrieveCredentials}
\usage{
RetrieveCredentials(credentialId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the credentials set.
}
\details{
Return the credential set matching the specified ID.

This method invokes \verb{GET /credentials/\{credentialId\}/} in the DataRobot Public API.
}
\seealso{
The method CredentialsRetrieve in \link{CredentialsApi}, which invokes the same underlying API endpoint.
}
