% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{RetrieveCustomModelsVersions}
\alias{RetrieveCustomModelsVersions}
\title{RetrieveCustomModelsVersions}
\usage{
RetrieveCustomModelsVersions(customModelId, customModelVersionId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get custom model version.
}
\details{
Display a requested version of a custom model along with the files attached to it.

This method invokes \verb{GET /customModels/\{customModelId\}/versions/\{customModelVersionId\}/} in the DataRobot Public API.
}
\seealso{
The method CustomModelsVersionsRetrieve in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
