% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{RetrieveCustomModelTests}
\alias{RetrieveCustomModelTests}
\title{RetrieveCustomModelTests}
\usage{
RetrieveCustomModelTests(customModelTestId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get custom model test.
}
\details{
Retrieve a specific testing history entry for a custom model.

This method invokes \verb{GET /customModelTests/\{customModelTestId\}/} in the DataRobot Public API.
}
\seealso{
The method CustomModelTestsRetrieve in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
