% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{RetrieveCustomTasks}
\alias{RetrieveCustomTasks}
\title{RetrieveCustomTasks}
\usage{
RetrieveCustomTasks(customTaskId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get custom task.
}
\details{
Retrieve metadata for a custom task

This method invokes \verb{GET /customTasks/\{customTaskId\}/} in the DataRobot Public API.
}
\seealso{
The method CustomTasksRetrieve in \link{CustomTasksApi}, which invokes the same underlying API endpoint.
}
