% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tasks_apiWrapper.R
\name{RetrieveCustomTasksVersions}
\alias{RetrieveCustomTasksVersions}
\title{RetrieveCustomTasksVersions}
\usage{
RetrieveCustomTasksVersions(customTaskVersionId, customTaskId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get custom task version.
}
\details{
Display a requested version of a custom task along with the files attached to it.

This method invokes \verb{GET /customTasks/\{customTaskId\}/versions/\{customTaskVersionId\}/} in the DataRobot Public API.
}
\seealso{
The method CustomTasksVersionsRetrieve in \link{CustomTasksApi}, which invokes the same underlying API endpoint.
}
