% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{RetrieveDataEngineWorkspaceStates}
\alias{RetrieveDataEngineWorkspaceStates}
\title{RetrieveDataEngineWorkspaceStates}
\usage{
RetrieveDataEngineWorkspaceStates(workspaceStateId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Read Data Engine workspace state
}
\details{
Read and return previously stored Data Engine workspace state.

This method invokes \verb{GET /dataEngineWorkspaceStates/\{workspaceStateId\}/} in the DataRobot Public API.
}
\seealso{
The method DataEngineWorkspaceStatesRetrieve in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
