% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{RetrieveDatasetsFeatureTransforms}
\alias{RetrieveDatasetsFeatureTransforms}
\title{RetrieveDatasetsFeatureTransforms}
\usage{
RetrieveDatasetsFeatureTransforms(featureName, datasetId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get dataset feature transform
}
\details{
Retrieve the specified feature with descriptive information.

This method invokes \verb{GET /datasets/\{datasetId\}/featureTransforms/\{featureName\}/} in the DataRobot Public API.
}
\seealso{
The method DatasetsFeatureTransformsRetrieve in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
