% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{RetrieveDatasetsRefreshJobs}
\alias{RetrieveDatasetsRefreshJobs}
\title{RetrieveDatasetsRefreshJobs}
\usage{
RetrieveDatasetsRefreshJobs(jobId, datasetId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Gets configuration of a user scheduled dataset refresh job by job ID
}
\details{
Gets configuration of a user scheduled dataset refresh job by job ID.

This method invokes \verb{GET /datasets/\{datasetId\}/refreshJobs/\{jobId\}/} in the DataRobot Public API.
}
\seealso{
The method DatasetsRefreshJobsRetrieve in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
