% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R
\name{RetrieveDatasetsVersions}
\alias{RetrieveDatasetsVersions}
\title{RetrieveDatasetsVersions}
\usage{
RetrieveDatasetsVersions(datasetId, datasetVersionId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get dataset details by version
}
\details{
Retrieves the details of the dataset with given ID and version ID.

This method invokes \verb{GET /datasets/\{datasetId\}/versions/\{datasetVersionId\}/} in the DataRobot Public API.
}
\seealso{
The method DatasetsVersionsRetrieve in \link{AiCatalogApi}, which invokes the same underlying API endpoint.
}
