% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{RetrieveDeploymentsChallengers}
\alias{RetrieveDeploymentsChallengers}
\title{RetrieveDeploymentsChallengers}
\usage{
RetrieveDeploymentsChallengers(deploymentId, challengerId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get challenger model
}
\details{
Retrieve challenger model.

This method invokes \verb{GET /deployments/\{deploymentId\}/challengers/\{challengerId\}/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsChallengersRetrieve in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
