% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{RetrieveDiscardedFeaturesInformation}
\alias{RetrieveDiscardedFeaturesInformation}
\title{RetrieveDiscardedFeaturesInformation}
\usage{
RetrieveDiscardedFeaturesInformation(projectId, search = NULL, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get discarded features.
}
\details{
Get features which were discarded during feature reduction process.

This method invokes \verb{GET /projects/\{projectId\}/discardedFeatures/} in the DataRobot Public API.
}
\seealso{
The method DiscardedFeaturesList in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
