% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_driver_response.R
\docType{class}
\name{RetrieveDriverResponse}
\alias{RetrieveDriverResponse}
\title{RetrieveDriverResponse}
\format{
An \code{R6Class} generator object
}
\description{
RetrieveDriverResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{associatedAuthTypes}}{list( character ) \link{optional}}

\item{\code{baseNames}}{list( character )}

\item{\code{canonicalName}}{character}

\item{\code{className}}{character}

\item{\code{configurationId}}{character \link{optional}}

\item{\code{creator}}{character}

\item{\code{id}}{character}

\item{\code{version}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RetrieveDriverResponse$new()}}
\item \href{#method-toJSON}{\code{RetrieveDriverResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{RetrieveDriverResponse$fromJSON()}}
\item \href{#method-clone}{\code{RetrieveDriverResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrieveDriverResponse$new(
  baseNames,
  canonicalName,
  className,
  creator,
  id,
  associatedAuthTypes = NULL,
  configurationId = NULL,
  version = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrieveDriverResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrieveDriverResponse$fromJSON(RetrieveDriverResponseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrieveDriverResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
