% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{RetrieveGroups}
\alias{RetrieveGroups}
\title{RetrieveGroups}
\usage{
RetrieveGroups(groupId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve user group
}
\details{
Retrieve the user group.

This method invokes \verb{GET /groups/\{groupId\}/} in the DataRobot Public API.
}
\seealso{
The method GroupsRetrieve in \link{UserManagementApi}, which invokes the same underlying API endpoint.
}
