% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R
\name{RetrieveImageAugmentationLists}
\alias{RetrieveImageAugmentationLists}
\title{RetrieveImageAugmentationLists}
\usage{
RetrieveImageAugmentationLists(augmentationId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Returns a single augmentation list with the specified id.
}
\details{
Returns a single augmentation list with the specified id.

This method invokes \verb{GET /imageAugmentationLists/\{augmentationId\}/} in the DataRobot Public API.
}
\seealso{
The method ImageAugmentationListsRetrieve in \link{ImagesApi}, which invokes the same underlying API endpoint.
}
