% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_apiWrapper.R
\name{RetrieveImageAugmentationOptions}
\alias{RetrieveImageAugmentationOptions}
\title{RetrieveImageAugmentationOptions}
\usage{
RetrieveImageAugmentationOptions(projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Augmentation list of all available transformations that are supported in the system.
}
\details{
Augmentation list of all available transformations that are supported in the system.

This method invokes \verb{GET /imageAugmentationOptions/\{projectId\}/} in the DataRobot Public API.
}
\seealso{
The method ImageAugmentationOptionsRetrieve in \link{ImagesApi}, which invokes the same underlying API endpoint.
}
