% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_apiWrapper.R
\name{RetrieveModelComplianceDocsInitializations}
\alias{RetrieveModelComplianceDocsInitializations}
\title{RetrieveModelComplianceDocsInitializations}
\usage{
RetrieveModelComplianceDocsInitializations(entityId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Check if compliance documentation pre-processing is initialized for the model
}
\details{
Check if compliance documentation pre-processing is initialized for the current model. This is only required for custom models.

This method invokes \verb{GET /modelComplianceDocsInitializations/\{entityId\}/} in the DataRobot Public API.
}
\seealso{
The method ModelComplianceDocsInitializationsRetrieve in \link{DocumentationApi}, which invokes the same underlying API endpoint.
}
