% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{RetrieveModelPackages}
\alias{RetrieveModelPackages}
\title{RetrieveModelPackages}
\usage{
RetrieveModelPackages(modelPackageId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve info about a model package.
}
\details{
Retrieve info about a model package.

This method invokes \verb{GET /modelPackages/\{modelPackageId\}/} in the DataRobot Public API.
}
\seealso{
The method ModelPackagesRetrieve in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
