% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{RetrieveNotificationChannels}
\alias{RetrieveNotificationChannels}
\title{RetrieveNotificationChannels}
\usage{
RetrieveNotificationChannels(channelId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve notification channel
}
\details{
Retrieve the notification channel.

This method invokes \verb{GET /notificationChannels/\{channelId\}/} in the DataRobot Public API.
}
\seealso{
The method NotificationChannelsRetrieve in \link{NotificationsApi}, which invokes the same underlying API endpoint.
}
