% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{RetrieveNotificationPolicies}
\alias{RetrieveNotificationPolicies}
\title{RetrieveNotificationPolicies}
\usage{
RetrieveNotificationPolicies(policyId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve notification policy
}
\details{
Retrieve the notification policy.

This method invokes \verb{GET /notificationPolicies/\{policyId\}/} in the DataRobot Public API.
}
\seealso{
The method NotificationPoliciesRetrieve in \link{NotificationsApi}, which invokes the same underlying API endpoint.
}
