% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_apiWrapper.R
\name{RetrieveNotificationWebhookChannelTests}
\alias{RetrieveNotificationWebhookChannelTests}
\title{RetrieveNotificationWebhookChannelTests}
\usage{
RetrieveNotificationWebhookChannelTests(notificationId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve status of notification channel test
}
\details{
Retrieve the status of the notification channel test.

This method invokes \verb{GET /notificationWebhookChannelTests/\{notificationId\}/} in the DataRobot Public API.
}
\seealso{
The method NotificationWebhookChannelTestsRetrieve in \link{NotificationsApi}, which invokes the same underlying API endpoint.
}
