% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{RetrieveOrganizationsUsers}
\alias{RetrieveOrganizationsUsers}
\title{RetrieveOrganizationsUsers}
\usage{
RetrieveOrganizationsUsers(organizationId, userId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve user from organization
}
\details{
Retrieve the user from this organization.

This method invokes \verb{GET /organizations/\{organizationId\}/users/\{userId\}/} in the DataRobot Public API.
}
\seealso{
The method OrganizationsUsersRetrieve in \link{UserManagementApi}, which invokes the same underlying API endpoint.
}
