% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_prediction_metadata_object.R
\docType{class}
\name{RetrievePredictionMetadataObject}
\alias{RetrievePredictionMetadataObject}
\title{RetrievePredictionMetadataObject}
\format{
An \code{R6Class} generator object
}
\description{
RetrievePredictionMetadataObject Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actualValueColumn}}{character \link{optional} For time series unsupervised projects only. Actual value column can be used to calculate the classification metrics and insights.}

\item{\code{datasetId}}{character \link{optional} Deprecated alias for `predictionDatasetId`.}

\item{\code{explanationAlgorithm}}{character \link{optional} The selected algorithm to use for prediction explanations. At present, the only acceptable value is `shap`, which selects the SHapley Additive exPlanations (SHAP) explainer. Defaults to null (no prediction explanations).}

\item{\code{featureDerivationWindowCounts}}{integer \link{optional} For time series projects with partial history only. Indicates how many points were used in during feature derivation.}

\item{\code{forecastPoint}}{character \link{optional} For time series projects only. The time in the dataset relative to which predictions were generated.}

\item{\code{id}}{character The id of the prediction record.}

\item{\code{includesPredictionIntervals}}{character Whether the predictions include prediction intervals.}

\item{\code{maxExplanations}}{integer \link{optional} The maximum number of prediction explanations values to be returned with each row in the `predictions` json array. Null indicates `no limit`. Will be present only if `explanationAlgorithm` was set.}

\item{\code{modelId}}{character The model id used for predictions.}

\item{\code{predictionDatasetId}}{character \link{optional} The dataset id where the prediction data comes from. The field is available via `/api/v2/projects/<projectId>/predictionsMetadata/` route and replaced on `datasetId`in deprecated `/api/v2/projects/<projectId>/predictions/` endpoint.}

\item{\code{predictionIntervalsSize}}{integer For time series projects only. If prediction intervals were computed, what percentile they represent. Will be ``None`` if ``includePredictionIntervals`` is ``False``.}

\item{\code{predictionThreshold}}{numeric \link{optional} Threshold used for binary classification in predictions.}

\item{\code{predictionsEndDate}}{character \link{optional} For time series projects only. The end date for bulk predictions, exclusive. Note that this parameter was used for generating historical predictions using the training data, not for future predictions.}

\item{\code{predictionsStartDate}}{character \link{optional} For time series projects only. The start date for bulk predictions. Note that this parameter was used for generating historical predictions using the training data, not for future predictions.}

\item{\code{projectId}}{character The project id of the predictions.}

\item{\code{shapWarnings}}{\link{ShapWarnings} \link{optional}}

\item{\code{url}}{character The url at which you can download the predictions.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RetrievePredictionMetadataObject$new()}}
\item \href{#method-validate}{\code{RetrievePredictionMetadataObject$validate()}}
\item \href{#method-toJSON}{\code{RetrievePredictionMetadataObject$toJSON()}}
\item \href{#method-fromJSON}{\code{RetrievePredictionMetadataObject$fromJSON()}}
\item \href{#method-clone}{\code{RetrievePredictionMetadataObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrievePredictionMetadataObject$new(
  id = NULL,
  includesPredictionIntervals = NULL,
  modelId = NULL,
  predictionIntervalsSize = NULL,
  projectId = NULL,
  url = NULL,
  actualValueColumn = NULL,
  datasetId = NULL,
  explanationAlgorithm = NULL,
  featureDerivationWindowCounts = NULL,
  forecastPoint = NULL,
  maxExplanations = NULL,
  predictionDatasetId = NULL,
  predictionThreshold = NULL,
  predictionsEndDate = NULL,
  predictionsStartDate = NULL,
  shapWarnings = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of the prediction record.}

\item{\code{includesPredictionIntervals}}{Whether the predictions include prediction intervals.}

\item{\code{modelId}}{The model id used for predictions.}

\item{\code{predictionIntervalsSize}}{For time series projects only. If prediction intervals were computed, what percentile they represent. Will be ``None`` if ``includePredictionIntervals`` is ``False``.}

\item{\code{projectId}}{The project id of the predictions.}

\item{\code{url}}{The url at which you can download the predictions.}

\item{\code{actualValueColumn}}{For time series unsupervised projects only. Actual value column can be used to calculate the classification metrics and insights.}

\item{\code{datasetId}}{Deprecated alias for `predictionDatasetId`.}

\item{\code{explanationAlgorithm}}{The selected algorithm to use for prediction explanations. At present, the only acceptable value is `shap`, which selects the SHapley Additive exPlanations (SHAP) explainer. Defaults to null (no prediction explanations).}

\item{\code{featureDerivationWindowCounts}}{For time series projects with partial history only. Indicates how many points were used in during feature derivation.}

\item{\code{forecastPoint}}{For time series projects only. The time in the dataset relative to which predictions were generated.}

\item{\code{maxExplanations}}{The maximum number of prediction explanations values to be returned with each row in the `predictions` json array. Null indicates `no limit`. Will be present only if `explanationAlgorithm` was set.}

\item{\code{predictionDatasetId}}{The dataset id where the prediction data comes from. The field is available via `/api/v2/projects/<projectId>/predictionsMetadata/` route and replaced on `datasetId`in deprecated `/api/v2/projects/<projectId>/predictions/` endpoint.}

\item{\code{predictionThreshold}}{Threshold used for binary classification in predictions.}

\item{\code{predictionsEndDate}}{For time series projects only. The end date for bulk predictions, exclusive. Note that this parameter was used for generating historical predictions using the training data, not for future predictions.}

\item{\code{predictionsStartDate}}{For time series projects only. The start date for bulk predictions. Note that this parameter was used for generating historical predictions using the training data, not for future predictions.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrievePredictionMetadataObject$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrievePredictionMetadataObject$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrievePredictionMetadataObject$fromJSON(
  RetrievePredictionMetadataObjectJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{RetrievePredictionMetadataObjectJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrievePredictionMetadataObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
