% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{RetrieveProjectCleanupJobs}
\alias{RetrieveProjectCleanupJobs}
\title{RetrieveProjectCleanupJobs}
\usage{
RetrieveProjectCleanupJobs(statusId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve Project Permadelete job status
}
\details{
Get async status of the project permadelete job

This method invokes \verb{GET /projectCleanupJobs/\{statusId\}/} in the DataRobot Public API.
}
\seealso{
The method CleanupJobsRetrieve in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
