% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{RetrieveProjectsCombinedModels}
\alias{RetrieveProjectsCombinedModels}
\title{RetrieveProjectsCombinedModels}
\usage{
RetrieveProjectsCombinedModels(combinedModelId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve an existing combined model.
}
\details{
Retrieve an existing combined model. If available, contains information on which champion model is used for each segment.

This method invokes \verb{GET /projects/\{projectId\}/combinedModels/\{combinedModelId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsCombinedModelsRetrieve in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
