% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{RetrieveProjectsDatetimeModelsDatasetAccuracyOverTimePlots}
\alias{RetrieveProjectsDatetimeModelsDatasetAccuracyOverTimePlots}
\title{RetrieveProjectsDatetimeModelsDatasetAccuracyOverTimePlots}
\usage{
RetrieveProjectsDatetimeModelsDatasetAccuracyOverTimePlots(
  endDate,
  modelId,
  datasetId,
  resolution,
  projectId,
  startDate,
  maxBinSize = 500,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the Accuracy Over Time (AOT) chart data for an external dataset for a project.
}
\details{
Retrieve the Accuracy Over Time (AOT) chart data for an external dataset for a project. Datetimes are specified in accordance with \verb{timestamp and duration formatting rules <time_format>}.

This method invokes \verb{GET /projects/\{projectId\}/datetimeModels/\{modelId\}/datasetAccuracyOverTimePlots/\{datasetId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsDatetimeModelsDatasetAccuracyOverTimePlotsRetrieve in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
