% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_apiWrapper.R
\name{RetrieveProjectsEureqaDistributionPlot}
\alias{RetrieveProjectsEureqaDistributionPlot}
\title{RetrieveProjectsEureqaDistributionPlot}
\usage{
RetrieveProjectsEureqaDistributionPlot(solutionId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve Eureqa model details plot.
}
\details{
Retrieve Eureqa model details plot.  Available for classification projects only

This method invokes \verb{GET /projects/\{projectId\}/eureqaDistributionPlot/\{solutionId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsEureqaDistributionPlotRetrieve in \link{ModelsApi}, which invokes the same underlying API endpoint.
}
