% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{RetrieveProjectsGeometryFeaturePlots}
\alias{RetrieveProjectsGeometryFeaturePlots}
\title{RetrieveProjectsGeometryFeaturePlots}
\usage{
RetrieveProjectsGeometryFeaturePlots(featureName, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a map of one location feature
}
\details{
Retrieve a map of one location feature

This method invokes \verb{GET /projects/\{projectId\}/geometryFeaturePlots/\{featureName\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsGeometryFeaturePlotsRetrieve in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
