% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{RetrieveProjectsModelsDatasetConfusionCharts}
\alias{RetrieveProjectsModelsDatasetConfusionCharts}
\title{RetrieveProjectsModelsDatasetConfusionCharts}
\usage{
RetrieveProjectsModelsDatasetConfusionCharts(
  colEnd,
  modelId,
  colStart,
  datasetId,
  rowEnd,
  rowStart,
  projectId,
  orientation = "actual",
  orderBy = "className",
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve Confusion Chart objects on external datasets for a project.
}
\details{
Retrieve Confusion Chart objects on external datasets for a project. Prediction dataset may have Confusion Chart for multiclass projects computed if it contained a target with actual values and insights were computed on this dataset. A confusion chart consists of the confusion matrix for all classes, classes frequencies and oneVsAll metrics for all classes. The confusion matrix can be requested in a particular sort order and oriented by rows or columns (zero-indexed). Available for multiclass projects only.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/datasetConfusionCharts/\{datasetId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsDatasetConfusionChartsRetrieve in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
