% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{RetrieveProjectsModelsLiftChart}
\alias{RetrieveProjectsModelsLiftChart}
\title{RetrieveProjectsModelsLiftChart}
\usage{
RetrieveProjectsModelsLiftChart(modelId, source, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the lift chart data from a single source.
}
\details{
Retrieve the lift chart data from a single source.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/liftChart/\{source\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsLiftChartRetrieve in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
