% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{RetrieveProjectsModelsRocCurve}
\alias{RetrieveProjectsModelsRocCurve}
\title{RetrieveProjectsModelsRocCurve}
\usage{
RetrieveProjectsModelsRocCurve(modelId, source, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve the ROC curve data from a single source.
}
\details{
Retrieve the ROC curve data from a single source. The response includes an array of pointsshowing the performance of the model at different thresholds for classification, and arrays of sample predictions for both the positive and negative classes.

This method invokes \verb{GET /projects/\{projectId\}/models/\{modelId\}/rocCurve/\{source\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsModelsRocCurveRetrieve in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
