% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{RetrieveProjectsPredictions}
\alias{RetrieveProjectsPredictions}
\title{RetrieveProjectsPredictions}
\usage{
RetrieveProjectsPredictions(
  shapMulticlassLevel,
  projectId,
  predictionId,
  accept,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get a completed set of predictions.
}
\details{
Retrieve predictions that have previously been computed. Training predictions encoded either as JSON or CSV. If CSV output was requested, the returned CSV data will contain the following columns:  - For regression projects: row_id and prediction. - For binary classification projects: row_id, prediction,   class\\\emph{\\<positive_class_label\\> and class\\}\\<negative_class_label\\>. - For multiclass projects: row_id, prediction and a   class\\\emph{\\<class_label\\> for each class. - For multilabel projects: row_id and for each class   prediction\\}\\<class_label\\> and class\\\emph{\\<class_label\\>. - For time-series, these additional columns will be added:   forecast_point, forecast_distance, timestamp, and series_id.  v2.21  - If explanationAlgorithm = 'shap', these additional columns will be   added: triplets of (Explanation\\}\\<i\\>\\\emph{feature_name,   Explanation\\}\\<i\\>\\\emph{feature_value, and Explanation\\}\\<i\\>\\_strength)   for i ranging from 1 to maxExplanations, shap_remaining_total and   shap_base_value. Binary classification projects will also have   explained_class, the class for which positive SHAP values imply an   increased probability.

This method invokes \verb{GET /projects/\{projectId\}/predictions/\{predictionId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPredictionsRetrieve in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
