% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{RetrieveProjectsPredictionsMetadata}
\alias{RetrieveProjectsPredictionsMetadata}
\title{RetrieveProjectsPredictionsMetadata}
\usage{
RetrieveProjectsPredictionsMetadata(projectId, predictionId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve metadata for a set of predictions.
}
\details{
Use the ID of a metadata object to get the complete set of predictions.

This method invokes \verb{GET /projects/\{projectId\}/predictionsMetadata/\{predictionId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPredictionsMetadataRetrieve in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
