% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{RetrieveProjectsSecondaryDatasetsConfigurations}
\alias{RetrieveProjectsSecondaryDatasetsConfigurations}
\title{RetrieveProjectsSecondaryDatasetsConfigurations}
\usage{
RetrieveProjectsSecondaryDatasetsConfigurations(
  secondaryDatasetConfigId,
  projectId,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve secondary dataset configuration by ID.
}
\details{
Retrieve secondary dataset configuration by ID.

This method invokes \verb{GET /projects/\{projectId\}/secondaryDatasetsConfigurations/\{secondaryDatasetConfigId\}/} in the DataRobot Public API.
}
\seealso{
The method SecondaryDatasetsConfigurationsRetrieve in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
