% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{RetrieveProjectsSegmentationTasks}
\alias{RetrieveProjectsSegmentationTasks}
\title{RetrieveProjectsSegmentationTasks}
\usage{
RetrieveProjectsSegmentationTasks(segmentationTaskId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve segmentation task.
}
\details{
Retrieve information about a segmentation task.

This method invokes \verb{GET /projects/\{projectId\}/segmentationTasks/\{segmentationTaskId\}/} in the DataRobot Public API.
}
\seealso{
The method SegmentationTasksRetrieve in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
