% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_apiWrapper.R
\name{RetrieveProjectsShapMatrices}
\alias{RetrieveProjectsShapMatrices}
\title{RetrieveProjectsShapMatrices}
\usage{
RetrieveProjectsShapMatrices(shapMatrixId, projectId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get matrix with SHAP prediction explanations scores.
}
\details{
Get matrix with SHAP prediction explanations scores.

This method invokes \verb{GET /projects/\{projectId\}/shapMatrices/\{shapMatrixId\}/} in the DataRobot Public API.
}
\seealso{
The method ProjectsShapMatricesRetrieve in \link{InsightsApi}, which invokes the same underlying API endpoint.
}
