% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{RetrieveScheduledJobs}
\alias{RetrieveScheduledJobs}
\title{RetrieveScheduledJobs}
\usage{
RetrieveScheduledJobs(jobId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
List a single deployment batch prediction job
}
\details{
Get a scheduled batch prediction job

This method invokes \verb{GET /scheduledJobs/\{jobId\}/} in the DataRobot Public API.
}
\seealso{
The method ScheduledJobsRetrieve in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
