% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sso_configuration_apiWrapper.R
\name{RetrieveSsoConfigurations}
\alias{RetrieveSsoConfigurations}
\title{RetrieveSsoConfigurations}
\usage{
RetrieveSsoConfigurations(configurationId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve SSO configuration of a specific organization.
}
\details{
Retrieve SSO configuration of a specific organization.

This method invokes \verb{GET /ssoConfigurations/\{configurationId\}/} in the DataRobot Public API.
}
\seealso{
The method SsoConfigurationsRetrieve in \link{SsoConfigurationApi}, which invokes the same underlying API endpoint.
}
