% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{RetrieveUseCases}
\alias{RetrieveUseCases}
\title{RetrieveUseCases}
\usage{
RetrieveUseCases(useCaseId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a use case.
}
\details{
Retrieve a use case.

This method invokes \verb{GET /useCases/\{useCaseId\}/} in the DataRobot Public API.
}
\seealso{
The method UseCasesRetrieve in \link{UseCaseApi}, which invokes the same underlying API endpoint.
}
