% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_apiWrapper.R
\name{RetrieveUseCaseValueTemplates}
\alias{RetrieveUseCaseValueTemplates}
\title{RetrieveUseCaseValueTemplates}
\usage{
RetrieveUseCaseValueTemplates(templateType, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Get an individual use case value template by its name.
}
\details{
Get an individual use case value template by its name.

This method invokes \verb{GET /useCaseValueTemplates/\{templateType\}/} in the DataRobot Public API.
}
\seealso{
The method UseCaseValueTemplatesRetrieve in \link{UseCaseApi}, which invokes the same underlying API endpoint.
}
