% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_catalog_apiWrapper.R,
%   R/blueprints_apiWrapper.R
\name{RetrieveUserBlueprints}
\alias{RetrieveUserBlueprints}
\title{RetrieveUserBlueprints}
\usage{
RetrieveUserBlueprints(
  userBlueprintId,
  projectId,
  isInplaceEditor = FALSE,
  getDynamicLabels = FALSE,
  decompressedBlueprint = FALSE,
  editMode = FALSE,
  ...
)

RetrieveUserBlueprints(
  userBlueprintId,
  projectId,
  isInplaceEditor = FALSE,
  getDynamicLabels = FALSE,
  decompressedBlueprint = FALSE,
  editMode = FALSE,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve a user blueprint.

Retrieve a user blueprint.
}
\details{
Retrieve a user blueprint

This method invokes \verb{GET /userBlueprints/\{userBlueprintId\}/} in the DataRobot Public API.

Retrieve a user blueprint

This method invokes \verb{GET /userBlueprints/\{userBlueprintId\}/} in the DataRobot Public API.
}
\seealso{
The method UserBlueprintsRetrieve in \link{AiCatalogApi}, which invokes the same underlying API endpoint.

The method UserBlueprintsRetrieve in \link{BlueprintsApi}, which invokes the same underlying API endpoint.
}
