% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_apiWrapper.R
\name{RetrieveUserCleanupJobs}
\alias{RetrieveUserCleanupJobs}
\title{RetrieveUserCleanupJobs}
\usage{
RetrieveUserCleanupJobs(statusId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Retrieve users perma-delete job status.
}
\details{
Get async status of the users perma-delete job.

This method invokes \verb{GET /userCleanupJobs/\{statusId\}/} in the DataRobot Public API.
}
\seealso{
The method UserCleanupJobsRetrieve in \link{UserManagementApi}, which invokes the same underlying API endpoint.
}
