% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_points_response.R
\docType{class}
\name{RocPointsResponse}
\alias{RocPointsResponse}
\title{RocPointsResponse}
\format{
An \code{R6Class} generator object
}
\description{
RocPointsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{accuracy}}{numeric Accuracy for given threshold.}

\item{\code{f1Score}}{numeric F1 score.}

\item{\code{falseNegativeScore}}{integer False negative score.}

\item{\code{falsePositiveRate}}{numeric False positive rate.}

\item{\code{falsePositiveScore}}{integer False positive score.}

\item{\code{fractionPredictedAsNegative}}{numeric Fraction of data that will be predicted as negative.}

\item{\code{fractionPredictedAsPositive}}{numeric Fraction of data that will be predicted as positive.}

\item{\code{liftNegative}}{numeric Lift for the negative class.}

\item{\code{liftPositive}}{numeric Lift for the positive class.}

\item{\code{matthewsCorrelationCoefficient}}{numeric Matthews correlation coefficient.}

\item{\code{negativePredictiveValue}}{numeric Negative predictive value.}

\item{\code{positivePredictiveValue}}{numeric Positive predictive value.}

\item{\code{threshold}}{numeric Value of threshold for this ROC point.}

\item{\code{trueNegativeRate}}{numeric True negative rate.}

\item{\code{trueNegativeScore}}{integer True negative score.}

\item{\code{truePositiveRate}}{numeric True positive rate.}

\item{\code{truePositiveScore}}{integer True positive score.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RocPointsResponse-new}{\code{RocPointsResponse$new()}}
\item \href{#method-RocPointsResponse-validate}{\code{RocPointsResponse$validate()}}
\item \href{#method-RocPointsResponse-toJSON}{\code{RocPointsResponse$toJSON()}}
\item \href{#method-RocPointsResponse-fromJSON}{\code{RocPointsResponse$fromJSON()}}
\item \href{#method-RocPointsResponse-clone}{\code{RocPointsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RocPointsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-RocPointsResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RocPointsResponse$new(
  accuracy = NULL,
  f1Score = NULL,
  falseNegativeScore = NULL,
  falsePositiveRate = NULL,
  falsePositiveScore = NULL,
  fractionPredictedAsNegative = NULL,
  fractionPredictedAsPositive = NULL,
  liftNegative = NULL,
  liftPositive = NULL,
  matthewsCorrelationCoefficient = NULL,
  negativePredictiveValue = NULL,
  positivePredictiveValue = NULL,
  threshold = NULL,
  trueNegativeRate = NULL,
  trueNegativeScore = NULL,
  truePositiveRate = NULL,
  truePositiveScore = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{accuracy}}{Accuracy for given threshold.}

\item{\code{f1Score}}{F1 score.}

\item{\code{falseNegativeScore}}{False negative score.}

\item{\code{falsePositiveRate}}{False positive rate.}

\item{\code{falsePositiveScore}}{False positive score.}

\item{\code{fractionPredictedAsNegative}}{Fraction of data that will be predicted as negative.}

\item{\code{fractionPredictedAsPositive}}{Fraction of data that will be predicted as positive.}

\item{\code{liftNegative}}{Lift for the negative class.}

\item{\code{liftPositive}}{Lift for the positive class.}

\item{\code{matthewsCorrelationCoefficient}}{Matthews correlation coefficient.}

\item{\code{negativePredictiveValue}}{Negative predictive value.}

\item{\code{positivePredictiveValue}}{Positive predictive value.}

\item{\code{threshold}}{Value of threshold for this ROC point.}

\item{\code{trueNegativeRate}}{True negative rate.}

\item{\code{trueNegativeScore}}{True negative score.}

\item{\code{truePositiveRate}}{True positive rate.}

\item{\code{truePositiveScore}}{True positive score.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RocPointsResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-RocPointsResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RocPointsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RocPointsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RocPointsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RocPointsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RocPointsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RocPointsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RocPointsResponse$fromJSON(RocPointsResponseJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{RocPointsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RocPointsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RocPointsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RocPointsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
