% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_credentials.R
\docType{class}
\name{S3Credentials}
\alias{S3Credentials}
\title{S3Credentials}
\format{
An \code{R6Class} generator object
}
\description{
S3Credentials Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{awsAccessKeyId}}{character \link{optional} The S3 AWS access key ID.}

\item{\code{awsSecretAccessKey}}{character \link{optional} The S3 AWS secret access key.}

\item{\code{awsSessionToken}}{character \link{optional} The S3 AWS session token.}

\item{\code{credentialType}}{character The type of these credentials, 's3' here.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-S3Credentials-new}{\code{S3Credentials$new()}}
\item \href{#method-S3Credentials-validate}{\code{S3Credentials$validate()}}
\item \href{#method-S3Credentials-toJSON}{\code{S3Credentials$toJSON()}}
\item \href{#method-S3Credentials-fromJSON}{\code{S3Credentials$fromJSON()}}
\item \href{#method-S3Credentials-clone}{\code{S3Credentials$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Credentials-new"></a>}}
\if{latex}{\out{\hypertarget{method-S3Credentials-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Credentials$new(
  credentialType = NULL,
  awsAccessKeyId = NULL,
  awsSecretAccessKey = NULL,
  awsSessionToken = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{credentialType}}{The type of these credentials, 's3' here.}

\item{\code{awsAccessKeyId}}{The S3 AWS access key ID.}

\item{\code{awsSecretAccessKey}}{The S3 AWS secret access key.}

\item{\code{awsSessionToken}}{The S3 AWS session token.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Credentials-validate"></a>}}
\if{latex}{\out{\hypertarget{method-S3Credentials-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Credentials$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Credentials-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-S3Credentials-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Credentials$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Credentials-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-S3Credentials-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Credentials$fromJSON(S3CredentialsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{S3CredentialsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Credentials-clone"></a>}}
\if{latex}{\out{\hypertarget{method-S3Credentials-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Credentials$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
