% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saml_advanced_configuration.R
\docType{class}
\name{SamlAdvancedConfiguration}
\alias{SamlAdvancedConfiguration}
\title{SamlAdvancedConfiguration}
\format{
An \code{R6Class} generator object
}
\description{
SamlAdvancedConfiguration Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{digestAlgorithm}}{character \link{optional} Algorithm for calculating digest.}

\item{\code{samlAttributesMapping}}{\link{SamlAttributeMapping}}

\item{\code{samlClientConfiguration}}{\link{SamlClientConfig}}

\item{\code{signatureAlgorithm}}{character \link{optional} Algorithm for calculating signature.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SamlAdvancedConfiguration$new()}}
\item \href{#method-validate}{\code{SamlAdvancedConfiguration$validate()}}
\item \href{#method-toJSON}{\code{SamlAdvancedConfiguration$toJSON()}}
\item \href{#method-fromJSON}{\code{SamlAdvancedConfiguration$fromJSON()}}
\item \href{#method-clone}{\code{SamlAdvancedConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlAdvancedConfiguration$new(
  samlAttributesMapping = NULL,
  samlClientConfiguration = NULL,
  digestAlgorithm = NULL,
  signatureAlgorithm = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digestAlgorithm}}{Algorithm for calculating digest.}

\item{\code{signatureAlgorithm}}{Algorithm for calculating signature.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlAdvancedConfiguration$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlAdvancedConfiguration$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlAdvancedConfiguration$fromJSON(
  SamlAdvancedConfigurationJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SamlAdvancedConfigurationJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlAdvancedConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
