% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saml_attribute_mapping.R
\docType{class}
\name{SamlAttributeMapping}
\alias{SamlAttributeMapping}
\title{SamlAttributeMapping}
\format{
An \code{R6Class} generator object
}
\description{
SamlAttributeMapping Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{displayName}}{character \link{optional} Display name.}

\item{\code{email}}{character \link{optional} Email.}

\item{\code{firstName}}{character \link{optional} First name.}

\item{\code{group}}{character \link{optional} Group.}

\item{\code{impersonationUser}}{character \link{optional} Impersonation user.}

\item{\code{lastName}}{character \link{optional} Last name.}

\item{\code{role}}{character \link{optional} Role.}

\item{\code{username}}{character \link{optional} Username.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SamlAttributeMapping$new()}}
\item \href{#method-validate}{\code{SamlAttributeMapping$validate()}}
\item \href{#method-toJSON}{\code{SamlAttributeMapping$toJSON()}}
\item \href{#method-fromJSON}{\code{SamlAttributeMapping$fromJSON()}}
\item \href{#method-clone}{\code{SamlAttributeMapping$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlAttributeMapping$new(
  displayName = NULL,
  email = NULL,
  firstName = NULL,
  group = NULL,
  impersonationUser = NULL,
  lastName = NULL,
  role = NULL,
  username = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{displayName}}{Display name.}

\item{\code{email}}{Email.}

\item{\code{firstName}}{First name.}

\item{\code{group}}{Group.}

\item{\code{impersonationUser}}{Impersonation user.}

\item{\code{lastName}}{Last name.}

\item{\code{role}}{Role.}

\item{\code{username}}{Username.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlAttributeMapping$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlAttributeMapping$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlAttributeMapping$fromJSON(SamlAttributeMappingJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SamlAttributeMappingJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlAttributeMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
