% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saml_security_parameters.R
\docType{class}
\name{SamlSecurityParameters}
\alias{SamlSecurityParameters}
\title{SamlSecurityParameters}
\format{
An \code{R6Class} generator object
}
\description{
SamlSecurityParameters Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{allowUnsolicited}}{character \link{optional} Allow unsolicited.}

\item{\code{authnRequestsSigned}}{character \link{optional} Sign auth requests.}

\item{\code{logoutRequestsSigned}}{character \link{optional} Sign logout requests.}

\item{\code{wantAssertionsSigned}}{character \link{optional} Sign assertions.}

\item{\code{wantResponseSigned}}{character \link{optional} Sign response.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SamlSecurityParameters$new()}}
\item \href{#method-validate}{\code{SamlSecurityParameters$validate()}}
\item \href{#method-toJSON}{\code{SamlSecurityParameters$toJSON()}}
\item \href{#method-fromJSON}{\code{SamlSecurityParameters$fromJSON()}}
\item \href{#method-clone}{\code{SamlSecurityParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlSecurityParameters$new(
  allowUnsolicited = NULL,
  authnRequestsSigned = NULL,
  logoutRequestsSigned = NULL,
  wantAssertionsSigned = NULL,
  wantResponseSigned = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allowUnsolicited}}{Allow unsolicited.}

\item{\code{authnRequestsSigned}}{Sign auth requests.}

\item{\code{logoutRequestsSigned}}{Sign logout requests.}

\item{\code{wantAssertionsSigned}}{Sign assertions.}

\item{\code{wantResponseSigned}}{Sign response.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlSecurityParameters$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlSecurityParameters$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlSecurityParameters$fromJSON(
  SamlSecurityParametersJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SamlSecurityParametersJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamlSecurityParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
