% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduled_job_payload.R
\docType{class}
\name{ScheduledJobPayload}
\alias{ScheduledJobPayload}
\title{ScheduledJobPayload}
\format{
An \code{R6Class} generator object
}
\description{
ScheduledJobPayload Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dayOfMonth}}{list( \link{AnyOfnumberstring} ) \link{optional} The date(s) of the month that the job will run. Allowed values are either ``\link{1 ... 31}`` or ``\link[=\\"*\\"]{\"*\"}`` for all days of the month. This field is additive with ``dayOfWeek``, meaning the job will run both on the date(s) defined in this field and the day specified by ``dayOfWeek`` (for example, dates 1st, 2nd, 3rd, plus every Tuesday). If ``dayOfMonth`` is set to ``\link[=\\"*\\"]{\"*\"}`` and ``dayOfWeek`` is defined, the scheduler will trigger on every day of the month that matches ``dayOfWeek`` (for example, Tuesday the 2nd, 9th, 16th, 23rd, 30th). Invalid dates such as February 31st are ignored.}

\item{\code{dayOfWeek}}{list( \link{AnyOfnumberstring} ) \link{optional} The day(s) of the week that the job will run. Allowed values are ``\link{0 .. 6}``, where (Sunday=0), or ``\link[=\\"*\\"]{\"*\"}``, for all days of the week. Strings, either 3-letter abbreviations or the full name of the day, can be used interchangeably (e.g., \"sunday\", \"Sunday\", \"sun\", or \"Sun\", all map to ``\link{0}``. This field is additive with ``dayOfMonth``, meaning the job will run both on the date specified by ``dayOfMonth`` and the day defined in this field.}

\item{\code{hour}}{list( \link{AnyOfnumberstring} ) \link{optional} The hour(s) of the day that the job will run. Allowed values are either ``\link[=\\"*\\"]{\"*\"}`` meaning every hour of the day or ``\link{0 ... 23}``.}

\item{\code{minute}}{\link{OneOfstringarray} \link{optional} The minute that will trigger the refresh, allowed values are 0 to 59.}

\item{\code{month}}{list( \link{AnyOfnumberstring} ) \link{optional} The month(s) of the year that the job will run. Allowed values are either ``\link{1 ... 12}`` or ``\link[=\\"*\\"]{\"*\"}`` for all months of the year. Strings, either 3-letter abbreviations or the full name of the month, can be used interchangeably (e.g., \"jan\" or \"october\"). Months that are not compatible with ``dayOfMonth`` are ignored, for example ``{\"dayOfMonth\": \link{31}, \"month\":\link[=\\"feb\\"]{\"feb\"}}``.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScheduledJobPayload$new()}}
\item \href{#method-validate}{\code{ScheduledJobPayload$validate()}}
\item \href{#method-toJSON}{\code{ScheduledJobPayload$toJSON()}}
\item \href{#method-fromJSON}{\code{ScheduledJobPayload$fromJSON()}}
\item \href{#method-clone}{\code{ScheduledJobPayload$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledJobPayload$new(
  dayOfMonth = c("*"),
  dayOfWeek = c("*"),
  hour = c("*"),
  minute = NULL,
  month = c("*"),
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dayOfMonth}}{The date(s) of the month that the job will run. Allowed values are either ``\link{1 ... 31}`` or ``\link[=\\"*\\"]{\"*\"}`` for all days of the month. This field is additive with ``dayOfWeek``, meaning the job will run both on the date(s) defined in this field and the day specified by ``dayOfWeek`` (for example, dates 1st, 2nd, 3rd, plus every Tuesday). If ``dayOfMonth`` is set to ``\link[=\\"*\\"]{\"*\"}`` and ``dayOfWeek`` is defined, the scheduler will trigger on every day of the month that matches ``dayOfWeek`` (for example, Tuesday the 2nd, 9th, 16th, 23rd, 30th). Invalid dates such as February 31st are ignored.}

\item{\code{dayOfWeek}}{The day(s) of the week that the job will run. Allowed values are ``\link{0 .. 6}``, where (Sunday=0), or ``\link[=\\"*\\"]{\"*\"}``, for all days of the week. Strings, either 3-letter abbreviations or the full name of the day, can be used interchangeably (e.g., \"sunday\", \"Sunday\", \"sun\", or \"Sun\", all map to ``\link{0}``. This field is additive with ``dayOfMonth``, meaning the job will run both on the date specified by ``dayOfMonth`` and the day defined in this field.}

\item{\code{hour}}{The hour(s) of the day that the job will run. Allowed values are either ``\link[=\\"*\\"]{\"*\"}`` meaning every hour of the day or ``\link{0 ... 23}``.}

\item{\code{minute}}{The minute that will trigger the refresh, allowed values are 0 to 59.}

\item{\code{month}}{The month(s) of the year that the job will run. Allowed values are either ``\link{1 ... 12}`` or ``\link[=\\"*\\"]{\"*\"}`` for all months of the year. Strings, either 3-letter abbreviations or the full name of the month, can be used interchangeably (e.g., \"jan\" or \"october\"). Months that are not compatible with ``dayOfMonth`` are ignored, for example ``{\"dayOfMonth\": \link{31}, \"month\":\link[=\\"feb\\"]{\"feb\"}}``.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledJobPayload$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledJobPayload$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledJobPayload$fromJSON(ScheduledJobPayloadJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ScheduledJobPayloadJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledJobPayload$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
