% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduled_job_status.R
\docType{class}
\name{ScheduledJobStatus}
\alias{ScheduledJobStatus}
\title{ScheduledJobStatus}
\format{
An \code{R6Class} generator object
}
\description{
ScheduledJobStatus Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lastFailedRun}}{character \link{optional} Date and time of the last failed run.}

\item{\code{lastSuccessfulRun}}{character \link{optional} Date and time of the last successful run.}

\item{\code{nextRunTime}}{character \link{optional} Date and time of the next run.}

\item{\code{queuePosition}}{integer \link{optional} Position of the job in the queue Job. The value will show 0 if the job is about to run, otherwise, the number will be greater than 0 if currently queued, or None if the job is not currently running.}

\item{\code{running}}{character `true` or `false` depending on whether the job is currently running.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ScheduledJobStatus-new}{\code{ScheduledJobStatus$new()}}
\item \href{#method-ScheduledJobStatus-validate}{\code{ScheduledJobStatus$validate()}}
\item \href{#method-ScheduledJobStatus-toJSON}{\code{ScheduledJobStatus$toJSON()}}
\item \href{#method-ScheduledJobStatus-fromJSON}{\code{ScheduledJobStatus$fromJSON()}}
\item \href{#method-ScheduledJobStatus-clone}{\code{ScheduledJobStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScheduledJobStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-ScheduledJobStatus-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledJobStatus$new(
  running = NULL,
  lastFailedRun = NULL,
  lastSuccessfulRun = NULL,
  nextRunTime = NULL,
  queuePosition = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{running}}{`true` or `false` depending on whether the job is currently running.}

\item{\code{lastFailedRun}}{Date and time of the last failed run.}

\item{\code{lastSuccessfulRun}}{Date and time of the last successful run.}

\item{\code{nextRunTime}}{Date and time of the next run.}

\item{\code{queuePosition}}{Position of the job in the queue Job. The value will show 0 if the job is about to run, otherwise, the number will be greater than 0 if currently queued, or None if the job is not currently running.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScheduledJobStatus-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ScheduledJobStatus-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledJobStatus$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScheduledJobStatus-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ScheduledJobStatus-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledJobStatus$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScheduledJobStatus-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ScheduledJobStatus-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledJobStatus$fromJSON(ScheduledJobStatusJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ScheduledJobStatusJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScheduledJobStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ScheduledJobStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledJobStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
