% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_dataset.R
\docType{class}
\name{SecondaryDataset}
\alias{SecondaryDataset}
\title{SecondaryDataset}
\format{
An \code{R6Class} generator object
}
\description{
SecondaryDataset Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalogId}}{character Id of the catalog item version.}

\item{\code{catalogVersionId}}{character Id of the catalog item.}

\item{\code{identifier}}{character Short name of this table (used directly as part of generated feature names).}

\item{\code{snapshotPolicy}}{character \link{optional} Type of snapshot policy to use by the dataset.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SecondaryDataset-new}{\code{SecondaryDataset$new()}}
\item \href{#method-SecondaryDataset-validate}{\code{SecondaryDataset$validate()}}
\item \href{#method-SecondaryDataset-toJSON}{\code{SecondaryDataset$toJSON()}}
\item \href{#method-SecondaryDataset-fromJSON}{\code{SecondaryDataset$fromJSON()}}
\item \href{#method-SecondaryDataset-clone}{\code{SecondaryDataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondaryDataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-SecondaryDataset-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryDataset$new(
  catalogId = NULL,
  catalogVersionId = NULL,
  identifier = NULL,
  snapshotPolicy = "latest",
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalogId}}{Id of the catalog item version.}

\item{\code{catalogVersionId}}{Id of the catalog item.}

\item{\code{identifier}}{Short name of this table (used directly as part of generated feature names).}

\item{\code{snapshotPolicy}}{Type of snapshot policy to use by the dataset.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondaryDataset-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SecondaryDataset-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryDataset$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondaryDataset-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SecondaryDataset-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryDataset$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondaryDataset-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SecondaryDataset-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryDataset$fromJSON(SecondaryDatasetJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SecondaryDatasetJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondaryDataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SecondaryDataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
