% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_dataset_config_update.R
\docType{class}
\name{SecondaryDatasetConfigUpdate}
\alias{SecondaryDatasetConfigUpdate}
\title{SecondaryDatasetConfigUpdate}
\format{
An \code{R6Class} generator object
}
\description{
SecondaryDatasetConfigUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{credentialsIds}}{list( \link{DatasetsCredential} ) \link{optional} List of credentials used by the secondary datasets}

\item{\code{secondaryDatasetConfigId}}{character ID of the secondary datasets configuration to be used at the time of prediction}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SecondaryDatasetConfigUpdate-new}{\code{SecondaryDatasetConfigUpdate$new()}}
\item \href{#method-SecondaryDatasetConfigUpdate-validate}{\code{SecondaryDatasetConfigUpdate$validate()}}
\item \href{#method-SecondaryDatasetConfigUpdate-toJSON}{\code{SecondaryDatasetConfigUpdate$toJSON()}}
\item \href{#method-SecondaryDatasetConfigUpdate-fromJSON}{\code{SecondaryDatasetConfigUpdate$fromJSON()}}
\item \href{#method-SecondaryDatasetConfigUpdate-clone}{\code{SecondaryDatasetConfigUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondaryDatasetConfigUpdate-new"></a>}}
\if{latex}{\out{\hypertarget{method-SecondaryDatasetConfigUpdate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryDatasetConfigUpdate$new(
  secondaryDatasetConfigId = NULL,
  credentialsIds = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{secondaryDatasetConfigId}}{ID of the secondary datasets configuration to be used at the time of prediction}

\item{\code{credentialsIds}}{List of credentials used by the secondary datasets}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondaryDatasetConfigUpdate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SecondaryDatasetConfigUpdate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryDatasetConfigUpdate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondaryDatasetConfigUpdate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SecondaryDatasetConfigUpdate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryDatasetConfigUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondaryDatasetConfigUpdate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SecondaryDatasetConfigUpdate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryDatasetConfigUpdate$fromJSON(
  SecondaryDatasetConfigUpdateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SecondaryDatasetConfigUpdateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondaryDatasetConfigUpdate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SecondaryDatasetConfigUpdate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryDatasetConfigUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
