% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_custom.R
\docType{class}
\name{SectionCustom}
\alias{SectionCustom}
\title{SectionCustom}
\format{
An \code{R6Class} generator object
}
\description{
SectionCustom Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{description}}{character \link{optional} Section description}

\item{\code{highlightedText}}{character Highlighted text of the section, optionally separated by ``\\n`` to split paragraphs.}

\item{\code{instructions}}{\link{InstructionsField} \link{optional}}

\item{\code{locked}}{character \link{optional} Locked section flag}

\item{\code{regularText}}{character Regular text of the section, optionally separated by ``\\n`` to split paragraphs.}

\item{\code{sections}}{list( object ) \link{optional} List of section objects representing the structure of the document. Each section can have sub-sections that have the same schema as the parent section, e.g. the structure is recursive. The limit of nesting sections is 5. Total number of sections is limited to 500.}

\item{\code{title}}{character Section Title}

\item{\code{type}}{character Section with user-defined content. It can be a section title or summary.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SectionCustom-new}{\code{SectionCustom$new()}}
\item \href{#method-SectionCustom-validate}{\code{SectionCustom$validate()}}
\item \href{#method-SectionCustom-toJSON}{\code{SectionCustom$toJSON()}}
\item \href{#method-SectionCustom-fromJSON}{\code{SectionCustom$fromJSON()}}
\item \href{#method-SectionCustom-clone}{\code{SectionCustom$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SectionCustom-new"></a>}}
\if{latex}{\out{\hypertarget{method-SectionCustom-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SectionCustom$new(
  highlightedText = NULL,
  regularText = NULL,
  title = NULL,
  type = NULL,
  description = NULL,
  instructions = NULL,
  locked = NULL,
  sections = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{highlightedText}}{Highlighted text of the section, optionally separated by ``\\n`` to split paragraphs.}

\item{\code{regularText}}{Regular text of the section, optionally separated by ``\\n`` to split paragraphs.}

\item{\code{title}}{Section Title}

\item{\code{type}}{Section with user-defined content. It can be a section title or summary.}

\item{\code{description}}{Section description}

\item{\code{locked}}{Locked section flag}

\item{\code{sections}}{List of section objects representing the structure of the document. Each section can have sub-sections that have the same schema as the parent section, e.g. the structure is recursive. The limit of nesting sections is 5. Total number of sections is limited to 500.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SectionCustom-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SectionCustom-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SectionCustom$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SectionCustom-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SectionCustom-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SectionCustom$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SectionCustom-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SectionCustom-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SectionCustom$fromJSON(SectionCustomJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SectionCustomJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SectionCustom-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SectionCustom-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SectionCustom$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
