% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_data_robot.R
\docType{class}
\name{SectionDataRobot}
\alias{SectionDataRobot}
\title{SectionDataRobot}
\format{
An \code{R6Class} generator object
}
\description{
SectionDataRobot Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{contentId}}{character The identifier of the content in the section. This attribute identifies what is going to be rendered in the section.}

\item{\code{description}}{character \link{optional} Section description}

\item{\code{instructions}}{\link{InstructionsField} \link{optional}}

\item{\code{locked}}{character \link{optional} Locked section flag}

\item{\code{sections}}{list( object ) \link{optional} List of section objects representing the structure of the document. Each section can have sub-sections that have the same schema as the parent section, e.g. the structure is recursive. The limit of nesting sections is 5. Total number of sections is limited to 500.}

\item{\code{title}}{character Section Title}

\item{\code{type}}{character Section owned by DataRobot. The content of this section type is controlled by DataRobot (see ``contentId`` attribute). Users can add sub-sections to it.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SectionDataRobot$new()}}
\item \href{#method-validate}{\code{SectionDataRobot$validate()}}
\item \href{#method-toJSON}{\code{SectionDataRobot$toJSON()}}
\item \href{#method-fromJSON}{\code{SectionDataRobot$fromJSON()}}
\item \href{#method-clone}{\code{SectionDataRobot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SectionDataRobot$new(
  contentId = NULL,
  title = NULL,
  type = NULL,
  description = NULL,
  instructions = NULL,
  locked = NULL,
  sections = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{contentId}}{The identifier of the content in the section. This attribute identifies what is going to be rendered in the section.}

\item{\code{title}}{Section Title}

\item{\code{type}}{Section owned by DataRobot. The content of this section type is controlled by DataRobot (see ``contentId`` attribute). Users can add sub-sections to it.}

\item{\code{description}}{Section description}

\item{\code{locked}}{Locked section flag}

\item{\code{sections}}{List of section objects representing the structure of the document. Each section can have sub-sections that have the same schema as the parent section, e.g. the structure is recursive. The limit of nesting sections is 5. Total number of sections is limited to 500.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SectionDataRobot$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SectionDataRobot$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SectionDataRobot$fromJSON(SectionDataRobotJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SectionDataRobotJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SectionDataRobot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
