% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_analysis.R
\docType{class}
\name{SegmentAnalysis}
\alias{SegmentAnalysis}
\title{SegmentAnalysis}
\format{
An \code{R6Class} generator object
}
\description{
SegmentAnalysis Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{attributes}}{list( character ) \link{optional} The list of segment attributes selected for tracking.}

\item{\code{enabled}}{character Indicates whether service health, drift, and accuracy are tracked for segments of prediction data.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SegmentAnalysis-new}{\code{SegmentAnalysis$new()}}
\item \href{#method-SegmentAnalysis-validate}{\code{SegmentAnalysis$validate()}}
\item \href{#method-SegmentAnalysis-toJSON}{\code{SegmentAnalysis$toJSON()}}
\item \href{#method-SegmentAnalysis-fromJSON}{\code{SegmentAnalysis$fromJSON()}}
\item \href{#method-SegmentAnalysis-clone}{\code{SegmentAnalysis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentAnalysis-new"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentAnalysis-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentAnalysis$new(
  enabled = NULL,
  attributes = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enabled}}{Indicates whether service health, drift, and accuracy are tracked for segments of prediction data.}

\item{\code{attributes}}{The list of segment attributes selected for tracking.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentAnalysis-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentAnalysis-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentAnalysis$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentAnalysis-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentAnalysis-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentAnalysis$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentAnalysis-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentAnalysis-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentAnalysis$fromJSON(SegmentAnalysisJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SegmentAnalysisJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentAnalysis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentAnalysis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentAnalysis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
