% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_analysis_update.R
\docType{class}
\name{SegmentAnalysisUpdate}
\alias{SegmentAnalysisUpdate}
\title{SegmentAnalysisUpdate}
\format{
An \code{R6Class} generator object
}
\description{
SegmentAnalysisUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{attributes}}{list( character ) \link{optional} The segment attributes to be tracked. Note that only categorical columns can be specified as tracked segment attributes. The target column may not be specified}

\item{\code{enabled}}{character Indicates whether service health, drift, and accuracy are tracked for segments of prediction data.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SegmentAnalysisUpdate-new}{\code{SegmentAnalysisUpdate$new()}}
\item \href{#method-SegmentAnalysisUpdate-validate}{\code{SegmentAnalysisUpdate$validate()}}
\item \href{#method-SegmentAnalysisUpdate-toJSON}{\code{SegmentAnalysisUpdate$toJSON()}}
\item \href{#method-SegmentAnalysisUpdate-fromJSON}{\code{SegmentAnalysisUpdate$fromJSON()}}
\item \href{#method-SegmentAnalysisUpdate-clone}{\code{SegmentAnalysisUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentAnalysisUpdate-new"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentAnalysisUpdate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentAnalysisUpdate$new(
  enabled = NULL,
  attributes = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enabled}}{Indicates whether service health, drift, and accuracy are tracked for segments of prediction data.}

\item{\code{attributes}}{The segment attributes to be tracked. Note that only categorical columns can be specified as tracked segment attributes. The target column may not be specified}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentAnalysisUpdate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentAnalysisUpdate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentAnalysisUpdate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentAnalysisUpdate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentAnalysisUpdate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentAnalysisUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentAnalysisUpdate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentAnalysisUpdate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentAnalysisUpdate$fromJSON(
  SegmentAnalysisUpdateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SegmentAnalysisUpdateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentAnalysisUpdate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentAnalysisUpdate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentAnalysisUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
