% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_data_mapping_response.R
\docType{class}
\name{SegmentationDataMappingResponse}
\alias{SegmentationDataMappingResponse}
\title{SegmentationDataMappingResponse}
\format{
An \code{R6Class} generator object
}
\description{
SegmentationDataMappingResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{segmentId}}{character The segment name associated with the multiseries ID column by the segmentation task}

\item{\code{seriesId}}{character The multiseries ID column used to identify series for segmentation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SegmentationDataMappingResponse-new}{\code{SegmentationDataMappingResponse$new()}}
\item \href{#method-SegmentationDataMappingResponse-validate}{\code{SegmentationDataMappingResponse$validate()}}
\item \href{#method-SegmentationDataMappingResponse-toJSON}{\code{SegmentationDataMappingResponse$toJSON()}}
\item \href{#method-SegmentationDataMappingResponse-fromJSON}{\code{SegmentationDataMappingResponse$fromJSON()}}
\item \href{#method-SegmentationDataMappingResponse-clone}{\code{SegmentationDataMappingResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentationDataMappingResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentationDataMappingResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationDataMappingResponse$new(
  segmentId = NULL,
  seriesId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{segmentId}}{The segment name associated with the multiseries ID column by the segmentation task}

\item{\code{seriesId}}{The multiseries ID column used to identify series for segmentation}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentationDataMappingResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentationDataMappingResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationDataMappingResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentationDataMappingResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentationDataMappingResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationDataMappingResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentationDataMappingResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentationDataMappingResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationDataMappingResponse$fromJSON(
  SegmentationDataMappingResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SegmentationDataMappingResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentationDataMappingResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentationDataMappingResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationDataMappingResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
