% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_eda_completed_response.R
\docType{class}
\name{SegmentationEDACompletedResponse}
\alias{SegmentationEDACompletedResponse}
\title{SegmentationEDACompletedResponse}
\format{
An \code{R6Class} generator object
}
\description{
SegmentationEDACompletedResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{maxDate}}{character Latest date in segment.}

\item{\code{minDate}}{character Earliest date in segment.}

\item{\code{name}}{character Name of the segment.}

\item{\code{numberOfRows}}{integer Number of rows in segment.}

\item{\code{sizeInBytes}}{integer Size of segment in bytes.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SegmentationEDACompletedResponse$new()}}
\item \href{#method-validate}{\code{SegmentationEDACompletedResponse$validate()}}
\item \href{#method-toJSON}{\code{SegmentationEDACompletedResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{SegmentationEDACompletedResponse$fromJSON()}}
\item \href{#method-clone}{\code{SegmentationEDACompletedResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationEDACompletedResponse$new(
  maxDate = NULL,
  minDate = NULL,
  name = NULL,
  numberOfRows = NULL,
  sizeInBytes = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maxDate}}{Latest date in segment.}

\item{\code{minDate}}{Earliest date in segment.}

\item{\code{name}}{Name of the segment.}

\item{\code{numberOfRows}}{Number of rows in segment.}

\item{\code{sizeInBytes}}{Size of segment in bytes.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationEDACompletedResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationEDACompletedResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationEDACompletedResponse$fromJSON(
  SegmentationEDACompletedResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SegmentationEDACompletedResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationEDACompletedResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
