% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_results_completed_response.R
\docType{class}
\name{SegmentationResultsCompletedResponse}
\alias{SegmentationResultsCompletedResponse}
\title{SegmentationResultsCompletedResponse}
\format{
An \code{R6Class} generator object
}
\description{
SegmentationResultsCompletedResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character Name of the segmentation task job}

\item{\code{segmentationTaskId}}{character Id of the completed segmentation task}

\item{\code{segmentsCount}}{integer Number of segments produced by the task}

\item{\code{segmentsEda}}{list( \link{SegmentationEDACompletedResponse} ) Array of segments EDA information.}

\item{\code{url}}{character Url to retrieve detailed information about the segmentation task}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SegmentationResultsCompletedResponse-new}{\code{SegmentationResultsCompletedResponse$new()}}
\item \href{#method-SegmentationResultsCompletedResponse-validate}{\code{SegmentationResultsCompletedResponse$validate()}}
\item \href{#method-SegmentationResultsCompletedResponse-toJSON}{\code{SegmentationResultsCompletedResponse$toJSON()}}
\item \href{#method-SegmentationResultsCompletedResponse-fromJSON}{\code{SegmentationResultsCompletedResponse$fromJSON()}}
\item \href{#method-SegmentationResultsCompletedResponse-clone}{\code{SegmentationResultsCompletedResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentationResultsCompletedResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentationResultsCompletedResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationResultsCompletedResponse$new(
  name = NULL,
  segmentationTaskId = NULL,
  segmentsCount = NULL,
  segmentsEda = NULL,
  url = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the segmentation task job}

\item{\code{segmentationTaskId}}{Id of the completed segmentation task}

\item{\code{segmentsCount}}{Number of segments produced by the task}

\item{\code{segmentsEda}}{Array of segments EDA information.}

\item{\code{url}}{Url to retrieve detailed information about the segmentation task}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentationResultsCompletedResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentationResultsCompletedResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationResultsCompletedResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentationResultsCompletedResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentationResultsCompletedResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationResultsCompletedResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentationResultsCompletedResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentationResultsCompletedResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationResultsCompletedResponse$fromJSON(
  SegmentationResultsCompletedResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SegmentationResultsCompletedResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentationResultsCompletedResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentationResultsCompletedResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationResultsCompletedResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
