% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_task_create.R
\docType{class}
\name{SegmentationTaskCreate}
\alias{SegmentationTaskCreate}
\title{SegmentationTaskCreate}
\format{
An \code{R6Class} generator object
}
\description{
SegmentationTaskCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datetimePartitionColumn}}{character \link{optional} The date column that will be used to identify the date in time series segmentation.}

\item{\code{modelPackageId}}{character \link{optional} Model package id for using an external model registry package.}

\item{\code{multiseriesIdColumns}}{list( character ) \link{optional} List of one or more names of multiseries id columns.}

\item{\code{target}}{character The target for the dataset}

\item{\code{useAutomatedSegmentation}}{character \link{optional} Enable the use of automated segmentation tasks}

\item{\code{useTimeSeries}}{character \link{optional} Enable time series based segmentation tasks}

\item{\code{userDefinedSegmentIdColumns}}{list( character ) \link{optional} List of one or more names of columns to be used for user-defined business rule segmentations.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SegmentationTaskCreate$new()}}
\item \href{#method-validate}{\code{SegmentationTaskCreate$validate()}}
\item \href{#method-toJSON}{\code{SegmentationTaskCreate$toJSON()}}
\item \href{#method-fromJSON}{\code{SegmentationTaskCreate$fromJSON()}}
\item \href{#method-clone}{\code{SegmentationTaskCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationTaskCreate$new(
  target = NULL,
  datetimePartitionColumn = NULL,
  modelPackageId = NULL,
  multiseriesIdColumns = NULL,
  useAutomatedSegmentation = FALSE,
  useTimeSeries = FALSE,
  userDefinedSegmentIdColumns = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target}}{The target for the dataset}

\item{\code{datetimePartitionColumn}}{The date column that will be used to identify the date in time series segmentation.}

\item{\code{modelPackageId}}{Model package id for using an external model registry package.}

\item{\code{multiseriesIdColumns}}{List of one or more names of multiseries id columns.}

\item{\code{useAutomatedSegmentation}}{Enable the use of automated segmentation tasks}

\item{\code{useTimeSeries}}{Enable time series based segmentation tasks}

\item{\code{userDefinedSegmentIdColumns}}{List of one or more names of columns to be used for user-defined business rule segmentations.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationTaskCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationTaskCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationTaskCreate$fromJSON(
  SegmentationTaskCreateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SegmentationTaskCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationTaskCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
