% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_task_data_response.R
\docType{class}
\name{SegmentationTaskDataResponse}
\alias{SegmentationTaskDataResponse}
\title{SegmentationTaskDataResponse}
\format{
An \code{R6Class} generator object
}
\description{
SegmentationTaskDataResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datetimePartitionColumn}}{character \link{optional} Name of the datetime partitioning column used by the segmentation task}

\item{\code{modelPackageId}}{character \link{optional} External model package id used by the segmentation task for automated segmentation}

\item{\code{multiseriesIdColumns}}{list( character ) \link{optional} The multiseries id columns used by the segmentation task}

\item{\code{userDefinedSegmentIdColumns}}{list( character ) \link{optional} The user defined segmentation columns used by the segmentation task}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SegmentationTaskDataResponse-new}{\code{SegmentationTaskDataResponse$new()}}
\item \href{#method-SegmentationTaskDataResponse-validate}{\code{SegmentationTaskDataResponse$validate()}}
\item \href{#method-SegmentationTaskDataResponse-toJSON}{\code{SegmentationTaskDataResponse$toJSON()}}
\item \href{#method-SegmentationTaskDataResponse-fromJSON}{\code{SegmentationTaskDataResponse$fromJSON()}}
\item \href{#method-SegmentationTaskDataResponse-clone}{\code{SegmentationTaskDataResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentationTaskDataResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentationTaskDataResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationTaskDataResponse$new(
  datetimePartitionColumn = NULL,
  modelPackageId = NULL,
  multiseriesIdColumns = NULL,
  userDefinedSegmentIdColumns = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datetimePartitionColumn}}{Name of the datetime partitioning column used by the segmentation task}

\item{\code{modelPackageId}}{External model package id used by the segmentation task for automated segmentation}

\item{\code{multiseriesIdColumns}}{The multiseries id columns used by the segmentation task}

\item{\code{userDefinedSegmentIdColumns}}{The user defined segmentation columns used by the segmentation task}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentationTaskDataResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentationTaskDataResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationTaskDataResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentationTaskDataResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentationTaskDataResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationTaskDataResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentationTaskDataResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentationTaskDataResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationTaskDataResponse$fromJSON(
  SegmentationTaskDataResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SegmentationTaskDataResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentationTaskDataResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentationTaskDataResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationTaskDataResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
