% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_task_metadata_response.R
\docType{class}
\name{SegmentationTaskMetadataResponse}
\alias{SegmentationTaskMetadataResponse}
\title{SegmentationTaskMetadataResponse}
\format{
An \code{R6Class} generator object
}
\description{
SegmentationTaskMetadataResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{useAutomatedSegmentation}}{character Segmentation task uses automated segmentation}

\item{\code{useMultiseriesIdColumns}}{character Segmentation task uses a multiseries column}

\item{\code{useTimeSeries}}{character Segmentation task is a time series task}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SegmentationTaskMetadataResponse-new}{\code{SegmentationTaskMetadataResponse$new()}}
\item \href{#method-SegmentationTaskMetadataResponse-validate}{\code{SegmentationTaskMetadataResponse$validate()}}
\item \href{#method-SegmentationTaskMetadataResponse-toJSON}{\code{SegmentationTaskMetadataResponse$toJSON()}}
\item \href{#method-SegmentationTaskMetadataResponse-fromJSON}{\code{SegmentationTaskMetadataResponse$fromJSON()}}
\item \href{#method-SegmentationTaskMetadataResponse-clone}{\code{SegmentationTaskMetadataResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentationTaskMetadataResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentationTaskMetadataResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationTaskMetadataResponse$new(
  useAutomatedSegmentation = NULL,
  useMultiseriesIdColumns = NULL,
  useTimeSeries = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{useAutomatedSegmentation}}{Segmentation task uses automated segmentation}

\item{\code{useMultiseriesIdColumns}}{Segmentation task uses a multiseries column}

\item{\code{useTimeSeries}}{Segmentation task is a time series task}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentationTaskMetadataResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentationTaskMetadataResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationTaskMetadataResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentationTaskMetadataResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentationTaskMetadataResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationTaskMetadataResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentationTaskMetadataResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentationTaskMetadataResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationTaskMetadataResponse$fromJSON(
  SegmentationTaskMetadataResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SegmentationTaskMetadataResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentationTaskMetadataResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentationTaskMetadataResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationTaskMetadataResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
