% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_task_response.R
\docType{class}
\name{SegmentationTaskResponse}
\alias{SegmentationTaskResponse}
\title{SegmentationTaskResponse}
\format{
An \code{R6Class} generator object
}
\description{
SegmentationTaskResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{created}}{character Date and time when the segmentation task was originally created}

\item{\code{data}}{\link{SegmentationTaskDataResponse}}

\item{\code{metadata}}{\link{SegmentationTaskMetadataResponse}}

\item{\code{name}}{character Name of the segmentation task}

\item{\code{projectId}}{character Id of the parent project associated with the segmentation task}

\item{\code{segmentationTaskId}}{character Id of the segmentation task}

\item{\code{segments}}{list( character ) Names of the unique segments generated by the segmentation task}

\item{\code{segmentsCount}}{integer Number of segments generated by the segmentation task}

\item{\code{segmentsEda}}{list( \link{SegmentationEDACompletedResponse} ) Array of segments EDA information.}

\item{\code{type}}{character Type of segmentation task (e.g. AutoML, AutoTS)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SegmentationTaskResponse$new()}}
\item \href{#method-validate}{\code{SegmentationTaskResponse$validate()}}
\item \href{#method-toJSON}{\code{SegmentationTaskResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{SegmentationTaskResponse$fromJSON()}}
\item \href{#method-clone}{\code{SegmentationTaskResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationTaskResponse$new(
  created = NULL,
  data = NULL,
  metadata = NULL,
  name = NULL,
  projectId = NULL,
  segmentationTaskId = NULL,
  segments = NULL,
  segmentsCount = NULL,
  segmentsEda = NULL,
  type = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{created}}{Date and time when the segmentation task was originally created}

\item{\code{name}}{Name of the segmentation task}

\item{\code{projectId}}{Id of the parent project associated with the segmentation task}

\item{\code{segmentationTaskId}}{Id of the segmentation task}

\item{\code{segments}}{Names of the unique segments generated by the segmentation task}

\item{\code{segmentsCount}}{Number of segments generated by the segmentation task}

\item{\code{segmentsEda}}{Array of segments EDA information.}

\item{\code{type}}{Type of segmentation task (e.g. AutoML, AutoTS)}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationTaskResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationTaskResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationTaskResponse$fromJSON(
  SegmentationTaskResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SegmentationTaskResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentationTaskResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
