% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_project_model_response.R
\docType{class}
\name{SegmentProjectModelResponse}
\alias{SegmentProjectModelResponse}
\title{SegmentProjectModelResponse}
\format{
An \code{R6Class} generator object
}
\description{
SegmentProjectModelResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{modelId}}{character ID of segment champion model.}

\item{\code{projectId}}{character ID of the project used for this segment.}

\item{\code{segment}}{character Segment name.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SegmentProjectModelResponse-new}{\code{SegmentProjectModelResponse$new()}}
\item \href{#method-SegmentProjectModelResponse-validate}{\code{SegmentProjectModelResponse$validate()}}
\item \href{#method-SegmentProjectModelResponse-toJSON}{\code{SegmentProjectModelResponse$toJSON()}}
\item \href{#method-SegmentProjectModelResponse-fromJSON}{\code{SegmentProjectModelResponse$fromJSON()}}
\item \href{#method-SegmentProjectModelResponse-clone}{\code{SegmentProjectModelResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentProjectModelResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentProjectModelResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentProjectModelResponse$new(
  modelId = NULL,
  projectId = NULL,
  segment = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{modelId}}{ID of segment champion model.}

\item{\code{projectId}}{ID of the project used for this segment.}

\item{\code{segment}}{Segment name.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentProjectModelResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentProjectModelResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentProjectModelResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentProjectModelResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentProjectModelResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentProjectModelResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentProjectModelResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentProjectModelResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentProjectModelResponse$fromJSON(
  SegmentProjectModelResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SegmentProjectModelResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SegmentProjectModelResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SegmentProjectModelResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentProjectModelResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
