% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_accuracy_retrieve_data_response.R
\docType{class}
\name{SeriesAccuracyRetrieveDataResponse}
\alias{SeriesAccuracyRetrieveDataResponse}
\title{SeriesAccuracyRetrieveDataResponse}
\format{
An \code{R6Class} generator object
}
\description{
SeriesAccuracyRetrieveDataResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{backtestingScore}}{numeric The backtesting score for this series. If backtesting has not been run for this model, this score will be null.}

\item{\code{cluster}}{character \link{optional} The cluster associated with this series.}

\item{\code{duration}}{character The duration of this series formatted as an ISO 8601 duration string.}

\item{\code{endDate}}{character \link{optional} The ISO-formatted end date of this series.}

\item{\code{holdoutScore}}{numeric \link{optional} The holdout set score for this series. If holdout is locked for the project, this score will be null.}

\item{\code{multiseriesId}}{character A DataRobot-generated ID corresponding to a single series in a multiseries dataset.}

\item{\code{multiseriesValues}}{list( character ) The actual values of series ID columns from the dataset.}

\item{\code{rowCount}}{integer The number of rows available for this series in the input dataset.}

\item{\code{startDate}}{character \link{optional} The ISO-formatted start date of this series.}

\item{\code{targetAverage}}{\link{OneOfstringnumber} \link{optional} For regression projects, this is the average (mean) value of target values for this series.For classification projects, this is the ratio of the positive class in the target for this series.}

\item{\code{validationScore}}{numeric The validation set score for this series}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeriesAccuracyRetrieveDataResponse-new}{\code{SeriesAccuracyRetrieveDataResponse$new()}}
\item \href{#method-SeriesAccuracyRetrieveDataResponse-validate}{\code{SeriesAccuracyRetrieveDataResponse$validate()}}
\item \href{#method-SeriesAccuracyRetrieveDataResponse-toJSON}{\code{SeriesAccuracyRetrieveDataResponse$toJSON()}}
\item \href{#method-SeriesAccuracyRetrieveDataResponse-fromJSON}{\code{SeriesAccuracyRetrieveDataResponse$fromJSON()}}
\item \href{#method-SeriesAccuracyRetrieveDataResponse-clone}{\code{SeriesAccuracyRetrieveDataResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeriesAccuracyRetrieveDataResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeriesAccuracyRetrieveDataResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeriesAccuracyRetrieveDataResponse$new(
  backtestingScore = NULL,
  duration = NULL,
  multiseriesId = NULL,
  multiseriesValues = NULL,
  rowCount = NULL,
  validationScore = NULL,
  cluster = NULL,
  endDate = NULL,
  holdoutScore = NULL,
  startDate = NULL,
  targetAverage = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{backtestingScore}}{The backtesting score for this series. If backtesting has not been run for this model, this score will be null.}

\item{\code{duration}}{The duration of this series formatted as an ISO 8601 duration string.}

\item{\code{multiseriesId}}{A DataRobot-generated ID corresponding to a single series in a multiseries dataset.}

\item{\code{multiseriesValues}}{The actual values of series ID columns from the dataset.}

\item{\code{rowCount}}{The number of rows available for this series in the input dataset.}

\item{\code{validationScore}}{The validation set score for this series}

\item{\code{cluster}}{The cluster associated with this series.}

\item{\code{endDate}}{The ISO-formatted end date of this series.}

\item{\code{holdoutScore}}{The holdout set score for this series. If holdout is locked for the project, this score will be null.}

\item{\code{startDate}}{The ISO-formatted start date of this series.}

\item{\code{targetAverage}}{For regression projects, this is the average (mean) value of target values for this series.For classification projects, this is the ratio of the positive class in the target for this series.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeriesAccuracyRetrieveDataResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SeriesAccuracyRetrieveDataResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeriesAccuracyRetrieveDataResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeriesAccuracyRetrieveDataResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SeriesAccuracyRetrieveDataResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeriesAccuracyRetrieveDataResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeriesAccuracyRetrieveDataResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SeriesAccuracyRetrieveDataResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeriesAccuracyRetrieveDataResponse$fromJSON(
  SeriesAccuracyRetrieveDataResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SeriesAccuracyRetrieveDataResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeriesAccuracyRetrieveDataResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeriesAccuracyRetrieveDataResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeriesAccuracyRetrieveDataResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
