% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_stats_metrics.R
\docType{class}
\name{ServiceStatsMetrics}
\alias{ServiceStatsMetrics}
\title{ServiceStatsMetrics}
\format{
An \code{R6Class} generator object
}
\description{
ServiceStatsMetrics Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cacheHitRatio}}{numeric Number of cache hits.}

\item{\code{executionTime}}{numeric Request execution time at `executionTimeQuantile` (in milliseconds).}

\item{\code{medianLoad}}{numeric Median of the request rate (in requests per minute).}

\item{\code{numConsumers}}{integer Number of unique users performing requests.}

\item{\code{peakLoad}}{numeric Maximum of the request rate (in requests per minute).}

\item{\code{responseTime}}{numeric Request response time at `responseTimeQuantile` (in milliseconds).}

\item{\code{serverErrorRate}}{numeric Ratio of server errors to the total number of requests.}

\item{\code{slowRequests}}{integer Number of requests with response time greater than `slowRequestsThreshold`}

\item{\code{totalPredictions}}{integer Total number of prediction rows.}

\item{\code{totalRequests}}{integer Total number of prediction requests performed.}

\item{\code{userErrorRate}}{numeric Ratio of user errors to the total number of requests.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ServiceStatsMetrics$new()}}
\item \href{#method-validate}{\code{ServiceStatsMetrics$validate()}}
\item \href{#method-toJSON}{\code{ServiceStatsMetrics$toJSON()}}
\item \href{#method-fromJSON}{\code{ServiceStatsMetrics$fromJSON()}}
\item \href{#method-clone}{\code{ServiceStatsMetrics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceStatsMetrics$new(
  cacheHitRatio = NULL,
  executionTime = NULL,
  medianLoad = NULL,
  numConsumers = NULL,
  peakLoad = NULL,
  responseTime = NULL,
  serverErrorRate = NULL,
  slowRequests = NULL,
  totalPredictions = NULL,
  totalRequests = NULL,
  userErrorRate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cacheHitRatio}}{Number of cache hits.}

\item{\code{executionTime}}{Request execution time at `executionTimeQuantile` (in milliseconds).}

\item{\code{medianLoad}}{Median of the request rate (in requests per minute).}

\item{\code{numConsumers}}{Number of unique users performing requests.}

\item{\code{peakLoad}}{Maximum of the request rate (in requests per minute).}

\item{\code{responseTime}}{Request response time at `responseTimeQuantile` (in milliseconds).}

\item{\code{serverErrorRate}}{Ratio of server errors to the total number of requests.}

\item{\code{slowRequests}}{Number of requests with response time greater than `slowRequestsThreshold`}

\item{\code{totalPredictions}}{Total number of prediction rows.}

\item{\code{totalRequests}}{Total number of prediction requests performed.}

\item{\code{userErrorRate}}{Ratio of user errors to the total number of requests.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceStatsMetrics$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceStatsMetrics$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceStatsMetrics$fromJSON(ServiceStatsMetricsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ServiceStatsMetricsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceStatsMetrics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
