% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_stats_retrieve_response.R
\docType{class}
\name{ServiceStatsRetrieveResponse}
\alias{ServiceStatsRetrieveResponse}
\title{ServiceStatsRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
ServiceStatsRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metrics}}{\link{ServiceStatsMetrics}}

\item{\code{modelId}}{character \link{optional} The id of the model for which metrics are being retrieved.}

\item{\code{period}}{\link{TimeRange}}

\item{\code{segmentAttribute}}{character \link{optional} The name of the segment on which segment analysis is being performed.}

\item{\code{segmentValue}}{character \link{optional} The value of the `segmentAttribute` to segment on.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ServiceStatsRetrieveResponse-new}{\code{ServiceStatsRetrieveResponse$new()}}
\item \href{#method-ServiceStatsRetrieveResponse-validate}{\code{ServiceStatsRetrieveResponse$validate()}}
\item \href{#method-ServiceStatsRetrieveResponse-toJSON}{\code{ServiceStatsRetrieveResponse$toJSON()}}
\item \href{#method-ServiceStatsRetrieveResponse-fromJSON}{\code{ServiceStatsRetrieveResponse$fromJSON()}}
\item \href{#method-ServiceStatsRetrieveResponse-clone}{\code{ServiceStatsRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceStatsRetrieveResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceStatsRetrieveResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceStatsRetrieveResponse$new(
  metrics = NULL,
  period = NULL,
  modelId = NULL,
  segmentAttribute = NULL,
  segmentValue = "",
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{modelId}}{The id of the model for which metrics are being retrieved.}

\item{\code{segmentAttribute}}{The name of the segment on which segment analysis is being performed.}

\item{\code{segmentValue}}{The value of the `segmentAttribute` to segment on.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceStatsRetrieveResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceStatsRetrieveResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceStatsRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceStatsRetrieveResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceStatsRetrieveResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceStatsRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceStatsRetrieveResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceStatsRetrieveResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceStatsRetrieveResponse$fromJSON(
  ServiceStatsRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ServiceStatsRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceStatsRetrieveResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceStatsRetrieveResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceStatsRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
