% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{SetupProject}
\alias{SetupProject}
\title{SetupProject}
\usage{
SetupProject(dataSource, projectName = NULL, maxWait = 60 * 60, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create project.
}
\details{
Create a new project.

This method invokes \verb{POST /projects/} in the DataRobot Public API.
}
\seealso{
The method Create in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
