% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shap_explanation_response.R
\docType{class}
\name{ShapExplanationResponse}
\alias{ShapExplanationResponse}
\title{ShapExplanationResponse}
\format{
An \code{R6Class} generator object
}
\description{
ShapExplanationResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{feature}}{character Feature name}

\item{\code{featureValue}}{character Feature value for this row. First 50 characters are returned.}

\item{\code{strength}}{numeric Shapley value for this feature and row.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ShapExplanationResponse-new}{\code{ShapExplanationResponse$new()}}
\item \href{#method-ShapExplanationResponse-validate}{\code{ShapExplanationResponse$validate()}}
\item \href{#method-ShapExplanationResponse-toJSON}{\code{ShapExplanationResponse$toJSON()}}
\item \href{#method-ShapExplanationResponse-fromJSON}{\code{ShapExplanationResponse$fromJSON()}}
\item \href{#method-ShapExplanationResponse-clone}{\code{ShapExplanationResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShapExplanationResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ShapExplanationResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapExplanationResponse$new(
  feature = NULL,
  featureValue = NULL,
  strength = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{Feature name}

\item{\code{featureValue}}{Feature value for this row. First 50 characters are returned.}

\item{\code{strength}}{Shapley value for this feature and row.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShapExplanationResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ShapExplanationResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapExplanationResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShapExplanationResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ShapExplanationResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapExplanationResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShapExplanationResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ShapExplanationResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapExplanationResponse$fromJSON(
  ShapExplanationResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ShapExplanationResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShapExplanationResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ShapExplanationResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapExplanationResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
