% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shap_explanations_metadata_values.R
\docType{class}
\name{ShapExplanationsMetadataValues}
\alias{ShapExplanationsMetadataValues}
\title{ShapExplanationsMetadataValues}
\format{
An \code{R6Class} generator object
}
\description{
ShapExplanationsMetadataValues Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{baseValue}}{numeric Will be present only if`explanationAlgorithm` = \"shap\". The model's average prediction over the training data. SHAP values are deviations from the base value.}

\item{\code{remainingTotal}}{numeric Will be present only if `explanationAlgorithm` = \"shap\" and `maxExplanations` is nonzero. The total of SHAP values for features beyond the `maxExplanations`. This can be identically 0 in all rows, if `maxExplanations` is greater than the number of features and thus all features are returned.}

\item{\code{warnings}}{list( \link{ShapExplanationsWarningsValues} ) Will be present if `explanationAlgorithm` was set to \"shap\" and there were additivity failures during SHAP values calculation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ShapExplanationsMetadataValues$new()}}
\item \href{#method-validate}{\code{ShapExplanationsMetadataValues$validate()}}
\item \href{#method-toJSON}{\code{ShapExplanationsMetadataValues$toJSON()}}
\item \href{#method-fromJSON}{\code{ShapExplanationsMetadataValues$fromJSON()}}
\item \href{#method-clone}{\code{ShapExplanationsMetadataValues$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapExplanationsMetadataValues$new(
  baseValue = NULL,
  remainingTotal = NULL,
  warnings = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{baseValue}}{Will be present only if`explanationAlgorithm` = \"shap\". The model's average prediction over the training data. SHAP values are deviations from the base value.}

\item{\code{remainingTotal}}{Will be present only if `explanationAlgorithm` = \"shap\" and `maxExplanations` is nonzero. The total of SHAP values for features beyond the `maxExplanations`. This can be identically 0 in all rows, if `maxExplanations` is greater than the number of features and thus all features are returned.}

\item{\code{warnings}}{Will be present if `explanationAlgorithm` was set to \"shap\" and there were additivity failures during SHAP values calculation.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapExplanationsMetadataValues$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapExplanationsMetadataValues$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapExplanationsMetadataValues$fromJSON(
  ShapExplanationsMetadataValuesJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ShapExplanationsMetadataValuesJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapExplanationsMetadataValues$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
