% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_roles_with_grant_list_response.R
\docType{class}
\name{SharedRolesWithGrantListResponse}
\alias{SharedRolesWithGrantListResponse}
\title{SharedRolesWithGrantListResponse}
\format{
An \code{R6Class} generator object
}
\description{
SharedRolesWithGrantListResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer The number of items returned.}

\item{\code{data}}{list( \link{AccessControlWithGrant} ) The access control list.}

\item{\code{next_}}{character URL pointing to the next page.}

\item{\code{previous}}{character URL pointing to the previous page.}

\item{\code{totalCount}}{integer Total number of items matching the condition.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SharedRolesWithGrantListResponse-new}{\code{SharedRolesWithGrantListResponse$new()}}
\item \href{#method-SharedRolesWithGrantListResponse-validate}{\code{SharedRolesWithGrantListResponse$validate()}}
\item \href{#method-SharedRolesWithGrantListResponse-toJSON}{\code{SharedRolesWithGrantListResponse$toJSON()}}
\item \href{#method-SharedRolesWithGrantListResponse-fromJSON}{\code{SharedRolesWithGrantListResponse$fromJSON()}}
\item \href{#method-SharedRolesWithGrantListResponse-clone}{\code{SharedRolesWithGrantListResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharedRolesWithGrantListResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-SharedRolesWithGrantListResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SharedRolesWithGrantListResponse$new(
  count = NULL,
  data = NULL,
  next_ = NULL,
  previous = NULL,
  totalCount = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{The number of items returned.}

\item{\code{data}}{The access control list.}

\item{\code{next_}}{URL pointing to the next page.}

\item{\code{previous}}{URL pointing to the previous page.}

\item{\code{totalCount}}{Total number of items matching the condition.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharedRolesWithGrantListResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SharedRolesWithGrantListResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SharedRolesWithGrantListResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharedRolesWithGrantListResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SharedRolesWithGrantListResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SharedRolesWithGrantListResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharedRolesWithGrantListResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SharedRolesWithGrantListResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SharedRolesWithGrantListResponse$fromJSON(
  SharedRolesWithGrantListResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SharedRolesWithGrantListResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharedRolesWithGrantListResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SharedRolesWithGrantListResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SharedRolesWithGrantListResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
