% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharing_role_update_data.R
\docType{class}
\name{SharingRoleUpdateData}
\alias{SharingRoleUpdateData}
\title{SharingRoleUpdateData}
\format{
An \code{R6Class} generator object
}
\description{
SharingRoleUpdateData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character \link{optional} The ID of the recipient}

\item{\code{role}}{character The assigned role}

\item{\code{shareRecipientType}}{character The recipient type}

\item{\code{username}}{character \link{optional} The name of the user, group, or organization}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SharingRoleUpdateData-new}{\code{SharingRoleUpdateData$new()}}
\item \href{#method-SharingRoleUpdateData-validate}{\code{SharingRoleUpdateData$validate()}}
\item \href{#method-SharingRoleUpdateData-toJSON}{\code{SharingRoleUpdateData$toJSON()}}
\item \href{#method-SharingRoleUpdateData-fromJSON}{\code{SharingRoleUpdateData$fromJSON()}}
\item \href{#method-SharingRoleUpdateData-clone}{\code{SharingRoleUpdateData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharingRoleUpdateData-new"></a>}}
\if{latex}{\out{\hypertarget{method-SharingRoleUpdateData-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SharingRoleUpdateData$new(
  role = NULL,
  shareRecipientType = NULL,
  id = NULL,
  username = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{The assigned role}

\item{\code{shareRecipientType}}{The recipient type}

\item{\code{id}}{The ID of the recipient}

\item{\code{username}}{The name of the user, group, or organization}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharingRoleUpdateData-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SharingRoleUpdateData-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SharingRoleUpdateData$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharingRoleUpdateData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SharingRoleUpdateData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SharingRoleUpdateData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharingRoleUpdateData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SharingRoleUpdateData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SharingRoleUpdateData$fromJSON(
  SharingRoleUpdateDataJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SharingRoleUpdateDataJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharingRoleUpdateData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SharingRoleUpdateData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SharingRoleUpdateData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
