% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharing_update_or_remove.R
\docType{class}
\name{SharingUpdateOrRemove}
\alias{SharingUpdateOrRemove}
\title{SharingUpdateOrRemove}
\format{
An \code{R6Class} generator object
}
\description{
SharingUpdateOrRemove Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{UpdateAccessControl} ) The role to set for the user.}

\item{\code{includeFeatureDiscoveryEntities}}{character \link{optional} Whether to share all the related entities.}

\item{\code{sendNotification}}{character \link{optional} Send an email notification.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SharingUpdateOrRemove-new}{\code{SharingUpdateOrRemove$new()}}
\item \href{#method-SharingUpdateOrRemove-validate}{\code{SharingUpdateOrRemove$validate()}}
\item \href{#method-SharingUpdateOrRemove-toJSON}{\code{SharingUpdateOrRemove$toJSON()}}
\item \href{#method-SharingUpdateOrRemove-fromJSON}{\code{SharingUpdateOrRemove$fromJSON()}}
\item \href{#method-SharingUpdateOrRemove-clone}{\code{SharingUpdateOrRemove$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharingUpdateOrRemove-new"></a>}}
\if{latex}{\out{\hypertarget{method-SharingUpdateOrRemove-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SharingUpdateOrRemove$new(
  data = NULL,
  includeFeatureDiscoveryEntities = FALSE,
  sendNotification = TRUE,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The role to set for the user.}

\item{\code{includeFeatureDiscoveryEntities}}{Whether to share all the related entities.}

\item{\code{sendNotification}}{Send an email notification.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharingUpdateOrRemove-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SharingUpdateOrRemove-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SharingUpdateOrRemove$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharingUpdateOrRemove-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SharingUpdateOrRemove-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SharingUpdateOrRemove$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharingUpdateOrRemove-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SharingUpdateOrRemove-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SharingUpdateOrRemove$fromJSON(
  SharingUpdateOrRemoveJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SharingUpdateOrRemoveJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SharingUpdateOrRemove-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SharingUpdateOrRemove-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SharingUpdateOrRemove$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
