% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowflake_data_streamer.R
\docType{class}
\name{SnowflakeDataStreamer}
\alias{SnowflakeDataStreamer}
\title{SnowflakeDataStreamer}
\format{
An \code{R6Class} generator object
}
\description{
SnowflakeDataStreamer Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalog}}{character \link{optional} The name of the specified database catalog to read input data from.}

\item{\code{cloudStorageCredentialId}}{\link{OneOfstringstring} \link{optional} Either the populated value of the field or \link{redacted} due to permission settings}

\item{\code{cloudStorageType}}{character \link{optional} Type name for cloud storage}

\item{\code{credentialId}}{\link{OneOfstringstring} \link{optional} Either the populated value of the field or \link{redacted} due to permission settings}

\item{\code{dataStoreId}}{\link{OneOfstringstring} Either the populated value of the field or \link{redacted} due to permission settings}

\item{\code{externalStage}}{character External storage}

\item{\code{query}}{character \link{optional} A self-supplied SELECT statement of the dataset you wish to score. Helpful for supplying a more fine-grained selection of data not achievable through specification of \"table\" and/or \"schema\" parameters exclusively.If this job is executed with a job definition, then template variables are available which will be substituted for timestamps: {{ current_run_timestamp }}, {{ last_completed_run_time }}, {{ last_scheduled_run_time }}, {{ next_scheduled_run_time }}, {{ current_run_time }}}

\item{\code{schema}}{character \link{optional} The name of the specified database schema to read input data from.}

\item{\code{table}}{character \link{optional} The name of the specified database table to read input data from.}

\item{\code{type}}{character Type name for this intake type}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SnowflakeDataStreamer$new()}}
\item \href{#method-validate}{\code{SnowflakeDataStreamer$validate()}}
\item \href{#method-toJSON}{\code{SnowflakeDataStreamer$toJSON()}}
\item \href{#method-fromJSON}{\code{SnowflakeDataStreamer$fromJSON()}}
\item \href{#method-clone}{\code{SnowflakeDataStreamer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnowflakeDataStreamer$new(
  dataStoreId = NULL,
  externalStage = NULL,
  type = NULL,
  catalog = NULL,
  cloudStorageCredentialId = NULL,
  cloudStorageType = "s3",
  credentialId = NULL,
  query = NULL,
  schema = NULL,
  table = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataStoreId}}{Either the populated value of the field or \link{redacted} due to permission settings}

\item{\code{externalStage}}{External storage}

\item{\code{type}}{Type name for this intake type}

\item{\code{catalog}}{The name of the specified database catalog to read input data from.}

\item{\code{cloudStorageCredentialId}}{Either the populated value of the field or \link{redacted} due to permission settings}

\item{\code{cloudStorageType}}{Type name for cloud storage}

\item{\code{credentialId}}{Either the populated value of the field or \link{redacted} due to permission settings}

\item{\code{query}}{A self-supplied SELECT statement of the dataset you wish to score. Helpful for supplying a more fine-grained selection of data not achievable through specification of \"table\" and/or \"schema\" parameters exclusively.If this job is executed with a job definition, then template variables are available which will be substituted for timestamps: {{ current_run_timestamp }}, {{ last_completed_run_time }}, {{ last_scheduled_run_time }}, {{ next_scheduled_run_time }}, {{ current_run_time }}}

\item{\code{schema}}{The name of the specified database schema to read input data from.}

\item{\code{table}}{The name of the specified database table to read input data from.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnowflakeDataStreamer$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnowflakeDataStreamer$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnowflakeDataStreamer$fromJSON(
  SnowflakeDataStreamerJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SnowflakeDataStreamerJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnowflakeDataStreamer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
