% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spooler_channel_event_data.R
\docType{class}
\name{SpoolerChannelEventData}
\alias{SpoolerChannelEventData}
\title{SpoolerChannelEventData}
\format{
An \code{R6Class} generator object
}
\description{
SpoolerChannelEventData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character Name or label of the spooler channel.}

\item{\code{type}}{character Type of the spooler channel.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SpoolerChannelEventData-new}{\code{SpoolerChannelEventData$new()}}
\item \href{#method-SpoolerChannelEventData-validate}{\code{SpoolerChannelEventData$validate()}}
\item \href{#method-SpoolerChannelEventData-toJSON}{\code{SpoolerChannelEventData$toJSON()}}
\item \href{#method-SpoolerChannelEventData-fromJSON}{\code{SpoolerChannelEventData$fromJSON()}}
\item \href{#method-SpoolerChannelEventData-clone}{\code{SpoolerChannelEventData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpoolerChannelEventData-new"></a>}}
\if{latex}{\out{\hypertarget{method-SpoolerChannelEventData-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpoolerChannelEventData$new(
  name = NULL,
  type = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name or label of the spooler channel.}

\item{\code{type}}{Type of the spooler channel.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpoolerChannelEventData-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SpoolerChannelEventData-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpoolerChannelEventData$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpoolerChannelEventData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SpoolerChannelEventData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpoolerChannelEventData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpoolerChannelEventData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SpoolerChannelEventData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpoolerChannelEventData$fromJSON(
  SpoolerChannelEventDataJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SpoolerChannelEventDataJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpoolerChannelEventData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SpoolerChannelEventData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpoolerChannelEventData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
