% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_coefficients.R
\docType{class}
\name{StageCoefficients}
\alias{StageCoefficients}
\title{StageCoefficients}
\format{
An \code{R6Class} generator object
}
\description{
StageCoefficients Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{coefficient}}{numeric The corresponding value of the coefficient for that stage.}

\item{\code{stage}}{character The name of the stage.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StageCoefficients-new}{\code{StageCoefficients$new()}}
\item \href{#method-StageCoefficients-validate}{\code{StageCoefficients$validate()}}
\item \href{#method-StageCoefficients-toJSON}{\code{StageCoefficients$toJSON()}}
\item \href{#method-StageCoefficients-fromJSON}{\code{StageCoefficients$fromJSON()}}
\item \href{#method-StageCoefficients-clone}{\code{StageCoefficients$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StageCoefficients-new"></a>}}
\if{latex}{\out{\hypertarget{method-StageCoefficients-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StageCoefficients$new(
  coefficient = NULL,
  stage = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coefficient}}{The corresponding value of the coefficient for that stage.}

\item{\code{stage}}{The name of the stage.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StageCoefficients-validate"></a>}}
\if{latex}{\out{\hypertarget{method-StageCoefficients-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StageCoefficients$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StageCoefficients-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StageCoefficients-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StageCoefficients$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StageCoefficients-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StageCoefficients-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StageCoefficients$fromJSON(StageCoefficientsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{StageCoefficientsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StageCoefficients-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StageCoefficients-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StageCoefficients$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
