% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{StartNewAutoPilot}
\alias{StartNewAutoPilot}
\title{StartNewAutoPilot}
\usage{
StartNewAutoPilot(project, featurelistId, mode = AutopilotMode$FullAuto, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Start autopilot
}
\details{
Start autopilot on provided featurelist.

This method invokes \verb{POST /projects/\{projectId\}/autopilots/} in the DataRobot Public API.
}
\seealso{
The method AutopilotsCreate in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
