% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_intake.R
\docType{class}
\name{SynapseIntake}
\alias{SynapseIntake}
\title{SynapseIntake}
\format{
An \code{R6Class} generator object
}
\description{
SynapseIntake Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cloudStorageCredentialId}}{character \link{optional} The ID of the Azure credential holding information about a user with read access to the cloud storage.}

\item{\code{credentialId}}{character \link{optional} The ID of the credential holding information about a user with read access to the JDBC data source.}

\item{\code{dataStoreId}}{character ID of the data store to connect to}

\item{\code{externalDataSource}}{character External datasource name}

\item{\code{query}}{character \link{optional} A self-supplied SELECT statement of the dataset you wish to score. Helpful for supplying a more fine-grained selection of data not achievable through specification of \"table\" and/or \"schema\" parameters exclusively.If this job is executed with a job definition, then template variables are available which will be substituted for timestamps: {{ current_run_timestamp }}, {{ last_completed_run_time }}, {{ last_scheduled_run_time }}, {{ next_scheduled_run_time }}, {{ current_run_time }}}

\item{\code{schema}}{character \link{optional} The name of the specified database schema to read input data from.}

\item{\code{table}}{character \link{optional} The name of the specified database table to read input data from.}

\item{\code{type}}{character Type name for this intake type}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SynapseIntake-new}{\code{SynapseIntake$new()}}
\item \href{#method-SynapseIntake-validate}{\code{SynapseIntake$validate()}}
\item \href{#method-SynapseIntake-toJSON}{\code{SynapseIntake$toJSON()}}
\item \href{#method-SynapseIntake-fromJSON}{\code{SynapseIntake$fromJSON()}}
\item \href{#method-SynapseIntake-clone}{\code{SynapseIntake$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SynapseIntake-new"></a>}}
\if{latex}{\out{\hypertarget{method-SynapseIntake-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SynapseIntake$new(
  dataStoreId = NULL,
  externalDataSource = NULL,
  type = NULL,
  cloudStorageCredentialId = NULL,
  credentialId = NULL,
  query = NULL,
  schema = NULL,
  table = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataStoreId}}{ID of the data store to connect to}

\item{\code{externalDataSource}}{External datasource name}

\item{\code{type}}{Type name for this intake type}

\item{\code{cloudStorageCredentialId}}{The ID of the Azure credential holding information about a user with read access to the cloud storage.}

\item{\code{credentialId}}{The ID of the credential holding information about a user with read access to the JDBC data source.}

\item{\code{query}}{A self-supplied SELECT statement of the dataset you wish to score. Helpful for supplying a more fine-grained selection of data not achievable through specification of \"table\" and/or \"schema\" parameters exclusively.If this job is executed with a job definition, then template variables are available which will be substituted for timestamps: {{ current_run_timestamp }}, {{ last_completed_run_time }}, {{ last_scheduled_run_time }}, {{ next_scheduled_run_time }}, {{ current_run_time }}}

\item{\code{schema}}{The name of the specified database schema to read input data from.}

\item{\code{table}}{The name of the specified database table to read input data from.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SynapseIntake-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SynapseIntake-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SynapseIntake$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SynapseIntake-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SynapseIntake-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SynapseIntake$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SynapseIntake-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SynapseIntake-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SynapseIntake$fromJSON(SynapseIntakeJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SynapseIntakeJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SynapseIntake-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SynapseIntake-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SynapseIntake$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
