% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_output.R
\docType{class}
\name{SynapseOutput}
\alias{SynapseOutput}
\title{SynapseOutput}
\format{
An \code{R6Class} generator object
}
\description{
SynapseOutput Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cloudStorageCredentialId}}{character \link{optional} The ID of the credential holding information about a user with write access to the cloud storage.}

\item{\code{createTableIfNotExists}}{character \link{optional} Attempt to create the table first if no existing one is detected, before writing data with the strategy defined in the `statementType` parameter.}

\item{\code{credentialId}}{character \link{optional} The ID of the credential holding information about a user with write access to the JDBC data source.}

\item{\code{dataStoreId}}{character ID of the data store to connect to}

\item{\code{externalDataSource}}{character External data source name}

\item{\code{schema}}{character \link{optional} The name of the specified database schema to write results to.}

\item{\code{statementType}}{character The statement type to use when writing the results.}

\item{\code{table}}{character The name of the specified database table to write results to.}

\item{\code{type}}{character Type name for this output type}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SynapseOutput$new()}}
\item \href{#method-validate}{\code{SynapseOutput$validate()}}
\item \href{#method-toJSON}{\code{SynapseOutput$toJSON()}}
\item \href{#method-fromJSON}{\code{SynapseOutput$fromJSON()}}
\item \href{#method-clone}{\code{SynapseOutput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SynapseOutput$new(
  dataStoreId = NULL,
  externalDataSource = NULL,
  statementType = NULL,
  table = NULL,
  type = NULL,
  cloudStorageCredentialId = NULL,
  createTableIfNotExists = FALSE,
  credentialId = NULL,
  schema = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataStoreId}}{ID of the data store to connect to}

\item{\code{externalDataSource}}{External data source name}

\item{\code{statementType}}{The statement type to use when writing the results.}

\item{\code{table}}{The name of the specified database table to write results to.}

\item{\code{type}}{Type name for this output type}

\item{\code{cloudStorageCredentialId}}{The ID of the credential holding information about a user with write access to the cloud storage.}

\item{\code{createTableIfNotExists}}{Attempt to create the table first if no existing one is detected, before writing data with the strategy defined in the `statementType` parameter.}

\item{\code{credentialId}}{The ID of the credential holding information about a user with write access to the JDBC data source.}

\item{\code{schema}}{The name of the specified database schema to write results to.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SynapseOutput$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SynapseOutput$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SynapseOutput$fromJSON(SynapseOutputJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SynapseOutputJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SynapseOutput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
