% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau.R
\docType{class}
\name{Tableau}
\alias{Tableau}
\title{Tableau}
\format{
An \code{R6Class} generator object
}
\description{
Tableau Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{contentUrl}}{character \link{optional} Deprecated, use `siteName` instead}

\item{\code{credentialId}}{character \link{optional} Use the specified credential to access the url}

\item{\code{dataSourceId}}{character The ID of your Tableau data source}

\item{\code{overwrite}}{character Should the dataset be overwritten or appended to}

\item{\code{siteName}}{character \link{optional} Your Tableau site name}

\item{\code{type}}{character Type name for this output type}

\item{\code{url}}{character The URL to your online Tableau server}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Tableau-new}{\code{Tableau$new()}}
\item \href{#method-Tableau-validate}{\code{Tableau$validate()}}
\item \href{#method-Tableau-toJSON}{\code{Tableau$toJSON()}}
\item \href{#method-Tableau-fromJSON}{\code{Tableau$fromJSON()}}
\item \href{#method-Tableau-clone}{\code{Tableau$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tableau-new"></a>}}
\if{latex}{\out{\hypertarget{method-Tableau-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tableau$new(
  dataSourceId = NULL,
  overwrite = NULL,
  type = NULL,
  url = NULL,
  contentUrl = NULL,
  credentialId = NULL,
  siteName = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataSourceId}}{The ID of your Tableau data source}

\item{\code{overwrite}}{Should the dataset be overwritten or appended to}

\item{\code{type}}{Type name for this output type}

\item{\code{url}}{The URL to your online Tableau server}

\item{\code{contentUrl}}{Deprecated, use `siteName` instead}

\item{\code{credentialId}}{Use the specified credential to access the url}

\item{\code{siteName}}{Your Tableau site name}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tableau-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Tableau-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tableau$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tableau-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Tableau-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tableau$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tableau-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Tableau-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tableau$fromJSON(TableauJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TableauJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tableau-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Tableau-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tableau$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
