% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau_server_prediction_target_payload.R
\docType{class}
\name{TableauServerPredictionTargetPayload}
\alias{TableauServerPredictionTargetPayload}
\title{TableauServerPredictionTargetPayload}
\format{
An \code{R6Class} generator object
}
\description{
TableauServerPredictionTargetPayload Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{contentUrl}}{character Site name of the Tableau online or Tableau server you wish to use. Required for Tableau online users.}

\item{\code{datasourceId}}{character Id of the DSS data source object}

\item{\code{password}}{character User password to connect to the data location}

\item{\code{typeId}}{character For Tableau should be set to `tableauServer`}

\item{\code{url}}{character Url of the Tableau online or Tableau server instance}

\item{\code{userName}}{character User name to connect to the data location}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TableauServerPredictionTargetPayload$new()}}
\item \href{#method-validate}{\code{TableauServerPredictionTargetPayload$validate()}}
\item \href{#method-toJSON}{\code{TableauServerPredictionTargetPayload$toJSON()}}
\item \href{#method-fromJSON}{\code{TableauServerPredictionTargetPayload$fromJSON()}}
\item \href{#method-clone}{\code{TableauServerPredictionTargetPayload$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableauServerPredictionTargetPayload$new(
  contentUrl = NULL,
  datasourceId = NULL,
  password = NULL,
  typeId = NULL,
  url = NULL,
  userName = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{contentUrl}}{Site name of the Tableau online or Tableau server you wish to use. Required for Tableau online users.}

\item{\code{datasourceId}}{Id of the DSS data source object}

\item{\code{password}}{User password to connect to the data location}

\item{\code{typeId}}{For Tableau should be set to `tableauServer`}

\item{\code{url}}{Url of the Tableau online or Tableau server instance}

\item{\code{userName}}{User name to connect to the data location}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableauServerPredictionTargetPayload$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableauServerPredictionTargetPayload$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableauServerPredictionTargetPayload$fromJSON(
  TableauServerPredictionTargetPayloadJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TableauServerPredictionTargetPayloadJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableauServerPredictionTargetPayload$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
