% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_permissions_response.R
\docType{class}
\name{TemplatePermissionsResponse}
\alias{TemplatePermissionsResponse}
\title{TemplatePermissionsResponse}
\format{
An \code{R6Class} generator object
}
\description{
TemplatePermissionsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CANUnderscoreDELETEUnderscoreTEMPLATE}}{character Whether the current user can delete this template}

\item{\code{CANUnderscoreEDITUnderscoreTEMPLATE}}{character Whether the current user can edit this template}

\item{\code{CANUnderscoreSHARE}}{character Whether the current user can share this template}

\item{\code{CANUnderscoreUSEUnderscoreTEMPLATE}}{character Whether the current user can generate documents with this template}

\item{\code{CANUnderscoreVIEW}}{character Whether the current user can view this template}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TemplatePermissionsResponse-new}{\code{TemplatePermissionsResponse$new()}}
\item \href{#method-TemplatePermissionsResponse-validate}{\code{TemplatePermissionsResponse$validate()}}
\item \href{#method-TemplatePermissionsResponse-toJSON}{\code{TemplatePermissionsResponse$toJSON()}}
\item \href{#method-TemplatePermissionsResponse-fromJSON}{\code{TemplatePermissionsResponse$fromJSON()}}
\item \href{#method-TemplatePermissionsResponse-clone}{\code{TemplatePermissionsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TemplatePermissionsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-TemplatePermissionsResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemplatePermissionsResponse$new(
  CANUnderscoreDELETEUnderscoreTEMPLATE = NULL,
  CANUnderscoreEDITUnderscoreTEMPLATE = NULL,
  CANUnderscoreSHARE = NULL,
  CANUnderscoreUSEUnderscoreTEMPLATE = NULL,
  CANUnderscoreVIEW = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CANUnderscoreDELETEUnderscoreTEMPLATE}}{Whether the current user can delete this template}

\item{\code{CANUnderscoreEDITUnderscoreTEMPLATE}}{Whether the current user can edit this template}

\item{\code{CANUnderscoreSHARE}}{Whether the current user can share this template}

\item{\code{CANUnderscoreUSEUnderscoreTEMPLATE}}{Whether the current user can generate documents with this template}

\item{\code{CANUnderscoreVIEW}}{Whether the current user can view this template}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TemplatePermissionsResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-TemplatePermissionsResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemplatePermissionsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TemplatePermissionsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TemplatePermissionsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemplatePermissionsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TemplatePermissionsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TemplatePermissionsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemplatePermissionsResponse$fromJSON(
  TemplatePermissionsResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TemplatePermissionsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TemplatePermissionsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TemplatePermissionsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemplatePermissionsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
