% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{TestDataStore}
\alias{TestDataStore}
\title{TestDataStore}
\usage{
TestDataStore(dataStoreId, username, password, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Tests data store connection.
}
\details{
Tests the ability to connect to a data store with specified authentication.

This method invokes \verb{POST /externalDataStores/\{dataStoreId\}/test/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataStoresTestCreate in \link{DataConnectivityApi}, which invokes the same underlying API endpoint.
}
