% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/time_series_feature_log_list_controller_response.R
\docType{class}
\name{TimeSeriesFeatureLogListControllerResponse}
\alias{TimeSeriesFeatureLogListControllerResponse}
\title{TimeSeriesFeatureLogListControllerResponse}
\format{
An \code{R6Class} generator object
}
\description{
TimeSeriesFeatureLogListControllerResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer The number of items returned on this page.}

\item{\code{featureLog}}{character The content of the feature log.}

\item{\code{next_}}{character URL pointing to the next page (if null, there is no next page).}

\item{\code{previous}}{character URL pointing to the previous page (if null, there is no previous page).}

\item{\code{totalLogLines}}{integer The total number of lines in feature derivation log.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TimeSeriesFeatureLogListControllerResponse-new}{\code{TimeSeriesFeatureLogListControllerResponse$new()}}
\item \href{#method-TimeSeriesFeatureLogListControllerResponse-validate}{\code{TimeSeriesFeatureLogListControllerResponse$validate()}}
\item \href{#method-TimeSeriesFeatureLogListControllerResponse-toJSON}{\code{TimeSeriesFeatureLogListControllerResponse$toJSON()}}
\item \href{#method-TimeSeriesFeatureLogListControllerResponse-fromJSON}{\code{TimeSeriesFeatureLogListControllerResponse$fromJSON()}}
\item \href{#method-TimeSeriesFeatureLogListControllerResponse-clone}{\code{TimeSeriesFeatureLogListControllerResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeSeriesFeatureLogListControllerResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-TimeSeriesFeatureLogListControllerResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeSeriesFeatureLogListControllerResponse$new(
  count = NULL,
  featureLog = NULL,
  next_ = NULL,
  previous = NULL,
  totalLogLines = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{The number of items returned on this page.}

\item{\code{featureLog}}{The content of the feature log.}

\item{\code{next_}}{URL pointing to the next page (if null, there is no next page).}

\item{\code{previous}}{URL pointing to the previous page (if null, there is no previous page).}

\item{\code{totalLogLines}}{The total number of lines in feature derivation log.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeSeriesFeatureLogListControllerResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-TimeSeriesFeatureLogListControllerResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeSeriesFeatureLogListControllerResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeSeriesFeatureLogListControllerResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TimeSeriesFeatureLogListControllerResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeSeriesFeatureLogListControllerResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeSeriesFeatureLogListControllerResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TimeSeriesFeatureLogListControllerResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeSeriesFeatureLogListControllerResponse$fromJSON(
  TimeSeriesFeatureLogListControllerResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TimeSeriesFeatureLogListControllerResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeSeriesFeatureLogListControllerResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeSeriesFeatureLogListControllerResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeSeriesFeatureLogListControllerResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
