% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_fields_payload.R
\docType{class}
\name{TimeSeriesFieldsPayload}
\alias{TimeSeriesFieldsPayload}
\title{TimeSeriesFieldsPayload}
\format{
An \code{R6Class} generator object
}
\description{
TimeSeriesFieldsPayload Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{forecastPoint}}{character \link{optional} A template of format for the expected day of prediction}

\item{\code{predictionEndDate}}{character \link{optional} A template of format for the end date of prediction}

\item{\code{predictionStartDate}}{character \link{optional} A template of format for the start date of prediction}

\item{\code{relaxKnownInAdvanceFeaturesCheck}}{character If `True`, missing values in the known in advance features are allowed in the forecast window at the prediction time. If omitted or `False`, missing values are not allowed.}

\item{\code{type}}{character One of forecast, historical}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TimeSeriesFieldsPayload$new()}}
\item \href{#method-validate}{\code{TimeSeriesFieldsPayload$validate()}}
\item \href{#method-toJSON}{\code{TimeSeriesFieldsPayload$toJSON()}}
\item \href{#method-fromJSON}{\code{TimeSeriesFieldsPayload$fromJSON()}}
\item \href{#method-clone}{\code{TimeSeriesFieldsPayload$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeSeriesFieldsPayload$new(
  relaxKnownInAdvanceFeaturesCheck = NULL,
  type = NULL,
  forecastPoint = NULL,
  predictionEndDate = NULL,
  predictionStartDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{relaxKnownInAdvanceFeaturesCheck}}{If `True`, missing values in the known in advance features are allowed in the forecast window at the prediction time. If omitted or `False`, missing values are not allowed.}

\item{\code{type}}{One of forecast, historical}

\item{\code{forecastPoint}}{A template of format for the expected day of prediction}

\item{\code{predictionEndDate}}{A template of format for the end date of prediction}

\item{\code{predictionStartDate}}{A template of format for the start date of prediction}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeSeriesFieldsPayload$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeSeriesFieldsPayload$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeSeriesFieldsPayload$fromJSON(
  TimeSeriesFieldsPayloadJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TimeSeriesFieldsPayloadJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeSeriesFieldsPayload$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
