% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_options.R
\docType{class}
\name{TimeSeriesOptions}
\alias{TimeSeriesOptions}
\title{TimeSeriesOptions}
\format{
An \code{R6Class} generator object
}
\description{
TimeSeriesOptions Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{calendarId}}{character \link{optional} The ID of the calendar to be used in this project.}

\item{\code{differencingMethod}}{character \link{optional} For time series projects only. Used to specify which differencing method to apply if the data is stationary. For classification problems `simple` and `seasonal` are not allowed. Parameter `periodicities` must be specified if `seasonal` is chosen. Defaults to `auto`.}

\item{\code{exponentiallyWeightedMovingAlpha}}{numeric \link{optional} Discount factor (alpha) used for exponentially weighted moving features}

\item{\code{periodicities}}{list( \link{Periodicity} ) \link{optional} A list of periodicities for time series projects only. For classification problems periodicities are not allowed. If this is provided, parameter 'differencing_method' will default to 'seasonal' if not provided or 'auto'.}

\item{\code{treatAsExponential}}{character \link{optional} For time series projects only. Used to specify whether to treat data as exponential trend and apply transformations like log-transform. For classification problems `always` is not allowed. Defaults to `auto`.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TimeSeriesOptions-new}{\code{TimeSeriesOptions$new()}}
\item \href{#method-TimeSeriesOptions-validate}{\code{TimeSeriesOptions$validate()}}
\item \href{#method-TimeSeriesOptions-toJSON}{\code{TimeSeriesOptions$toJSON()}}
\item \href{#method-TimeSeriesOptions-fromJSON}{\code{TimeSeriesOptions$fromJSON()}}
\item \href{#method-TimeSeriesOptions-clone}{\code{TimeSeriesOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeSeriesOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-TimeSeriesOptions-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeSeriesOptions$new(
  calendarId = NULL,
  differencingMethod = NULL,
  exponentiallyWeightedMovingAlpha = NULL,
  periodicities = NULL,
  treatAsExponential = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{calendarId}}{The ID of the calendar to be used in this project.}

\item{\code{differencingMethod}}{For time series projects only. Used to specify which differencing method to apply if the data is stationary. For classification problems `simple` and `seasonal` are not allowed. Parameter `periodicities` must be specified if `seasonal` is chosen. Defaults to `auto`.}

\item{\code{exponentiallyWeightedMovingAlpha}}{Discount factor (alpha) used for exponentially weighted moving features}

\item{\code{periodicities}}{A list of periodicities for time series projects only. For classification problems periodicities are not allowed. If this is provided, parameter 'differencing_method' will default to 'seasonal' if not provided or 'auto'.}

\item{\code{treatAsExponential}}{For time series projects only. Used to specify whether to treat data as exponential trend and apply transformations like log-transform. For classification problems `always` is not allowed. Defaults to `auto`.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeSeriesOptions-validate"></a>}}
\if{latex}{\out{\hypertarget{method-TimeSeriesOptions-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeSeriesOptions$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeSeriesOptions-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TimeSeriesOptions-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeSeriesOptions$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeSeriesOptions-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TimeSeriesOptions-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeSeriesOptions$fromJSON(TimeSeriesOptionsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TimeSeriesOptionsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeSeriesOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeSeriesOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeSeriesOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
