% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_datetime_frozen_model.R
\docType{class}
\name{TrainDatetimeFrozenModel}
\alias{TrainDatetimeFrozenModel}
\title{TrainDatetimeFrozenModel}
\format{
An \code{R6Class} generator object
}
\description{
TrainDatetimeFrozenModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{modelId}}{character The ID of an existing model to use as the source for the training parameters.}

\item{\code{samplingMethod}}{character \link{optional} Defines how training data is selected if subsampling is used (e.g., if `timeWindowSamplePct` is specified). Can be either ``random`` or ``latest``. If omitted, defaults to ``latest`` if `trainingRowCount` is used and ``random`` for other cases (e.g., if `trainingDuration` or `useProjectSettings` is specified). May only be specified for OTV projects.}

\item{\code{timeWindowSamplePct}}{integer \link{optional} An integer between 1 and 99 indicating the percentage of sampling within the time window. The points kept are determined by the value provided for the `samplingMethod` option. If specified, `trainingRowCount` may not be specified, and the specified model must either be a duration or selectedDateRange model, or one of `trainingDuration` or `trainingStartDate` and `trainingEndDate` must be specified.}

\item{\code{trainingDuration}}{character \link{optional} A duration string representing the training duration for the submitted model. Only one of `trainingDuration`, `trainingRowCount`, `trainingStartDate` and `trainingEndDate`, or `useProjectSettings` may be specified.}

\item{\code{trainingEndDate}}{character \link{optional} A datetime string representing the end date of the data to use for training this model. If specified, `trainingStartDate` must also be specified. Only one of `trainingDuration`, `trainingRowCount`, `trainingStartDate` and `trainingEndDate`, or `useProjectSettings` may be specified.}

\item{\code{trainingRowCount}}{integer \link{optional} The number of rows of data that should be used when training this model. Only one of `trainingDuration`, `trainingRowCount`, `trainingStartDate` and `trainingEndDate`, or `useProjectSettings` may be specified.}

\item{\code{trainingStartDate}}{character \link{optional} A datetime string representing the start date of the data to use for training this model. If specified, `trainingEndDate` must also be specified. Only one of `trainingDuration`, `trainingRowCount`, `trainingStartDate` and `trainingEndDate`, or `useProjectSettings` may be specified.}

\item{\code{useProjectSettings}}{character \link{optional} If ``True``, the model will be trained using the previously-specified custom backtest training settings. Only one of `trainingDuration`, `trainingRowCount`, `trainingStartDate` and `trainingEndDate`, or `useProjectSettings` may be specified.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TrainDatetimeFrozenModel$new()}}
\item \href{#method-validate}{\code{TrainDatetimeFrozenModel$validate()}}
\item \href{#method-toJSON}{\code{TrainDatetimeFrozenModel$toJSON()}}
\item \href{#method-fromJSON}{\code{TrainDatetimeFrozenModel$fromJSON()}}
\item \href{#method-clone}{\code{TrainDatetimeFrozenModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainDatetimeFrozenModel$new(
  modelId = NULL,
  samplingMethod = NULL,
  timeWindowSamplePct = NULL,
  trainingDuration = NULL,
  trainingEndDate = NULL,
  trainingRowCount = NULL,
  trainingStartDate = NULL,
  useProjectSettings = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{modelId}}{The ID of an existing model to use as the source for the training parameters.}

\item{\code{samplingMethod}}{Defines how training data is selected if subsampling is used (e.g., if `timeWindowSamplePct` is specified). Can be either ``random`` or ``latest``. If omitted, defaults to ``latest`` if `trainingRowCount` is used and ``random`` for other cases (e.g., if `trainingDuration` or `useProjectSettings` is specified). May only be specified for OTV projects.}

\item{\code{timeWindowSamplePct}}{An integer between 1 and 99 indicating the percentage of sampling within the time window. The points kept are determined by the value provided for the `samplingMethod` option. If specified, `trainingRowCount` may not be specified, and the specified model must either be a duration or selectedDateRange model, or one of `trainingDuration` or `trainingStartDate` and `trainingEndDate` must be specified.}

\item{\code{trainingDuration}}{A duration string representing the training duration for the submitted model. Only one of `trainingDuration`, `trainingRowCount`, `trainingStartDate` and `trainingEndDate`, or `useProjectSettings` may be specified.}

\item{\code{trainingEndDate}}{A datetime string representing the end date of the data to use for training this model. If specified, `trainingStartDate` must also be specified. Only one of `trainingDuration`, `trainingRowCount`, `trainingStartDate` and `trainingEndDate`, or `useProjectSettings` may be specified.}

\item{\code{trainingRowCount}}{The number of rows of data that should be used when training this model. Only one of `trainingDuration`, `trainingRowCount`, `trainingStartDate` and `trainingEndDate`, or `useProjectSettings` may be specified.}

\item{\code{trainingStartDate}}{A datetime string representing the start date of the data to use for training this model. If specified, `trainingEndDate` must also be specified. Only one of `trainingDuration`, `trainingRowCount`, `trainingStartDate` and `trainingEndDate`, or `useProjectSettings` may be specified.}

\item{\code{useProjectSettings}}{If ``True``, the model will be trained using the previously-specified custom backtest training settings. Only one of `trainingDuration`, `trainingRowCount`, `trainingStartDate` and `trainingEndDate`, or `useProjectSettings` may be specified.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainDatetimeFrozenModel$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainDatetimeFrozenModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainDatetimeFrozenModel$fromJSON(
  TrainDatetimeFrozenModelJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TrainDatetimeFrozenModelJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainDatetimeFrozenModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
