% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_history_entry.R
\docType{class}
\name{TrainingHistoryEntry}
\alias{TrainingHistoryEntry}
\title{TrainingHistoryEntry}
\format{
An \code{R6Class} generator object
}
\description{
TrainingHistoryEntry Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{creationDate}}{character ISO-8601 timestamp of when the project the blueprint was trained on was created.}

\item{\code{lid}}{character The leaderboard ID the blueprint was trained on.}

\item{\code{pid}}{character The project ID the blueprint was trained on.}

\item{\code{projectModelsCount}}{integer Number of models in the project the blueprint was trained on.}

\item{\code{projectName}}{character The project name the blueprint was trained on.}

\item{\code{targetName}}{character The target name of the project the blueprint was trained on.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TrainingHistoryEntry-new}{\code{TrainingHistoryEntry$new()}}
\item \href{#method-TrainingHistoryEntry-validate}{\code{TrainingHistoryEntry$validate()}}
\item \href{#method-TrainingHistoryEntry-toJSON}{\code{TrainingHistoryEntry$toJSON()}}
\item \href{#method-TrainingHistoryEntry-fromJSON}{\code{TrainingHistoryEntry$fromJSON()}}
\item \href{#method-TrainingHistoryEntry-clone}{\code{TrainingHistoryEntry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainingHistoryEntry-new"></a>}}
\if{latex}{\out{\hypertarget{method-TrainingHistoryEntry-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingHistoryEntry$new(
  creationDate = NULL,
  lid = NULL,
  pid = NULL,
  projectModelsCount = NULL,
  projectName = NULL,
  targetName = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{creationDate}}{ISO-8601 timestamp of when the project the blueprint was trained on was created.}

\item{\code{lid}}{The leaderboard ID the blueprint was trained on.}

\item{\code{pid}}{The project ID the blueprint was trained on.}

\item{\code{projectModelsCount}}{Number of models in the project the blueprint was trained on.}

\item{\code{projectName}}{The project name the blueprint was trained on.}

\item{\code{targetName}}{The target name of the project the blueprint was trained on.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainingHistoryEntry-validate"></a>}}
\if{latex}{\out{\hypertarget{method-TrainingHistoryEntry-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingHistoryEntry$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainingHistoryEntry-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TrainingHistoryEntry-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingHistoryEntry$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainingHistoryEntry-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TrainingHistoryEntry-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingHistoryEntry$fromJSON(TrainingHistoryEntryJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TrainingHistoryEntryJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainingHistoryEntry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TrainingHistoryEntry-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingHistoryEntry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
