% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_info_response.R
\docType{class}
\name{TrainingInfoResponse}
\alias{TrainingInfoResponse}
\title{TrainingInfoResponse}
\format{
An \code{R6Class} generator object
}
\description{
TrainingInfoResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{holdoutTrainingDuration}}{character the duration of the data used to train a model to score the holdout}

\item{\code{holdoutTrainingEndDate}}{character the end date of the data used to train a model to score the holdout}

\item{\code{holdoutTrainingRowCount}}{integer the number of rows used to train a model to score the holdout}

\item{\code{holdoutTrainingStartDate}}{character the start date of data used to train a model to score the holdout}

\item{\code{predictionTrainingDuration}}{character the duration of the data used to train a model to make predictions}

\item{\code{predictionTrainingEndDate}}{character the end date of the data used to train a model to make predictions}

\item{\code{predictionTrainingRowCount}}{integer the number of rows used to train a model to make predictions}

\item{\code{predictionTrainingStartDate}}{character the start date of data used to train a model to make predictions}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TrainingInfoResponse-new}{\code{TrainingInfoResponse$new()}}
\item \href{#method-TrainingInfoResponse-validate}{\code{TrainingInfoResponse$validate()}}
\item \href{#method-TrainingInfoResponse-toJSON}{\code{TrainingInfoResponse$toJSON()}}
\item \href{#method-TrainingInfoResponse-fromJSON}{\code{TrainingInfoResponse$fromJSON()}}
\item \href{#method-TrainingInfoResponse-clone}{\code{TrainingInfoResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainingInfoResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-TrainingInfoResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingInfoResponse$new(
  holdoutTrainingDuration = NULL,
  holdoutTrainingEndDate = NULL,
  holdoutTrainingRowCount = NULL,
  holdoutTrainingStartDate = NULL,
  predictionTrainingDuration = NULL,
  predictionTrainingEndDate = NULL,
  predictionTrainingRowCount = NULL,
  predictionTrainingStartDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{holdoutTrainingDuration}}{the duration of the data used to train a model to score the holdout}

\item{\code{holdoutTrainingEndDate}}{the end date of the data used to train a model to score the holdout}

\item{\code{holdoutTrainingRowCount}}{the number of rows used to train a model to score the holdout}

\item{\code{holdoutTrainingStartDate}}{the start date of data used to train a model to score the holdout}

\item{\code{predictionTrainingDuration}}{the duration of the data used to train a model to make predictions}

\item{\code{predictionTrainingEndDate}}{the end date of the data used to train a model to make predictions}

\item{\code{predictionTrainingRowCount}}{the number of rows used to train a model to make predictions}

\item{\code{predictionTrainingStartDate}}{the start date of data used to train a model to make predictions}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainingInfoResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-TrainingInfoResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingInfoResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainingInfoResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TrainingInfoResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingInfoResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainingInfoResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TrainingInfoResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingInfoResponse$fromJSON(TrainingInfoResponseJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TrainingInfoResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainingInfoResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TrainingInfoResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingInfoResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
