% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\docType{class}
\name{TrainModel}
\alias{TrainModel}
\title{TrainModel}
\format{
An \code{R6Class} generator object
}
\description{
TrainModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{blueprintId}}{character The ID of a blueprint to use to generate the model. Allowed blueprints can be retrieved using :http:get:`/api/v2/projects/(projectId)/blueprints/` or taken from existing models.}

\item{\code{featurelistId}}{character \link{optional} If specified, the model will be trained using this featurelist. If not specified, the recommended featurelist for the specified blueprint will be used. If there is no recommended featurelist, the project's default will be used.}

\item{\code{monotonicDecreasingFeaturelistId}}{character \link{optional} The ID of the featurelist that defines the set of features with a monotonically decreasing relationship to the target. If ``null``, no constraints will be enforced. If omitted, the project default is used.}

\item{\code{monotonicIncreasingFeaturelistId}}{character \link{optional} The ID of the featurelist that defines the set of features with a monotonically increasing relationship to the target. If ``null``, no constraints will be enforced. If omitted, the project default is used.}

\item{\code{nClusters}}{integer \link{optional} The number of clusters to use in the specified unsupervised clustering model. Only valid in unsupervised clustering projects.}

\item{\code{samplePct}}{numeric \link{optional} The percentage of the dataset to use with the model. Only one of `samplePct` and `trainingRowCount` should be specified. The specified percentage should be between 0 and 100.}

\item{\code{scoringType}}{character \link{optional} Validation is available for any partitioning. If the project uses cross validation, `crossValidation` may be used to indicate that all available training/validation combinations should be used.}

\item{\code{sourceProjectId}}{character \link{optional} The project the blueprint comes from. Required only if the `blueprintId` comes from a different project.}

\item{\code{trainingRowCount}}{integer \link{optional} An integer representing the number of rows of the dataset to use with the model. Only one of `samplePct` and `trainingRowCount` should be specified.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TrainModel-new}{\code{TrainModel$new()}}
\item \href{#method-TrainModel-validate}{\code{TrainModel$validate()}}
\item \href{#method-TrainModel-toJSON}{\code{TrainModel$toJSON()}}
\item \href{#method-TrainModel-fromJSON}{\code{TrainModel$fromJSON()}}
\item \href{#method-TrainModel-clone}{\code{TrainModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-TrainModel-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainModel$new(
  blueprintId = NULL,
  featurelistId = NULL,
  monotonicDecreasingFeaturelistId = NULL,
  monotonicIncreasingFeaturelistId = NULL,
  nClusters = NULL,
  samplePct = NULL,
  scoringType = NULL,
  sourceProjectId = NULL,
  trainingRowCount = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blueprintId}}{The ID of a blueprint to use to generate the model. Allowed blueprints can be retrieved using :http:get:`/api/v2/projects/(projectId)/blueprints/` or taken from existing models.}

\item{\code{featurelistId}}{If specified, the model will be trained using this featurelist. If not specified, the recommended featurelist for the specified blueprint will be used. If there is no recommended featurelist, the project's default will be used.}

\item{\code{monotonicDecreasingFeaturelistId}}{The ID of the featurelist that defines the set of features with a monotonically decreasing relationship to the target. If ``null``, no constraints will be enforced. If omitted, the project default is used.}

\item{\code{monotonicIncreasingFeaturelistId}}{The ID of the featurelist that defines the set of features with a monotonically increasing relationship to the target. If ``null``, no constraints will be enforced. If omitted, the project default is used.}

\item{\code{nClusters}}{The number of clusters to use in the specified unsupervised clustering model. Only valid in unsupervised clustering projects.}

\item{\code{samplePct}}{The percentage of the dataset to use with the model. Only one of `samplePct` and `trainingRowCount` should be specified. The specified percentage should be between 0 and 100.}

\item{\code{scoringType}}{Validation is available for any partitioning. If the project uses cross validation, `crossValidation` may be used to indicate that all available training/validation combinations should be used.}

\item{\code{sourceProjectId}}{The project the blueprint comes from. Required only if the `blueprintId` comes from a different project.}

\item{\code{trainingRowCount}}{An integer representing the number of rows of the dataset to use with the model. Only one of `samplePct` and `trainingRowCount` should be specified.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainModel-validate"></a>}}
\if{latex}{\out{\hypertarget{method-TrainModel-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainModel$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainModel-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TrainModel-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainModel-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TrainModel-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainModel$fromJSON(TrainModelJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TrainModelJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TrainModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
