% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traning_prediction_row.R
\docType{class}
\name{TraningPredictionRow}
\alias{TraningPredictionRow}
\title{TraningPredictionRow}
\format{
An \code{R6Class} generator object
}
\description{
TraningPredictionRow Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{forecastDistance}}{integer \link{optional} (if time series project) The number of time units this prediction is away from the forecastPoint. The unit of time is determined by the timeUnit of the datetime partition column.}

\item{\code{forecastPoint}}{character \link{optional} (if time series project) The forecastPoint of the predictions. Either provided or inferred.}

\item{\code{partitionId}}{character The partition used for the prediction record}

\item{\code{prediction}}{\link{OneOfnumberstringarray} The prediction of the model.}

\item{\code{predictionExplanations}}{list( \link{PredictionExplanationsObject} ) \link{optional} Array contains `predictionExplanation` objects. The total elements in the array are bounded by maxExplanations and feature count. It will be present only if `explanationAlgorithm` is not null (prediction explanations were requested).}

\item{\code{predictionThreshold}}{numeric \link{optional} Threshold used for binary classification in predictions.}

\item{\code{predictionValues}}{list( \link{PredictionArrayObjectValues} ) \link{optional} A list of predicted values for this row.}

\item{\code{rowId}}{integer The row in the prediction dataset this prediction corresponds to.}

\item{\code{seriesId}}{character \link{optional} The ID of the series value for a multiseries project. For time series projects that are not a multiseries this will be a NaN.}

\item{\code{shapMetadata}}{\link{TraningPredictionShapMetadata} \link{optional}}

\item{\code{timestamp}}{character \link{optional} (if time series project) The timestamp of this row in the prediction dataset.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TraningPredictionRow-new}{\code{TraningPredictionRow$new()}}
\item \href{#method-TraningPredictionRow-validate}{\code{TraningPredictionRow$validate()}}
\item \href{#method-TraningPredictionRow-toJSON}{\code{TraningPredictionRow$toJSON()}}
\item \href{#method-TraningPredictionRow-fromJSON}{\code{TraningPredictionRow$fromJSON()}}
\item \href{#method-TraningPredictionRow-clone}{\code{TraningPredictionRow$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraningPredictionRow-new"></a>}}
\if{latex}{\out{\hypertarget{method-TraningPredictionRow-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraningPredictionRow$new(
  partitionId = NULL,
  prediction = NULL,
  rowId = NULL,
  forecastDistance = NULL,
  forecastPoint = NULL,
  predictionExplanations = NULL,
  predictionThreshold = NULL,
  predictionValues = NULL,
  seriesId = NULL,
  shapMetadata = NULL,
  timestamp = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{partitionId}}{The partition used for the prediction record}

\item{\code{prediction}}{The prediction of the model.}

\item{\code{rowId}}{The row in the prediction dataset this prediction corresponds to.}

\item{\code{forecastDistance}}{(if time series project) The number of time units this prediction is away from the forecastPoint. The unit of time is determined by the timeUnit of the datetime partition column.}

\item{\code{forecastPoint}}{(if time series project) The forecastPoint of the predictions. Either provided or inferred.}

\item{\code{predictionExplanations}}{Array contains `predictionExplanation` objects. The total elements in the array are bounded by maxExplanations and feature count. It will be present only if `explanationAlgorithm` is not null (prediction explanations were requested).}

\item{\code{predictionThreshold}}{Threshold used for binary classification in predictions.}

\item{\code{predictionValues}}{A list of predicted values for this row.}

\item{\code{seriesId}}{The ID of the series value for a multiseries project. For time series projects that are not a multiseries this will be a NaN.}

\item{\code{timestamp}}{(if time series project) The timestamp of this row in the prediction dataset.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraningPredictionRow-validate"></a>}}
\if{latex}{\out{\hypertarget{method-TraningPredictionRow-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraningPredictionRow$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraningPredictionRow-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TraningPredictionRow-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraningPredictionRow$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraningPredictionRow-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TraningPredictionRow-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraningPredictionRow$fromJSON(TraningPredictionRowJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TraningPredictionRowJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraningPredictionRow-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TraningPredictionRow-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraningPredictionRow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
