% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traning_predictions.R
\docType{class}
\name{TraningPredictions}
\alias{TraningPredictions}
\title{TraningPredictions}
\format{
An \code{R6Class} generator object
}
\description{
TraningPredictions Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dataSubset}}{character Subset of data predicted on}

\item{\code{explanationAlgorithm}}{character \link{optional} The method used for calculating prediction explanations}

\item{\code{id}}{character ID of the training prediction job}

\item{\code{maxExplanations}}{integer \link{optional} the number of top contributors that are included in prediction explanations. Defaults to null for datasets narrower than 100 columns, defaults to 100 for datasets wider than 100 columns}

\item{\code{modelId}}{character ID of the model}

\item{\code{shapWarnings}}{list( \link{ShapWarning} ) \link{optional} Will be present if \"explanationAlgorithm\" was set to \"shap\" and there were additivity failures during SHAP values calculation}

\item{\code{url}}{character The location of these predictions}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TraningPredictions-new}{\code{TraningPredictions$new()}}
\item \href{#method-TraningPredictions-validate}{\code{TraningPredictions$validate()}}
\item \href{#method-TraningPredictions-toJSON}{\code{TraningPredictions$toJSON()}}
\item \href{#method-TraningPredictions-fromJSON}{\code{TraningPredictions$fromJSON()}}
\item \href{#method-TraningPredictions-clone}{\code{TraningPredictions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraningPredictions-new"></a>}}
\if{latex}{\out{\hypertarget{method-TraningPredictions-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraningPredictions$new(
  dataSubset = NULL,
  id = NULL,
  modelId = NULL,
  url = NULL,
  explanationAlgorithm = NULL,
  maxExplanations = NULL,
  shapWarnings = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataSubset}}{Subset of data predicted on}

\item{\code{id}}{ID of the training prediction job}

\item{\code{modelId}}{ID of the model}

\item{\code{url}}{The location of these predictions}

\item{\code{explanationAlgorithm}}{The method used for calculating prediction explanations}

\item{\code{maxExplanations}}{the number of top contributors that are included in prediction explanations. Defaults to null for datasets narrower than 100 columns, defaults to 100 for datasets wider than 100 columns}

\item{\code{shapWarnings}}{Will be present if \"explanationAlgorithm\" was set to \"shap\" and there were additivity failures during SHAP values calculation}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraningPredictions-validate"></a>}}
\if{latex}{\out{\hypertarget{method-TraningPredictions-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraningPredictions$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraningPredictions-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TraningPredictions-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraningPredictions$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraningPredictions-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TraningPredictions-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraningPredictions$fromJSON(TraningPredictionsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TraningPredictionsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TraningPredictions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TraningPredictions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraningPredictions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
