% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigger.R
\docType{class}
\name{Trigger}
\alias{Trigger}
\title{Trigger}
\format{
An \code{R6Class} generator object
}
\description{
Trigger Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{schedule}}{\link{Schedule}}

\item{\code{statusDeclinesToFailing}}{character Identifies when trigger type is based on deployment a health status, whether the policy will run when health status declines to failing.}

\item{\code{statusDeclinesToWarning}}{character Identifies when trigger type is based on deployment a health status, whether the policy will run when health status declines to warning.}

\item{\code{type}}{character Type of retraining policy trigger.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Trigger-new}{\code{Trigger$new()}}
\item \href{#method-Trigger-validate}{\code{Trigger$validate()}}
\item \href{#method-Trigger-toJSON}{\code{Trigger$toJSON()}}
\item \href{#method-Trigger-fromJSON}{\code{Trigger$fromJSON()}}
\item \href{#method-Trigger-clone}{\code{Trigger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trigger-new"></a>}}
\if{latex}{\out{\hypertarget{method-Trigger-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trigger$new(
  schedule = NULL,
  statusDeclinesToFailing = NULL,
  statusDeclinesToWarning = NULL,
  type = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{statusDeclinesToFailing}}{Identifies when trigger type is based on deployment a health status, whether the policy will run when health status declines to failing.}

\item{\code{statusDeclinesToWarning}}{Identifies when trigger type is based on deployment a health status, whether the policy will run when health status declines to warning.}

\item{\code{type}}{Type of retraining policy trigger.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trigger-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Trigger-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trigger$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trigger-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Trigger-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trigger$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trigger-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Trigger-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trigger$fromJSON(TriggerJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TriggerJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trigger-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Trigger-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trigger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
