% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_parameters.R
\docType{class}
\name{TuningParameters}
\alias{TuningParameters}
\title{TuningParameters}
\format{
An \code{R6Class} generator object
}
\description{
TuningParameters Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{constraints}}{\link{Constraints}}

\item{\code{currentValue}}{\link{OneOfAnyTypearray} Value of this parameter used by the current model.}

\item{\code{defaultValue}}{\link{OneOfAnyTypearray} Default value of this parameter.}

\item{\code{parameterId}}{character Unique (per-blueprint) identifier of this parameter. This is the identifier used to specify which parameter to tune when make a new advanced tuning request.}

\item{\code{parameterName}}{character Name of the parameter.}

\item{\code{taskName}}{character Human-readable name of the task that this parameter belongs to.}

\item{\code{vertexId}}{character Id of the vertex this parameter belongs to.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TuningParameters-new}{\code{TuningParameters$new()}}
\item \href{#method-TuningParameters-validate}{\code{TuningParameters$validate()}}
\item \href{#method-TuningParameters-toJSON}{\code{TuningParameters$toJSON()}}
\item \href{#method-TuningParameters-fromJSON}{\code{TuningParameters$fromJSON()}}
\item \href{#method-TuningParameters-clone}{\code{TuningParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TuningParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-TuningParameters-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TuningParameters$new(
  constraints = NULL,
  currentValue = NULL,
  defaultValue = NULL,
  parameterId = NULL,
  parameterName = NULL,
  taskName = NULL,
  vertexId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{currentValue}}{Value of this parameter used by the current model.}

\item{\code{defaultValue}}{Default value of this parameter.}

\item{\code{parameterId}}{Unique (per-blueprint) identifier of this parameter. This is the identifier used to specify which parameter to tune when make a new advanced tuning request.}

\item{\code{parameterName}}{Name of the parameter.}

\item{\code{taskName}}{Human-readable name of the task that this parameter belongs to.}

\item{\code{vertexId}}{Id of the vertex this parameter belongs to.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TuningParameters-validate"></a>}}
\if{latex}{\out{\hypertarget{method-TuningParameters-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TuningParameters$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TuningParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TuningParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TuningParameters$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TuningParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TuningParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TuningParameters$fromJSON(TuningParametersJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TuningParametersJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TuningParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TuningParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TuningParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
