% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{UpdateCalendar}
\alias{UpdateCalendar}
\title{UpdateCalendar}
\usage{
UpdateCalendar(calendarId, name = NULL, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update a calendar's name
}
\details{
Update a calendar's name

This method invokes \verb{PATCH /calendars/\{calendarId\}/} in the DataRobot Public API.
}
\seealso{
The method CalendarsPatch in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
