% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectivity_apiWrapper.R
\name{UpdateDataSource}
\alias{UpdateDataSource}
\title{UpdateDataSource}
\usage{
UpdateDataSource(
  dataSourceId,
  canonicalName = NULL,
  dataStoreId = NULL,
  query = NULL,
  table = NULL,
  schema = NULL,
  partitionColumn = NULL,
  fetchSize = NULL,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update the data source.
}
\details{
Update the data source with given ID. To test the SQL query before updating the data source,use :\url{http:post}:/api/v2/externalDataStores/(dataStoreId)/verifySQL/.

This method invokes \verb{PATCH /externalDataSources/\{dataSourceId\}/} in the DataRobot Public API.
}
\seealso{
The method ExternalDataSourcesPatch in \link{DataConnectivityApi}, which invokes the same underlying API endpoint.
}
