% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_driver.R
\docType{class}
\name{UpdateDriver}
\alias{UpdateDriver}
\title{UpdateDriver}
\format{
An \code{R6Class} generator object
}
\description{
UpdateDriver Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{baseNames}}{list( character ) \link{optional}}

\item{\code{canonicalName}}{character \link{optional}}

\item{\code{className}}{character \link{optional}}

\item{\code{configurationId}}{character \link{optional}}

\item{\code{localJarUrls}}{list( character ) \link{optional}}

\item{\code{removeConfig}}{character \link{optional}}

\item{\code{skipConfigVerification}}{character \link{optional}}

\item{\code{version}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UpdateDriver$new()}}
\item \href{#method-toJSON}{\code{UpdateDriver$toJSON()}}
\item \href{#method-fromJSON}{\code{UpdateDriver$fromJSON()}}
\item \href{#method-clone}{\code{UpdateDriver$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateDriver$new(
  baseNames = NULL,
  canonicalName = NULL,
  className = NULL,
  configurationId = NULL,
  localJarUrls = NULL,
  removeConfig = NULL,
  skipConfigVerification = NULL,
  version = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateDriver$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateDriver$fromJSON(UpdateDriverJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateDriver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
