% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_apiWrapper.R
\name{UpdateProject}
\alias{UpdateProject}
\title{UpdateProject}
\usage{
UpdateProject(
  project,
  newProjectName = NULL,
  workerCount = NULL,
  holdoutUnlocked = NULL,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Update project
}
\details{
Change project name, worker count, or unlock the holdout.   If any of the optional json arguments are not provided, that aspect of the project will not be altered.

This method invokes \verb{PATCH /projects/\{projectId\}/} in the DataRobot Public API.
}
\seealso{
The method Patch in \link{ProjectsApi}, which invokes the same underlying API endpoint.
}
