% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_apiWrapper.R
\name{UploadComplianceDocTemplate}
\alias{UploadComplianceDocTemplate}
\title{UploadComplianceDocTemplate}
\usage{
UploadComplianceDocTemplate(name, filename = NULL, sections = NULL, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Create a new compliance documentation template
}
\details{
Create a new compliance documentation template. One can retrieve the default DataRobot template via \verb{GET /api/v2/complianceDocTemplates/default/} endpoint.

This method invokes \verb{POST /complianceDocTemplates/} in the DataRobot Public API.
}
\seealso{
The method ComplianceDocTemplatesCreate in \link{DocumentationApi}, which invokes the same underlying API endpoint.
}
