% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{UploadPredictionDataset}
\alias{UploadPredictionDataset}
\title{UploadPredictionDataset}
\usage{
UploadPredictionDataset(
  project,
  dataSource,
  forecastPoint = NULL,
  predictionsStartDate = NULL,
  predictionsEndDate = NULL,
  relaxKIAFeaturesCheck = NULL,
  maxWait = 600,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Upload a file for predictions from a URL.
}
\details{
Upload a file for predictions from a URL.

This method invokes \verb{POST /projects/\{projectId\}/predictionDatasets/urlUploads/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPredictionDatasetsUrlUploadsCreate in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
