% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_apiWrapper.R
\name{UploadPredictionDatasetFromDataSource}
\alias{UploadPredictionDatasetFromDataSource}
\title{UploadPredictionDatasetFromDataSource}
\usage{
UploadPredictionDatasetFromDataSource(
  project,
  dataSourceId,
  username,
  password,
  forecastPoint = NULL,
  maxWait = 600,
  relaxKIAFeaturesCheck = NULL,
  ...
)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Upload a dataset for predictions from a \code{DataSource}.
}
\details{
Upload a dataset for predictions from a \code{DataSource}.

This method invokes \verb{POST /projects/\{projectId\}/predictionDatasets/dataSourceUploads/} in the DataRobot Public API.
}
\seealso{
The method ProjectsPredictionDatasetsDataSourceUploadsCreate in \link{PredictionsApi}, which invokes the same underlying API endpoint.
}
