% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage_data_export.R
\docType{class}
\name{UsageDataExport}
\alias{UsageDataExport}
\title{UsageDataExport}
\format{
An \code{R6Class} generator object
}
\description{
UsageDataExport Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{end}}{character \link{optional} The upper bound of stored events timestamp to include within the artifact.}

\item{\code{include}}{list( character ) \link{optional} Additional fields to be included.}

\item{\code{includeIdentifyingFields}}{character Indicates if identifying information like user names, project names, etc. should be included or not. Defaults to True.}

\item{\code{includeReport}}{list( character ) \link{optional} The list of reports that should be generated. Will default to `None` if not specified.}

\item{\code{noCache}}{character \link{optional} Switches off caching.}

\item{\code{projectId}}{character \link{optional} Only actions that are connected with the project will be retrieved.}

\item{\code{start}}{character \link{optional} The lower bound of stored events timestamp to include within the artifact.}

\item{\code{userId}}{character \link{optional} Only actions performed by this user will be retrieved.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UsageDataExport-new}{\code{UsageDataExport$new()}}
\item \href{#method-UsageDataExport-validate}{\code{UsageDataExport$validate()}}
\item \href{#method-UsageDataExport-toJSON}{\code{UsageDataExport$toJSON()}}
\item \href{#method-UsageDataExport-fromJSON}{\code{UsageDataExport$fromJSON()}}
\item \href{#method-UsageDataExport-clone}{\code{UsageDataExport$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageDataExport-new"></a>}}
\if{latex}{\out{\hypertarget{method-UsageDataExport-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageDataExport$new(
  includeIdentifyingFields = NULL,
  end = NULL,
  include = NULL,
  includeReport = NULL,
  noCache = NULL,
  projectId = NULL,
  start = NULL,
  userId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{includeIdentifyingFields}}{Indicates if identifying information like user names, project names, etc. should be included or not. Defaults to True.}

\item{\code{end}}{The upper bound of stored events timestamp to include within the artifact.}

\item{\code{include}}{Additional fields to be included.}

\item{\code{includeReport}}{The list of reports that should be generated. Will default to `None` if not specified.}

\item{\code{noCache}}{Switches off caching.}

\item{\code{projectId}}{Only actions that are connected with the project will be retrieved.}

\item{\code{start}}{The lower bound of stored events timestamp to include within the artifact.}

\item{\code{userId}}{Only actions performed by this user will be retrieved.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageDataExport-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UsageDataExport-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageDataExport$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageDataExport-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsageDataExport-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageDataExport$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageDataExport-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsageDataExport-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageDataExport$fromJSON(UsageDataExportJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UsageDataExportJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageDataExport-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UsageDataExport-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageDataExport$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
