% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case.R
\docType{class}
\name{UseCase}
\alias{UseCase}
\title{UseCase}
\format{
An \code{R6Class} generator object
}
\description{
UseCase Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{businessImpact}}{integer \link{optional} Expected effects on overall business operations}

\item{\code{commentId}}{character \link{optional} The id for this comment}

\item{\code{content}}{character \link{optional} A string}

\item{\code{description}}{character \link{optional} Use case description}

\item{\code{feasibility}}{integer \link{optional} Assessment of how the use case can be accomplished across multiple dimensions}

\item{\code{id}}{character \link{optional} The id of this description item}

\item{\code{mentions}}{list( \link{DrUserId} ) \link{optional} A list of user objects}

\item{\code{name}}{character Name of the use case}

\item{\code{notes}}{character \link{optional} User notes}

\item{\code{owner}}{\link{DrUserId} \link{optional}}

\item{\code{permissions}}{list( character ) \link{optional} Permissions of current user}

\item{\code{potentialValue}}{\link{MonetaryValue} \link{optional}}

\item{\code{potentialValueTemplate}}{\link{AnyOfUseCaseValueTemplateClassificationUseCaseValueTemplateRegression} \link{optional} Optional, Contains Template type and parameter information}

\item{\code{predictionTargets}}{list( character ) \link{optional} An array of strings prediction target names}

\item{\code{realizedValue}}{\link{AnyOfstringMonetaryValue} \link{optional} Optional, Contains MonetaryValue objects}

\item{\code{stage}}{character \link{optional} Current stage of use case}

\item{\code{targetDates}}{list( \link{TargetDate} ) \link{optional} Array of TargetDate objects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UseCase-new}{\code{UseCase$new()}}
\item \href{#method-UseCase-validate}{\code{UseCase$validate()}}
\item \href{#method-UseCase-toJSON}{\code{UseCase$toJSON()}}
\item \href{#method-UseCase-fromJSON}{\code{UseCase$fromJSON()}}
\item \href{#method-UseCase-clone}{\code{UseCase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCase-new"></a>}}
\if{latex}{\out{\hypertarget{method-UseCase-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCase$new(
  name = NULL,
  businessImpact = NULL,
  commentId = NULL,
  content = NULL,
  description = NULL,
  feasibility = NULL,
  id = NULL,
  mentions = NULL,
  notes = NULL,
  owner = NULL,
  permissions = NULL,
  potentialValue = NULL,
  potentialValueTemplate = NULL,
  predictionTargets = NULL,
  realizedValue = NULL,
  stage = NULL,
  targetDates = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the use case}

\item{\code{businessImpact}}{Expected effects on overall business operations}

\item{\code{commentId}}{The id for this comment}

\item{\code{content}}{A string}

\item{\code{description}}{Use case description}

\item{\code{feasibility}}{Assessment of how the use case can be accomplished across multiple dimensions}

\item{\code{id}}{The id of this description item}

\item{\code{mentions}}{A list of user objects}

\item{\code{notes}}{User notes}

\item{\code{permissions}}{Permissions of current user}

\item{\code{potentialValueTemplate}}{Optional, Contains Template type and parameter information}

\item{\code{predictionTargets}}{An array of strings prediction target names}

\item{\code{realizedValue}}{Optional, Contains MonetaryValue objects}

\item{\code{stage}}{Current stage of use case}

\item{\code{targetDates}}{Array of TargetDate objects}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCase-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UseCase-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCase$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCase-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UseCase-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCase$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCase-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UseCase-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCase$fromJSON(UseCaseJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UseCaseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UseCase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
