% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_activity_list_response.R
\docType{class}
\name{UseCaseActivityListResponse}
\alias{UseCaseActivityListResponse}
\title{UseCaseActivityListResponse}
\format{
An \code{R6Class} generator object
}
\description{
UseCaseActivityListResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer The number of items returned on this page}

\item{\code{data}}{list( \link{UseCaseActivity} ) The activities of a use case}

\item{\code{next_}}{character URL pointing to the next page (if null, there is no next page)}

\item{\code{previous}}{character URL pointing to the previous page (if null, there is no previous page)}

\item{\code{totalCount}}{integer The number of items matching the query condition}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UseCaseActivityListResponse-new}{\code{UseCaseActivityListResponse$new()}}
\item \href{#method-UseCaseActivityListResponse-validate}{\code{UseCaseActivityListResponse$validate()}}
\item \href{#method-UseCaseActivityListResponse-toJSON}{\code{UseCaseActivityListResponse$toJSON()}}
\item \href{#method-UseCaseActivityListResponse-fromJSON}{\code{UseCaseActivityListResponse$fromJSON()}}
\item \href{#method-UseCaseActivityListResponse-clone}{\code{UseCaseActivityListResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseActivityListResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseActivityListResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseActivityListResponse$new(
  count = NULL,
  data = NULL,
  next_ = NULL,
  previous = NULL,
  totalCount = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{The number of items returned on this page}

\item{\code{data}}{The activities of a use case}

\item{\code{next_}}{URL pointing to the next page (if null, there is no next page)}

\item{\code{previous}}{URL pointing to the previous page (if null, there is no previous page)}

\item{\code{totalCount}}{The number of items matching the query condition}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseActivityListResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseActivityListResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseActivityListResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseActivityListResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseActivityListResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseActivityListResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseActivityListResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseActivityListResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseActivityListResponse$fromJSON(
  UseCaseActivityListResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UseCaseActivityListResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseActivityListResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseActivityListResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseActivityListResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
