% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_attachment_custom_model_overview.R
\docType{class}
\name{UseCaseAttachmentCustomModelOverview}
\alias{UseCaseAttachmentCustomModelOverview}
\title{UseCaseAttachmentCustomModelOverview}
\format{
An \code{R6Class} generator object
}
\description{
UseCaseAttachmentCustomModelOverview Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{createdAt}}{character Time when custom model was created}

\item{\code{modifiedAt}}{character Time of the last modification}

\item{\code{name}}{character Name of the custom model}

\item{\code{owner}}{character Name of the custom model owner}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UseCaseAttachmentCustomModelOverview-new}{\code{UseCaseAttachmentCustomModelOverview$new()}}
\item \href{#method-UseCaseAttachmentCustomModelOverview-validate}{\code{UseCaseAttachmentCustomModelOverview$validate()}}
\item \href{#method-UseCaseAttachmentCustomModelOverview-toJSON}{\code{UseCaseAttachmentCustomModelOverview$toJSON()}}
\item \href{#method-UseCaseAttachmentCustomModelOverview-fromJSON}{\code{UseCaseAttachmentCustomModelOverview$fromJSON()}}
\item \href{#method-UseCaseAttachmentCustomModelOverview-clone}{\code{UseCaseAttachmentCustomModelOverview$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseAttachmentCustomModelOverview-new"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseAttachmentCustomModelOverview-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentCustomModelOverview$new(
  createdAt = NULL,
  modifiedAt = NULL,
  name = NULL,
  owner = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{createdAt}}{Time when custom model was created}

\item{\code{modifiedAt}}{Time of the last modification}

\item{\code{name}}{Name of the custom model}

\item{\code{owner}}{Name of the custom model owner}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseAttachmentCustomModelOverview-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseAttachmentCustomModelOverview-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentCustomModelOverview$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseAttachmentCustomModelOverview-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseAttachmentCustomModelOverview-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentCustomModelOverview$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseAttachmentCustomModelOverview-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseAttachmentCustomModelOverview-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentCustomModelOverview$fromJSON(
  UseCaseAttachmentCustomModelOverviewJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UseCaseAttachmentCustomModelOverviewJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseAttachmentCustomModelOverview-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseAttachmentCustomModelOverview-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentCustomModelOverview$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
