% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_attachment_dataset_overview.R
\docType{class}
\name{UseCaseAttachmentDatasetOverview}
\alias{UseCaseAttachmentDatasetOverview}
\title{UseCaseAttachmentDatasetOverview}
\format{
An \code{R6Class} generator object
}
\description{
UseCaseAttachmentDatasetOverview Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{createdAt}}{character Time when dataset was created}

\item{\code{createdBy}}{character Name of user who created the dataset}

\item{\code{lastModifiedAt}}{character Time of the last modification}

\item{\code{lastModifiedBy}}{character Name of user who did last modification}

\item{\code{name}}{character Name of the dataset}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UseCaseAttachmentDatasetOverview$new()}}
\item \href{#method-validate}{\code{UseCaseAttachmentDatasetOverview$validate()}}
\item \href{#method-toJSON}{\code{UseCaseAttachmentDatasetOverview$toJSON()}}
\item \href{#method-fromJSON}{\code{UseCaseAttachmentDatasetOverview$fromJSON()}}
\item \href{#method-clone}{\code{UseCaseAttachmentDatasetOverview$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentDatasetOverview$new(
  createdAt = NULL,
  createdBy = NULL,
  lastModifiedAt = NULL,
  lastModifiedBy = NULL,
  name = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{createdAt}}{Time when dataset was created}

\item{\code{createdBy}}{Name of user who created the dataset}

\item{\code{lastModifiedAt}}{Time of the last modification}

\item{\code{lastModifiedBy}}{Name of user who did last modification}

\item{\code{name}}{Name of the dataset}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentDatasetOverview$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentDatasetOverview$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentDatasetOverview$fromJSON(
  UseCaseAttachmentDatasetOverviewJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UseCaseAttachmentDatasetOverviewJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentDatasetOverview$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
