% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_attachment_list_response.R
\docType{class}
\name{UseCaseAttachmentListResponse}
\alias{UseCaseAttachmentListResponse}
\title{UseCaseAttachmentListResponse}
\format{
An \code{R6Class} generator object
}
\description{
UseCaseAttachmentListResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer Number of results returned}

\item{\code{data}}{list( \link{UseCaseAttachmentResponse} ) An array of attachment objects}

\item{\code{next_}}{character Link to next page}

\item{\code{previous}}{character Link to previous page}

\item{\code{totalCount}}{integer Total number of results available}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UseCaseAttachmentListResponse-new}{\code{UseCaseAttachmentListResponse$new()}}
\item \href{#method-UseCaseAttachmentListResponse-validate}{\code{UseCaseAttachmentListResponse$validate()}}
\item \href{#method-UseCaseAttachmentListResponse-toJSON}{\code{UseCaseAttachmentListResponse$toJSON()}}
\item \href{#method-UseCaseAttachmentListResponse-fromJSON}{\code{UseCaseAttachmentListResponse$fromJSON()}}
\item \href{#method-UseCaseAttachmentListResponse-clone}{\code{UseCaseAttachmentListResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseAttachmentListResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseAttachmentListResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentListResponse$new(
  count = NULL,
  data = NULL,
  next_ = NULL,
  previous = NULL,
  totalCount = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{Number of results returned}

\item{\code{data}}{An array of attachment objects}

\item{\code{next_}}{Link to next page}

\item{\code{previous}}{Link to previous page}

\item{\code{totalCount}}{Total number of results available}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseAttachmentListResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseAttachmentListResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentListResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseAttachmentListResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseAttachmentListResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentListResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseAttachmentListResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseAttachmentListResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentListResponse$fromJSON(
  UseCaseAttachmentListResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UseCaseAttachmentListResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseAttachmentListResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseAttachmentListResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentListResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
