% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_attachment_modeling_project_overview.R
\docType{class}
\name{UseCaseAttachmentModelingProjectOverview}
\alias{UseCaseAttachmentModelingProjectOverview}
\title{UseCaseAttachmentModelingProjectOverview}
\format{
An \code{R6Class} generator object
}
\description{
UseCaseAttachmentModelingProjectOverview Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{createdAt}}{character Time when project was created}

\item{\code{datasetName}}{character Name of default dataset}

\item{\code{name}}{character Name of the project}

\item{\code{numberOfModels}}{integer Number of the models in the project}

\item{\code{owner}}{character Name of the modeling project owner}

\item{\code{targetName}}{character Name of modeling target}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UseCaseAttachmentModelingProjectOverview$new()}}
\item \href{#method-validate}{\code{UseCaseAttachmentModelingProjectOverview$validate()}}
\item \href{#method-toJSON}{\code{UseCaseAttachmentModelingProjectOverview$toJSON()}}
\item \href{#method-fromJSON}{\code{UseCaseAttachmentModelingProjectOverview$fromJSON()}}
\item \href{#method-clone}{\code{UseCaseAttachmentModelingProjectOverview$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentModelingProjectOverview$new(
  createdAt = NULL,
  datasetName = NULL,
  name = NULL,
  numberOfModels = NULL,
  owner = NULL,
  targetName = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{createdAt}}{Time when project was created}

\item{\code{datasetName}}{Name of default dataset}

\item{\code{name}}{Name of the project}

\item{\code{numberOfModels}}{Number of the models in the project}

\item{\code{owner}}{Name of the modeling project owner}

\item{\code{targetName}}{Name of modeling target}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentModelingProjectOverview$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentModelingProjectOverview$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentModelingProjectOverview$fromJSON(
  UseCaseAttachmentModelingProjectOverviewJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UseCaseAttachmentModelingProjectOverviewJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentModelingProjectOverview$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
