% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_attachment_response.R
\docType{class}
\name{UseCaseAttachmentResponse}
\alias{UseCaseAttachmentResponse}
\title{UseCaseAttachmentResponse}
\format{
An \code{R6Class} generator object
}
\description{
UseCaseAttachmentResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{attachedObjectId}}{character The id of the attached object}

\item{\code{canView}}{character Indicates if user has view access to the attached object}

\item{\code{id}}{character The id of the use case}

\item{\code{isAccessRequested}}{character Indicates if user has requested access to the attached object}

\item{\code{overview}}{\link{OneOfUseCaseAttachmentDeploymentOverviewUseCaseAttachmentDatasetOverviewUseCaseAttachmentModelingProjectOverviewUseCaseAttachmentCustomModelOverviewUseCaseAttachmentModelPackageOverviewUseCaseAttachmentApplicationOverview} Detailed information depending on the type of the attachment}

\item{\code{type}}{character Use case attachment type}

\item{\code{useCaseId}}{character The id of the use case}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UseCaseAttachmentResponse-new}{\code{UseCaseAttachmentResponse$new()}}
\item \href{#method-UseCaseAttachmentResponse-validate}{\code{UseCaseAttachmentResponse$validate()}}
\item \href{#method-UseCaseAttachmentResponse-toJSON}{\code{UseCaseAttachmentResponse$toJSON()}}
\item \href{#method-UseCaseAttachmentResponse-fromJSON}{\code{UseCaseAttachmentResponse$fromJSON()}}
\item \href{#method-UseCaseAttachmentResponse-clone}{\code{UseCaseAttachmentResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseAttachmentResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseAttachmentResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentResponse$new(
  attachedObjectId = NULL,
  canView = NULL,
  id = NULL,
  isAccessRequested = NULL,
  overview = NULL,
  type = NULL,
  useCaseId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attachedObjectId}}{The id of the attached object}

\item{\code{canView}}{Indicates if user has view access to the attached object}

\item{\code{id}}{The id of the use case}

\item{\code{isAccessRequested}}{Indicates if user has requested access to the attached object}

\item{\code{overview}}{Detailed information depending on the type of the attachment}

\item{\code{type}}{Use case attachment type}

\item{\code{useCaseId}}{The id of the use case}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseAttachmentResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseAttachmentResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseAttachmentResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseAttachmentResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseAttachmentResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseAttachmentResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentResponse$fromJSON(
  UseCaseAttachmentResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UseCaseAttachmentResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseAttachmentResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseAttachmentResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
