% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_bucket_period.R
\docType{class}
\name{UseCaseBucketPeriod}
\alias{UseCaseBucketPeriod}
\title{UseCaseBucketPeriod}
\format{
An \code{R6Class} generator object
}
\description{
UseCaseBucketPeriod Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{end}}{character RFC3339 datetime. End of time period to retrieve.}

\item{\code{start}}{character RFC3339 datetime. Start of time period to retrieve.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UseCaseBucketPeriod-new}{\code{UseCaseBucketPeriod$new()}}
\item \href{#method-UseCaseBucketPeriod-validate}{\code{UseCaseBucketPeriod$validate()}}
\item \href{#method-UseCaseBucketPeriod-toJSON}{\code{UseCaseBucketPeriod$toJSON()}}
\item \href{#method-UseCaseBucketPeriod-fromJSON}{\code{UseCaseBucketPeriod$fromJSON()}}
\item \href{#method-UseCaseBucketPeriod-clone}{\code{UseCaseBucketPeriod$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseBucketPeriod-new"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseBucketPeriod-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseBucketPeriod$new(end = NULL, start = NULL, validateParams = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{end}}{RFC3339 datetime. End of time period to retrieve.}

\item{\code{start}}{RFC3339 datetime. Start of time period to retrieve.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseBucketPeriod-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseBucketPeriod-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseBucketPeriod$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseBucketPeriod-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseBucketPeriod-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseBucketPeriod$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseBucketPeriod-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseBucketPeriod-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseBucketPeriod$fromJSON(UseCaseBucketPeriodJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UseCaseBucketPeriodJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseBucketPeriod-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseBucketPeriod-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseBucketPeriod$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
