% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_summary.R
\docType{class}
\name{UseCaseSummary}
\alias{UseCaseSummary}
\title{UseCaseSummary}
\format{
An \code{R6Class} generator object
}
\description{
UseCaseSummary Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{averagePredictionsCount}}{integer Average number of predictions per bucket during this time period}

\item{\code{averageRealizedValue}}{integer \link{optional} Average amount of value per bucket realised during this time period}

\item{\code{period}}{\link{UseCaseBucketPeriod}}

\item{\code{totalPredictionsCount}}{integer \link{optional} Total number of predictions that happened during this time period}

\item{\code{totalRealizedValue}}{integer Total amount of value realised during this time period}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UseCaseSummary$new()}}
\item \href{#method-validate}{\code{UseCaseSummary$validate()}}
\item \href{#method-toJSON}{\code{UseCaseSummary$toJSON()}}
\item \href{#method-fromJSON}{\code{UseCaseSummary$fromJSON()}}
\item \href{#method-clone}{\code{UseCaseSummary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseSummary$new(
  averagePredictionsCount = NULL,
  period = NULL,
  totalRealizedValue = NULL,
  averageRealizedValue = NULL,
  totalPredictionsCount = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{averagePredictionsCount}}{Average number of predictions per bucket during this time period}

\item{\code{totalRealizedValue}}{Total amount of value realised during this time period}

\item{\code{averageRealizedValue}}{Average amount of value per bucket realised during this time period}

\item{\code{totalPredictionsCount}}{Total number of predictions that happened during this time period}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseSummary$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseSummary$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseSummary$fromJSON(UseCaseSummaryJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UseCaseSummaryJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseSummary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
