% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_value_template_schema.R
\docType{class}
\name{UseCaseValueTemplateSchema}
\alias{UseCaseValueTemplateSchema}
\title{UseCaseValueTemplateSchema}
\format{
An \code{R6Class} generator object
}
\description{
UseCaseValueTemplateSchema Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{elementType}}{character Possible types used for template variables}

\item{\code{label}}{character Label describing the represented value}

\item{\code{maximum}}{integer Sets Maximum value if given}

\item{\code{minimum}}{integer Sets Minimum value if given}

\item{\code{placeholder}}{character Text that can be used to prefill an input filed for the value}

\item{\code{unit}}{character Unit type. e.g. \%}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UseCaseValueTemplateSchema-new}{\code{UseCaseValueTemplateSchema$new()}}
\item \href{#method-UseCaseValueTemplateSchema-validate}{\code{UseCaseValueTemplateSchema$validate()}}
\item \href{#method-UseCaseValueTemplateSchema-toJSON}{\code{UseCaseValueTemplateSchema$toJSON()}}
\item \href{#method-UseCaseValueTemplateSchema-fromJSON}{\code{UseCaseValueTemplateSchema$fromJSON()}}
\item \href{#method-UseCaseValueTemplateSchema-clone}{\code{UseCaseValueTemplateSchema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseValueTemplateSchema-new"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseValueTemplateSchema-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseValueTemplateSchema$new(
  elementType = NULL,
  label = NULL,
  maximum = NULL,
  minimum = NULL,
  placeholder = NULL,
  unit = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{elementType}}{Possible types used for template variables}

\item{\code{label}}{Label describing the represented value}

\item{\code{maximum}}{Sets Maximum value if given}

\item{\code{minimum}}{Sets Minimum value if given}

\item{\code{placeholder}}{Text that can be used to prefill an input filed for the value}

\item{\code{unit}}{Unit type. e.g. \%}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseValueTemplateSchema-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseValueTemplateSchema-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseValueTemplateSchema$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseValueTemplateSchema-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseValueTemplateSchema-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseValueTemplateSchema$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseValueTemplateSchema-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseValueTemplateSchema-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseValueTemplateSchema$fromJSON(
  UseCaseValueTemplateSchemaJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UseCaseValueTemplateSchemaJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseValueTemplateSchema-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseValueTemplateSchema-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseValueTemplateSchema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
