% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_blueprint_added_to_menu_item.R
\docType{class}
\name{UserBlueprintAddedToMenuItem}
\alias{UserBlueprintAddedToMenuItem}
\title{UserBlueprintAddedToMenuItem}
\format{
An \code{R6Class} generator object
}
\description{
UserBlueprintAddedToMenuItem Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{blueprintId}}{character The blueprintId representing the blueprint which was added to the project repository.}

\item{\code{userBlueprintId}}{character The userBlueprintId associated with the blueprintId added to the project repository.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserBlueprintAddedToMenuItem-new}{\code{UserBlueprintAddedToMenuItem$new()}}
\item \href{#method-UserBlueprintAddedToMenuItem-validate}{\code{UserBlueprintAddedToMenuItem$validate()}}
\item \href{#method-UserBlueprintAddedToMenuItem-toJSON}{\code{UserBlueprintAddedToMenuItem$toJSON()}}
\item \href{#method-UserBlueprintAddedToMenuItem-fromJSON}{\code{UserBlueprintAddedToMenuItem$fromJSON()}}
\item \href{#method-UserBlueprintAddedToMenuItem-clone}{\code{UserBlueprintAddedToMenuItem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintAddedToMenuItem-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintAddedToMenuItem-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintAddedToMenuItem$new(
  blueprintId = NULL,
  userBlueprintId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blueprintId}}{The blueprintId representing the blueprint which was added to the project repository.}

\item{\code{userBlueprintId}}{The userBlueprintId associated with the blueprintId added to the project repository.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintAddedToMenuItem-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintAddedToMenuItem-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintAddedToMenuItem$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintAddedToMenuItem-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintAddedToMenuItem-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintAddedToMenuItem$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintAddedToMenuItem-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintAddedToMenuItem-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintAddedToMenuItem$fromJSON(
  UserBlueprintAddedToMenuItemJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserBlueprintAddedToMenuItemJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintAddedToMenuItem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintAddedToMenuItem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintAddedToMenuItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
