% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_blueprint_add_to_menu.R
\docType{class}
\name{UserBlueprintAddToMenu}
\alias{UserBlueprintAddToMenu}
\title{UserBlueprintAddToMenu}
\format{
An \code{R6Class} generator object
}
\description{
UserBlueprintAddToMenu Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{deleteAfter}}{character Whether to delete the user blueprint(s) after adding it (them) to the project menu.}

\item{\code{describeFailures}}{character Whether to include extra fields to describe why any blueprints were not added to the chosen project.}

\item{\code{projectId}}{character The projectId of the project for the repository to add the specified user blueprints to.}

\item{\code{userBlueprintIds}}{list( character ) The ids of the user blueprints to add to the specified project's repository.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserBlueprintAddToMenu-new}{\code{UserBlueprintAddToMenu$new()}}
\item \href{#method-UserBlueprintAddToMenu-validate}{\code{UserBlueprintAddToMenu$validate()}}
\item \href{#method-UserBlueprintAddToMenu-toJSON}{\code{UserBlueprintAddToMenu$toJSON()}}
\item \href{#method-UserBlueprintAddToMenu-fromJSON}{\code{UserBlueprintAddToMenu$fromJSON()}}
\item \href{#method-UserBlueprintAddToMenu-clone}{\code{UserBlueprintAddToMenu$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintAddToMenu-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintAddToMenu-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintAddToMenu$new(
  deleteAfter = NULL,
  describeFailures = NULL,
  projectId = NULL,
  userBlueprintIds = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deleteAfter}}{Whether to delete the user blueprint(s) after adding it (them) to the project menu.}

\item{\code{describeFailures}}{Whether to include extra fields to describe why any blueprints were not added to the chosen project.}

\item{\code{projectId}}{The projectId of the project for the repository to add the specified user blueprints to.}

\item{\code{userBlueprintIds}}{The ids of the user blueprints to add to the specified project's repository.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintAddToMenu-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintAddToMenu-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintAddToMenu$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintAddToMenu-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintAddToMenu-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintAddToMenu$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintAddToMenu-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintAddToMenu-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintAddToMenu$fromJSON(
  UserBlueprintAddToMenuJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserBlueprintAddToMenuJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintAddToMenu-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintAddToMenu-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintAddToMenu$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
