% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_blueprint_add_to_menu_response.R
\docType{class}
\name{UserBlueprintAddToMenuResponse}
\alias{UserBlueprintAddToMenuResponse}
\title{UserBlueprintAddToMenuResponse}
\format{
An \code{R6Class} generator object
}
\description{
UserBlueprintAddToMenuResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{addedToMenu}}{list( \link{UserBlueprintAddedToMenuItem} ) The list of userBlueprintId and blueprintId pairs representing blueprints successfully added to the project repository.}

\item{\code{message}}{character \link{optional} A success message or a list of reasons why the list of blueprints could not be added to the project repository.}

\item{\code{notAddedToMenu}}{list( \link{UserBlueprintFailedToAddToMenuItem} ) \link{optional} The list of userBlueprintId and error message representing blueprints which failed to be added to the project repository.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UserBlueprintAddToMenuResponse$new()}}
\item \href{#method-validate}{\code{UserBlueprintAddToMenuResponse$validate()}}
\item \href{#method-toJSON}{\code{UserBlueprintAddToMenuResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{UserBlueprintAddToMenuResponse$fromJSON()}}
\item \href{#method-clone}{\code{UserBlueprintAddToMenuResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintAddToMenuResponse$new(
  addedToMenu = NULL,
  message = NULL,
  notAddedToMenu = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{addedToMenu}}{The list of userBlueprintId and blueprintId pairs representing blueprints successfully added to the project repository.}

\item{\code{message}}{A success message or a list of reasons why the list of blueprints could not be added to the project repository.}

\item{\code{notAddedToMenu}}{The list of userBlueprintId and error message representing blueprints which failed to be added to the project repository.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintAddToMenuResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintAddToMenuResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintAddToMenuResponse$fromJSON(
  UserBlueprintAddToMenuResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserBlueprintAddToMenuResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintAddToMenuResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
