% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/user_blueprint_create_from_custom_task_version_id_payload.R
\docType{class}
\name{UserBlueprintCreateFromCustomTaskVersionIdPayload}
\alias{UserBlueprintCreateFromCustomTaskVersionIdPayload}
\title{UserBlueprintCreateFromCustomTaskVersionIdPayload}
\format{
An \code{R6Class} generator object
}
\description{
UserBlueprintCreateFromCustomTaskVersionIdPayload Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{customTaskVersionId}}{character The ID of a custom task version.}

\item{\code{decompressedBlueprint}}{character Whether to retrieve the blueprint in the decompressed format.}

\item{\code{description}}{character \link{optional} The description for the user blueprint that will be created from this CustomTaskVersion.}

\item{\code{saveToCatalog}}{character Whether to save the blueprint to the catalog.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserBlueprintCreateFromCustomTaskVersionIdPayload-new}{\code{UserBlueprintCreateFromCustomTaskVersionIdPayload$new()}}
\item \href{#method-UserBlueprintCreateFromCustomTaskVersionIdPayload-validate}{\code{UserBlueprintCreateFromCustomTaskVersionIdPayload$validate()}}
\item \href{#method-UserBlueprintCreateFromCustomTaskVersionIdPayload-toJSON}{\code{UserBlueprintCreateFromCustomTaskVersionIdPayload$toJSON()}}
\item \href{#method-UserBlueprintCreateFromCustomTaskVersionIdPayload-fromJSON}{\code{UserBlueprintCreateFromCustomTaskVersionIdPayload$fromJSON()}}
\item \href{#method-UserBlueprintCreateFromCustomTaskVersionIdPayload-clone}{\code{UserBlueprintCreateFromCustomTaskVersionIdPayload$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintCreateFromCustomTaskVersionIdPayload-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintCreateFromCustomTaskVersionIdPayload-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintCreateFromCustomTaskVersionIdPayload$new(
  customTaskVersionId = NULL,
  decompressedBlueprint = NULL,
  saveToCatalog = NULL,
  description = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{customTaskVersionId}}{The ID of a custom task version.}

\item{\code{decompressedBlueprint}}{Whether to retrieve the blueprint in the decompressed format.}

\item{\code{saveToCatalog}}{Whether to save the blueprint to the catalog.}

\item{\code{description}}{The description for the user blueprint that will be created from this CustomTaskVersion.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintCreateFromCustomTaskVersionIdPayload-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintCreateFromCustomTaskVersionIdPayload-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintCreateFromCustomTaskVersionIdPayload$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintCreateFromCustomTaskVersionIdPayload-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintCreateFromCustomTaskVersionIdPayload-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintCreateFromCustomTaskVersionIdPayload$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintCreateFromCustomTaskVersionIdPayload-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintCreateFromCustomTaskVersionIdPayload-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintCreateFromCustomTaskVersionIdPayload$fromJSON(
  UserBlueprintCreateFromCustomTaskVersionIdPayloadJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserBlueprintCreateFromCustomTaskVersionIdPayloadJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintCreateFromCustomTaskVersionIdPayload-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintCreateFromCustomTaskVersionIdPayload-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintCreateFromCustomTaskVersionIdPayload$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
