% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_blueprint_create_from_user_blueprint_id.R
\docType{class}
\name{UserBlueprintCreateFromUserBlueprintId}
\alias{UserBlueprintCreateFromUserBlueprintId}
\title{UserBlueprintCreateFromUserBlueprintId}
\format{
An \code{R6Class} generator object
}
\description{
UserBlueprintCreateFromUserBlueprintId Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{decompressedBlueprint}}{character Whether to retrieve the blueprint in the decompressed format.}

\item{\code{description}}{character \link{optional} The description to give to the blueprint.}

\item{\code{getDynamicLabels}}{character \link{optional} Whether to add dynamic labels to a decompressed blueprint.}

\item{\code{isInplaceEditor}}{character Whether the request is sent from the in place user BP editor.}

\item{\code{modelType}}{character \link{optional} The title to give to the blueprint.}

\item{\code{projectId}}{character \link{optional} String representation of ObjectId for the currently active project. The user blueprint is created when this project is active. Necessary in the event of project specific tasks, such as column selection tasks.}

\item{\code{saveToCatalog}}{character Whether to save the blueprint to the catalog.}

\item{\code{userBlueprintId}}{character The id of the existing user blueprint to copy.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserBlueprintCreateFromUserBlueprintId-new}{\code{UserBlueprintCreateFromUserBlueprintId$new()}}
\item \href{#method-UserBlueprintCreateFromUserBlueprintId-validate}{\code{UserBlueprintCreateFromUserBlueprintId$validate()}}
\item \href{#method-UserBlueprintCreateFromUserBlueprintId-toJSON}{\code{UserBlueprintCreateFromUserBlueprintId$toJSON()}}
\item \href{#method-UserBlueprintCreateFromUserBlueprintId-fromJSON}{\code{UserBlueprintCreateFromUserBlueprintId$fromJSON()}}
\item \href{#method-UserBlueprintCreateFromUserBlueprintId-clone}{\code{UserBlueprintCreateFromUserBlueprintId$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintCreateFromUserBlueprintId-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintCreateFromUserBlueprintId-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintCreateFromUserBlueprintId$new(
  decompressedBlueprint = NULL,
  isInplaceEditor = NULL,
  saveToCatalog = NULL,
  userBlueprintId = NULL,
  description = NULL,
  getDynamicLabels = FALSE,
  modelType = NULL,
  projectId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{decompressedBlueprint}}{Whether to retrieve the blueprint in the decompressed format.}

\item{\code{isInplaceEditor}}{Whether the request is sent from the in place user BP editor.}

\item{\code{saveToCatalog}}{Whether to save the blueprint to the catalog.}

\item{\code{userBlueprintId}}{The id of the existing user blueprint to copy.}

\item{\code{description}}{The description to give to the blueprint.}

\item{\code{getDynamicLabels}}{Whether to add dynamic labels to a decompressed blueprint.}

\item{\code{modelType}}{The title to give to the blueprint.}

\item{\code{projectId}}{String representation of ObjectId for the currently active project. The user blueprint is created when this project is active. Necessary in the event of project specific tasks, such as column selection tasks.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintCreateFromUserBlueprintId-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintCreateFromUserBlueprintId-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintCreateFromUserBlueprintId$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintCreateFromUserBlueprintId-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintCreateFromUserBlueprintId-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintCreateFromUserBlueprintId$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintCreateFromUserBlueprintId-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintCreateFromUserBlueprintId-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintCreateFromUserBlueprintId$fromJSON(
  UserBlueprintCreateFromUserBlueprintIdJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserBlueprintCreateFromUserBlueprintIdJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintCreateFromUserBlueprintId-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintCreateFromUserBlueprintId-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintCreateFromUserBlueprintId$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
