% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_blueprints_blueprint_task_data.R
\docType{class}
\name{UserBlueprintsBlueprintTaskData}
\alias{UserBlueprintsBlueprintTaskData}
\title{UserBlueprintsBlueprintTaskData}
\format{
An \code{R6Class} generator object
}
\description{
UserBlueprintsBlueprintTaskData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{inputs}}{list( character ) The ids or input data types which will be inputs to the task.}

\item{\code{outputMethod}}{character The method which the task will use to produce output.}

\item{\code{outputMethodParameters}}{list( \link{ParamValuePair} ) The parameters which further define how output will be produced.}

\item{\code{taskCode}}{character The unique code representing the python class which will be instantiated and executed.}

\item{\code{taskParameters}}{list( \link{ParamValuePair} ) The parameters which further define the behavior of the task.}

\item{\code{xTransformations}}{list( \link{ParamValuePair} ) Transformations to apply to the input data before fitting or predicting.}

\item{\code{yTransformations}}{list( \link{ParamValuePair} ) Transformations to apply to the input target before fitting or predicting.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserBlueprintsBlueprintTaskData-new}{\code{UserBlueprintsBlueprintTaskData$new()}}
\item \href{#method-UserBlueprintsBlueprintTaskData-validate}{\code{UserBlueprintsBlueprintTaskData$validate()}}
\item \href{#method-UserBlueprintsBlueprintTaskData-toJSON}{\code{UserBlueprintsBlueprintTaskData$toJSON()}}
\item \href{#method-UserBlueprintsBlueprintTaskData-fromJSON}{\code{UserBlueprintsBlueprintTaskData$fromJSON()}}
\item \href{#method-UserBlueprintsBlueprintTaskData-clone}{\code{UserBlueprintsBlueprintTaskData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintsBlueprintTaskData-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintsBlueprintTaskData-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintsBlueprintTaskData$new(
  inputs = NULL,
  outputMethod = NULL,
  outputMethodParameters = NULL,
  taskCode = NULL,
  taskParameters = NULL,
  xTransformations = NULL,
  yTransformations = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputs}}{The ids or input data types which will be inputs to the task.}

\item{\code{outputMethod}}{The method which the task will use to produce output.}

\item{\code{outputMethodParameters}}{The parameters which further define how output will be produced.}

\item{\code{taskCode}}{The unique code representing the python class which will be instantiated and executed.}

\item{\code{taskParameters}}{The parameters which further define the behavior of the task.}

\item{\code{xTransformations}}{Transformations to apply to the input data before fitting or predicting.}

\item{\code{yTransformations}}{Transformations to apply to the input target before fitting or predicting.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintsBlueprintTaskData-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintsBlueprintTaskData-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintsBlueprintTaskData$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintsBlueprintTaskData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintsBlueprintTaskData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintsBlueprintTaskData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintsBlueprintTaskData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintsBlueprintTaskData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintsBlueprintTaskData$fromJSON(
  UserBlueprintsBlueprintTaskDataJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserBlueprintsBlueprintTaskDataJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintsBlueprintTaskData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintsBlueprintTaskData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintsBlueprintTaskData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
