% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_blueprint_shared_roles_response.R
\docType{class}
\name{UserBlueprintSharedRolesResponse}
\alias{UserBlueprintSharedRolesResponse}
\title{UserBlueprintSharedRolesResponse}
\format{
An \code{R6Class} generator object
}
\description{
UserBlueprintSharedRolesResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character The ID of the recipient organization, group or user.}

\item{\code{name}}{character The name of the recipient organization, group or user.}

\item{\code{role}}{character The role of the org/group/user on this dataset or \"NO_ROLE\" for removing access when used with route to modify access.}

\item{\code{shareRecipientType}}{character Describes the recipient type, either user, group, or organization.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UserBlueprintSharedRolesResponse$new()}}
\item \href{#method-validate}{\code{UserBlueprintSharedRolesResponse$validate()}}
\item \href{#method-toJSON}{\code{UserBlueprintSharedRolesResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{UserBlueprintSharedRolesResponse$fromJSON()}}
\item \href{#method-clone}{\code{UserBlueprintSharedRolesResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintSharedRolesResponse$new(
  id = NULL,
  name = NULL,
  role = NULL,
  shareRecipientType = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The ID of the recipient organization, group or user.}

\item{\code{name}}{The name of the recipient organization, group or user.}

\item{\code{role}}{The role of the org/group/user on this dataset or \"NO_ROLE\" for removing access when used with route to modify access.}

\item{\code{shareRecipientType}}{Describes the recipient type, either user, group, or organization.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintSharedRolesResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintSharedRolesResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintSharedRolesResponse$fromJSON(
  UserBlueprintSharedRolesResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserBlueprintSharedRolesResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintSharedRolesResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
