% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_blueprints_hex_column_name_lookup_entry.R
\docType{class}
\name{UserBlueprintsHexColumnNameLookupEntry}
\alias{UserBlueprintsHexColumnNameLookupEntry}
\title{UserBlueprintsHexColumnNameLookupEntry}
\format{
An \code{R6Class} generator object
}
\description{
UserBlueprintsHexColumnNameLookupEntry Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{colname}}{character The name of the column.}

\item{\code{hex}}{character A safe hex representation of the column name.}

\item{\code{projectId}}{character \link{optional} The id of the project from which the column name originates.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserBlueprintsHexColumnNameLookupEntry-new}{\code{UserBlueprintsHexColumnNameLookupEntry$new()}}
\item \href{#method-UserBlueprintsHexColumnNameLookupEntry-validate}{\code{UserBlueprintsHexColumnNameLookupEntry$validate()}}
\item \href{#method-UserBlueprintsHexColumnNameLookupEntry-toJSON}{\code{UserBlueprintsHexColumnNameLookupEntry$toJSON()}}
\item \href{#method-UserBlueprintsHexColumnNameLookupEntry-fromJSON}{\code{UserBlueprintsHexColumnNameLookupEntry$fromJSON()}}
\item \href{#method-UserBlueprintsHexColumnNameLookupEntry-clone}{\code{UserBlueprintsHexColumnNameLookupEntry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintsHexColumnNameLookupEntry-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintsHexColumnNameLookupEntry-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintsHexColumnNameLookupEntry$new(
  colname = NULL,
  hex = NULL,
  projectId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{colname}}{The name of the column.}

\item{\code{hex}}{A safe hex representation of the column name.}

\item{\code{projectId}}{The id of the project from which the column name originates.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintsHexColumnNameLookupEntry-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintsHexColumnNameLookupEntry-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintsHexColumnNameLookupEntry$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintsHexColumnNameLookupEntry-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintsHexColumnNameLookupEntry-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintsHexColumnNameLookupEntry$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintsHexColumnNameLookupEntry-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintsHexColumnNameLookupEntry-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintsHexColumnNameLookupEntry$fromJSON(
  UserBlueprintsHexColumnNameLookupEntryJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserBlueprintsHexColumnNameLookupEntryJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintsHexColumnNameLookupEntry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintsHexColumnNameLookupEntry-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintsHexColumnNameLookupEntry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
