% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_blueprints_list_response.R
\docType{class}
\name{UserBlueprintsListResponse}
\alias{UserBlueprintsListResponse}
\title{UserBlueprintsListResponse}
\format{
An \code{R6Class} generator object
}
\description{
UserBlueprintsListResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer Number of records on this page}

\item{\code{data}}{list( \link{UserBlueprintsListItem} ) List of user blueprints}

\item{\code{next_}}{character URL to the next page, or `null` if there is no such page}

\item{\code{previous}}{character URL to the previous page, or `null` if there is no such page}

\item{\code{totalCount}}{integer \link{optional} Total number of records}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserBlueprintsListResponse-new}{\code{UserBlueprintsListResponse$new()}}
\item \href{#method-UserBlueprintsListResponse-validate}{\code{UserBlueprintsListResponse$validate()}}
\item \href{#method-UserBlueprintsListResponse-toJSON}{\code{UserBlueprintsListResponse$toJSON()}}
\item \href{#method-UserBlueprintsListResponse-fromJSON}{\code{UserBlueprintsListResponse$fromJSON()}}
\item \href{#method-UserBlueprintsListResponse-clone}{\code{UserBlueprintsListResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintsListResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintsListResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintsListResponse$new(
  count = NULL,
  data = NULL,
  next_ = NULL,
  previous = NULL,
  totalCount = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{Number of records on this page}

\item{\code{data}}{List of user blueprints}

\item{\code{next_}}{URL to the next page, or `null` if there is no such page}

\item{\code{previous}}{URL to the previous page, or `null` if there is no such page}

\item{\code{totalCount}}{Total number of records}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintsListResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintsListResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintsListResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintsListResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintsListResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintsListResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintsListResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintsListResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintsListResponse$fromJSON(
  UserBlueprintsListResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserBlueprintsListResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintsListResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintsListResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintsListResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
