% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_blueprint_task.R
\docType{class}
\name{UserBlueprintTask}
\alias{UserBlueprintTask}
\title{UserBlueprintTask}
\format{
An \code{R6Class} generator object
}
\description{
UserBlueprintTask Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{arguments}}{list( \link{UserBlueprintTaskArgument} ) A list of definitions of each argument which can be set for the task.}

\item{\code{categories}}{list( character ) The categories which the task is in.}

\item{\code{colnamesAndTypes}}{list( \link{ColnameAndType} ) \link{optional} The column names, their types, and their hex representation, available in the specified project for the task.}

\item{\code{customTaskId}}{character \link{optional} The id of the custom task, if it is a custom task.}

\item{\code{customTaskVersions}}{list( \link{UserBlueprintTaskCustomTaskMetadataWithArguments} ) \link{optional} Metadata for all of the custom task's versions.}

\item{\code{description}}{character A description of the task.}

\item{\code{icon}}{integer The integer representing the id to be displayed when the blueprint is trained.}

\item{\code{isCommonTask}}{character \link{optional} Whether the task is a common task.}

\item{\code{isCustomTask}}{character \link{optional} Whether the task is custom code written by the user.}

\item{\code{isVisibleInComposableMl}}{character \link{optional} Whether the task is visible in the ComposableML menu}

\item{\code{label}}{character The generic / default title or label for the task.}

\item{\code{outputMethods}}{list( character ) The methods which the task can use to produce output.}

\item{\code{supportsScoringCode}}{character Whether the task supports Scoring Code.}

\item{\code{taskCode}}{character The unique code which represents the task to be constructed and executed}

\item{\code{timeSeriesOnly}}{character Whether the task can only be used with time series projects.}

\item{\code{url}}{\link{OneOfobjectstring} The URL of the documentation of the task.}

\item{\code{validInputs}}{list( character ) The supported input types of the task.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UserBlueprintTask$new()}}
\item \href{#method-validate}{\code{UserBlueprintTask$validate()}}
\item \href{#method-toJSON}{\code{UserBlueprintTask$toJSON()}}
\item \href{#method-fromJSON}{\code{UserBlueprintTask$fromJSON()}}
\item \href{#method-clone}{\code{UserBlueprintTask$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTask$new(
  arguments = NULL,
  categories = NULL,
  description = NULL,
  icon = NULL,
  label = NULL,
  outputMethods = NULL,
  supportsScoringCode = NULL,
  taskCode = NULL,
  timeSeriesOnly = NULL,
  url = NULL,
  validInputs = NULL,
  colnamesAndTypes = NULL,
  customTaskId = NULL,
  customTaskVersions = NULL,
  isCommonTask = FALSE,
  isCustomTask = NULL,
  isVisibleInComposableMl = TRUE,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{arguments}}{A list of definitions of each argument which can be set for the task.}

\item{\code{categories}}{The categories which the task is in.}

\item{\code{description}}{A description of the task.}

\item{\code{icon}}{The integer representing the id to be displayed when the blueprint is trained.}

\item{\code{label}}{The generic / default title or label for the task.}

\item{\code{outputMethods}}{The methods which the task can use to produce output.}

\item{\code{supportsScoringCode}}{Whether the task supports Scoring Code.}

\item{\code{taskCode}}{The unique code which represents the task to be constructed and executed}

\item{\code{timeSeriesOnly}}{Whether the task can only be used with time series projects.}

\item{\code{url}}{The URL of the documentation of the task.}

\item{\code{validInputs}}{The supported input types of the task.}

\item{\code{colnamesAndTypes}}{The column names, their types, and their hex representation, available in the specified project for the task.}

\item{\code{customTaskId}}{The id of the custom task, if it is a custom task.}

\item{\code{customTaskVersions}}{Metadata for all of the custom task's versions.}

\item{\code{isCommonTask}}{Whether the task is a common task.}

\item{\code{isCustomTask}}{Whether the task is custom code written by the user.}

\item{\code{isVisibleInComposableMl}}{Whether the task is visible in the ComposableML menu}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTask$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTask$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTask$fromJSON(UserBlueprintTaskJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserBlueprintTaskJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
