% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_blueprint_task_argument.R
\docType{class}
\name{UserBlueprintTaskArgument}
\alias{UserBlueprintTaskArgument}
\title{UserBlueprintTaskArgument}
\format{
An \code{R6Class} generator object
}
\description{
UserBlueprintTaskArgument Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{argument}}{\link{OneOfUserBlueprintTaskArgumentDefinition} The definition of a task argument, used to specify a certain aspect of the task.}

\item{\code{key}}{character The unique key of the argument}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserBlueprintTaskArgument-new}{\code{UserBlueprintTaskArgument$new()}}
\item \href{#method-UserBlueprintTaskArgument-validate}{\code{UserBlueprintTaskArgument$validate()}}
\item \href{#method-UserBlueprintTaskArgument-toJSON}{\code{UserBlueprintTaskArgument$toJSON()}}
\item \href{#method-UserBlueprintTaskArgument-fromJSON}{\code{UserBlueprintTaskArgument$fromJSON()}}
\item \href{#method-UserBlueprintTaskArgument-clone}{\code{UserBlueprintTaskArgument$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintTaskArgument-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintTaskArgument-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskArgument$new(
  argument = NULL,
  key = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{argument}}{The definition of a task argument, used to specify a certain aspect of the task.}

\item{\code{key}}{The unique key of the argument}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintTaskArgument-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintTaskArgument-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskArgument$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintTaskArgument-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintTaskArgument-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskArgument$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintTaskArgument-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintTaskArgument-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskArgument$fromJSON(
  UserBlueprintTaskArgumentJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserBlueprintTaskArgumentJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintTaskArgument-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintTaskArgument-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskArgument$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
