% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_blueprint_task_argument_definition.R
\docType{class}
\name{UserBlueprintTaskArgumentDefinition}
\alias{UserBlueprintTaskArgumentDefinition}
\title{UserBlueprintTaskArgumentDefinition}
\format{
An \code{R6Class} generator object
}
\description{
UserBlueprintTaskArgumentDefinition Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{default}}{\link{OneOfAnyTypearray} \link{optional} The default value of the argument.}

\item{\code{name}}{character The name of the argument.}

\item{\code{recommended}}{\link{OneOfAnyTypearray} \link{optional} The recommended value, based on frequently used values.}

\item{\code{tunable}}{character \link{optional} Whether the argument is tunable by the end-user.}

\item{\code{type}}{character The type of the argument (e.g. \"int\", \"float\", \"select\", \"intgrid\", \"multi\", etc.)}

\item{\code{values}}{\link{OneOfAnyTypearrayobject} The possible values of the argument, which may be a range or list, or dictionary of ranges or lists keyed by type.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserBlueprintTaskArgumentDefinition-new}{\code{UserBlueprintTaskArgumentDefinition$new()}}
\item \href{#method-UserBlueprintTaskArgumentDefinition-validate}{\code{UserBlueprintTaskArgumentDefinition$validate()}}
\item \href{#method-UserBlueprintTaskArgumentDefinition-toJSON}{\code{UserBlueprintTaskArgumentDefinition$toJSON()}}
\item \href{#method-UserBlueprintTaskArgumentDefinition-fromJSON}{\code{UserBlueprintTaskArgumentDefinition$fromJSON()}}
\item \href{#method-UserBlueprintTaskArgumentDefinition-clone}{\code{UserBlueprintTaskArgumentDefinition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintTaskArgumentDefinition-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintTaskArgumentDefinition-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskArgumentDefinition$new(
  name = NULL,
  type = NULL,
  values = NULL,
  default = NULL,
  recommended = NULL,
  tunable = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the argument.}

\item{\code{type}}{The type of the argument (e.g. \"int\", \"float\", \"select\", \"intgrid\", \"multi\", etc.)}

\item{\code{values}}{The possible values of the argument, which may be a range or list, or dictionary of ranges or lists keyed by type.}

\item{\code{default}}{The default value of the argument.}

\item{\code{recommended}}{The recommended value, based on frequently used values.}

\item{\code{tunable}}{Whether the argument is tunable by the end-user.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintTaskArgumentDefinition-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintTaskArgumentDefinition-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskArgumentDefinition$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintTaskArgumentDefinition-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintTaskArgumentDefinition-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskArgumentDefinition$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintTaskArgumentDefinition-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintTaskArgumentDefinition-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskArgumentDefinition$fromJSON(
  UserBlueprintTaskArgumentDefinitionJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserBlueprintTaskArgumentDefinitionJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintTaskArgumentDefinition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintTaskArgumentDefinition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskArgumentDefinition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
