% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_blueprint_task_lookup_entry.R
\docType{class}
\name{UserBlueprintTaskLookupEntry}
\alias{UserBlueprintTaskLookupEntry}
\title{UserBlueprintTaskLookupEntry}
\format{
An \code{R6Class} generator object
}
\description{
UserBlueprintTaskLookupEntry Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{taskCode}}{character The unique code which represents the task to be constructed and executed}

\item{\code{taskDefinition}}{\link{OneOfUserBlueprintTask} A definition of a task in terms of label, arguments, description, and other metadata.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserBlueprintTaskLookupEntry-new}{\code{UserBlueprintTaskLookupEntry$new()}}
\item \href{#method-UserBlueprintTaskLookupEntry-validate}{\code{UserBlueprintTaskLookupEntry$validate()}}
\item \href{#method-UserBlueprintTaskLookupEntry-toJSON}{\code{UserBlueprintTaskLookupEntry$toJSON()}}
\item \href{#method-UserBlueprintTaskLookupEntry-fromJSON}{\code{UserBlueprintTaskLookupEntry$fromJSON()}}
\item \href{#method-UserBlueprintTaskLookupEntry-clone}{\code{UserBlueprintTaskLookupEntry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintTaskLookupEntry-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintTaskLookupEntry-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskLookupEntry$new(
  taskCode = NULL,
  taskDefinition = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{taskCode}}{The unique code which represents the task to be constructed and executed}

\item{\code{taskDefinition}}{A definition of a task in terms of label, arguments, description, and other metadata.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintTaskLookupEntry-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintTaskLookupEntry-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskLookupEntry$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintTaskLookupEntry-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintTaskLookupEntry-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskLookupEntry$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintTaskLookupEntry-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintTaskLookupEntry-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskLookupEntry$fromJSON(
  UserBlueprintTaskLookupEntryJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserBlueprintTaskLookupEntryJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintTaskLookupEntry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintTaskLookupEntry-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskLookupEntry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
