% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_blueprint_task_parameter_validation.R
\docType{class}
\name{UserBlueprintTaskParameterValidation}
\alias{UserBlueprintTaskParameterValidation}
\title{UserBlueprintTaskParameterValidation}
\format{
An \code{R6Class} generator object
}
\description{
UserBlueprintTaskParameterValidation Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{outputMethod}}{character The method representing how the task will output data.}

\item{\code{projectId}}{character \link{optional} The projectId representing the project where this user blueprint is edited.}

\item{\code{taskCode}}{character The task code representing the task to validate parameter values.}

\item{\code{taskParameters}}{list( \link{UserBlueprintTaskParameterValidationRequestParamItem} ) A list of task parameters and proposed values to be validated.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserBlueprintTaskParameterValidation-new}{\code{UserBlueprintTaskParameterValidation$new()}}
\item \href{#method-UserBlueprintTaskParameterValidation-validate}{\code{UserBlueprintTaskParameterValidation$validate()}}
\item \href{#method-UserBlueprintTaskParameterValidation-toJSON}{\code{UserBlueprintTaskParameterValidation$toJSON()}}
\item \href{#method-UserBlueprintTaskParameterValidation-fromJSON}{\code{UserBlueprintTaskParameterValidation$fromJSON()}}
\item \href{#method-UserBlueprintTaskParameterValidation-clone}{\code{UserBlueprintTaskParameterValidation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintTaskParameterValidation-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintTaskParameterValidation-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskParameterValidation$new(
  outputMethod = NULL,
  taskCode = NULL,
  taskParameters = NULL,
  projectId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outputMethod}}{The method representing how the task will output data.}

\item{\code{taskCode}}{The task code representing the task to validate parameter values.}

\item{\code{taskParameters}}{A list of task parameters and proposed values to be validated.}

\item{\code{projectId}}{The projectId representing the project where this user blueprint is edited.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintTaskParameterValidation-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintTaskParameterValidation-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskParameterValidation$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintTaskParameterValidation-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintTaskParameterValidation-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskParameterValidation$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintTaskParameterValidation-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintTaskParameterValidation-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskParameterValidation$fromJSON(
  UserBlueprintTaskParameterValidationJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserBlueprintTaskParameterValidationJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserBlueprintTaskParameterValidation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserBlueprintTaskParameterValidation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskParameterValidation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
