% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_blueprint_validation.R
\docType{class}
\name{UserBlueprintValidation}
\alias{UserBlueprintValidation}
\title{UserBlueprintValidation}
\format{
An \code{R6Class} generator object
}
\description{
UserBlueprintValidation Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{blueprint}}{\link{OneOfarrayobject} The representation of a directed acyclic graph defining a pipeline of data through tasks and a final estimator.}

\item{\code{isInplaceEditor}}{character Whether the request is sent from the in place user BP editor.}

\item{\code{projectId}}{character \link{optional} String representation of ObjectId for the currently active project. The user blueprint is validated when this project is active. Necessary in the event of project specific tasks, such as column selection tasks.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UserBlueprintValidation$new()}}
\item \href{#method-validate}{\code{UserBlueprintValidation$validate()}}
\item \href{#method-toJSON}{\code{UserBlueprintValidation$toJSON()}}
\item \href{#method-fromJSON}{\code{UserBlueprintValidation$fromJSON()}}
\item \href{#method-clone}{\code{UserBlueprintValidation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintValidation$new(
  blueprint = NULL,
  isInplaceEditor = NULL,
  projectId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blueprint}}{The representation of a directed acyclic graph defining a pipeline of data through tasks and a final estimator.}

\item{\code{isInplaceEditor}}{Whether the request is sent from the in place user BP editor.}

\item{\code{projectId}}{String representation of ObjectId for the currently active project. The user blueprint is validated when this project is active. Necessary in the event of project specific tasks, such as column selection tasks.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintValidation$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintValidation$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintValidation$fromJSON(
  UserBlueprintValidationJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserBlueprintValidationJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintValidation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
