% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_create.R
\docType{class}
\name{UserCreate}
\alias{UserCreate}
\title{UserCreate}
\format{
An \code{R6Class} generator object
}
\description{
UserCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{country}}{character \link{optional} Country where the user lives.}

\item{\code{defaultCatalogDatasetSampleSize}}{\link{AdminSampleSize} \link{optional}}

\item{\code{expirationDate}}{character \link{optional} Datetime, RFC3339, at which the user should expire.}

\item{\code{externalId}}{character \link{optional} User ID in external IdP.}

\item{\code{firstName}}{character First name of the user being created.}

\item{\code{language}}{character \link{optional} Sets the language in the app for the user, and the language of the invite email, if the user is being invited (i.e. the password is omitted). Value must be a valid ISO-639-1 language code. Available options are: `en`, `ja`, `fr`, `ko`. System's default language will be used if omitted.}

\item{\code{lastName}}{character \link{optional} Last name of the user being created.}

\item{\code{maxGpuWorkers}}{integer \link{optional} Amount of user's available GPU workers.}

\item{\code{maxIdleWorkers}}{integer \link{optional} Amount of org workers that the user can utilize when idle.}

\item{\code{maxUploadSize}}{integer \link{optional} The upper limit for the allowed upload size for this user.}

\item{\code{maxUploadSizeCatalog}}{integer \link{optional} The upper limit for the allowed upload size in the AI catalog for this user.}

\item{\code{maxWorkers}}{integer \link{optional} Amount of user's available workers.}

\item{\code{organizationId}}{character \link{optional} ID of the organization to add the user to.}

\item{\code{password}}{character \link{optional} Creates user with this password, if blank sends username invite.}

\item{\code{requireClickthroughAgreement}}{character \link{optional} Boolean to require the user to agree to a clickthrough.}

\item{\code{userType}}{character External name for UserType such as AcademicUser, BasicUser, ProUser.}

\item{\code{username}}{character Email address of new user.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserCreate-new}{\code{UserCreate$new()}}
\item \href{#method-UserCreate-validate}{\code{UserCreate$validate()}}
\item \href{#method-UserCreate-toJSON}{\code{UserCreate$toJSON()}}
\item \href{#method-UserCreate-fromJSON}{\code{UserCreate$fromJSON()}}
\item \href{#method-UserCreate-clone}{\code{UserCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserCreate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserCreate$new(
  firstName = NULL,
  userType = NULL,
  username = NULL,
  country = NULL,
  defaultCatalogDatasetSampleSize = NULL,
  expirationDate = NULL,
  externalId = NULL,
  language = NULL,
  lastName = NULL,
  maxGpuWorkers = NULL,
  maxIdleWorkers = NULL,
  maxUploadSize = NULL,
  maxUploadSizeCatalog = NULL,
  maxWorkers = NULL,
  organizationId = NULL,
  password = NULL,
  requireClickthroughAgreement = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{firstName}}{First name of the user being created.}

\item{\code{userType}}{External name for UserType such as AcademicUser, BasicUser, ProUser.}

\item{\code{username}}{Email address of new user.}

\item{\code{country}}{Country where the user lives.}

\item{\code{expirationDate}}{Datetime, RFC3339, at which the user should expire.}

\item{\code{externalId}}{User ID in external IdP.}

\item{\code{language}}{Sets the language in the app for the user, and the language of the invite email, if the user is being invited (i.e. the password is omitted). Value must be a valid ISO-639-1 language code. Available options are: `en`, `ja`, `fr`, `ko`. System's default language will be used if omitted.}

\item{\code{lastName}}{Last name of the user being created.}

\item{\code{maxGpuWorkers}}{Amount of user's available GPU workers.}

\item{\code{maxIdleWorkers}}{Amount of org workers that the user can utilize when idle.}

\item{\code{maxUploadSize}}{The upper limit for the allowed upload size for this user.}

\item{\code{maxUploadSizeCatalog}}{The upper limit for the allowed upload size in the AI catalog for this user.}

\item{\code{maxWorkers}}{Amount of user's available workers.}

\item{\code{organizationId}}{ID of the organization to add the user to.}

\item{\code{password}}{Creates user with this password, if blank sends username invite.}

\item{\code{requireClickthroughAgreement}}{Boolean to require the user to agree to a clickthrough.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserCreate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UserCreate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserCreate$fromJSON(UserCreateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
