% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_group_response.R
\docType{class}
\name{UserGroupResponse}
\alias{UserGroupResponse}
\title{UserGroupResponse}
\format{
An \code{R6Class} generator object
}
\description{
UserGroupResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{accessRoleId}}{character The identifier of the access role assigned to the group.}

\item{\code{accessRoleName}}{character The name of the access role assigned to the group.}

\item{\code{accountPermissions}}{named list( character ) Account permissions of this user group. Each key is a permission name.}

\item{\code{createdBy}}{character The identifier of the user who created this user group.}

\item{\code{description}}{character The description of this user group.}

\item{\code{email}}{character The email of this user group.}

\item{\code{id}}{character The identifier of the user group.}

\item{\code{maxAppCount}}{integer The maximum allowed applications count for users of this group.}

\item{\code{maxAppCountLimit}}{integer The upper limit for the maximum allowed applications count. The limit is defined if the group is part of an organization, and the attribute is set on the organization level.}

\item{\code{maxCustomDeployments}}{integer The number of maximum custom deployments available for users of this group.}

\item{\code{maxCustomDeploymentsLimit}}{integer The upper limit for the number of maximum custom deployments. The limit is defined if the group is part of an organization, and the attribute is set on the organization level.}

\item{\code{maxEdaWorkers}}{integer The upper limit for a number of EDA workers to be run for users of this user group.}

\item{\code{maxRam}}{integer The upper limit for amount of RAM available to users of this user group.}

\item{\code{maxUploadCatalogSizeLimit}}{integer The upper limit for the maximum upload size in the AI catalog. The limit is defined if the group is part of an organization, and the attribute is set on the organization level.}

\item{\code{maxUploadSize}}{integer The maximum allowed upload size for users of this group.}

\item{\code{maxUploadSizeCatalog}}{integer The maximum allowed upload size in the AI catalog for users in this group.}

\item{\code{maxUploadSizeLimit}}{integer The upper limit for the maximum allowed upload size. The limit is defined if the group is part of an organization, and the attribute is set on the organization level.}

\item{\code{maxWorkers}}{integer The upper limit for a number of workers to be run for users of this user group.}

\item{\code{membersCount}}{integer The number of members in this user group.}

\item{\code{name}}{character The name of the user group.}

\item{\code{orgId}}{character The identifier of the organization the user group belongs to.}

\item{\code{orgName}}{character The name of the organization this user group belongs to.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UserGroupResponse$new()}}
\item \href{#method-validate}{\code{UserGroupResponse$validate()}}
\item \href{#method-toJSON}{\code{UserGroupResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{UserGroupResponse$fromJSON()}}
\item \href{#method-clone}{\code{UserGroupResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserGroupResponse$new(
  accessRoleId = NULL,
  accessRoleName = NULL,
  accountPermissions = NULL,
  createdBy = NULL,
  description = NULL,
  email = NULL,
  id = NULL,
  maxAppCount = NULL,
  maxAppCountLimit = NULL,
  maxCustomDeployments = NULL,
  maxCustomDeploymentsLimit = NULL,
  maxEdaWorkers = NULL,
  maxRam = NULL,
  maxUploadCatalogSizeLimit = NULL,
  maxUploadSize = NULL,
  maxUploadSizeCatalog = NULL,
  maxUploadSizeLimit = NULL,
  maxWorkers = NULL,
  membersCount = NULL,
  name = NULL,
  orgId = NULL,
  orgName = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{accessRoleId}}{The identifier of the access role assigned to the group.}

\item{\code{accessRoleName}}{The name of the access role assigned to the group.}

\item{\code{accountPermissions}}{Account permissions of this user group. Each key is a permission name.}

\item{\code{createdBy}}{The identifier of the user who created this user group.}

\item{\code{description}}{The description of this user group.}

\item{\code{email}}{The email of this user group.}

\item{\code{id}}{The identifier of the user group.}

\item{\code{maxAppCount}}{The maximum allowed applications count for users of this group.}

\item{\code{maxAppCountLimit}}{The upper limit for the maximum allowed applications count. The limit is defined if the group is part of an organization, and the attribute is set on the organization level.}

\item{\code{maxCustomDeployments}}{The number of maximum custom deployments available for users of this group.}

\item{\code{maxCustomDeploymentsLimit}}{The upper limit for the number of maximum custom deployments. The limit is defined if the group is part of an organization, and the attribute is set on the organization level.}

\item{\code{maxEdaWorkers}}{The upper limit for a number of EDA workers to be run for users of this user group.}

\item{\code{maxRam}}{The upper limit for amount of RAM available to users of this user group.}

\item{\code{maxUploadCatalogSizeLimit}}{The upper limit for the maximum upload size in the AI catalog. The limit is defined if the group is part of an organization, and the attribute is set on the organization level.}

\item{\code{maxUploadSize}}{The maximum allowed upload size for users of this group.}

\item{\code{maxUploadSizeCatalog}}{The maximum allowed upload size in the AI catalog for users in this group.}

\item{\code{maxUploadSizeLimit}}{The upper limit for the maximum allowed upload size. The limit is defined if the group is part of an organization, and the attribute is set on the organization level.}

\item{\code{maxWorkers}}{The upper limit for a number of workers to be run for users of this user group.}

\item{\code{membersCount}}{The number of members in this user group.}

\item{\code{name}}{The name of the user group.}

\item{\code{orgId}}{The identifier of the organization the user group belongs to.}

\item{\code{orgName}}{The name of the organization this user group belongs to.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserGroupResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserGroupResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserGroupResponse$fromJSON(UserGroupResponseJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserGroupResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserGroupResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
