% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_group_update.R
\docType{class}
\name{UserGroupUpdate}
\alias{UserGroupUpdate}
\title{UserGroupUpdate}
\format{
An \code{R6Class} generator object
}
\description{
UserGroupUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{accessRoleId}}{character \link{optional} The identifier of the access role to assign to the group.}

\item{\code{accountPermissions}}{named list( character ) \link{optional} Account permissions to set for this user group. Each key is a permission name.}

\item{\code{description}}{character \link{optional} The description of this user group.}

\item{\code{email}}{character \link{optional} The email of this user group.}

\item{\code{maxAppCount}}{integer \link{optional} The upper limit for the allowed applications count for the user group.}

\item{\code{maxCustomDeployments}}{integer \link{optional} Maximum allowed value for maxCustomDeployment for the user group.}

\item{\code{maxEdaWorkers}}{integer \link{optional} The upper limit for a number of EDA workers to be run for users of this user group.}

\item{\code{maxRam}}{integer \link{optional} The upper limit for amount of RAM available to users of this user group.}

\item{\code{maxUploadSize}}{integer \link{optional} The upper limit for the allowed upload size for the user group.}

\item{\code{maxUploadSizeCatalog}}{integer \link{optional} The upper limit for the allowed upload size in the AI catalog for this user group.}

\item{\code{maxWorkers}}{integer \link{optional} The upper limit for a number of workers to be run for users of this user group.}

\item{\code{name}}{character \link{optional} The new name for the user group.}

\item{\code{orgId}}{character \link{optional} The id of the organization to assign the user group to.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UserGroupUpdate$new()}}
\item \href{#method-validate}{\code{UserGroupUpdate$validate()}}
\item \href{#method-toJSON}{\code{UserGroupUpdate$toJSON()}}
\item \href{#method-fromJSON}{\code{UserGroupUpdate$fromJSON()}}
\item \href{#method-clone}{\code{UserGroupUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserGroupUpdate$new(
  accessRoleId = NULL,
  accountPermissions = NULL,
  description = NULL,
  email = NULL,
  maxAppCount = NULL,
  maxCustomDeployments = NULL,
  maxEdaWorkers = NULL,
  maxRam = NULL,
  maxUploadSize = NULL,
  maxUploadSizeCatalog = NULL,
  maxWorkers = NULL,
  name = NULL,
  orgId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{accessRoleId}}{The identifier of the access role to assign to the group.}

\item{\code{accountPermissions}}{Account permissions to set for this user group. Each key is a permission name.}

\item{\code{description}}{The description of this user group.}

\item{\code{email}}{The email of this user group.}

\item{\code{maxAppCount}}{The upper limit for the allowed applications count for the user group.}

\item{\code{maxCustomDeployments}}{Maximum allowed value for maxCustomDeployment for the user group.}

\item{\code{maxEdaWorkers}}{The upper limit for a number of EDA workers to be run for users of this user group.}

\item{\code{maxRam}}{The upper limit for amount of RAM available to users of this user group.}

\item{\code{maxUploadSize}}{The upper limit for the allowed upload size for the user group.}

\item{\code{maxUploadSizeCatalog}}{The upper limit for the allowed upload size in the AI catalog for this user group.}

\item{\code{maxWorkers}}{The upper limit for a number of workers to be run for users of this user group.}

\item{\code{name}}{The new name for the user group.}

\item{\code{orgId}}{The id of the organization to assign the user group to.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserGroupUpdate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserGroupUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserGroupUpdate$fromJSON(UserGroupUpdateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserGroupUpdateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserGroupUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
