% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/username_password_catalog_data_source_payload.R
\docType{class}
\name{UsernamePasswordCatalogDataSourcePayload}
\alias{UsernamePasswordCatalogDataSourcePayload}
\title{UsernamePasswordCatalogDataSourcePayload}
\format{
An \code{R6Class} generator object
}
\description{
UsernamePasswordCatalogDataSourcePayload Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datasourceId}}{character Id of the Catalog object}

\item{\code{password}}{character User password to connect to the data location}

\item{\code{typeId}}{character Id of the target, currently only `catalog`}

\item{\code{userName}}{character User name to connect to the data location}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UsernamePasswordCatalogDataSourcePayload$new()}}
\item \href{#method-validate}{\code{UsernamePasswordCatalogDataSourcePayload$validate()}}
\item \href{#method-toJSON}{\code{UsernamePasswordCatalogDataSourcePayload$toJSON()}}
\item \href{#method-fromJSON}{\code{UsernamePasswordCatalogDataSourcePayload$fromJSON()}}
\item \href{#method-clone}{\code{UsernamePasswordCatalogDataSourcePayload$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsernamePasswordCatalogDataSourcePayload$new(
  datasourceId = NULL,
  password = NULL,
  typeId = NULL,
  userName = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasourceId}}{Id of the Catalog object}

\item{\code{password}}{User password to connect to the data location}

\item{\code{typeId}}{Id of the target, currently only `catalog`}

\item{\code{userName}}{User name to connect to the data location}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsernamePasswordCatalogDataSourcePayload$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsernamePasswordCatalogDataSourcePayload$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsernamePasswordCatalogDataSourcePayload$fromJSON(
  UsernamePasswordCatalogDataSourcePayloadJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UsernamePasswordCatalogDataSourcePayloadJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsernamePasswordCatalogDataSourcePayload$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
