% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_notification_related_comment_response.R
\docType{class}
\name{UserNotificationRelatedCommentResponse}
\alias{UserNotificationRelatedCommentResponse}
\title{UserNotificationRelatedCommentResponse}
\format{
An \code{R6Class} generator object
}
\description{
UserNotificationRelatedCommentResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{commentId}}{character The ID of the comment}

\item{\code{commentLink}}{character \link{optional} The link to the comment.}

\item{\code{entityId}}{character The ID of the commented entity}

\item{\code{entityType}}{character \link{optional} The type of the commented entity.}

\item{\code{inactive}}{character \link{optional} True if the comment was deleted.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserNotificationRelatedCommentResponse-new}{\code{UserNotificationRelatedCommentResponse$new()}}
\item \href{#method-UserNotificationRelatedCommentResponse-validate}{\code{UserNotificationRelatedCommentResponse$validate()}}
\item \href{#method-UserNotificationRelatedCommentResponse-toJSON}{\code{UserNotificationRelatedCommentResponse$toJSON()}}
\item \href{#method-UserNotificationRelatedCommentResponse-fromJSON}{\code{UserNotificationRelatedCommentResponse$fromJSON()}}
\item \href{#method-UserNotificationRelatedCommentResponse-clone}{\code{UserNotificationRelatedCommentResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserNotificationRelatedCommentResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserNotificationRelatedCommentResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationRelatedCommentResponse$new(
  commentId = NULL,
  entityId = NULL,
  commentLink = NULL,
  entityType = NULL,
  inactive = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commentId}}{The ID of the comment}

\item{\code{entityId}}{The ID of the commented entity}

\item{\code{commentLink}}{The link to the comment.}

\item{\code{entityType}}{The type of the commented entity.}

\item{\code{inactive}}{True if the comment was deleted.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserNotificationRelatedCommentResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UserNotificationRelatedCommentResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationRelatedCommentResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserNotificationRelatedCommentResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserNotificationRelatedCommentResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationRelatedCommentResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserNotificationRelatedCommentResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserNotificationRelatedCommentResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationRelatedCommentResponse$fromJSON(
  UserNotificationRelatedCommentResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserNotificationRelatedCommentResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserNotificationRelatedCommentResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserNotificationRelatedCommentResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationRelatedCommentResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
