% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_notification_related_deployment_response.R
\docType{class}
\name{UserNotificationRelatedDeploymentResponse}
\alias{UserNotificationRelatedDeploymentResponse}
\title{UserNotificationRelatedDeploymentResponse}
\format{
An \code{R6Class} generator object
}
\description{
UserNotificationRelatedDeploymentResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{deploymentId}}{character The ID of the deployment}

\item{\code{deploymentName}}{character \link{optional} Deployment's label}

\item{\code{deploymentUrl}}{character \link{optional} The link to the deployment}

\item{\code{inactive}}{character \link{optional} True if the deployment was deleted.}

\item{\code{modelId}}{character \link{optional} The ID of the related model}

\item{\code{projectId}}{character \link{optional} The ID of the related project}

\item{\code{userId}}{character \link{optional} The ID of the related user}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UserNotificationRelatedDeploymentResponse$new()}}
\item \href{#method-validate}{\code{UserNotificationRelatedDeploymentResponse$validate()}}
\item \href{#method-toJSON}{\code{UserNotificationRelatedDeploymentResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{UserNotificationRelatedDeploymentResponse$fromJSON()}}
\item \href{#method-clone}{\code{UserNotificationRelatedDeploymentResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationRelatedDeploymentResponse$new(
  deploymentId = NULL,
  deploymentName = NULL,
  deploymentUrl = NULL,
  inactive = NULL,
  modelId = NULL,
  projectId = NULL,
  userId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deploymentId}}{The ID of the deployment}

\item{\code{deploymentName}}{Deployment's label}

\item{\code{deploymentUrl}}{The link to the deployment}

\item{\code{inactive}}{True if the deployment was deleted.}

\item{\code{modelId}}{The ID of the related model}

\item{\code{projectId}}{The ID of the related project}

\item{\code{userId}}{The ID of the related user}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationRelatedDeploymentResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationRelatedDeploymentResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationRelatedDeploymentResponse$fromJSON(
  UserNotificationRelatedDeploymentResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserNotificationRelatedDeploymentResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationRelatedDeploymentResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
