% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_notification_related_project_response.R
\docType{class}
\name{UserNotificationRelatedProjectResponse}
\alias{UserNotificationRelatedProjectResponse}
\title{UserNotificationRelatedProjectResponse}
\format{
An \code{R6Class} generator object
}
\description{
UserNotificationRelatedProjectResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{inactive}}{character \link{optional} True if the project was deleted.}

\item{\code{pid}}{character The ID of the project}

\item{\code{projectLink}}{character \link{optional} The link to the project.}

\item{\code{projectName}}{character \link{optional} Project name.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserNotificationRelatedProjectResponse-new}{\code{UserNotificationRelatedProjectResponse$new()}}
\item \href{#method-UserNotificationRelatedProjectResponse-validate}{\code{UserNotificationRelatedProjectResponse$validate()}}
\item \href{#method-UserNotificationRelatedProjectResponse-toJSON}{\code{UserNotificationRelatedProjectResponse$toJSON()}}
\item \href{#method-UserNotificationRelatedProjectResponse-fromJSON}{\code{UserNotificationRelatedProjectResponse$fromJSON()}}
\item \href{#method-UserNotificationRelatedProjectResponse-clone}{\code{UserNotificationRelatedProjectResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserNotificationRelatedProjectResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserNotificationRelatedProjectResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationRelatedProjectResponse$new(
  pid = NULL,
  inactive = NULL,
  projectLink = NULL,
  projectName = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pid}}{The ID of the project}

\item{\code{inactive}}{True if the project was deleted.}

\item{\code{projectLink}}{The link to the project.}

\item{\code{projectName}}{Project name.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserNotificationRelatedProjectResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UserNotificationRelatedProjectResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationRelatedProjectResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserNotificationRelatedProjectResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserNotificationRelatedProjectResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationRelatedProjectResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserNotificationRelatedProjectResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserNotificationRelatedProjectResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationRelatedProjectResponse$fromJSON(
  UserNotificationRelatedProjectResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserNotificationRelatedProjectResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserNotificationRelatedProjectResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserNotificationRelatedProjectResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationRelatedProjectResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
