% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_notification_related_user_response.R
\docType{class}
\name{UserNotificationRelatedUserResponse}
\alias{UserNotificationRelatedUserResponse}
\title{UserNotificationRelatedUserResponse}
\format{
An \code{R6Class} generator object
}
\description{
UserNotificationRelatedUserResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fullName}}{character \link{optional} User's full name.}

\item{\code{gravatarHash}}{character \link{optional} User's gravatar hash.}

\item{\code{inactive}}{character \link{optional} True if the user was deleted.}

\item{\code{uid}}{character The ID of the user}

\item{\code{username}}{character \link{optional} Username of the user.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserNotificationRelatedUserResponse-new}{\code{UserNotificationRelatedUserResponse$new()}}
\item \href{#method-UserNotificationRelatedUserResponse-validate}{\code{UserNotificationRelatedUserResponse$validate()}}
\item \href{#method-UserNotificationRelatedUserResponse-toJSON}{\code{UserNotificationRelatedUserResponse$toJSON()}}
\item \href{#method-UserNotificationRelatedUserResponse-fromJSON}{\code{UserNotificationRelatedUserResponse$fromJSON()}}
\item \href{#method-UserNotificationRelatedUserResponse-clone}{\code{UserNotificationRelatedUserResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserNotificationRelatedUserResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserNotificationRelatedUserResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationRelatedUserResponse$new(
  uid = NULL,
  fullName = NULL,
  gravatarHash = NULL,
  inactive = NULL,
  username = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uid}}{The ID of the user}

\item{\code{fullName}}{User's full name.}

\item{\code{gravatarHash}}{User's gravatar hash.}

\item{\code{inactive}}{True if the user was deleted.}

\item{\code{username}}{Username of the user.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserNotificationRelatedUserResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UserNotificationRelatedUserResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationRelatedUserResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserNotificationRelatedUserResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserNotificationRelatedUserResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationRelatedUserResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserNotificationRelatedUserResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserNotificationRelatedUserResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationRelatedUserResponse$fromJSON(
  UserNotificationRelatedUserResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserNotificationRelatedUserResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserNotificationRelatedUserResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserNotificationRelatedUserResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationRelatedUserResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
