% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_notification_response.R
\docType{class}
\name{UserNotificationResponse}
\alias{UserNotificationResponse}
\title{UserNotificationResponse}
\format{
An \code{R6Class} generator object
}
\description{
UserNotificationResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{callerUser}}{\link{UserNotificationRelatedUserResponse}}

\item{\code{created}}{character The ISO 8601 formatted date and time when the notification was created.}

\item{\code{description}}{character Notification's description}

\item{\code{eventType}}{character The type of the notification.}

\item{\code{isRead}}{character True if the notification is already read.}

\item{\code{link}}{character Call-to-action link for the notification}

\item{\code{pushNotificationSent}}{character True if the notification was also sent via push notifications.}

\item{\code{relatedComment}}{\link{UserNotificationRelatedCommentResponse} \link{optional}}

\item{\code{relatedDeployment}}{\link{UserNotificationRelatedDeploymentResponse} \link{optional}}

\item{\code{relatedProject}}{\link{UserNotificationRelatedProjectResponse} \link{optional}}

\item{\code{relatedUsersDelete}}{\link{UserNotificationRelatedUsersDeleteResponse} \link{optional}}

\item{\code{sharedUsers}}{list( \link{UserNotificationRelatedUserResponse} ) \link{optional} List of the user details a resource was shared with.}

\item{\code{statusId}}{character \link{optional} Asynchronous job status ID}

\item{\code{title}}{character Notification's title}

\item{\code{updated}}{character The ISO 8601 formatted date and time when the notification was updated.}

\item{\code{userNotificationId}}{character The ID of the notification}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserNotificationResponse-new}{\code{UserNotificationResponse$new()}}
\item \href{#method-UserNotificationResponse-validate}{\code{UserNotificationResponse$validate()}}
\item \href{#method-UserNotificationResponse-toJSON}{\code{UserNotificationResponse$toJSON()}}
\item \href{#method-UserNotificationResponse-fromJSON}{\code{UserNotificationResponse$fromJSON()}}
\item \href{#method-UserNotificationResponse-clone}{\code{UserNotificationResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserNotificationResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserNotificationResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationResponse$new(
  callerUser = NULL,
  created = NULL,
  description = NULL,
  eventType = NULL,
  isRead = NULL,
  link = NULL,
  pushNotificationSent = NULL,
  title = NULL,
  updated = NULL,
  userNotificationId = NULL,
  relatedComment = NULL,
  relatedDeployment = NULL,
  relatedProject = NULL,
  relatedUsersDelete = NULL,
  sharedUsers = NULL,
  statusId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{created}}{The ISO 8601 formatted date and time when the notification was created.}

\item{\code{description}}{Notification's description}

\item{\code{eventType}}{The type of the notification.}

\item{\code{isRead}}{True if the notification is already read.}

\item{\code{link}}{Call-to-action link for the notification}

\item{\code{pushNotificationSent}}{True if the notification was also sent via push notifications.}

\item{\code{title}}{Notification's title}

\item{\code{updated}}{The ISO 8601 formatted date and time when the notification was updated.}

\item{\code{userNotificationId}}{The ID of the notification}

\item{\code{sharedUsers}}{List of the user details a resource was shared with.}

\item{\code{statusId}}{Asynchronous job status ID}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserNotificationResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UserNotificationResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserNotificationResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserNotificationResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserNotificationResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserNotificationResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationResponse$fromJSON(
  UserNotificationResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserNotificationResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserNotificationResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserNotificationResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserNotificationResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
