% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_operations.R
\docType{class}
\name{UserOperations}
\alias{UserOperations}
\title{UserOperations}
\format{
An \code{R6Class} generator object
}
\description{
UserOperations Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{canCancel}}{character Whether the user can cancel the Change Request.}

\item{\code{canComment}}{character Whether the user can create commenting review on the Change Request.}

\item{\code{canResolve}}{character Whether the user can resolve the Change Request.}

\item{\code{canReview}}{character Whether the user can review (approve or request updates) the Change Request.}

\item{\code{canUpdate}}{character Whether the user can update the Change Request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UserOperations$new()}}
\item \href{#method-validate}{\code{UserOperations$validate()}}
\item \href{#method-toJSON}{\code{UserOperations$toJSON()}}
\item \href{#method-fromJSON}{\code{UserOperations$fromJSON()}}
\item \href{#method-clone}{\code{UserOperations$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserOperations$new(
  canCancel = NULL,
  canComment = NULL,
  canResolve = NULL,
  canReview = NULL,
  canUpdate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{canCancel}}{Whether the user can cancel the Change Request.}

\item{\code{canComment}}{Whether the user can create commenting review on the Change Request.}

\item{\code{canResolve}}{Whether the user can resolve the Change Request.}

\item{\code{canReview}}{Whether the user can review (approve or request updates) the Change Request.}

\item{\code{canUpdate}}{Whether the user can update the Change Request.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserOperations$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserOperations$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserOperations$fromJSON(UserOperationsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserOperationsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserOperations$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
