% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_role_with_grant.R
\docType{class}
\name{UserRoleWithGrant}
\alias{UserRoleWithGrant}
\title{UserRoleWithGrant}
\format{
An \code{R6Class} generator object
}
\description{
UserRoleWithGrant Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{canShare}}{character \link{optional} Whether the org/group/user should be able to share with others.If true, the org/group/user will be able to grant any role up to and includingtheir own to other orgs/groups/user. If `role` is `NO_ROLE` `canShare` is ignored.}

\item{\code{role}}{character The role to set on the entity. When it is None, the role of this user will be removedfrom this entity.}

\item{\code{username}}{character Username of the user to update the access role for.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserRoleWithGrant-new}{\code{UserRoleWithGrant$new()}}
\item \href{#method-UserRoleWithGrant-validate}{\code{UserRoleWithGrant$validate()}}
\item \href{#method-UserRoleWithGrant-toJSON}{\code{UserRoleWithGrant$toJSON()}}
\item \href{#method-UserRoleWithGrant-fromJSON}{\code{UserRoleWithGrant$fromJSON()}}
\item \href{#method-UserRoleWithGrant-clone}{\code{UserRoleWithGrant$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserRoleWithGrant-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserRoleWithGrant-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserRoleWithGrant$new(
  role = NULL,
  username = NULL,
  canShare = TRUE,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{The role to set on the entity. When it is None, the role of this user will be removedfrom this entity.}

\item{\code{username}}{Username of the user to update the access role for.}

\item{\code{canShare}}{Whether the org/group/user should be able to share with others.If true, the org/group/user will be able to grant any role up to and includingtheir own to other orgs/groups/user. If `role` is `NO_ROLE` `canShare` is ignored.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserRoleWithGrant-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UserRoleWithGrant-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserRoleWithGrant$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserRoleWithGrant-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserRoleWithGrant-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserRoleWithGrant$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserRoleWithGrant-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserRoleWithGrant-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserRoleWithGrant$fromJSON(UserRoleWithGrantJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserRoleWithGrantJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserRoleWithGrant-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserRoleWithGrant-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserRoleWithGrant$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
