% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_permadelete_job_status_response.R
\docType{class}
\name{UsersPermadeleteJobStatusResponse}
\alias{UsersPermadeleteJobStatusResponse}
\title{UsersPermadeleteJobStatusResponse}
\format{
An \code{R6Class} generator object
}
\description{
UsersPermadeleteJobStatusResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{created}}{character The time the status record was created.}

\item{\code{data}}{\link{SummaryPermadeleteStatus}}

\item{\code{message}}{character May contain further information about the status.}

\item{\code{status}}{character The processing state of the job.}

\item{\code{statusId}}{character The ID of the status object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UsersPermadeleteJobStatusResponse$new()}}
\item \href{#method-validate}{\code{UsersPermadeleteJobStatusResponse$validate()}}
\item \href{#method-toJSON}{\code{UsersPermadeleteJobStatusResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{UsersPermadeleteJobStatusResponse$fromJSON()}}
\item \href{#method-clone}{\code{UsersPermadeleteJobStatusResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersPermadeleteJobStatusResponse$new(
  created = NULL,
  data = NULL,
  message = NULL,
  status = NULL,
  statusId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{created}}{The time the status record was created.}

\item{\code{message}}{May contain further information about the status.}

\item{\code{status}}{The processing state of the job.}

\item{\code{statusId}}{The ID of the status object.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersPermadeleteJobStatusResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersPermadeleteJobStatusResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersPermadeleteJobStatusResponse$fromJSON(
  UsersPermadeleteJobStatusResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UsersPermadeleteJobStatusResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersPermadeleteJobStatusResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
