% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_permadelete_preview_statistics.R
\docType{class}
\name{UsersPermadeletePreviewStatistics}
\alias{UsersPermadeletePreviewStatistics}
\title{UsersPermadeletePreviewStatistics}
\format{
An \code{R6Class} generator object
}
\description{
UsersPermadeletePreviewStatistics Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{deploymentsWillBeAssignedNewOwner}}{integer Number of deployments that will be assigned a new owner.}

\item{\code{deploymentsWillBeDeleted}}{integer Number of deployments that will be deleted.}

\item{\code{deploymentsWillNotBeDeleted}}{integer Number of deployments that will not be deleted.}

\item{\code{errors}}{integer Number of errors encountered during preview construction.}

\item{\code{projectsWillBeAssignedNewOwner}}{integer Number of projects that will be assigned a new owner.}

\item{\code{projectsWillBeDeleted}}{integer Number of projects that will be deleted.}

\item{\code{projectsWillNotBeDeleted}}{integer Number of projects that will not be deleted.}

\item{\code{usersWillBeDeleted}}{integer Number of users that will be deleted.}

\item{\code{usersWillNotBeDeleted}}{integer Number of users that will not be deleted.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UsersPermadeletePreviewStatistics-new}{\code{UsersPermadeletePreviewStatistics$new()}}
\item \href{#method-UsersPermadeletePreviewStatistics-validate}{\code{UsersPermadeletePreviewStatistics$validate()}}
\item \href{#method-UsersPermadeletePreviewStatistics-toJSON}{\code{UsersPermadeletePreviewStatistics$toJSON()}}
\item \href{#method-UsersPermadeletePreviewStatistics-fromJSON}{\code{UsersPermadeletePreviewStatistics$fromJSON()}}
\item \href{#method-UsersPermadeletePreviewStatistics-clone}{\code{UsersPermadeletePreviewStatistics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersPermadeletePreviewStatistics-new"></a>}}
\if{latex}{\out{\hypertarget{method-UsersPermadeletePreviewStatistics-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersPermadeletePreviewStatistics$new(
  deploymentsWillBeAssignedNewOwner = NULL,
  deploymentsWillBeDeleted = NULL,
  deploymentsWillNotBeDeleted = NULL,
  errors = NULL,
  projectsWillBeAssignedNewOwner = NULL,
  projectsWillBeDeleted = NULL,
  projectsWillNotBeDeleted = NULL,
  usersWillBeDeleted = NULL,
  usersWillNotBeDeleted = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deploymentsWillBeAssignedNewOwner}}{Number of deployments that will be assigned a new owner.}

\item{\code{deploymentsWillBeDeleted}}{Number of deployments that will be deleted.}

\item{\code{deploymentsWillNotBeDeleted}}{Number of deployments that will not be deleted.}

\item{\code{errors}}{Number of errors encountered during preview construction.}

\item{\code{projectsWillBeAssignedNewOwner}}{Number of projects that will be assigned a new owner.}

\item{\code{projectsWillBeDeleted}}{Number of projects that will be deleted.}

\item{\code{projectsWillNotBeDeleted}}{Number of projects that will not be deleted.}

\item{\code{usersWillBeDeleted}}{Number of users that will be deleted.}

\item{\code{usersWillNotBeDeleted}}{Number of users that will not be deleted.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersPermadeletePreviewStatistics-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UsersPermadeletePreviewStatistics-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersPermadeletePreviewStatistics$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersPermadeletePreviewStatistics-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersPermadeletePreviewStatistics-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersPermadeletePreviewStatistics$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersPermadeletePreviewStatistics-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersPermadeletePreviewStatistics-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersPermadeletePreviewStatistics$fromJSON(
  UsersPermadeletePreviewStatisticsJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UsersPermadeletePreviewStatisticsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersPermadeletePreviewStatistics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UsersPermadeletePreviewStatistics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersPermadeletePreviewStatistics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
