% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_permadelete_summary_report_statistics.R
\docType{class}
\name{UsersPermadeleteSummaryReportStatistics}
\alias{UsersPermadeleteSummaryReportStatistics}
\title{UsersPermadeleteSummaryReportStatistics}
\format{
An \code{R6Class} generator object
}
\description{
UsersPermadeleteSummaryReportStatistics Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{deploymentsAssignedNewOwner}}{integer Number of deployments that are assigned a new owner.}

\item{\code{deploymentsDeleted}}{integer Number of deployments that are deleted.}

\item{\code{deploymentsNotDeleted}}{integer Number of deployments that are not deleted.}

\item{\code{errors}}{integer Number of errors encountered during users deletion.}

\item{\code{projectsAssignedNewOwner}}{integer Number of projects that are assigned a new owner.}

\item{\code{projectsDeleted}}{integer Number of projects that are deleted.}

\item{\code{projectsNotDeleted}}{integer Number of projects that are not deleted.}

\item{\code{usersDeleted}}{integer Number of users that are deleted.}

\item{\code{usersNotDeleted}}{integer Number of users that are not deleted.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UsersPermadeleteSummaryReportStatistics$new()}}
\item \href{#method-validate}{\code{UsersPermadeleteSummaryReportStatistics$validate()}}
\item \href{#method-toJSON}{\code{UsersPermadeleteSummaryReportStatistics$toJSON()}}
\item \href{#method-fromJSON}{\code{UsersPermadeleteSummaryReportStatistics$fromJSON()}}
\item \href{#method-clone}{\code{UsersPermadeleteSummaryReportStatistics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersPermadeleteSummaryReportStatistics$new(
  deploymentsAssignedNewOwner = NULL,
  deploymentsDeleted = NULL,
  deploymentsNotDeleted = NULL,
  errors = NULL,
  projectsAssignedNewOwner = NULL,
  projectsDeleted = NULL,
  projectsNotDeleted = NULL,
  usersDeleted = NULL,
  usersNotDeleted = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deploymentsAssignedNewOwner}}{Number of deployments that are assigned a new owner.}

\item{\code{deploymentsDeleted}}{Number of deployments that are deleted.}

\item{\code{deploymentsNotDeleted}}{Number of deployments that are not deleted.}

\item{\code{errors}}{Number of errors encountered during users deletion.}

\item{\code{projectsAssignedNewOwner}}{Number of projects that are assigned a new owner.}

\item{\code{projectsDeleted}}{Number of projects that are deleted.}

\item{\code{projectsNotDeleted}}{Number of projects that are not deleted.}

\item{\code{usersDeleted}}{Number of users that are deleted.}

\item{\code{usersNotDeleted}}{Number of users that are not deleted.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersPermadeleteSummaryReportStatistics$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersPermadeleteSummaryReportStatistics$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersPermadeleteSummaryReportStatistics$fromJSON(
  UsersPermadeleteSummaryReportStatisticsJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UsersPermadeleteSummaryReportStatisticsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersPermadeleteSummaryReportStatistics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
