% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments_apiWrapper.R
\name{ValidateReplaceDeployedModel}
\alias{ValidateReplaceDeployedModel}
\title{ValidateReplaceDeployedModel}
\usage{
ValidateReplaceDeployedModel(deploymentId, newModelId, ...)
}
\arguments{
\item{...}{Additional keyword arguments to be passed on to the \code{datarobot.apicore} and \code{httr} libraries.}
}
\description{
Model Replacement Validation.
}
\details{
Validate that a model can be used to replace the current model of the deployment.

This method invokes \verb{POST /deployments/\{deploymentId\}/model/validation/} in the DataRobot Public API.
}
\seealso{
The method DeploymentsModelValidationCreate in \link{DeploymentsApi}, which invokes the same underlying API endpoint.
}
