% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_retrieve_response.R
\docType{class}
\name{VersionRetrieveResponse}
\alias{VersionRetrieveResponse}
\title{VersionRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
VersionRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{major}}{integer The major version number.}

\item{\code{minor}}{integer The minor version number.}

\item{\code{versionString}}{character The full version string.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VersionRetrieveResponse-new}{\code{VersionRetrieveResponse$new()}}
\item \href{#method-VersionRetrieveResponse-validate}{\code{VersionRetrieveResponse$validate()}}
\item \href{#method-VersionRetrieveResponse-toJSON}{\code{VersionRetrieveResponse$toJSON()}}
\item \href{#method-VersionRetrieveResponse-fromJSON}{\code{VersionRetrieveResponse$fromJSON()}}
\item \href{#method-VersionRetrieveResponse-clone}{\code{VersionRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VersionRetrieveResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-VersionRetrieveResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionRetrieveResponse$new(
  major = NULL,
  minor = NULL,
  versionString = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{major}}{The major version number.}

\item{\code{minor}}{The minor version number.}

\item{\code{versionString}}{The full version string.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VersionRetrieveResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-VersionRetrieveResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VersionRetrieveResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-VersionRetrieveResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VersionRetrieveResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-VersionRetrieveResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionRetrieveResponse$fromJSON(
  VersionRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{VersionRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VersionRetrieveResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VersionRetrieveResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VersionRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
