% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_cloud_ngram.R
\docType{class}
\name{WordCloudNgram}
\alias{WordCloudNgram}
\title{WordCloudNgram}
\format{
An \code{R6Class} generator object
}
\description{
WordCloudNgram Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{class}}{character \link{optional} For classification - values of the target class for corresponding word or ngram. For regression - null.}

\item{\code{coefficient}}{numeric Describes effect of this ngram on the target. A large negative value means a strong effect toward the negative class in classification projects and a smaller predicted target value in regression projects. A large positive value means a strong effect toward the positive class and a larger predicted target value respectively.}

\item{\code{count}}{integer Number of rows in the training sample where this ngram appears.}

\item{\code{frequency}}{numeric Frequency of this ngram relative to the most common ngram.}

\item{\code{isStopword}}{character True for ngrams that DataRobot evaluates as stopwords.}

\item{\code{ngram}}{character Word or ngram value.}

\item{\code{variable}}{character String representation of the ngram source - contains column name and, for some models, preprocessing details. E.g. NGRAM_OCCUR_L2_colname will be for ngram occurrences count using L2 normalization from the colname column.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WordCloudNgram$new()}}
\item \href{#method-validate}{\code{WordCloudNgram$validate()}}
\item \href{#method-toJSON}{\code{WordCloudNgram$toJSON()}}
\item \href{#method-fromJSON}{\code{WordCloudNgram$fromJSON()}}
\item \href{#method-clone}{\code{WordCloudNgram$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WordCloudNgram$new(
  coefficient = NULL,
  count = NULL,
  frequency = NULL,
  isStopword = NULL,
  ngram = NULL,
  variable = NULL,
  class = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coefficient}}{Describes effect of this ngram on the target. A large negative value means a strong effect toward the negative class in classification projects and a smaller predicted target value in regression projects. A large positive value means a strong effect toward the positive class and a larger predicted target value respectively.}

\item{\code{count}}{Number of rows in the training sample where this ngram appears.}

\item{\code{frequency}}{Frequency of this ngram relative to the most common ngram.}

\item{\code{isStopword}}{True for ngrams that DataRobot evaluates as stopwords.}

\item{\code{ngram}}{Word or ngram value.}

\item{\code{variable}}{String representation of the ngram source - contains column name and, for some models, preprocessing details. E.g. NGRAM_OCCUR_L2_colname will be for ngram occurrences count using L2 normalization from the colname column.}

\item{\code{class}}{For classification - values of the target class for corresponding word or ngram. For regression - null.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WordCloudNgram$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WordCloudNgram$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WordCloudNgram$fromJSON(WordCloudNgramJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{WordCloudNgramJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WordCloudNgram$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
