% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_source_dataset_with_creds_response.R
\docType{class}
\name{WorkspaceSourceDatasetWithCredsResponse}
\alias{WorkspaceSourceDatasetWithCredsResponse}
\title{WorkspaceSourceDatasetWithCredsResponse}
\format{
An \code{R6Class} generator object
}
\description{
WorkspaceSourceDatasetWithCredsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{alias}}{character Alias to be used as the table name.}

\item{\code{datasetId}}{character ID of a dataset in the catalog.}

\item{\code{datasetVersionId}}{character ID of a dataset version in the catalog.}

\item{\code{needsCredentials}}{character Whether a user must provide credentials for source datasets.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WorkspaceSourceDatasetWithCredsResponse$new()}}
\item \href{#method-validate}{\code{WorkspaceSourceDatasetWithCredsResponse$validate()}}
\item \href{#method-toJSON}{\code{WorkspaceSourceDatasetWithCredsResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{WorkspaceSourceDatasetWithCredsResponse$fromJSON()}}
\item \href{#method-clone}{\code{WorkspaceSourceDatasetWithCredsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkspaceSourceDatasetWithCredsResponse$new(
  alias = NULL,
  datasetId = NULL,
  datasetVersionId = NULL,
  needsCredentials = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alias}}{Alias to be used as the table name.}

\item{\code{datasetId}}{ID of a dataset in the catalog.}

\item{\code{datasetVersionId}}{ID of a dataset version in the catalog.}

\item{\code{needsCredentials}}{Whether a user must provide credentials for source datasets.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkspaceSourceDatasetWithCredsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkspaceSourceDatasetWithCredsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkspaceSourceDatasetWithCredsResponse$fromJSON(
  WorkspaceSourceDatasetWithCredsResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{WorkspaceSourceDatasetWithCredsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkspaceSourceDatasetWithCredsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
