% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/workspace_state_created_from_query_generator_response.R
\docType{class}
\name{WorkspaceStateCreatedFromQueryGeneratorResponse}
\alias{WorkspaceStateCreatedFromQueryGeneratorResponse}
\title{WorkspaceStateCreatedFromQueryGeneratorResponse}
\format{
An \code{R6Class} generator object
}
\description{
WorkspaceStateCreatedFromQueryGeneratorResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datasets}}{list( \link{DataEngineDataset} ) Source datasets in the Data Engine workspace.}

\item{\code{language}}{character Language of the Data Engine query.}

\item{\code{query}}{character Actual body of the Data Engine query.}

\item{\code{queryGeneratorId}}{character Query generator id.}

\item{\code{workspaceStateId}}{character Data Engine workspace state ID.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WorkspaceStateCreatedFromQueryGeneratorResponse-new}{\code{WorkspaceStateCreatedFromQueryGeneratorResponse$new()}}
\item \href{#method-WorkspaceStateCreatedFromQueryGeneratorResponse-validate}{\code{WorkspaceStateCreatedFromQueryGeneratorResponse$validate()}}
\item \href{#method-WorkspaceStateCreatedFromQueryGeneratorResponse-toJSON}{\code{WorkspaceStateCreatedFromQueryGeneratorResponse$toJSON()}}
\item \href{#method-WorkspaceStateCreatedFromQueryGeneratorResponse-fromJSON}{\code{WorkspaceStateCreatedFromQueryGeneratorResponse$fromJSON()}}
\item \href{#method-WorkspaceStateCreatedFromQueryGeneratorResponse-clone}{\code{WorkspaceStateCreatedFromQueryGeneratorResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkspaceStateCreatedFromQueryGeneratorResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-WorkspaceStateCreatedFromQueryGeneratorResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkspaceStateCreatedFromQueryGeneratorResponse$new(
  datasets = NULL,
  language = NULL,
  query = NULL,
  queryGeneratorId = NULL,
  workspaceStateId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasets}}{Source datasets in the Data Engine workspace.}

\item{\code{language}}{Language of the Data Engine query.}

\item{\code{query}}{Actual body of the Data Engine query.}

\item{\code{queryGeneratorId}}{Query generator id.}

\item{\code{workspaceStateId}}{Data Engine workspace state ID.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkspaceStateCreatedFromQueryGeneratorResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-WorkspaceStateCreatedFromQueryGeneratorResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkspaceStateCreatedFromQueryGeneratorResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkspaceStateCreatedFromQueryGeneratorResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-WorkspaceStateCreatedFromQueryGeneratorResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkspaceStateCreatedFromQueryGeneratorResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkspaceStateCreatedFromQueryGeneratorResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-WorkspaceStateCreatedFromQueryGeneratorResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkspaceStateCreatedFromQueryGeneratorResponse$fromJSON(
  WorkspaceStateCreatedFromQueryGeneratorResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{WorkspaceStateCreatedFromQueryGeneratorResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkspaceStateCreatedFromQueryGeneratorResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WorkspaceStateCreatedFromQueryGeneratorResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkspaceStateCreatedFromQueryGeneratorResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
