% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_state_response.R
\docType{class}
\name{WorkspaceStateResponse}
\alias{WorkspaceStateResponse}
\title{WorkspaceStateResponse}
\format{
An \code{R6Class} generator object
}
\description{
WorkspaceStateResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datasets}}{list( \link{WorkspaceSourceDatasetWithCredsResponse} ) Source datasets in the Data Engine workspace.}

\item{\code{language}}{character Language of the Data Engine query.}

\item{\code{query}}{character Actual body of the Data Engine query.}

\item{\code{queryGeneratorId}}{character Query generator id.}

\item{\code{runTime}}{numeric Data Engine query execution time.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WorkspaceStateResponse$new()}}
\item \href{#method-validate}{\code{WorkspaceStateResponse$validate()}}
\item \href{#method-toJSON}{\code{WorkspaceStateResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{WorkspaceStateResponse$fromJSON()}}
\item \href{#method-clone}{\code{WorkspaceStateResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkspaceStateResponse$new(
  datasets = NULL,
  language = NULL,
  query = NULL,
  queryGeneratorId = NULL,
  runTime = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasets}}{Source datasets in the Data Engine workspace.}

\item{\code{language}}{Language of the Data Engine query.}

\item{\code{query}}{Actual body of the Data Engine query.}

\item{\code{queryGeneratorId}}{Query generator id.}

\item{\code{runTime}}{Data Engine query execution time.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkspaceStateResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkspaceStateResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkspaceStateResponse$fromJSON(
  WorkspaceStateResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{WorkspaceStateResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkspaceStateResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
