% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_insight_table_data.R
\docType{class}
\name{AnomalyInsightTableData}
\alias{AnomalyInsightTableData}
\title{AnomalyInsightTableData}
\format{
An \code{R6Class} generator object
}
\description{
AnomalyInsightTableData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{columns}}{list( character ) array of columns that contain columns from training dataset and `anomalyScore` column.}

\item{\code{data}}{list( numeric ) array of arrays with actual data. Order in each array corresponds to order in columns array.}

\item{\code{rowId}}{list( integer ) index 0-based array. Each rowId corresponds to the actual row number of training data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnomalyInsightTableData-new}{\code{AnomalyInsightTableData$new()}}
\item \href{#method-AnomalyInsightTableData-validate}{\code{AnomalyInsightTableData$validate()}}
\item \href{#method-AnomalyInsightTableData-toJSON}{\code{AnomalyInsightTableData$toJSON()}}
\item \href{#method-AnomalyInsightTableData-fromJSON}{\code{AnomalyInsightTableData$fromJSON()}}
\item \href{#method-AnomalyInsightTableData-clone}{\code{AnomalyInsightTableData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyInsightTableData-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyInsightTableData-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyInsightTableData$new(
  columns = NULL,
  data = NULL,
  rowId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{array of columns that contain columns from training dataset and `anomalyScore` column.}

\item{\code{data}}{array of arrays with actual data. Order in each array corresponds to order in columns array.}

\item{\code{rowId}}{index 0-based array. Each rowId corresponds to the actual row number of training data}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyInsightTableData-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyInsightTableData-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyInsightTableData$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyInsightTableData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyInsightTableData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyInsightTableData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyInsightTableData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyInsightTableData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyInsightTableData$fromJSON(
  AnomalyInsightTableDataJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AnomalyInsightTableDataJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnomalyInsightTableData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnomalyInsightTableData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnomalyInsightTableData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
