% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_create_response.R
\docType{class}
\name{ApplicationCreateResponse}
\alias{ApplicationCreateResponse}
\title{ApplicationCreateResponse}
\format{
An \code{R6Class} generator object
}
\description{
ApplicationCreateResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{applicationTypeId}}{character The ID of the type of the application}

\item{\code{createdAt}}{character The timestamp when the application was created}

\item{\code{createdBy}}{character The username of who created the application}

\item{\code{deactivationStatusId}}{character The ID of the status object to track the asynchronous app deactivation process status. Will be null if the app was never deactivated.}

\item{\code{deploymentIds}}{list( character ) The list of model deployment IDs associated with the application}

\item{\code{deploymentStatusId}}{character The ID of the status object to track the asynchronous deployment process status.}

\item{\code{id}}{character The ID of the created application}

\item{\code{modelDeploymentId}}{character The ID of the associated model deployment}

\item{\code{name}}{character The name of the application}

\item{\code{permissions}}{character The list of permitted actions, which the authenticated user can perform on this application.}

\item{\code{updatedAt}}{character The timestamp when the application was updated}

\item{\code{userId}}{character The ID of the user which created the application}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApplicationCreateResponse-new}{\code{ApplicationCreateResponse$new()}}
\item \href{#method-ApplicationCreateResponse-validate}{\code{ApplicationCreateResponse$validate()}}
\item \href{#method-ApplicationCreateResponse-toJSON}{\code{ApplicationCreateResponse$toJSON()}}
\item \href{#method-ApplicationCreateResponse-fromJSON}{\code{ApplicationCreateResponse$fromJSON()}}
\item \href{#method-ApplicationCreateResponse-clone}{\code{ApplicationCreateResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationCreateResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationCreateResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationCreateResponse$new(
  applicationTypeId = NULL,
  createdAt = NULL,
  createdBy = NULL,
  deactivationStatusId = NULL,
  deploymentIds = NULL,
  deploymentStatusId = NULL,
  id = NULL,
  modelDeploymentId = NULL,
  name = NULL,
  permissions = NULL,
  updatedAt = NULL,
  userId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{applicationTypeId}}{The ID of the type of the application}

\item{\code{createdAt}}{The timestamp when the application was created}

\item{\code{createdBy}}{The username of who created the application}

\item{\code{deactivationStatusId}}{The ID of the status object to track the asynchronous app deactivation process status. Will be null if the app was never deactivated.}

\item{\code{deploymentIds}}{The list of model deployment IDs associated with the application}

\item{\code{deploymentStatusId}}{The ID of the status object to track the asynchronous deployment process status.}

\item{\code{id}}{The ID of the created application}

\item{\code{modelDeploymentId}}{The ID of the associated model deployment}

\item{\code{name}}{The name of the application}

\item{\code{permissions}}{The list of permitted actions, which the authenticated user can perform on this application.}

\item{\code{updatedAt}}{The timestamp when the application was updated}

\item{\code{userId}}{The ID of the user which created the application}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationCreateResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationCreateResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationCreateResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationCreateResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationCreateResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationCreateResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationCreateResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationCreateResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationCreateResponse$fromJSON(
  ApplicationCreateResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ApplicationCreateResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationCreateResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationCreateResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationCreateResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
