% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_doc_item.R
\docType{class}
\name{AutomatedDocItem}
\alias{AutomatedDocItem}
\title{AutomatedDocItem}
\format{
An \code{R6Class} generator object
}
\description{
AutomatedDocItem Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{createdAt}}{character Timestamp for the document creation time.}

\item{\code{docTypeSpecificInfo}}{\link{DocTypeSpecificInfo} \link{optional}}

\item{\code{documentType}}{character Type of the generated document.}

\item{\code{entityId}}{character Unique identifier of the entity the document was generated for.}

\item{\code{id}}{character Unique identifier of the generated document.}

\item{\code{locale}}{character \link{optional} Locale of the generated document.}

\item{\code{outputFormat}}{character File format of the generated document.}

\item{\code{templateId}}{character Unique identifier of the template used for the document outline.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AutomatedDocItem$new()}}
\item \href{#method-validate}{\code{AutomatedDocItem$validate()}}
\item \href{#method-toJSON}{\code{AutomatedDocItem$toJSON()}}
\item \href{#method-fromJSON}{\code{AutomatedDocItem$fromJSON()}}
\item \href{#method-clone}{\code{AutomatedDocItem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedDocItem$new(
  createdAt = NULL,
  documentType = NULL,
  entityId = NULL,
  id = NULL,
  outputFormat = NULL,
  templateId = NULL,
  docTypeSpecificInfo = NULL,
  locale = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{createdAt}}{Timestamp for the document creation time.}

\item{\code{documentType}}{Type of the generated document.}

\item{\code{entityId}}{Unique identifier of the entity the document was generated for.}

\item{\code{id}}{Unique identifier of the generated document.}

\item{\code{outputFormat}}{File format of the generated document.}

\item{\code{templateId}}{Unique identifier of the template used for the document outline.}

\item{\code{locale}}{Locale of the generated document.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedDocItem$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedDocItem$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedDocItem$fromJSON(AutomatedDocItemJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AutomatedDocItemJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutomatedDocItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
