% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blender_info_retrieve_response.R
\docType{class}
\name{BlenderInfoRetrieveResponse}
\alias{BlenderInfoRetrieveResponse}
\title{BlenderInfoRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
BlenderInfoRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{blendable}}{character If True, the models can be blended.}

\item{\code{reason}}{character Useful info as to why a model can't be blended.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BlenderInfoRetrieveResponse-new}{\code{BlenderInfoRetrieveResponse$new()}}
\item \href{#method-BlenderInfoRetrieveResponse-validate}{\code{BlenderInfoRetrieveResponse$validate()}}
\item \href{#method-BlenderInfoRetrieveResponse-toJSON}{\code{BlenderInfoRetrieveResponse$toJSON()}}
\item \href{#method-BlenderInfoRetrieveResponse-fromJSON}{\code{BlenderInfoRetrieveResponse$fromJSON()}}
\item \href{#method-BlenderInfoRetrieveResponse-clone}{\code{BlenderInfoRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlenderInfoRetrieveResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-BlenderInfoRetrieveResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlenderInfoRetrieveResponse$new(
  blendable = NULL,
  reason = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blendable}}{If True, the models can be blended.}

\item{\code{reason}}{Useful info as to why a model can't be blended.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlenderInfoRetrieveResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BlenderInfoRetrieveResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlenderInfoRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlenderInfoRetrieveResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BlenderInfoRetrieveResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlenderInfoRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlenderInfoRetrieveResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BlenderInfoRetrieveResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlenderInfoRetrieveResponse$fromJSON(
  BlenderInfoRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{BlenderInfoRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlenderInfoRetrieveResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BlenderInfoRetrieveResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlenderInfoRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
