% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_user_role_record_response.R
\docType{class}
\name{CalendarUserRoleRecordResponse}
\alias{CalendarUserRoleRecordResponse}
\title{CalendarUserRoleRecordResponse}
\format{
An \code{R6Class} generator object
}
\description{
CalendarUserRoleRecordResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{canShare}}{character Whether this user can share this calendar with other users.}

\item{\code{role}}{character The role of the user on this calendar.}

\item{\code{userId}}{character The ID of the user.}

\item{\code{username}}{character The username of a user with access to this calendar.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CalendarUserRoleRecordResponse-new}{\code{CalendarUserRoleRecordResponse$new()}}
\item \href{#method-CalendarUserRoleRecordResponse-validate}{\code{CalendarUserRoleRecordResponse$validate()}}
\item \href{#method-CalendarUserRoleRecordResponse-toJSON}{\code{CalendarUserRoleRecordResponse$toJSON()}}
\item \href{#method-CalendarUserRoleRecordResponse-fromJSON}{\code{CalendarUserRoleRecordResponse$fromJSON()}}
\item \href{#method-CalendarUserRoleRecordResponse-clone}{\code{CalendarUserRoleRecordResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarUserRoleRecordResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarUserRoleRecordResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarUserRoleRecordResponse$new(
  canShare = NULL,
  role = NULL,
  userId = NULL,
  username = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{canShare}}{Whether this user can share this calendar with other users.}

\item{\code{role}}{The role of the user on this calendar.}

\item{\code{userId}}{The ID of the user.}

\item{\code{username}}{The username of a user with access to this calendar.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarUserRoleRecordResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarUserRoleRecordResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarUserRoleRecordResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarUserRoleRecordResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarUserRoleRecordResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarUserRoleRecordResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarUserRoleRecordResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarUserRoleRecordResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarUserRoleRecordResponse$fromJSON(
  CalendarUserRoleRecordResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CalendarUserRoleRecordResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalendarUserRoleRecordResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CalendarUserRoleRecordResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalendarUserRoleRecordResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
