% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/challenger_create.R
\docType{class}
\name{ChallengerCreate}
\alias{ChallengerCreate}
\title{ChallengerCreate}
\format{
An \code{R6Class} generator object
}
\description{
ChallengerCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{modelPackageId}}{character ID of the model package to add as a challenger.}

\item{\code{name}}{character Human-readable name for the challenger.}

\item{\code{predictionEnvironmentId}}{character \link{optional} ID of the Prediction Environment the challenger should use. If prediction environments are enabled, this is required}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChallengerCreate-new}{\code{ChallengerCreate$new()}}
\item \href{#method-ChallengerCreate-validate}{\code{ChallengerCreate$validate()}}
\item \href{#method-ChallengerCreate-toJSON}{\code{ChallengerCreate$toJSON()}}
\item \href{#method-ChallengerCreate-fromJSON}{\code{ChallengerCreate$fromJSON()}}
\item \href{#method-ChallengerCreate-clone}{\code{ChallengerCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChallengerCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-ChallengerCreate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChallengerCreate$new(
  modelPackageId = NULL,
  name = NULL,
  predictionEnvironmentId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{modelPackageId}}{ID of the model package to add as a challenger.}

\item{\code{name}}{Human-readable name for the challenger.}

\item{\code{predictionEnvironmentId}}{ID of the Prediction Environment the challenger should use. If prediction environments are enabled, this is required}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChallengerCreate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ChallengerCreate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChallengerCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChallengerCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ChallengerCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChallengerCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChallengerCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ChallengerCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChallengerCreate$fromJSON(ChallengerCreateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ChallengerCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChallengerCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChallengerCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChallengerCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
