% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_user.R
\docType{class}
\name{CommentUser}
\alias{CommentUser}
\title{CommentUser}
\format{
An \code{R6Class} generator object
}
\description{
CommentUser Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{firstName}}{character First name of the commenter}

\item{\code{id}}{character User ID of the commenter}

\item{\code{lastName}}{character Last name of the commenter}

\item{\code{username}}{character Username of the commenter}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CommentUser-new}{\code{CommentUser$new()}}
\item \href{#method-CommentUser-validate}{\code{CommentUser$validate()}}
\item \href{#method-CommentUser-toJSON}{\code{CommentUser$toJSON()}}
\item \href{#method-CommentUser-fromJSON}{\code{CommentUser$fromJSON()}}
\item \href{#method-CommentUser-clone}{\code{CommentUser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommentUser-new"></a>}}
\if{latex}{\out{\hypertarget{method-CommentUser-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentUser$new(
  firstName = NULL,
  id = NULL,
  lastName = NULL,
  username = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{firstName}}{First name of the commenter}

\item{\code{id}}{User ID of the commenter}

\item{\code{lastName}}{Last name of the commenter}

\item{\code{username}}{Username of the commenter}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommentUser-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CommentUser-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentUser$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommentUser-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CommentUser-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentUser$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommentUser-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CommentUser-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentUser$fromJSON(CommentUserJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CommentUserJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommentUser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CommentUser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentUser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
