% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_get_and_list_combined_model.R
\docType{class}
\name{CommonGetAndListCombinedModel}
\alias{CommonGetAndListCombinedModel}
\title{CommonGetAndListCombinedModel}
\format{
An \code{R6Class} generator object
}
\description{
CommonGetAndListCombinedModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{combinedModelId}}{character The ID of combined model.}

\item{\code{isActiveCombinedModel}}{character Indicates whether this model is the active one in segmented modeling project.}

\item{\code{modelCategory}}{character Indicates what kind of model this is. Will be ``combined`` for combined models.}

\item{\code{projectId}}{character The ID of the project.}

\item{\code{segmentationTaskId}}{character The ID of the segmentation task used to generate this combined model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CommonGetAndListCombinedModel-new}{\code{CommonGetAndListCombinedModel$new()}}
\item \href{#method-CommonGetAndListCombinedModel-validate}{\code{CommonGetAndListCombinedModel$validate()}}
\item \href{#method-CommonGetAndListCombinedModel-toJSON}{\code{CommonGetAndListCombinedModel$toJSON()}}
\item \href{#method-CommonGetAndListCombinedModel-fromJSON}{\code{CommonGetAndListCombinedModel$fromJSON()}}
\item \href{#method-CommonGetAndListCombinedModel-clone}{\code{CommonGetAndListCombinedModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommonGetAndListCombinedModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-CommonGetAndListCombinedModel-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommonGetAndListCombinedModel$new(
  combinedModelId = NULL,
  isActiveCombinedModel = NULL,
  modelCategory = NULL,
  projectId = NULL,
  segmentationTaskId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{combinedModelId}}{The ID of combined model.}

\item{\code{isActiveCombinedModel}}{Indicates whether this model is the active one in segmented modeling project.}

\item{\code{modelCategory}}{Indicates what kind of model this is. Will be ``combined`` for combined models.}

\item{\code{projectId}}{The ID of the project.}

\item{\code{segmentationTaskId}}{The ID of the segmentation task used to generate this combined model.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommonGetAndListCombinedModel-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CommonGetAndListCombinedModel-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommonGetAndListCombinedModel$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommonGetAndListCombinedModel-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CommonGetAndListCombinedModel-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommonGetAndListCombinedModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommonGetAndListCombinedModel-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CommonGetAndListCombinedModel-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommonGetAndListCombinedModel$fromJSON(
  CommonGetAndListCombinedModelJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CommonGetAndListCombinedModelJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommonGetAndListCombinedModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CommonGetAndListCombinedModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommonGetAndListCombinedModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
