% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_workspace_state_from_query_generator.R
\docType{class}
\name{CreateWorkspaceStateFromQueryGenerator}
\alias{CreateWorkspaceStateFromQueryGenerator}
\title{CreateWorkspaceStateFromQueryGenerator}
\format{
An \code{R6Class} generator object
}
\description{
CreateWorkspaceStateFromQueryGenerator Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datasetId}}{character \link{optional} The ID of the dataset.}

\item{\code{datasetVersionId}}{character \link{optional} The ID of the dataset version.}

\item{\code{queryGeneratorId}}{character The ID of the query generator.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CreateWorkspaceStateFromQueryGenerator$new()}}
\item \href{#method-validate}{\code{CreateWorkspaceStateFromQueryGenerator$validate()}}
\item \href{#method-toJSON}{\code{CreateWorkspaceStateFromQueryGenerator$toJSON()}}
\item \href{#method-fromJSON}{\code{CreateWorkspaceStateFromQueryGenerator$fromJSON()}}
\item \href{#method-clone}{\code{CreateWorkspaceStateFromQueryGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateWorkspaceStateFromQueryGenerator$new(
  queryGeneratorId = NULL,
  datasetId = NULL,
  datasetVersionId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{queryGeneratorId}}{The ID of the query generator.}

\item{\code{datasetId}}{The ID of the dataset.}

\item{\code{datasetVersionId}}{The ID of the dataset version.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateWorkspaceStateFromQueryGenerator$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateWorkspaceStateFromQueryGenerator$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateWorkspaceStateFromQueryGenerator$fromJSON(
  CreateWorkspaceStateFromQueryGeneratorJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CreateWorkspaceStateFromQueryGeneratorJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateWorkspaceStateFromQueryGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
