% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential_id.R
\docType{class}
\name{CredentialId}
\alias{CredentialId}
\title{CredentialId}
\format{
An \code{R6Class} generator object
}
\description{
CredentialId Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalogVersionId}}{character \link{optional} The ID of the latest version of the catalog entry.}

\item{\code{credentialId}}{character The ID of the set of credentials to use instead of user and password. Note that with this change, username and password will become optional.}

\item{\code{url}}{character \link{optional} The link to retrieve more detailed information about the entity that uses this catalog dataset.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CredentialId-new}{\code{CredentialId$new()}}
\item \href{#method-CredentialId-validate}{\code{CredentialId$validate()}}
\item \href{#method-CredentialId-toJSON}{\code{CredentialId$toJSON()}}
\item \href{#method-CredentialId-fromJSON}{\code{CredentialId$fromJSON()}}
\item \href{#method-CredentialId-clone}{\code{CredentialId$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialId-new"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialId-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialId$new(
  credentialId = NULL,
  catalogVersionId = NULL,
  url = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{credentialId}}{The ID of the set of credentials to use instead of user and password. Note that with this change, username and password will become optional.}

\item{\code{catalogVersionId}}{The ID of the latest version of the catalog entry.}

\item{\code{url}}{The link to retrieve more detailed information about the entity that uses this catalog dataset.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialId-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialId-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialId$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialId-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialId-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialId$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialId-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialId-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialId$fromJSON(CredentialIdJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CredentialIdJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialId-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialId-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialId$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
