% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_body.R
\docType{class}
\name{CredentialsBody}
\alias{CredentialsBody}
\title{CredentialsBody}
\format{
An \code{R6Class} generator object
}
\description{
CredentialsBody Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{credentialType}}{character \link{optional} Credentials type.}

\item{\code{description}}{character \link{optional} Credentials description.}

\item{\code{name}}{character Credentials name.}

\item{\code{password}}{character Password to store for this credentials.}

\item{\code{snowflakeAccountName}}{character \link{optional} Snowflake account name.}

\item{\code{user}}{character Username to store for this credentials.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CredentialsBody-new}{\code{CredentialsBody$new()}}
\item \href{#method-CredentialsBody-validate}{\code{CredentialsBody$validate()}}
\item \href{#method-CredentialsBody-toJSON}{\code{CredentialsBody$toJSON()}}
\item \href{#method-CredentialsBody-fromJSON}{\code{CredentialsBody$fromJSON()}}
\item \href{#method-CredentialsBody-clone}{\code{CredentialsBody$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsBody-new"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsBody-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsBody$new(
  name = NULL,
  password = NULL,
  user = NULL,
  credentialType = NULL,
  description = NULL,
  snowflakeAccountName = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Credentials name.}

\item{\code{password}}{Password to store for this credentials.}

\item{\code{user}}{Username to store for this credentials.}

\item{\code{credentialType}}{Credentials type.}

\item{\code{description}}{Credentials description.}

\item{\code{snowflakeAccountName}}{Snowflake account name.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsBody-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsBody-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsBody$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsBody-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsBody-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsBody$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsBody-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsBody-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsBody$fromJSON(CredentialsBodyJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CredentialsBodyJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CredentialsBody-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CredentialsBody-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CredentialsBody$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
