% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cross_series_group_by_column_retrieve_response.R
\docType{class}
\name{CrossSeriesGroupByColumnRetrieveResponse}
\alias{CrossSeriesGroupByColumnRetrieveResponse}
\title{CrossSeriesGroupByColumnRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
CrossSeriesGroupByColumnRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{crossSeriesGroupByColumns}}{list( \link{CrossSeriesGroupByColumnsListItem} ) A list of columns with information about each column's eligibility as a cross-series group-by column.}

\item{\code{multiseriesId}}{character The name of the multiseries ID column.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CrossSeriesGroupByColumnRetrieveResponse$new()}}
\item \href{#method-validate}{\code{CrossSeriesGroupByColumnRetrieveResponse$validate()}}
\item \href{#method-toJSON}{\code{CrossSeriesGroupByColumnRetrieveResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{CrossSeriesGroupByColumnRetrieveResponse$fromJSON()}}
\item \href{#method-clone}{\code{CrossSeriesGroupByColumnRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSeriesGroupByColumnRetrieveResponse$new(
  crossSeriesGroupByColumns = NULL,
  multiseriesId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{crossSeriesGroupByColumns}}{A list of columns with information about each column's eligibility as a cross-series group-by column.}

\item{\code{multiseriesId}}{The name of the multiseries ID column.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSeriesGroupByColumnRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSeriesGroupByColumnRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSeriesGroupByColumnRetrieveResponse$fromJSON(
  CrossSeriesGroupByColumnRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CrossSeriesGroupByColumnRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSeriesGroupByColumnRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
