% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_model_resource_limits.R
\docType{class}
\name{CustomModelResourceLimits}
\alias{CustomModelResourceLimits}
\title{CustomModelResourceLimits}
\format{
An \code{R6Class} generator object
}
\description{
CustomModelResourceLimits Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{desiredCustomModelContainerSize}}{integer \link{optional} The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{maxCustomModelContainerSize}}{integer \link{optional} The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{maxCustomModelReplicasPerDeployment}}{integer \link{optional} A fixed number of replicas that will be set for the given custom-model.}

\item{\code{maxCustomModelTestingParallelUsers}}{integer The maximum number of parallel users that can be used for Custom Model Testing checks}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CustomModelResourceLimits-new}{\code{CustomModelResourceLimits$new()}}
\item \href{#method-CustomModelResourceLimits-validate}{\code{CustomModelResourceLimits$validate()}}
\item \href{#method-CustomModelResourceLimits-toJSON}{\code{CustomModelResourceLimits$toJSON()}}
\item \href{#method-CustomModelResourceLimits-fromJSON}{\code{CustomModelResourceLimits$fromJSON()}}
\item \href{#method-CustomModelResourceLimits-clone}{\code{CustomModelResourceLimits$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomModelResourceLimits-new"></a>}}
\if{latex}{\out{\hypertarget{method-CustomModelResourceLimits-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelResourceLimits$new(
  maxCustomModelTestingParallelUsers = NULL,
  desiredCustomModelContainerSize = NULL,
  maxCustomModelContainerSize = NULL,
  maxCustomModelReplicasPerDeployment = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maxCustomModelTestingParallelUsers}}{The maximum number of parallel users that can be used for Custom Model Testing checks}

\item{\code{desiredCustomModelContainerSize}}{The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{maxCustomModelContainerSize}}{The maximum memory that might be allocated by the custom-model. If exceeded, the custom-model will be killed}

\item{\code{maxCustomModelReplicasPerDeployment}}{A fixed number of replicas that will be set for the given custom-model.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomModelResourceLimits-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CustomModelResourceLimits-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelResourceLimits$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomModelResourceLimits-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomModelResourceLimits-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelResourceLimits$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomModelResourceLimits-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CustomModelResourceLimits-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelResourceLimits$fromJSON(
  CustomModelResourceLimitsJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CustomModelResourceLimitsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomModelResourceLimits-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CustomModelResourceLimits-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomModelResourceLimits$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
