% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_relationship_response.R
\docType{class}
\name{DatasetRelationshipResponse}
\alias{DatasetRelationshipResponse}
\title{DatasetRelationshipResponse}
\format{
An \code{R6Class} generator object
}
\description{
DatasetRelationshipResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{createdBy}}{character The username of the user that created this relationship.}

\item{\code{creationDate}}{character ISO-8601 formatted time/date that this record was created.}

\item{\code{id}}{character ID of the dataset relationship}

\item{\code{linkedDatasetId}}{character ID of the linked dataset.}

\item{\code{linkedFeatures}}{list( character ) List of features belonging to the linked dataset.}

\item{\code{modificationDate}}{character ISO-8601 formatted time/date that this record was last updated.}

\item{\code{modifiedBy}}{character The username of the user that modified this relationship most recently.}

\item{\code{sourceDatasetId}}{character ID of the source dataset.}

\item{\code{sourceFeatures}}{list( character ) List of features belonging to the source dataset.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DatasetRelationshipResponse$new()}}
\item \href{#method-validate}{\code{DatasetRelationshipResponse$validate()}}
\item \href{#method-toJSON}{\code{DatasetRelationshipResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{DatasetRelationshipResponse$fromJSON()}}
\item \href{#method-clone}{\code{DatasetRelationshipResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRelationshipResponse$new(
  createdBy = NULL,
  creationDate = NULL,
  id = NULL,
  linkedDatasetId = NULL,
  linkedFeatures = NULL,
  modificationDate = NULL,
  modifiedBy = NULL,
  sourceDatasetId = NULL,
  sourceFeatures = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{createdBy}}{The username of the user that created this relationship.}

\item{\code{creationDate}}{ISO-8601 formatted time/date that this record was created.}

\item{\code{id}}{ID of the dataset relationship}

\item{\code{linkedDatasetId}}{ID of the linked dataset.}

\item{\code{linkedFeatures}}{List of features belonging to the linked dataset.}

\item{\code{modificationDate}}{ISO-8601 formatted time/date that this record was last updated.}

\item{\code{modifiedBy}}{The username of the user that modified this relationship most recently.}

\item{\code{sourceDatasetId}}{ID of the source dataset.}

\item{\code{sourceFeatures}}{List of features belonging to the source dataset.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRelationshipResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRelationshipResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRelationshipResponse$fromJSON(
  DatasetRelationshipResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatasetRelationshipResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetRelationshipResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
