% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source_retrieve_response.R
\docType{class}
\name{DataSourceRetrieveResponse}
\alias{DataSourceRetrieveResponse}
\title{DataSourceRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
DataSourceRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{canonicalName}}{character Data source canonical name.}

\item{\code{creator}}{character ID of the user who created the data source.}

\item{\code{id}}{character Data source ID.}

\item{\code{params}}{\link{OneOfJDBCTableDataSourceJDBCQueryDataSourceDRConnectorV1DataSource} Data source configuration.}

\item{\code{role}}{character Role of the user making the request on this data source.}

\item{\code{type}}{character Data source type.}

\item{\code{updated}}{character ISO 8601-formatted date/time of the data source update.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataSourceRetrieveResponse$new()}}
\item \href{#method-validate}{\code{DataSourceRetrieveResponse$validate()}}
\item \href{#method-toJSON}{\code{DataSourceRetrieveResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{DataSourceRetrieveResponse$fromJSON()}}
\item \href{#method-clone}{\code{DataSourceRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSourceRetrieveResponse$new(
  canonicalName = NULL,
  creator = NULL,
  id = NULL,
  params = NULL,
  role = NULL,
  type = NULL,
  updated = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{canonicalName}}{Data source canonical name.}

\item{\code{creator}}{ID of the user who created the data source.}

\item{\code{id}}{Data source ID.}

\item{\code{params}}{Data source configuration.}

\item{\code{role}}{Role of the user making the request on this data source.}

\item{\code{type}}{Data source type.}

\item{\code{updated}}{ISO 8601-formatted date/time of the data source update.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSourceRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSourceRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSourceRetrieveResponse$fromJSON(
  DataSourceRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DataSourceRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSourceRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
