% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_trend_plots_preview_response.R
\docType{class}
\name{DatetimeTrendPlotsPreviewResponse}
\alias{DatetimeTrendPlotsPreviewResponse}
\title{DatetimeTrendPlotsPreviewResponse}
\format{
An \code{R6Class} generator object
}
\description{
DatetimeTrendPlotsPreviewResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bins}}{list( \link{DatetimeTrendPlotsPreviewBins} ) An array of bins for the retrieved plots.}

\item{\code{endDate}}{character The datetime of the end of the chart data (exclusive).}

\item{\code{startDate}}{character The datetime of the start of the chart data (inclusive).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DatetimeTrendPlotsPreviewResponse$new()}}
\item \href{#method-validate}{\code{DatetimeTrendPlotsPreviewResponse$validate()}}
\item \href{#method-toJSON}{\code{DatetimeTrendPlotsPreviewResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{DatetimeTrendPlotsPreviewResponse$fromJSON()}}
\item \href{#method-clone}{\code{DatetimeTrendPlotsPreviewResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatetimeTrendPlotsPreviewResponse$new(
  bins = NULL,
  endDate = NULL,
  startDate = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bins}}{An array of bins for the retrieved plots.}

\item{\code{endDate}}{The datetime of the end of the chart data (exclusive).}

\item{\code{startDate}}{The datetime of the start of the chart data (inclusive).}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatetimeTrendPlotsPreviewResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatetimeTrendPlotsPreviewResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatetimeTrendPlotsPreviewResponse$fromJSON(
  DatetimeTrendPlotsPreviewResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DatetimeTrendPlotsPreviewResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatetimeTrendPlotsPreviewResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
