% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_cleanup_stats_change.R
\docType{class}
\name{DeploymentCleanupStatsChange}
\alias{DeploymentCleanupStatsChange}
\title{DeploymentCleanupStatsChange}
\format{
An \code{R6Class} generator object
}
\description{
DeploymentCleanupStatsChange Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dataType}}{character Type of stats to cleanup.}

\item{\code{end}}{character If specified, the stats will be cleaned up to this timestamp. If ``null`` all stats till the deployment end forecast date will be cleaned up. Defaults to ``null``.}

\item{\code{modelId}}{character ID of the model to remove deployment stats for. If ``null``, the stats will be cleaned up for all models in the mathcing period. Defaults to ``null``.}

\item{\code{start}}{character If specified, the stats will be cleaned up from this timestamp. If ``null`` all stats from the deployment start forecast date will be cleaned up. Defaults to ``null``.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeploymentCleanupStatsChange-new}{\code{DeploymentCleanupStatsChange$new()}}
\item \href{#method-DeploymentCleanupStatsChange-validate}{\code{DeploymentCleanupStatsChange$validate()}}
\item \href{#method-DeploymentCleanupStatsChange-toJSON}{\code{DeploymentCleanupStatsChange$toJSON()}}
\item \href{#method-DeploymentCleanupStatsChange-fromJSON}{\code{DeploymentCleanupStatsChange$fromJSON()}}
\item \href{#method-DeploymentCleanupStatsChange-clone}{\code{DeploymentCleanupStatsChange$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentCleanupStatsChange-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentCleanupStatsChange-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentCleanupStatsChange$new(
  dataType = NULL,
  end = NULL,
  modelId = NULL,
  start = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataType}}{Type of stats to cleanup.}

\item{\code{end}}{If specified, the stats will be cleaned up to this timestamp. If ``null`` all stats till the deployment end forecast date will be cleaned up. Defaults to ``null``.}

\item{\code{modelId}}{ID of the model to remove deployment stats for. If ``null``, the stats will be cleaned up for all models in the mathcing period. Defaults to ``null``.}

\item{\code{start}}{If specified, the stats will be cleaned up from this timestamp. If ``null`` all stats from the deployment start forecast date will be cleaned up. Defaults to ``null``.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentCleanupStatsChange-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentCleanupStatsChange-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentCleanupStatsChange$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentCleanupStatsChange-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentCleanupStatsChange-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentCleanupStatsChange$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentCleanupStatsChange-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentCleanupStatsChange-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentCleanupStatsChange$fromJSON(
  DeploymentCleanupStatsChangeJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DeploymentCleanupStatsChangeJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeploymentCleanupStatsChange-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeploymentCleanupStatsChange-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeploymentCleanupStatsChange$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
