% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.setPrimitiveProperty}
\alias{.setPrimitiveProperty}
\title{Validate and set primitive types}
\usage{
.setPrimitiveProperty(typeList, propertyData)
}
\arguments{
\item{typeList}{A vector of R6 class generator names.}

\item{propertyData}{A deserialized JSON string.}
}
\value{
The value of \code{propertyData} if the value is one of the types listed in \code{typeList}.
A helper method to validate simple types for anyOf and oneOf properties in an R6 object.
}
\description{
A helper method to validate simple types for anyOf and oneOf properties in an R6 object.
}
