% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_connector_v1_field.R
\docType{class}
\name{DRConnectorV1Field}
\alias{DRConnectorV1Field}
\title{DRConnectorV1Field}
\format{
An \code{R6Class} generator object
}
\description{
DRConnectorV1Field Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character \link{optional} Connector field name.}

\item{\code{name}}{character User-friendly displayable connector field name.}

\item{\code{value}}{character Connector field value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DRConnectorV1Field-new}{\code{DRConnectorV1Field$new()}}
\item \href{#method-DRConnectorV1Field-validate}{\code{DRConnectorV1Field$validate()}}
\item \href{#method-DRConnectorV1Field-toJSON}{\code{DRConnectorV1Field$toJSON()}}
\item \href{#method-DRConnectorV1Field-fromJSON}{\code{DRConnectorV1Field$fromJSON()}}
\item \href{#method-DRConnectorV1Field-clone}{\code{DRConnectorV1Field$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DRConnectorV1Field-new"></a>}}
\if{latex}{\out{\hypertarget{method-DRConnectorV1Field-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DRConnectorV1Field$new(
  name = NULL,
  value = NULL,
  id = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{User-friendly displayable connector field name.}

\item{\code{value}}{Connector field value.}

\item{\code{id}}{Connector field name.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DRConnectorV1Field-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DRConnectorV1Field-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DRConnectorV1Field$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DRConnectorV1Field-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DRConnectorV1Field-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DRConnectorV1Field$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DRConnectorV1Field-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DRConnectorV1Field-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DRConnectorV1Field$fromJSON(DRConnectorV1FieldJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DRConnectorV1FieldJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DRConnectorV1Field-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DRConnectorV1Field-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DRConnectorV1Field$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
