% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_user_id.R
\docType{class}
\name{DrUserId}
\alias{DrUserId}
\title{DrUserId}
\format{
An \code{R6Class} generator object
}
\description{
DrUserId Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{firstName}}{character \link{optional} First name of use case owner}

\item{\code{id}}{character The DR user ID}

\item{\code{lastName}}{character \link{optional} Last of use case owner}

\item{\code{username}}{character \link{optional} Username of use case owner}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DrUserId-new}{\code{DrUserId$new()}}
\item \href{#method-DrUserId-validate}{\code{DrUserId$validate()}}
\item \href{#method-DrUserId-toJSON}{\code{DrUserId$toJSON()}}
\item \href{#method-DrUserId-fromJSON}{\code{DrUserId$fromJSON()}}
\item \href{#method-DrUserId-clone}{\code{DrUserId$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrUserId-new"></a>}}
\if{latex}{\out{\hypertarget{method-DrUserId-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrUserId$new(
  id = NULL,
  firstName = NULL,
  lastName = NULL,
  username = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DR user ID}

\item{\code{firstName}}{First name of use case owner}

\item{\code{lastName}}{Last of use case owner}

\item{\code{username}}{Username of use case owner}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrUserId-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DrUserId-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrUserId$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrUserId-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DrUserId-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrUserId$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrUserId-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DrUserId-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrUserId$fromJSON(DrUserIdJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DrUserIdJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrUserId-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DrUserId-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrUserId$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
