% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption_keypairs.R
\docType{class}
\name{EncryptionKeypairs}
\alias{EncryptionKeypairs}
\title{EncryptionKeypairs}
\format{
An \code{R6Class} generator object
}
\description{
EncryptionKeypairs Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{certUnderscorefile}}{character \link{optional} Path to the pem file with a single certificate.}

\item{\code{certUnderscorefileUnderscorevalue}}{character \link{optional} A single certificate pem file content as a single string. Has priority over cert_file.}

\item{\code{keyUnderscorefile}}{character \link{optional} Path to the private key pem file.}

\item{\code{keyUnderscorefileUnderscorevalue}}{character \link{optional} The private key pem file content as a single string. Has priority over key_file.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EncryptionKeypairs$new()}}
\item \href{#method-validate}{\code{EncryptionKeypairs$validate()}}
\item \href{#method-toJSON}{\code{EncryptionKeypairs$toJSON()}}
\item \href{#method-fromJSON}{\code{EncryptionKeypairs$fromJSON()}}
\item \href{#method-clone}{\code{EncryptionKeypairs$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EncryptionKeypairs$new(
  certUnderscorefile = NULL,
  certUnderscorefileUnderscorevalue = NULL,
  keyUnderscorefile = NULL,
  keyUnderscorefileUnderscorevalue = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{certUnderscorefile}}{Path to the pem file with a single certificate.}

\item{\code{certUnderscorefileUnderscorevalue}}{A single certificate pem file content as a single string. Has priority over cert_file.}

\item{\code{keyUnderscorefile}}{Path to the private key pem file.}

\item{\code{keyUnderscorefileUnderscorevalue}}{The private key pem file content as a single string. Has priority over key_file.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EncryptionKeypairs$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EncryptionKeypairs$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EncryptionKeypairs$fromJSON(EncryptionKeypairsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{EncryptionKeypairsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EncryptionKeypairs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
