% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eureqa_leaderboard_entry_payload.R
\docType{class}
\name{EureqaLeaderboardEntryPayload}
\alias{EureqaLeaderboardEntryPayload}
\title{EureqaLeaderboardEntryPayload}
\format{
An \code{R6Class} generator object
}
\description{
EureqaLeaderboardEntryPayload Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{parentModelId}}{character \link{optional} The ID of the model to clone from. If omitted, will automatically search for and find the first leaderboard model created by the blueprint run that also created the solution associated with `solutionId`.}

\item{\code{solutionId}}{character the ID of the solution to be cloned.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EureqaLeaderboardEntryPayload$new()}}
\item \href{#method-validate}{\code{EureqaLeaderboardEntryPayload$validate()}}
\item \href{#method-toJSON}{\code{EureqaLeaderboardEntryPayload$toJSON()}}
\item \href{#method-fromJSON}{\code{EureqaLeaderboardEntryPayload$fromJSON()}}
\item \href{#method-clone}{\code{EureqaLeaderboardEntryPayload$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EureqaLeaderboardEntryPayload$new(
  solutionId = NULL,
  parentModelId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{solutionId}}{the ID of the solution to be cloned.}

\item{\code{parentModelId}}{The ID of the model to clone from. If omitted, will automatically search for and find the first leaderboard model created by the blueprint run that also created the solution associated with `solutionId`.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EureqaLeaderboardEntryPayload$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EureqaLeaderboardEntryPayload$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EureqaLeaderboardEntryPayload$fromJSON(
  EureqaLeaderboardEntryPayloadJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{EureqaLeaderboardEntryPayloadJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EureqaLeaderboardEntryPayload$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
