% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\docType{class}
\name{Feature}
\alias{Feature}
\title{Feature}
\format{
An \code{R6Class} generator object
}
\description{
Feature Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dateFormat}}{character The date format string for how this feature was interpreted.}

\item{\code{featureType}}{character Feature type.}

\item{\code{importance}}{numeric Numeric measure of the relationship strength between the feature and target (independent of model or other features).}

\item{\code{knownInAdvance}}{character Whether the feature was selected as known in advance in a time-series model, false for non-time-series models.}

\item{\code{name}}{character Feature name.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Feature-new}{\code{Feature$new()}}
\item \href{#method-Feature-validate}{\code{Feature$validate()}}
\item \href{#method-Feature-toJSON}{\code{Feature$toJSON()}}
\item \href{#method-Feature-fromJSON}{\code{Feature$fromJSON()}}
\item \href{#method-Feature-clone}{\code{Feature$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-new"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$new(
  dateFormat = NULL,
  featureType = NULL,
  importance = NULL,
  knownInAdvance = NULL,
  name = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dateFormat}}{The date format string for how this feature was interpreted.}

\item{\code{featureType}}{Feature type.}

\item{\code{importance}}{Numeric measure of the relationship strength between the feature and target (independent of model or other features).}

\item{\code{knownInAdvance}}{Whether the feature was selected as known in advance in a time-series model, false for non-time-series models.}

\item{\code{name}}{Feature name.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$fromJSON(FeatureJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Feature-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Feature-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
