% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_flag_object.R
\docType{class}
\name{FeatureFlagObject}
\alias{FeatureFlagObject}
\title{FeatureFlagObject}
\format{
An \code{R6Class} generator object
}
\description{
FeatureFlagObject Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uiLabel}}{character \link{optional} String representation of the feature flag}

\item{\code{uiTooltip}}{character \link{optional} Detailed description of the feature flag}

\item{\code{value}}{character Value of the feature flag}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureFlagObject-new}{\code{FeatureFlagObject$new()}}
\item \href{#method-FeatureFlagObject-validate}{\code{FeatureFlagObject$validate()}}
\item \href{#method-FeatureFlagObject-toJSON}{\code{FeatureFlagObject$toJSON()}}
\item \href{#method-FeatureFlagObject-fromJSON}{\code{FeatureFlagObject$fromJSON()}}
\item \href{#method-FeatureFlagObject-clone}{\code{FeatureFlagObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureFlagObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureFlagObject-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureFlagObject$new(
  value = NULL,
  uiLabel = NULL,
  uiTooltip = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Value of the feature flag}

\item{\code{uiLabel}}{String representation of the feature flag}

\item{\code{uiTooltip}}{Detailed description of the feature flag}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureFlagObject-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureFlagObject-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureFlagObject$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureFlagObject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureFlagObject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureFlagObject$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureFlagObject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureFlagObject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureFlagObject$fromJSON(FeatureFlagObjectJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureFlagObjectJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureFlagObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureFlagObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureFlagObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
