% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_system_data_streamer.R
\docType{class}
\name{FileSystemDataStreamer}
\alias{FileSystemDataStreamer}
\title{FileSystemDataStreamer}
\format{
An \code{R6Class} generator object
}
\description{
FileSystemDataStreamer Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{character Path to data on host filesystem}

\item{\code{type}}{character Type name for this intake type}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FileSystemDataStreamer-new}{\code{FileSystemDataStreamer$new()}}
\item \href{#method-FileSystemDataStreamer-validate}{\code{FileSystemDataStreamer$validate()}}
\item \href{#method-FileSystemDataStreamer-toJSON}{\code{FileSystemDataStreamer$toJSON()}}
\item \href{#method-FileSystemDataStreamer-fromJSON}{\code{FileSystemDataStreamer$fromJSON()}}
\item \href{#method-FileSystemDataStreamer-clone}{\code{FileSystemDataStreamer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileSystemDataStreamer-new"></a>}}
\if{latex}{\out{\hypertarget{method-FileSystemDataStreamer-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileSystemDataStreamer$new(
  path = NULL,
  type = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to data on host filesystem}

\item{\code{type}}{Type name for this intake type}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileSystemDataStreamer-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FileSystemDataStreamer-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileSystemDataStreamer$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileSystemDataStreamer-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FileSystemDataStreamer-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileSystemDataStreamer$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileSystemDataStreamer-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FileSystemDataStreamer-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileSystemDataStreamer$fromJSON(
  FileSystemDataStreamerJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FileSystemDataStreamerJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileSystemDataStreamer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FileSystemDataStreamer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileSystemDataStreamer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
