% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift_chartbins.R
\docType{class}
\name{LiftChartbins}
\alias{LiftChartbins}
\title{LiftChartbins}
\format{
An \code{R6Class} generator object
}
\description{
LiftChartbins Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual}}{numeric The average of the actual target values for the rows in the bin.}

\item{\code{binWeight}}{numeric How much data is in the bin. For projects with weights, it is the sum of the weights of all rows in the bins; otherwise, it is the number of rows in the bin.}

\item{\code{predicted}}{numeric The average of predicted values of the target for the rows in the bin.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LiftChartbins-new}{\code{LiftChartbins$new()}}
\item \href{#method-LiftChartbins-validate}{\code{LiftChartbins$validate()}}
\item \href{#method-LiftChartbins-toJSON}{\code{LiftChartbins$toJSON()}}
\item \href{#method-LiftChartbins-fromJSON}{\code{LiftChartbins$fromJSON()}}
\item \href{#method-LiftChartbins-clone}{\code{LiftChartbins$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LiftChartbins-new"></a>}}
\if{latex}{\out{\hypertarget{method-LiftChartbins-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LiftChartbins$new(
  actual = NULL,
  binWeight = NULL,
  predicted = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{actual}}{The average of the actual target values for the rows in the bin.}

\item{\code{binWeight}}{How much data is in the bin. For projects with weights, it is the sum of the weights of all rows in the bins; otherwise, it is the number of rows in the bin.}

\item{\code{predicted}}{The average of predicted values of the target for the rows in the bin.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LiftChartbins-validate"></a>}}
\if{latex}{\out{\hypertarget{method-LiftChartbins-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LiftChartbins$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LiftChartbins-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LiftChartbins-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LiftChartbins$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LiftChartbins-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LiftChartbins-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LiftChartbins$fromJSON(LiftChartbinsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{LiftChartbinsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LiftChartbins-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LiftChartbins-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LiftChartbins$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
