% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/model_confusion_chart_metadata_retrieve_response.R
\docType{class}
\name{ModelConfusionChartMetadataRetrieveResponse}
\alias{ModelConfusionChartMetadataRetrieveResponse}
\title{ModelConfusionChartMetadataRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
ModelConfusionChartMetadataRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{classNames}}{list( character ) List of all class names in the full confusion matrix, sorted by the `orderBy` parameter}

\item{\code{relevantClassesPositions}}{list( \link{array[integer]} ) Matrix to highlight important cell blocks in the confusion chart.  Intended to represent a thumbnail view, where the relevantClassesPositions array has a 1 in thumbnail cells that are of interest, and 0 otherwise.  The dimensions of the implied thumbnail will not match those of the confusion matrix, e.g. a twenty-class confusion matrix may have a 2x2 thumbnail.}

\item{\code{source}}{character Source of the chart.}

\item{\code{totalMatrixSum}}{integer Sum of all values in the full confusion matrix (equal to the number of points considered)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelConfusionChartMetadataRetrieveResponse-new}{\code{ModelConfusionChartMetadataRetrieveResponse$new()}}
\item \href{#method-ModelConfusionChartMetadataRetrieveResponse-validate}{\code{ModelConfusionChartMetadataRetrieveResponse$validate()}}
\item \href{#method-ModelConfusionChartMetadataRetrieveResponse-toJSON}{\code{ModelConfusionChartMetadataRetrieveResponse$toJSON()}}
\item \href{#method-ModelConfusionChartMetadataRetrieveResponse-fromJSON}{\code{ModelConfusionChartMetadataRetrieveResponse$fromJSON()}}
\item \href{#method-ModelConfusionChartMetadataRetrieveResponse-clone}{\code{ModelConfusionChartMetadataRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfusionChartMetadataRetrieveResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfusionChartMetadataRetrieveResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfusionChartMetadataRetrieveResponse$new(
  classNames = NULL,
  relevantClassesPositions = NULL,
  source = NULL,
  totalMatrixSum = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{classNames}}{List of all class names in the full confusion matrix, sorted by the `orderBy` parameter}

\item{\code{relevantClassesPositions}}{Matrix to highlight important cell blocks in the confusion chart.  Intended to represent a thumbnail view, where the relevantClassesPositions array has a 1 in thumbnail cells that are of interest, and 0 otherwise.  The dimensions of the implied thumbnail will not match those of the confusion matrix, e.g. a twenty-class confusion matrix may have a 2x2 thumbnail.}

\item{\code{source}}{Source of the chart.}

\item{\code{totalMatrixSum}}{Sum of all values in the full confusion matrix (equal to the number of points considered)}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfusionChartMetadataRetrieveResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfusionChartMetadataRetrieveResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfusionChartMetadataRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfusionChartMetadataRetrieveResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfusionChartMetadataRetrieveResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfusionChartMetadataRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfusionChartMetadataRetrieveResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfusionChartMetadataRetrieveResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfusionChartMetadataRetrieveResponse$fromJSON(
  ModelConfusionChartMetadataRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ModelConfusionChartMetadataRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelConfusionChartMetadataRetrieveResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelConfusionChartMetadataRetrieveResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfusionChartMetadataRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
