% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_feature_list_response.R
\docType{class}
\name{ModelFeatureListResponse}
\alias{ModelFeatureListResponse}
\title{ModelFeatureListResponse}
\format{
An \code{R6Class} generator object
}
\description{
ModelFeatureListResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aPrioriFeatureNames}}{list( character ) (Deprecated in version v2.11) Renamed to `knownInAdvanceFeatureNames`. This parameter always has the same value as `knownInAdvanceFeatureNames` and will be removed in a future release.}

\item{\code{featureNames}}{list( character ) An array of the names of all features used by the specified model.}

\item{\code{knownInAdvanceFeatureNames}}{list( character ) An array of the names of time series known-in-advance features used by the specified model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelFeatureListResponse-new}{\code{ModelFeatureListResponse$new()}}
\item \href{#method-ModelFeatureListResponse-validate}{\code{ModelFeatureListResponse$validate()}}
\item \href{#method-ModelFeatureListResponse-toJSON}{\code{ModelFeatureListResponse$toJSON()}}
\item \href{#method-ModelFeatureListResponse-fromJSON}{\code{ModelFeatureListResponse$fromJSON()}}
\item \href{#method-ModelFeatureListResponse-clone}{\code{ModelFeatureListResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelFeatureListResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelFeatureListResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFeatureListResponse$new(
  aPrioriFeatureNames = NULL,
  featureNames = NULL,
  knownInAdvanceFeatureNames = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aPrioriFeatureNames}}{(Deprecated in version v2.11) Renamed to `knownInAdvanceFeatureNames`. This parameter always has the same value as `knownInAdvanceFeatureNames` and will be removed in a future release.}

\item{\code{featureNames}}{An array of the names of all features used by the specified model.}

\item{\code{knownInAdvanceFeatureNames}}{An array of the names of time series known-in-advance features used by the specified model.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelFeatureListResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ModelFeatureListResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFeatureListResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelFeatureListResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelFeatureListResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFeatureListResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelFeatureListResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelFeatureListResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFeatureListResponse$fromJSON(
  ModelFeatureListResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ModelFeatureListResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelFeatureListResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelFeatureListResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFeatureListResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
