% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_feature_effect_create.R
\docType{class}
\name{MulticlassFeatureEffectCreate}
\alias{MulticlassFeatureEffectCreate}
\title{MulticlassFeatureEffectCreate}
\format{
An \code{R6Class} generator object
}
\description{
MulticlassFeatureEffectCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{features}}{list( character ) \link{optional} The list of features to use to calculate feature effects.}

\item{\code{rowCount}}{integer \link{optional} The number of rows from dataset to use for Feature Impact calculation.}

\item{\code{topNFeatures}}{integer \link{optional} Number of top features (ranked by feature impact) to use to calculate feature effects.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MulticlassFeatureEffectCreate$new()}}
\item \href{#method-validate}{\code{MulticlassFeatureEffectCreate$validate()}}
\item \href{#method-toJSON}{\code{MulticlassFeatureEffectCreate$toJSON()}}
\item \href{#method-fromJSON}{\code{MulticlassFeatureEffectCreate$fromJSON()}}
\item \href{#method-clone}{\code{MulticlassFeatureEffectCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureEffectCreate$new(
  features = NULL,
  rowCount = NULL,
  topNFeatures = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{features}}{The list of features to use to calculate feature effects.}

\item{\code{rowCount}}{The number of rows from dataset to use for Feature Impact calculation.}

\item{\code{topNFeatures}}{Number of top features (ranked by feature impact) to use to calculate feature effects.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureEffectCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureEffectCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureEffectCreate$fromJSON(
  MulticlassFeatureEffectCreateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MulticlassFeatureEffectCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MulticlassFeatureEffectCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
