% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilabel_insights_response.R
\docType{class}
\name{MultilabelInsightsResponse}
\alias{MultilabelInsightsResponse}
\title{MultilabelInsightsResponse}
\format{
An \code{R6Class} generator object
}
\description{
MultilabelInsightsResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{multilabelInsightsKey}}{character Key for multilabel insights, unique per project, feature, and EDA stage. The response will contain the key for the most recent, finished EDA stage.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultilabelInsightsResponse-new}{\code{MultilabelInsightsResponse$new()}}
\item \href{#method-MultilabelInsightsResponse-validate}{\code{MultilabelInsightsResponse$validate()}}
\item \href{#method-MultilabelInsightsResponse-toJSON}{\code{MultilabelInsightsResponse$toJSON()}}
\item \href{#method-MultilabelInsightsResponse-fromJSON}{\code{MultilabelInsightsResponse$fromJSON()}}
\item \href{#method-MultilabelInsightsResponse-clone}{\code{MultilabelInsightsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultilabelInsightsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultilabelInsightsResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultilabelInsightsResponse$new(
  multilabelInsightsKey = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{multilabelInsightsKey}}{Key for multilabel insights, unique per project, feature, and EDA stage. The response will contain the key for the most recent, finished EDA stage.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultilabelInsightsResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-MultilabelInsightsResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultilabelInsightsResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultilabelInsightsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MultilabelInsightsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultilabelInsightsResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultilabelInsightsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MultilabelInsightsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultilabelInsightsResponse$fromJSON(
  MultilabelInsightsResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MultilabelInsightsResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultilabelInsightsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultilabelInsightsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultilabelInsightsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
