% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notification_log_list_response.R
\docType{class}
\name{NotificationLogListResponse}
\alias{NotificationLogListResponse}
\title{NotificationLogListResponse}
\format{
An \code{R6Class} generator object
}
\description{
NotificationLogListResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer Number of items returned on this page}

\item{\code{data}}{list( \link{NotificationLogListDataResponse} ) An array notification logs}

\item{\code{next_}}{character URL pointing to the next page (if null, there is no next page)}

\item{\code{previous}}{character URL pointing to the previous page (if null, there is no previous page)}

\item{\code{totalCount}}{integer \link{optional} The total number of notifications}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NotificationLogListResponse-new}{\code{NotificationLogListResponse$new()}}
\item \href{#method-NotificationLogListResponse-validate}{\code{NotificationLogListResponse$validate()}}
\item \href{#method-NotificationLogListResponse-toJSON}{\code{NotificationLogListResponse$toJSON()}}
\item \href{#method-NotificationLogListResponse-fromJSON}{\code{NotificationLogListResponse$fromJSON()}}
\item \href{#method-NotificationLogListResponse-clone}{\code{NotificationLogListResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationLogListResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationLogListResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationLogListResponse$new(
  count = NULL,
  data = NULL,
  next_ = NULL,
  previous = NULL,
  totalCount = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{Number of items returned on this page}

\item{\code{data}}{An array notification logs}

\item{\code{next_}}{URL pointing to the next page (if null, there is no next page)}

\item{\code{previous}}{URL pointing to the previous page (if null, there is no previous page)}

\item{\code{totalCount}}{The total number of notifications}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationLogListResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationLogListResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationLogListResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationLogListResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationLogListResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationLogListResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationLogListResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationLogListResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationLogListResponse$fromJSON(
  NotificationLogListResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{NotificationLogListResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotificationLogListResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NotificationLogListResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotificationLogListResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
