% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_iterations_trained_response.R
\docType{class}
\name{NumIterationsTrainedResponse}
\alias{NumIterationsTrainedResponse}
\title{NumIterationsTrainedResponse}
\format{
An \code{R6Class} generator object
}
\description{
NumIterationsTrainedResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{NumIterationsTrainedData} ) Number of estimators or iterations for a single model stage}

\item{\code{modelId}}{character The model ID}

\item{\code{projectId}}{character The project ID}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NumIterationsTrainedResponse-new}{\code{NumIterationsTrainedResponse$new()}}
\item \href{#method-NumIterationsTrainedResponse-validate}{\code{NumIterationsTrainedResponse$validate()}}
\item \href{#method-NumIterationsTrainedResponse-toJSON}{\code{NumIterationsTrainedResponse$toJSON()}}
\item \href{#method-NumIterationsTrainedResponse-fromJSON}{\code{NumIterationsTrainedResponse$fromJSON()}}
\item \href{#method-NumIterationsTrainedResponse-clone}{\code{NumIterationsTrainedResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumIterationsTrainedResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-NumIterationsTrainedResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NumIterationsTrainedResponse$new(
  data = NULL,
  modelId = NULL,
  projectId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Number of estimators or iterations for a single model stage}

\item{\code{modelId}}{The model ID}

\item{\code{projectId}}{The project ID}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumIterationsTrainedResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-NumIterationsTrainedResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NumIterationsTrainedResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumIterationsTrainedResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NumIterationsTrainedResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NumIterationsTrainedResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumIterationsTrainedResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NumIterationsTrainedResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NumIterationsTrainedResponse$fromJSON(
  NumIterationsTrainedResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{NumIterationsTrainedResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NumIterationsTrainedResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NumIterationsTrainedResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NumIterationsTrainedResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
