% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_jdbc_table_data_source.R
\docType{class}
\name{OptionalJDBCTableDataSource}
\alias{OptionalJDBCTableDataSource}
\title{OptionalJDBCTableDataSource}
\format{
An \code{R6Class} generator object
}
\description{
OptionalJDBCTableDataSource Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalog}}{character \link{optional} Catalog name in the database if supported.}

\item{\code{dataStoreId}}{character \link{optional} Data store ID for this data source.}

\item{\code{fetchSize}}{integer \link{optional} User-specified fetch size.}

\item{\code{partitionColumn}}{character \link{optional} The name of the partition column. It is needed to allow parallel execution for the 10GB+ projects.}

\item{\code{schema}}{character \link{optional} Schema associated with the table or view in the database if supported.}

\item{\code{table}}{character \link{optional} Table or view name in the database.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OptionalJDBCTableDataSource$new()}}
\item \href{#method-validate}{\code{OptionalJDBCTableDataSource$validate()}}
\item \href{#method-toJSON}{\code{OptionalJDBCTableDataSource$toJSON()}}
\item \href{#method-fromJSON}{\code{OptionalJDBCTableDataSource$fromJSON()}}
\item \href{#method-clone}{\code{OptionalJDBCTableDataSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionalJDBCTableDataSource$new(
  catalog = NULL,
  dataStoreId = NULL,
  fetchSize = NULL,
  partitionColumn = NULL,
  schema = NULL,
  table = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalog}}{Catalog name in the database if supported.}

\item{\code{dataStoreId}}{Data store ID for this data source.}

\item{\code{fetchSize}}{User-specified fetch size.}

\item{\code{partitionColumn}}{The name of the partition column. It is needed to allow parallel execution for the 10GB+ projects.}

\item{\code{schema}}{Schema associated with the table or view in the database if supported.}

\item{\code{table}}{Table or view name in the database.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionalJDBCTableDataSource$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionalJDBCTableDataSource$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionalJDBCTableDataSource$fromJSON(
  OptionalJDBCTableDataSourceJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{OptionalJDBCTableDataSourceJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptionalJDBCTableDataSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
