% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization_user_patch.R
\docType{class}
\name{OrganizationUserPatch}
\alias{OrganizationUserPatch}
\title{OrganizationUserPatch}
\format{
An \code{R6Class} generator object
}
\description{
OrganizationUserPatch Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{accessRoleId}}{character \link{optional} \link{DEPRECATED - Use accessRoleIds instead} ID of the access role assigned to the user}

\item{\code{accessRoleIds}}{list( character ) \link{optional} List of access role ids assigned to the user}

\item{\code{activated}}{character \link{optional} Whether the organization user is activated.}

\item{\code{expirationDate}}{character \link{optional} User expiration date}

\item{\code{maxWorkers}}{integer \link{optional} User max workers}

\item{\code{orgAdmin}}{character \link{optional} Mark user as an organizational admin}

\item{\code{organizationId}}{character \link{optional} The organization to move the user to}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OrganizationUserPatch$new()}}
\item \href{#method-validate}{\code{OrganizationUserPatch$validate()}}
\item \href{#method-toJSON}{\code{OrganizationUserPatch$toJSON()}}
\item \href{#method-fromJSON}{\code{OrganizationUserPatch$fromJSON()}}
\item \href{#method-clone}{\code{OrganizationUserPatch$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUserPatch$new(
  accessRoleId = NULL,
  accessRoleIds = NULL,
  activated = NULL,
  expirationDate = NULL,
  maxWorkers = NULL,
  orgAdmin = NULL,
  organizationId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{accessRoleId}}{\link{DEPRECATED - Use accessRoleIds instead} ID of the access role assigned to the user}

\item{\code{accessRoleIds}}{List of access role ids assigned to the user}

\item{\code{activated}}{Whether the organization user is activated.}

\item{\code{expirationDate}}{User expiration date}

\item{\code{maxWorkers}}{User max workers}

\item{\code{orgAdmin}}{Mark user as an organizational admin}

\item{\code{organizationId}}{The organization to move the user to}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUserPatch$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUserPatch$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUserPatch$fromJSON(
  OrganizationUserPatchJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{OrganizationUserPatchJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrganizationUserPatch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
