% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependence_data.R
\docType{class}
\name{PartialDependenceData}
\alias{PartialDependenceData}
\title{PartialDependenceData}
\format{
An \code{R6Class} generator object
}
\description{
PartialDependenceData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dependence}}{numeric The value of partial dependence.}

\item{\code{label}}{character Contains the label for categorical and numeric features as string.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PartialDependenceData-new}{\code{PartialDependenceData$new()}}
\item \href{#method-PartialDependenceData-validate}{\code{PartialDependenceData$validate()}}
\item \href{#method-PartialDependenceData-toJSON}{\code{PartialDependenceData$toJSON()}}
\item \href{#method-PartialDependenceData-fromJSON}{\code{PartialDependenceData$fromJSON()}}
\item \href{#method-PartialDependenceData-clone}{\code{PartialDependenceData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialDependenceData-new"></a>}}
\if{latex}{\out{\hypertarget{method-PartialDependenceData-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialDependenceData$new(
  dependence = NULL,
  label = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dependence}}{The value of partial dependence.}

\item{\code{label}}{Contains the label for categorical and numeric features as string.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialDependenceData-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PartialDependenceData-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialDependenceData$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialDependenceData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PartialDependenceData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialDependenceData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialDependenceData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PartialDependenceData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialDependenceData$fromJSON(
  PartialDependenceDataJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PartialDependenceDataJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialDependenceData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PartialDependenceData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialDependenceData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
