% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoff_matrices_create.R
\docType{class}
\name{PayoffMatricesCreate}
\alias{PayoffMatricesCreate}
\title{PayoffMatricesCreate}
\format{
An \code{R6Class} generator object
}
\description{
PayoffMatricesCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{falseNegativeValue}}{numeric False negative value to use for profit curve calculation.}

\item{\code{falsePositiveValue}}{numeric False positive value to use for profit curve calculation.}

\item{\code{name}}{character Name of the payoff matrix to be created.}

\item{\code{trueNegativeValue}}{numeric True negative value to use for profit curve calculation.}

\item{\code{truePositiveValue}}{numeric True positive value to use for profit curve calculation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PayoffMatricesCreate-new}{\code{PayoffMatricesCreate$new()}}
\item \href{#method-PayoffMatricesCreate-validate}{\code{PayoffMatricesCreate$validate()}}
\item \href{#method-PayoffMatricesCreate-toJSON}{\code{PayoffMatricesCreate$toJSON()}}
\item \href{#method-PayoffMatricesCreate-fromJSON}{\code{PayoffMatricesCreate$fromJSON()}}
\item \href{#method-PayoffMatricesCreate-clone}{\code{PayoffMatricesCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PayoffMatricesCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-PayoffMatricesCreate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PayoffMatricesCreate$new(
  falseNegativeValue = NULL,
  falsePositiveValue = NULL,
  name = NULL,
  trueNegativeValue = NULL,
  truePositiveValue = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{falseNegativeValue}}{False negative value to use for profit curve calculation.}

\item{\code{falsePositiveValue}}{False positive value to use for profit curve calculation.}

\item{\code{name}}{Name of the payoff matrix to be created.}

\item{\code{trueNegativeValue}}{True negative value to use for profit curve calculation.}

\item{\code{truePositiveValue}}{True positive value to use for profit curve calculation.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PayoffMatricesCreate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PayoffMatricesCreate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PayoffMatricesCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PayoffMatricesCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PayoffMatricesCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PayoffMatricesCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PayoffMatricesCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PayoffMatricesCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PayoffMatricesCreate$fromJSON(PayoffMatricesCreateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PayoffMatricesCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PayoffMatricesCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PayoffMatricesCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PayoffMatricesCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
