% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_class_fairness.R
\docType{class}
\name{PerClassFairness}
\alias{PerClassFairness}
\title{PerClassFairness}
\format{
An \code{R6Class} generator object
}
\description{
PerClassFairness Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{absoluteValue}}{numeric The absolute fairness score for the class.}

\item{\code{className}}{character Name of the protected class the score is calculated for.}

\item{\code{entriesCount}}{integer The number of entries of the class in the analysed data.}

\item{\code{isStatisticallySignificant}}{character Flag to tell whether the score can be treated as statistically significant. In other words, whether we are confident enough with the score for this protected class.}

\item{\code{value}}{numeric The relative fairness score for the class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PerClassFairness-new}{\code{PerClassFairness$new()}}
\item \href{#method-PerClassFairness-validate}{\code{PerClassFairness$validate()}}
\item \href{#method-PerClassFairness-toJSON}{\code{PerClassFairness$toJSON()}}
\item \href{#method-PerClassFairness-fromJSON}{\code{PerClassFairness$fromJSON()}}
\item \href{#method-PerClassFairness-clone}{\code{PerClassFairness$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerClassFairness-new"></a>}}
\if{latex}{\out{\hypertarget{method-PerClassFairness-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerClassFairness$new(
  absoluteValue = NULL,
  className = NULL,
  entriesCount = NULL,
  isStatisticallySignificant = NULL,
  value = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{absoluteValue}}{The absolute fairness score for the class.}

\item{\code{className}}{Name of the protected class the score is calculated for.}

\item{\code{entriesCount}}{The number of entries of the class in the analysed data.}

\item{\code{isStatisticallySignificant}}{Flag to tell whether the score can be treated as statistically significant. In other words, whether we are confident enough with the score for this protected class.}

\item{\code{value}}{The relative fairness score for the class.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerClassFairness-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PerClassFairness-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerClassFairness$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerClassFairness-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PerClassFairness-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerClassFairness$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerClassFairness-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PerClassFairness-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerClassFairness$fromJSON(PerClassFairnessJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PerClassFairnessJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PerClassFairness-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PerClassFairness-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PerClassFairness$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
