% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicted_vs_actual_data.R
\docType{class}
\name{PredictedVsActualData}
\alias{PredictedVsActualData}
\title{PredictedVsActualData}
\format{
An \code{R6Class} generator object
}
\description{
PredictedVsActualData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual}}{numeric Actual value. `null` for 0 rows bins and unsupervised time series models.}

\item{\code{bin}}{list( character ) \link{optional} For numeric features contains labels for left and right bin limits}

\item{\code{label}}{character Contains label for categorical features; For numeric features contains range or numeric value.}

\item{\code{predicted}}{numeric Predicted value. `null` for 0 rows bins.}

\item{\code{rowCount}}{integer Number of rows for the label and bin.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictedVsActualData-new}{\code{PredictedVsActualData$new()}}
\item \href{#method-PredictedVsActualData-validate}{\code{PredictedVsActualData$validate()}}
\item \href{#method-PredictedVsActualData-toJSON}{\code{PredictedVsActualData$toJSON()}}
\item \href{#method-PredictedVsActualData-fromJSON}{\code{PredictedVsActualData$fromJSON()}}
\item \href{#method-PredictedVsActualData-clone}{\code{PredictedVsActualData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictedVsActualData-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictedVsActualData-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictedVsActualData$new(
  actual = NULL,
  label = NULL,
  predicted = NULL,
  rowCount = NULL,
  bin = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{actual}}{Actual value. `null` for 0 rows bins and unsupervised time series models.}

\item{\code{label}}{Contains label for categorical features; For numeric features contains range or numeric value.}

\item{\code{predicted}}{Predicted value. `null` for 0 rows bins.}

\item{\code{rowCount}}{Number of rows for the label and bin.}

\item{\code{bin}}{For numeric features contains labels for left and right bin limits}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictedVsActualData-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PredictedVsActualData-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictedVsActualData$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictedVsActualData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictedVsActualData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictedVsActualData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictedVsActualData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictedVsActualData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictedVsActualData$fromJSON(
  PredictedVsActualDataJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictedVsActualDataJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictedVsActualData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictedVsActualData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictedVsActualData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
