% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_explanations_object.R
\docType{class}
\name{PredictionExplanationsObject}
\alias{PredictionExplanationsObject}
\title{PredictionExplanationsObject}
\format{
An \code{R6Class} generator object
}
\description{
PredictionExplanationsObject Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{feature}}{character The name of the feature contributing to the prediction.}

\item{\code{featureValue}}{\link{OneOfintegerbooleanstringnumber} The value the feature took on for this row. The type corresponds to the feature (bool, int, float, str, etc.).}

\item{\code{label}}{\link{OneOfstringnumber} Describes what output was driven by this prediction explanation. For regression projects, it is the name of the target feature. For classification projects, it is the class whose probability increasing would correspond to a positive strength of this prediction explanation. For predictions made using anomaly detection models, it is the `Anomaly Score`.}

\item{\code{strength}}{numeric \link{optional} Algorithm-specific explanation value attributed to `feature` in this row. If `explanationAlgorithm` = `shap`, this is the SHAP value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictionExplanationsObject-new}{\code{PredictionExplanationsObject$new()}}
\item \href{#method-PredictionExplanationsObject-validate}{\code{PredictionExplanationsObject$validate()}}
\item \href{#method-PredictionExplanationsObject-toJSON}{\code{PredictionExplanationsObject$toJSON()}}
\item \href{#method-PredictionExplanationsObject-fromJSON}{\code{PredictionExplanationsObject$fromJSON()}}
\item \href{#method-PredictionExplanationsObject-clone}{\code{PredictionExplanationsObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsObject-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsObject$new(
  feature = NULL,
  featureValue = NULL,
  label = NULL,
  strength = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{The name of the feature contributing to the prediction.}

\item{\code{featureValue}}{The value the feature took on for this row. The type corresponds to the feature (bool, int, float, str, etc.).}

\item{\code{label}}{Describes what output was driven by this prediction explanation. For regression projects, it is the name of the target feature. For classification projects, it is the class whose probability increasing would correspond to a positive strength of this prediction explanation. For predictions made using anomaly detection models, it is the `Anomaly Score`.}

\item{\code{strength}}{Algorithm-specific explanation value attributed to `feature` in this row. If `explanationAlgorithm` = `shap`, this is the SHAP value.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsObject-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsObject-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsObject$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsObject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsObject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsObject$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsObject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsObject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsObject$fromJSON(
  PredictionExplanationsObjectJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionExplanationsObjectJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExplanationsObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExplanationsObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExplanationsObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
