% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_intervals.R
\docType{class}
\name{PredictionIntervals}
\alias{PredictionIntervals}
\title{PredictionIntervals}
\format{
An \code{R6Class} generator object
}
\description{
PredictionIntervals Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{enabled}}{character Indicates whether prediction intervals are enabled for this deployment.}

\item{\code{percentiles}}{list( integer ) \link{optional} The percentiles used for this deployment. Currently, we support at most one percentile at a time.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictionIntervals-new}{\code{PredictionIntervals$new()}}
\item \href{#method-PredictionIntervals-validate}{\code{PredictionIntervals$validate()}}
\item \href{#method-PredictionIntervals-toJSON}{\code{PredictionIntervals$toJSON()}}
\item \href{#method-PredictionIntervals-fromJSON}{\code{PredictionIntervals$fromJSON()}}
\item \href{#method-PredictionIntervals-clone}{\code{PredictionIntervals$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionIntervals-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionIntervals-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionIntervals$new(
  enabled = NULL,
  percentiles = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enabled}}{Indicates whether prediction intervals are enabled for this deployment.}

\item{\code{percentiles}}{The percentiles used for this deployment. Currently, we support at most one percentile at a time.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionIntervals-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionIntervals-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionIntervals$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionIntervals-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionIntervals-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionIntervals$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionIntervals-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionIntervals-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionIntervals$fromJSON(PredictionIntervalsJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PredictionIntervalsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionIntervals-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionIntervals-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionIntervals$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
