% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime_file_create.R
\docType{class}
\name{PrimeFileCreate}
\alias{PrimeFileCreate}
\title{PrimeFileCreate}
\format{
An \code{R6Class} generator object
}
\description{
PrimeFileCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{language}}{character The desired language of the generated code}

\item{\code{modelId}}{character the Prime model to generate code for}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PrimeFileCreate-new}{\code{PrimeFileCreate$new()}}
\item \href{#method-PrimeFileCreate-validate}{\code{PrimeFileCreate$validate()}}
\item \href{#method-PrimeFileCreate-toJSON}{\code{PrimeFileCreate$toJSON()}}
\item \href{#method-PrimeFileCreate-fromJSON}{\code{PrimeFileCreate$fromJSON()}}
\item \href{#method-PrimeFileCreate-clone}{\code{PrimeFileCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimeFileCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-PrimeFileCreate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimeFileCreate$new(
  language = NULL,
  modelId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{language}}{The desired language of the generated code}

\item{\code{modelId}}{the Prime model to generate code for}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimeFileCreate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PrimeFileCreate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimeFileCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimeFileCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PrimeFileCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimeFileCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimeFileCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PrimeFileCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimeFileCreate$fromJSON(PrimeFileCreateJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PrimeFileCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimeFileCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PrimeFileCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimeFileCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
