% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime_file_response.R
\docType{class}
\name{PrimeFileResponse}
\alias{PrimeFileResponse}
\title{PrimeFileResponse}
\format{
An \code{R6Class} generator object
}
\description{
PrimeFileResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character the ID of the file}

\item{\code{isValid}}{character whether the code passed basic validation checks}

\item{\code{language}}{character the language the code is written in (e.g. Python}

\item{\code{modelId}}{character the Id of the Prime model}

\item{\code{parentModelId}}{character the ID of the model this code approximates}

\item{\code{projectId}}{character the ID of the project the file belongs to}

\item{\code{rulesetId}}{integer the ID of the ruleset this code uses to approximate the parent model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PrimeFileResponse$new()}}
\item \href{#method-validate}{\code{PrimeFileResponse$validate()}}
\item \href{#method-toJSON}{\code{PrimeFileResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{PrimeFileResponse$fromJSON()}}
\item \href{#method-clone}{\code{PrimeFileResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimeFileResponse$new(
  id = NULL,
  isValid = NULL,
  language = NULL,
  modelId = NULL,
  parentModelId = NULL,
  projectId = NULL,
  rulesetId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the ID of the file}

\item{\code{isValid}}{whether the code passed basic validation checks}

\item{\code{language}}{the language the code is written in (e.g. Python}

\item{\code{modelId}}{the Id of the Prime model}

\item{\code{parentModelId}}{the ID of the model this code approximates}

\item{\code{projectId}}{the ID of the project the file belongs to}

\item{\code{rulesetId}}{the ID of the ruleset this code uses to approximate the parent model}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimeFileResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimeFileResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimeFileResponse$fromJSON(PrimeFileResponseJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PrimeFileResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrimeFileResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
