% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_secondary_dataset_config_response.R
\docType{class}
\name{ProjectSecondaryDatasetConfigResponse}
\alias{ProjectSecondaryDatasetConfigResponse}
\title{ProjectSecondaryDatasetConfigResponse}
\format{
An \code{R6Class} generator object
}
\description{
ProjectSecondaryDatasetConfigResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{config}}{list( \link{SecondaryDatasetConfig} ) \link{optional} Graph-specific secondary datasets. Deprecated in version v2.23.}

\item{\code{created}}{character DR-formatted datetime, null for legacy (before DR 6.0) db records.}

\item{\code{creatorFullName}}{character Fullname or email of the user created this config. null for legacy (before DR 6.0) db records.}

\item{\code{creatorUserId}}{character ID of the user created this config, null for legacy (before DR 6.0) db records.}

\item{\code{credentialIds}}{list( \link{DatasetsCredential} ) \link{optional} List of credentials used by the secondary datasets if the datasets used in the configuration are from datasource.}

\item{\code{featurelistId}}{character \link{optional} Id of the feature list.}

\item{\code{id}}{character ID of the secondary datasets configuration.}

\item{\code{isDefault}}{character Secondary datasets config is default config or not.}

\item{\code{name}}{character Name of the secondary datasets config.}

\item{\code{projectId}}{character ID of the project.}

\item{\code{projectVersion}}{character \link{optional} DataRobot project version.}

\item{\code{secondaryDatasets}}{list( \link{SecondaryDatasetResponse} ) \link{optional} List of secondary datasets used in the config.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProjectSecondaryDatasetConfigResponse-new}{\code{ProjectSecondaryDatasetConfigResponse$new()}}
\item \href{#method-ProjectSecondaryDatasetConfigResponse-validate}{\code{ProjectSecondaryDatasetConfigResponse$validate()}}
\item \href{#method-ProjectSecondaryDatasetConfigResponse-toJSON}{\code{ProjectSecondaryDatasetConfigResponse$toJSON()}}
\item \href{#method-ProjectSecondaryDatasetConfigResponse-fromJSON}{\code{ProjectSecondaryDatasetConfigResponse$fromJSON()}}
\item \href{#method-ProjectSecondaryDatasetConfigResponse-clone}{\code{ProjectSecondaryDatasetConfigResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSecondaryDatasetConfigResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSecondaryDatasetConfigResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSecondaryDatasetConfigResponse$new(
  created = NULL,
  creatorFullName = NULL,
  creatorUserId = NULL,
  id = NULL,
  isDefault = NULL,
  name = NULL,
  projectId = NULL,
  config = NULL,
  credentialIds = NULL,
  featurelistId = NULL,
  projectVersion = NULL,
  secondaryDatasets = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{created}}{DR-formatted datetime, null for legacy (before DR 6.0) db records.}

\item{\code{creatorFullName}}{Fullname or email of the user created this config. null for legacy (before DR 6.0) db records.}

\item{\code{creatorUserId}}{ID of the user created this config, null for legacy (before DR 6.0) db records.}

\item{\code{id}}{ID of the secondary datasets configuration.}

\item{\code{isDefault}}{Secondary datasets config is default config or not.}

\item{\code{name}}{Name of the secondary datasets config.}

\item{\code{projectId}}{ID of the project.}

\item{\code{config}}{Graph-specific secondary datasets. Deprecated in version v2.23.}

\item{\code{credentialIds}}{List of credentials used by the secondary datasets if the datasets used in the configuration are from datasource.}

\item{\code{featurelistId}}{Id of the feature list.}

\item{\code{projectVersion}}{DataRobot project version.}

\item{\code{secondaryDatasets}}{List of secondary datasets used in the config.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSecondaryDatasetConfigResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSecondaryDatasetConfigResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSecondaryDatasetConfigResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSecondaryDatasetConfigResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSecondaryDatasetConfigResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSecondaryDatasetConfigResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSecondaryDatasetConfigResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSecondaryDatasetConfigResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSecondaryDatasetConfigResponse$fromJSON(
  ProjectSecondaryDatasetConfigResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ProjectSecondaryDatasetConfigResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectSecondaryDatasetConfigResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectSecondaryDatasetConfigResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectSecondaryDatasetConfigResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
