% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating_table_create_response.R
\docType{class}
\name{RatingTableCreateResponse}
\alias{RatingTableCreateResponse}
\title{RatingTableCreateResponse}
\format{
An \code{R6Class} generator object
}
\description{
RatingTableCreateResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ratingTableId}}{character the ID of the newly created rating table}

\item{\code{ratingTableName}}{character The name that was used for the rating table. May differ from the ratingTableName in the request, as names are trimmed and a suffix added to ensure all rating tables derived from the same model have unique names}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RatingTableCreateResponse-new}{\code{RatingTableCreateResponse$new()}}
\item \href{#method-RatingTableCreateResponse-validate}{\code{RatingTableCreateResponse$validate()}}
\item \href{#method-RatingTableCreateResponse-toJSON}{\code{RatingTableCreateResponse$toJSON()}}
\item \href{#method-RatingTableCreateResponse-fromJSON}{\code{RatingTableCreateResponse$fromJSON()}}
\item \href{#method-RatingTableCreateResponse-clone}{\code{RatingTableCreateResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RatingTableCreateResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-RatingTableCreateResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RatingTableCreateResponse$new(
  ratingTableId = NULL,
  ratingTableName = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ratingTableId}}{the ID of the newly created rating table}

\item{\code{ratingTableName}}{The name that was used for the rating table. May differ from the ratingTableName in the request, as names are trimmed and a suffix added to ensure all rating tables derived from the same model have unique names}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RatingTableCreateResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-RatingTableCreateResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RatingTableCreateResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RatingTableCreateResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RatingTableCreateResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RatingTableCreateResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RatingTableCreateResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RatingTableCreateResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RatingTableCreateResponse$fromJSON(
  RatingTableCreateResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{RatingTableCreateResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RatingTableCreateResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RatingTableCreateResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RatingTableCreateResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
