% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationships_config_create.R
\docType{class}
\name{RelationshipsConfigCreate}
\alias{RelationshipsConfigCreate}
\title{RelationshipsConfigCreate}
\format{
An \code{R6Class} generator object
}
\description{
RelationshipsConfigCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datasetDefinitions}}{list( \link{DatasetDefinition} ) List of dataset definitions.}

\item{\code{featureDiscoveryMode}}{character \link{optional} Mode of feature discovery. Supported values are 'default' and 'manual'.}

\item{\code{featureDiscoverySettings}}{list( \link{FeatureDiscoverySetting} ) \link{optional} List of feature discovery settings used to customize the feature discovery process. Applicable when feature_discovery_mode is 'manual'.}

\item{\code{relationships}}{list( \link{Relationship} ) List of relationship between datasets specified in datasetDefinitions.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RelationshipsConfigCreate-new}{\code{RelationshipsConfigCreate$new()}}
\item \href{#method-RelationshipsConfigCreate-validate}{\code{RelationshipsConfigCreate$validate()}}
\item \href{#method-RelationshipsConfigCreate-toJSON}{\code{RelationshipsConfigCreate$toJSON()}}
\item \href{#method-RelationshipsConfigCreate-fromJSON}{\code{RelationshipsConfigCreate$fromJSON()}}
\item \href{#method-RelationshipsConfigCreate-clone}{\code{RelationshipsConfigCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RelationshipsConfigCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-RelationshipsConfigCreate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationshipsConfigCreate$new(
  datasetDefinitions = NULL,
  relationships = NULL,
  featureDiscoveryMode = NULL,
  featureDiscoverySettings = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasetDefinitions}}{List of dataset definitions.}

\item{\code{relationships}}{List of relationship between datasets specified in datasetDefinitions.}

\item{\code{featureDiscoveryMode}}{Mode of feature discovery. Supported values are 'default' and 'manual'.}

\item{\code{featureDiscoverySettings}}{List of feature discovery settings used to customize the feature discovery process. Applicable when feature_discovery_mode is 'manual'.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RelationshipsConfigCreate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-RelationshipsConfigCreate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationshipsConfigCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RelationshipsConfigCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RelationshipsConfigCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationshipsConfigCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RelationshipsConfigCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RelationshipsConfigCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationshipsConfigCreate$fromJSON(
  RelationshipsConfigCreateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{RelationshipsConfigCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RelationshipsConfigCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RelationshipsConfigCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationshipsConfigCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
