% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_metadata_value.R
\docType{class}
\name{RequiredMetadataValue}
\alias{RequiredMetadataValue}
\title{RequiredMetadataValue}
\format{
An \code{R6Class} generator object
}
\description{
RequiredMetadataValue Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fieldName}}{character The required field name. This value will be added as an environment vairable when running custom models.}

\item{\code{value}}{character The value for the given field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RequiredMetadataValue-new}{\code{RequiredMetadataValue$new()}}
\item \href{#method-RequiredMetadataValue-validate}{\code{RequiredMetadataValue$validate()}}
\item \href{#method-RequiredMetadataValue-toJSON}{\code{RequiredMetadataValue$toJSON()}}
\item \href{#method-RequiredMetadataValue-fromJSON}{\code{RequiredMetadataValue$fromJSON()}}
\item \href{#method-RequiredMetadataValue-clone}{\code{RequiredMetadataValue$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequiredMetadataValue-new"></a>}}
\if{latex}{\out{\hypertarget{method-RequiredMetadataValue-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequiredMetadataValue$new(
  fieldName = NULL,
  value = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fieldName}}{The required field name. This value will be added as an environment vairable when running custom models.}

\item{\code{value}}{The value for the given field.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequiredMetadataValue-validate"></a>}}
\if{latex}{\out{\hypertarget{method-RequiredMetadataValue-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequiredMetadataValue$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequiredMetadataValue-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RequiredMetadataValue-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequiredMetadataValue$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequiredMetadataValue-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RequiredMetadataValue-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequiredMetadataValue$fromJSON(
  RequiredMetadataValueJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{RequiredMetadataValueJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RequiredMetadataValue-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RequiredMetadataValue-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RequiredMetadataValue$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
