% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data_engine_query_response.R
\docType{class}
\name{RetrieveDataEngineQueryResponse}
\alias{RetrieveDataEngineQueryResponse}
\title{RetrieveDataEngineQueryResponse}
\format{
An \code{R6Class} generator object
}
\description{
RetrieveDataEngineQueryResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datasets}}{list( \link{DataEngineDataset} ) Source datasets in the Data Engine workspace.}

\item{\code{generatorSettings}}{\link{GeneratorSettings}}

\item{\code{generatorType}}{character Type of data engine query generator}

\item{\code{id}}{character The ID of the data engine query generator.}

\item{\code{query}}{character Generated SparkSQL query.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RetrieveDataEngineQueryResponse$new()}}
\item \href{#method-validate}{\code{RetrieveDataEngineQueryResponse$validate()}}
\item \href{#method-toJSON}{\code{RetrieveDataEngineQueryResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{RetrieveDataEngineQueryResponse$fromJSON()}}
\item \href{#method-clone}{\code{RetrieveDataEngineQueryResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrieveDataEngineQueryResponse$new(
  datasets = NULL,
  generatorSettings = NULL,
  generatorType = NULL,
  id = NULL,
  query = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasets}}{Source datasets in the Data Engine workspace.}

\item{\code{generatorType}}{Type of data engine query generator}

\item{\code{id}}{The ID of the data engine query generator.}

\item{\code{query}}{Generated SparkSQL query.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrieveDataEngineQueryResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrieveDataEngineQueryResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrieveDataEngineQueryResponse$fromJSON(
  RetrieveDataEngineQueryResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{RetrieveDataEngineQueryResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RetrieveDataEngineQueryResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
