% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_dataset_create.R
\docType{class}
\name{SecondaryDatasetCreate}
\alias{SecondaryDatasetCreate}
\title{SecondaryDatasetCreate}
\format{
An \code{R6Class} generator object
}
\description{
SecondaryDatasetCreate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{config}}{list( \link{SecondaryDatasetConfig} ) \link{optional} Graph-specific secondary datasets. Deprecated in version v2.23}

\item{\code{featurelistId}}{character \link{optional} Feature list ID of the configuration.}

\item{\code{modelId}}{character \link{optional} ID of the model.}

\item{\code{name}}{character Name of the configuration.}

\item{\code{save}}{character \link{optional} ID of the model.}

\item{\code{secondaryDatasets}}{list( \link{SecondaryDataset} ) \link{optional} Whether to save the configuration. If set to False the configuration is validated but not saved. Defaults to True.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SecondaryDatasetCreate-new}{\code{SecondaryDatasetCreate$new()}}
\item \href{#method-SecondaryDatasetCreate-validate}{\code{SecondaryDatasetCreate$validate()}}
\item \href{#method-SecondaryDatasetCreate-toJSON}{\code{SecondaryDatasetCreate$toJSON()}}
\item \href{#method-SecondaryDatasetCreate-fromJSON}{\code{SecondaryDatasetCreate$fromJSON()}}
\item \href{#method-SecondaryDatasetCreate-clone}{\code{SecondaryDatasetCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondaryDatasetCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-SecondaryDatasetCreate-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryDatasetCreate$new(
  name = NULL,
  config = NULL,
  featurelistId = NULL,
  modelId = NULL,
  save = TRUE,
  secondaryDatasets = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the configuration.}

\item{\code{config}}{Graph-specific secondary datasets. Deprecated in version v2.23}

\item{\code{featurelistId}}{Feature list ID of the configuration.}

\item{\code{modelId}}{ID of the model.}

\item{\code{save}}{ID of the model.}

\item{\code{secondaryDatasets}}{Whether to save the configuration. If set to False the configuration is validated but not saved. Defaults to True.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondaryDatasetCreate-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SecondaryDatasetCreate-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryDatasetCreate$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondaryDatasetCreate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SecondaryDatasetCreate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryDatasetCreate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondaryDatasetCreate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SecondaryDatasetCreate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryDatasetCreate$fromJSON(
  SecondaryDatasetCreateJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SecondaryDatasetCreateJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecondaryDatasetCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SecondaryDatasetCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryDatasetCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
