% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_dataset_response.R
\docType{class}
\name{SecondaryDatasetResponse}
\alias{SecondaryDatasetResponse}
\title{SecondaryDatasetResponse}
\format{
An \code{R6Class} generator object
}
\description{
SecondaryDatasetResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalogId}}{character ID of the catalog item.}

\item{\code{catalogVersionId}}{character ID of the catalog version.}

\item{\code{identifier}}{character Short name of this table (used directly as part of generated feature names).}

\item{\code{requiredFeatures}}{list( character ) \link{optional} List of required feature names used by the table.}

\item{\code{snapshotPolicy}}{character Policy to be used by a dataset while making prediction.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SecondaryDatasetResponse$new()}}
\item \href{#method-validate}{\code{SecondaryDatasetResponse$validate()}}
\item \href{#method-toJSON}{\code{SecondaryDatasetResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{SecondaryDatasetResponse$fromJSON()}}
\item \href{#method-clone}{\code{SecondaryDatasetResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryDatasetResponse$new(
  catalogId = NULL,
  catalogVersionId = NULL,
  identifier = NULL,
  snapshotPolicy = NULL,
  requiredFeatures = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalogId}}{ID of the catalog item.}

\item{\code{catalogVersionId}}{ID of the catalog version.}

\item{\code{identifier}}{Short name of this table (used directly as part of generated feature names).}

\item{\code{snapshotPolicy}}{Policy to be used by a dataset while making prediction.}

\item{\code{requiredFeatures}}{List of required feature names used by the table.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryDatasetResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryDatasetResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryDatasetResponse$fromJSON(
  SecondaryDatasetResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SecondaryDatasetResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecondaryDatasetResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
