% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_accuracy_retrieve_response.R
\docType{class}
\name{SeriesAccuracyRetrieveResponse}
\alias{SeriesAccuracyRetrieveResponse}
\title{SeriesAccuracyRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
SeriesAccuracyRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{SeriesAccuracyRetrieveDataResponse} ) An array of available multiseries identifiers and column values.}

\item{\code{next_}}{character \link{optional} A URL pointing to the next page (if null, there is no next page).}

\item{\code{previous}}{character \link{optional} A URL pointing to the previous page (if null, there is no previous page).}

\item{\code{querySeriesCount}}{integer The total number of series after filtering is applied.}

\item{\code{totalSeriesCount}}{integer The total number of series in the project dataset.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SeriesAccuracyRetrieveResponse$new()}}
\item \href{#method-validate}{\code{SeriesAccuracyRetrieveResponse$validate()}}
\item \href{#method-toJSON}{\code{SeriesAccuracyRetrieveResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{SeriesAccuracyRetrieveResponse$fromJSON()}}
\item \href{#method-clone}{\code{SeriesAccuracyRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeriesAccuracyRetrieveResponse$new(
  data = NULL,
  querySeriesCount = NULL,
  totalSeriesCount = NULL,
  next_ = NULL,
  previous = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{An array of available multiseries identifiers and column values.}

\item{\code{querySeriesCount}}{The total number of series after filtering is applied.}

\item{\code{totalSeriesCount}}{The total number of series in the project dataset.}

\item{\code{next_}}{A URL pointing to the next page (if null, there is no next page).}

\item{\code{previous}}{A URL pointing to the previous page (if null, there is no previous page).}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeriesAccuracyRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeriesAccuracyRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeriesAccuracyRetrieveResponse$fromJSON(
  SeriesAccuracyRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SeriesAccuracyRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeriesAccuracyRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
