% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shap_impact_retrieve_response.R
\docType{class}
\name{ShapImpactRetrieveResponse}
\alias{ShapImpactRetrieveResponse}
\title{ShapImpactRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
ShapImpactRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{integer The number of shapImpact object returned}

\item{\code{rowCount}}{integer \link{optional} The number of rows from dataset to use.}

\item{\code{shapImpacts}}{list( \link{ShapImpactsResponse} ) A list which contains shap impact scores for top 1000 features used by a model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ShapImpactRetrieveResponse-new}{\code{ShapImpactRetrieveResponse$new()}}
\item \href{#method-ShapImpactRetrieveResponse-validate}{\code{ShapImpactRetrieveResponse$validate()}}
\item \href{#method-ShapImpactRetrieveResponse-toJSON}{\code{ShapImpactRetrieveResponse$toJSON()}}
\item \href{#method-ShapImpactRetrieveResponse-fromJSON}{\code{ShapImpactRetrieveResponse$fromJSON()}}
\item \href{#method-ShapImpactRetrieveResponse-clone}{\code{ShapImpactRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShapImpactRetrieveResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ShapImpactRetrieveResponse-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapImpactRetrieveResponse$new(
  count = NULL,
  shapImpacts = NULL,
  rowCount = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{The number of shapImpact object returned}

\item{\code{shapImpacts}}{A list which contains shap impact scores for top 1000 features used by a model}

\item{\code{rowCount}}{The number of rows from dataset to use.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShapImpactRetrieveResponse-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ShapImpactRetrieveResponse-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapImpactRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShapImpactRetrieveResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ShapImpactRetrieveResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapImpactRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShapImpactRetrieveResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ShapImpactRetrieveResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapImpactRetrieveResponse$fromJSON(
  ShapImpactRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ShapImpactRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShapImpactRetrieveResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ShapImpactRetrieveResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapImpactRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
