% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shap_matrix_list_data_field.R
\docType{class}
\name{ShapMatrixListDataField}
\alias{ShapMatrixListDataField}
\title{ShapMatrixListDataField}
\format{
An \code{R6Class} generator object
}
\description{
ShapMatrixListDataField Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datasetId}}{character The dataset ID.}

\item{\code{id}}{character The ID of the SHAP matrix record.}

\item{\code{metadata}}{\link{ShapMatrixMetadataField}}

\item{\code{modelId}}{character The model ID.}

\item{\code{projectId}}{character The project ID.}

\item{\code{url}}{character The url at which you can retrieve the SHAP matrix.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ShapMatrixListDataField$new()}}
\item \href{#method-validate}{\code{ShapMatrixListDataField$validate()}}
\item \href{#method-toJSON}{\code{ShapMatrixListDataField$toJSON()}}
\item \href{#method-fromJSON}{\code{ShapMatrixListDataField$fromJSON()}}
\item \href{#method-clone}{\code{ShapMatrixListDataField$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapMatrixListDataField$new(
  datasetId = NULL,
  id = NULL,
  metadata = NULL,
  modelId = NULL,
  projectId = NULL,
  url = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasetId}}{The dataset ID.}

\item{\code{id}}{The ID of the SHAP matrix record.}

\item{\code{modelId}}{The model ID.}

\item{\code{projectId}}{The project ID.}

\item{\code{url}}{The url at which you can retrieve the SHAP matrix.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapMatrixListDataField$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapMatrixListDataField$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapMatrixListDataField$fromJSON(
  ShapMatrixListDataFieldJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ShapMatrixListDataFieldJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShapMatrixListDataField$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
