% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_retrieve_response.R
\docType{class}
\name{StatusRetrieveResponse}
\alias{StatusRetrieveResponse}
\title{StatusRetrieveResponse}
\format{
An \code{R6Class} generator object
}
\description{
StatusRetrieveResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{code}}{integer If no error occurred, 0.  Otherwise, may contain a status code reflecting the error.}

\item{\code{created}}{character The time the status record was created.}

\item{\code{description}}{character A description of the Task being performed, if applicable.}

\item{\code{message}}{character May contain further information about the status.}

\item{\code{status}}{character The status of the task.}

\item{\code{statusId}}{character The ID of the status object.}

\item{\code{statusType}}{character \link{optional} The type of thing being created by the Task, if applicable.}

\item{\code{type}}{character \link{optional} The type of thing being created by the Task, if applicable.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StatusRetrieveResponse$new()}}
\item \href{#method-validate}{\code{StatusRetrieveResponse$validate()}}
\item \href{#method-toJSON}{\code{StatusRetrieveResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{StatusRetrieveResponse$fromJSON()}}
\item \href{#method-clone}{\code{StatusRetrieveResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatusRetrieveResponse$new(
  code = NULL,
  created = NULL,
  description = NULL,
  message = NULL,
  status = NULL,
  statusId = NULL,
  statusType = NULL,
  type = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{If no error occurred, 0.  Otherwise, may contain a status code reflecting the error.}

\item{\code{created}}{The time the status record was created.}

\item{\code{description}}{A description of the Task being performed, if applicable.}

\item{\code{message}}{May contain further information about the status.}

\item{\code{status}}{The status of the task.}

\item{\code{statusId}}{The ID of the status object.}

\item{\code{statusType}}{The type of thing being created by the Task, if applicable.}

\item{\code{type}}{The type of thing being created by the Task, if applicable.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatusRetrieveResponse$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatusRetrieveResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatusRetrieveResponse$fromJSON(
  StatusRetrieveResponseJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{StatusRetrieveResponseJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatusRetrieveResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
