% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\docType{class}
\name{Transformation}
\alias{Transformation}
\title{Transformation}
\format{
An \code{R6Class} generator object
}
\description{
Transformation Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{enabled}}{character \link{optional} Whether this transformation is enabled by default}

\item{\code{name}}{character Transformation name}

\item{\code{params}}{list( \link{TransformationParam} ) \link{optional} Config values for transformation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Transformation-new}{\code{Transformation$new()}}
\item \href{#method-Transformation-validate}{\code{Transformation$validate()}}
\item \href{#method-Transformation-toJSON}{\code{Transformation$toJSON()}}
\item \href{#method-Transformation-fromJSON}{\code{Transformation$fromJSON()}}
\item \href{#method-Transformation-clone}{\code{Transformation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Transformation-new"></a>}}
\if{latex}{\out{\hypertarget{method-Transformation-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transformation$new(
  name = NULL,
  enabled = FALSE,
  params = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Transformation name}

\item{\code{enabled}}{Whether this transformation is enabled by default}

\item{\code{params}}{Config values for transformation}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Transformation-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Transformation-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transformation$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Transformation-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Transformation-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transformation$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Transformation-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Transformation-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transformation$fromJSON(TransformationJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TransformationJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Transformation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Transformation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transformation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
