% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_activity.R
\docType{class}
\name{UseCaseActivity}
\alias{UseCaseActivity}
\title{UseCaseActivity}
\format{
An \code{R6Class} generator object
}
\description{
UseCaseActivity Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{eventDescription}}{\link{EventDescriptions}}

\item{\code{eventType}}{character The type of activity}

\item{\code{firstName}}{character \link{optional} First name of use case owner}

\item{\code{id}}{character \link{optional} The id of the use case activity}

\item{\code{lastName}}{character \link{optional} Last of use case owner}

\item{\code{timestamp}}{character The timestamp of the activity}

\item{\code{user}}{\link{DrUserId}}

\item{\code{username}}{character \link{optional} Username of use case owner}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UseCaseActivity$new()}}
\item \href{#method-validate}{\code{UseCaseActivity$validate()}}
\item \href{#method-toJSON}{\code{UseCaseActivity$toJSON()}}
\item \href{#method-fromJSON}{\code{UseCaseActivity$fromJSON()}}
\item \href{#method-clone}{\code{UseCaseActivity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseActivity$new(
  eventDescription = NULL,
  eventType = NULL,
  timestamp = NULL,
  user = NULL,
  firstName = NULL,
  id = NULL,
  lastName = NULL,
  username = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{eventType}}{The type of activity}

\item{\code{timestamp}}{The timestamp of the activity}

\item{\code{firstName}}{First name of use case owner}

\item{\code{id}}{The id of the use case activity}

\item{\code{lastName}}{Last of use case owner}

\item{\code{username}}{Username of use case owner}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseActivity$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseActivity$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseActivity$fromJSON(UseCaseActivityJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UseCaseActivityJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseActivity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
