% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_attachment_application_overview.R
\docType{class}
\name{UseCaseAttachmentApplicationOverview}
\alias{UseCaseAttachmentApplicationOverview}
\title{UseCaseAttachmentApplicationOverview}
\format{
An \code{R6Class} generator object
}
\description{
UseCaseAttachmentApplicationOverview Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{appType}}{character Type of the application}

\item{\code{createdAt}}{character Time when application was created}

\item{\code{lastModifiedAt}}{character Time of the last modification}

\item{\code{name}}{character Name of the application}

\item{\code{version}}{character Version string of the app}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UseCaseAttachmentApplicationOverview-new}{\code{UseCaseAttachmentApplicationOverview$new()}}
\item \href{#method-UseCaseAttachmentApplicationOverview-validate}{\code{UseCaseAttachmentApplicationOverview$validate()}}
\item \href{#method-UseCaseAttachmentApplicationOverview-toJSON}{\code{UseCaseAttachmentApplicationOverview$toJSON()}}
\item \href{#method-UseCaseAttachmentApplicationOverview-fromJSON}{\code{UseCaseAttachmentApplicationOverview$fromJSON()}}
\item \href{#method-UseCaseAttachmentApplicationOverview-clone}{\code{UseCaseAttachmentApplicationOverview$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseAttachmentApplicationOverview-new"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseAttachmentApplicationOverview-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentApplicationOverview$new(
  appType = NULL,
  createdAt = NULL,
  lastModifiedAt = NULL,
  name = NULL,
  version = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{appType}}{Type of the application}

\item{\code{createdAt}}{Time when application was created}

\item{\code{lastModifiedAt}}{Time of the last modification}

\item{\code{name}}{Name of the application}

\item{\code{version}}{Version string of the app}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseAttachmentApplicationOverview-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseAttachmentApplicationOverview-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentApplicationOverview$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseAttachmentApplicationOverview-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseAttachmentApplicationOverview-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentApplicationOverview$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseAttachmentApplicationOverview-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseAttachmentApplicationOverview-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentApplicationOverview$fromJSON(
  UseCaseAttachmentApplicationOverviewJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UseCaseAttachmentApplicationOverviewJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UseCaseAttachmentApplicationOverview-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UseCaseAttachmentApplicationOverview-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentApplicationOverview$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
