% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_case_attachment_model_package_overview.R
\docType{class}
\name{UseCaseAttachmentModelPackageOverview}
\alias{UseCaseAttachmentModelPackageOverview}
\title{UseCaseAttachmentModelPackageOverview}
\format{
An \code{R6Class} generator object
}
\description{
UseCaseAttachmentModelPackageOverview Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{createdAt}}{character Time when model package was created}

\item{\code{deploymentCount}}{integer Number of deployments}

\item{\code{executionType}}{character Type of execution}

\item{\code{name}}{character Name of the model package}

\item{\code{targetName}}{character Name of modeling target}

\item{\code{targetType}}{character Type of modeling target}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UseCaseAttachmentModelPackageOverview$new()}}
\item \href{#method-validate}{\code{UseCaseAttachmentModelPackageOverview$validate()}}
\item \href{#method-toJSON}{\code{UseCaseAttachmentModelPackageOverview$toJSON()}}
\item \href{#method-fromJSON}{\code{UseCaseAttachmentModelPackageOverview$fromJSON()}}
\item \href{#method-clone}{\code{UseCaseAttachmentModelPackageOverview$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentModelPackageOverview$new(
  createdAt = NULL,
  deploymentCount = NULL,
  executionType = NULL,
  name = NULL,
  targetName = NULL,
  targetType = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{createdAt}}{Time when model package was created}

\item{\code{deploymentCount}}{Number of deployments}

\item{\code{executionType}}{Type of execution}

\item{\code{name}}{Name of the model package}

\item{\code{targetName}}{Name of modeling target}

\item{\code{targetType}}{Type of modeling target}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentModelPackageOverview$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentModelPackageOverview$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentModelPackageOverview$fromJSON(
  UseCaseAttachmentModelPackageOverviewJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UseCaseAttachmentModelPackageOverviewJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UseCaseAttachmentModelPackageOverview$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
