% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_blueprint_bulk_validation_request.R
\docType{class}
\name{UserBlueprintBulkValidationRequest}
\alias{UserBlueprintBulkValidationRequest}
\title{UserBlueprintBulkValidationRequest}
\format{
An \code{R6Class} generator object
}
\description{
UserBlueprintBulkValidationRequest Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{projectId}}{character \link{optional} String representation of ObjectId for the currently active project. The user blueprint is validated when this project is active. Necessary in the event of project specific tasks, such as column selection tasks.}

\item{\code{userBlueprintIds}}{list( character ) The ids of the user blueprints to validate in bulk.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UserBlueprintBulkValidationRequest$new()}}
\item \href{#method-validate}{\code{UserBlueprintBulkValidationRequest$validate()}}
\item \href{#method-toJSON}{\code{UserBlueprintBulkValidationRequest$toJSON()}}
\item \href{#method-fromJSON}{\code{UserBlueprintBulkValidationRequest$fromJSON()}}
\item \href{#method-clone}{\code{UserBlueprintBulkValidationRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintBulkValidationRequest$new(
  userBlueprintIds = NULL,
  projectId = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userBlueprintIds}}{The ids of the user blueprints to validate in bulk.}

\item{\code{projectId}}{String representation of ObjectId for the currently active project. The user blueprint is validated when this project is active. Necessary in the event of project specific tasks, such as column selection tasks.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintBulkValidationRequest$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintBulkValidationRequest$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintBulkValidationRequest$fromJSON(
  UserBlueprintBulkValidationRequestJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserBlueprintBulkValidationRequestJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintBulkValidationRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
