% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/user_blueprint_task_custom_task_metadata_with_arguments.R
\docType{class}
\name{UserBlueprintTaskCustomTaskMetadataWithArguments}
\alias{UserBlueprintTaskCustomTaskMetadataWithArguments}
\title{UserBlueprintTaskCustomTaskMetadataWithArguments}
\format{
An \code{R6Class} generator object
}
\description{
UserBlueprintTaskCustomTaskMetadataWithArguments Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character Id of the custom task version. The ID can be latest_<task_id> which implies to use the latest version of that custom task.}

\item{\code{label}}{character The name of the custom task version.}

\item{\code{versionMajor}}{integer Major version of the custom task.}

\item{\code{versionMinor}}{integer Minor version of the custom task.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UserBlueprintTaskCustomTaskMetadataWithArguments$new()}}
\item \href{#method-validate}{\code{UserBlueprintTaskCustomTaskMetadataWithArguments$validate()}}
\item \href{#method-toJSON}{\code{UserBlueprintTaskCustomTaskMetadataWithArguments$toJSON()}}
\item \href{#method-fromJSON}{\code{UserBlueprintTaskCustomTaskMetadataWithArguments$fromJSON()}}
\item \href{#method-clone}{\code{UserBlueprintTaskCustomTaskMetadataWithArguments$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskCustomTaskMetadataWithArguments$new(
  id = NULL,
  label = NULL,
  versionMajor = NULL,
  versionMinor = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Id of the custom task version. The ID can be latest_<task_id> which implies to use the latest version of that custom task.}

\item{\code{label}}{The name of the custom task version.}

\item{\code{versionMajor}}{Major version of the custom task.}

\item{\code{versionMinor}}{Minor version of the custom task.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskCustomTaskMetadataWithArguments$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskCustomTaskMetadataWithArguments$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskCustomTaskMetadataWithArguments$fromJSON(
  UserBlueprintTaskCustomTaskMetadataWithArgumentsJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserBlueprintTaskCustomTaskMetadataWithArgumentsJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskCustomTaskMetadataWithArguments$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
