% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/user_blueprint_task_parameter_validation_request_param_item.R
\docType{class}
\name{UserBlueprintTaskParameterValidationRequestParamItem}
\alias{UserBlueprintTaskParameterValidationRequestParamItem}
\title{UserBlueprintTaskParameterValidationRequestParamItem}
\format{
An \code{R6Class} generator object
}
\description{
UserBlueprintTaskParameterValidationRequestParamItem Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{newValue}}{\link{OneOfAnyTypearray} The proposed value for the task parameter.}

\item{\code{paramName}}{character The name of the task parameter to be validated.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UserBlueprintTaskParameterValidationRequestParamItem$new()}}
\item \href{#method-validate}{\code{UserBlueprintTaskParameterValidationRequestParamItem$validate()}}
\item \href{#method-toJSON}{\code{UserBlueprintTaskParameterValidationRequestParamItem$toJSON()}}
\item \href{#method-fromJSON}{\code{UserBlueprintTaskParameterValidationRequestParamItem$fromJSON()}}
\item \href{#method-clone}{\code{UserBlueprintTaskParameterValidationRequestParamItem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskParameterValidationRequestParamItem$new(
  newValue = NULL,
  paramName = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The proposed value for the task parameter.}

\item{\code{paramName}}{The name of the task parameter to be validated.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskParameterValidationRequestParamItem$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskParameterValidationRequestParamItem$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskParameterValidationRequestParamItem$fromJSON(
  UserBlueprintTaskParameterValidationRequestParamItemJson,
  validateParams = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserBlueprintTaskParameterValidationRequestParamItemJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserBlueprintTaskParameterValidationRequestParamItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
