% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_in_group.R
\docType{class}
\name{UserInGroup}
\alias{UserInGroup}
\title{UserInGroup}
\format{
An \code{R6Class} generator object
}
\description{
UserInGroup Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{expirationDate}}{character \link{optional} The expiration date of the user account, if applicable}

\item{\code{firstName}}{character The first name of the user.}

\item{\code{lastName}}{character The last name of the user.}

\item{\code{organization}}{character The name of the organization the user is part of.}

\item{\code{scheduledForDeletion}}{character \link{optional} If the user is scheduled for deletion. Will be returned when an appropriate FF is enabled.}

\item{\code{status}}{character The status of the user, `active` if the has been activated, `inactive` otherwise.}

\item{\code{userId}}{character The identifier of the user.}

\item{\code{username}}{character The name of the user.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserInGroup-new}{\code{UserInGroup$new()}}
\item \href{#method-UserInGroup-validate}{\code{UserInGroup$validate()}}
\item \href{#method-UserInGroup-toJSON}{\code{UserInGroup$toJSON()}}
\item \href{#method-UserInGroup-fromJSON}{\code{UserInGroup$fromJSON()}}
\item \href{#method-UserInGroup-clone}{\code{UserInGroup$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserInGroup-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserInGroup-new}{}}}
\subsection{Method \code{new()}}{
A function used to initialize an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserInGroup$new(
  firstName = NULL,
  lastName = NULL,
  organization = NULL,
  status = NULL,
  userId = NULL,
  username = NULL,
  expirationDate = NULL,
  scheduledForDeletion = NULL,
  validateParams = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{firstName}}{The first name of the user.}

\item{\code{lastName}}{The last name of the user.}

\item{\code{organization}}{The name of the organization the user is part of.}

\item{\code{status}}{The status of the user, `active` if the has been activated, `inactive` otherwise.}

\item{\code{userId}}{The identifier of the user.}

\item{\code{username}}{The name of the user.}

\item{\code{expirationDate}}{The expiration date of the user account, if applicable}

\item{\code{scheduledForDeletion}}{If the user is scheduled for deletion. Will be returned when an appropriate FF is enabled.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters before initialization. Default FALSE.}

\item{\code{...}}{Any additional keyword arguments to be passed into this object for initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserInGroup-validate"></a>}}
\if{latex}{\out{\hypertarget{method-UserInGroup-validate}{}}}
\subsection{Method \code{validate()}}{
A helper function that provides public access to the private validateProps function. This allows users the ability
to programmatically validate objects before sending them to DataRobot.
checking this objects set properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserInGroup$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserInGroup-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserInGroup-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
A helper function that serializes this object into a JSON encoded string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserInGroup$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserInGroup-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserInGroup-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
A helper function that deserializes a JSON string into an instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserInGroup$fromJSON(UserInGroupJson, validateParams = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UserInGroupJson}}{A JSON encoded string representation of a class instance.}

\item{\code{validateParams}}{An optional param for auto validating this object's parameters after deserialization. Default FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserInGroup-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserInGroup-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserInGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
